Struct qt_gui::abstract_text_document_layout::PaintContext

pub struct PaintContext {
    // some fields omitted
}

C++ type: QAbstractTextDocumentLayout::PaintContext

C++ documentation:

The QAbstractTextDocumentLayout::PaintContext class is a convenience class defining the parameters used when painting a document's layout.

A paint context is used when rendering custom layouts for QTextDocuments with the QAbstractTextDocumentLayout::draw() function. It is specified by a cursor position, default text color, clip rectangle and a collection of selections.

Methods

impl PaintContext

fn clip<'l0>(&'l0 self) -> &'l0 RectF

C++ method: const QRectF& QAbstractTextDocumentLayout::PaintContext::clip() const

fn clip_mut<'l0>(&'l0 mut self) -> &'l0 mut RectF

C++ method: QRectF& QAbstractTextDocumentLayout::PaintContext::clip_mut()

fn cursor_position(&self) -> c_int

C++ method: int QAbstractTextDocumentLayout::PaintContext::cursorPosition() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: PaintContextNewArgs

C++ method: QAbstractTextDocumentLayout::PaintContext::PaintContext

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::abstract_text_document_layout::PaintContext
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::abstract_text_document_layout::PaintContext>
C++ method: [constructor] void QAbstractTextDocumentLayout::PaintContext::PaintContext()

fn palette<'l0>(&'l0 self) -> &'l0 Palette

C++ method: const QPalette& QAbstractTextDocumentLayout::PaintContext::palette() const

fn palette_mut<'l0>(&'l0 mut self) -> &'l0 mut Palette

C++ method: QPalette& QAbstractTextDocumentLayout::PaintContext::palette_mut()

fn selections<'l0>(&'l0 self) -> &'l0 VectorSelectionRef

C++ method: const QVector<QAbstractTextDocumentLayout::Selection>& QAbstractTextDocumentLayout::PaintContext::selections() const

fn selections_mut<'l0>(&'l0 mut self) -> &'l0 mut VectorSelectionRef

C++ method: QVector<QAbstractTextDocumentLayout::Selection>& QAbstractTextDocumentLayout::PaintContext::selections_mut()

fn set_clip(&mut self, value: &RectF)

C++ method: void QAbstractTextDocumentLayout::PaintContext::set_clip(QRectF value)

fn set_cursor_position(&mut self, value: c_int)

C++ method: void QAbstractTextDocumentLayout::PaintContext::set_cursorPosition(int value)

fn set_palette(&mut self, value: &Palette)

C++ method: void QAbstractTextDocumentLayout::PaintContext::set_palette(QPalette value)

fn set_selections(&mut self, value: &VectorSelectionRef)

C++ method: void QAbstractTextDocumentLayout::PaintContext::set_selections(QVector<QAbstractTextDocumentLayout::Selection> value)