Enum qt_gui::accessible::Event

pub enum Event {
    SoundPlayed,
    Alert,
    ForegroundChanged,
    MenuStart,
    MenuEnd,
    PopupMenuStart,
    PopupMenuEnd,
    ContextHelpStart,
    ContextHelpEnd,
    DragDropStart,
    DragDropEnd,
    DialogStart,
    DialogEnd,
    ScrollingStart,
    ScrollingEnd,
    MenuCommand,
    ActionChanged,
    ActiveDescendantChanged,
    AttributeChanged,
    DocumentContentChanged,
    DocumentLoadComplete,
    DocumentLoadStopped,
    DocumentReload,
    HyperlinkEndIndexChanged,
    HyperlinkNumberOfAnchorsChanged,
    HyperlinkSelectedLinkChanged,
    HypertextLinkActivated,
    HypertextLinkSelected,
    HyperlinkStartIndexChanged,
    HypertextChanged,
    HypertextNLinksChanged,
    ObjectAttributeChanged,
    PageChanged,
    SectionChanged,
    TableCaptionChanged,
    TableColumnDescriptionChanged,
    TableColumnHeaderChanged,
    TableModelChanged,
    TableRowDescriptionChanged,
    TableRowHeaderChanged,
    TableSummaryChanged,
    TextAttributeChanged,
    TextCaretMoved,
    TextColumnChanged,
    TextInserted,
    TextRemoved,
    TextUpdated,
    TextSelectionChanged,
    VisibleDataChanged,
    ObjectCreated,
    ObjectDestroyed,
    ObjectShow,
    ObjectHide,
    ObjectReorder,
    Focus,
    Selection,
    SelectionAdd,
    SelectionRemove,
    SelectionWithin,
    StateChanged,
    LocationChanged,
    NameChanged,
    DescriptionChanged,
    ValueChanged,
    ParentChanged,
    HelpChanged,
    DefaultActionChanged,
    AcceleratorChanged,
    InvalidEvent,
}

C++ type: QAccessible::Event

C++ documentation:

This enum type defines accessible event types.

Internal: Used when creating subclasses of QAccessibleEvent.

The values for this enum are defined to be the same as those defined in the IAccessible2 and MSAA specifications.

Variants
SoundPlayed

A sound has been played by an object The QAccessible::State of an object has changed. This value is used internally for the QAccessibleStateChangeEvent. (C++ enum variant: SoundPlayed = 1)

Alert

A system alert (e.g., a message from a QMessageBox) (C++ enum variant: Alert = 2)

ForegroundChanged

A window has been activated (i.e., a new window has gained focus on the desktop). (C++ enum variant: ForegroundChanged = 3)

MenuStart

A menu has been opened on the menubar (Qt uses PopupMenuStart for all menus). (C++ enum variant: MenuStart = 4)

MenuEnd

A menu has been closed (Qt uses PopupMenuEnd for all menus). (C++ enum variant: MenuEnd = 5)

PopupMenuStart

A pop-up menu has opened. (C++ enum variant: PopupMenuStart = 6)

PopupMenuEnd

A pop-up menu has closed. (C++ enum variant: PopupMenuEnd = 7)

ContextHelpStart

Context help (QWhatsThis) for an object is initiated. (C++ enum variant: ContextHelpStart = 12)

ContextHelpEnd

Context help (QWhatsThis) for an object is finished. (C++ enum variant: ContextHelpEnd = 13)

DragDropStart

A drag and drop operation is about to be initiated. (C++ enum variant: DragDropStart = 14)

DragDropEnd

A drag and drop operation is about to finished. (C++ enum variant: DragDropEnd = 15)

DialogStart

A dialog (QDialog) has been set visible. (C++ enum variant: DialogStart = 16)

DialogEnd

A dialog (QDialog) has been hidden (C++ enum variant: DialogEnd = 17)

ScrollingStart

A scrollbar scroll operation is about to start; this may be caused by a mouse press on the slider handle, for example. (C++ enum variant: ScrollingStart = 18)

ScrollingEnd

A scrollbar scroll operation has ended (the mouse has released the slider handle). (C++ enum variant: ScrollingEnd = 19)

MenuCommand

A menu item is triggered. (C++ enum variant: MenuCommand = 24)

ActionChanged

An action has been changed. (C++ enum variant: ActionChanged = 257)

ActiveDescendantChanged

(C++ enum variant: ActiveDescendantChanged = 258)

AttributeChanged

(C++ enum variant: AttributeChanged = 259)

DocumentContentChanged

The contents of a text document have changed. (C++ enum variant: DocumentContentChanged = 260)

DocumentLoadComplete

A document has been loaded. (C++ enum variant: DocumentLoadComplete = 261)

DocumentLoadStopped

A document load has been stopped. (C++ enum variant: DocumentLoadStopped = 262)

DocumentReload

A document reload has been initiated. (C++ enum variant: DocumentReload = 263)

HyperlinkEndIndexChanged

The end position of the display text for a hypertext link has changed. (C++ enum variant: HyperlinkEndIndexChanged = 264)

HyperlinkNumberOfAnchorsChanged

The number of anchors in a hypertext link has changed, perhaps because the display text has been split to provide more than one link. (C++ enum variant: HyperlinkNumberOfAnchorsChanged = 265)

HyperlinkSelectedLinkChanged

The link for the selected hypertext link has changed. (C++ enum variant: HyperlinkSelectedLinkChanged = 266)

HypertextLinkActivated

A hypertext link has been activated, perhaps by being clicked or via a key press. (C++ enum variant: HypertextLinkActivated = 267)

HypertextLinkSelected

A hypertext link has been selected. (C++ enum variant: HypertextLinkSelected = 268)

HyperlinkStartIndexChanged

The start position of the display text for a hypertext link has changed. (C++ enum variant: HyperlinkStartIndexChanged = 269)

HypertextChanged

The display text for a hypertext link has changed. (C++ enum variant: HypertextChanged = 270)

HypertextNLinksChanged

(C++ enum variant: HypertextNLinksChanged = 271)

ObjectAttributeChanged

(C++ enum variant: ObjectAttributeChanged = 272)

PageChanged

(C++ enum variant: PageChanged = 273)

SectionChanged

(C++ enum variant: SectionChanged = 274)

TableCaptionChanged

A table caption has been changed. (C++ enum variant: TableCaptionChanged = 275)

TableColumnDescriptionChanged

The description of a table column, typically found in the column's header, has been changed. (C++ enum variant: TableColumnDescriptionChanged = 276)

TableColumnHeaderChanged

A table column header has been changed. The model providing data for a table has been changed. (C++ enum variant: TableColumnHeaderChanged = 277)

TableModelChanged

C++ enum variant: TableModelChanged = 278

TableRowDescriptionChanged

The description of a table row, typically found in the row's header, has been changed. (C++ enum variant: TableRowDescriptionChanged = 279)

TableRowHeaderChanged

A table row header has been changed. (C++ enum variant: TableRowHeaderChanged = 280)

TableSummaryChanged

The summary of a table has been changed. The caret has moved in an editable widget. The caret represents the cursor position in an editable widget with the input focus. (C++ enum variant: TableSummaryChanged = 281)

TextAttributeChanged

C++ enum variant: TextAttributeChanged = 282

TextCaretMoved

C++ enum variant: TextCaretMoved = 283

TextColumnChanged

A text column has been changed. Text has been inserted into an editable widget. Text has been removed from an editable widget. The selected text has changed in an editable widget. The text has been update in an editable widget. The QAccessible::Value of an object has changed. (C++ enum variant: TextColumnChanged = 285)

TextInserted

C++ enum variant: TextInserted = 286

TextRemoved

C++ enum variant: TextRemoved = 287

TextUpdated

C++ enum variant: TextUpdated = 288

TextSelectionChanged

C++ enum variant: TextSelectionChanged = 289

VisibleDataChanged

(C++ enum variant: VisibleDataChanged = 290)

ObjectCreated

A new object is created. (C++ enum variant: ObjectCreated = 32768)

ObjectDestroyed

An object is deleted. (C++ enum variant: ObjectDestroyed = 32769)

ObjectShow

An object is displayed; for example, with QWidget::show(). (C++ enum variant: ObjectShow = 32770)

ObjectHide

An object is hidden; for example, with QWidget::hide(). Any children the object that is hidden has do not send this event. It is not sent when an object is hidden as it is being obcured by others. (C++ enum variant: ObjectHide = 32771)

ObjectReorder

A layout or item view has added, removed, or moved an object (Qt does not use this event). (C++ enum variant: ObjectReorder = 32772)

Focus

An object has gained keyboard focus. (C++ enum variant: Focus = 32773)

Selection

The selection has changed in a menu or item view. (C++ enum variant: Selection = 32774)

SelectionAdd

An item has been added to the selection in an item view. (C++ enum variant: SelectionAdd = 32775)

SelectionRemove

An item has been removed from an item view selection. (C++ enum variant: SelectionRemove = 32776)

SelectionWithin

Several changes to a selection has occurred in an item view. (C++ enum variant: SelectionWithin = 32777)

StateChanged

C++ enum variant: StateChanged = 32778

LocationChanged

An object's location on the screen has changed. (C++ enum variant: LocationChanged = 32779)

NameChanged

The QAccessible::Name property of an object has changed. (C++ enum variant: NameChanged = 32780)

DescriptionChanged

The object's QAccessible::Description changed. (C++ enum variant: DescriptionChanged = 32781)

ValueChanged

C++ enum variant: ValueChanged = 32782

ParentChanged

An object's parent object changed. (C++ enum variant: ParentChanged = 32783)

HelpChanged

The QAccessible::Help text property of an object has changed. (C++ enum variant: HelpChanged = 32928)

DefaultActionChanged

The default QAccessible::Action for the accessible object has changed. (C++ enum variant: DefaultActionChanged = 32944)

AcceleratorChanged

The keyboard accelerator for an action has been changed. (C++ enum variant: AcceleratorChanged = 32960)

InvalidEvent

C++ enum variant: InvalidEvent = 32961