Enum qt_gui::accessible::Role

pub enum Role {
    NoRole,
    TitleBar,
    MenuBar,
    ScrollBar,
    Grip,
    Sound,
    Cursor,
    Caret,
    AlertMessage,
    Window,
    Client,
    PopupMenu,
    MenuItem,
    ToolTip,
    Application,
    Document,
    Pane,
    Chart,
    Dialog,
    Border,
    Grouping,
    Separator,
    ToolBar,
    StatusBar,
    Table,
    ColumnHeader,
    RowHeader,
    Column,
    Row,
    Cell,
    Link,
    HelpBalloon,
    Assistant,
    List,
    ListItem,
    Tree,
    TreeItem,
    PageTab,
    PropertyPage,
    Indicator,
    Graphic,
    StaticText,
    EditableText,
    Button,
    CheckBox,
    RadioButton,
    ComboBox,
    ProgressBar,
    Dial,
    HotkeyField,
    Slider,
    SpinBox,
    Canvas,
    Animation,
    Equation,
    ButtonDropDown,
    ButtonMenu,
    ButtonDropGrid,
    Whitespace,
    PageTabList,
    Clock,
    Splitter,
    LayeredPane,
    Terminal,
    Desktop,
    Paragraph,
    WebDocument,
    Section,
    ColorChooser,
    Footer,
    Form,
    Heading,
    Note,
    ComplementaryContent,
    UserRole,
}

C++ type: QAccessible::Role

C++ documentation:

This enum defines the role of an accessible object. The roles are:

Variants
NoRole

The object has no role. This usually indicates an invalid object. (C++ enum variant: NoRole = 0)

TitleBar

The title bar caption of a window. (C++ enum variant: TitleBar = 1)

MenuBar

A menu bar from which menus are opened by the user. (C++ enum variant: MenuBar = 2)

ScrollBar

A scroll bar, which allows the user to scroll the visible area. (C++ enum variant: ScrollBar = 3)

Grip

A grip that the user can drag to change the size of widgets. (C++ enum variant: Grip = 4)

Sound

An object that represents a sound. (C++ enum variant: Sound = 5)

Cursor

An object that represents the mouse cursor. (C++ enum variant: Cursor = 6)

Caret

An object that represents the system caret (text cursor). (C++ enum variant: Caret = 7)

AlertMessage

An object that is used to alert the user. (C++ enum variant: AlertMessage = 8)

Window

A top level window. (C++ enum variant: Window = 9)

Client

The client area in a window. (C++ enum variant: Client = 10)

PopupMenu

A menu which lists options that the user can select to perform an action. (C++ enum variant: PopupMenu = 11)

MenuItem

An item in a menu or menu bar. (C++ enum variant: MenuItem = 12)

ToolTip

A tool tip which provides information about other objects. (C++ enum variant: ToolTip = 13)

Application

The application's main window. (C++ enum variant: Application = 14)

Document

A document, for example in an office application. (C++ enum variant: Document = 15)

Pane

A generic container. (C++ enum variant: Pane = 16)

Chart

An object that displays a graphical representation of data. (C++ enum variant: Chart = 17)

Dialog

A dialog box. (C++ enum variant: Dialog = 18)

Border

An object that represents a border. (C++ enum variant: Border = 19)

Grouping

An object that represents a logical grouping of other objects. (C++ enum variant: Grouping = 20)

Separator

A separator that divides space into logical areas. (C++ enum variant: Separator = 21)

ToolBar

A tool bar, which groups widgets that the user accesses frequently. (C++ enum variant: ToolBar = 22)

StatusBar

A status bar. (C++ enum variant: StatusBar = 23)

Table

A table representing data in a grid of rows and columns. (C++ enum variant: Table = 24)

ColumnHeader

A header for a column of data. (C++ enum variant: ColumnHeader = 25)

RowHeader

A header for a row of data. (C++ enum variant: RowHeader = 26)

Column

A column of cells, usually within a table. (C++ enum variant: Column = 27)

Row

A row of cells, usually within a table. (C++ enum variant: Row = 28)

Cell

A cell in a table. (C++ enum variant: Cell = 29)

Link

A link to something else. (C++ enum variant: Link = 30)

HelpBalloon

An object that displays help in a separate, short lived window. (C++ enum variant: HelpBalloon = 31)

Assistant

An object that provids interactive help. (C++ enum variant: Assistant = 32)

List

A list of items, from which the user can select one or more items. (C++ enum variant: List = 33)

ListItem

An item in a list of items. (C++ enum variant: ListItem = 34)

Tree

A list of items in a tree structure. (C++ enum variant: Tree = 35)

TreeItem

An item in a tree structure. (C++ enum variant: TreeItem = 36)

PageTab

A page tab that the user can select to switch to a different page in a dialog. (C++ enum variant: PageTab = 37)

PropertyPage

A property page where the user can change options and settings. (C++ enum variant: PropertyPage = 38)

Indicator

An indicator that represents a current value or item. (C++ enum variant: Indicator = 39)

Graphic

A graphic or picture, e.g. an icon. (C++ enum variant: Graphic = 40)

StaticText

Static text, such as labels for other widgets. (C++ enum variant: StaticText = 41)

EditableText

Editable text such as a line or text edit. (C++ enum variant: EditableText = 42)

Button

This variant corresponds to multiple C++ enum variants with the same value:

Button = 43: A button.

PushButton = 43

CheckBox

An object that represents an option that can be checked or unchecked. Some options provide a "mixed" state, e.g. neither checked nor unchecked. (C++ enum variant: CheckBox = 44)

RadioButton

An object that represents an option that is mutually exclusive with other options. (C++ enum variant: RadioButton = 45)

ComboBox

A list of choices that the user can select from. (C++ enum variant: ComboBox = 46)

ProgressBar

The object displays the progress of an operation in progress. (C++ enum variant: ProgressBar = 48)

Dial

An object that represents a dial or knob. (C++ enum variant: Dial = 49)

HotkeyField

A hotkey field that allows the user to enter a key sequence. (C++ enum variant: HotkeyField = 50)

Slider

A slider that allows the user to select a value within a given range. (C++ enum variant: Slider = 51)

SpinBox

A spin box widget that allows the user to enter a value within a given range. (C++ enum variant: SpinBox = 52)

Canvas

An object that displays graphics that the user can interact with. (C++ enum variant: Canvas = 53)

Animation

An object that displays an animation. (C++ enum variant: Animation = 54)

Equation

An object that represents a mathematical equation. (C++ enum variant: Equation = 55)

ButtonDropDown

A button that drops down a list of items. (C++ enum variant: ButtonDropDown = 56)

ButtonMenu

A button that drops down a menu. (C++ enum variant: ButtonMenu = 57)

ButtonDropGrid

A button that drops down a grid. (C++ enum variant: ButtonDropGrid = 58)

Whitespace

Blank space between other objects. (C++ enum variant: Whitespace = 59)

PageTabList

A list of page tabs. (C++ enum variant: PageTabList = 60)

Clock

A clock displaying time. (C++ enum variant: Clock = 61)

Splitter

A splitter distributing available space between its child widgets. (C++ enum variant: Splitter = 62)

LayeredPane

An object that can contain layered children, e.g. in a stack. (C++ enum variant: LayeredPane = 128)

Terminal

A terminal or command line interface. (C++ enum variant: Terminal = 129)

Desktop

The object represents the desktop or workspace. (C++ enum variant: Desktop = 130)

Paragraph

A paragraph of text (usually found in documents). (C++ enum variant: Paragraph = 131)

WebDocument

HTML document, usually in a browser. (C++ enum variant: WebDocument = 132)

Section

A section (in a document). (C++ enum variant: Section = 133)

ColorChooser

A dialog that lets the user choose a color. (C++ enum variant: ColorChooser = 1028)

Footer

A footer in a page (usually in documents). (C++ enum variant: Footer = 1038)

Form

A web form containing controls. (C++ enum variant: Form = 1040)

Heading

A heading in a document. (C++ enum variant: Heading = 1044)

Note

A section whose content is parenthetic or ancillary to the main content of the resource. (C++ enum variant: Note = 1051)

ComplementaryContent

A part of the document or web page that is complementary to the main content, usually a landmark (see WAI-ARIA). (C++ enum variant: ComplementaryContent = 1068)

UserRole

The first value to be used for user defined roles. (C++ enum variant: UserRole = 65535)