Enum qt_gui::accessible::Text

pub enum Text {
    Name,
    Description,
    Value,
    Help,
    Accelerator,
    DebugDescription,
    UserText,
}

C++ type: QAccessible::Text

C++ documentation:

This enum specifies string information that an accessible object returns.

Variants
Name

The name of the object. This can be used both as an identifier or a short description by accessible clients. (C++ enum variant: Name = 0)

Description

A short text describing the object. (C++ enum variant: Description = 1)

Value

The value of the object. (C++ enum variant: Value = 2)

Help

A longer text giving information about how to use the object. (C++ enum variant: Help = 3)

Accelerator

The keyboard shortcut that executes the object's default action. (C++ enum variant: Accelerator = 4)

DebugDescription

C++ enum variant: DebugDescription = 5

UserText

The first value to be used for user defined text. (C++ enum variant: UserText = 65535)