Struct qt_gui::accessible::State

pub struct State {
    // some fields omitted
}

C++ type: QAccessible::State

C++ documentation:

This structure defines bit flags that indicate the state of an accessible object. The values are:

Constant	Description

active	The object is the active window or the active sub-element in a container (that would get focus when focusing the container).

adjustable	The object represents an adjustable value, e.g. sliders.

animated	The object's appearance changes frequently.

busy	The object cannot accept input at the moment.

checkable	The object is checkable.

checked	The object's check box is checked.

checkStateMixed	The third state of checkboxes (half checked in tri-state check boxes).

collapsed	The object is collapsed, e.g. a closed listview item, or an iconified window.

defaultButton	The object represents the default button in a dialog.

defunct	The object no longer exists.

editable	The object has a text carret (and often implements the text interface).

expandable	The object is expandable, mostly used for cells in a tree view.

expanded	The object is expanded, currently its children are visible.

extSelectable	The object supports extended selection.

focusable	The object can receive focus. Only objects in the active window can receive focus.

focused	The object has keyboard focus.

hasPopup	The object opens a popup.

hotTracked	The object's appearance is sensitive to the mouse cursor position.

invalid	The object is no longer valid (because it has been deleted).

invalidEntry	Input validation current input invalid.

invisible	The object is not visible to the user.

linked	The object is linked to another object, e.g. a hyperlink.

marqueed	The object displays scrolling contents, e.g. a log view.

modal	The object blocks input from other objects.

movable	The object can be moved.

multiLine	The object has multiple lines of text (word wrap), as opposed to a single line.

multiSelectable	The object supports multiple selected items.

offscreen	The object is clipped by the visible area. Objects that are off screen are also invisible.

passwordEdit	The object is a password field, e.g. a line edit for entering a Password.

playsSound	The object produces sound when interacted with.

pressed	The object is pressed.

readOnly	The object can usually be edited, but is explicitly set to read-only.

searchEdit	The object is a line edit that is the input for search queries.

selectable	The object is selectable.

selectableText	The object has text which can be selected. This is different from selectable which refers to the object's children.

selected	The object is selected, this is independent of text selection.

selfVoicing	The object describes itself through speech or sound.

sizeable	The object can be resized, e.g. top-level windows.

summaryElement	The object summarizes the state of the window and should be treated with priority.

supportsAutoCompletion	The object has auto-completion, for example in line edits or combo boxes.

traversed	The object is linked and has been visited.

updatesFrequently	The object changes frequently and needs to be refreshed when accessing it.

disabled	The object is unavailable to the user, e.g. a disabled widget.

Implementations of QAccessibleInterface::state() return a combination of these flags.

Methods

impl State

fn active(&self) -> u64

C++ method: quint64 QAccessible::State::active() const

fn animated(&self) -> u64

C++ method: quint64 QAccessible::State::animated() const

fn busy(&self) -> u64

C++ method: quint64 QAccessible::State::busy() const

fn check_state_mixed(&self) -> u64

C++ method: quint64 QAccessible::State::checkStateMixed() const

fn checkable(&self) -> u64

C++ method: quint64 QAccessible::State::checkable() const

fn checked(&self) -> u64

C++ method: quint64 QAccessible::State::checked() const

fn collapsed(&self) -> u64

C++ method: quint64 QAccessible::State::collapsed() const

fn default_button(&self) -> u64

C++ method: quint64 QAccessible::State::defaultButton() const

fn disabled(&self) -> u64

C++ method: quint64 QAccessible::State::disabled() const

fn editable(&self) -> u64

C++ method: quint64 QAccessible::State::editable() const

fn expandable(&self) -> u64

C++ method: quint64 QAccessible::State::expandable() const

fn expanded(&self) -> u64

C++ method: quint64 QAccessible::State::expanded() const

fn ext_selectable(&self) -> u64

C++ method: quint64 QAccessible::State::extSelectable() const

fn focusable(&self) -> u64

C++ method: quint64 QAccessible::State::focusable() const

fn focused(&self) -> u64

C++ method: quint64 QAccessible::State::focused() const

fn has_popup(&self) -> u64

C++ method: quint64 QAccessible::State::hasPopup() const

fn hot_tracked(&self) -> u64

C++ method: quint64 QAccessible::State::hotTracked() const

fn invalid(&self) -> u64

C++ method: quint64 QAccessible::State::invalid() const

fn invisible(&self) -> u64

C++ method: quint64 QAccessible::State::invisible() const

fn linked(&self) -> u64

C++ method: quint64 QAccessible::State::linked() const

fn marqueed(&self) -> u64

C++ method: quint64 QAccessible::State::marqueed() const

fn modal(&self) -> u64

C++ method: quint64 QAccessible::State::modal() const

fn movable(&self) -> u64

C++ method: quint64 QAccessible::State::movable() const

fn multi_line(&self) -> u64

C++ method: quint64 QAccessible::State::multiLine() const

fn multi_selectable(&self) -> u64

C++ method: quint64 QAccessible::State::multiSelectable() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: StateNewArgs

C++ method: QAccessible::State::State

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::accessible::State
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::accessible::State>
C++ method: [constructor] void QAccessible::State::State()

fn offscreen(&self) -> u64

C++ method: quint64 QAccessible::State::offscreen() const

fn password_edit(&self) -> u64

C++ method: quint64 QAccessible::State::passwordEdit() const

fn pressed(&self) -> u64

C++ method: quint64 QAccessible::State::pressed() const

fn read_only(&self) -> u64

C++ method: quint64 QAccessible::State::readOnly() const

fn search_edit(&self) -> u64

C++ method: quint64 QAccessible::State::searchEdit() const

fn selectable(&self) -> u64

C++ method: quint64 QAccessible::State::selectable() const

fn selectable_text(&self) -> u64

C++ method: quint64 QAccessible::State::selectableText() const

fn selected(&self) -> u64

C++ method: quint64 QAccessible::State::selected() const

fn self_voicing(&self) -> u64

C++ method: quint64 QAccessible::State::selfVoicing() const

fn set_active(&mut self, value: u64)

C++ method: void QAccessible::State::set_active(quint64 value)

fn set_animated(&mut self, value: u64)

C++ method: void QAccessible::State::set_animated(quint64 value)

fn set_busy(&mut self, value: u64)

C++ method: void QAccessible::State::set_busy(quint64 value)

fn set_check_state_mixed(&mut self, value: u64)

C++ method: void QAccessible::State::set_checkStateMixed(quint64 value)

fn set_checkable(&mut self, value: u64)

C++ method: void QAccessible::State::set_checkable(quint64 value)

fn set_checked(&mut self, value: u64)

C++ method: void QAccessible::State::set_checked(quint64 value)

fn set_collapsed(&mut self, value: u64)

C++ method: void QAccessible::State::set_collapsed(quint64 value)

fn set_default_button(&mut self, value: u64)

C++ method: void QAccessible::State::set_defaultButton(quint64 value)

fn set_disabled(&mut self, value: u64)

C++ method: void QAccessible::State::set_disabled(quint64 value)

fn set_editable(&mut self, value: u64)

C++ method: void QAccessible::State::set_editable(quint64 value)

fn set_expandable(&mut self, value: u64)

C++ method: void QAccessible::State::set_expandable(quint64 value)

fn set_expanded(&mut self, value: u64)

C++ method: void QAccessible::State::set_expanded(quint64 value)

fn set_ext_selectable(&mut self, value: u64)

C++ method: void QAccessible::State::set_extSelectable(quint64 value)

fn set_focusable(&mut self, value: u64)

C++ method: void QAccessible::State::set_focusable(quint64 value)

fn set_focused(&mut self, value: u64)

C++ method: void QAccessible::State::set_focused(quint64 value)

fn set_has_popup(&mut self, value: u64)

C++ method: void QAccessible::State::set_hasPopup(quint64 value)

fn set_hot_tracked(&mut self, value: u64)

C++ method: void QAccessible::State::set_hotTracked(quint64 value)

fn set_invalid(&mut self, value: u64)

C++ method: void QAccessible::State::set_invalid(quint64 value)

fn set_invisible(&mut self, value: u64)

C++ method: void QAccessible::State::set_invisible(quint64 value)

fn set_linked(&mut self, value: u64)

C++ method: void QAccessible::State::set_linked(quint64 value)

fn set_marqueed(&mut self, value: u64)

C++ method: void QAccessible::State::set_marqueed(quint64 value)

fn set_modal(&mut self, value: u64)

C++ method: void QAccessible::State::set_modal(quint64 value)

fn set_movable(&mut self, value: u64)

C++ method: void QAccessible::State::set_movable(quint64 value)

fn set_multi_line(&mut self, value: u64)

C++ method: void QAccessible::State::set_multiLine(quint64 value)

fn set_multi_selectable(&mut self, value: u64)

C++ method: void QAccessible::State::set_multiSelectable(quint64 value)

fn set_offscreen(&mut self, value: u64)

C++ method: void QAccessible::State::set_offscreen(quint64 value)

fn set_password_edit(&mut self, value: u64)

C++ method: void QAccessible::State::set_passwordEdit(quint64 value)

fn set_pressed(&mut self, value: u64)

C++ method: void QAccessible::State::set_pressed(quint64 value)

fn set_read_only(&mut self, value: u64)

C++ method: void QAccessible::State::set_readOnly(quint64 value)

fn set_search_edit(&mut self, value: u64)

C++ method: void QAccessible::State::set_searchEdit(quint64 value)

fn set_selectable(&mut self, value: u64)

C++ method: void QAccessible::State::set_selectable(quint64 value)

fn set_selectable_text(&mut self, value: u64)

C++ method: void QAccessible::State::set_selectableText(quint64 value)

fn set_selected(&mut self, value: u64)

C++ method: void QAccessible::State::set_selected(quint64 value)

fn set_self_voicing(&mut self, value: u64)

C++ method: void QAccessible::State::set_selfVoicing(quint64 value)

fn set_sizeable(&mut self, value: u64)

C++ method: void QAccessible::State::set_sizeable(quint64 value)

fn set_supports_auto_completion(&mut self, value: u64)

C++ method: void QAccessible::State::set_supportsAutoCompletion(quint64 value)

fn set_traversed(&mut self, value: u64)

C++ method: void QAccessible::State::set_traversed(quint64 value)

fn sizeable(&self) -> u64

C++ method: quint64 QAccessible::State::sizeable() const

fn supports_auto_completion(&self) -> u64

C++ method: quint64 QAccessible::State::supportsAutoCompletion() const

fn traversed(&self) -> u64

C++ method: quint64 QAccessible::State::traversed() const