Module qt_gui::accessible_action_interface::overloading

Traits

AccessibleActionInterfaceActionNamesArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::action_names method.

AccessibleActionInterfaceKeyBindingsForActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::key_bindings_for_action method.

AccessibleActionInterfaceLocalizedActionDescriptionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::localized_action_description method.

AccessibleActionInterfaceLocalizedActionNameArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::localized_action_name method.

AccessibleActionInterfaceNextPageActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::next_page_action method.

AccessibleActionInterfacePreviousPageActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::previous_page_action method.

AccessibleActionInterfaceScrollDownActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::scroll_down_action method.

AccessibleActionInterfaceScrollLeftActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::scroll_left_action method.

AccessibleActionInterfaceScrollRightActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::scroll_right_action method.

AccessibleActionInterfaceScrollUpActionArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::scroll_up_action method.

AccessibleActionInterfaceTrArgs
This trait represents a set of arguments accepted by AccessibleActionInterface::tr method.

AccessibleActionInterfaceTrUtf8Args
This trait represents a set of arguments accepted by AccessibleActionInterface::tr_utf8 method.