Struct qt_gui::accessible_action_interface::AccessibleActionInterface

pub struct AccessibleActionInterface {
    // some fields omitted
}

C++ type: QAccessibleActionInterface

C++ documentation:

The QAccessibleActionInterface class implements support for invocable actions in the interface.

Accessible objects should implement the action interface if they support user interaction. Usually this interface is implemented by classes that also implement QAccessibleInterface.

The supported actions should use the predefined actions offered in this class unless they do not fit a predefined action. In that case a custom action can be added.

When subclassing QAccessibleActionInterface you need to provide a list of actionNames which is the primary means to discover the available actions. Action names are never localized. In order to present actions to the user there are two functions that need to return localized versions of the name and give a description of the action. For the predefined action names use QAccessibleActionInterface::localizedActionName() and QAccessibleActionInterface::localizedActionDescription() to return their localized counterparts.

In general you should use one of the predefined action names, unless describing an action that does not fit these:

Action name	Description

toggleAction()	toggles the item (checkbox, radio button, switch, ...)

decreaseAction()	decrease the value of the accessible (e.g. spinbox)

increaseAction()	increase the value of the accessible (e.g. spinbox)

pressAction()	press or click or activate the accessible (should correspont to clicking the object with the mouse)

setFocusAction()	set the focus to this accessible

showMenuAction()	show a context menu, corresponds to right-clicks

In order to invoke the action, doAction() is called with an action name.

Most widgets will simply implement pressAction(). This is what happens when the widget is activated by being clicked, space pressed or similar.

IAccessible2 Specification

Methods

impl AccessibleActionInterface

fn action_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceActionNamesArgs<'largs>

C++ method: QAccessibleActionInterface::actionNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn action_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn action_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: pure virtual QStringList QAccessibleActionInterface::actionNames() const

C++ documentation:

Returns the list of actions supported by this accessible object. The actions returned should be in preferred order, i.e. the action that the user most likely wants to trigger should be returned first, while the least likely action should be returned last.

The list does only contain actions that can be invoked. It won't return disabled actions, or actions associated with disabled UI controls.

The list can be empty.

Note that this list is not localized. For a localized representation re-implement localizedActionName() and localizedActionDescription()

See also doAction(), localizedActionName(), and localizedActionDescription().

fn decrease_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::decreaseAction()

C++ documentation:

Returns the name of the decrease default action.

See also actionNames() and localizedActionName().

fn do_action(&mut self, action_name: &String)

C++ method: pure virtual void QAccessibleActionInterface::doAction(const QString& actionName)

C++ documentation:

Invokes the action specified by actionName. Note that actionName is the non-localized name as returned by actionNames() This function is usually implemented by calling the same functions that other user interaction, such as clicking the object, would trigger.

See also actionNames().

fn increase_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::increaseAction()

C++ documentation:

Returns the name of the increase default action.

See also actionNames() and localizedActionName().

fn key_bindings_for_action<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceKeyBindingsForActionArgs<'largs>

C++ method: QAccessibleActionInterface::keyBindingsForAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn key_bindings_for_action(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: pure virtual QStringList QAccessibleActionInterface::keyBindingsForAction(const QString& actionName) const

C++ documentation:

Returns a list of the keyboard shortcuts available for invoking the action named actionName.

This is important to let users learn alternative ways of using the application by emphasizing the keyboard.

See also actionNames().

fn localized_action_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceLocalizedActionDescriptionArgs<'largs>

C++ method: QAccessibleActionInterface::localizedActionDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_description(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleActionInterface::localizedActionDescription(const QString& name) const

C++ documentation:

Returns a localized action description of the action actionName.

When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionName().

fn localized_action_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceLocalizedActionNameArgs<'largs>

C++ method: QAccessibleActionInterface::localizedActionName

This is an overloaded function. Available variants:

Rust arguments:
1) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn localized_action_name(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleActionInterface::localizedActionName(const QString& name) const

C++ documentation:

Returns a localized action name of actionName.

For custom actions this function has to be re-implemented. When using one of the default names, you can call this function in QAccessibleActionInterface to get the localized string.

See also actionNames() and localizedActionDescription().

fn next_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceNextPageActionArgs

C++ method: QAccessibleActionInterface::nextPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn next_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::nextPageAction()

C++ documentation:

Returns the name of the next page default action.

See also actionNames() and localizedActionName().

fn press_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::pressAction()

C++ documentation:

Returns the name of the press default action.

See also actionNames() and localizedActionName().

fn previous_page_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfacePreviousPageActionArgs

C++ method: QAccessibleActionInterface::previousPageAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_page_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn previous_page_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::previousPageAction()

C++ documentation:

Returns the name of the previous page default action.

See also actionNames() and localizedActionName().

fn scroll_down_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceScrollDownActionArgs

C++ method: QAccessibleActionInterface::scrollDownAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_down_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_down_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::scrollDownAction()

C++ documentation:

Returns the name of the scroll down default action.

See also actionNames() and localizedActionName().

fn scroll_left_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceScrollLeftActionArgs

C++ method: QAccessibleActionInterface::scrollLeftAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_left_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_left_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::scrollLeftAction()

C++ documentation:

Returns the name of the scroll left default action.

See also actionNames() and localizedActionName().

fn scroll_right_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceScrollRightActionArgs

C++ method: QAccessibleActionInterface::scrollRightAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_right_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_right_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::scrollRightAction()

C++ documentation:

Returns the name of the scroll right default action.

See also actionNames() and localizedActionName().

fn scroll_up_action<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceScrollUpActionArgs

C++ method: QAccessibleActionInterface::scrollUpAction

This is an overloaded function. Available variants:

Rust arguments:
1) fn scroll_up_action(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn scroll_up_action(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::scrollUpAction()

C++ documentation:

Returns the name of the scroll up default action.

See also actionNames() and localizedActionName().

fn set_focus_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::setFocusAction()

C++ documentation:

Returns the name of the set focus default action.

See also actionNames() and localizedActionName().

fn show_menu_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::showMenuAction()

C++ documentation:

Returns the name of the show menu default action.

See also actionNames() and localizedActionName().

fn toggle_action() -> &'static String

C++ method: static const QString& QAccessibleActionInterface::toggleAction()

C++ documentation:

Returns the name of the toggle default action.

See also actionNames() and localizedActionName().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceTrArgs

C++ method: QAccessibleActionInterface::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: AccessibleActionInterfaceTrUtf8Args

C++ method: QAccessibleActionInterface::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QAccessibleActionInterface::trUtf8(const char* sourceText, const char* disambiguation, int n)