Struct qt_gui::accessible_application::AccessibleApplication

pub struct AccessibleApplication {
    // some fields omitted
}

C++ type: QAccessibleApplication

Methods

impl AccessibleApplication

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleApplication::actionInterface()

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationBackgroundColorArgs<'largs>

C++ method: QAccessibleApplication::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: virtual QColor QAccessibleApplication::backgroundColor() const

C++ documentation:

Returns the accessible's background color if applicable or an invalid QColor.

See also foregroundColor().

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleApplication::child(int index) const

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleApplication::childAt(int x, int y) const

C++ documentation:

Reimplemented from QAccessibleInterface::childAt().

fn child_count(&self) -> c_int

C++ method: virtual int QAccessibleApplication::childCount() const

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleApplication::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleApplication::focusChild() const

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationForegroundColorArgs<'largs>

C++ method: QAccessibleApplication::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: virtual QColor QAccessibleApplication::foregroundColor() const

C++ documentation:

Returns the accessible's foreground color if applicable or an invalid QColor.

See also backgroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleApplication::imageInterface()

fn index_of_child(&self, arg1: *const AccessibleInterface) -> c_int

C++ method: virtual int QAccessibleApplication::indexOfChild(const QAccessibleInterface* arg1) const

fn interface_cast(&mut self, arg1: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleApplication::interface_cast(QAccessible::InterfaceType arg1)

C++ documentation:

Returns a specialized accessibility interface type from the generic QAccessibleInterface.

This function must be reimplemented when providing more information about a widget or object through the specialized interfaces. For example a line edit should implement the QAccessibleTextInterface.

Qt's QLineEdit for example has its accessibility support implemented in QAccessibleLineEdit.

void QAccessibleLineEdit::interface_cast(QAccessible::InterfaceType t)
{
    if (t == QAccessible::TextInterface)
        return static_cast<QAccessibleTextInterface>(this);
    return QAccessibleWidget::interface_cast(t);
}

See also QAccessible::InterfaceType, QAccessibleTextInterface, QAccessibleValueInterface, QAccessibleActionInterface, QAccessibleTableInterface, and QAccessibleTableCellInterface.

fn is_valid(&self) -> bool

C++ method: virtual bool QAccessibleApplication::isValid() const

C++ documentation:

Reimplemented from QAccessibleInterface::isValid().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleApplicationNewArgs

C++ method: QAccessibleApplication::QAccessibleApplication

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::accessible_application::AccessibleApplication
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::accessible_application::AccessibleApplication>
C++ method: [constructor] void QAccessibleApplication::QAccessibleApplication()

fn object(&self) -> *mut Object

C++ method: virtual QObject* QAccessibleApplication::object() const

C++ documentation:

Reimplemented from QAccessibleInterface::object().

fn parent(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleApplication::parent() const

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationRectArgs<'largs>

C++ method: QAccessibleApplication::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QAccessibleApplication::rect() const

C++ documentation:

Reimplemented from QAccessibleInterface::rect().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationRelationsArgs<'largs>

C++ method: QAccessibleApplication::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsStruct)) -> ::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> ::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleApplication::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Returns the meaningful relations to other widgets. Usually this will not return parent/child relations, unless they are handled in a specific way such as in tree views. It will typically return the labelled-by and label relations.

It is possible to filter the relations by using match. It should never return itself.

See also parent() and child().

fn role(&self) -> Role

C++ method: virtual QAccessible::Role QAccessibleApplication::role() const

fn set_text(&mut self, t: Text, text: &String)

C++ method: virtual void QAccessibleApplication::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Reimplemented from QAccessibleInterface::setText().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationStateArgs<'largs>

C++ method: QAccessibleApplication::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> ::accessible::State
C++ method: virtual QAccessible::State QAccessibleApplication::state() const

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleApplication::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleApplication::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleApplicationTextArgs<'largs>

C++ method: QAccessibleApplication::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleApplication::text(QAccessible::Text t) const

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleApplication::textInterface()

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleApplication::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleApplication::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleApplication::window() const