Struct qt_gui::accessible_editable_text_interface::AccessibleEditableTextInterface

pub struct AccessibleEditableTextInterface {
    // some fields omitted
}

C++ type: QAccessibleEditableTextInterface

C++ documentation:

The QAccessibleEditableTextInterface class implements support for objects with editable text.

When implementing this interface you will almost certainly also want to implement QAccessibleTextInterface.

IAccessible2 Specification

Methods

impl AccessibleEditableTextInterface

fn delete_text(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: pure virtual void QAccessibleEditableTextInterface::deleteText(int startOffset, int endOffset)

C++ documentation:

Deletes the text from startOffset to endOffset.

fn insert_text(&mut self, offset: c_int, text: &String)

C++ method: pure virtual void QAccessibleEditableTextInterface::insertText(int offset, const QString& text)

C++ documentation:

Inserts text at position offset.

fn replace_text(&mut self, start_offset: c_int, end_offset: c_int, text: &String)

C++ method: pure virtual void QAccessibleEditableTextInterface::replaceText(int startOffset, int endOffset, const QString& text)

C++ documentation:

Removes the text from startOffset to endOffset and instead inserts text.