Struct qt_gui::accessible_event::AccessibleEvent

pub struct AccessibleEvent {
    // some fields omitted
}

C++ type: QAccessibleEvent

C++ documentation:

The QAccessibleEvent class is the base class for accessibility notifications.

This class is used with QAccessible::updateAccessibility().

The event type is one of the values of QAccessible::Event. There are a number of subclasses that should be used to provide more details about the event.

For example to notify about a focus change when re-implementing QWidget::setFocus, the event could be used as follows:

void MyWidget::setFocus(Qt::FocusReason reason)
{
    // handle custom focus setting...
    QAccessibleEvent event(f, QAccessible::Focus);
    QAccessible::updateAccessibility(&event);
}

To enable in process screen readers, all events must be sent after the change has happened.

Methods

impl AccessibleEvent

fn accessible_interface(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleEvent::accessibleInterface() const

C++ documentation:

Returns the QAccessibleInterface associated with the event. The caller of this function takes ownership of the returned interface.

fn child(&self) -> c_int

C++ method: int QAccessibleEvent::child() const

C++ documentation:

Returns the child index.

See also setChild().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleEventNewArgs

C++ method: QAccessibleEvent::QAccessibleEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::accessible_interface::AccessibleInterface, ::accessible::Event, cpp_utils::AsStruct)) -> ::accessible_event::AccessibleEvent
2) fn new((*mut ::accessible_interface::AccessibleInterface, ::accessible::Event, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_event::AccessibleEvent>
C++ method: [constructor] void QAccessibleEvent::QAccessibleEvent(QAccessibleInterface* iface, QAccessible::Event typ)

Variant 2

Rust arguments:
1) fn new((*mut qt_core::object::Object, ::accessible::Event, cpp_utils::AsStruct)) -> ::accessible_event::AccessibleEvent
2) fn new((*mut qt_core::object::Object, ::accessible::Event, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_event::AccessibleEvent>
C++ method: [constructor] void QAccessibleEvent::QAccessibleEvent(QObject* obj, QAccessible::Event typ)

fn object(&self) -> *mut Object

C++ method: QObject* QAccessibleEvent::object() const

C++ documentation:

Returns the event object.

fn set_child(&mut self, chld: c_int)

C++ method: void QAccessibleEvent::setChild(int chld)

C++ documentation:

Sets the child index to child.

See also child().

fn type_(&self) -> Event

C++ method: QAccessible::Event QAccessibleEvent::type() const

C++ documentation:

Returns the event type.

fn unique_id(&self) -> c_uint

C++ method: unsigned int QAccessibleEvent::uniqueId() const