Struct qt_gui::accessible_interface::AccessibleInterface

pub struct AccessibleInterface {
    // some fields omitted
}

C++ type: QAccessibleInterface

C++ documentation:

The QAccessibleInterface class defines an interface that exposes information about accessible objects.

This class is part of Accessibility for QWidget Applications.

Accessibility tools (also called AT Clients), such as screen readers or braille displays, require high-level information about accessible objects in an application. Accessible objects provide specialized input and output methods, making it possible for users to use accessibility tools with enabled applications (AT Servers).

Every element that the user needs to interact with or react to is an accessible object, and should provide this information. These are mainly visual objects, such as widgets and widget elements, but can also be content, such as sounds.

The AT client uses three basic concepts to acquire information about any accessible object in an application:

Properties The client can read information about accessible objects. In some cases the client can also modify these properties; such as text in a line edit.

Actions The client can invoke actions like pressing a button or .

Relationships and Navigation The client can traverse from one accessible object to another, using the relationships between objects.

The QAccessibleInterface defines the API for these three concepts.

Methods

impl AccessibleInterface

fn action_interface(&mut self) -> *mut AccessibleActionInterface

C++ method: QAccessibleActionInterface* QAccessibleInterface::actionInterface()

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceBackgroundColorArgs<'largs>

C++ method: QAccessibleInterface::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: virtual QColor QAccessibleInterface::backgroundColor() const

C++ documentation:

Returns the accessible's background color if applicable or an invalid QColor.

See also foregroundColor().

fn child(&self, index: c_int) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleInterface::child(int index) const

C++ documentation:

Returns the accessible child with index index. 0-based index. The number of children of an object can be checked with childCount.

Returns 0 when asking for an invalid child (e.g. when the child became invalid in the meantime).

See also childCount() and parent().

fn child_at(&self, x: c_int, y: c_int) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleInterface::childAt(int x, int y) const

C++ documentation:

Returns the QAccessibleInterface of a child that contains the screen coordinates (x, y). If there are no children at this position this function returns 0. The returned accessible must be a child, but not necessarily a direct child.

This function is only relyable for visible objects (invisible object might not be laid out correctly).

All visual objects provide this information.

A default implementation is provided for objects inheriting QAccessibleObject. This will iterate over all children. If the widget manages its children (e.g. a table) it will be more efficient to write a specialized implementation.

See also rect().

fn child_count(&self) -> c_int

C++ method: pure virtual int QAccessibleInterface::childCount() const

C++ documentation:

Returns the number of children that belong to this object. A child can provide accessibility information on its own (e.g. a child widget), or be a sub-element of this accessible object.

All objects provide this information.

See also indexOfChild().

fn editable_text_interface(&mut self) -> *mut AccessibleEditableTextInterface

C++ method: QAccessibleEditableTextInterface* QAccessibleInterface::editableTextInterface()

fn focus_child(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleInterface::focusChild() const

C++ documentation:

Returns the object that has the keyboard focus.

The object returned can be any descendant, including itself.

fn foreground_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceForegroundColorArgs<'largs>

C++ method: QAccessibleInterface::foregroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn foreground_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: virtual QColor QAccessibleInterface::foregroundColor() const

C++ documentation:

Returns the accessible's foreground color if applicable or an invalid QColor.

See also backgroundColor().

fn image_interface(&mut self) -> *mut AccessibleImageInterface

C++ method: QAccessibleImageInterface* QAccessibleInterface::imageInterface()

fn index_of_child(&self, arg1: *const AccessibleInterface) -> c_int

C++ method: pure virtual int QAccessibleInterface::indexOfChild(const QAccessibleInterface* arg1) const

C++ documentation:

Returns the 0-based index of the object child in this object's children list, or -1 if child is not a child of this object.

All objects provide this information about their children.

See also childCount().

fn interface_cast(&mut self, arg1: InterfaceType) -> *mut c_void

C++ method: virtual void* QAccessibleInterface::interface_cast(QAccessible::InterfaceType arg1)

C++ documentation:

Returns a specialized accessibility interface type from the generic QAccessibleInterface.

This function must be reimplemented when providing more information about a widget or object through the specialized interfaces. For example a line edit should implement the QAccessibleTextInterface.

Qt's QLineEdit for example has its accessibility support implemented in QAccessibleLineEdit.

void QAccessibleLineEdit::interface_cast(QAccessible::InterfaceType t)
{
    if (t == QAccessible::TextInterface)
        return static_cast<QAccessibleTextInterface>(this);
    return QAccessibleWidget::interface_cast(t);
}

See also QAccessible::InterfaceType, QAccessibleTextInterface, QAccessibleValueInterface, QAccessibleActionInterface, QAccessibleTableInterface, and QAccessibleTableCellInterface.

fn is_valid(&self) -> bool

C++ method: pure virtual bool QAccessibleInterface::isValid() const

C++ documentation:

Returns true if all the data necessary to use this interface implementation is valid (e.g. all pointers are non-null); otherwise returns false.

See also object().

fn object(&self) -> *mut Object

C++ method: pure virtual QObject* QAccessibleInterface::object() const

C++ documentation:

Returns a pointer to the QObject this interface implementation provides information for.

See also isValid().

fn parent(&self) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleInterface::parent() const

C++ documentation:

Returns the QAccessibleInterface of the parent in the accessible object hierarchy.

Returns 0 if no parent exists (e.g. for the top level application object).

See also child().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceRectArgs<'largs>

C++ method: QAccessibleInterface::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: pure virtual QRect QAccessibleInterface::rect() const

C++ documentation:

Returns the geometry of the object. The geometry is in screen coordinates.

This function is only reliable for visible objects (invisible objects might not be laid out correctly).

All visual objects provide this information.

See also childAt().

fn relations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceRelationsArgs<'largs>

C++ method: QAccessibleInterface::relations

This is an overloaded function. Available variants:

Rust arguments:
1) fn relations(&self, (qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
2) fn relations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef>
3) fn relations(&self, (qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsStruct)) -> ::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
4) fn relations(&self, cpp_utils::AsStruct) -> ::vector::VectorPairAccessibleInterfaceMutPtrFlagsRelationFlagRef
C++ method: virtual QVector<QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>> QAccessibleInterface::relations(QFlags<QAccessible::RelationFlag> match = ?) const

C++ documentation:

Returns the meaningful relations to other widgets. Usually this will not return parent/child relations, unless they are handled in a specific way such as in tree views. It will typically return the labelled-by and label relations.

It is possible to filter the relations by using match. It should never return itself.

See also parent() and child().

fn role(&self) -> Role

C++ method: pure virtual QAccessible::Role QAccessibleInterface::role() const

C++ documentation:

Returns the role of the object. The role of an object is usually static.

All accessible objects have a role.

See also text() and state().

fn set_text(&mut self, t: Text, text: &String)

C++ method: pure virtual void QAccessibleInterface::setText(QAccessible::Text t, const QString& text)

C++ documentation:

Sets the text property t of the object to text.

Note that the text properties of most objects are read-only so calling this function might have no effect.

See also text().

fn state<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceStateArgs<'largs>

C++ method: QAccessibleInterface::state

This is an overloaded function. Available variants:

Rust arguments:
1) fn state(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::accessible::State>
2) fn state(&self, cpp_utils::AsStruct) -> ::accessible::State
C++ method: pure virtual QAccessible::State QAccessibleInterface::state() const

fn table_cell_interface(&mut self) -> *mut AccessibleTableCellInterface

C++ method: QAccessibleTableCellInterface* QAccessibleInterface::tableCellInterface()

fn table_interface(&mut self) -> *mut AccessibleTableInterface

C++ method: QAccessibleTableInterface* QAccessibleInterface::tableInterface()

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleInterfaceTextArgs<'largs>

C++ method: QAccessibleInterface::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (::accessible::Text, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (::accessible::Text, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: pure virtual QString QAccessibleInterface::text(QAccessible::Text t) const

C++ documentation:

Returns the value of the text property t of the object.

The QAccessible::Name is a string used by clients to identify, find, or announce an accessible object for the user. All objects must have a name that is unique within their container. The name can be used differently by clients, so the name should both give a short description of the object and be unique.

An accessible object's QAccessible::Description provides textual information about an object's visual appearance. The description is primarily used to provide greater context for vision-impaired users, but is also used for context searching or other applications. Not all objects have a description. An "OK" button would not need a description, but a tool button that shows a picture of a smiley would.

The QAccessible::Value of an accessible object represents visual information contained by the object, e.g. the text in a line edit. Usually, the value can be modified by the user. Not all objects have a value, e.g. static text labels don't, and some objects have a state that already is the value, e.g. toggle buttons.

The QAccessible::Help text provides information about the function and usage of an accessible object. Not all objects provide this information.

The QAccessible::Accelerator is a keyboard shortcut that activates the object's default action. A keyboard shortcut is the underlined character in the text of a menu, menu item or widget, and is either the character itself, or a combination of this character and a modifier key like Alt, Ctrl or Shift. Command controls like tool buttons also have shortcut keys and usually display them in their tooltip.

All objects provide a string for QAccessible::Name.

See also setText(), role(), and state().

fn text_interface(&mut self) -> *mut AccessibleTextInterface

C++ method: QAccessibleTextInterface* QAccessibleInterface::textInterface()

fn value_interface(&mut self) -> *mut AccessibleValueInterface

C++ method: QAccessibleValueInterface* QAccessibleInterface::valueInterface()

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QAccessibleInterface::virtual_hook(int id, void* data)

fn window(&self) -> *mut Window

C++ method: virtual QWindow* QAccessibleInterface::window() const

C++ documentation:

Returns the window associated with the underlying object. For instance, QAccessibleWidget reimplements this and returns the windowHandle() of the QWidget.

It is used on some platforms to be able to notify the AT client about state changes. The backend will traverse up all ancestors until it finds a window. (This means that at least one interface among the ancestors should return a valid QWindow pointer).

The default implementation returns 0.