Module qt_gui::accessible_object::overloading

Traits

AccessibleObjectBackgroundColorArgs
This trait represents a set of arguments accepted by AccessibleObject::background_color method.

AccessibleObjectForegroundColorArgs
This trait represents a set of arguments accepted by AccessibleObject::foreground_color method.

AccessibleObjectRectArgs
This trait represents a set of arguments accepted by AccessibleObject::rect method.

AccessibleObjectRelationsArgs
This trait represents a set of arguments accepted by AccessibleObject::relations method.

AccessibleObjectStateArgs
This trait represents a set of arguments accepted by AccessibleObject::state method.

AccessibleObjectTextArgs
This trait represents a set of arguments accepted by AccessibleObject::text method.