Struct qt_gui::accessible_state_change_event::AccessibleStateChangeEvent

pub struct AccessibleStateChangeEvent {
    // some fields omitted
}

C++ type: QAccessibleStateChangeEvent

C++ documentation:

The QAccessibleStateChangeEvent notfies the accessibility framework that the state of an object has changed.

This class is used with QAccessible::updateAccessibility().

Methods

impl AccessibleStateChangeEvent

fn accessible_interface(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleStateChangeEvent::accessibleInterface() const

C++ documentation:

Returns the QAccessibleInterface associated with the event. The caller of this function takes ownership of the returned interface.

fn changed_states<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleStateChangeEventChangedStatesArgs<'largs>

C++ method: QAccessibleStateChangeEvent::changedStates

This is an overloaded function. Available variants:

Rust arguments:
1) fn changed_states(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::accessible::State>
2) fn changed_states(&self, cpp_utils::AsStruct) -> ::accessible::State
C++ method: QAccessible::State QAccessibleStateChangeEvent::changedStates() const

C++ documentation:

Returns the states that have been changed.

Be aware that the returned states are the ones that have changed, to find out about the state of an object, use QAccessibleInterface::state().

For example, if an object used to have the focus but loses it, the object's state will have focused set to false. This event on the other hand tells about the change and has focused set to true since the focus state is changed from true to false.

fn child(&self) -> c_int

C++ method: int QAccessibleStateChangeEvent::child() const

C++ documentation:

Returns the child index.

See also setChild().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleStateChangeEventNewArgs

C++ method: QAccessibleStateChangeEvent::QAccessibleStateChangeEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::accessible_interface::AccessibleInterface, &::accessible::State, cpp_utils::AsStruct)) -> ::accessible_state_change_event::AccessibleStateChangeEvent
2) fn new((*mut ::accessible_interface::AccessibleInterface, &::accessible::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_state_change_event::AccessibleStateChangeEvent>
C++ method: [constructor] void QAccessibleStateChangeEvent::QAccessibleStateChangeEvent(QAccessibleInterface* iface, QAccessible::State state)

Variant 2

Rust arguments:
1) fn new((*mut qt_core::object::Object, &::accessible::State, cpp_utils::AsStruct)) -> ::accessible_state_change_event::AccessibleStateChangeEvent
2) fn new((*mut qt_core::object::Object, &::accessible::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_state_change_event::AccessibleStateChangeEvent>
C++ method: [constructor] void QAccessibleStateChangeEvent::QAccessibleStateChangeEvent(QObject* obj, QAccessible::State state)

fn object(&self) -> *mut Object

C++ method: QObject* QAccessibleStateChangeEvent::object() const

C++ documentation:

Returns the event object.

fn set_child(&mut self, chld: c_int)

C++ method: void QAccessibleStateChangeEvent::setChild(int chld)

C++ documentation:

Sets the child index to child.

See also child().

fn type_(&self) -> Event

C++ method: QAccessible::Event QAccessibleStateChangeEvent::type() const

C++ documentation:

Returns the event type.

fn unique_id(&self) -> c_uint

C++ method: unsigned int QAccessibleStateChangeEvent::uniqueId() const