Struct qt_gui::accessible_table_cell_interface::AccessibleTableCellInterface

pub struct AccessibleTableCellInterface {
    // some fields omitted
}

C++ type: QAccessibleTableCellInterface

C++ documentation:

The QAccessibleTableCellInterface class implements support for the IAccessibleTable2 Cell interface.

IAccessible2 Specification

Methods

impl AccessibleTableCellInterface

fn column_extent(&self) -> c_int

C++ method: pure virtual int QAccessibleTableCellInterface::columnExtent() const

C++ documentation:

Returns the number of columns occupied by this cell accessible.

fn column_header_cells<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCellInterfaceColumnHeaderCellsArgs<'largs>

C++ method: QAccessibleTableCellInterface::columnHeaderCells

This is an overloaded function. Available variants:

Rust arguments:
1) fn column_header_cells(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListAccessibleInterfaceMutPtr>
2) fn column_header_cells(&self, cpp_utils::AsStruct) -> ::list::ListAccessibleInterfaceMutPtr
C++ method: pure virtual QList<QAccessibleInterface*> QAccessibleTableCellInterface::columnHeaderCells() const

C++ documentation:

Returns the column headers as an array of cell accessibles.

fn column_index(&self) -> c_int

C++ method: pure virtual int QAccessibleTableCellInterface::columnIndex() const

C++ documentation:

Translates this cell accessible into the corresponding column index.

fn is_selected(&self) -> bool

C++ method: pure virtual bool QAccessibleTableCellInterface::isSelected() const

C++ documentation:

Returns a boolean value indicating whether this cell is selected.

fn row_extent(&self) -> c_int

C++ method: pure virtual int QAccessibleTableCellInterface::rowExtent() const

C++ documentation:

Returns the number of rows occupied by this cell accessible.

fn row_header_cells<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableCellInterfaceRowHeaderCellsArgs<'largs>

C++ method: QAccessibleTableCellInterface::rowHeaderCells

This is an overloaded function. Available variants:

Rust arguments:
1) fn row_header_cells(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListAccessibleInterfaceMutPtr>
2) fn row_header_cells(&self, cpp_utils::AsStruct) -> ::list::ListAccessibleInterfaceMutPtr
C++ method: pure virtual QList<QAccessibleInterface*> QAccessibleTableCellInterface::rowHeaderCells() const

C++ documentation:

Returns the row headers as an array of cell accessibles.

fn row_index(&self) -> c_int

C++ method: pure virtual int QAccessibleTableCellInterface::rowIndex() const

C++ documentation:

Translates this cell accessible into the corresponding row index.

fn table(&self) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleTableCellInterface::table() const

C++ documentation:

Returns the QAccessibleInterface of the table containing this cell.