Struct qt_gui::accessible_table_interface::AccessibleTableInterface

pub struct AccessibleTableInterface {
    // some fields omitted
}

C++ type: QAccessibleTableInterface

C++ documentation:

The QAccessibleTableInterface class implements support for the IAccessibleTable2 interface.

IAccessible2 Specification

Methods

impl AccessibleTableInterface

fn caption(&self) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleTableInterface::caption() const

C++ documentation:

Returns the caption for the table.

fn cell_at(&self, row: c_int, column: c_int) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleTableInterface::cellAt(int row, int column) const

C++ documentation:

Returns the cell at the specified row and column in the table.

fn column_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTableInterface::columnCount() const

C++ documentation:

Returns the total number of columns in table.

fn column_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableInterfaceColumnDescriptionArgs<'largs>

C++ method: QAccessibleTableInterface::columnDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn column_description(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn column_description(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: pure virtual QString QAccessibleTableInterface::columnDescription(int column) const

C++ documentation:

Returns the description text of the specified column in the table.

fn is_column_selected(&self, column: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::isColumnSelected(int column) const

C++ documentation:

Returns a boolean value indicating whether the specified column is completely selected.

fn is_row_selected(&self, row: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::isRowSelected(int row) const

C++ documentation:

Returns a boolean value indicating whether the specified row is completely selected.

fn model_change(&mut self, event: *mut AccessibleTableModelChangeEvent)

C++ method: pure virtual void QAccessibleTableInterface::modelChange(QAccessibleTableModelChangeEvent* event)

C++ documentation:

Informs about a change in the model's layout. The event contains the details.

See also QAccessibleTableModelChangeEvent.

fn row_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTableInterface::rowCount() const

C++ documentation:

Returns the total number of rows in table.

fn row_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableInterfaceRowDescriptionArgs<'largs>

C++ method: QAccessibleTableInterface::rowDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn row_description(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn row_description(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: pure virtual QString QAccessibleTableInterface::rowDescription(int row) const

C++ documentation:

Returns the description text of the specified row in the table.

fn select_column(&mut self, column: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::selectColumn(int column)

C++ documentation:

Selects column. This function might unselect all previously selected columns. Returns true if the selection was successful.

fn select_row(&mut self, row: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::selectRow(int row)

C++ documentation:

Selects row. This function might unselect all previously selected rows. Returns true if the selection was successful.

fn selected_cell_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTableInterface::selectedCellCount() const

C++ documentation:

Returns the total number of selected cells.

fn selected_cells<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableInterfaceSelectedCellsArgs<'largs>

C++ method: QAccessibleTableInterface::selectedCells

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_cells(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListAccessibleInterfaceMutPtr>
2) fn selected_cells(&self, cpp_utils::AsStruct) -> ::list::ListAccessibleInterfaceMutPtr
C++ method: pure virtual QList<QAccessibleInterface*> QAccessibleTableInterface::selectedCells() const

C++ documentation:

Returns the list of selected cell (by their index as QAccessibleInterface::child() accepts).

fn selected_column_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTableInterface::selectedColumnCount() const

C++ documentation:

Returns the total number of selected columns.

fn selected_columns<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableInterfaceSelectedColumnsArgs<'largs>

C++ method: QAccessibleTableInterface::selectedColumns

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_columns(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn selected_columns(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: pure virtual QList<int> QAccessibleTableInterface::selectedColumns() const

C++ documentation:

Returns the list of currently selected columns.

fn selected_row_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTableInterface::selectedRowCount() const

C++ documentation:

Returns the total number of selected rows.

fn selected_rows<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTableInterfaceSelectedRowsArgs<'largs>

C++ method: QAccessibleTableInterface::selectedRows

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_rows(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn selected_rows(&self, cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: pure virtual QList<int> QAccessibleTableInterface::selectedRows() const

C++ documentation:

Returns the list of currently selected columns.

fn summary(&self) -> *mut AccessibleInterface

C++ method: pure virtual QAccessibleInterface* QAccessibleTableInterface::summary() const

C++ documentation:

Returns a QAccessibleInterface that represents a summary of the table. This function may return 0 if no such interface exists.

fn unselect_column(&mut self, column: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::unselectColumn(int column)

C++ documentation:

Unselects column, leaving other selected columns selected (if any). Returns true if the selection was successful.

fn unselect_row(&mut self, row: c_int) -> bool

C++ method: pure virtual bool QAccessibleTableInterface::unselectRow(int row)

C++ documentation:

Unselects row, leaving other selected rows selected (if any). Returns true if the selection was successful.