Struct qt_gui::accessible_table_model_change_event::AccessibleTableModelChangeEvent

pub struct AccessibleTableModelChangeEvent {
    // some fields omitted
}

C++ type: QAccessibleTableModelChangeEvent

C++ documentation:

The QAccessibleTableModelChangeEvent signifies a change in a table, list, or tree where cells are added or removed. If the change affected a number of rows, firstColumn and lastColumn will return -1. Likewise for columns, the row functions may return -1.

This class is used with QAccessible::updateAccessibility().

Methods

impl AccessibleTableModelChangeEvent

fn accessible_interface(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTableModelChangeEvent::accessibleInterface() const

C++ documentation:

Returns the QAccessibleInterface associated with the event. The caller of this function takes ownership of the returned interface.

fn child(&self) -> c_int

C++ method: int QAccessibleTableModelChangeEvent::child() const

C++ documentation:

Returns the child index.

See also setChild().

fn first_column(&self) -> c_int

C++ method: int QAccessibleTableModelChangeEvent::firstColumn() const

C++ documentation:

Returns the first changed column.

See also setFirstColumn().

fn first_row(&self) -> c_int

C++ method: int QAccessibleTableModelChangeEvent::firstRow() const

C++ documentation:

Returns the first changed row.

See also setFirstRow().

fn last_column(&self) -> c_int

C++ method: int QAccessibleTableModelChangeEvent::lastColumn() const

C++ documentation:

Returns the last changed column.

See also setLastColumn().

fn last_row(&self) -> c_int

C++ method: int QAccessibleTableModelChangeEvent::lastRow() const

C++ documentation:

Returns the last changed row.

See also setLastRow().

fn model_change_type(&self) -> ModelChangeType

C++ method: QAccessibleTableModelChangeEvent::ModelChangeType QAccessibleTableModelChangeEvent::modelChangeType() const

C++ documentation:

Returns the type of change.

See also setModelChangeType().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleTableModelChangeEventNewArgs

C++ method: QAccessibleTableModelChangeEvent::QAccessibleTableModelChangeEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::accessible_interface::AccessibleInterface, ::accessible_table_model_change_event::ModelChangeType, cpp_utils::AsStruct)) -> ::accessible_table_model_change_event::AccessibleTableModelChangeEvent
2) fn new((*mut ::accessible_interface::AccessibleInterface, ::accessible_table_model_change_event::ModelChangeType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_table_model_change_event::AccessibleTableModelChangeEvent>
C++ method: [constructor] void QAccessibleTableModelChangeEvent::QAccessibleTableModelChangeEvent(QAccessibleInterface* iface, QAccessibleTableModelChangeEvent::ModelChangeType changeType)

Variant 2

Rust arguments:
1) fn new((*mut qt_core::object::Object, ::accessible_table_model_change_event::ModelChangeType, cpp_utils::AsStruct)) -> ::accessible_table_model_change_event::AccessibleTableModelChangeEvent
2) fn new((*mut qt_core::object::Object, ::accessible_table_model_change_event::ModelChangeType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_table_model_change_event::AccessibleTableModelChangeEvent>
C++ method: [constructor] void QAccessibleTableModelChangeEvent::QAccessibleTableModelChangeEvent(QObject* obj, QAccessibleTableModelChangeEvent::ModelChangeType changeType)

fn object(&self) -> *mut Object

C++ method: QObject* QAccessibleTableModelChangeEvent::object() const

C++ documentation:

Returns the event object.

fn set_child(&mut self, chld: c_int)

C++ method: void QAccessibleTableModelChangeEvent::setChild(int chld)

C++ documentation:

Sets the child index to child.

See also child().

fn set_first_column(&mut self, col: c_int)

C++ method: void QAccessibleTableModelChangeEvent::setFirstColumn(int col)

C++ documentation:

Sets the first changed column.

See also firstColumn().

fn set_first_row(&mut self, row: c_int)

C++ method: void QAccessibleTableModelChangeEvent::setFirstRow(int row)

C++ documentation:

Sets the first changed row.

See also firstRow().

fn set_last_column(&mut self, col: c_int)

C++ method: void QAccessibleTableModelChangeEvent::setLastColumn(int col)

C++ documentation:

Sets the last changed column.

See also lastColumn().

fn set_last_row(&mut self, row: c_int)

C++ method: void QAccessibleTableModelChangeEvent::setLastRow(int row)

C++ documentation:

Sets the last changed row.

See also lastRow().

fn set_model_change_type(&mut self, change_type: ModelChangeType)

C++ method: void QAccessibleTableModelChangeEvent::setModelChangeType(QAccessibleTableModelChangeEvent::ModelChangeType changeType)

C++ documentation:

Sets the type of change to changeType.

See also modelChangeType().

fn type_(&self) -> Event

C++ method: QAccessible::Event QAccessibleTableModelChangeEvent::type() const

C++ documentation:

Returns the event type.

fn unique_id(&self) -> c_uint

C++ method: unsigned int QAccessibleTableModelChangeEvent::uniqueId() const