Module qt_gui::accessible_text_interface::overloading

Traits

AccessibleTextInterfaceAttributesArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::attributes method.

AccessibleTextInterfaceCharacterRectArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::character_rect method.

AccessibleTextInterfaceTextAfterOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::text_after_offset method.

AccessibleTextInterfaceTextArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::text method.

AccessibleTextInterfaceTextAtOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::text_at_offset method.

AccessibleTextInterfaceTextBeforeOffsetArgs
This trait represents a set of arguments accepted by AccessibleTextInterface::text_before_offset method.