Struct qt_gui::accessible_text_interface::AccessibleTextInterface

pub struct AccessibleTextInterface {
    // some fields omitted
}

C++ type: QAccessibleTextInterface

C++ documentation:

The QAccessibleTextInterface class implements support for text handling.

This interface corresponds to the IAccessibleText interface. It should be implemented for widgets that display more text than a plain label. Labels should be represented by only QAccessibleInterface and return their text as name (QAccessibleInterface::text() with QAccessible::Name as type). The QAccessibleTextInterface is typically for text that a screen reader might want to read line by line, and for widgets that support text selection and input. This interface is, for example, implemented for QLineEdit.

IAccessible2 Specification

Methods

impl AccessibleTextInterface

fn add_selection(&mut self, start_offset: c_int, end_offset: c_int)

C++ method: pure virtual void QAccessibleTextInterface::addSelection(int startOffset, int endOffset)

C++ documentation:

Select the text from startOffset to endOffset. The startOffset is the first character that will be selected. The endOffset is the first character that will not be selected.

When the object supports multiple selections (e.g. in a word processor), this adds a new selection, otherwise it replaces the previous selection.

The selection will be endOffset - startOffset characters long.

fn attributes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceAttributesArgs<'largs>

C++ method: QAccessibleTextInterface::attributes

This is an overloaded function. Available variants:

Rust arguments:
1) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn attributes(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: pure virtual QString QAccessibleTextInterface::attributes(int offset, int* startOffset, int* endOffset) const

C++ documentation:

Returns the text attributes at the position offset. In addition the range of the attributes is returned in startOffset and endOffset.

fn character_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTextInterface::characterCount() const

C++ documentation:

Returns the length of the text (total size including spaces).

fn character_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceCharacterRectArgs<'largs>

C++ method: QAccessibleTextInterface::characterRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn character_rect(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn character_rect(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: pure virtual QRect QAccessibleTextInterface::characterRect(int offset) const

C++ documentation:

Returns the position and size of the character at position offset in screen coordinates.

fn cursor_position(&self) -> c_int

C++ method: pure virtual int QAccessibleTextInterface::cursorPosition() const

C++ documentation:

Returns the current cursor position.

See also setCursorPosition().

fn offset_at_point(&self, point: &Point) -> c_int

C++ method: pure virtual int QAccessibleTextInterface::offsetAtPoint(const QPoint& point) const

C++ documentation:

Returns the offset of the character at the point in screen coordinates.

fn remove_selection(&mut self, selection_index: c_int)

C++ method: pure virtual void QAccessibleTextInterface::removeSelection(int selectionIndex)

C++ documentation:

Clears the selection with index selectionIndex.

fn scroll_to_substring(&mut self, start_index: c_int, end_index: c_int)

C++ method: pure virtual void QAccessibleTextInterface::scrollToSubstring(int startIndex, int endIndex)

C++ documentation:

Ensures that the text between startIndex and endIndex is visible.

fn selection(&self, selection_index: c_int, start_offset: *mut c_int, end_offset: *mut c_int)

C++ method: pure virtual void QAccessibleTextInterface::selection(int selectionIndex, int* startOffset, int* endOffset) const

C++ documentation:

Returns a selection. The size of the selection is returned in startOffset and endOffset. If there is no selection both startOffset and endOffset are 0.

The accessibility APIs support multiple selections. For most widgets though, only one selection is supported with selectionIndex equal to 0.

See also setSelection().

fn selection_count(&self) -> c_int

C++ method: pure virtual int QAccessibleTextInterface::selectionCount() const

C++ documentation:

Returns the number of selections in this text.

fn set_cursor_position(&mut self, position: c_int)

C++ method: pure virtual void QAccessibleTextInterface::setCursorPosition(int position)

C++ documentation:

Moves the cursor to position.

See also cursorPosition().

fn set_selection(&mut self, selection_index: c_int, start_offset: c_int, end_offset: c_int)

C++ method: pure virtual void QAccessibleTextInterface::setSelection(int selectionIndex, int startOffset, int endOffset)

C++ documentation:

Set the selection selectionIndex to the range from startOffset to endOffset.

See also selection(), addSelection(), and removeSelection().

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceTextArgs<'largs>

C++ method: QAccessibleTextInterface::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: pure virtual QString QAccessibleTextInterface::text(int startOffset, int endOffset) const

C++ documentation:

Returns the text from startOffset to endOffset. The startOffset is the first character that will be returned. The endOffset is the first character that will not be returned.

fn text_after_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceTextAfterOffsetArgs<'largs>

C++ method: QAccessibleTextInterface::textAfterOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_after_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_after_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTextInterface::textAfterOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

C++ documentation:

Returns the text item of type boundaryType that is right after offset offset and sets startOffset and endOffset values to the start and end positions of that item; returns an empty string if there is no such an item. Sets startOffset and endOffset values to -1 on error.

This default implementation is provided for small text edits. A word processor or text editor should provide their own efficient implementations. This function makes no distinction between paragraphs and lines.

Note: this function can not take the cursor position into account. By convention an offset of -2 means that this function should use the cursor position as offset. Thus an offset of -2 must be converted to the cursor position before calling this function. An offset of -1 is used for the text length and custom implementations of this function have to return the result as if the length was passed in as offset.

fn text_at_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceTextAtOffsetArgs<'largs>

C++ method: QAccessibleTextInterface::textAtOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_at_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_at_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTextInterface::textAtOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

C++ documentation:

Returns the text item of type boundaryType at offset offset and sets startOffset and endOffset values to the start and end positions of that item; returns an empty string if there is no such an item. Sets startOffset and endOffset values to -1 on error.

This default implementation is provided for small text edits. A word processor or text editor should provide their own efficient implementations. This function makes no distinction between paragraphs and lines.

Note: this function can not take the cursor position into account. By convention an offset of -2 means that this function should use the cursor position as offset. Thus an offset of -2 must be converted to the cursor position before calling this function. An offset of -1 is used for the text length and custom implementations of this function have to return the result as if the length was passed in as offset.

fn text_before_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextInterfaceTextBeforeOffsetArgs<'largs>

C++ method: QAccessibleTextInterface::textBeforeOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_before_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_before_offset(&self, (libc::c_int, ::accessible::TextBoundaryType, *mut libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: virtual QString QAccessibleTextInterface::textBeforeOffset(int offset, QAccessible::TextBoundaryType boundaryType, int* startOffset, int* endOffset) const

C++ documentation:

Returns the text item of type boundaryType that is close to offset offset and sets startOffset and endOffset values to the start and end positions of that item; returns an empty string if there is no such an item. Sets startOffset and endOffset values to -1 on error.

This default implementation is provided for small text edits. A word processor or text editor should provide their own efficient implementations. This function makes no distinction between paragraphs and lines.

Note: this function can not take the cursor position into account. By convention an offset of -2 means that this function should use the cursor position as offset. Thus an offset of -2 must be converted to the cursor position before calling this function. An offset of -1 is used for the text length and custom implementations of this function have to return the result as if the length was passed in as offset.