Struct qt_gui::accessible_text_update_event::AccessibleTextUpdateEvent

pub struct AccessibleTextUpdateEvent {
    // some fields omitted
}

C++ type: QAccessibleTextUpdateEvent

C++ documentation:

The QAccessibleTextUpdateEvent class notifies about text changes. This is for accessibles that support editable text such as line edits. This event occurs for example when a portion of selected text gets replaced by pasting a new text or in override mode of editors.

This class is used with QAccessible::updateAccessibility().

Methods

impl AccessibleTextUpdateEvent

fn accessible_interface(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleTextUpdateEvent::accessibleInterface() const

C++ documentation:

Returns the QAccessibleInterface associated with the event. The caller of this function takes ownership of the returned interface.

fn change_position(&self) -> c_int

C++ method: int QAccessibleTextUpdateEvent::changePosition() const

C++ documentation:

Returns where the change took place.

fn child(&self) -> c_int

C++ method: int QAccessibleTextUpdateEvent::child() const

C++ documentation:

Returns the child index.

See also setChild().

fn cursor_position(&self) -> c_int

C++ method: int QAccessibleTextUpdateEvent::cursorPosition() const

C++ documentation:

Returns the cursor position.

See also setCursorPosition().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleTextUpdateEventNewArgs

C++ method: QAccessibleTextUpdateEvent::QAccessibleTextUpdateEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::accessible_interface::AccessibleInterface, libc::c_int, &qt_core::string::String, &qt_core::string::String, cpp_utils::AsStruct)) -> ::accessible_text_update_event::AccessibleTextUpdateEvent
2) fn new((*mut ::accessible_interface::AccessibleInterface, libc::c_int, &qt_core::string::String, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_text_update_event::AccessibleTextUpdateEvent>
C++ method: [constructor] void QAccessibleTextUpdateEvent::QAccessibleTextUpdateEvent(QAccessibleInterface* iface, int position, const QString& oldText, const QString& text)

Variant 2

Rust arguments:
1) fn new((*mut qt_core::object::Object, libc::c_int, &qt_core::string::String, &qt_core::string::String, cpp_utils::AsStruct)) -> ::accessible_text_update_event::AccessibleTextUpdateEvent
2) fn new((*mut qt_core::object::Object, libc::c_int, &qt_core::string::String, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_text_update_event::AccessibleTextUpdateEvent>
C++ method: [constructor] void QAccessibleTextUpdateEvent::QAccessibleTextUpdateEvent(QObject* obj, int position, const QString& oldText, const QString& text)

fn object(&self) -> *mut Object

C++ method: QObject* QAccessibleTextUpdateEvent::object() const

C++ documentation:

Returns the event object.

fn set_child(&mut self, chld: c_int)

C++ method: void QAccessibleTextUpdateEvent::setChild(int chld)

C++ documentation:

Sets the child index to child.

See also child().

fn set_cursor_position(&mut self, position: c_int)

C++ method: void QAccessibleTextUpdateEvent::setCursorPosition(int position)

C++ documentation:

Sets the cursor position for this event.

See also cursorPosition().

fn text_inserted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextUpdateEventTextInsertedArgs<'largs>

C++ method: QAccessibleTextUpdateEvent::textInserted

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_inserted(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_inserted(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QAccessibleTextUpdateEvent::textInserted() const

C++ documentation:

Returns the inserted text.

fn text_removed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleTextUpdateEventTextRemovedArgs<'largs>

C++ method: QAccessibleTextUpdateEvent::textRemoved

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_removed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text_removed(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QAccessibleTextUpdateEvent::textRemoved() const

C++ documentation:

Returns the removed text.

fn type_(&self) -> Event

C++ method: QAccessible::Event QAccessibleTextUpdateEvent::type() const

C++ documentation:

Returns the event type.

fn unique_id(&self) -> c_uint

C++ method: unsigned int QAccessibleTextUpdateEvent::uniqueId() const