Struct qt_gui::accessible_value_change_event::AccessibleValueChangeEvent

pub struct AccessibleValueChangeEvent {
    // some fields omitted
}

C++ type: QAccessibleValueChangeEvent

C++ documentation:

The QAccessibleValueChangeEvent describes a change in value for an accessible object.

It contains the new value.

This class is used with QAccessible::updateAccessibility().

Methods

impl AccessibleValueChangeEvent

fn accessible_interface(&self) -> *mut AccessibleInterface

C++ method: virtual QAccessibleInterface* QAccessibleValueChangeEvent::accessibleInterface() const

C++ documentation:

Returns the QAccessibleInterface associated with the event. The caller of this function takes ownership of the returned interface.

fn child(&self) -> c_int

C++ method: int QAccessibleValueChangeEvent::child() const

C++ documentation:

Returns the child index.

See also setChild().

fn new<Args>(args: Args) -> Args::ReturnType where Args: AccessibleValueChangeEventNewArgs

C++ method: QAccessibleValueChangeEvent::QAccessibleValueChangeEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::accessible_interface::AccessibleInterface, &qt_core::variant::Variant, cpp_utils::AsStruct)) -> ::accessible_value_change_event::AccessibleValueChangeEvent
2) fn new((*mut ::accessible_interface::AccessibleInterface, &qt_core::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_value_change_event::AccessibleValueChangeEvent>
C++ method: [constructor] void QAccessibleValueChangeEvent::QAccessibleValueChangeEvent(QAccessibleInterface* iface, const QVariant& val)

Variant 2

Rust arguments:
1) fn new((*mut qt_core::object::Object, &qt_core::variant::Variant, cpp_utils::AsStruct)) -> ::accessible_value_change_event::AccessibleValueChangeEvent
2) fn new((*mut qt_core::object::Object, &qt_core::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::accessible_value_change_event::AccessibleValueChangeEvent>
C++ method: [constructor] void QAccessibleValueChangeEvent::QAccessibleValueChangeEvent(QObject* obj, const QVariant& val)

fn object(&self) -> *mut Object

C++ method: QObject* QAccessibleValueChangeEvent::object() const

C++ documentation:

Returns the event object.

fn set_child(&mut self, chld: c_int)

C++ method: void QAccessibleValueChangeEvent::setChild(int chld)

C++ documentation:

Sets the child index to child.

See also child().

fn set_value(&mut self, val: &Variant)

C++ method: void QAccessibleValueChangeEvent::setValue(const QVariant& val)

C++ documentation:

Sets the new value for this event.

See also value().

fn type_(&self) -> Event

C++ method: QAccessible::Event QAccessibleValueChangeEvent::type() const

C++ documentation:

Returns the event type.

fn unique_id(&self) -> c_uint

C++ method: unsigned int QAccessibleValueChangeEvent::uniqueId() const

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleValueChangeEventValueArgs<'largs>

C++ method: QAccessibleValueChangeEvent::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QAccessibleValueChangeEvent::value() const

C++ documentation:

Returns the new value of the accessible object of this event.

See also setValue().