Struct qt_gui::accessible_value_interface::AccessibleValueInterface

pub struct AccessibleValueInterface {
    // some fields omitted
}

C++ type: QAccessibleValueInterface

C++ documentation:

The QAccessibleValueInterface class implements support for objects that manipulate a value.

This interface should be implemented by accessible objects that represent a value. Examples are spinner, slider, dial and scroll bar.

Instead of forcing the user to deal with the individual parts of the widgets, this interface gives an easier approach to the kind of widget it represents.

Usually this interface is implemented by classes that also implement QAccessibleInterface.

IAccessible2 Specification

Methods

impl AccessibleValueInterface

fn current_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleValueInterfaceCurrentValueArgs<'largs>

C++ method: QAccessibleValueInterface::currentValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn current_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: pure virtual QVariant QAccessibleValueInterface::currentValue() const

C++ documentation:

Returns the current value of the widget. This is usually a double or int.

See also setCurrentValue().

fn maximum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleValueInterfaceMaximumValueArgs<'largs>

C++ method: QAccessibleValueInterface::maximumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn maximum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: pure virtual QVariant QAccessibleValueInterface::maximumValue() const

C++ documentation:

Returns the maximum value this object accepts.

See also minimumValue() and currentValue().

fn minimum_step_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleValueInterfaceMinimumStepSizeArgs<'largs>

C++ method: QAccessibleValueInterface::minimumStepSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_step_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_step_size(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: pure virtual QVariant QAccessibleValueInterface::minimumStepSize() const

C++ documentation:

Returns the minimum step size for the accessible. This is the smalles increment that makes sense when changing the value. When programatically changing the value it should always be a multiple of the minimum step size.

Some tools use this value even when the setCurrentValue does not perform any action. Progress bars for example are read-only but should return their range divided by 100.

fn minimum_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AccessibleValueInterfaceMinimumValueArgs<'largs>

C++ method: QAccessibleValueInterface::minimumValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn minimum_value(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: pure virtual QVariant QAccessibleValueInterface::minimumValue() const

C++ documentation:

Returns the minimum value this object accepts.

See also maximumValue() and currentValue().

fn set_current_value(&mut self, value: &Variant)

C++ method: pure virtual void QAccessibleValueInterface::setCurrentValue(const QVariant& value)

C++ documentation:

Sets the value. If the desired value is out of the range of permissible values, this call will be ignored.

See also currentValue(), minimumValue(), and maximumValue().