Struct qt_gui::backing_store::BackingStore

pub struct BackingStore {
    // some fields omitted
}

C++ type: QBackingStore

C++ documentation:

The QBackingStore class provides a drawing area for QWindow.

QBackingStore enables the use of QPainter to paint on a QWindow with type RasterSurface. The other way of rendering to a QWindow is through the use of OpenGL with QOpenGLContext.

A QBackingStore contains a buffered representation of the window contents, and thus supports partial updates by using QPainter to only update a sub region of the window contents.

QBackingStore might be used by an application that wants to use QPainter without OpenGL acceleration and without the extra overhead of using the QWidget or QGraphicsView UI stacks. For an example of how to use QBackingStore see the Raster Window Example.

Methods

impl BackingStore

fn begin_paint(&mut self, arg1: &Region)

C++ method: void QBackingStore::beginPaint(const QRegion& arg1)

C++ documentation:

This function is called before painting onto the surface begins, with the region in which the painting will occur.

See also endPaint() and paintDevice().

fn end_paint(&mut self)

C++ method: void QBackingStore::endPaint()

C++ documentation:

This function is called after painting onto the surface has ended.

See also beginPaint() and paintDevice().

fn flush<'largs, Args>(&'largs mut self, args: Args) where Args: BackingStoreFlushArgs<'largs>

C++ method: QBackingStore::flush

This is an overloaded function. Available variants:

Rust arguments:
1) fn flush(&mut self, &::region::Region) -> ()
2) fn flush(&mut self, (&::region::Region, *mut ::window::Window)) -> ()
3) fn flush(&mut self, (&::region::Region, *mut ::window::Window, &qt_core::point::Point)) -> ()
C++ method: void QBackingStore::flush(const QRegion& region, QWindow* window = ?, const QPoint& offset = ?)

C++ documentation:

Flushes the given region from the specified window win onto the screen.

Note that the offset parameter is currently unused.

fn has_static_contents(&self) -> bool

C++ method: bool QBackingStore::hasStaticContents() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: BackingStoreNewArgs

C++ method: QBackingStore::QBackingStore

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::window::Window, cpp_utils::AsStruct)) -> ::backing_store::BackingStore
2) fn new((*mut ::window::Window, cpp_utils::AsBox)) -> cpp_utils::CppBox<::backing_store::BackingStore>
C++ method: [constructor] void QBackingStore::QBackingStore(QWindow* window)

fn paint_device(&mut self) -> *mut PaintDevice

C++ method: QPaintDevice* QBackingStore::paintDevice()

C++ documentation:

Implement this function to return the appropriate paint device.

fn resize(&mut self, size: &Size)

C++ method: void QBackingStore::resize(const QSize& size)

C++ documentation:

Sets the size of the windowsurface to be size.

See also size().

fn scroll(&mut self, area: &Region, dx: c_int, dy: c_int) -> bool

C++ method: bool QBackingStore::scroll(const QRegion& area, int dx, int dy)

C++ documentation:

Scrolls the given area dx pixels to the right and dy downward; both dx and dy may be negative.

Returns true if the area was scrolled successfully; false otherwise.

fn set_static_contents(&mut self, region: &Region)

C++ method: void QBackingStore::setStaticContents(const QRegion& region)

C++ documentation:

See also staticContents().

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: BackingStoreSizeArgs<'largs>

C++ method: QBackingStore::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QBackingStore::size() const

C++ documentation:

Returns the current size of the windowsurface.

fn static_contents<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: BackingStoreStaticContentsArgs<'largs>

C++ method: QBackingStore::staticContents

This is an overloaded function. Available variants:

Rust arguments:
1) fn static_contents(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::region::Region>
2) fn static_contents(&self, cpp_utils::AsStruct) -> ::region::Region
C++ method: QRegion QBackingStore::staticContents() const

C++ documentation:

See also setStaticContents().

fn window(&self) -> *mut Window

C++ method: QWindow* QBackingStore::window() const

C++ documentation:

Returns a pointer to the top-level window associated with this surface.