Module qt_gui::bitmap::overloading

Traits

BitmapAsQVariantArgs
This trait represents a set of arguments accepted by Bitmap::as_q_variant method.

BitmapConvertFromImageArgs
This trait represents a set of arguments accepted by Bitmap::convert_from_image method.

BitmapCopyArgs
This trait represents a set of arguments accepted by Bitmap::copy method.

BitmapCreateHeuristicMaskArgs
This trait represents a set of arguments accepted by Bitmap::create_heuristic_mask method.

BitmapCreateMaskFromColorArgs
This trait represents a set of arguments accepted by Bitmap::create_mask_from_color method.

BitmapFillArgs
This trait represents a set of arguments accepted by Bitmap::fill method.

BitmapFromDataArgs
This trait represents a set of arguments accepted by Bitmap::from_data method.

BitmapFromImageArgs
This trait represents a set of arguments accepted by Bitmap::from_image method.

BitmapFromImageReaderArgs
This trait represents a set of arguments accepted by Bitmap::from_image_reader method.

BitmapGrabWidgetArgs
This trait represents a set of arguments accepted by Bitmap::grab_widget method.

BitmapGrabWindowArgs
This trait represents a set of arguments accepted by Bitmap::grab_window method.

BitmapLoadArgs
This trait represents a set of arguments accepted by Bitmap::load method.

BitmapLoadFromDataArgs
This trait represents a set of arguments accepted by Bitmap::load_from_data method.

BitmapMaskArgs
This trait represents a set of arguments accepted by Bitmap::mask method.

BitmapNewArgs
This trait represents a set of arguments accepted by Bitmap::new method.

BitmapRectArgs
This trait represents a set of arguments accepted by Bitmap::rect method.

BitmapSaveArgs
This trait represents a set of arguments accepted by Bitmap::save method.

BitmapScaledArgs
This trait represents a set of arguments accepted by Bitmap::scaled method.

BitmapScaledToHeightArgs
This trait represents a set of arguments accepted by Bitmap::scaled_to_height method.

BitmapScaledToWidthArgs
This trait represents a set of arguments accepted by Bitmap::scaled_to_width method.

BitmapScrollArgs
This trait represents a set of arguments accepted by Bitmap::scroll method.

BitmapSizeArgs
This trait represents a set of arguments accepted by Bitmap::size method.

BitmapToImageArgs
This trait represents a set of arguments accepted by Bitmap::to_image method.

BitmapTransformedArgs
This trait represents a set of arguments accepted by Bitmap::transformed method.

BitmapTrueMatrixArgs
This trait represents a set of arguments accepted by Bitmap::true_matrix method.