Module qt_gui::brush::overloading

Traits

BrushAsQVariantArgs
This trait represents a set of arguments accepted by Brush::as_q_variant method.

BrushNewArgs
This trait represents a set of arguments accepted by Brush::new method.

BrushSetColorArgs
This trait represents a set of arguments accepted by Brush::set_color method.

BrushTextureArgs
This trait represents a set of arguments accepted by Brush::texture method.

BrushTextureImageArgs
This trait represents a set of arguments accepted by Brush::texture_image method.

BrushTransformArgs
This trait represents a set of arguments accepted by Brush::transform method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.