Module qt_gui::clipboard::overloading

Traits

ClipboardClearArgs
This trait represents a set of arguments accepted by Clipboard::clear method.

ClipboardConnectArgs
This trait represents a set of arguments accepted by Clipboard::connect method.

ClipboardConnectStaticArgs
This trait represents a set of arguments accepted by Clipboard::connect_static method.

ClipboardDestroyedArgs
This trait represents a set of arguments accepted by Clipboard::destroyed method.

ClipboardDisconnectArgs
This trait represents a set of arguments accepted by Clipboard::disconnect method.

ClipboardDisconnectStaticArgs
This trait represents a set of arguments accepted by Clipboard::disconnect_static method.

ClipboardDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Clipboard::dynamic_property_names method.

ClipboardImageArgs
This trait represents a set of arguments accepted by Clipboard::image method.

ClipboardMimeDataArgs
This trait represents a set of arguments accepted by Clipboard::mime_data method.

ClipboardObjectNameArgs
This trait represents a set of arguments accepted by Clipboard::object_name method.

ClipboardPixmapArgs
This trait represents a set of arguments accepted by Clipboard::pixmap method.

ClipboardPropertyArgs
This trait represents a set of arguments accepted by Clipboard::property method.

ClipboardSetImageArgs
This trait represents a set of arguments accepted by Clipboard::set_image method.

ClipboardSetMimeDataArgs
This trait represents a set of arguments accepted by Clipboard::set_mime_data method.

ClipboardSetPixmapArgs
This trait represents a set of arguments accepted by Clipboard::set_pixmap method.

ClipboardSetTextArgs
This trait represents a set of arguments accepted by Clipboard::set_text method.

ClipboardStartTimerArgs
This trait represents a set of arguments accepted by Clipboard::start_timer method.

ClipboardTextArgs
This trait represents a set of arguments accepted by Clipboard::text method.

ClipboardTrArgs
This trait represents a set of arguments accepted by Clipboard::tr method.

ClipboardTrUtf8Args
This trait represents a set of arguments accepted by Clipboard::tr_utf8 method.