Module qt_gui::color::overloading

Traits

ColorAsQVariantArgs
This trait represents a set of arguments accepted by Color::as_q_variant method.

ColorColorNamesArgs
This trait represents a set of arguments accepted by Color::color_names method.

ColorConvertToArgs
This trait represents a set of arguments accepted by Color::convert_to method.

ColorDarkArgs
This trait represents a set of arguments accepted by Color::dark method.

ColorDarkerArgs
This trait represents a set of arguments accepted by Color::darker method.

ColorFromCmykArgs
This trait represents a set of arguments accepted by Color::from_cmyk method.

ColorFromCmykFArgs
This trait represents a set of arguments accepted by Color::from_cmyk_f method.

ColorFromHslArgs
This trait represents a set of arguments accepted by Color::from_hsl method.

ColorFromHslFArgs
This trait represents a set of arguments accepted by Color::from_hsl_f method.

ColorFromHsvArgs
This trait represents a set of arguments accepted by Color::from_hsv method.

ColorFromHsvFArgs
This trait represents a set of arguments accepted by Color::from_hsv_f method.

ColorFromRgbArgs
This trait represents a set of arguments accepted by Color::from_rgb method.

ColorFromRgbFArgs
This trait represents a set of arguments accepted by Color::from_rgb_f method.

ColorFromRgbaArgs
This trait represents a set of arguments accepted by Color::from_rgba method.

ColorGetCmykArgs
This trait represents a set of arguments accepted by Color::get_cmyk method.

ColorGetCmykFArgs
This trait represents a set of arguments accepted by Color::get_cmyk_f method.

ColorGetHslArgs
This trait represents a set of arguments accepted by Color::get_hsl method.

ColorGetHslFArgs
This trait represents a set of arguments accepted by Color::get_hsl_f method.

ColorGetHsvArgs
This trait represents a set of arguments accepted by Color::get_hsv method.

ColorGetHsvFArgs
This trait represents a set of arguments accepted by Color::get_hsv_f method.

ColorGetRgbArgs
This trait represents a set of arguments accepted by Color::get_rgb method.

ColorGetRgbFArgs
This trait represents a set of arguments accepted by Color::get_rgb_f method.

ColorLightArgs
This trait represents a set of arguments accepted by Color::light method.

ColorLighterArgs
This trait represents a set of arguments accepted by Color::lighter method.

ColorNameArgs
This trait represents a set of arguments accepted by Color::name method.

ColorNewArgs
This trait represents a set of arguments accepted by Color::new method.

ColorOpAssignArgs
This trait represents a set of arguments accepted by Color::op_assign method.

ColorSetCmykArgs
This trait represents a set of arguments accepted by Color::set_cmyk method.

ColorSetCmykFArgs
This trait represents a set of arguments accepted by Color::set_cmyk_f method.

ColorSetHslArgs
This trait represents a set of arguments accepted by Color::set_hsl method.

ColorSetHslFArgs
This trait represents a set of arguments accepted by Color::set_hsl_f method.

ColorSetHsvArgs
This trait represents a set of arguments accepted by Color::set_hsv method.

ColorSetHsvFArgs
This trait represents a set of arguments accepted by Color::set_hsv_f method.

ColorSetRgbArgs
This trait represents a set of arguments accepted by Color::set_rgb method.

ColorSetRgbFArgs
This trait represents a set of arguments accepted by Color::set_rgb_f method.

ColorToCmykArgs
This trait represents a set of arguments accepted by Color::to_cmyk method.

ColorToHslArgs
This trait represents a set of arguments accepted by Color::to_hsl method.

ColorToHsvArgs
This trait represents a set of arguments accepted by Color::to_hsv method.

ColorToRgbArgs
This trait represents a set of arguments accepted by Color::to_rgb method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.