Struct qt_gui::cursor::Cursor

pub struct Cursor {
    // some fields omitted
}

C++ type: QCursor

C++ documentation:

The QCursor class provides a mouse cursor with an arbitrary shape.

This class is mainly used to create mouse cursors that are associated with particular widgets and to get and set the position of the mouse cursor.

Qt has a number of standard cursor shapes, but you can also make custom cursor shapes based on a QBitmap, a mask and a hotspot.

To associate a cursor with a widget, use QWidget::setCursor(). To associate a cursor with all widgets (normally for a short period of time), use QGuiApplication::setOverrideCursor().

To set a cursor shape use QCursor::setShape() or use the QCursor constructor which takes the shape as argument, or you can use one of the predefined cursors defined in the Qt::CursorShape enum.

If you want to create a cursor with your own bitmap, either use the QCursor constructor which takes a bitmap and a mask or the constructor which takes a pixmap as arguments.

To set or get the position of the mouse cursor use the static methods QCursor::pos() and QCursor::setPos().

Note: It is possible to create a QCursor before QGuiApplication, but it is not useful except as a place-holder for a real QCursor created after QGuiApplication. Attempting to use a QCursor that was created before QGuiApplication will result in a crash.

Methods

impl Cursor

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CursorAsQVariantArgs<'largs>

C++ method: QCursor::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QCursor::operator QVariant() const

C++ documentation:

Returns the cursor as a QVariant.

fn bitmap(&self) -> *const Bitmap

C++ method: const QBitmap* QCursor::bitmap() const

C++ documentation:

Returns the cursor bitmap, or 0 if it is one of the standard cursors.

fn hot_spot<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CursorHotSpotArgs<'largs>

C++ method: QCursor::hotSpot

This is an overloaded function. Available variants:

Rust arguments:
1) fn hot_spot(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn hot_spot(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QCursor::hotSpot() const

C++ documentation:

Returns the cursor hot spot, or (0, 0) if it is one of the standard cursors.

fn mask(&self) -> *const Bitmap

C++ method: const QBitmap* QCursor::mask() const

C++ documentation:

Returns the cursor bitmap mask, or 0 if it is one of the standard cursors.

fn new<Args>(args: Args) -> Args::ReturnType where Args: CursorNewArgs

C++ method: QCursor::QCursor

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::cursor::Cursor
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::cursor::Cursor>
C++ method: [constructor] void QCursor::QCursor()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::CursorShape, cpp_utils::AsStruct)) -> ::cursor::Cursor
2) fn new((qt_core::qt::CursorShape, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
C++ method: [constructor] void QCursor::QCursor(Qt::CursorShape shape)

Variant 3

Rust arguments:
1) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, cpp_utils::AsStruct)) -> ::cursor::Cursor
2) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, libc::c_int, cpp_utils::AsStruct)) -> ::cursor::Cursor
3) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::cursor::Cursor
4) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
5) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
6) fn new((&::bitmap::Bitmap, &::bitmap::Bitmap, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
C++ method: [constructor] void QCursor::QCursor(const QBitmap& bitmap, const QBitmap& mask, int hotX = ?, int hotY = ?)

Variant 4

Rust arguments:
1) fn new((&::cursor::Cursor, cpp_utils::AsStruct)) -> ::cursor::Cursor
2) fn new((&::cursor::Cursor, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
C++ method: [constructor] void QCursor::QCursor(const QCursor& cursor)

Variant 5

Rust arguments:
1) fn new((&::pixmap::Pixmap, cpp_utils::AsStruct)) -> ::cursor::Cursor
2) fn new((&::pixmap::Pixmap, libc::c_int, cpp_utils::AsStruct)) -> ::cursor::Cursor
3) fn new((&::pixmap::Pixmap, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::cursor::Cursor
4) fn new((&::pixmap::Pixmap, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
5) fn new((&::pixmap::Pixmap, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
6) fn new((&::pixmap::Pixmap, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cursor::Cursor>
C++ method: [constructor] void QCursor::QCursor(const QPixmap& pixmap, int hotX = ?, int hotY = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, cursor: &'l1 Cursor) -> &'l0 mut Cursor

C++ method: QCursor& QCursor::operator=(const QCursor& cursor)

C++ documentation:

Assigns c to this cursor and returns a reference to this cursor.

fn pixmap<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CursorPixmapArgs<'largs>

C++ method: QCursor::pixmap

This is an overloaded function. Available variants:

Rust arguments:
1) fn pixmap(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&self, cpp_utils::AsStruct) -> ::pixmap::Pixmap
C++ method: QPixmap QCursor::pixmap() const

C++ documentation:

Returns the cursor pixmap. This is only valid if the cursor is a pixmap cursor.

fn pos<Args>(args: Args) -> Args::ReturnType where Args: CursorPosArgs

C++ method: QCursor::pos

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn pos(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: static QPoint QCursor::pos()

C++ documentation:

Returns the position of the cursor (hot spot) of the primary screen in global screen coordinates.

You can call QWidget::mapFromGlobal() to translate it to widget coordinates.

Note: The position is queried from the windowing system. If mouse events are generated via other means (e.g., via QWindowSystemInterface in a unit test), those fake mouse moves will not be reflected in the returned value.

Note: On platforms where there is no windowing system or cursors are not available, the returned position is based on the mouse move events generated via QWindowSystemInterface.

See also setPos(), QWidget::mapFromGlobal(), QWidget::mapToGlobal(), and QGuiApplication::primaryScreen().

Variant 2

Rust arguments:
1) fn pos((*const ::screen::Screen, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos((*const ::screen::Screen, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: static QPoint QCursor::pos(const QScreen* screen)

C++ documentation:

Returns the position of the cursor (hot spot) of the screen in global screen coordinates.

You can call QWidget::mapFromGlobal() to translate it to widget coordinates.

See also setPos(), QWidget::mapFromGlobal(), and QWidget::mapToGlobal().

fn set_pos<Args>(args: Args) where Args: CursorSetPosArgs

C++ method: QCursor::setPos

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_pos((*mut ::screen::Screen, &qt_core::point::Point)) -> ()
C++ method: static void QCursor::setPos(QScreen* screen, const QPoint& p)

C++ documentation:

This is an overloaded function.

Moves the cursor (hot spot) to the global screen position of the screen at point p.

Variant 2

Rust arguments: fn set_pos((*mut ::screen::Screen, libc::c_int, libc::c_int)) -> ()
C++ method: static void QCursor::setPos(QScreen* screen, int x, int y)

C++ documentation:

Moves the cursor (hot spot) of the screen to the global screen position (x, y).

You can call QWidget::mapToGlobal() to translate widget coordinates to global screen coordinates.

Note: Calling this function results in changing the cursor position through the windowing system. The windowing system will typically respond by sending mouse events to the application's window. This means that the usage of this function should be avoided in unit tests and everywhere where fake mouse events are being injected via QWindowSystemInterface because the windowing system's mouse state (with regards to buttons for example) may not match the state in the application-generated events.

Note: On platforms where there is no windowing system or cursors are not available, this function may do nothing.

See also pos(), QWidget::mapFromGlobal(), and QWidget::mapToGlobal().

Variant 3

Rust arguments: fn set_pos(&qt_core::point::Point) -> ()
C++ method: static void QCursor::setPos(const QPoint& p)

C++ documentation:

This is an overloaded function.

Moves the cursor (hot spot) to the global screen position at point p.

Variant 4

Rust arguments: fn set_pos((libc::c_int, libc::c_int)) -> ()
C++ method: static void QCursor::setPos(int x, int y)

C++ documentation:

Moves the cursor (hot spot) of the primary screen to the global screen position (x, y).

You can call QWidget::mapToGlobal() to translate widget coordinates to global screen coordinates.

See also pos(), QWidget::mapFromGlobal(), QWidget::mapToGlobal(), and QGuiApplication::primaryScreen().

fn set_shape(&mut self, new_shape: CursorShape)

C++ method: void QCursor::setShape(Qt::CursorShape newShape)

C++ documentation:

Sets the cursor to the shape identified by shape.

See Qt::CursorShape for the list of cursor shapes.

See also shape().

fn shape(&self) -> CursorShape

C++ method: Qt::CursorShape QCursor::shape() const

C++ documentation:

Returns the cursor shape identifier. The return value is one of the Qt::CursorShape enum values (cast to an int).

See also setShape().