Module qt_gui::drag::overloading

Traits

DragConnectArgs
This trait represents a set of arguments accepted by Drag::connect method.

DragConnectStaticArgs
This trait represents a set of arguments accepted by Drag::connect_static method.

DragDestroyedArgs
This trait represents a set of arguments accepted by Drag::destroyed method.

DragDisconnectArgs
This trait represents a set of arguments accepted by Drag::disconnect method.

DragDisconnectStaticArgs
This trait represents a set of arguments accepted by Drag::disconnect_static method.

DragDragCursorArgs
This trait represents a set of arguments accepted by Drag::drag_cursor method.

DragDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Drag::dynamic_property_names method.

DragExecArgs
This trait represents a set of arguments accepted by Drag::exec method.

DragHotSpotArgs
This trait represents a set of arguments accepted by Drag::hot_spot method.

DragNewArgs
This trait represents a set of arguments accepted by Drag::new method.

DragObjectNameArgs
This trait represents a set of arguments accepted by Drag::object_name method.

DragPixmapArgs
This trait represents a set of arguments accepted by Drag::pixmap method.

DragPropertyArgs
This trait represents a set of arguments accepted by Drag::property method.

DragStartArgs
This trait represents a set of arguments accepted by Drag::start method.

DragStartTimerArgs
This trait represents a set of arguments accepted by Drag::start_timer method.

DragTrArgs
This trait represents a set of arguments accepted by Drag::tr method.

DragTrUtf8Args
This trait represents a set of arguments accepted by Drag::tr_utf8 method.