Struct qt_gui::drag_enter_event::DragEnterEvent

pub struct DragEnterEvent {
    // some fields omitted
}

C++ type: QDragEnterEvent

C++ documentation:

The QDragEnterEvent class provides an event which is sent to a widget when a drag and drop action enters it.

A widget must accept this event in order to receive the drag move events that are sent while the drag and drop action is in progress. The drag enter event is always immediately followed by a drag move event.

QDragEnterEvent inherits most of its functionality from QDragMoveEvent, which in turn inherits most of its functionality from QDropEvent.

Methods

impl DragEnterEvent

fn accept<'largs, Args>(&'largs mut self, args: Args) where Args: DragEnterEventAcceptArgs<'largs>

C++ method: QDragEnterEvent::accept

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn accept(&mut self, ()) -> ()
C++ method: void QDragEnterEvent::accept()

C++ documentation:

This is an overloaded function.

Calls QDropEvent::accept().

Variant 2

Rust arguments: fn accept(&mut self, &qt_core::rect::Rect) -> ()
C++ method: void QDragEnterEvent::accept(const QRect& r)

C++ documentation:

The same as accept(), but also notifies that future moves will also be acceptable if they remain within the rectangle given on the widget. This can improve performance, but may also be ignored by the underlying system.

If the rectangle is empty, drag move events will be sent continuously. This is useful if the source is scrolling in a timer event.

fn accept_proposed_action(&mut self)

C++ method: void QDragEnterEvent::acceptProposedAction()

C++ documentation:

Sets the drop action to be the proposed action.

See also setDropAction(), proposedAction(), and accept().

fn answer_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DragEnterEventAnswerRectArgs<'largs>

C++ method: QDragEnterEvent::answerRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn answer_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn answer_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QDragEnterEvent::answerRect() const

C++ documentation:

Returns the rectangle in the widget where the drop will occur if accepted. You can use this information to restrict drops to certain places on the widget.

fn drop_action(&self) -> DropAction

C++ method: Qt::DropAction QDragEnterEvent::dropAction() const

C++ documentation:

Returns the action to be performed on the data by the target. This may be different from the action supplied in proposedAction() if you have called setDropAction() to explicitly choose a drop action.

See also setDropAction().

fn ignore<'largs, Args>(&'largs mut self, args: Args) where Args: DragEnterEventIgnoreArgs<'largs>

C++ method: QDragEnterEvent::ignore

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn ignore(&mut self, ()) -> ()
C++ method: void QDragEnterEvent::ignore()

C++ documentation:

This is an overloaded function.

Calls QDropEvent::ignore().

Variant 2

Rust arguments: fn ignore(&mut self, &qt_core::rect::Rect) -> ()
C++ method: void QDragEnterEvent::ignore(const QRect& r)

C++ documentation:

The opposite of the accept(const QRect&) function. Moves within the rectangle are not acceptable, and will be ignored.

fn is_accepted(&self) -> bool

C++ method: bool QDragEnterEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn keyboard_modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QDragEnterEvent::keyboardModifiers() const

C++ documentation:

Returns the modifier keys that are pressed.

fn mime_data(&self) -> *const MimeData

C++ method: const QMimeData* QDragEnterEvent::mimeData() const

C++ documentation:

Returns the data that was dropped on the widget and its associated MIME type information.

fn mouse_buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QDragEnterEvent::mouseButtons() const

C++ documentation:

Returns the mouse buttons that are pressed..

fn new<Args>(args: Args) -> Args::ReturnType where Args: DragEnterEventNewArgs

C++ method: QDragEnterEvent::QDragEnterEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&qt_core::point::Point, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::drag_enter_event::DragEnterEvent
2) fn new((&qt_core::point::Point, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::drag_enter_event::DragEnterEvent>
C++ method: [constructor] void QDragEnterEvent::QDragEnterEvent(const QPoint& pos, QFlags<Qt::DropAction> actions, const QMimeData* data, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DragEnterEventPosArgs<'largs>

C++ method: QDragEnterEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QDragEnterEvent::pos() const

C++ documentation:

Returns the position where the drop was made.

fn pos_f<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QDragEnterEvent::posF() const

C++ documentation:

Returns the position where the drop was made.

fn possible_actions(&self) -> Flags<DropAction>

C++ method: QFlags<Qt::DropAction> QDragEnterEvent::possibleActions() const

C++ documentation:

Returns an OR-combination of possible drop actions.

See also dropAction().

fn proposed_action(&self) -> DropAction

C++ method: Qt::DropAction QDragEnterEvent::proposedAction() const

C++ documentation:

Returns the proposed drop action.

See also dropAction().

fn register_event_type<Args>(args: Args) -> c_int where Args: DragEnterEventRegisterEventTypeArgs

C++ method: QDragEnterEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QDragEnterEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QDragEnterEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_drop_action(&mut self, action: DropAction)

C++ method: void QDragEnterEvent::setDropAction(Qt::DropAction action)

C++ documentation:

Sets the action to be performed on the data by the target. Use this to override the proposed action with one of the possible actions.

If you set a drop action that is not one of the possible actions, the drag and drop operation will default to a copy operation.

Once you have supplied a replacement drop action, call accept() instead of acceptProposedAction().

See also dropAction().

fn source(&self) -> *mut Object

C++ method: QObject* QDragEnterEvent::source() const

C++ documentation:

If the source of the drag operation is a widget in this application, this function returns that source; otherwise it returns 0. The source of the operation is the first parameter to the QDrag object used instantiate the drag.

This is useful if your widget needs special behavior when dragging to itself.

See also QDrag::QDrag().

fn spontaneous(&self) -> bool

C++ method: bool QDragEnterEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QDragEnterEvent::type() const

C++ documentation:

Returns the event type.