Struct qt_gui::drop_event::DropEvent

pub struct DropEvent {
    // some fields omitted
}

C++ type: QDropEvent

C++ documentation:

The QDropEvent class provides an event which is sent when a drag and drop action is completed.

When a widget accepts drop events, it will receive this event if it has accepted the most recent QDragEnterEvent or QDragMoveEvent sent to it.

The drop event contains a proposed action, available from proposedAction(), for the widget to either accept or ignore. If the action can be handled by the widget, you should call the acceptProposedAction() function. Since the proposed action can be a combination of Qt::DropAction values, it may be useful to either select one of these values as a default action or ask the user to select their preferred action.

If the proposed drop action is not suitable, perhaps because your custom widget does not support that action, you can replace it with any of the possible drop actions by calling setDropAction() with your preferred action. If you set a value that is not present in the bitwise OR combination of values returned by possibleActions(), the default copy action will be used. Once a replacement drop action has been set, call accept() instead of acceptProposedAction() to complete the drop operation.

The mimeData() function provides the data dropped on the widget in a QMimeData object. This contains information about the MIME type of the data in addition to the data itself.

Methods

impl DropEvent

fn accept(&mut self)

C++ method: void QDropEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn accept_proposed_action(&mut self)

C++ method: void QDropEvent::acceptProposedAction()

C++ documentation:

Sets the drop action to be the proposed action.

See also setDropAction(), proposedAction(), and accept().

fn drop_action(&self) -> DropAction

C++ method: Qt::DropAction QDropEvent::dropAction() const

C++ documentation:

Returns the action to be performed on the data by the target. This may be different from the action supplied in proposedAction() if you have called setDropAction() to explicitly choose a drop action.

See also setDropAction().

fn ignore(&mut self)

C++ method: void QDropEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QDropEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn keyboard_modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QDropEvent::keyboardModifiers() const

C++ documentation:

Returns the modifier keys that are pressed.

fn mime_data(&self) -> *const MimeData

C++ method: const QMimeData* QDropEvent::mimeData() const

C++ documentation:

Returns the data that was dropped on the widget and its associated MIME type information.

fn mouse_buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QDropEvent::mouseButtons() const

C++ documentation:

Returns the mouse buttons that are pressed..

fn new<Args>(args: Args) -> Args::ReturnType where Args: DropEventNewArgs

C++ method: QDropEvent::QDropEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&qt_core::point_f::PointF, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::drop_event::DropEvent
2) fn new((&qt_core::point_f::PointF, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::event::Type, cpp_utils::AsStruct)) -> ::drop_event::DropEvent
3) fn new((&qt_core::point_f::PointF, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::drop_event::DropEvent>
4) fn new((&qt_core::point_f::PointF, qt_core::flags::Flags<qt_core::qt::DropAction>, *const qt_core::mime_data::MimeData, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::event::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::drop_event::DropEvent>
C++ method: [constructor] void QDropEvent::QDropEvent(const QPointF& pos, QFlags<Qt::DropAction> actions, const QMimeData* data, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers, QEvent::Type type = ?)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DropEventPosArgs<'largs>

C++ method: QDropEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QDropEvent::pos() const

C++ documentation:

Returns the position where the drop was made.

fn pos_f<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QDropEvent::posF() const

C++ documentation:

Returns the position where the drop was made.

fn possible_actions(&self) -> Flags<DropAction>

C++ method: QFlags<Qt::DropAction> QDropEvent::possibleActions() const

C++ documentation:

Returns an OR-combination of possible drop actions.

See also dropAction().

fn proposed_action(&self) -> DropAction

C++ method: Qt::DropAction QDropEvent::proposedAction() const

C++ documentation:

Returns the proposed drop action.

See also dropAction().

fn register_event_type<Args>(args: Args) -> c_int where Args: DropEventRegisterEventTypeArgs

C++ method: QDropEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QDropEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QDropEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_drop_action(&mut self, action: DropAction)

C++ method: void QDropEvent::setDropAction(Qt::DropAction action)

C++ documentation:

Sets the action to be performed on the data by the target. Use this to override the proposed action with one of the possible actions.

If you set a drop action that is not one of the possible actions, the drag and drop operation will default to a copy operation.

Once you have supplied a replacement drop action, call accept() instead of acceptProposedAction().

See also dropAction().

fn source(&self) -> *mut Object

C++ method: QObject* QDropEvent::source() const

C++ documentation:

If the source of the drag operation is a widget in this application, this function returns that source; otherwise it returns 0. The source of the operation is the first parameter to the QDrag object used instantiate the drag.

This is useful if your widget needs special behavior when dragging to itself.

See also QDrag::QDrag().

fn spontaneous(&self) -> bool

C++ method: bool QDropEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QDropEvent::type() const

C++ documentation:

Returns the event type.