Enum qt_gui::font::StyleHint

pub enum StyleHint {
    Helvetica,
    Times,
    Courier,
    OldEnglish,
    System,
    AnyStyle,
    Cursive,
    Monospace,
    Fantasy,
}

C++ type: QFont::StyleHint

C++ documentation:

Style hints are used by the font matching algorithm to find an appropriate default family if a selected font family is not available.

Variants
Helvetica

This variant corresponds to multiple C++ enum variants with the same value:

Helvetica = 0: is a synonym for SansSerif.

SansSerif = 0: the font matcher prefer sans serif fonts.

Times

This variant corresponds to multiple C++ enum variants with the same value:

Times = 1: is a synonym for Serif.

Serif = 1: the font matcher prefers serif fonts.

Courier

This variant corresponds to multiple C++ enum variants with the same value:

Courier = 2: a synonym for TypeWriter.

TypeWriter = 2: the font matcher prefers fixed pitch fonts.

OldEnglish

This variant corresponds to multiple C++ enum variants with the same value:

OldEnglish = 3: the font matcher prefers decorative fonts.

Decorative = 3: is a synonym for OldEnglish.

System

the font matcher prefers system fonts. (C++ enum variant: System = 4)

AnyStyle

leaves the font matching algorithm to choose the family. This is the default. (C++ enum variant: AnyStyle = 5)

Cursive

the font matcher prefers fonts that map to the CSS generic font-family 'cursive'. (C++ enum variant: Cursive = 6)

Monospace

the font matcher prefers fonts that map to the CSS generic font-family 'monospace'. (C++ enum variant: Monospace = 7)

Fantasy

the font matcher prefers fonts that map to the CSS generic font-family 'fantasy'. (C++ enum variant: Fantasy = 8)