Enum qt_gui::font::StyleStrategy

pub enum StyleStrategy {
    PreferDefault,
    PreferBitmap,
    PreferDevice,
    PreferOutline,
    ForceOutline,
    PreferMatch,
    PreferQuality,
    PreferAntialias,
    NoAntialias,
    OpenGLCompatible,
    ForceIntegerMetrics,
    NoSubpixelAntialias,
    NoFontMerging,
}

C++ type: QFont::StyleStrategy

C++ documentation:

The style strategy tells the font matching algorithm what type of fonts should be used to find an appropriate default family.

The following strategies are available:

Any of these may be OR-ed with one of these flags:

Variants
PreferDefault

the default style strategy. It does not prefer any type of font. (C++ enum variant: PreferDefault = 1)

PreferBitmap

prefers bitmap fonts (as opposed to outline fonts). (C++ enum variant: PreferBitmap = 2)

PreferDevice

prefers device fonts. (C++ enum variant: PreferDevice = 4)

PreferOutline

prefers outline fonts (as opposed to bitmap fonts). (C++ enum variant: PreferOutline = 8)

ForceOutline

forces the use of outline fonts. (C++ enum variant: ForceOutline = 16)

PreferMatch

prefer an exact match. The font matcher will try to use the exact font size that has been specified. (C++ enum variant: PreferMatch = 32)

PreferQuality

prefer the best quality font. The font matcher will use the nearest standard point size that the font supports. (C++ enum variant: PreferQuality = 64)

PreferAntialias

antialias if possible. (C++ enum variant: PreferAntialias = 128)

NoAntialias

don't antialias the fonts. (C++ enum variant: NoAntialias = 256)

OpenGLCompatible

forces the use of OpenGL compatible fonts. (C++ enum variant: OpenGLCompatible = 512)

ForceIntegerMetrics

forces the use of integer values in font engines that support fractional font metrics. (C++ enum variant: ForceIntegerMetrics = 1024)

NoSubpixelAntialias

avoid subpixel antialiasing on the fonts if possible. (C++ enum variant: NoSubpixelAntialias = 2048)

NoFontMerging

If the font selected for a certain writing system does not contain a character requested to draw, then Qt automatically chooses a similar looking font that contains the character. The NoFontMerging flag disables this feature. Please note that enabling this flag will not prevent Qt from automatically picking a suitable font when the selected font does not support the writing system of the text. (C++ enum variant: NoFontMerging = 32768)