Struct qt_gui::font::Font

pub struct Font {
    // some fields omitted
}

C++ type: QFont

C++ documentation:

The QFont class specifies a font used for drawing text.

When you create a QFont object you specify various attributes that you want the font to have. Qt will use the font with the specified attributes, or if no matching font exists, Qt will use the closest matching installed font. The attributes of the font that is actually used are retrievable from a QFontInfo object. If the window system provides an exact match exactMatch() returns true. Use QFontMetrics to get measurements, e.g. the pixel length of a string using QFontMetrics::width().

Note that a QGuiApplication instance must exist before a QFont can be used. You can set the application's default font with QGuiApplication::setFont().

If a chosen font does not include all the characters that need to be displayed, QFont will try to find the characters in the nearest equivalent fonts. When a QPainter draws a character from a font the QFont will report whether or not it has the character; if it does not, QPainter will draw an unfilled square.

Create QFonts like this:

QFont serifFont("Times", 10, QFont::Bold);
QFont sansFont("Helvetica [Cronyx]", 12);

The attributes set in the constructor can also be set later, e.g. setFamily(), setPointSize(), setPointSizeF(), setWeight() and setItalic(). The remaining attributes must be set after contstruction, e.g. setBold(), setUnderline(), setOverline(), setStrikeOut() and setFixedPitch(). QFontInfo objects should be created after the font's attributes have been set. A QFontInfo object will not change, even if you change the font's attributes. The corresponding "get" functions, e.g. family(), pointSize(), etc., return the values that were set, even though the values used may differ. The actual values are available from a QFontInfo object.

If the requested font family is unavailable you can influence the font matching algorithm by choosing a particular QFont::StyleHint and QFont::StyleStrategy with setStyleHint(). The default family (corresponding to the current style hint) is returned by defaultFamily().

The font-matching algorithm has a lastResortFamily() and lastResortFont() in cases where a suitable match cannot be found. You can provide substitutions for font family names using insertSubstitution() and insertSubstitutions(). Substitutions can be removed with removeSubstitutions(). Use substitute() to retrieve a family's first substitute, or the family name itself if it has no substitutes. Use substitutes() to retrieve a list of a family's substitutes (which may be empty).

Every QFont has a key() which you can use, for example, as the key in a cache or dictionary. If you want to store a user's font preferences you could use QSettings, writing the font information with toString() and reading it back with fromString(). The operator<<() and operator>>() functions are also available, but they work on a data stream.

It is possible to set the height of characters shown on the screen to a specified number of pixels with setPixelSize(); however using setPointSize() has a similar effect and provides device independence.

Loading fonts can be expensive, especially on X11. QFont contains extensive optimizations to make the copying of QFont objects fast, and to cache the results of the slow window system functions it depends upon.

The font matching algorithm works as follows:

The specified font family is searched for.

If not found, the styleHint() is used to select a replacement family.

Each replacement font family is searched for.

If none of these are found or there was no styleHint(), "helvetica" will be searched for.

If "helvetica" isn't found Qt will try the lastResortFamily().

If the lastResortFamily() isn't found Qt will try the lastResortFont() which will always return a name of some kind.

Note that the actual font matching algorithm varies from platform to platform.

In Windows a request for the "Courier" font is automatically changed to "Courier New", an improved version of Courier that allows for smooth scaling. The older "Courier" bitmap font can be selected by setting the PreferBitmap style strategy (see setStyleStrategy()).

Once a font is found, the remaining attributes are matched in order of priority:

fixedPitch()

pointSize() (see below)

weight()

style()

If you have a font which matches on family, even if none of the other attributes match, this font will be chosen in preference to a font which doesn't match on family but which does match on the other attributes. This is because font family is the dominant search criteria.

The point size is defined to match if it is within 20% of the requested point size. When several fonts match and are only distinguished by point size, the font with the closest point size to the one requested will be chosen.

The actual family, font size, weight and other font attributes used for drawing text will depend on what's available for the chosen family under the window system. A QFontInfo object can be used to determine the actual values used for drawing the text.

Examples:

QFont f("Helvetica");

If you had both an Adobe and a Cronyx Helvetica, you might get either.

QFont f("Helvetica [Cronyx]");

You can specify the foundry you want in the family name. The font f in the above example will be set to "Helvetica [Cronyx]".

To determine the attributes of the font actually used in the window system, use a QFontInfo object, e.g.

QFontInfo info(f1);
QString family = info.family();

To find out font metrics use a QFontMetrics object, e.g.

QFontMetrics fm(f1);
int textWidthInPixels = fm.width("How many pixels wide is this text?");
int textHeightInPixels = fm.height();

For more general information on fonts, see the comp.fonts FAQ. Information on encodings can be found from Roman Czyborra's page.

Methods

impl Font

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontAsQVariantArgs<'largs>

C++ method: QFont::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QFont::operator QVariant() const

C++ documentation:

Returns the font as a QVariant

fn bold(&self) -> bool

C++ method: bool QFont::bold() const

C++ documentation:

Returns true if weight() is a value greater than QFont::Medium; otherwise returns false.

See also weight(), setBold(), and QFontInfo::bold().

fn cache_statistics()

C++ method: static void QFont::cacheStatistics()

fn capitalization(&self) -> Capitalization

C++ method: QFont::Capitalization QFont::capitalization() const

C++ documentation:

Returns the current capitalization type of the font.

This function was introduced in Qt 4.4.

See also setCapitalization().

fn cleanup()

C++ method: static void QFont::cleanup()

fn default_family<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontDefaultFamilyArgs<'largs>

C++ method: QFont::defaultFamily

This is an overloaded function. Available variants:

Rust arguments:
1) fn default_family(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn default_family(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::defaultFamily() const

C++ documentation:

Returns the family name that corresponds to the current style hint.

See also StyleHint, styleHint(), and setStyleHint().

fn exact_match(&self) -> bool

C++ method: bool QFont::exactMatch() const

C++ documentation:

Returns true if a window system font exactly matching the settings of this font is available.

See also QFontInfo.

fn family<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontFamilyArgs<'largs>

C++ method: QFont::family

This is an overloaded function. Available variants:

Rust arguments:
1) fn family(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn family(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::family() const

C++ documentation:

Returns the requested font family name, i.e. the name set in the constructor or the last setFont() call.

See also setFamily(), substitutes(), and substitute().

fn fixed_pitch(&self) -> bool

C++ method: bool QFont::fixedPitch() const

C++ documentation:

Returns true if fixed pitch has been set; otherwise returns false.

See also setFixedPitch() and QFontInfo::fixedPitch().

fn from_string(&mut self, arg1: &String) -> bool

C++ method: bool QFont::fromString(const QString& arg1)

C++ documentation:

Sets this font to match the description descrip. The description is a comma-separated list of the font attributes, as returned by toString().

See also toString().

fn hinting_preference(&self) -> HintingPreference

C++ method: QFont::HintingPreference QFont::hintingPreference() const

C++ documentation:

Returns the currently preferred hinting level for glyphs rendered with this font.

This function was introduced in Qt 4.8.

See also setHintingPreference().

fn initialize()

C++ method: static void QFont::initialize()

fn insert_substitution(arg1: &String, arg2: &String)

C++ method: static void QFont::insertSubstitution(const QString& arg1, const QString& arg2)

C++ documentation:

Inserts substituteName into the substitution table for the family familyName.

See also insertSubstitutions(), removeSubstitutions(), substitutions(), substitute(), and substitutes().

fn insert_substitutions(arg1: &String, arg2: &StringList)

C++ method: static void QFont::insertSubstitutions(const QString& arg1, const QStringList& arg2)

C++ documentation:

Inserts the list of families substituteNames into the substitution list for familyName.

See also insertSubstitution(), removeSubstitutions(), substitutions(), and substitute().

fn is_copy_of(&self, arg1: &Font) -> bool

C++ method: bool QFont::isCopyOf(const QFont& arg1) const

C++ documentation:

Returns true if this font and f are copies of each other, i.e. one of them was created as a copy of the other and neither has been modified since. This is much stricter than equality.

See also operator=() and operator==().

fn italic(&self) -> bool

C++ method: bool QFont::italic() const

C++ documentation:

Returns true if the style() of the font is not QFont::StyleNormal

See also setItalic() and style().

fn kerning(&self) -> bool

C++ method: bool QFont::kerning() const

C++ documentation:

Returns true if kerning should be used when drawing text with this font.

See also setKerning().

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontKeyArgs<'largs>

C++ method: QFont::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn key(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::key() const

C++ documentation:

Returns the font's key, a textual representation of a font. It is typically used as the key for a cache or dictionary of fonts.

See also QMap.

fn last_resort_family<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontLastResortFamilyArgs<'largs>

C++ method: QFont::lastResortFamily

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_resort_family(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn last_resort_family(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::lastResortFamily() const

C++ documentation:

Returns the "last resort" font family name.

The current implementation tries a wide variety of common fonts, returning the first one it finds. Is is possible that no family is found in which case an empty string is returned.

See also lastResortFont().

fn last_resort_font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontLastResortFontArgs<'largs>

C++ method: QFont::lastResortFont

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_resort_font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn last_resort_font(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::lastResortFont() const

C++ documentation:

Returns a "last resort" font name for the font matching algorithm. This is used if the last resort family is not available. It will always return a name, if necessary returning something like "fixed" or "system".

The current implementation tries a wide variety of common fonts, returning the first one it finds. The implementation may change at any time, but this function will always return a string containing something.

It is theoretically possible that there really isn't a lastResortFont() in which case Qt will abort with an error message. We have not been able to identify a case where this happens. Please report it as a bug if it does, preferably with a list of the fonts you have installed.

See also lastResortFamily().

fn letter_spacing(&self) -> c_double

C++ method: double QFont::letterSpacing() const

C++ documentation:

Returns the letter spacing for the font.

This function was introduced in Qt 4.4.

See also setLetterSpacing(), letterSpacingType(), and setWordSpacing().

fn letter_spacing_type(&self) -> SpacingType

C++ method: QFont::SpacingType QFont::letterSpacingType() const

C++ documentation:

Returns the spacing type used for letter spacing.

This function was introduced in Qt 4.4.

See also letterSpacing(), setLetterSpacing(), and setWordSpacing().

fn new<Args>(args: Args) -> Args::ReturnType where Args: FontNewArgs

C++ method: QFont::QFont

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::font::Font
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
C++ method: [constructor] void QFont::QFont()

Variant 2

Rust arguments:
1) fn new((&::font::Font, cpp_utils::AsStruct)) -> ::font::Font
2) fn new((&::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
C++ method: [constructor] void QFont::QFont(const QFont& arg1)

Variant 3

Rust arguments:
1) fn new((&::font::Font, *mut ::paint_device::PaintDevice, cpp_utils::AsStruct)) -> ::font::Font
2) fn new((&::font::Font, *mut ::paint_device::PaintDevice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
C++ method: [constructor] void QFont::QFont(const QFont& arg1, QPaintDevice* pd)

Variant 4

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::font::Font
2) fn new((&qt_core::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::font::Font
3) fn new((&qt_core::string::String, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::font::Font
4) fn new((&qt_core::string::String, libc::c_int, libc::c_int, bool, cpp_utils::AsStruct)) -> ::font::Font
5) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
6) fn new((&qt_core::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
7) fn new((&qt_core::string::String, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
8) fn new((&qt_core::string::String, libc::c_int, libc::c_int, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
C++ method: [constructor] void QFont::QFont(const QString& family, int pointSize = ?, int weight = ?, bool italic = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Font) -> &'l0 mut Font

C++ method: QFont& QFont::operator=(const QFont& arg1)

C++ documentation:

Assigns font to this font and returns a reference to it.

fn op_eq(&self, arg1: &Font) -> bool

C++ method: bool QFont::operator==(const QFont& arg1) const

C++ documentation:

Returns true if this font is equal to f; otherwise returns false.

Two QFonts are considered equal if their font attributes are equal.

See also operator!=() and isCopyOf().

fn op_lt(&self, arg1: &Font) -> bool

C++ method: bool QFont::operator<(const QFont& arg1) const

C++ documentation:

Provides an arbitrary comparison of this font and font f. All that is guaranteed is that the operator returns false if both fonts are equal and that (f1 < f2) == !(f2 < f1) if the fonts are not equal.

This function is useful in some circumstances, for example if you want to use QFont objects as keys in a QMap.

See also operator==(), operator!=(), and isCopyOf().

fn op_neq(&self, arg1: &Font) -> bool

C++ method: bool QFont::operator!=(const QFont& arg1) const

C++ documentation:

Returns true if this font is different from f; otherwise returns false.

Two QFonts are considered to be different if their font attributes are different.

See also operator==().

fn overline(&self) -> bool

C++ method: bool QFont::overline() const

C++ documentation:

Returns true if overline has been set; otherwise returns false.

See also setOverline().

fn pixel_size(&self) -> c_int

C++ method: int QFont::pixelSize() const

C++ documentation:

Returns the pixel size of the font if it was set with setPixelSize(). Returns -1 if the size was set with setPointSize() or setPointSizeF().

See also setPixelSize(), pointSize(), QFontInfo::pointSize(), and QFontInfo::pixelSize().

fn point_size(&self) -> c_int

C++ method: int QFont::pointSize() const

C++ documentation:

Returns the point size of the font. Returns -1 if the font size was specified in pixels.

See also setPointSize() and pointSizeF().

fn point_size_f(&self) -> c_double

C++ method: double QFont::pointSizeF() const

C++ documentation:

Returns the point size of the font. Returns -1 if the font size was specified in pixels.

See also pointSize(), setPointSizeF(), pixelSize(), QFontInfo::pointSize(), and QFontInfo::pixelSize().

fn raw_mode(&self) -> bool

C++ method: bool QFont::rawMode() const

C++ documentation:

Returns true if raw mode is used for font name matching; otherwise returns false.

See also setRawMode().

fn raw_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontRawNameArgs<'largs>

C++ method: QFont::rawName

This is an overloaded function. Available variants:

Rust arguments:
1) fn raw_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn raw_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::rawName() const

C++ documentation:

Returns the name of the font within the underlying window system.

On X11, this function will return an empty string.

Using the return value of this function is usually not portable.

See also setRawName().

fn remove_substitutions(arg1: &String)

C++ method: static void QFont::removeSubstitutions(const QString& arg1)

C++ documentation:

Removes all the substitutions for familyName.

This function was introduced in Qt 5.0.

See also insertSubstitutions(), insertSubstitution(), substitutions(), and substitute().

fn resolve<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontResolveArgs<'largs>

C++ method: QFont::resolve

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn resolve(&self, (&::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
2) fn resolve(&self, (&::font::Font, cpp_utils::AsStruct)) -> ::font::Font
C++ method: QFont QFont::resolve(const QFont& arg1) const

C++ documentation:

Returns a new QFont that has attributes copied from other that have not been previously set on this font.

Variant 2

Rust arguments: fn resolve(&self, ()) -> libc::c_uint
C++ method: unsigned int QFont::resolve() const

fn resolve_mut(&mut self, mask: c_uint)

C++ method: void QFont::resolve(unsigned int mask)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for QFont QFont::resolve(const QFont & other) const:

Returns a new QFont that has attributes copied from other that have not been previously set on this font.

fn set_bold(&mut self, arg1: bool)

C++ method: void QFont::setBold(bool arg1)

C++ documentation:

If enable is true sets the font's weight to QFont::Bold; otherwise sets the weight to QFont::Normal.

For finer boldness control use setWeight().

See also bold() and setWeight().

fn set_capitalization(&mut self, arg1: Capitalization)

C++ method: void QFont::setCapitalization(QFont::Capitalization arg1)

C++ documentation:

Sets the capitalization of the text in this font to caps.

A font's capitalization makes the text appear in the selected capitalization mode.

This function was introduced in Qt 4.4.

See also capitalization().

fn set_family(&mut self, arg1: &String)

C++ method: void QFont::setFamily(const QString& arg1)

C++ documentation:

Sets the family name of the font. The name is case insensitive and may include a foundry name.

The family name may optionally also include a foundry name, e.g. "Helvetica [Cronyx]". If the family is available from more than one foundry and the foundry isn't specified, an arbitrary foundry is chosen. If the family isn't available a family will be set using the font matching algorithm.

See also family(), setStyleHint(), and QFontInfo.

fn set_fixed_pitch(&mut self, arg1: bool)

C++ method: void QFont::setFixedPitch(bool arg1)

C++ documentation:

If enable is true, sets fixed pitch on; otherwise sets fixed pitch off.

See also fixedPitch() and QFontInfo.

fn set_hinting_preference(&mut self, hinting_preference: HintingPreference)

C++ method: void QFont::setHintingPreference(QFont::HintingPreference hintingPreference)

C++ documentation:

Set the preference for the hinting level of the glyphs to hintingPreference. This is a hint to the underlying font rendering system to use a certain level of hinting, and has varying support across platforms. See the table in the documentation for QFont::HintingPreference for more details.

The default hinting preference is QFont::PreferDefaultHinting.

This function was introduced in Qt 4.8.

See also hintingPreference().

fn set_italic(&mut self, b: bool)

C++ method: void QFont::setItalic(bool b)

C++ documentation:

Sets the style() of the font to QFont::StyleItalic if enable is true; otherwise the style is set to QFont::StyleNormal.

See also italic() and QFontInfo.

fn set_kerning(&mut self, arg1: bool)

C++ method: void QFont::setKerning(bool arg1)

C++ documentation:

Enables kerning for this font if enable is true; otherwise disables it. By default, kerning is enabled.

When kerning is enabled, glyph metrics do not add up anymore, even for Latin text. In other words, the assumption that width('a') + width('b') is equal to width("ab") is not necessarily true.

See also kerning() and QFontMetrics.

fn set_letter_spacing(&mut self, type_: SpacingType, spacing: c_double)

C++ method: void QFont::setLetterSpacing(QFont::SpacingType type, double spacing)

C++ documentation:

Sets the letter spacing for the font to spacing and the type of spacing to type.

Letter spacing changes the default spacing between individual letters in the font. The spacing between the letters can be made smaller as well as larger either in percentage of the character width or in pixels, depending on the selected spacing type.

This function was introduced in Qt 4.4.

See also letterSpacing(), letterSpacingType(), and setWordSpacing().

fn set_overline(&mut self, arg1: bool)

C++ method: void QFont::setOverline(bool arg1)

C++ documentation:

If enable is true, sets overline on; otherwise sets overline off.

See also overline() and QFontInfo.

fn set_pixel_size(&mut self, arg1: c_int)

C++ method: void QFont::setPixelSize(int arg1)

C++ documentation:

Sets the font size to pixelSize pixels.

Using this function makes the font device dependent. Use setPointSize() or setPointSizeF() to set the size of the font in a device independent manner.

See also pixelSize().

fn set_point_size(&mut self, arg1: c_int)

C++ method: void QFont::setPointSize(int arg1)

C++ documentation:

Sets the point size to pointSize. The point size must be greater than zero.

See also pointSize() and setPointSizeF().

fn set_point_size_f(&mut self, arg1: c_double)

C++ method: void QFont::setPointSizeF(double arg1)

C++ documentation:

Sets the point size to pointSize. The point size must be greater than zero. The requested precision may not be achieved on all platforms.

See also pointSizeF(), setPointSize(), and setPixelSize().

fn set_raw_mode(&mut self, arg1: bool)

C++ method: void QFont::setRawMode(bool arg1)

C++ documentation:

If enable is true, turns raw mode on; otherwise turns raw mode off. This function only has an effect under X11.

If raw mode is enabled, Qt will search for an X font with a complete font name matching the family name, ignoring all other values set for the QFont. If the font name matches several fonts, Qt will use the first font returned by X. QFontInfo cannot be used to fetch information about a QFont using raw mode (it will return the values set in the QFont for all parameters, including the family name).

Warning: Enabling raw mode has no effect since Qt 5.0.

See also rawMode().

fn set_raw_name(&mut self, arg1: &String)

C++ method: void QFont::setRawName(const QString& arg1)

C++ documentation:

Sets a font by its system specific name.

A font set with setRawName() is still a full-featured QFont. It can be queried (for example with italic()) or modified (for example with setItalic()) and is therefore also suitable for rendering rich text.

If Qt's internal font database cannot resolve the raw name, the font becomes a raw font with name as its family.

See also rawName() and setFamily().

fn set_stretch(&mut self, arg1: c_int)

C++ method: void QFont::setStretch(int arg1)

C++ documentation:

Sets the stretch factor for the font.

The stretch factor changes the width of all characters in the font by factor percent. For example, setting factor to 150 results in all characters in the font being 1.5 times (ie. 150%) wider. The default stretch factor is 100. The minimum stretch factor is 1, and the maximum stretch factor is 4000.

The stretch factor is only applied to outline fonts. The stretch factor is ignored for bitmap fonts.

See also stretch() and QFont::Stretch.

fn set_strike_out(&mut self, arg1: bool)

C++ method: void QFont::setStrikeOut(bool arg1)

C++ documentation:

If enable is true, sets strikeout on; otherwise sets strikeout off.

See also strikeOut() and QFontInfo.

fn set_style(&mut self, style: Style)

C++ method: void QFont::setStyle(QFont::Style style)

C++ documentation:

Sets the style of the font to style.

See also style(), italic(), and QFontInfo.

fn set_style_hint<'largs, Args>(&'largs mut self, args: Args) where Args: FontSetStyleHintArgs<'largs>

C++ method: QFont::setStyleHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_style_hint(&mut self, ::font::StyleHint) -> ()
2) fn set_style_hint(&mut self, (::font::StyleHint, ::font::StyleStrategy)) -> ()
C++ method: void QFont::setStyleHint(QFont::StyleHint arg1, QFont::StyleStrategy arg2 = ?)

C++ documentation:

Sets the style hint and strategy to hint and strategy, respectively.

If these aren't set explicitly the style hint will default to AnyStyle and the style strategy to PreferDefault.

Qt does not support style hints on X11 since this information is not provided by the window system.

See also StyleHint, styleHint(), StyleStrategy, styleStrategy(), and QFontInfo.

fn set_style_name(&mut self, arg1: &String)

C++ method: void QFont::setStyleName(const QString& arg1)

C++ documentation:

Sets the style name of the font to styleName. When set, other style properties like style() and weight() will be ignored for font matching.

This function was introduced in Qt 4.8.

See also styleName().

fn set_style_strategy(&mut self, s: StyleStrategy)

C++ method: void QFont::setStyleStrategy(QFont::StyleStrategy s)

C++ documentation:

Sets the style strategy for the font to s.

See also styleStrategy() and QFont::StyleStrategy.

fn set_underline(&mut self, arg1: bool)

C++ method: void QFont::setUnderline(bool arg1)

C++ documentation:

If enable is true, sets underline on; otherwise sets underline off.

See also underline() and QFontInfo.

fn set_weight(&mut self, arg1: c_int)

C++ method: void QFont::setWeight(int arg1)

C++ documentation:

Sets the weight of the font to weight, using the scale defined by QFont::Weight enumeration.

See also weight() and QFontInfo.

fn set_word_spacing(&mut self, spacing: c_double)

C++ method: void QFont::setWordSpacing(double spacing)

C++ documentation:

Sets the word spacing for the font to spacing.

Word spacing changes the default spacing between individual words. A positive value increases the word spacing by a corresponding amount of pixels, while a negative value decreases the inter-word spacing accordingly.

Word spacing will not apply to writing systems, where indiviaul words are not separated by white space.

This function was introduced in Qt 4.4.

See also wordSpacing() and setLetterSpacing().

fn stretch(&self) -> c_int

C++ method: int QFont::stretch() const

C++ documentation:

Returns the stretch factor for the font.

See also setStretch().

fn strike_out(&self) -> bool

C++ method: bool QFont::strikeOut() const

C++ documentation:

Returns true if strikeout has been set; otherwise returns false.

See also setStrikeOut().

fn style(&self) -> Style

C++ method: QFont::Style QFont::style() const

C++ documentation:

Returns the style of the font.

See also setStyle().

fn style_hint(&self) -> StyleHint

C++ method: QFont::StyleHint QFont::styleHint() const

C++ documentation:

Returns the StyleHint.

The style hint affects the font matching algorithm. See QFont::StyleHint for the list of available hints.

See also setStyleHint(), QFont::StyleStrategy, and QFontInfo::styleHint().

fn style_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontStyleNameArgs<'largs>

C++ method: QFont::styleName

This is an overloaded function. Available variants:

Rust arguments:
1) fn style_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn style_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::styleName() const

C++ documentation:

Returns the requested font style name, it will be used to match the font with irregular styles (that can't be normalized in other style properties). It depends on system font support, thus only works for Mac OS X and X11 so far. On Windows irregular styles will be added as separate font families so there is no need for this.

This function was introduced in Qt 4.8.

See also setStyleName(), setFamily(), and setStyle().

fn style_strategy(&self) -> StyleStrategy

C++ method: QFont::StyleStrategy QFont::styleStrategy() const

C++ documentation:

Returns the StyleStrategy.

The style strategy affects the font matching algorithm. See QFont::StyleStrategy for the list of available strategies.

See also setStyleStrategy(), setStyleHint(), and QFont::StyleHint.

fn substitute<Args>(args: Args) -> Args::ReturnType where Args: FontSubstituteArgs

C++ method: QFont::substitute

This is an overloaded function. Available variants:

Rust arguments:
1) fn substitute((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn substitute((&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QFont::substitute(const QString& arg1)

C++ documentation:

Returns the first family name to be used whenever familyName is specified. The lookup is case insensitive.

If there is no substitution for familyName, familyName is returned.

To obtain a list of substitutions use substitutes().

See also setFamily(), insertSubstitutions(), insertSubstitution(), and removeSubstitutions().

fn substitutes<Args>(args: Args) -> Args::ReturnType where Args: FontSubstitutesArgs

C++ method: QFont::substitutes

This is an overloaded function. Available variants:

Rust arguments:
1) fn substitutes((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn substitutes((&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: static QStringList QFont::substitutes(const QString& arg1)

C++ documentation:

Returns a list of family names to be used whenever familyName is specified. The lookup is case insensitive.

If there is no substitution for familyName, an empty list is returned.

See also substitute(), insertSubstitutions(), insertSubstitution(), and removeSubstitutions().

fn substitutions<Args>(args: Args) -> Args::ReturnType where Args: FontSubstitutionsArgs

C++ method: QFont::substitutions

This is an overloaded function. Available variants:

Rust arguments:
1) fn substitutions(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn substitutions(cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: static QStringList QFont::substitutions()

C++ documentation:

Returns a sorted list of substituted family names.

See also insertSubstitution(), removeSubstitution(), and substitute().

fn swap(&mut self, other: &mut Font)

C++ method: void QFont::swap(QFont& other)

C++ documentation:

Swaps this font instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontToStringArgs<'largs>

C++ method: QFont::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn to_string(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFont::toString() const

C++ documentation:

Returns a description of the font. The description is a comma-separated list of the attributes, perfectly suited for use in QSettings.

See also fromString().

fn underline(&self) -> bool

C++ method: bool QFont::underline() const

C++ documentation:

Returns true if underline has been set; otherwise returns false.

See also setUnderline().

fn weight(&self) -> c_int

C++ method: int QFont::weight() const

C++ documentation:

Returns the weight of the font, using the same scale as the QFont::Weight enumeration.

See also setWeight(), Weight, and QFontInfo.

fn word_spacing(&self) -> c_double

C++ method: double QFont::wordSpacing() const

C++ documentation:

Returns the word spacing for the font.

This function was introduced in Qt 4.4.

See also setWordSpacing() and setLetterSpacing().