Module qt_gui::font_database::overloading

Traits

FontDatabaseApplicationFontFamiliesArgs
This trait represents a set of arguments accepted by FontDatabase::application_font_families method.

FontDatabaseFamiliesArgs
This trait represents a set of arguments accepted by FontDatabase::families method.

FontDatabaseFontArgs
This trait represents a set of arguments accepted by FontDatabase::font method.

FontDatabaseIsBitmapScalableArgs
This trait represents a set of arguments accepted by FontDatabase::is_bitmap_scalable method.

FontDatabaseIsFixedPitchArgs
This trait represents a set of arguments accepted by FontDatabase::is_fixed_pitch method.

FontDatabaseIsScalableArgs
This trait represents a set of arguments accepted by FontDatabase::is_scalable method.

FontDatabaseIsSmoothlyScalableArgs
This trait represents a set of arguments accepted by FontDatabase::is_smoothly_scalable method.

FontDatabaseNewArgs
This trait represents a set of arguments accepted by FontDatabase::new method.

FontDatabasePointSizesArgs
This trait represents a set of arguments accepted by FontDatabase::point_sizes method.

FontDatabaseSmoothSizesArgs
This trait represents a set of arguments accepted by FontDatabase::smooth_sizes method.

FontDatabaseStandardSizesArgs
This trait represents a set of arguments accepted by FontDatabase::standard_sizes method.

FontDatabaseStyleStringArgs
This trait represents a set of arguments accepted by FontDatabase::style_string method.

FontDatabaseStylesArgs
This trait represents a set of arguments accepted by FontDatabase::styles method.

FontDatabaseSystemFontArgs
This trait represents a set of arguments accepted by FontDatabase::system_font method.

FontDatabaseWritingSystemNameArgs
This trait represents a set of arguments accepted by FontDatabase::writing_system_name method.

FontDatabaseWritingSystemSampleArgs
This trait represents a set of arguments accepted by FontDatabase::writing_system_sample method.

FontDatabaseWritingSystemsArgs
This trait represents a set of arguments accepted by FontDatabase::writing_systems method.