Struct qt_gui::font_database::FontDatabase

pub struct FontDatabase {
    // some fields omitted
}

C++ type: QFontDatabase

C++ documentation:

The QFontDatabase class provides information about the fonts available in the underlying window system.

The most common uses of this class are to query the database for the list of font families() and for the pointSizes() and styles() that are available for each family. An alternative to pointSizes() is smoothSizes() which returns the sizes at which a given family and style will look attractive.

If the font family is available from two or more foundries the foundry name is included in the family name; for example: "Helvetica [Adobe]" and "Helvetica [Cronyx]". When you specify a family, you can either use the old hyphenated "foundry-family" format or the bracketed "family [foundry]" format; for example: "Cronyx-Helvetica" or "Helvetica [Cronyx]". If the family has a foundry it is always returned using the bracketed format, as is the case with the value returned by families().

The font() function returns a QFont given a family, style and point size.

A family and style combination can be checked to see if it is italic() or bold(), and to retrieve its weight(). Similarly we can call isBitmapScalable(), isSmoothlyScalable(), isScalable() and isFixedPitch().

Use the styleString() to obtain a text version of a style.

The QFontDatabase class also supports some static functions, for example, standardSizes(). You can retrieve the description of a writing system using writingSystemName(), and a sample of characters in a writing system with writingSystemSample().

Example:

    QFontDatabase database;
    QTreeWidget fontTree;
    fontTree.setColumnCount(2);
    fontTree.setHeaderLabels(QStringList() << "Font" << "Smooth Sizes");

foreach (const QString &family, database.families()) { QTreeWidgetItem *familyItem = new QTreeWidgetItem(&fontTree); familyItem->setText(0, family);

foreach (const QString &style, database.styles(family)) { QTreeWidgetItem *styleItem = new QTreeWidgetItem(familyItem); styleItem->setText(0, style);

QString sizes; foreach (int points, database.smoothSizes(family, style)) sizes += QString::number(points) + " ";

styleItem->setText(1, sizes.trimmed()); } }
This example gets the list of font families, the list of styles for each family, and the point sizes that are available for each combination of family and style, displaying this information in a tree view.

Methods

impl FontDatabase

fn add_application_font(file_name: &String) -> c_int

C++ method: static int QFontDatabase::addApplicationFont(const QString& fileName)

C++ documentation:

Loads the font from the file specified by fileName and makes it available to the application. An ID is returned that can be used to remove the font again with removeApplicationFont() or to retrieve the list of family names contained in the font.

The function returns -1 if the font could not be loaded.

Currently only TrueType fonts, TrueType font collections, and OpenType fonts are supported.

Note: Adding application fonts on Unix/X11 platforms without fontconfig is currently not supported.

This function was introduced in Qt 4.2.

See also addApplicationFontFromData(), applicationFontFamilies(), and removeApplicationFont().

fn add_application_font_from_data(font_data: &ByteArray) -> c_int

C++ method: static int QFontDatabase::addApplicationFontFromData(const QByteArray& fontData)

C++ documentation:

Loads the font from binary data specified by fontData and makes it available to the application. An ID is returned that can be used to remove the font again with removeApplicationFont() or to retrieve the list of family names contained in the font.

The function returns -1 if the font could not be loaded.

Currently only TrueType fonts and TrueType font collections are supported.

Note: Adding application fonts on Unix/X11 platforms without fontconfig is currently not supported.

This function was introduced in Qt 4.2.

See also addApplicationFont(), applicationFontFamilies(), and removeApplicationFont().

fn application_font_families<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseApplicationFontFamiliesArgs

C++ method: QFontDatabase::applicationFontFamilies

This is an overloaded function. Available variants:

Rust arguments:
1) fn application_font_families((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn application_font_families((libc::c_int, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: static QStringList QFontDatabase::applicationFontFamilies(int id)

C++ documentation:

Returns a list of font families for the given application font identified by id.

This function was introduced in Qt 4.2.

See also addApplicationFont() and addApplicationFontFromData().

fn bold(&self, family: &String, style: &String) -> bool

C++ method: bool QFontDatabase::bold(const QString& family, const QString& style) const

C++ documentation:

Returns true if the font that has family family and style style is bold; otherwise returns false.

See also italic() and weight().

fn families<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontDatabaseFamiliesArgs<'largs>

C++ method: QFontDatabase::families

This is an overloaded function. Available variants:

Rust arguments:
1) fn families(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn families(&self, (::font_database::WritingSystem, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
3) fn families(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
4) fn families(&self, (::font_database::WritingSystem, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: QStringList QFontDatabase::families(QFontDatabase::WritingSystem writingSystem = ?) const

C++ documentation:

Returns a sorted list of the available font families which support the writingSystem.

If a family exists in several foundries, the returned name for that font is in the form "family [foundry]". Examples: "Times [Adobe]", "Times [Cronyx]", "Palatino".

See also writingSystems().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontDatabaseFontArgs<'largs>

C++ method: QFontDatabase::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, (&qt_core::string::String, &qt_core::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, (&qt_core::string::String, &qt_core::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::font::Font
C++ method: QFont QFontDatabase::font(const QString& family, const QString& style, int pointSize) const

C++ documentation:

Returns a QFont object that has family family, style style and point size pointSize. If no matching font could be created, a QFont object that uses the application's default font is returned.

fn has_family(&self, family: &String) -> bool

C++ method: bool QFontDatabase::hasFamily(const QString& family) const

fn is_bitmap_scalable<'largs, Args>(&'largs self, args: Args) -> bool where Args: FontDatabaseIsBitmapScalableArgs<'largs>

C++ method: QFontDatabase::isBitmapScalable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_bitmap_scalable(&self, &qt_core::string::String) -> bool
2) fn is_bitmap_scalable(&self, (&qt_core::string::String, &qt_core::string::String)) -> bool
C++ method: bool QFontDatabase::isBitmapScalable(const QString& family, const QString& style = ?) const

C++ documentation:

Returns true if the font that has family family and style style is a scalable bitmap font; otherwise returns false. Scaling a bitmap font usually produces an unattractive hardly readable result, because the pixels of the font are scaled. If you need to scale a bitmap font it is better to scale it to one of the fixed sizes returned by smoothSizes().

See also isScalable() and isSmoothlyScalable().

fn is_fixed_pitch<'largs, Args>(&'largs self, args: Args) -> bool where Args: FontDatabaseIsFixedPitchArgs<'largs>

C++ method: QFontDatabase::isFixedPitch

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_fixed_pitch(&self, &qt_core::string::String) -> bool
2) fn is_fixed_pitch(&self, (&qt_core::string::String, &qt_core::string::String)) -> bool
C++ method: bool QFontDatabase::isFixedPitch(const QString& family, const QString& style = ?) const

C++ documentation:

Returns true if the font that has family family and style style is fixed pitch; otherwise returns false.

fn is_private_family(&self, family: &String) -> bool

C++ method: bool QFontDatabase::isPrivateFamily(const QString& family) const

C++ documentation:

Returns true if and only if the family font family is private.

This happens, for instance, on OS X and iOS, where the system UI fonts are not accessible to the user. For completeness, QFontDatabase::families() returns all font families, including the private ones. You should use this function if you are developing a font selection control in order to keep private fonts hidden.

This function was introduced in Qt 5.5.

See also families().

fn is_scalable<'largs, Args>(&'largs self, args: Args) -> bool where Args: FontDatabaseIsScalableArgs<'largs>

C++ method: QFontDatabase::isScalable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_scalable(&self, &qt_core::string::String) -> bool
2) fn is_scalable(&self, (&qt_core::string::String, &qt_core::string::String)) -> bool
C++ method: bool QFontDatabase::isScalable(const QString& family, const QString& style = ?) const

C++ documentation:

Returns true if the font that has family family and style style is scalable; otherwise returns false.

See also isBitmapScalable() and isSmoothlyScalable().

fn is_smoothly_scalable<'largs, Args>(&'largs self, args: Args) -> bool where Args: FontDatabaseIsSmoothlyScalableArgs<'largs>

C++ method: QFontDatabase::isSmoothlyScalable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_smoothly_scalable(&self, &qt_core::string::String) -> bool
2) fn is_smoothly_scalable(&self, (&qt_core::string::String, &qt_core::string::String)) -> bool
C++ method: bool QFontDatabase::isSmoothlyScalable(const QString& family, const QString& style = ?) const

C++ documentation:

Returns true if the font that has family family and style style is smoothly scalable; otherwise returns false. If this function returns true, it's safe to scale this font to any size, and the result will always look attractive.

See also isScalable() and isBitmapScalable().

fn italic(&self, family: &String, style: &String) -> bool

C++ method: bool QFontDatabase::italic(const QString& family, const QString& style) const

C++ documentation:

Returns true if the font that has family family and style style is italic; otherwise returns false.

See also weight() and bold().

fn new<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseNewArgs

C++ method: QFontDatabase::QFontDatabase

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::font_database::FontDatabase
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::font_database::FontDatabase>
C++ method: [constructor] void QFontDatabase::QFontDatabase()

fn point_sizes<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: FontDatabasePointSizesArgs<'largs>

C++ method: QFontDatabase::pointSizes

This is an overloaded function. Available variants:

Rust arguments:
1) fn point_sizes(&mut self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn point_sizes(&mut self, (&qt_core::string::String, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::list::ListCInt>
3) fn point_sizes(&mut self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::list::ListCInt
4) fn point_sizes(&mut self, (&qt_core::string::String, &qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::list::ListCInt
C++ method: QList<int> QFontDatabase::pointSizes(const QString& family, const QString& style = ?)

C++ documentation:

Returns a list of the point sizes available for the font that has family family and style styleName. The list may be empty.

See also smoothSizes() and standardSizes().

fn remove_all_application_fonts() -> bool

C++ method: static bool QFontDatabase::removeAllApplicationFonts()

C++ documentation:

Removes all application-local fonts previously added using addApplicationFont() and addApplicationFontFromData().

Returns true if unloading of the fonts succeeded; otherwise returns false.

This function was introduced in Qt 4.2.

See also removeApplicationFont(), addApplicationFont(), and addApplicationFontFromData().

fn remove_application_font(id: c_int) -> bool

C++ method: static bool QFontDatabase::removeApplicationFont(int id)

C++ documentation:

Removes the previously loaded application font identified by id. Returns true if unloading of the font succeeded; otherwise returns false.

This function was introduced in Qt 4.2.

See also removeAllApplicationFonts(), addApplicationFont(), and addApplicationFontFromData().

fn smooth_sizes<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: FontDatabaseSmoothSizesArgs<'largs>

C++ method: QFontDatabase::smoothSizes

This is an overloaded function. Available variants:

Rust arguments:
1) fn smooth_sizes(&mut self, (&qt_core::string::String, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn smooth_sizes(&mut self, (&qt_core::string::String, &qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::list::ListCInt
C++ method: QList<int> QFontDatabase::smoothSizes(const QString& family, const QString& style)

C++ documentation:

Returns the point sizes of a font that has family family and style styleName that will look attractive. The list may be empty. For non-scalable fonts and bitmap scalable fonts, this function is equivalent to pointSizes().

See also pointSizes() and standardSizes().

fn standard_sizes<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseStandardSizesArgs

C++ method: QFontDatabase::standardSizes

This is an overloaded function. Available variants:

Rust arguments:
1) fn standard_sizes(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListCInt>
2) fn standard_sizes(cpp_utils::AsStruct) -> qt_core::list::ListCInt
C++ method: static QList<int> QFontDatabase::standardSizes()

C++ documentation:

Returns a list of standard font sizes.

See also smoothSizes() and pointSizes().

fn style_string<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: FontDatabaseStyleStringArgs<'largs>

C++ method: QFontDatabase::styleString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn style_string(&mut self, (&::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn style_string(&mut self, (&::font::Font, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QFontDatabase::styleString(const QFont& font)

C++ documentation:

Returns a string that describes the style of the font. For example, "Bold Italic", "Bold", "Italic" or "Normal". An empty string may be returned.

Variant 2

Rust arguments:
1) fn style_string(&mut self, (&::font_info::FontInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn style_string(&mut self, (&::font_info::FontInfo, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QFontDatabase::styleString(const QFontInfo& fontInfo)

C++ documentation:

Returns a string that describes the style of the fontInfo. For example, "Bold Italic", "Bold", "Italic" or "Normal". An empty string may be returned.

fn styles<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontDatabaseStylesArgs<'largs>

C++ method: QFontDatabase::styles

This is an overloaded function. Available variants:

Rust arguments:
1) fn styles(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn styles(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string_list::StringList
C++ method: QStringList QFontDatabase::styles(const QString& family) const

C++ documentation:

Returns a list of the styles available for the font family family. Some example styles: "Light", "Light Italic", "Bold", "Oblique", "Demi". The list may be empty.

See also families().

fn supports_threaded_font_rendering() -> bool

C++ method: static bool QFontDatabase::supportsThreadedFontRendering()

C++ documentation:

Returns true if font rendering is supported outside the GUI thread, false otherwise. In other words, a return value of false means that all QPainter::drawText() calls outside the GUI thread will not produce readable output.

As of 5.0, always returns true.

This function was introduced in Qt 4.4.

See also Painting In Threads.

fn system_font<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseSystemFontArgs

C++ method: QFontDatabase::systemFont

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_font((::font_database::SystemFont, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font::Font>
2) fn system_font((::font_database::SystemFont, cpp_utils::AsStruct)) -> ::font::Font
C++ method: static QFont QFontDatabase::systemFont(QFontDatabase::SystemFont type)

C++ documentation:

Returns the most adequate font for a given type case for proper integration with the system's look and feel.

This function was introduced in Qt 5.2.

See also QGuiApplication::font().

fn weight(&self, family: &String, style: &String) -> c_int

C++ method: int QFontDatabase::weight(const QString& family, const QString& style) const

C++ documentation:

Returns the weight of the font that has family family and style style. If there is no such family and style combination, returns -1.

See also italic() and bold().

fn writing_system_name<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseWritingSystemNameArgs

C++ method: QFontDatabase::writingSystemName

This is an overloaded function. Available variants:

Rust arguments:
1) fn writing_system_name((::font_database::WritingSystem, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn writing_system_name((::font_database::WritingSystem, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QFontDatabase::writingSystemName(QFontDatabase::WritingSystem writingSystem)

C++ documentation:

Returns the names the writingSystem (e.g. for displaying to the user in a dialog).

fn writing_system_sample<Args>(args: Args) -> Args::ReturnType where Args: FontDatabaseWritingSystemSampleArgs

C++ method: QFontDatabase::writingSystemSample

This is an overloaded function. Available variants:

Rust arguments:
1) fn writing_system_sample((::font_database::WritingSystem, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn writing_system_sample((::font_database::WritingSystem, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QFontDatabase::writingSystemSample(QFontDatabase::WritingSystem writingSystem)

C++ documentation:

Returns a string with sample characters from writingSystem.

fn writing_systems<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontDatabaseWritingSystemsArgs<'largs>

C++ method: QFontDatabase::writingSystems

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn writing_systems(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListWritingSystem>
2) fn writing_systems(&self, cpp_utils::AsStruct) -> ::list::ListWritingSystem
C++ method: QList<QFontDatabase::WritingSystem> QFontDatabase::writingSystems() const

C++ documentation:

Returns a sorted list of the available writing systems. This is list generated from information about all installed fonts on the system.

See also families().

Variant 2

Rust arguments:
1) fn writing_systems(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListWritingSystem>
2) fn writing_systems(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> ::list::ListWritingSystem
C++ method: QList<QFontDatabase::WritingSystem> QFontDatabase::writingSystems(const QString& family) const

C++ documentation:

Returns a sorted list of the writing systems supported by a given font family.

See also families().