Struct qt_gui::font_info::FontInfo

pub struct FontInfo {
    // some fields omitted
}

C++ type: QFontInfo

C++ documentation:

The QFontInfo class provides general information about fonts.

The QFontInfo class provides the same access functions as QFont, e.g. family(), pointSize(), italic(), weight(), fixedPitch(), styleHint() etc. But whilst the QFont access functions return the values that were set, a QFontInfo object returns the values that apply to the font that will actually be used to draw the text.

For example, when the program asks for a 25pt Courier font on a machine that has a non-scalable 24pt Courier font, QFont will (normally) use the 24pt Courier for rendering. In this case, QFont::pointSize() returns 25 and QFontInfo::pointSize() returns 24.

There are three ways to create a QFontInfo object.

Calling the QFontInfo constructor with a QFont creates a font info object for a screen-compatible font, i.e. the font cannot be a printer font. If the font is changed later, the font info object is not updated.
(Note: If you use a printer font the values returned may be inaccurate. Printer fonts are not always accessible so the nearest screen font is used if a printer font is supplied.)

QWidget::fontInfo() returns the font info for a widget's font. This is equivalent to calling QFontInfo(widget->font()). If the widget's font is changed later, the font info object is not updated.

QPainter::fontInfo() returns the font info for a painter's current font. If the painter's font is changed later, the font info object is not updated.

Methods

impl FontInfo

fn bold(&self) -> bool

C++ method: bool QFontInfo::bold() const

C++ documentation:

Returns true if weight() would return a value greater than QFont::Normal; otherwise returns false.

See also weight() and QFont::bold().

fn exact_match(&self) -> bool

C++ method: bool QFontInfo::exactMatch() const

C++ documentation:

Returns true if the matched window system font is exactly the same as the one specified by the font; otherwise returns false.

See also QFont::exactMatch().

fn family<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontInfoFamilyArgs<'largs>

C++ method: QFontInfo::family

This is an overloaded function. Available variants:

Rust arguments:
1) fn family(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn family(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFontInfo::family() const

C++ documentation:

Returns the family name of the matched window system font.

See also QFont::family().

fn fixed_pitch(&self) -> bool

C++ method: bool QFontInfo::fixedPitch() const

C++ documentation:

Returns the fixed pitch value of the matched window system font.

See also QFont::fixedPitch().

fn italic(&self) -> bool

C++ method: bool QFontInfo::italic() const

C++ documentation:

Returns the italic value of the matched window system font.

See also QFont::italic().

fn new<Args>(args: Args) -> Args::ReturnType where Args: FontInfoNewArgs

C++ method: QFontInfo::QFontInfo

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::font::Font, cpp_utils::AsStruct)) -> ::font_info::FontInfo
2) fn new((&::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font_info::FontInfo>
C++ method: [constructor] void QFontInfo::QFontInfo(const QFont& arg1)

Variant 2

Rust arguments:
1) fn new((&::font_info::FontInfo, cpp_utils::AsStruct)) -> ::font_info::FontInfo
2) fn new((&::font_info::FontInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<::font_info::FontInfo>
C++ method: [constructor] void QFontInfo::QFontInfo(const QFontInfo& arg1)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 FontInfo) -> &'l0 mut FontInfo

C++ method: QFontInfo& QFontInfo::operator=(const QFontInfo& arg1)

C++ documentation:

Assigns the font info in fi.

fn overline(&self) -> bool

C++ method: bool QFontInfo::overline() const

fn pixel_size(&self) -> c_int

C++ method: int QFontInfo::pixelSize() const

C++ documentation:

Returns the pixel size of the matched window system font.

See also QFont::pointSize().

fn point_size(&self) -> c_int

C++ method: int QFontInfo::pointSize() const

C++ documentation:

Returns the point size of the matched window system font.

See also pointSizeF() and QFont::pointSize().

fn point_size_f(&self) -> c_double

C++ method: double QFontInfo::pointSizeF() const

C++ documentation:

Returns the point size of the matched window system font.

See also QFont::pointSizeF().

fn raw_mode(&self) -> bool

C++ method: bool QFontInfo::rawMode() const

C++ documentation:

Returns true if the font is a raw mode font; otherwise returns false.

If it is a raw mode font, all other functions in QFontInfo will return the same values set in the QFont, regardless of the font actually used.

See also QFont::rawMode().

fn strike_out(&self) -> bool

C++ method: bool QFontInfo::strikeOut() const

fn style(&self) -> Style

C++ method: QFont::Style QFontInfo::style() const

C++ documentation:

Returns the style value of the matched window system font.

See also QFont::style().

fn style_hint(&self) -> StyleHint

C++ method: QFont::StyleHint QFontInfo::styleHint() const

C++ documentation:

Returns the style of the matched window system font.

Currently only returns the style hint set in QFont.

See also QFont::styleHint() and QFont::StyleHint.

fn style_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FontInfoStyleNameArgs<'largs>

C++ method: QFontInfo::styleName

This is an overloaded function. Available variants:

Rust arguments:
1) fn style_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn style_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QFontInfo::styleName() const

C++ documentation:

Returns the style name of the matched window system font on systems that support it.

This function was introduced in Qt 4.8.

See also QFont::styleName().

fn swap(&mut self, other: &mut FontInfo)

C++ method: void QFontInfo::swap(QFontInfo& other)

C++ documentation:

Swaps this font info instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn underline(&self) -> bool

C++ method: bool QFontInfo::underline() const

fn weight(&self) -> c_int

C++ method: int QFontInfo::weight() const

C++ documentation:

Returns the weight of the matched window system font.

See also QFont::weight() and bold().