Module qt_gui::generic_plugin::overloading

Traits

GenericPluginConnectArgs
This trait represents a set of arguments accepted by GenericPlugin::connect method.

GenericPluginConnectStaticArgs
This trait represents a set of arguments accepted by GenericPlugin::connect_static method.

GenericPluginDestroyedArgs
This trait represents a set of arguments accepted by GenericPlugin::destroyed method.

GenericPluginDisconnectArgs
This trait represents a set of arguments accepted by GenericPlugin::disconnect method.

GenericPluginDisconnectStaticArgs
This trait represents a set of arguments accepted by GenericPlugin::disconnect_static method.

GenericPluginDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GenericPlugin::dynamic_property_names method.

GenericPluginObjectNameArgs
This trait represents a set of arguments accepted by GenericPlugin::object_name method.

GenericPluginPropertyArgs
This trait represents a set of arguments accepted by GenericPlugin::property method.

GenericPluginStartTimerArgs
This trait represents a set of arguments accepted by GenericPlugin::start_timer method.

GenericPluginTrArgs
This trait represents a set of arguments accepted by GenericPlugin::tr method.

GenericPluginTrUtf8Args
This trait represents a set of arguments accepted by GenericPlugin::tr_utf8 method.