Enum qt_gui::glyph_run::GlyphRunFlag

pub enum GlyphRunFlag {
    Overline,
    Underline,
    StrikeOut,
    RightToLeft,
    SplitLigature,
}

C++ type: QGlyphRun::GlyphRunFlag

C++ documentation:

This enum describes flags that alter the way the run of glyphs might be presented or behave in a visual layout. The layout which generates the glyph runs can set these flags based on relevant internal data, to retain information needed to present the text as intended by the user of the layout.

This enum was introduced or modified in Qt 5.0.

The GlyphRunFlags type is a typedef for QFlags<GlyphRunFlag>. It stores an OR combination of GlyphRunFlag values.

Variants
Overline

Indicates that the glyphs should be visualized together with an overline. (C++ enum variant: Overline = 1)

Underline

Indicates that the glyphs should be visualized together with an underline. (C++ enum variant: Underline = 2)

StrikeOut

Indicates that the glyphs should be struck out visually. (C++ enum variant: StrikeOut = 4)

RightToLeft

Indicates that the glyphs are ordered right to left. This can affect the positioning of other screen elements that are relative to the glyph run, such as an inline text object. (C++ enum variant: RightToLeft = 8)

SplitLigature

Indicates that the glyph run splits a ligature glyph. This means that a ligature glyph is included in the run, but the characters represented by it corresponds only to part of that ligature. The glyph run's boundingRect() function can in this case be used to retrieve the area covered by glyphs that correspond to the characters represented by the glyph run. When visualizing the glyphs, care needs to be taken to clip to this bounding rect to ensure that only the corresponding part of the ligature is painted. In particular, this can be the case when retrieving a glyph run from a QTextLayout for a specific character range, e.g. when retrieving the selected area of a QTextLayout. (C++ enum variant: SplitLigature = 16)