Module qt_gui::glyph_run::overloading

Traits

GlyphRunBoundingRectArgs
This trait represents a set of arguments accepted by GlyphRun::bounding_rect method.

GlyphRunGlyphIndexesArgs
This trait represents a set of arguments accepted by GlyphRun::glyph_indexes method.

GlyphRunNewArgs
This trait represents a set of arguments accepted by GlyphRun::new method.

GlyphRunPositionsArgs
This trait represents a set of arguments accepted by GlyphRun::positions method.

GlyphRunRawFontArgs
This trait represents a set of arguments accepted by GlyphRun::raw_font method.

GlyphRunSetFlagArgs
This trait represents a set of arguments accepted by GlyphRun::set_flag method.