Struct qt_gui::glyph_run::GlyphRun

pub struct GlyphRun {
    // some fields omitted
}

C++ type: QGlyphRun

C++ documentation:

The QGlyphRun class provides direct access to the internal glyphs in a font.

When Qt displays a string of text encoded in Unicode, it will first convert the Unicode points into a list of glyph indexes and a list of positions based on one or more fonts. The Unicode representation of the text and the QFont object will in this case serve as a convenient abstraction that hides the details of what actually takes place when displaying the text on-screen. For instance, by the time the text actually reaches the screen, it may be represented by a set of fonts in addition to the one specified by the user, e.g. in case the originally selected font did not support all the writing systems contained in the text.

Under certain circumstances, it can be useful as an application developer to have more low-level control over which glyphs in a specific font are drawn to the screen. This could for instance be the case in applications that use an external font engine and text shaper together with Qt. QGlyphRun provides an interface to the raw data needed to get text on the screen. It contains a list of glyph indexes, a position for each glyph and a font.

It is the user's responsibility to ensure that the selected font actually contains the provided glyph indexes.

QTextLayout::glyphRuns() or QTextFragment::glyphRuns() can be used to convert unicode encoded text into a list of QGlyphRun objects, and QPainter::drawGlyphRun() can be used to draw the glyphs.

Note: Please note that QRawFont is considered local to the thread in which it is constructed. This in turn means that a new QRawFont will have to be created and set on the QGlyphRun if it is moved to a different thread. If the QGlyphRun contains a reference to a QRawFont from a different thread than the current, it will not be possible to draw the glyphs using a QPainter, as the QRawFont is considered invalid and inaccessible in this case.

Methods

impl GlyphRun

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GlyphRunBoundingRectArgs<'largs>

C++ method: QGlyphRun::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn bounding_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QGlyphRun::boundingRect() const

C++ documentation:

Returns the smallest rectangle that contains all glyphs in this QGlyphRun. If a bounding rect has been set using setBoundingRect(), then this will be returned. Otherwise the bounding rect will be calculated based on the font metrics of the glyphs in the glyph run.

This function was introduced in Qt 5.0.

See also setBoundingRect().

fn clear(&mut self)

C++ method: void QGlyphRun::clear()

C++ documentation:

Clears all data in the QGlyphRun object.

fn flags(&self) -> Flags<GlyphRunFlag>

C++ method: QFlags<QGlyphRun::GlyphRunFlag> QGlyphRun::flags() const

C++ documentation:

Returns the flags set for this QGlyphRun.

This function was introduced in Qt 5.0.

See also setFlags(), setFlag(), and setFlag().

fn glyph_indexes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GlyphRunGlyphIndexesArgs<'largs>

C++ method: QGlyphRun::glyphIndexes

This is an overloaded function. Available variants:

Rust arguments:
1) fn glyph_indexes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::vector::VectorCUint>
2) fn glyph_indexes(&self, cpp_utils::AsStruct) -> qt_core::vector::VectorCUint
C++ method: QVector<unsigned int> QGlyphRun::glyphIndexes() const

C++ documentation:

Returns the glyph indexes for this QGlyphRun object.

See also setGlyphIndexes() and setPositions().

fn is_empty(&self) -> bool

C++ method: bool QGlyphRun::isEmpty() const

C++ documentation:

Returns true if the QGlyphRun does not contain any glyphs.

This function was introduced in Qt 5.0.

fn is_right_to_left(&self) -> bool

C++ method: bool QGlyphRun::isRightToLeft() const

C++ documentation:

Returns true if this QGlyphRun contains glyphs that are painted from the right to the left.

This function was introduced in Qt 5.0.

See also setRightToLeft() and flags().

fn new<Args>(args: Args) -> Args::ReturnType where Args: GlyphRunNewArgs

C++ method: QGlyphRun::QGlyphRun

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::glyph_run::GlyphRun
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::glyph_run::GlyphRun>
C++ method: [constructor] void QGlyphRun::QGlyphRun()

Variant 2

Rust arguments:
1) fn new((&::glyph_run::GlyphRun, cpp_utils::AsStruct)) -> ::glyph_run::GlyphRun
2) fn new((&::glyph_run::GlyphRun, cpp_utils::AsBox)) -> cpp_utils::CppBox<::glyph_run::GlyphRun>
C++ method: [constructor] void QGlyphRun::QGlyphRun(const QGlyphRun& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 GlyphRun) -> &'l0 mut GlyphRun

C++ method: QGlyphRun& QGlyphRun::operator=(const QGlyphRun& other)

C++ documentation:

Assigns other to this QGlyphRun object.

fn op_eq(&self, other: &GlyphRun) -> bool

C++ method: bool QGlyphRun::operator==(const QGlyphRun& other) const

C++ documentation:

Compares other to this QGlyphRun object. Returns true if the list of glyph indexes, the list of positions and the font are all equal, otherwise returns false.

fn op_neq(&self, other: &GlyphRun) -> bool

C++ method: bool QGlyphRun::operator!=(const QGlyphRun& other) const

C++ documentation:

Compares other to this QGlyphRun object. Returns true if any of the list of glyph indexes, the list of positions or the font are different, otherwise returns false.

fn overline(&self) -> bool

C++ method: bool QGlyphRun::overline() const

C++ documentation:

Returns true if this QGlyphRun should be painted with an overline decoration.

See also setOverline() and flags().

fn positions<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GlyphRunPositionsArgs<'largs>

C++ method: QGlyphRun::positions

This is an overloaded function. Available variants:

Rust arguments:
1) fn positions(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn positions(&self, cpp_utils::AsStruct) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QGlyphRun::positions() const

C++ documentation:

Returns the position of the edge of the baseline for each glyph in this set of glyph indexes.

See also setPositions().

fn raw_font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: GlyphRunRawFontArgs<'largs>

C++ method: QGlyphRun::rawFont

This is an overloaded function. Available variants:

Rust arguments:
1) fn raw_font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::raw_font::RawFont>
2) fn raw_font(&self, cpp_utils::AsStruct) -> ::raw_font::RawFont
C++ method: QRawFont QGlyphRun::rawFont() const

C++ documentation:

Returns the font selected for this QGlyphRun object.

See also setRawFont().

fn set_bounding_rect(&mut self, bounding_rect: &RectF)

C++ method: void QGlyphRun::setBoundingRect(const QRectF& boundingRect)

C++ documentation:

Sets the bounding rect of the glyphs in this QGlyphRun to be boundingRect. This rectangle will be returned by boundingRect() unless it is empty, in which case the bounding rectangle of the glyphs in the glyph run will be returned instead.

Note: Unless you are implementing text shaping, you should not have to use this function. It is used specifically when the QGlyphRun should represent an area which is smaller than the area of the glyphs it contains. This could happen e.g. if the glyph run is retrieved by calling QTextLayout::glyphRuns() and the specified range only includes part of a ligature (where two or more characters are combined to a single glyph.) When this is the case, the bounding rect should only include the appropriate part of the ligature glyph, based on a calculation of the average width of the characters in the ligature.

In order to support such a case (an example is selections which should be drawn with a different color than the main text color), it is necessary to clip the painting mechanism to the rectangle returned from boundingRect() to avoid drawing the entire ligature glyph.

This function was introduced in Qt 5.0.

See also boundingRect().

fn set_flag<'largs, Args>(&'largs mut self, args: Args) where Args: GlyphRunSetFlagArgs<'largs>

C++ method: QGlyphRun::setFlag

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_flag(&mut self, ::glyph_run::GlyphRunFlag) -> ()
2) fn set_flag(&mut self, (::glyph_run::GlyphRunFlag, bool)) -> ()
C++ method: void QGlyphRun::setFlag(QGlyphRun::GlyphRunFlag flag, bool enabled = ?)

C++ documentation:

If enabled is true, then flag is enabled; otherwise, it is disabled.

This function was introduced in Qt 5.0.

See also flags() and setFlags().

fn set_flags(&mut self, flags: Flags<GlyphRunFlag>)

C++ method: void QGlyphRun::setFlags(QFlags<QGlyphRun::GlyphRunFlag> flags)

C++ documentation:

Sets the flags of this QGlyphRun to flags.

This function was introduced in Qt 5.0.

See also setFlag() and flags().

fn set_glyph_indexes(&mut self, glyph_indexes: &VectorCUint)

C++ method: void QGlyphRun::setGlyphIndexes(const QVector<unsigned int>& glyphIndexes)

C++ documentation:

Set the glyph indexes for this QGlyphRun object to glyphIndexes. The glyph indexes must be valid for the selected font.

See also glyphIndexes().

fn set_overline(&mut self, overline: bool)

C++ method: void QGlyphRun::setOverline(bool overline)

C++ documentation:

Indicates that this QGlyphRun should be painted with an overline decoration if overline is true. Otherwise the QGlyphRun should be painted with no overline decoration.

See also overline(), setFlag(), and setFlags().

fn set_positions(&mut self, positions: &VectorPointFRef)

C++ method: void QGlyphRun::setPositions(const QVector<QPointF>& positions)

C++ documentation:

Sets the positions of the edge of the baseline for each glyph in this set of glyph indexes to positions.

See also positions().

fn set_raw_data(&mut self, glyph_index_array: *const c_uint, glyph_position_array: *const PointF, size: c_int)

C++ method: void QGlyphRun::setRawData(const unsigned int* glyphIndexArray, const QPointF* glyphPositionArray, int size)

C++ documentation:

Sets the glyph indexes and positions of this QGlyphRun to use the first size elements in the arrays glyphIndexArray and glyphPositionArray. The data is not copied. The caller must guarantee that the arrays are not deleted as long as this QGlyphRun and any copies of it exists.

See also setGlyphIndexes() and setPositions().

fn set_raw_font(&mut self, raw_font: &RawFont)

C++ method: void QGlyphRun::setRawFont(const QRawFont& rawFont)

C++ documentation:

Sets the font in which to look up the glyph indexes to the rawFont specified.

See also rawFont() and setGlyphIndexes().

fn set_right_to_left(&mut self, on: bool)

C++ method: void QGlyphRun::setRightToLeft(bool on)

C++ documentation:

Indicates that this QGlyphRun contains glyphs that should be ordered from the right to left if rightToLeft is true. Otherwise the order of the glyphs is assumed to be left to right.

This function was introduced in Qt 5.0.

See also isRightToLeft(), setFlag(), and setFlags().

fn set_strike_out(&mut self, strike_out: bool)

C++ method: void QGlyphRun::setStrikeOut(bool strikeOut)

C++ documentation:

Indicates that this QGlyphRun should be painted with an strike out decoration if strikeOut is true. Otherwise the QGlyphRun should be painted with no strike out decoration.

See also strikeOut(), setFlag(), and setFlags().

fn set_underline(&mut self, underline: bool)

C++ method: void QGlyphRun::setUnderline(bool underline)

C++ documentation:

Indicates that this QGlyphRun should be painted with an underline decoration if underline is true. Otherwise the QGlyphRun should be painted with no underline decoration.

See also underline(), setFlag(), and setFlags().

fn strike_out(&self) -> bool

C++ method: bool QGlyphRun::strikeOut() const

C++ documentation:

Returns true if this QGlyphRun should be painted with a strike out decoration.

See also setStrikeOut() and flags().

fn swap(&mut self, other: &mut GlyphRun)

C++ method: void QGlyphRun::swap(QGlyphRun& other)

C++ documentation:

Swaps this glyph run instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn underline(&self) -> bool

C++ method: bool QGlyphRun::underline() const

C++ documentation:

Returns true if this QGlyphRun should be painted with an underline decoration.

See also setUnderline() and flags().