Module qt_gui::gui_application::overloading

Traits

GuiApplicationAllWindowsArgs
This trait represents a set of arguments accepted by GuiApplication::all_windows method.

GuiApplicationApplicationDirPathArgs
This trait represents a set of arguments accepted by GuiApplication::application_dir_path method.

GuiApplicationApplicationDisplayNameArgs
This trait represents a set of arguments accepted by GuiApplication::application_display_name method.

GuiApplicationApplicationFilePathArgs
This trait represents a set of arguments accepted by GuiApplication::application_file_path method.

GuiApplicationApplicationNameArgs
This trait represents a set of arguments accepted by GuiApplication::application_name method.

GuiApplicationApplicationVersionArgs
This trait represents a set of arguments accepted by GuiApplication::application_version method.

GuiApplicationArgumentsArgs
This trait represents a set of arguments accepted by GuiApplication::arguments method.

GuiApplicationConnectArgs
This trait represents a set of arguments accepted by GuiApplication::connect method.

GuiApplicationConnectStaticArgs
This trait represents a set of arguments accepted by GuiApplication::connect_static method.

GuiApplicationDestroyedArgs
This trait represents a set of arguments accepted by GuiApplication::destroyed method.

GuiApplicationDisconnectArgs
This trait represents a set of arguments accepted by GuiApplication::disconnect method.

GuiApplicationDisconnectStaticArgs
This trait represents a set of arguments accepted by GuiApplication::disconnect_static method.

GuiApplicationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by GuiApplication::dynamic_property_names method.

GuiApplicationExitArgs
This trait represents a set of arguments accepted by GuiApplication::exit method.

GuiApplicationFontArgs
This trait represents a set of arguments accepted by GuiApplication::font method.

GuiApplicationLibraryPathsArgs
This trait represents a set of arguments accepted by GuiApplication::library_paths method.

GuiApplicationNewArgs
This trait represents a set of arguments accepted by GuiApplication::new method.

GuiApplicationObjectNameArgs
This trait represents a set of arguments accepted by GuiApplication::object_name method.

GuiApplicationOrganizationDomainArgs
This trait represents a set of arguments accepted by GuiApplication::organization_domain method.

GuiApplicationOrganizationNameArgs
This trait represents a set of arguments accepted by GuiApplication::organization_name method.

GuiApplicationPaletteArgs
This trait represents a set of arguments accepted by GuiApplication::palette method.

GuiApplicationPlatformNameArgs
This trait represents a set of arguments accepted by GuiApplication::platform_name method.

GuiApplicationPostEventArgs
This trait represents a set of arguments accepted by GuiApplication::post_event method.

GuiApplicationProcessEventsArgs
This trait represents a set of arguments accepted by GuiApplication::process_events method.

GuiApplicationPropertyArgs
This trait represents a set of arguments accepted by GuiApplication::property method.

GuiApplicationRemovePostedEventsArgs
This trait represents a set of arguments accepted by GuiApplication::remove_posted_events method.

GuiApplicationScreensArgs
This trait represents a set of arguments accepted by GuiApplication::screens method.

GuiApplicationSendPostedEventsArgs
This trait represents a set of arguments accepted by GuiApplication::send_posted_events method.

GuiApplicationSessionIdArgs
This trait represents a set of arguments accepted by GuiApplication::session_id method.

GuiApplicationSessionKeyArgs
This trait represents a set of arguments accepted by GuiApplication::session_key method.

GuiApplicationSetAttributeArgs
This trait represents a set of arguments accepted by GuiApplication::set_attribute method.

GuiApplicationStartTimerArgs
This trait represents a set of arguments accepted by GuiApplication::start_timer method.

GuiApplicationTopLevelWindowsArgs
This trait represents a set of arguments accepted by GuiApplication::top_level_windows method.

GuiApplicationTrArgs
This trait represents a set of arguments accepted by GuiApplication::tr method.

GuiApplicationTrUtf8Args
This trait represents a set of arguments accepted by GuiApplication::tr_utf8 method.

GuiApplicationTranslateArgs
This trait represents a set of arguments accepted by GuiApplication::translate method.

GuiApplicationWindowIconArgs
This trait represents a set of arguments accepted by GuiApplication::window_icon method.