Module qt_gui::icon::overloading

Traits

IconActualSizeArgs
This trait represents a set of arguments accepted by Icon::actual_size method.

IconAddFileArgs
This trait represents a set of arguments accepted by Icon::add_file method.

IconAddPixmapArgs
This trait represents a set of arguments accepted by Icon::add_pixmap method.

IconAsQVariantArgs
This trait represents a set of arguments accepted by Icon::as_q_variant method.

IconAvailableSizesArgs
This trait represents a set of arguments accepted by Icon::available_sizes method.

IconFromThemeArgs
This trait represents a set of arguments accepted by Icon::from_theme method.

IconNameArgs
This trait represents a set of arguments accepted by Icon::name method.

IconNewArgs
This trait represents a set of arguments accepted by Icon::new method.

IconPaintArgs
This trait represents a set of arguments accepted by Icon::paint method.

IconPixmapArgs
This trait represents a set of arguments accepted by Icon::pixmap method.

IconThemeNameArgs
This trait represents a set of arguments accepted by Icon::theme_name method.

IconThemeSearchPathsArgs
This trait represents a set of arguments accepted by Icon::theme_search_paths method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.