Struct qt_gui::icon::Icon

pub struct Icon {
    // some fields omitted
}

C++ type: QIcon

C++ documentation:

The QIcon class provides scalable icons in different modes and states.

A QIcon can generate smaller, larger, active, and disabled pixmaps from the set of pixmaps it is given. Such pixmaps are used by Qt widgets to show an icon representing a particular action.

The simplest use of QIcon is to create one from a QPixmap file or resource, and then use it, allowing Qt to work out all the required icon styles and sizes. For example:

QToolButton *button = new QToolButton;
button->setIcon(QIcon("open.xpm"));

To undo a QIcon, simply set a null icon in its place:

button->setIcon(QIcon());

Use the QImageReader::supportedImageFormats() and QImageWriter::supportedImageFormats() functions to retrieve a complete list of the supported file formats.

When you retrieve a pixmap using pixmap(QSize, Mode, State), and no pixmap for this given size, mode and state has been added with addFile() or addPixmap(), then QIcon will generate one on the fly. This pixmap generation happens in a QIconEngineV2. The default engine scales pixmaps down if required, but never up, and it uses the current style to calculate a disabled appearance. By using custom icon engines, you can customize every aspect of generated icons. With QIconEnginePluginV2 it is possible to register different icon engines for different file suffixes, making it possible for third parties to provide additional icon engines to those included with Qt.

Note: Since Qt 4.2, an icon engine that supports SVG is included.

Methods

impl Icon

fn actual_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconActualSizeArgs<'largs>

C++ method: QIcon::actualSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
3) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
4) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, cpp_utils::AsStruct)) -> qt_core::size::Size
5) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, cpp_utils::AsStruct)) -> qt_core::size::Size
6) fn actual_size(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: QSize QIcon::actualSize(QWindow* window, const QSize& size, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns the actual size of the icon for the requested window size, mode, and state.

The pixmap can be smaller than the requested size. The returned size is in device-independent pixels (This is relevant for high-dpi pixmaps.)

This function was introduced in Qt 5.1.

See also actualSize(), pixmap(), and paint().

Variant 2

Rust arguments:
1) fn actual_size(&self, (&qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn actual_size(&self, (&qt_core::size::Size, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
3) fn actual_size(&self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
4) fn actual_size(&self, (&qt_core::size::Size, cpp_utils::AsStruct)) -> qt_core::size::Size
5) fn actual_size(&self, (&qt_core::size::Size, ::icon::Mode, cpp_utils::AsStruct)) -> qt_core::size::Size
6) fn actual_size(&self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: QSize QIcon::actualSize(const QSize& size, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns the actual size of the icon for the requested size, mode, and state. The result might be smaller than requested, but never larger. The returned size is in device-independent pixels (This is relevant for high-dpi pixmaps.)

See also pixmap() and paint().

fn add_file<'largs, Args>(&'largs mut self, args: Args) where Args: IconAddFileArgs<'largs>

C++ method: QIcon::addFile

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_file(&mut self, &qt_core::string::String) -> ()
2) fn add_file(&mut self, (&qt_core::string::String, &qt_core::size::Size)) -> ()
3) fn add_file(&mut self, (&qt_core::string::String, &qt_core::size::Size, ::icon::Mode)) -> ()
4) fn add_file(&mut self, (&qt_core::string::String, &qt_core::size::Size, ::icon::Mode, ::icon::State)) -> ()
C++ method: void QIcon::addFile(const QString& fileName, const QSize& size = ?, QIcon::Mode mode = ?, QIcon::State state = ?)

C++ documentation:

Adds an image from the file with the given fileName to the icon, as a specialization for size, mode and state. The file will be loaded on demand. Note: custom icon engines are free to ignore additionally added pixmaps.

If fileName contains a relative path (e.g. the filename only) the relevant file must be found relative to the runtime working directory.

The file name can be either refer to an actual file on disk or to one of the application's embedded resources. See the Resource System overview for details on how to embed images and other resource files in the application's executable.

Use the QImageReader::supportedImageFormats() and QImageWriter::supportedImageFormats() functions to retrieve a complete list of the supported file formats.

If a high resolution version of the image exists (identified by the suffix @2x on the base name), it is automatically loaded and added with the device pixel ratio set to a value of 2. This can be disabled by setting the environment variable QT_HIGHDPI_DISABLE_2X_IMAGE_LOADING (see QImageReader).

Note: When you add a non-empty filename to a QIcon, the icon becomes non-null, even if the file doesn't exist or points to a corrupt file.

See also addPixmap() and QPixmap::devicePixelRatio().

fn add_pixmap<'largs, Args>(&'largs mut self, args: Args) where Args: IconAddPixmapArgs<'largs>

C++ method: QIcon::addPixmap

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_pixmap(&mut self, &::pixmap::Pixmap) -> ()
2) fn add_pixmap(&mut self, (&::pixmap::Pixmap, ::icon::Mode)) -> ()
3) fn add_pixmap(&mut self, (&::pixmap::Pixmap, ::icon::Mode, ::icon::State)) -> ()
C++ method: void QIcon::addPixmap(const QPixmap& pixmap, QIcon::Mode mode = ?, QIcon::State state = ?)

C++ documentation:

Adds pixmap to the icon, as a specialization for mode and state.

Custom icon engines are free to ignore additionally added pixmaps.

See also addFile().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconAsQVariantArgs<'largs>

C++ method: QIcon::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QIcon::operator QVariant() const

C++ documentation:

Returns the icon as a QVariant.

fn available_sizes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconAvailableSizesArgs<'largs>

C++ method: QIcon::availableSizes

This is an overloaded function. Available variants:

Rust arguments:
1) fn available_sizes(&self, (::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListSizeRef>
2) fn available_sizes(&self, (::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListSizeRef>
3) fn available_sizes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListSizeRef>
4) fn available_sizes(&self, (::icon::Mode, cpp_utils::AsStruct)) -> ::list::ListSizeRef
5) fn available_sizes(&self, (::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::list::ListSizeRef
6) fn available_sizes(&self, cpp_utils::AsStruct) -> ::list::ListSizeRef
C++ method: QList<QSize> QIcon::availableSizes(QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns a list of available icon sizes for the specified mode and state.

This function was introduced in Qt 4.5.

fn cache_key(&self) -> i64

C++ method: qint64 QIcon::cacheKey() const

C++ documentation:

Returns a number that identifies the contents of this QIcon object. Distinct QIcon objects can have the same key if they refer to the same contents.

The cacheKey() will change when the icon is altered via addPixmap() or addFile().

Cache keys are mostly useful in conjunction with caching.

This function was introduced in Qt 4.3.

See also QPixmap::cacheKey().

fn detach(&mut self)

C++ method: void QIcon::detach()

fn from_theme<Args>(args: Args) -> Args::ReturnType where Args: IconFromThemeArgs

C++ method: QIcon::fromTheme

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_theme((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
2) fn from_theme((&qt_core::string::String, &::icon::Icon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
3) fn from_theme((&qt_core::string::String, cpp_utils::AsStruct)) -> ::icon::Icon
4) fn from_theme((&qt_core::string::String, &::icon::Icon, cpp_utils::AsStruct)) -> ::icon::Icon
C++ method: static QIcon QIcon::fromTheme(const QString& name, const QIcon& fallback = ?)

C++ documentation:

Returns the QIcon corresponding to name in the current icon theme. If no such icon is found in the current theme fallback is returned instead.

The latest version of the freedesktop icon specification and naming specification can be obtained here:

http://standards.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html

http://standards.freedesktop.org/icon-naming-spec/icon-naming-spec-latest.html

To fetch an icon from the current icon theme:

    QIcon undoicon = QIcon::fromTheme("edit-undo");

Or if you want to provide a guaranteed fallback for platforms that do not support theme icons, you can use the second argument:

    QIcon undoicon = QIcon::fromTheme("edit-undo", QIcon(":/undo.png"));

Note: By default, only X11 will support themed icons. In order to use themed icons on Mac and Windows, you will have to bundle a compliant theme in one of your themeSearchPaths() and set the appropriate themeName().

This function was introduced in Qt 4.6.

See also themeName(), setThemeName(), and themeSearchPaths().

fn has_theme_icon(name: &String) -> bool

C++ method: static bool QIcon::hasThemeIcon(const QString& name)

C++ documentation:

Returns true if there is an icon available for name in the current icon theme, otherwise returns false.

This function was introduced in Qt 4.6.

See also themeSearchPaths(), fromTheme(), and setThemeName().

fn is_detached(&self) -> bool

C++ method: bool QIcon::isDetached() const

fn is_null(&self) -> bool

C++ method: bool QIcon::isNull() const

C++ documentation:

Returns true if the icon is empty; otherwise returns false.

An icon is empty if it has neither a pixmap nor a filename.

Note: Even a non-null icon might not be able to create valid pixmaps, eg. if the file does not exist or cannot be read.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconNameArgs<'largs>

C++ method: QIcon::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QIcon::name() const

C++ documentation:

Returns the name used to create the icon, if available.

Depending on the way the icon was created, it may have an associated name. This is the case for icons created with fromTheme() or icons using a QIconEngine which supports the QIconEngineV2::IconNameHook.

This function was introduced in Qt 4.7.

See also fromTheme() and QIconEngine.

fn new<Args>(args: Args) -> Args::ReturnType where Args: IconNewArgs

C++ method: QIcon::QIcon

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::icon::Icon
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::icon::Icon>
C++ method: [constructor] void QIcon::QIcon()

Variant 2

Rust arguments:
1) fn new((*mut ::icon_engine::IconEngine, cpp_utils::AsStruct)) -> ::icon::Icon
2) fn new((*mut ::icon_engine::IconEngine, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
C++ method: [constructor] void QIcon::QIcon(QIconEngine* engine)

Variant 3

Rust arguments:
1) fn new((&::icon::Icon, cpp_utils::AsStruct)) -> ::icon::Icon
2) fn new((&::icon::Icon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
C++ method: [constructor] void QIcon::QIcon(const QIcon& other)

Variant 4

Rust arguments:
1) fn new((&::pixmap::Pixmap, cpp_utils::AsStruct)) -> ::icon::Icon
2) fn new((&::pixmap::Pixmap, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
C++ method: [constructor] void QIcon::QIcon(const QPixmap& pixmap)

Variant 5

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::icon::Icon
2) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::icon::Icon>
C++ method: [constructor] void QIcon::QIcon(const QString& fileName)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Icon) -> &'l0 mut Icon

C++ method: QIcon& QIcon::operator=(const QIcon& other)

C++ documentation:

Assigns the other icon to this icon and returns a reference to this icon.

fn paint<'largs, Args>(&'largs self, args: Args) where Args: IconPaintArgs<'largs>

C++ method: QIcon::paint

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn paint(&self, (*mut ::painter::Painter, &qt_core::rect::Rect)) -> ()
2) fn paint(&self, (*mut ::painter::Painter, &qt_core::rect::Rect, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> ()
3) fn paint(&self, (*mut ::painter::Painter, &qt_core::rect::Rect, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, ::icon::Mode)) -> ()
4) fn paint(&self, (*mut ::painter::Painter, &qt_core::rect::Rect, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, ::icon::Mode, ::icon::State)) -> ()
C++ method: void QIcon::paint(QPainter* painter, const QRect& rect, QFlags<Qt::AlignmentFlag> alignment = ?, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Uses the painter to paint the icon with specified alignment, required mode, and state into the rectangle rect.

See also actualSize() and pixmap().

Variant 2

Rust arguments:
1) fn paint(&self, (*mut ::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> ()
2) fn paint(&self, (*mut ::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>)) -> ()
3) fn paint(&self, (*mut ::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, ::icon::Mode)) -> ()
4) fn paint(&self, (*mut ::painter::Painter, libc::c_int, libc::c_int, libc::c_int, libc::c_int, qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, ::icon::Mode, ::icon::State)) -> ()
C++ method: void QIcon::paint(QPainter* painter, int x, int y, int w, int h, QFlags<Qt::AlignmentFlag> alignment = ?, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

This is an overloaded function.

Paints the icon into the rectangle QRect(x, y, w, h).

fn pixmap<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconPixmapArgs<'largs>

C++ method: QIcon::pixmap

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
3) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
4) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
5) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
6) fn pixmap(&self, (*mut ::window::Window, &qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
C++ method: QPixmap QIcon::pixmap(QWindow* window, const QSize& size, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns a pixmap with the requested window size, mode, and state, generating one if necessary.

The pixmap can be smaller than the requested size. If window is on a high-dpi display the pixmap can be larger. In that case it will have a devicePixelRatio larger than 1.

This function was introduced in Qt 5.1.

See also actualSize() and paint().

Variant 2

Rust arguments:
1) fn pixmap(&self, (&qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&self, (&qt_core::size::Size, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
3) fn pixmap(&self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
4) fn pixmap(&self, (&qt_core::size::Size, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
5) fn pixmap(&self, (&qt_core::size::Size, ::icon::Mode, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
6) fn pixmap(&self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
C++ method: QPixmap QIcon::pixmap(const QSize& size, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns a pixmap with the requested size, mode, and state, generating one if necessary. The pixmap might be smaller than requested, but never larger.

Setting the Qt::AA_UseHighDpiPixmaps application attribute enables this function to return pixmaps that are larger than the requested size. Such images will have a devicePixelRatio larger than 1.

See also actualSize() and paint().

Variant 3

Rust arguments:
1) fn pixmap(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&self, (libc::c_int, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
3) fn pixmap(&self, (libc::c_int, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
4) fn pixmap(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
5) fn pixmap(&self, (libc::c_int, ::icon::Mode, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
6) fn pixmap(&self, (libc::c_int, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
C++ method: QPixmap QIcon::pixmap(int extent, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

This is an overloaded function.

Returns a pixmap of size QSize(extent, extent). The pixmap might be smaller than requested, but never larger.

Setting the Qt::AA_UseHighDpiPixmaps application attribute enables this function to return pixmaps that are larger than the requested size. Such images will have a devicePixelRatio larger than 1.

Variant 4

Rust arguments:
1) fn pixmap(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&self, (libc::c_int, libc::c_int, ::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
3) fn pixmap(&self, (libc::c_int, libc::c_int, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
4) fn pixmap(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
5) fn pixmap(&self, (libc::c_int, libc::c_int, ::icon::Mode, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
6) fn pixmap(&self, (libc::c_int, libc::c_int, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
C++ method: QPixmap QIcon::pixmap(int w, int h, QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

This is an overloaded function.

Returns a pixmap of size QSize(w, h). The pixmap might be smaller than requested, but never larger.

Setting the Qt::AA_UseHighDpiPixmaps application attribute enables this function to return pixmaps that are larger than the requested size. Such images will have a devicePixelRatio larger than 1.

fn set_theme_name(path: &String)

C++ method: static void QIcon::setThemeName(const QString& path)

C++ documentation:

Sets the current icon theme to name.

The name should correspond to a directory name in the themeSearchPath() containing an index.theme file describing it's contents.

This function was introduced in Qt 4.6.

See also themeSearchPaths() and themeName().

fn set_theme_search_paths(searchpath: &StringList)

C++ method: static void QIcon::setThemeSearchPaths(const QStringList& searchpath)

C++ documentation:

Sets the search paths for icon themes to paths.

This function was introduced in Qt 4.6.

See also themeSearchPaths(), fromTheme(), and setThemeName().

fn swap(&mut self, other: &mut Icon)

C++ method: void QIcon::swap(QIcon& other)

C++ documentation:

Swaps icon other with this icon. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn theme_name<Args>(args: Args) -> Args::ReturnType where Args: IconThemeNameArgs

C++ method: QIcon::themeName

This is an overloaded function. Available variants:

Rust arguments:
1) fn theme_name(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn theme_name(cpp_utils::AsStruct) -> qt_core::string::String
C++ method: static QString QIcon::themeName()

C++ documentation:

Returns the name of the current icon theme.

On X11, the current icon theme depends on your desktop settings. On other platforms it is not set by default.

This function was introduced in Qt 4.6.

See also setThemeName(), themeSearchPaths(), fromTheme(), and hasThemeIcon().

fn theme_search_paths<Args>(args: Args) -> Args::ReturnType where Args: IconThemeSearchPathsArgs

C++ method: QIcon::themeSearchPaths

This is an overloaded function. Available variants:

Rust arguments:
1) fn theme_search_paths(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn theme_search_paths(cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: static QStringList QIcon::themeSearchPaths()

C++ documentation:

Returns the search paths for icon themes.

The default value will depend on the platform:

On X11, the search path will use the XDG_DATA_DIRS environment variable if available.

By default all platforms will have the resource directory :\icons as a fallback. You can use "rcc -project" to generate a resource file from your icon theme.

This function was introduced in Qt 4.6.

See also setThemeSearchPaths(), fromTheme(), and setThemeName().