Module qt_gui::icon_engine::overloading

Traits

IconEngineActualSizeArgs
This trait represents a set of arguments accepted by IconEngine::actual_size method.

IconEngineAvailableSizesArgs
This trait represents a set of arguments accepted by IconEngine::available_sizes method.

IconEngineIconNameArgs
This trait represents a set of arguments accepted by IconEngine::icon_name method.

IconEngineKeyArgs
This trait represents a set of arguments accepted by IconEngine::key method.

IconEnginePixmapArgs
This trait represents a set of arguments accepted by IconEngine::pixmap method.