Struct qt_gui::icon_engine::IconEngine

pub struct IconEngine {
    // some fields omitted
}

C++ type: QIconEngine

C++ documentation:

The QIconEngine class provides an abstract base class for QIcon renderers.

An icon engine provides the rendering functions for a QIcon. Each icon has a corresponding icon engine that is responsible for drawing the icon with a requested size, mode and state.

The icon is rendered by the paint() function, and the icon can additionally be obtained as a pixmap with the pixmap() function (the default implementation simply uses paint() to achieve this). The addPixmap() function can be used to add new pixmaps to the icon engine, and is used by QIcon to add specialized custom pixmaps.

The paint(), pixmap(), and addPixmap() functions are all virtual, and can therefore be reimplemented in subclasses of QIconEngine.

Methods

impl IconEngine

fn actual_size<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IconEngineActualSizeArgs<'largs>

C++ method: QIconEngine::actualSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn actual_size(&mut self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn actual_size(&mut self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: virtual QSize QIconEngine::actualSize(const QSize& size, QIcon::Mode mode, QIcon::State state)

C++ documentation:

Returns the actual size of the icon the engine provides for the requested size, mode and state. The default implementation returns the given size.

fn add_file(&mut self, file_name: &String, size: &Size, mode: Mode, state: State)

C++ method: virtual void QIconEngine::addFile(const QString& fileName, const QSize& size, QIcon::Mode mode, QIcon::State state)

C++ documentation:

Called by QIcon::addFile(). Adds a specialized pixmap from the file with the given fileName, size, mode and state. The default pixmap-based engine stores any supplied file names, and it loads the pixmaps on demand instead of using scaled pixmaps if the size of a pixmap matches the size of icon requested. Custom icon engines that implement scalable vector formats are free to ignores any extra files.

fn add_pixmap(&mut self, pixmap: &Pixmap, mode: Mode, state: State)

C++ method: virtual void QIconEngine::addPixmap(const QPixmap& pixmap, QIcon::Mode mode, QIcon::State state)

C++ documentation:

Called by QIcon::addPixmap(). Adds a specialized pixmap for the given mode and state. The default pixmap-based engine stores any supplied pixmaps, and it uses them instead of scaled pixmaps if the size of a pixmap matches the size of icon requested. Custom icon engines that implement scalable vector formats are free to ignores any extra pixmaps.

fn available_sizes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconEngineAvailableSizesArgs<'largs>

C++ method: QIconEngine::availableSizes

This is an overloaded function. Available variants:

Rust arguments:
1) fn available_sizes(&self, (::icon::Mode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListSizeRef>
2) fn available_sizes(&self, (::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListSizeRef>
3) fn available_sizes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListSizeRef>
4) fn available_sizes(&self, (::icon::Mode, cpp_utils::AsStruct)) -> ::list::ListSizeRef
5) fn available_sizes(&self, (::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::list::ListSizeRef
6) fn available_sizes(&self, cpp_utils::AsStruct) -> ::list::ListSizeRef
C++ method: virtual QList<QSize> QIconEngine::availableSizes(QIcon::Mode mode = ?, QIcon::State state = ?) const

C++ documentation:

Returns sizes of all images that are contained in the engine for the specific mode and state.

Note: This is a helper method and the actual work is done by virtual_hook() method, hence this method depends on icon engine support and may not work with all icon engines.

This function was introduced in Qt 4.5.

fn clone(&self) -> *mut IconEngine

C++ method: pure virtual QIconEngine* QIconEngine::clone() const

C++ documentation:

Reimplement this method to return a clone of this icon engine.

fn icon_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconEngineIconNameArgs<'largs>

C++ method: QIconEngine::iconName

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn icon_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: virtual QString QIconEngine::iconName() const

C++ documentation:

Returns the name used to create the engine, if available.

Note: This is a helper method and the actual work is done by virtual_hook() method, hence this method depends on icon engine support and may not work with all icon engines.

This function was introduced in Qt 4.7.

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IconEngineKeyArgs<'largs>

C++ method: QIconEngine::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn key(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: virtual QString QIconEngine::key() const

C++ documentation:

Returns a key that identifies this icon engine.

fn paint(&mut self, painter: *mut Painter, rect: &Rect, mode: Mode, state: State)

C++ method: pure virtual void QIconEngine::paint(QPainter* painter, const QRect& rect, QIcon::Mode mode, QIcon::State state)

C++ documentation:

Uses the given painter to paint the icon with the required mode and state into the rectangle rect.

fn pixmap<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IconEnginePixmapArgs<'largs>

C++ method: QIconEngine::pixmap

This is an overloaded function. Available variants:

Rust arguments:
1) fn pixmap(&mut self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn pixmap(&mut self, (&qt_core::size::Size, ::icon::Mode, ::icon::State, cpp_utils::AsStruct)) -> ::pixmap::Pixmap
C++ method: virtual QPixmap QIconEngine::pixmap(const QSize& size, QIcon::Mode mode, QIcon::State state)

C++ documentation:

Returns the icon as a pixmap with the required size, mode, and state. The default implementation creates a new pixmap and calls paint() to fill it.

fn read(&mut self, in_: &mut DataStream) -> bool

C++ method: virtual bool QIconEngine::read(QDataStream& in)

C++ documentation:

Reads icon engine contents from the QDataStream in. Returns true if the contents were read; otherwise returns false.

QIconEngine's default implementation always return false.

fn virtual_hook(&mut self, id: c_int, data: *mut c_void)

C++ method: virtual void QIconEngine::virtual_hook(int id, void* data)

C++ documentation:

Additional method to allow extending QIconEngine without adding new virtual methods (and without breaking binary compatibility). The actual action and format of data depends on id argument which is in fact a constant from IconEngineHook enum.

This function was introduced in Qt 4.5.

See also IconEngineHook.

fn write(&self, out: &mut DataStream) -> bool

C++ method: virtual bool QIconEngine::write(QDataStream& out) const

C++ documentation:

Writes the contents of this engine to the QDataStream out. Returns true if the contents were written; otherwise returns false.

QIconEngine's default implementation always return false.