Module qt_gui::icon_engine_plugin::overloading

Traits

IconEnginePluginConnectArgs
This trait represents a set of arguments accepted by IconEnginePlugin::connect method.

IconEnginePluginConnectStaticArgs
This trait represents a set of arguments accepted by IconEnginePlugin::connect_static method.

IconEnginePluginCreateArgs
This trait represents a set of arguments accepted by IconEnginePlugin::create method.

IconEnginePluginDestroyedArgs
This trait represents a set of arguments accepted by IconEnginePlugin::destroyed method.

IconEnginePluginDisconnectArgs
This trait represents a set of arguments accepted by IconEnginePlugin::disconnect method.

IconEnginePluginDisconnectStaticArgs
This trait represents a set of arguments accepted by IconEnginePlugin::disconnect_static method.

IconEnginePluginDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by IconEnginePlugin::dynamic_property_names method.

IconEnginePluginObjectNameArgs
This trait represents a set of arguments accepted by IconEnginePlugin::object_name method.

IconEnginePluginPropertyArgs
This trait represents a set of arguments accepted by IconEnginePlugin::property method.

IconEnginePluginStartTimerArgs
This trait represents a set of arguments accepted by IconEnginePlugin::start_timer method.

IconEnginePluginTrArgs
This trait represents a set of arguments accepted by IconEnginePlugin::tr method.

IconEnginePluginTrUtf8Args
This trait represents a set of arguments accepted by IconEnginePlugin::tr_utf8 method.