Enum qt_gui::image::Format

pub enum Format {
    FormatInvalid,
    FormatMono,
    FormatMonoLSB,
    FormatIndexed8,
    FormatRGB32,
    FormatARGB32,
    FormatARGB32Premultiplied,
    FormatRGB16,
    FormatARGB8565Premultiplied,
    FormatRGB666,
    FormatARGB6666Premultiplied,
    FormatRGB555,
    FormatARGB8555Premultiplied,
    FormatRGB888,
    FormatRGB444,
    FormatARGB4444Premultiplied,
    FormatRGBX8888,
    FormatRGBA8888,
    FormatRGBA8888Premultiplied,
    FormatBGR30,
    FormatA2BGR30Premultiplied,
    FormatRGB30,
    FormatA2RGB30Premultiplied,
    FormatAlpha8,
    FormatGrayscale8,
    NImageFormats,
}

C++ type: QImage::Format

C++ documentation:

The following image formats are available in Qt. Values from Format_ARGB8565_Premultiplied to Format_ARGB4444_Premultiplied were added in Qt 4.4. Values Format_RGBX8888, Format_RGBA8888 and Format_RGBA8888_Premultiplied were added in Qt 5.2. Values Format_BGR30, Format_A2BGR30_Premultiplied, Format_RGB30, Format_A2RGB30_Premultiplied were added in Qt 5.4. Format_Alpha8 and Format_Grayscale8 were added in Qt 5.5. See the notes after the table.

Note: Drawing into a QImage with QImage::Format_Indexed8 is not supported.

Note: Do not render into ARGB32 images using QPainter. Using QImage::Format_ARGB32_Premultiplied is significantly faster.

Note: Formats with more than 8 bit per color channel will only be processed by the raster engine using 8 bit per color.

See also format() and convertToFormat().

Variants
FormatInvalid

The image is invalid. (C++ enum variant: Format_Invalid = 0)

FormatMono

The image is stored using 1-bit per pixel. Bytes are packed with the most significant bit (MSB) first. (C++ enum variant: Format_Mono = 1)

FormatMonoLSB

The image is stored using 1-bit per pixel. Bytes are packed with the less significant bit (LSB) first. (C++ enum variant: Format_MonoLSB = 2)

FormatIndexed8

The image is stored using 8-bit indexes into a colormap. (C++ enum variant: Format_Indexed8 = 3)

FormatRGB32

The image is stored using a 32-bit RGB format (0xffRRGGBB). (C++ enum variant: Format_RGB32 = 4)

FormatARGB32

The image is stored using a 32-bit ARGB format (0xAARRGGBB). (C++ enum variant: Format_ARGB32 = 5)

FormatARGB32Premultiplied

The image is stored using a premultiplied 32-bit ARGB format (0xAARRGGBB), i.e. the red, green, and blue channels are multiplied by the alpha component divided by 255. (If RR, GG, or BB has a higher value than the alpha channel, the results are undefined.) Certain operations (such as image composition using alpha blending) are faster using premultiplied ARGB32 than with plain ARGB32. (C++ enum variant: Format_ARGB32_Premultiplied = 6)

FormatRGB16

The image is stored using a 16-bit RGB format (5-6-5). (C++ enum variant: Format_RGB16 = 7)

FormatARGB8565Premultiplied

The image is stored using a premultiplied 24-bit ARGB format (8-5-6-5). (C++ enum variant: Format_ARGB8565_Premultiplied = 8)

FormatRGB666

The image is stored using a 24-bit RGB format (6-6-6). The unused most significant bits is always zero. (C++ enum variant: Format_RGB666 = 9)

FormatARGB6666Premultiplied

The image is stored using a premultiplied 24-bit ARGB format (6-6-6-6). (C++ enum variant: Format_ARGB6666_Premultiplied = 10)

FormatRGB555

The image is stored using a 16-bit RGB format (5-5-5). The unused most significant bit is always zero. (C++ enum variant: Format_RGB555 = 11)

FormatARGB8555Premultiplied

The image is stored using a premultiplied 24-bit ARGB format (8-5-5-5). (C++ enum variant: Format_ARGB8555_Premultiplied = 12)

FormatRGB888

The image is stored using a 24-bit RGB format (8-8-8). (C++ enum variant: Format_RGB888 = 13)

FormatRGB444

The image is stored using a 16-bit RGB format (4-4-4). The unused bits are always zero. (C++ enum variant: Format_RGB444 = 14)

FormatARGB4444Premultiplied

The image is stored using a premultiplied 16-bit ARGB format (4-4-4-4). (C++ enum variant: Format_ARGB4444_Premultiplied = 15)

FormatRGBX8888

The image is stored using a 32-bit byte-ordered RGB(x) format (8-8-8-8). This is the same as the Format_RGBA8888 except alpha must always be 255. (C++ enum variant: Format_RGBX8888 = 16)

FormatRGBA8888

The image is stored using a 32-bit byte-ordered RGBA format (8-8-8-8). Unlike ARGB32 this is a byte-ordered format, which means the 32bit encoding differs between big endian and little endian architectures, being respectively (0xRRGGBBAA) and (0xAABBGGRR). The order of the colors is the same on any architecture if read as bytes 0xRR,0xGG,0xBB,0xAA. (C++ enum variant: Format_RGBA8888 = 17)

FormatRGBA8888Premultiplied

The image is stored using a premultiplied 32-bit byte-ordered RGBA format (8-8-8-8). (C++ enum variant: Format_RGBA8888_Premultiplied = 18)

FormatBGR30

The image is stored using a 32-bit BGR format (x-10-10-10). (C++ enum variant: Format_BGR30 = 19)

FormatA2BGR30Premultiplied

The image is stored using a 32-bit premultiplied ABGR format (2-10-10-10). (C++ enum variant: Format_A2BGR30_Premultiplied = 20)

FormatRGB30

The image is stored using a 32-bit RGB format (x-10-10-10). (C++ enum variant: Format_RGB30 = 21)

FormatA2RGB30Premultiplied

The image is stored using a 32-bit premultiplied ARGB format (2-10-10-10). (C++ enum variant: Format_A2RGB30_Premultiplied = 22)

FormatAlpha8

The image is stored using an 8-bit alpha only format. (C++ enum variant: Format_Alpha8 = 23)

FormatGrayscale8

The image is stored using an 8-bit grayscale format. (C++ enum variant: Format_Grayscale8 = 24)

NImageFormats

C++ enum variant: NImageFormats = 25