Module qt_gui::image::overloading

Traits

ImageAlphaChannelArgs
This trait represents a set of arguments accepted by Image::alpha_channel method.

ImageAsQVariantArgs
This trait represents a set of arguments accepted by Image::as_q_variant method.

ImageColorTableArgs
This trait represents a set of arguments accepted by Image::color_table method.

ImageConvertToFormatArgs
This trait represents a set of arguments accepted by Image::convert_to_format method.

ImageConvertToFormatMutArgs
This trait represents a set of arguments accepted by Image::convert_to_format_mut method.

ImageCopyArgs
This trait represents a set of arguments accepted by Image::copy method.

ImageCreateAlphaMaskArgs
This trait represents a set of arguments accepted by Image::create_alpha_mask method.

ImageCreateHeuristicMaskArgs
This trait represents a set of arguments accepted by Image::create_heuristic_mask method.

ImageCreateMaskFromColorArgs
This trait represents a set of arguments accepted by Image::create_mask_from_color method.

ImageFillArgs
This trait represents a set of arguments accepted by Image::fill method.

ImageFromDataArgs
This trait represents a set of arguments accepted by Image::from_data method.

ImageInvertPixelsArgs
This trait represents a set of arguments accepted by Image::invert_pixels method.

ImageLoadArgs
This trait represents a set of arguments accepted by Image::load method.

ImageLoadFromDataArgs
This trait represents a set of arguments accepted by Image::load_from_data method.

ImageMirroredArgs
This trait represents a set of arguments accepted by Image::mirrored method.

ImageNewArgs
This trait represents a set of arguments accepted by Image::new method.

ImageOffsetArgs
This trait represents a set of arguments accepted by Image::offset method.

ImagePixelArgs
This trait represents a set of arguments accepted by Image::pixel method.

ImagePixelFormatArgs
This trait represents a set of arguments accepted by Image::pixel_format method.

ImagePixelIndexArgs
This trait represents a set of arguments accepted by Image::pixel_index method.

ImageRectArgs
This trait represents a set of arguments accepted by Image::rect method.

ImageRgbSwappedArgs
This trait represents a set of arguments accepted by Image::rgb_swapped method.

ImageSaveArgs
This trait represents a set of arguments accepted by Image::save method.

ImageScaledArgs
This trait represents a set of arguments accepted by Image::scaled method.

ImageScaledToHeightArgs
This trait represents a set of arguments accepted by Image::scaled_to_height method.

ImageScaledToWidthArgs
This trait represents a set of arguments accepted by Image::scaled_to_width method.

ImageSetPixelArgs
This trait represents a set of arguments accepted by Image::set_pixel method.

ImageSizeArgs
This trait represents a set of arguments accepted by Image::size method.

ImageTextArgs
This trait represents a set of arguments accepted by Image::text method.

ImageTextKeysArgs
This trait represents a set of arguments accepted by Image::text_keys method.

ImageToPixelFormatArgs
This trait represents a set of arguments accepted by Image::to_pixel_format method.

ImageTransformedArgs
This trait represents a set of arguments accepted by Image::transformed method.

ImageTrueMatrixArgs
This trait represents a set of arguments accepted by Image::true_matrix method.

ImageValidArgs
This trait represents a set of arguments accepted by Image::valid method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.