Struct qt_gui::image_i_o_handler::ImageIOHandler

pub struct ImageIOHandler {
    // some fields omitted
}

C++ type: QImageIOHandler

C++ documentation:

The QImageIOHandler class defines the common image I/O interface for all image formats in Qt.

Qt uses QImageIOHandler for reading and writing images through QImageReader and QImageWriter. You can also derive from this class to write your own image format handler using Qt's plugin mechanism.

Call setDevice() to assign a device to the handler, and setFormat() to assign a format to it. One QImageIOHandler may support more than one image format. canRead() returns true if an image can be read from the device, and read() and write() return true if reading or writing an image was completed successfully.

QImageIOHandler also has support for animations formats, through the functions loopCount(), imageCount(), nextImageDelay() and currentImageNumber().

In order to determine what options an image handler supports, Qt will call supportsOption() and setOption(). Make sure to reimplement these functions if you can provide support for any of the options in the ImageOption enum.

To write your own image handler, you must at least reimplement canRead() and read(). Then create a QImageIOPlugin that can create the handler. Finally, install your plugin, and QImageReader and QImageWriter will then automatically load the plugin, and start using it.

Methods

impl ImageIOHandler

fn can_read(&self) -> bool

C++ method: pure virtual bool QImageIOHandler::canRead() const

C++ documentation:

Returns true if an image can be read from the device (i.e., the image format is supported, the device can be read from and the initial header information suggests that the image can be read); otherwise returns false.

When reimplementing canRead(), make sure that the I/O device (device()) is left in its original state (e.g., by using peek() rather than read()).

See also read() and QIODevice::peek().

fn current_image_number(&self) -> c_int

C++ method: virtual int QImageIOHandler::currentImageNumber() const

C++ documentation:

For image formats that support animation, this function returns the sequence number of the current image in the animation. If this function is called before any image is read(), -1 is returned. The number of the first image in the sequence is 0.

If the image format does not support animation, 0 is returned.

See also read().

fn current_image_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageIOHandlerCurrentImageRectArgs<'largs>

C++ method: QImageIOHandler::currentImageRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_image_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn current_image_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: virtual QRect QImageIOHandler::currentImageRect() const

C++ documentation:

Returns the rect of the current image. If no rect is defined for the image, and empty QRect() is returned.

This function is useful for animations, where only parts of the frame may be updated at a time.

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QImageIOHandler::device() const

C++ documentation:

Returns the device currently assigned to the QImageIOHandler. If not device has been assigned, 0 is returned.

See also setDevice().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageIOHandlerFormatArgs<'largs>

C++ method: QImageIOHandler::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn format(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QImageIOHandler::format() const

C++ documentation:

Returns the format that is currently assigned to QImageIOHandler. If no format has been assigned, an empty string is returned.

See also setFormat().

fn image_count(&self) -> c_int

C++ method: virtual int QImageIOHandler::imageCount() const

C++ documentation:

For image formats that support animation, this function returns the number of images in the animation. If the image format does not support animation, or if it is unable to determine the number of images, 0 is returned.

The default implementation returns 1 if canRead() returns true; otherwise 0 is returned.

fn jump_to_image(&mut self, image_number: c_int) -> bool

C++ method: virtual bool QImageIOHandler::jumpToImage(int imageNumber)

C++ documentation:

For image formats that support animation, this function jumps to the image whose sequence number is imageNumber. The next call to read() will attempt to read this image.

The default implementation does nothing, and returns false.

fn jump_to_next_image(&mut self) -> bool

C++ method: virtual bool QImageIOHandler::jumpToNextImage()

C++ documentation:

For image formats that support animation, this function jumps to the next image.

The default implementation does nothing, and returns false.

fn loop_count(&self) -> c_int

C++ method: virtual int QImageIOHandler::loopCount() const

C++ documentation:

For image formats that support animation, this function returns the number of times the animation should loop. If the image format does not support animation, 0 is returned.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageIOHandlerNameArgs<'largs>

C++ method: QImageIOHandler::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn name(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: virtual QByteArray QImageIOHandler::name() const

C++ documentation:

Use format() instead.

fn next_image_delay(&self) -> c_int

C++ method: virtual int QImageIOHandler::nextImageDelay() const

C++ documentation:

For image formats that support animation, this function returns the number of milliseconds to wait until reading the next image. If the image format does not support animation, 0 is returned.

fn option<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageIOHandlerOptionArgs<'largs>

C++ method: QImageIOHandler::option

This is an overloaded function. Available variants:

Rust arguments:
1) fn option(&self, (::image_i_o_handler::ImageOption, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn option(&self, (::image_i_o_handler::ImageOption, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: virtual QVariant QImageIOHandler::option(QImageIOHandler::ImageOption option) const

C++ documentation:

Returns the value assigned to option as a QVariant. The type of the value depends on the option. For example, option(Size) returns a QSize variant.

See also setOption() and supportsOption().

fn read(&mut self, image: *mut Image) -> bool

C++ method: pure virtual bool QImageIOHandler::read(QImage* image)

C++ documentation:

Read an image from the device, and stores it in image. Returns true if the image is successfully read; otherwise returns false.

For image formats that support incremental loading, and for animation formats, the image handler can assume that image points to the previous frame.

See also canRead().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QImageIOHandler::setDevice(QIODevice* device)

C++ documentation:

Sets the device of the QImageIOHandler to device. The image handler will use this device when reading and writing images.

The device can only be set once and must be set before calling canRead(), read(), write(), etc. If you need to read multiple files, construct multiple instances of the appropriate QImageIOHandler subclass.

See also device().

fn set_format(&self, format: &ByteArray)

C++ method: void QImageIOHandler::setFormat(const QByteArray& format) const

C++ documentation:

Sets the format of the QImageIOHandler to format. The format is most useful for handlers that support multiple image formats.

This function is declared const so that it can be called from canRead().

See also format().

fn set_format_mut(&mut self, format: &ByteArray)

C++ method: void QImageIOHandler::setFormat(const QByteArray& format)

C++ documentation:

Sets the format of the QImageIOHandler to format. The format is most useful for handlers that support multiple image formats.

See also format().

fn set_option(&mut self, option: ImageOption, value: &Variant)

C++ method: virtual void QImageIOHandler::setOption(QImageIOHandler::ImageOption option, const QVariant& value)

C++ documentation:

Sets the option option with the value value.

See also option() and ImageOption.

fn supports_option(&self, option: ImageOption) -> bool

C++ method: virtual bool QImageIOHandler::supportsOption(QImageIOHandler::ImageOption option) const

C++ documentation:

Returns true if the QImageIOHandler supports the option option; otherwise returns false. For example, if the QImageIOHandler supports the Size option, supportsOption(Size) must return true.

See also setOption() and option().

fn write(&mut self, image: &Image) -> bool

C++ method: virtual bool QImageIOHandler::write(const QImage& image)

C++ documentation:

Writes the image image to the assigned device. Returns true on success; otherwise returns false.

The default implementation does nothing, and simply returns false.