Module qt_gui::image_i_o_plugin::overloading

Traits

ImageIOPluginConnectArgs
This trait represents a set of arguments accepted by ImageIOPlugin::connect method.

ImageIOPluginConnectStaticArgs
This trait represents a set of arguments accepted by ImageIOPlugin::connect_static method.

ImageIOPluginCreateArgs
This trait represents a set of arguments accepted by ImageIOPlugin::create method.

ImageIOPluginDestroyedArgs
This trait represents a set of arguments accepted by ImageIOPlugin::destroyed method.

ImageIOPluginDisconnectArgs
This trait represents a set of arguments accepted by ImageIOPlugin::disconnect method.

ImageIOPluginDisconnectStaticArgs
This trait represents a set of arguments accepted by ImageIOPlugin::disconnect_static method.

ImageIOPluginDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ImageIOPlugin::dynamic_property_names method.

ImageIOPluginObjectNameArgs
This trait represents a set of arguments accepted by ImageIOPlugin::object_name method.

ImageIOPluginPropertyArgs
This trait represents a set of arguments accepted by ImageIOPlugin::property method.

ImageIOPluginStartTimerArgs
This trait represents a set of arguments accepted by ImageIOPlugin::start_timer method.

ImageIOPluginTrArgs
This trait represents a set of arguments accepted by ImageIOPlugin::tr method.

ImageIOPluginTrUtf8Args
This trait represents a set of arguments accepted by ImageIOPlugin::tr_utf8 method.