Module qt_gui::image_reader::overloading

Traits

ImageReaderBackgroundColorArgs
This trait represents a set of arguments accepted by ImageReader::background_color method.

ImageReaderClipRectArgs
This trait represents a set of arguments accepted by ImageReader::clip_rect method.

ImageReaderCurrentImageRectArgs
This trait represents a set of arguments accepted by ImageReader::current_image_rect method.

ImageReaderErrorStringArgs
This trait represents a set of arguments accepted by ImageReader::error_string method.

ImageReaderFileNameArgs
This trait represents a set of arguments accepted by ImageReader::file_name method.

ImageReaderFormatArgs
This trait represents a set of arguments accepted by ImageReader::format method.

ImageReaderImageFormatStaticArgs
This trait represents a set of arguments accepted by ImageReader::image_format_static method.

ImageReaderNewArgs
This trait represents a set of arguments accepted by ImageReader::new method.

ImageReaderReadArgs
This trait represents a set of arguments accepted by ImageReader::read method.

ImageReaderScaledClipRectArgs
This trait represents a set of arguments accepted by ImageReader::scaled_clip_rect method.

ImageReaderScaledSizeArgs
This trait represents a set of arguments accepted by ImageReader::scaled_size method.

ImageReaderSizeArgs
This trait represents a set of arguments accepted by ImageReader::size method.

ImageReaderSubTypeArgs
This trait represents a set of arguments accepted by ImageReader::sub_type method.

ImageReaderSupportedImageFormatsArgs
This trait represents a set of arguments accepted by ImageReader::supported_image_formats method.

ImageReaderSupportedMimeTypesArgs
This trait represents a set of arguments accepted by ImageReader::supported_mime_types method.

ImageReaderSupportedSubTypesArgs
This trait represents a set of arguments accepted by ImageReader::supported_sub_types method.

ImageReaderTextArgs
This trait represents a set of arguments accepted by ImageReader::text method.

ImageReaderTextKeysArgs
This trait represents a set of arguments accepted by ImageReader::text_keys method.

ImageReaderTrArgs
This trait represents a set of arguments accepted by ImageReader::tr method.

ImageReaderTrUtf8Args
This trait represents a set of arguments accepted by ImageReader::tr_utf8 method.