Struct qt_gui::image_reader::ImageReader

pub struct ImageReader {
    // some fields omitted
}

C++ type: QImageReader

C++ documentation:

The QImageReader class provides a format independent interface for reading images from files or other devices.

The most common way to read images is through QImage and QPixmap's constructors, or by calling QImage::load() and QPixmap::load(). QImageReader is a specialized class which gives you more control when reading images. For example, you can read an image into a specific size by calling setScaledSize(), and you can select a clip rect, effectively loading only parts of an image, by calling setClipRect(). Depending on the underlying support in the image format, this can save memory and speed up loading of images.

To read an image, you start by constructing a QImageReader object. Pass either a file name or a device pointer, and the image format to QImageReader's constructor. You can then set several options, such as the clip rect (by calling setClipRect()) and scaled size (by calling setScaledSize()). canRead() returns the image if the QImageReader can read the image (i.e., the image format is supported and the device is open for reading). Call read() to read the image.

If any error occurs when reading the image, read() will return a null QImage. You can then call error() to find the type of error that occurred, or errorString() to get a human readable description of what went wrong.

Methods

impl ImageReader

fn auto_detect_image_format(&self) -> bool

C++ method: bool QImageReader::autoDetectImageFormat() const

C++ documentation:

Returns true if image format autodetection is enabled on this image reader; otherwise returns false. By default, autodetection is enabled.

See also setAutoDetectImageFormat().

fn auto_transform(&self) -> bool

C++ method: bool QImageReader::autoTransform() const

C++ documentation:

Returns true if the image handler will apply transformation metadata on read().

This function was introduced in Qt 5.5.

See also setAutoTransform(), transformation(), and read().

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderBackgroundColorArgs<'largs>

C++ method: QImageReader::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: QColor QImageReader::backgroundColor() const

C++ documentation:

Returns the background color that's used when reading an image. If the image format does not support setting the background color an invalid color is returned.

This function was introduced in Qt 4.1.

See also setBackgroundColor() and read().

fn can_read(&self) -> bool

C++ method: bool QImageReader::canRead() const

C++ documentation:

Returns true if an image can be read for the device (i.e., the image format is supported, and the device seems to contain valid data); otherwise returns false.

canRead() is a lightweight function that only does a quick test to see if the image data is valid. read() may still return false after canRead() returns true, if the image data is corrupt.

For images that support animation, canRead() returns false when all frames have been read.

See also read() and supportedImageFormats().

fn clip_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderClipRectArgs<'largs>

C++ method: QImageReader::clipRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn clip_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn clip_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QImageReader::clipRect() const

C++ documentation:

Returns the clip rect (also known as the ROI, or Region Of Interest) of the image. If no clip rect has been set, an invalid QRect is returned.

See also setClipRect().

fn current_image_number(&self) -> c_int

C++ method: int QImageReader::currentImageNumber() const

C++ documentation:

For image formats that support animation, this function returns the sequence number of the current frame. If the image format doesn't support animation, 0 is returned.

This function returns -1 if an error occurred.

See also supportsAnimation(), QImageIOHandler::currentImageNumber(), and canRead().

fn current_image_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderCurrentImageRectArgs<'largs>

C++ method: QImageReader::currentImageRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_image_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn current_image_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QImageReader::currentImageRect() const

C++ documentation:

For image formats that support animation, this function returns the rect for the current frame. Otherwise, a null rect is returned.

See also supportsAnimation() and QImageIOHandler::currentImageRect().

fn decide_format_from_content(&self) -> bool

C++ method: bool QImageReader::decideFormatFromContent() const

C++ documentation:

Returns whether the image reader should decide which plugin to use only based on the contents of the datastream rather than on the file extension.

See also setDecideFormatFromContent().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QImageReader::device() const

C++ documentation:

Returns the device currently assigned to QImageReader, or 0 if no device has been assigned.

See also setDevice().

fn error(&self) -> ImageReaderError

C++ method: QImageReader::ImageReaderError QImageReader::error() const

C++ documentation:

Returns the type of error that occurred last.

See also ImageReaderError and errorString().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderErrorStringArgs<'largs>

C++ method: QImageReader::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QImageReader::errorString() const

C++ documentation:

Returns a human readable description of the last error that occurred.

See also error().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderFileNameArgs<'largs>

C++ method: QImageReader::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QImageReader::fileName() const

C++ documentation:

If the currently assigned device is a QFile, or if setFileName() has been called, this function returns the name of the file QImageReader reads from. Otherwise (i.e., if no device has been assigned or the device is not a QFile), an empty QString is returned.

See also setFileName() and setDevice().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderFormatArgs<'largs>

C++ method: QImageReader::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn format(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QImageReader::format() const

C++ documentation:

Returns the format QImageReader uses for reading images.

You can call this function after assigning a device to the reader to determine the format of the device. For example:

QImageReader reader("image.png");
// reader.format() == "png"

If the reader cannot read any image from the device (e.g., there is no image there, or the image has already been read), or if the format is unsupported, this function returns an empty QByteArray().

See also setFormat() and supportedImageFormats().

fn image_count(&self) -> c_int

C++ method: int QImageReader::imageCount() const

C++ documentation:

For image formats that support animation, this function returns the total number of images in the animation. If the format does not support animation, 0 is returned.

This function returns -1 if an error occurred.

See also supportsAnimation(), QImageIOHandler::imageCount(), and canRead().

fn image_format(&self) -> Format

C++ method: QImage::Format QImageReader::imageFormat() const

C++ documentation:

Returns the format of the image, without actually reading the image contents. The format describes the image format QImageReader::read() returns, not the format of the actual image.

If the image format does not support this feature, this function returns an invalid format.

This function was introduced in Qt 4.5.

See also QImageIOHandler::ImageOption, QImageIOHandler::option(), and QImageIOHandler::supportsOption().

fn image_format_static<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderImageFormatStaticArgs

C++ method: QImageReader::imageFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn image_format_static((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn image_format_static((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: static QByteArray QImageReader::imageFormat(QIODevice* device)

C++ documentation:

If supported, this function returns the image format of the device device. Otherwise, an empty string is returned.

See also QImageReader::autoDetectImageFormat().

Variant 2

Rust arguments:
1) fn image_format_static((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn image_format_static((&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: static QByteArray QImageReader::imageFormat(const QString& fileName)

C++ documentation:

If supported, this function returns the image format of the file fileName. Otherwise, an empty string is returned.

fn jump_to_image(&mut self, image_number: c_int) -> bool

C++ method: bool QImageReader::jumpToImage(int imageNumber)

C++ documentation:

For image formats that support animation, this function skips to the image whose sequence number is imageNumber, returning true if successful or false if the corresponding image cannot be found.

The next call to read() will attempt to read this image.

See also jumpToNextImage() and QImageIOHandler::jumpToImage().

fn jump_to_next_image(&mut self) -> bool

C++ method: bool QImageReader::jumpToNextImage()

C++ documentation:

For image formats that support animation, this function steps over the current image, returning true if successful or false if there is no following image in the animation.

The default implementation calls read(), then discards the resulting image, but the image handler may have a more efficient way of implementing this operation.

See also jumpToImage() and QImageIOHandler::jumpToNextImage().

fn loop_count(&self) -> c_int

C++ method: int QImageReader::loopCount() const

C++ documentation:

For image formats that support animation, this function returns the number of times the animation should loop. If this function returns -1, it can either mean the animation should loop forever, or that an error occurred. If an error occurred, canRead() will return false.

See also supportsAnimation(), QImageIOHandler::loopCount(), and canRead().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderNewArgs

C++ method: QImageReader::QImageReader

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::image_reader::ImageReader
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::image_reader::ImageReader>
C++ method: [constructor] void QImageReader::QImageReader()

Variant 2

Rust arguments:
1) fn new((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::image_reader::ImageReader
2) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::image_reader::ImageReader
3) fn new((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_reader::ImageReader>
4) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_reader::ImageReader>
C++ method: [constructor] void QImageReader::QImageReader(QIODevice* device, const QByteArray& format = ?)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::image_reader::ImageReader
2) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::image_reader::ImageReader
3) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_reader::ImageReader>
4) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_reader::ImageReader>
C++ method: [constructor] void QImageReader::QImageReader(const QString& fileName, const QByteArray& format = ?)

fn next_image_delay(&self) -> c_int

C++ method: int QImageReader::nextImageDelay() const

C++ documentation:

For image formats that support animation, this function returns the number of milliseconds to wait until displaying the next frame in the animation. If the image format doesn't support animation, 0 is returned.

This function returns -1 if an error occurred.

See also supportsAnimation(), QImageIOHandler::nextImageDelay(), and canRead().

fn quality(&self) -> c_int

C++ method: int QImageReader::quality() const

C++ documentation:

Returns the quality setting of the image format.

This function was introduced in Qt 4.2.

See also setQuality().

fn read<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ImageReaderReadArgs<'largs>

C++ method: QImageReader::read

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn read(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::image::Image>
2) fn read(&mut self, cpp_utils::AsStruct) -> ::image::Image
C++ method: QImage QImageReader::read()

C++ documentation:

Reads an image from the device. On success, the image that was read is returned; otherwise, a null QImage is returned. You can then call error() to find the type of error that occurred, or errorString() to get a human readable description of the error.

For image formats that support animation, calling read() repeatedly will return the next frame. When all frames have been read, a null image will be returned.

See also canRead(), supportedImageFormats(), supportsAnimation(), and QMovie.

Variant 2

Rust arguments: fn read(&mut self, *mut ::image::Image) -> bool
C++ method: bool QImageReader::read(QImage* image)

C++ documentation:

This is an overloaded function.

Reads an image from the device into image, which must point to a QImage. Returns true on success; otherwise, returns false.

If image has same format and size as the image data that is about to be read, this function may not need to allocate a new image before reading. Because of this, it can be faster than the other read() overload, which always constructs a new image; especially when reading several images with the same format and size.

QImage icon(64, 64, QImage::Format_RGB32);
QImageReader reader("icon_64x64.bmp");
if (reader.read(&icon)) {
    // Display icon
}

For image formats that support animation, calling read() repeatedly will return the next frame. When all frames have been read, a null image will be returned.

See also canRead(), supportedImageFormats(), supportsAnimation(), and QMovie.

fn scaled_clip_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderScaledClipRectArgs<'largs>

C++ method: QImageReader::scaledClipRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn scaled_clip_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn scaled_clip_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QImageReader::scaledClipRect() const

C++ documentation:

Returns the scaled clip rect of the image.

See also setScaledClipRect().

fn scaled_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderScaledSizeArgs<'largs>

C++ method: QImageReader::scaledSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn scaled_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn scaled_size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QImageReader::scaledSize() const

C++ documentation:

Returns the scaled size of the image.

See also setScaledSize().

fn set_auto_detect_image_format(&mut self, enabled: bool)

C++ method: void QImageReader::setAutoDetectImageFormat(bool enabled)

C++ documentation:

If enabled is true, image format autodetection is enabled; otherwise, it is disabled. By default, autodetection is enabled.

QImageReader uses an extensive approach to detecting the image format; firstly, if you pass a file name to QImageReader, it will attempt to detect the file extension if the given file name does not point to an existing file, by appending supported default extensions to the given file name, one at a time. It then uses the following approach to detect the image format:

Image plugins are queried first, based on either the optional format string, or the file name suffix (if the source device is a file). No content detection is done at this stage. QImageReader will choose the first plugin that supports reading for this format.

If no plugin supports the image format, Qt's built-in handlers are checked based on either the optional format string, or the file name suffix.

If no capable plugins or built-in handlers are found, each plugin is tested by inspecting the content of the data stream.

If no plugins could detect the image format based on data contents, each built-in image handler is tested by inspecting the contents.

Finally, if all above approaches fail, QImageReader will report failure when trying to read the image.

By disabling image format autodetection, QImageReader will only query the plugins and built-in handlers based on the format string (i.e., no file name extensions are tested).

See also autoDetectImageFormat(), QImageIOHandler::canRead(), and QImageIOPlugin::capabilities().

fn set_auto_transform(&mut self, enabled: bool)

C++ method: void QImageReader::setAutoTransform(bool enabled)

C++ documentation:

Sets if images returned by read() should have transformation metadata automatically applied.

This function was introduced in Qt 5.5.

See also autoTransform(), transform(), and read().

fn set_background_color(&mut self, color: &Color)

C++ method: void QImageReader::setBackgroundColor(const QColor& color)

C++ documentation:

Sets the background color to color. Image formats that support this operation are expected to initialize the background to color before reading an image.

This function was introduced in Qt 4.1.

See also backgroundColor() and read().

fn set_clip_rect(&mut self, rect: &Rect)

C++ method: void QImageReader::setClipRect(const QRect& rect)

C++ documentation:

Sets the image clip rect (also known as the ROI, or Region Of Interest) to rect. The coordinates of rect are relative to the untransformed image size, as returned by size().

See also clipRect(), setScaledSize(), and setScaledClipRect().

fn set_decide_format_from_content(&mut self, ignored: bool)

C++ method: void QImageReader::setDecideFormatFromContent(bool ignored)

C++ documentation:

If ignored is set to true, then the image reader will ignore specified formats or file extensions and decide which plugin to use only based on the contents in the datastream.

Setting this flag means that all image plugins gets loaded. Each plugin will read the first bytes in the image data and decide if the plugin is compatible or not.

This also disables auto detecting the image format.

See also decideFormatFromContent().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QImageReader::setDevice(QIODevice* device)

C++ documentation:

Sets QImageReader's device to device. If a device has already been set, the old device is removed from QImageReader and is otherwise left unchanged.

If the device is not already open, QImageReader will attempt to open the device in QIODevice::ReadOnly mode by calling open(). Note that this does not work for certain devices, such as QProcess, QTcpSocket and QUdpSocket, where more logic is required to open the device.

See also device() and setFileName().

fn set_file_name(&mut self, file_name: &String)

C++ method: void QImageReader::setFileName(const QString& fileName)

C++ documentation:

Sets the file name of QImageReader to fileName. Internally, QImageReader will create a QFile object and open it in QIODevice::ReadOnly mode, and use this when reading images.

If fileName does not include a file extension (e.g., .png or .bmp), QImageReader will cycle through all supported extensions until it finds a matching file.

See also fileName(), setDevice(), and supportedImageFormats().

fn set_format(&mut self, format: &ByteArray)

C++ method: void QImageReader::setFormat(const QByteArray& format)

C++ documentation:

Sets the format QImageReader will use when reading images, to format. format is a case insensitive text string. Example:

QImageReader reader;
reader.setFormat("png"); // same as reader.setFormat("PNG");

You can call supportedImageFormats() for the full list of formats QImageReader supports.

See also format().

fn set_quality(&mut self, quality: c_int)

C++ method: void QImageReader::setQuality(int quality)

C++ documentation:

Sets the quality setting of the image format to quality.

Some image formats, in particular lossy ones, entail a tradeoff between a) visual quality of the resulting image, and b) decoding execution time. This function sets the level of that tradeoff for image formats that support it.

In case of scaled image reading, the quality setting may also influence the tradeoff level between visual quality and execution speed of the scaling algorithm.

The value range of quality depends on the image format. For example, the "jpeg" format supports a quality range from 0 (low visual quality) to 100 (high visual quality).

This function was introduced in Qt 4.2.

See also quality() and setScaledSize().

fn set_scaled_clip_rect(&mut self, rect: &Rect)

C++ method: void QImageReader::setScaledClipRect(const QRect& rect)

C++ documentation:

Sets the scaled clip rect to rect. The scaled clip rect is the clip rect (also known as ROI, or Region Of Interest) that is applied after the image has been scaled.

See also scaledClipRect() and setScaledSize().

fn set_scaled_size(&mut self, size: &Size)

C++ method: void QImageReader::setScaledSize(const QSize& size)

C++ documentation:

Sets the scaled size of the image to size. The scaling is performed after the initial clip rect, but before the scaled clip rect is applied. The algorithm used for scaling depends on the image format. By default (i.e., if the image format does not support scaling), QImageReader will use QImage::scale() with Qt::SmoothScaling.

See also scaledSize(), setClipRect(), and setScaledClipRect().

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderSizeArgs<'largs>

C++ method: QImageReader::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QImageReader::size() const

C++ documentation:

Returns the size of the image, without actually reading the image contents.

If the image format does not support this feature, this function returns an invalid size. Qt's built-in image handlers all support this feature, but custom image format plugins are not required to do so.

See also QImageIOHandler::ImageOption, QImageIOHandler::option(), and QImageIOHandler::supportsOption().

fn sub_type<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderSubTypeArgs<'largs>

C++ method: QImageReader::subType

This is an overloaded function. Available variants:

Rust arguments:
1) fn sub_type(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn sub_type(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QImageReader::subType() const

C++ documentation:

Returns the subtype of the image.

This function was introduced in Qt 5.4.

fn supported_image_formats<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderSupportedImageFormatsArgs

C++ method: QImageReader::supportedImageFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_image_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_image_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QImageReader::supportedImageFormats()

C++ documentation:

Returns the list of image formats supported by QImageReader.

By default, Qt can read the following formats:

Format	MIME type	Description

BMP	image/bmp	Windows Bitmap

GIF	image/gif	Graphic Interchange Format (optional)

JPG	image/jpeg	Joint Photographic Experts Group

PNG	image/png	Portable Network Graphics

PBM	image/x-portable-bitmap	Portable Bitmap

PGM	image/x-portable-graymap	Portable Graymap

PPM	image/x-portable-pixmap	Portable Pixmap

XBM	image/x-xbitmap	X11 Bitmap

XPM	image/x-xpixmap	X11 Pixmap

SVG	image/svg+xml	Scalable Vector Graphics

Reading and writing SVG files is supported through the Qt SVG module. The Qt Image Formats module provides support for additional image formats.

Note that the QApplication instance must be created before this function is called.

See also setFormat(), QImageWriter::supportedImageFormats(), and QImageIOPlugin.

fn supported_mime_types<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderSupportedMimeTypesArgs

C++ method: QImageReader::supportedMimeTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_mime_types(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_mime_types(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QImageReader::supportedMimeTypes()

C++ documentation:

Returns the list of MIME types supported by QImageReader.

Note that the QApplication instance must be created before this function is called.

See also supportedImageFormats() and QImageWriter::supportedMimeTypes().

fn supported_sub_types<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderSupportedSubTypesArgs<'largs>

C++ method: QImageReader::supportedSubTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_sub_types(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_sub_types(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QImageReader::supportedSubTypes() const

C++ documentation:

Returns the list of subtypes supported by an image.

This function was introduced in Qt 5.4.

fn supports_animation(&self) -> bool

C++ method: bool QImageReader::supportsAnimation() const

C++ documentation:

Returns true if the image format supports animation; otherwise, false is returned.

This function was introduced in Qt 4.1.

See also QMovie::supportedFormats().

fn supports_option(&self, option: ImageOption) -> bool

C++ method: bool QImageReader::supportsOption(QImageIOHandler::ImageOption option) const

C++ documentation:

Returns true if the reader supports option; otherwise returns false.

Different image formats support different options. Call this function to determine whether a certain option is supported by the current format. For example, the PNG format allows you to embed text into the image's metadata (see text()), and the BMP format allows you to determine the image's size without loading the whole image into memory (see size()).

QImageReader reader(":/image.png");
if (reader.supportsOption(QImageIOHandler::Size))
    qDebug() << "Size:" << reader.size();

This function was introduced in Qt 4.2.

See also QImageWriter::supportsOption().

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderTextArgs<'largs>

C++ method: QImageReader::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QImageReader::text(const QString& key) const

C++ documentation:

Returns the image text associated with key.

Support for this option is implemented through QImageIOHandler::Description.

This function was introduced in Qt 4.1.

See also textKeys() and QImageWriter::setText().

fn text_keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageReaderTextKeysArgs<'largs>

C++ method: QImageReader::textKeys

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn text_keys(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: QStringList QImageReader::textKeys() const

C++ documentation:

Returns the text keys for this image. You can use these keys with text() to list the image text for a certain key.

Support for this option is implemented through QImageIOHandler::Description.

This function was introduced in Qt 4.1.

See also text(), QImageWriter::setText(), and QImage::textKeys().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderTrArgs

C++ method: QImageReader::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QImageReader::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: ImageReaderTrUtf8Args

C++ method: QImageReader::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QImageReader::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn transformation(&self) -> Flags<Transformation>

C++ method: QFlags<QImageIOHandler::Transformation> QImageReader::transformation() const

C++ documentation:

Returns the transformation metadata of the image, including image orientation. If the format does not support transformation metadata QImageIOHandler::Transformation_None is returned.

This function was introduced in Qt 5.5.

See also setAutoTransform() and autoTransform().