Module qt_gui::image_writer::overloading

Traits

ImageWriterDescriptionArgs
This trait represents a set of arguments accepted by ImageWriter::description method.

ImageWriterErrorStringArgs
This trait represents a set of arguments accepted by ImageWriter::error_string method.

ImageWriterFileNameArgs
This trait represents a set of arguments accepted by ImageWriter::file_name method.

ImageWriterFormatArgs
This trait represents a set of arguments accepted by ImageWriter::format method.

ImageWriterNewArgs
This trait represents a set of arguments accepted by ImageWriter::new method.

ImageWriterSubTypeArgs
This trait represents a set of arguments accepted by ImageWriter::sub_type method.

ImageWriterSupportedImageFormatsArgs
This trait represents a set of arguments accepted by ImageWriter::supported_image_formats method.

ImageWriterSupportedMimeTypesArgs
This trait represents a set of arguments accepted by ImageWriter::supported_mime_types method.

ImageWriterSupportedSubTypesArgs
This trait represents a set of arguments accepted by ImageWriter::supported_sub_types method.

ImageWriterTrArgs
This trait represents a set of arguments accepted by ImageWriter::tr method.

ImageWriterTrUtf8Args
This trait represents a set of arguments accepted by ImageWriter::tr_utf8 method.