Struct qt_gui::image_writer::ImageWriter

pub struct ImageWriter {
    // some fields omitted
}

C++ type: QImageWriter

C++ documentation:

The QImageWriter class provides a format independent interface for writing images to files or other devices.

QImageWriter supports setting format specific options, such as the gamma level, compression level and quality, prior to storing the image. If you do not need such options, you can use QImage::save() or QPixmap::save() instead.

To store an image, you start by constructing a QImageWriter object. Pass either a file name or a device pointer, and the image format to QImageWriter's constructor. You can then set several options, such as the gamma level (by calling setGamma()) and quality (by calling setQuality()). canWrite() returns true if QImageWriter can write the image (i.e., the image format is supported and the device is open for writing). Call write() to write the image to the device.

If any error occurs when writing the image, write() will return false. You can then call error() to find the type of error that occurred, or errorString() to get a human readable description of what went wrong.

Call supportedImageFormats() for a list of formats that QImageWriter can write. QImageWriter supports all built-in image formats, in addition to any image format plugins that support writing.

Methods

impl ImageWriter

fn can_write(&self) -> bool

C++ method: bool QImageWriter::canWrite() const

C++ documentation:

Returns true if QImageWriter can write the image; i.e., the image format is supported and the assigned device is open for reading.

See also write(), setDevice(), and setFormat().

fn compression(&self) -> c_int

C++ method: int QImageWriter::compression() const

C++ documentation:

Returns the compression of the image.

See also setCompression().

fn description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterDescriptionArgs<'largs>

C++ method: QImageWriter::description

This is an overloaded function. Available variants:

Rust arguments:
1) fn description(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn description(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QImageWriter::description() const

C++ documentation:

Use QImageReader::text() instead.

Returns the description of the image.

See also setDescription().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QImageWriter::device() const

C++ documentation:

Returns the device currently assigned to QImageWriter, or 0 if no device has been assigned.

See also setDevice().

fn error(&self) -> ImageWriterError

C++ method: QImageWriter::ImageWriterError QImageWriter::error() const

C++ documentation:

Returns the type of error that last occurred.

See also ImageWriterError and errorString().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterErrorStringArgs<'largs>

C++ method: QImageWriter::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QImageWriter::errorString() const

C++ documentation:

Returns a human readable description of the last error that occurred.

See also error().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterFileNameArgs<'largs>

C++ method: QImageWriter::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QImageWriter::fileName() const

C++ documentation:

If the currently assigned device is a QFile, or if setFileName() has been called, this function returns the name of the file QImageWriter writes to. Otherwise (i.e., if no device has been assigned or the device is not a QFile), an empty QString is returned.

See also setFileName() and setDevice().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterFormatArgs<'largs>

C++ method: QImageWriter::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn format(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QImageWriter::format() const

C++ documentation:

Returns the format QImageWriter uses for writing images.

See also setFormat().

fn gamma(&self) -> c_float

C++ method: float QImageWriter::gamma() const

C++ documentation:

Returns the gamma level of the image.

See also setGamma().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ImageWriterNewArgs

C++ method: QImageWriter::QImageWriter

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::image_writer::ImageWriter
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::image_writer::ImageWriter>
C++ method: [constructor] void QImageWriter::QImageWriter()

Variant 2

Rust arguments:
1) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::image_writer::ImageWriter
2) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_writer::ImageWriter>
C++ method: [constructor] void QImageWriter::QImageWriter(QIODevice* device, const QByteArray& format)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::image_writer::ImageWriter
2) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::image_writer::ImageWriter
3) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_writer::ImageWriter>
4) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image_writer::ImageWriter>
C++ method: [constructor] void QImageWriter::QImageWriter(const QString& fileName, const QByteArray& format = ?)

fn optimized_write(&self) -> bool

C++ method: bool QImageWriter::optimizedWrite() const

C++ documentation:

Returns whether optimization has been turned on for writing the image.

This function was introduced in Qt 5.5.

See also setOptimizedWrite().

fn progressive_scan_write(&self) -> bool

C++ method: bool QImageWriter::progressiveScanWrite() const

C++ documentation:

Returns whether the image should be written as a progressive image.

This function was introduced in Qt 5.5.

See also setProgressiveScanWrite().

fn quality(&self) -> c_int

C++ method: int QImageWriter::quality() const

C++ documentation:

Returns the quality setting of the image format.

See also setQuality().

fn set_compression(&mut self, compression: c_int)

C++ method: void QImageWriter::setCompression(int compression)

C++ documentation:

This is an image format specific function that set the compression of an image. For image formats that do not support setting the compression, this value is ignored.

The value range of compression depends on the image format. For example, the "tiff" format supports two values, 0(no compression) and 1(LZW-compression).

See also compression().

fn set_description(&mut self, description: &String)

C++ method: void QImageWriter::setDescription(const QString& description)

C++ documentation:

Use setText() instead.

This is an image format specific function that sets the description of the image to description. For image formats that do not support setting the description, this value is ignored.

The contents of description depends on the image format.

See also description().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QImageWriter::setDevice(QIODevice* device)

C++ documentation:

Sets QImageWriter's device to device. If a device has already been set, the old device is removed from QImageWriter and is otherwise left unchanged.

If the device is not already open, QImageWriter will attempt to open the device in QIODevice::WriteOnly mode by calling open(). Note that this does not work for certain devices, such as QProcess, QTcpSocket and QUdpSocket, where more logic is required to open the device.

See also device() and setFileName().

fn set_file_name(&mut self, file_name: &String)

C++ method: void QImageWriter::setFileName(const QString& fileName)

C++ documentation:

Sets the file name of QImageWriter to fileName. Internally, QImageWriter will create a QFile and open it in QIODevice::WriteOnly mode, and use this file when writing images.

See also fileName() and setDevice().

fn set_format(&mut self, format: &ByteArray)

C++ method: void QImageWriter::setFormat(const QByteArray& format)

C++ documentation:

Sets the format QImageWriter will use when writing images, to format. format is a case insensitive text string. Example:

QImageWriter writer;
writer.setFormat("png"); // same as writer.setFormat("PNG");

You can call supportedImageFormats() for the full list of formats QImageWriter supports.

See also format().

fn set_gamma(&mut self, gamma: c_float)

C++ method: void QImageWriter::setGamma(float gamma)

C++ documentation:

This is an image format specific function that sets the gamma level of the image to gamma. For image formats that do not support setting the gamma level, this value is ignored.

The value range of gamma depends on the image format. For example, the "png" format supports a gamma range from 0.0 to 1.0.

See also gamma() and quality().

fn set_optimized_write(&mut self, optimize: bool)

C++ method: void QImageWriter::setOptimizedWrite(bool optimize)

C++ documentation:

This is an image format-specific function which sets the optimize flags when writing images. For image formats that do not support setting an optimize flag, this value is ignored.

The default is false.

This function was introduced in Qt 5.5.

See also optimizedWrite().

fn set_progressive_scan_write(&mut self, progressive: bool)

C++ method: void QImageWriter::setProgressiveScanWrite(bool progressive)

C++ documentation:

This is an image format-specific function which turns on progressive scanning when writing images. For image formats that do not support setting a progressive scan flag, this value is ignored.

The default is false.

This function was introduced in Qt 5.5.

See also progressiveScanWrite().

fn set_quality(&mut self, quality: c_int)

C++ method: void QImageWriter::setQuality(int quality)

C++ documentation:

Sets the quality setting of the image format to quality.

Some image formats, in particular lossy ones, entail a tradeoff between a) visual quality of the resulting image, and b) encoding execution time and compression level. This function sets the level of that tradeoff for image formats that support it. For other formats, this value is ignored.

The value range of quality depends on the image format. For example, the "jpeg" format supports a quality range from 0 (low visual quality, high compression) to 100 (high visual quality, low compression).

See also quality().

fn set_sub_type(&mut self, type_: &ByteArray)

C++ method: void QImageWriter::setSubType(const QByteArray& type)

C++ documentation:

This is an image format specific function that sets the subtype of the image to type. Subtype can be used by a handler to determine which format it should use while saving the image.

For example, saving an image in DDS format with A8R8G8R8 subtype:

QImageWriter writer("some/image.dds");
if (writer.supportsOption(QImageIOHandler::SubType))
    writer.setSubType("A8R8G8B8");
writer.write(image);

This function was introduced in Qt 5.4.

See also subType().

fn set_text(&mut self, key: &String, text: &String)

C++ method: void QImageWriter::setText(const QString& key, const QString& text)

C++ documentation:

Sets the image text associated with the key key to text. This is useful for storing copyright information or other information about the image. Example:

QImage image("some/image.jpeg");
QImageWriter writer("images/outimage.png", "png");
writer.setText("Author", "John Smith");
writer.write(image);

If you want to store a single block of data (e.g., a comment), you can pass an empty key, or use a generic key like "Description".

The key and text will be embedded into the image data after calling write().

Support for this option is implemented through QImageIOHandler::Description.

This function was introduced in Qt 4.1.

See also QImage::setText() and QImageReader::text().

fn set_transformation(&mut self, orientation: Flags<Transformation>)

C++ method: void QImageWriter::setTransformation(QFlags<QImageIOHandler::Transformation> orientation)

C++ documentation:

Sets the image transformations metadata including orientation.

If transformation metadata is not supported by the image format, the transform is applied before writing.

This function was introduced in Qt 5.5.

See also transformation() and write().

fn sub_type<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterSubTypeArgs<'largs>

C++ method: QImageWriter::subType

This is an overloaded function. Available variants:

Rust arguments:
1) fn sub_type(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn sub_type(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QImageWriter::subType() const

C++ documentation:

Returns the subtype of the image.

This function was introduced in Qt 5.4.

See also setSubType().

fn supported_image_formats<Args>(args: Args) -> Args::ReturnType where Args: ImageWriterSupportedImageFormatsArgs

C++ method: QImageWriter::supportedImageFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_image_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_image_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QImageWriter::supportedImageFormats()

C++ documentation:

Returns the list of image formats supported by QImageWriter.

By default, Qt can write the following formats:

Format	MIME type	Description

BMP	image/bmp	Windows Bitmap

JPG	image/jpeg	Joint Photographic Experts Group

PNG	image/png	Portable Network Graphics

PBM	image/x-portable-bitmap	Portable Bitmap

PGM	image/x-portable-graymap	Portable Graymap

PPM	image/x-portable-pixmap	Portable Pixmap

XBM	image/x-xbitmap	X11 Bitmap

XPM	image/x-xpixmap	X11 Pixmap

Reading and writing SVG files is supported through the Qt SVG module. The Qt Image Formats module provides support for additional image formats.

Note that the QApplication instance must be created before this function is called.

See also setFormat(), QImageReader::supportedImageFormats(), and QImageIOPlugin.

fn supported_mime_types<Args>(args: Args) -> Args::ReturnType where Args: ImageWriterSupportedMimeTypesArgs

C++ method: QImageWriter::supportedMimeTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_mime_types(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_mime_types(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QImageWriter::supportedMimeTypes()

C++ documentation:

Returns the list of MIME types supported by QImageWriter.

Note that the QApplication instance must be created before this function is called.

See also supportedImageFormats() and QImageReader::supportedMimeTypes().

fn supported_sub_types<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ImageWriterSupportedSubTypesArgs<'largs>

C++ method: QImageWriter::supportedSubTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_sub_types(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_sub_types(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QImageWriter::supportedSubTypes() const

C++ documentation:

Returns the list of subtypes supported by an image.

This function was introduced in Qt 5.4.

fn supports_option(&self, option: ImageOption) -> bool

C++ method: bool QImageWriter::supportsOption(QImageIOHandler::ImageOption option) const

C++ documentation:

Returns true if the writer supports option; otherwise returns false.

Different image formats support different options. Call this function to determine whether a certain option is supported by the current format. For example, the PNG format allows you to embed text into the image's metadata (see text()).

QImageWriter writer(fileName);
if (writer.supportsOption(QImageIOHandler::Description))
    writer.setText("Author", "John Smith");

Options can be tested after the writer has been associated with a format.

This function was introduced in Qt 4.2.

See also QImageReader::supportsOption() and setFormat().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: ImageWriterTrArgs

C++ method: QImageWriter::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QImageWriter::tr(const char* sourceText, const char* disambiguation, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: ImageWriterTrUtf8Args

C++ method: QImageWriter::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QImageWriter::trUtf8(const char* sourceText, const char* disambiguation, int n)

fn transformation(&self) -> Flags<Transformation>

C++ method: QFlags<QImageIOHandler::Transformation> QImageWriter::transformation() const

C++ documentation:

Returns the transformation and orientation the image has been set to written with.

This function was introduced in Qt 5.5.

See also setTransformation().

fn write(&mut self, image: &Image) -> bool

C++ method: bool QImageWriter::write(const QImage& image)

C++ documentation:

Writes the image image to the assigned device or file name. Returns true on success; otherwise returns false. If the operation fails, you can call error() to find the type of error that occurred, or errorString() to get a human readable description of the error.

See also canWrite(), error(), and errorString().