Struct qt_gui::input_method_event::InputMethodEvent

pub struct InputMethodEvent {
    // some fields omitted
}

C++ type: QInputMethodEvent

C++ documentation:

The QInputMethodEvent class provides parameters for input method events.

Input method events are sent to widgets when an input method is used to enter text into a widget. Input methods are widely used to enter text for languages with non-Latin alphabets.

Note that when creating custom text editing widgets, the Qt::WA_InputMethodEnabled window attribute must be set explicitly (using the QWidget::setAttribute() function) in order to receive input method events.

The events are of interest to authors of keyboard entry widgets who want to be able to correctly handle languages with complex character input. Text input in such languages is usually a three step process:

Starting to Compose
When the user presses the first key on a keyboard, an input context is created. This input context will contain a string of the typed characters.

Composing
With every new key pressed, the input method will try to create a matching string for the text typed so far called preedit string. While the input context is active, the user can only move the cursor inside the string belonging to this input context.

Completing
At some point, the user will activate a user interface component (perhaps using a particular key) where they can choose from a number of strings matching the text they have typed so far. The user can either confirm their choice cancel the input; in either case the input context will be closed.

QInputMethodEvent models these three stages, and transfers the information needed to correctly render the intermediate result. A QInputMethodEvent has two main parameters: preeditString() and commitString(). The preeditString() parameter gives the currently active preedit string. The commitString() parameter gives a text that should get added to (or replace parts of) the text of the editor widget. It usually is a result of the input operations and has to be inserted to the widgets text directly before the preedit string.

If the commitString() should replace parts of the of the text in the editor, replacementLength() will contain the number of characters to be replaced. replacementStart() contains the position at which characters are to be replaced relative from the start of the preedit string.

A number of attributes control the visual appearance of the preedit string (the visual appearance of text outside the preedit string is controlled by the widget only). The AttributeType enum describes the different attributes that can be set.

A class implementing QWidget::inputMethodEvent() or QGraphicsItem::inputMethodEvent() should at least understand and honor the TextFormat and Cursor attributes.

Since input methods need to be able to query certain properties from the widget or graphics item, subclasses must also implement QWidget::inputMethodQuery() and QGraphicsItem::inputMethodQuery(), respectively.

When receiving an input method event, the text widget has to performs the following steps:

If the widget has selected text, the selected text should get removed.

Remove the text starting at replacementStart() with length replacementLength() and replace it by the commitString(). If replacementLength() is 0, replacementStart() gives the insertion position for the commitString().
When doing replacement the area of the preedit string is ignored, thus a replacement starting at -1 with a length of 2 will remove the last character before the preedit string and the first character afterwards, and insert the commit string directly before the preedit string.

If the widget implements undo/redo, this operation gets added to the undo stack.

If there is no current preedit string, insert the preeditString() at the current cursor position; otherwise replace the previous preeditString with the one received from this event.
If the widget implements undo/redo, the preeditString() should not influence the undo/redo stack in any way.

The widget should examine the list of attributes to apply to the preedit string. It has to understand at least the TextFormat and Cursor attributes and render them as specified.

Methods

impl InputMethodEvent

fn accept(&mut self)

C++ method: void QInputMethodEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn attributes<'l0>(&'l0 self) -> &'l0 ListAttributeRef

C++ method: const QList<QInputMethodEvent::Attribute>& QInputMethodEvent::attributes() const

C++ documentation:

Returns the list of attributes passed to the QInputMethodEvent constructor. The attributes control the visual appearance of the preedit string (the visual appearance of text outside the preedit string is controlled by the widget only).

See also preeditString() and Attribute.

fn commit_string<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QInputMethodEvent::commitString() const

C++ documentation:

Returns the text that should get added to (or replace parts of) the text of the editor widget. It usually is a result of the input operations and has to be inserted to the widgets text directly before the preedit string.

See also setCommitString(), preeditString(), replacementStart(), and replacementLength().

fn ignore(&mut self)

C++ method: void QInputMethodEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QInputMethodEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: InputMethodEventNewArgs

C++ method: QInputMethodEvent::QInputMethodEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::input_method_event::InputMethodEvent
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::input_method_event::InputMethodEvent>
C++ method: [constructor] void QInputMethodEvent::QInputMethodEvent()

Variant 2

Rust arguments:
1) fn new((&::input_method_event::InputMethodEvent, cpp_utils::AsStruct)) -> ::input_method_event::InputMethodEvent
2) fn new((&::input_method_event::InputMethodEvent, cpp_utils::AsBox)) -> cpp_utils::CppBox<::input_method_event::InputMethodEvent>
C++ method: [constructor] void QInputMethodEvent::QInputMethodEvent(const QInputMethodEvent& other)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, &::list::ListAttributeRef, cpp_utils::AsStruct)) -> ::input_method_event::InputMethodEvent
2) fn new((&qt_core::string::String, &::list::ListAttributeRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::input_method_event::InputMethodEvent>
C++ method: [constructor] void QInputMethodEvent::QInputMethodEvent(const QString& preeditText, const QList<QInputMethodEvent::Attribute>& attributes)

fn preedit_string<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QInputMethodEvent::preeditString() const

C++ documentation:

Returns the preedit text, i.e. the text before the user started editing it.

See also commitString() and attributes().

fn register_event_type<Args>(args: Args) -> c_int where Args: InputMethodEventRegisterEventTypeArgs

C++ method: QInputMethodEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QInputMethodEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn replacement_length(&self) -> c_int

C++ method: int QInputMethodEvent::replacementLength() const

C++ documentation:

Returns the number of characters to be replaced in the preedit string.

See also replacementStart() and setCommitString().

fn replacement_start(&self) -> c_int

C++ method: int QInputMethodEvent::replacementStart() const

C++ documentation:

Returns the position at which characters are to be replaced relative from the start of the preedit string.

See also replacementLength() and setCommitString().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QInputMethodEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_commit_string<'largs, Args>(&'largs mut self, args: Args) where Args: InputMethodEventSetCommitStringArgs<'largs>

C++ method: QInputMethodEvent::setCommitString

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_commit_string(&mut self, &qt_core::string::String) -> ()
2) fn set_commit_string(&mut self, (&qt_core::string::String, libc::c_int)) -> ()
3) fn set_commit_string(&mut self, (&qt_core::string::String, libc::c_int, libc::c_int)) -> ()
C++ method: void QInputMethodEvent::setCommitString(const QString& commitString, int replaceFrom = ?, int replaceLength = ?)

C++ documentation:

Sets the commit string to commitString.

The commit string is the text that should get added to (or replace parts of) the text of the editor widget. It usually is a result of the input operations and has to be inserted to the widgets text directly before the preedit string.

If the commit string should replace parts of the of the text in the editor, replaceLength specifies the number of characters to be replaced. replaceFrom specifies the position at which characters are to be replaced relative from the start of the preedit string.

See also commitString(), replacementStart(), and replacementLength().

fn spontaneous(&self) -> bool

C++ method: bool QInputMethodEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QInputMethodEvent::type() const

C++ documentation:

Returns the event type.