Struct qt_gui::input_method_query_event::InputMethodQueryEvent

pub struct InputMethodQueryEvent {
    // some fields omitted
}

C++ type: QInputMethodQueryEvent

C++ documentation:

The QInputMethodQueryEvent class provides an event sent by the input context to input objects.

It is used by the input method to query a set of properties of the object to be able to support complex input method operations as support for surrounding text and reconversions.

queries() specifies which properties are queried.

The object should call setValue() on the event to fill in the requested data before calling accept().

Methods

impl InputMethodQueryEvent

fn accept(&mut self)

C++ method: void QInputMethodQueryEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn ignore(&mut self)

C++ method: void QInputMethodQueryEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QInputMethodQueryEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: InputMethodQueryEventNewArgs

C++ method: QInputMethodQueryEvent::QInputMethodQueryEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((qt_core::flags::Flags<qt_core::qt::InputMethodQuery>, cpp_utils::AsStruct)) -> ::input_method_query_event::InputMethodQueryEvent
2) fn new((qt_core::flags::Flags<qt_core::qt::InputMethodQuery>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::input_method_query_event::InputMethodQueryEvent>
C++ method: [constructor] void QInputMethodQueryEvent::QInputMethodQueryEvent(QFlags<Qt::InputMethodQuery> queries)

fn queries(&self) -> Flags<InputMethodQuery>

C++ method: QFlags<Qt::InputMethodQuery> QInputMethodQueryEvent::queries() const

C++ documentation:

Returns the properties queried by the event.

fn register_event_type<Args>(args: Args) -> c_int where Args: InputMethodQueryEventRegisterEventTypeArgs

C++ method: QInputMethodQueryEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QInputMethodQueryEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QInputMethodQueryEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_value(&mut self, query: InputMethodQuery, value: &Variant)

C++ method: void QInputMethodQueryEvent::setValue(Qt::InputMethodQuery query, const QVariant& value)

C++ documentation:

Sets property query to value.

See also value().

fn spontaneous(&self) -> bool

C++ method: bool QInputMethodQueryEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QInputMethodQueryEvent::type() const

C++ documentation:

Returns the event type.

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: InputMethodQueryEventValueArgs<'largs>

C++ method: QInputMethodQueryEvent::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, (qt_core::qt::InputMethodQuery, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn value(&self, (qt_core::qt::InputMethodQuery, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QInputMethodQueryEvent::value(Qt::InputMethodQuery query) const

C++ documentation:

Returns value of the property query.

See also setValue().