Struct qt_gui::key_event::KeyEvent

pub struct KeyEvent {
    // some fields omitted
}

C++ type: QKeyEvent

C++ documentation:

The QKeyEvent class describes a key event.

Key events are sent to the widget with keyboard input focus when keys are pressed or released.

A key event contains a special accept flag that indicates whether the receiver will handle the key event. This flag is set by default, so there is no need to call accept() when acting on a key event. Calling ignore() on a key event will propagate it to the parent widget. The event is propagated up the parent widget chain until a widget accepts it or an event filter consumes it.

The QWidget::setEnable() function can be used to enable or disable mouse and keyboard events for a widget.

The event handlers QWidget::keyPressEvent(), QWidget::keyReleaseEvent(), QGraphicsItem::keyPressEvent() and QGraphicsItem::keyReleaseEvent() receive key events.

Methods

impl KeyEvent

fn accept(&mut self)

C++ method: void QKeyEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn count(&self) -> c_int

C++ method: int QKeyEvent::count() const

C++ documentation:

Returns the number of keys involved in this event. If text() is not empty, this is simply the length of the string.

See also Qt::WA_KeyCompression.

fn ignore(&mut self)

C++ method: void QKeyEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QKeyEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn is_auto_repeat(&self) -> bool

C++ method: bool QKeyEvent::isAutoRepeat() const

C++ documentation:

Returns true if this event comes from an auto-repeating key; returns false if it comes from an initial key press.

Note that if the event is a multiple-key compressed event that is partly due to auto-repeat, this function could return either true or false indeterminately.

fn key(&self) -> c_int

C++ method: int QKeyEvent::key() const

C++ documentation:

Returns the code of the key that was pressed or released.

See Qt::Key for the list of keyboard codes. These codes are independent of the underlying window system. Note that this function does not distinguish between capital and non-capital letters, use the text() function (returning the Unicode text the key generated) for this purpose.

A value of either 0 or Qt::Key_unknown means that the event is not the result of a known key; for example, it may be the result of a compose sequence, a keyboard macro, or due to key event compression.

See also Qt::WA_KeyCompression.

fn matches(&self, key: StandardKey) -> bool

C++ method: bool QKeyEvent::matches(QKeySequence::StandardKey key) const

C++ documentation:

Returns true if the key event matches the given standard key; otherwise returns false.

This function was introduced in Qt 4.2.

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QKeyEvent::modifiers() const

C++ documentation:

Returns the keyboard modifier flags that existed immediately after the event occurred.

Warning: This function cannot always be trusted. The user can confuse it by pressing both Shift keys simultaneously and releasing one of them, for example.

See also QGuiApplication::keyboardModifiers().

fn native_modifiers(&self) -> u32

C++ method: quint32 QKeyEvent::nativeModifiers() const

C++ documentation:

Returns the native modifiers of a key event. If the key event does not contain this data 0 is returned.

Note: The native modifiers may be 0, even if the key event contains extended information.

This function was introduced in Qt 4.2.

fn native_scan_code(&self) -> u32

C++ method: quint32 QKeyEvent::nativeScanCode() const

C++ documentation:

Returns the native scan code of the key event. If the key event does not contain this data 0 is returned.

Note: The native scan code may be 0, even if the key event contains extended information.

Note: On Mac OS/X, this function is not useful, because there is no way to get the scan code from Carbon or Cocoa. The function always returns 1 (or 0 in the case explained above).

This function was introduced in Qt 4.2.

fn native_virtual_key(&self) -> u32

C++ method: quint32 QKeyEvent::nativeVirtualKey() const

C++ documentation:

Returns the native virtual key, or key sym of the key event. If the key event does not contain this data 0 is returned.

Note: The native virtual key may be 0, even if the key event contains extended information.

This function was introduced in Qt 4.2.

fn new<Args>(args: Args) -> Args::ReturnType where Args: KeyEventNewArgs

C++ method: QKeyEvent::QKeyEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
2) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
3) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, bool, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
4) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, bool, libc::c_ushort, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
5) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
6) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
7) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
8) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, &qt_core::string::String, bool, libc::c_ushort, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
C++ method: [constructor] void QKeyEvent::QKeyEvent(QEvent::Type type, int key, QFlags<Qt::KeyboardModifier> modifiers, const QString& text = ?, bool autorep = ?, unsigned short count = ?)

Variant 2

Rust arguments:
1) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
2) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
3) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, bool, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
4) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, bool, libc::c_ushort, cpp_utils::AsStruct)) -> ::key_event::KeyEvent
5) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
6) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
7) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
8) fn new((qt_core::event::Type, libc::c_int, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, u32, u32, u32, &qt_core::string::String, bool, libc::c_ushort, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_event::KeyEvent>
C++ method: [constructor] void QKeyEvent::QKeyEvent(QEvent::Type type, int key, QFlags<Qt::KeyboardModifier> modifiers, quint32 nativeScanCode, quint32 nativeVirtualKey, quint32 nativeModifiers, const QString& text = ?, bool autorep = ?, unsigned short count = ?)

fn register_event_type<Args>(args: Args) -> c_int where Args: KeyEventRegisterEventTypeArgs

C++ method: QKeyEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QKeyEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QKeyEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_modifiers(&mut self, amodifiers: Flags<KeyboardModifier>)

C++ method: void QKeyEvent::setModifiers(QFlags<Qt::KeyboardModifier> amodifiers)

fn set_timestamp(&mut self, atimestamp: c_ulong)

C++ method: void QKeyEvent::setTimestamp(unsigned long atimestamp)

fn spontaneous(&self) -> bool

C++ method: bool QKeyEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: KeyEventTextArgs<'largs>

C++ method: QKeyEvent::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QKeyEvent::text() const

C++ documentation:

Returns the Unicode text that this key generated.

Return values when modifier keys such as Shift, Control, Alt, and Meta are pressed differ among platforms and could return an empty string.

Note: key() will always return a valid value, independent of modifier keys.

See also Qt::WA_KeyCompression.

fn timestamp(&self) -> c_ulong

C++ method: unsigned long QKeyEvent::timestamp() const

C++ documentation:

Returns the window system's timestamp for this event.

fn type_(&self) -> Type

C++ method: QEvent::Type QKeyEvent::type() const

C++ documentation:

Returns the event type.