Enum qt_gui::key_sequence::StandardKey

pub enum StandardKey {
    UnknownKey,
    HelpContents,
    WhatsThis,
    Open,
    Close,
    Save,
    New,
    Delete,
    Cut,
    Copy,
    Paste,
    Undo,
    Redo,
    Back,
    Forward,
    Refresh,
    ZoomIn,
    ZoomOut,
    Print,
    AddTab,
    NextChild,
    PreviousChild,
    Find,
    FindNext,
    FindPrevious,
    Replace,
    SelectAll,
    Bold,
    Italic,
    Underline,
    MoveToNextChar,
    MoveToPreviousChar,
    MoveToNextWord,
    MoveToPreviousWord,
    MoveToNextLine,
    MoveToPreviousLine,
    MoveToNextPage,
    MoveToPreviousPage,
    MoveToStartOfLine,
    MoveToEndOfLine,
    MoveToStartOfBlock,
    MoveToEndOfBlock,
    MoveToStartOfDocument,
    MoveToEndOfDocument,
    SelectNextChar,
    SelectPreviousChar,
    SelectNextWord,
    SelectPreviousWord,
    SelectNextLine,
    SelectPreviousLine,
    SelectNextPage,
    SelectPreviousPage,
    SelectStartOfLine,
    SelectEndOfLine,
    SelectStartOfBlock,
    SelectEndOfBlock,
    SelectStartOfDocument,
    SelectEndOfDocument,
    DeleteStartOfWord,
    DeleteEndOfWord,
    DeleteEndOfLine,
    InsertParagraphSeparator,
    InsertLineSeparator,
    SaveAs,
    Preferences,
    Quit,
    FullScreen,
    Deselect,
    DeleteCompleteLine,
    Backspace,
}

C++ type: QKeySequence::StandardKey

C++ documentation:

This enum represent standard key bindings. They can be used to assign platform dependent keyboard shortcuts to a QAction.

Note that the key bindings are platform dependent. The currently bound shortcuts can be queried using keyBindings().

This enum was introduced or modified in Qt 4.2.

Variants
UnknownKey

Unbound key. (C++ enum variant: UnknownKey = 0)

HelpContents

Open help contents. (C++ enum variant: HelpContents = 1)

WhatsThis

Activate "what's this". (C++ enum variant: WhatsThis = 2)

Open

Open document. (C++ enum variant: Open = 3)

Close

Close document/tab. (C++ enum variant: Close = 4)

Save

Save document. (C++ enum variant: Save = 5)

New

Create new document. (C++ enum variant: New = 6)

Delete

Delete. (C++ enum variant: Delete = 7)

Cut

Cut. (C++ enum variant: Cut = 8)

Copy

Copy. (C++ enum variant: Copy = 9)

Paste

Paste. (C++ enum variant: Paste = 10)

Undo

Undo. (C++ enum variant: Undo = 11)

Redo

Redo. (C++ enum variant: Redo = 12)

Back

Navigate back. (C++ enum variant: Back = 13)

Forward

Navigate forward. (C++ enum variant: Forward = 14)

Refresh

Refresh or reload current document. (C++ enum variant: Refresh = 15)

ZoomIn

Zoom in. (C++ enum variant: ZoomIn = 16)

ZoomOut

Zoom out. (C++ enum variant: ZoomOut = 17)

Print

Print document. (C++ enum variant: Print = 18)

AddTab

Add new tab. (C++ enum variant: AddTab = 19)

NextChild

Navigate to next tab or child window. (C++ enum variant: NextChild = 20)

PreviousChild

Navigate to previous tab or child window. (C++ enum variant: PreviousChild = 21)

Find

Find in document. (C++ enum variant: Find = 22)

FindNext

Find next result. (C++ enum variant: FindNext = 23)

FindPrevious

Find previous result. (C++ enum variant: FindPrevious = 24)

Replace

Find and replace. (C++ enum variant: Replace = 25)

SelectAll

Select all text. (C++ enum variant: SelectAll = 26)

Bold

Bold text. (C++ enum variant: Bold = 27)

Italic

Italic text. (C++ enum variant: Italic = 28)

Underline

Underline text. (C++ enum variant: Underline = 29)

MoveToNextChar

Move cursor to next character. (C++ enum variant: MoveToNextChar = 30)

MoveToPreviousChar

Move cursor to previous character. (C++ enum variant: MoveToPreviousChar = 31)

MoveToNextWord

Move cursor to next word. (C++ enum variant: MoveToNextWord = 32)

MoveToPreviousWord

Move cursor to previous word. (C++ enum variant: MoveToPreviousWord = 33)

MoveToNextLine

Move cursor to next line. (C++ enum variant: MoveToNextLine = 34)

MoveToPreviousLine

Move cursor to previous line. (C++ enum variant: MoveToPreviousLine = 35)

MoveToNextPage

Move cursor to next page. (C++ enum variant: MoveToNextPage = 36)

MoveToPreviousPage

Move cursor to previous page. (C++ enum variant: MoveToPreviousPage = 37)

MoveToStartOfLine

Move cursor to start of line. (C++ enum variant: MoveToStartOfLine = 38)

MoveToEndOfLine

Move cursor to end of line. (C++ enum variant: MoveToEndOfLine = 39)

MoveToStartOfBlock

Move cursor to start of a block. This shortcut is only used on OS X. (C++ enum variant: MoveToStartOfBlock = 40)

MoveToEndOfBlock

Move cursor to end of block. This shortcut is only used on the OS X. (C++ enum variant: MoveToEndOfBlock = 41)

MoveToStartOfDocument

Move cursor to start of document. (C++ enum variant: MoveToStartOfDocument = 42)

MoveToEndOfDocument

Move cursor to end of document. (C++ enum variant: MoveToEndOfDocument = 43)

SelectNextChar

Extend selection to next character. (C++ enum variant: SelectNextChar = 44)

SelectPreviousChar

Extend selection to previous character. (C++ enum variant: SelectPreviousChar = 45)

SelectNextWord

Extend selection to next word. (C++ enum variant: SelectNextWord = 46)

SelectPreviousWord

Extend selection to previous word. (C++ enum variant: SelectPreviousWord = 47)

SelectNextLine

Extend selection to next line. (C++ enum variant: SelectNextLine = 48)

SelectPreviousLine

Extend selection to previous line. (C++ enum variant: SelectPreviousLine = 49)

SelectNextPage

Extend selection to next page. (C++ enum variant: SelectNextPage = 50)

SelectPreviousPage

Extend selection to previous page. (C++ enum variant: SelectPreviousPage = 51)

SelectStartOfLine

Extend selection to start of line. (C++ enum variant: SelectStartOfLine = 52)

SelectEndOfLine

Extend selection to end of line. (C++ enum variant: SelectEndOfLine = 53)

SelectStartOfBlock

Extend selection to the start of a text block. This shortcut is only used on OS X. (C++ enum variant: SelectStartOfBlock = 54)

SelectEndOfBlock

Extend selection to the end of a text block. This shortcut is only used on OS X. (C++ enum variant: SelectEndOfBlock = 55)

SelectStartOfDocument

Extend selection to start of document. (C++ enum variant: SelectStartOfDocument = 56)

SelectEndOfDocument

Extend selection to end of document. (C++ enum variant: SelectEndOfDocument = 57)

DeleteStartOfWord

Delete the beginning of a word up to the cursor. (C++ enum variant: DeleteStartOfWord = 58)

DeleteEndOfWord

Delete word from the end of the cursor. (C++ enum variant: DeleteEndOfWord = 59)

DeleteEndOfLine

Delete end of line. (C++ enum variant: DeleteEndOfLine = 60)

InsertParagraphSeparator

Insert a new paragraph. (C++ enum variant: InsertParagraphSeparator = 61)

InsertLineSeparator

Insert a new line. (C++ enum variant: InsertLineSeparator = 62)

SaveAs

Save document after prompting the user for a file name. (C++ enum variant: SaveAs = 63)

Preferences

Open the preferences dialog. (C++ enum variant: Preferences = 64)

Quit

Quit the application. (C++ enum variant: Quit = 65)

FullScreen

Toggle the window state to/from full screen. (C++ enum variant: FullScreen = 66)

Deselect

Deselect text. Since 5.1 (C++ enum variant: Deselect = 67)

DeleteCompleteLine

Delete the entire line. (C++ enum variant: DeleteCompleteLine = 68)

Backspace

Delete previous character. (C++ enum variant: Backspace = 69)