Struct qt_gui::key_sequence::KeySequence

pub struct KeySequence {
    // some fields omitted
}

C++ type: QKeySequence

C++ documentation:

The QKeySequence class encapsulates a key sequence as used by shortcuts.

In its most common form, a key sequence describes a combination of keys that must be used together to perform some action. Key sequences are used with QAction objects to specify which keyboard shortcuts can be used to trigger actions.

Key sequences can be constructed for use as keyboard shortcuts in three different ways:

For standard shortcuts, a standard key can be used to request the platform-specific key sequence associated with each shortcut.

For custom shortcuts, human-readable strings such as "Ctrl+X" can be used, and these can be translated into the appropriate shortcuts for users of different languages. Translations are made in the "QShortcut" context.

For hard-coded shortcuts, integer key codes can be specified with a combination of values defined by the Qt::Key and Qt::Modifier enum values. Each key code consists of a single Qt::Key value and zero or more modifiers, such as Qt::SHIFT, Qt::CTRL, Qt::ALT and Qt::META.

For example, Ctrl P might be a sequence used as a shortcut for printing a document, and can be specified in any of the following ways:

QKeySequence(QKeySequence::Print);
QKeySequence(tr("Ctrl+P"));
QKeySequence(tr("Ctrl+p"));
QKeySequence(Qt::CTRL + Qt::Key_P);

Note that, for letters, the case used in the specification string does not matter. In the above examples, the user does not need to hold down the Shift key to activate a shortcut specified with "Ctrl+P". However, for other keys, the use of Shift as an unspecified extra modifier key can lead to confusion for users of an application whose keyboards have different layouts to those used by the developers. See the Keyboard Layout Issues section below for more details.

It is preferable to use standard shortcuts where possible. When creating key sequences for non-standard shortcuts, you should use human-readable strings in preference to hard-coded integer values.

QKeySequence objects can be cast to a QString to obtain a human-readable translated version of the sequence. Similarly, the toString() function produces human-readable strings for use in menus. On Mac OS X, the appropriate symbols are used to describe keyboard shortcuts using special keys on the Macintosh keyboard.

An alternative way to specify hard-coded key codes is to use the Unicode code point of the character; for example, 'A' gives the same key sequence as Qt::Key_A.

Note: On Mac OS X, references to "Ctrl", Qt::CTRL, Qt::Control and Qt::ControlModifier correspond to the Command keys on the Macintosh keyboard, and references to "Meta", Qt::META, Qt::Meta and Qt::MetaModifier correspond to the Control keys. Developers on Mac OS X can use the same shortcut descriptions across all platforms, and their applications will automatically work as expected on Mac OS X.

Methods

impl KeySequence

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: KeySequenceAsQVariantArgs<'largs>

C++ method: QKeySequence::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QKeySequence::operator QVariant() const

C++ documentation:

Returns the key sequence as a QVariant

fn count(&self) -> c_int

C++ method: int QKeySequence::count() const

C++ documentation:

Returns the number of keys in the key sequence. The maximum is 4.

fn from_string<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceFromStringArgs

C++ method: QKeySequence::fromString

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_string((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
2) fn from_string((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
3) fn from_string((&qt_core::string::String, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
4) fn from_string((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
C++ method: static QKeySequence QKeySequence::fromString(const QString& str, QKeySequence::SequenceFormat format = ?)

C++ documentation:

Return a QKeySequence from the string str based on format.

This function was introduced in Qt 4.1.

See also toString().

fn is_detached(&self) -> bool

C++ method: bool QKeySequence::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QKeySequence::isEmpty() const

C++ documentation:

Returns true if the key sequence is empty; otherwise returns false.

fn key_bindings<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceKeyBindingsArgs

C++ method: QKeySequence::keyBindings

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_bindings((::key_sequence::StandardKey, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListKeySequenceRef>
2) fn key_bindings((::key_sequence::StandardKey, cpp_utils::AsStruct)) -> ::list::ListKeySequenceRef
C++ method: static QList<QKeySequence> QKeySequence::keyBindings(QKeySequence::StandardKey key)

C++ documentation:

Returns a list of key bindings for the given key. The result of calling this function will vary based on the target platform. The first element of the list indicates the primary shortcut for the given platform. If the result contains more than one result, these can be considered alternative shortcuts on the same platform for the given key.

This function was introduced in Qt 4.2.

fn list_from_string<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceListFromStringArgs

C++ method: QKeySequence::listFromString

This is an overloaded function. Available variants:

Rust arguments:
1) fn list_from_string((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListKeySequenceRef>
2) fn list_from_string((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListKeySequenceRef>
3) fn list_from_string((&qt_core::string::String, cpp_utils::AsStruct)) -> ::list::ListKeySequenceRef
4) fn list_from_string((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsStruct)) -> ::list::ListKeySequenceRef
C++ method: static QList<QKeySequence> QKeySequence::listFromString(const QString& str, QKeySequence::SequenceFormat format = ?)

C++ documentation:

Return a list of QKeySequence from the string str based on format.

This function was introduced in Qt 5.1.

See also fromString() and listToString().

fn list_to_string<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceListToStringArgs

C++ method: QKeySequence::listToString

This is an overloaded function. Available variants:

Rust arguments:
1) fn list_to_string((&::list::ListKeySequenceRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn list_to_string((&::list::ListKeySequenceRef, ::key_sequence::SequenceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
3) fn list_to_string((&::list::ListKeySequenceRef, cpp_utils::AsStruct)) -> qt_core::string::String
4) fn list_to_string((&::list::ListKeySequenceRef, ::key_sequence::SequenceFormat, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QKeySequence::listToString(const QList<QKeySequence>& list, QKeySequence::SequenceFormat format = ?)

C++ documentation:

Return a string representation of list based on format.

This function was introduced in Qt 5.1.

See also toString() and listFromString().

fn matches(&self, seq: &KeySequence) -> SequenceMatch

C++ method: QKeySequence::SequenceMatch QKeySequence::matches(const QKeySequence& seq) const

C++ documentation:

Matches the sequence with seq. Returns ExactMatch if successful, PartialMatch if seq matches incompletely, and NoMatch if the sequences have nothing in common. Returns NoMatch if seq is shorter.

fn mnemonic<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceMnemonicArgs

C++ method: QKeySequence::mnemonic

This is an overloaded function. Available variants:

Rust arguments:
1) fn mnemonic((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
2) fn mnemonic((&qt_core::string::String, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
C++ method: static QKeySequence QKeySequence::mnemonic(const QString& text)

C++ documentation:

Returns the shortcut key sequence for the mnemonic in text, or an empty key sequence if no mnemonics are found.

For example, mnemonic("E&xit") returns Qt::ALT+Qt::Key_X, mnemonic("&Quit") returns ALT+Key_Q, and mnemonic("Quit") returns an empty QKeySequence.

We provide a list of common mnemonics in English. At the time of writing, Microsoft and Open Group do not appear to have issued equivalent recommendations for other languages.

fn new<Args>(args: Args) -> Args::ReturnType where Args: KeySequenceNewArgs

C++ method: QKeySequence::QKeySequence

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::key_sequence::KeySequence
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::key_sequence::KeySequence>
C++ method: [constructor] void QKeySequence::QKeySequence()

Variant 2

Rust arguments:
1) fn new((::key_sequence::StandardKey, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
2) fn new((::key_sequence::StandardKey, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
C++ method: [constructor] void QKeySequence::QKeySequence(QKeySequence::StandardKey key)

Variant 3

Rust arguments:
1) fn new((&::key_sequence::KeySequence, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
2) fn new((&::key_sequence::KeySequence, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
C++ method: [constructor] void QKeySequence::QKeySequence(const QKeySequence& ks)

Variant 4

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
2) fn new((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
3) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
4) fn new((&qt_core::string::String, ::key_sequence::SequenceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
C++ method: [constructor] void QKeySequence::QKeySequence(const QString& key, QKeySequence::SequenceFormat format = ?)

Variant 5

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
2) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
3) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
4) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::key_sequence::KeySequence
5) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
6) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
7) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
8) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::key_sequence::KeySequence>
C++ method: [constructor] void QKeySequence::QKeySequence(int k1, int k2 = ?, int k3 = ?, int k4 = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 KeySequence) -> &'l0 mut KeySequence

C++ method: QKeySequence& QKeySequence::operator=(const QKeySequence& other)

C++ documentation:

Assignment operator. Assigns the other key sequence to this object.

fn op_eq(&self, other: &KeySequence) -> bool

C++ method: bool QKeySequence::operator==(const QKeySequence& other) const

C++ documentation:

Returns true if this key sequence is equal to the other key sequence; otherwise returns false.

fn op_ge(&self, other: &KeySequence) -> bool

C++ method: bool QKeySequence::operator>=(const QKeySequence& other) const

C++ documentation:

Returns true if this key sequence is larger or equal to the other key sequence; otherwise returns false.

See also operator==(), operator!=(), operator<(), operator>(), and operator<=().

fn op_gt(&self, other: &KeySequence) -> bool

C++ method: bool QKeySequence::operator>(const QKeySequence& other) const

C++ documentation:

Returns true if this key sequence is larger than the other key sequence; otherwise returns false.

See also operator==(), operator!=(), operator<(), operator<=(), and operator>=().

fn op_index(&self, i: c_uint) -> c_int

C++ method: int QKeySequence::operator[](unsigned int i) const

C++ documentation:

Returns a reference to the element at position index in the key sequence. This can only be used to read an element.

fn op_le(&self, other: &KeySequence) -> bool

C++ method: bool QKeySequence::operator<=(const QKeySequence& other) const

C++ documentation:

Returns true if this key sequence is smaller or equal to the other key sequence; otherwise returns false.

See also operator==(), operator!=(), operator<(), operator>(), and operator>=().

fn op_lt(&self, ks: &KeySequence) -> bool

C++ method: bool QKeySequence::operator<(const QKeySequence& ks) const

C++ documentation:

Provides an arbitrary comparison of this key sequence and other key sequence. All that is guaranteed is that the operator returns false if both key sequences are equal and that (ks1 < ks2) == !( ks2 < ks1) if the key sequences are not equal.

This function is useful in some circumstances, for example if you want to use QKeySequence objects as keys in a QMap.

See also operator==(), operator!=(), operator>(), operator<=(), and operator>=().

fn op_neq(&self, other: &KeySequence) -> bool

C++ method: bool QKeySequence::operator!=(const QKeySequence& other) const

C++ documentation:

Returns true if this key sequence is not equal to the other key sequence; otherwise returns false.

fn swap(&mut self, other: &mut KeySequence)

C++ method: void QKeySequence::swap(QKeySequence& other)

C++ documentation:

Swaps key sequence other with this key sequence. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: KeySequenceToStringArgs<'largs>

C++ method: QKeySequence::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, (::key_sequence::SequenceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
3) fn to_string(&self, (::key_sequence::SequenceFormat, cpp_utils::AsStruct)) -> qt_core::string::String
4) fn to_string(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QKeySequence::toString(QKeySequence::SequenceFormat format = ?) const

C++ documentation:

Return a string representation of the key sequence, based on format.

For example, the value Qt::CTRL+Qt::Key_O results in "Ctrl+O". If the key sequence has multiple key codes, each is separated by commas in the string returned, such as "Alt+X, Ctrl+Y, Z". The strings, "Ctrl", "Shift", etc. are translated using QObject::tr() in the "QShortcut" context.

If the key sequence has no keys, an empty string is returned.

On Mac OS X, the string returned resembles the sequence that is shown in the menu bar.

This function was introduced in Qt 4.1.

See also fromString().