Module qt_gui::matrix::overloading

Traits

MatrixAsQVariantArgs
This trait represents a set of arguments accepted by Matrix::as_q_variant method.

MatrixInvertedArgs
This trait represents a set of arguments accepted by Matrix::inverted method.

MatrixMapArgs
This trait represents a set of arguments accepted by Matrix::map method.

MatrixMapRectArgs
This trait represents a set of arguments accepted by Matrix::map_rect method.

MatrixMapToPolygonArgs
This trait represents a set of arguments accepted by Matrix::map_to_polygon method.

MatrixNewArgs
This trait represents a set of arguments accepted by Matrix::new method.

MatrixOpMulArgs
This trait represents a set of arguments accepted by Matrix::op_mul method.

OpMulArgs
This trait represents a set of arguments accepted by op_mul method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.