Struct qt_gui::matrix::Matrix

pub struct Matrix {
    // some fields omitted
}

C++ type: QMatrix

C++ documentation:

The QMatrix class specifies 2D transformations of a coordinate system.

A matrix specifies how to translate, scale, shear or rotate the coordinate system, and is typically used when rendering graphics. QMatrix, in contrast to QTransform, does not allow perspective transformations. QTransform is the recommended transformation class in Qt.

A QMatrix object can be built using the setMatrix(), scale(), rotate(), translate() and shear() functions. Alternatively, it can be built by applying basic matrix operations. The matrix can also be defined when constructed, and it can be reset to the identity matrix (the default) using the reset() function.

The QMatrix class supports mapping of graphic primitives: A given point, line, polygon, region, or painter path can be mapped to the coordinate system defined by this matrix using the map() function. In case of a rectangle, its coordinates can be transformed using the mapRect() function. A rectangle can also be transformed into a polygon (mapped to the coordinate system defined by this matrix), using the mapToPolygon() function.

QMatrix provides the isIdentity() function which returns true if the matrix is the identity matrix, and the isInvertible() function which returns true if the matrix is non-singular (i.e. AB = BA = I). The inverted() function returns an inverted copy of this matrix if it is invertible (otherwise it returns the identity matrix). In addition, QMatrix provides the determinant() function returning the matrix's determinant.

Finally, the QMatrix class supports matrix multiplication, and objects of the class can be streamed as well as compared.

Methods

impl Matrix

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixAsQVariantArgs<'largs>

C++ method: QMatrix::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QMatrix::operator QVariant() const

C++ documentation:

Returns the matrix as a QVariant.

This function was introduced in Qt 4.2.

fn determinant(&self) -> c_double

C++ method: double QMatrix::determinant() const

C++ documentation:

Returns the matrix's determinant.

This function was introduced in Qt 4.6.

fn dx(&self) -> c_double

C++ method: double QMatrix::dx() const

C++ documentation:

Returns the horizontal translation factor.

See also translate() and Basic Matrix Operations.

fn dy(&self) -> c_double

C++ method: double QMatrix::dy() const

C++ documentation:

Returns the vertical translation factor.

See also translate() and Basic Matrix Operations.

fn inverted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixInvertedArgs<'largs>

C++ method: QMatrix::inverted

This is an overloaded function. Available variants:

Rust arguments:
1) fn inverted(&self, (*mut bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::matrix::Matrix>
2) fn inverted(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::matrix::Matrix>
3) fn inverted(&self, (*mut bool, cpp_utils::AsStruct)) -> ::matrix::Matrix
4) fn inverted(&self, cpp_utils::AsStruct) -> ::matrix::Matrix
C++ method: QMatrix QMatrix::inverted(bool* invertible = ?) const

C++ documentation:

Returns an inverted copy of this matrix.

If the matrix is singular (not invertible), the returned matrix is the identity matrix. If invertible is valid (i.e. not 0), its value is set to true if the matrix is invertible, otherwise it is set to false.

See also isInvertible().

fn is_identity(&self) -> bool

C++ method: bool QMatrix::isIdentity() const

C++ documentation:

Returns true if the matrix is the identity matrix, otherwise returns false.

See also reset().

fn is_invertible(&self) -> bool

C++ method: bool QMatrix::isInvertible() const

C++ documentation:

Returns true if the matrix is invertible, otherwise returns false.

See also inverted().

fn m11(&self) -> c_double

C++ method: double QMatrix::m11() const

C++ documentation:

Returns the horizontal scaling factor.

See also scale() and Basic Matrix Operations.

fn m12(&self) -> c_double

C++ method: double QMatrix::m12() const

C++ documentation:

Returns the vertical shearing factor.

See also shear() and Basic Matrix Operations.

fn m21(&self) -> c_double

C++ method: double QMatrix::m21() const

C++ documentation:

Returns the horizontal shearing factor.

See also shear() and Basic Matrix Operations.

fn m22(&self) -> c_double

C++ method: double QMatrix::m22() const

C++ documentation:

Returns the vertical scaling factor.

See also scale() and Basic Matrix Operations.

fn map<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixMapArgs<'largs>

C++ method: QMatrix::map

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map(&self, (&qt_core::line::Line, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::line::Line>
2) fn map(&self, (&qt_core::line::Line, cpp_utils::AsStruct)) -> qt_core::line::Line
C++ method: QLine QMatrix::map(const QLine& l) const

C++ documentation:

This is an overloaded function.

Creates and returns a QLine object that is a copy of the given line, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 2

Rust arguments:
1) fn map(&self, (&qt_core::line_f::LineF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::line_f::LineF>
2) fn map(&self, (&qt_core::line_f::LineF, cpp_utils::AsStruct)) -> qt_core::line_f::LineF
C++ method: QLineF QMatrix::map(const QLineF& l) const

C++ documentation:

This is an overloaded function.

Creates and returns a QLineF object that is a copy of the given line, mapped into the coordinate system defined by this matrix.

Variant 3

Rust arguments:
1) fn map(&self, (&::painter_path::PainterPath, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter_path::PainterPath>
2) fn map(&self, (&::painter_path::PainterPath, cpp_utils::AsStruct)) -> ::painter_path::PainterPath
C++ method: QPainterPath QMatrix::map(const QPainterPath& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPainterPath object that is a copy of the given path, mapped into the coordinate system defined by this matrix.

Variant 4

Rust arguments:
1) fn map(&self, (&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn map(&self, (&qt_core::point::Point, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QMatrix::map(const QPoint& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPoint object that is a copy of the given point, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 5

Rust arguments:
1) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QMatrix::map(const QPointF& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPointF object that is a copy of the given point, mapped into the coordinate system defined by this matrix.

Variant 6

Rust arguments:
1) fn map(&self, (&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn map(&self, (&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QMatrix::map(const QPolygon& a) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPolygon object that is a copy of the given polygon, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 7

Rust arguments:
1) fn map(&self, (&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn map(&self, (&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QMatrix::map(const QPolygonF& a) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPolygonF object that is a copy of the given polygon, mapped into the coordinate system defined by this matrix.

Variant 8

Rust arguments:
1) fn map(&self, (&::region::Region, cpp_utils::AsBox)) -> cpp_utils::CppBox<::region::Region>
2) fn map(&self, (&::region::Region, cpp_utils::AsStruct)) -> ::region::Region
C++ method: QRegion QMatrix::map(const QRegion& r) const

C++ documentation:

This is an overloaded function.

Creates and returns a QRegion object that is a copy of the given region, mapped into the coordinate system defined by this matrix.

Calling this method can be rather expensive if rotations or shearing are used.

Variant 9

Rust arguments: fn map(&self, (libc::c_double, libc::c_double, *mut libc::c_double, *mut libc::c_double)) -> ()
C++ method: void QMatrix::map(double x, double y, double* tx, double* ty) const

C++ documentation:

Maps the given coordinates x and y into the coordinate system defined by this matrix. The resulting values are put in *tx and *ty, respectively.

The coordinates are transformed using the following formulas:

x' = m11*x + m21*y + dx
y' = m22y + m12x + dy

The point (x, y) is the original point, and (x', y') is the transformed point.

See also Basic Matrix Operations.

Variant 10

Rust arguments: fn map(&self, (libc::c_int, libc::c_int, *mut libc::c_int, *mut libc::c_int)) -> ()
C++ method: void QMatrix::map(int x, int y, int* tx, int* ty) const

C++ documentation:

This is an overloaded function.

Maps the given coordinates x and y into the coordinate system defined by this matrix. The resulting values are put in *tx and *ty, respectively. Note that the transformed coordinates are rounded to the nearest integer.

fn map_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixMapRectArgs<'largs>

C++ method: QMatrix::mapRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QMatrix::mapRect(const QRect& arg1) const

C++ documentation:

This is an overloaded function.

Creates and returns a QRect object that is a copy of the given rectangle, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 2

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QMatrix::mapRect(const QRectF& arg1) const

C++ documentation:

Creates and returns a QRectF object that is a copy of the given rectangle, mapped into the coordinate system defined by this matrix.

The rectangle's coordinates are transformed using the following formulas:

x' = m11*x + m21*y + dx
y' = m22y + m12x + dy

If rotation or shearing has been specified, this function returns the bounding rectangle. To retrieve the exact region the given rectangle maps to, use the mapToPolygon() function instead.

See also mapToPolygon() and Basic Matrix Operations.

fn map_to_polygon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixMapToPolygonArgs<'largs>

C++ method: QMatrix::mapToPolygon

This is an overloaded function. Available variants:

Rust arguments:
1) fn map_to_polygon(&self, (&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn map_to_polygon(&self, (&qt_core::rect::Rect, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QMatrix::mapToPolygon(const QRect& r) const

C++ documentation:

Creates and returns a QPolygon representation of the given rectangle, mapped into the coordinate system defined by this matrix.

The rectangle's coordinates are transformed using the following formulas:

x' = m11*x + m21*y + dx
y' = m22y + m12x + dy

Polygons and rectangles behave slightly differently when transformed (due to integer rounding), so matrix.map(QPolygon(rectangle)) is not always the same as matrix.mapToPolygon(rectangle).

See also mapRect() and Basic Matrix Operations.

fn new<Args>(args: Args) -> Args::ReturnType where Args: MatrixNewArgs

C++ method: QMatrix::QMatrix

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::matrix::Matrix
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::matrix::Matrix>
C++ method: [constructor] void QMatrix::QMatrix()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::matrix::Matrix
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::matrix::Matrix>
C++ method: [constructor] void QMatrix::QMatrix(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&::matrix::Matrix, cpp_utils::AsStruct)) -> ::matrix::Matrix
2) fn new((&::matrix::Matrix, cpp_utils::AsBox)) -> cpp_utils::CppBox<::matrix::Matrix>
C++ method: [constructor] void QMatrix::QMatrix(const QMatrix& matrix)

Variant 4

Rust arguments:
1) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::matrix::Matrix
2) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::matrix::Matrix>
C++ method: [constructor] void QMatrix::QMatrix(double m11, double m12, double m21, double m22, double dx, double dy)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Matrix) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::operator=(const QMatrix& arg1)

C++ documentation:

Assigns the given matrix's values to this matrix.

fn op_eq(&self, arg1: &Matrix) -> bool

C++ method: bool QMatrix::operator==(const QMatrix& arg1) const

C++ documentation:

Returns true if this matrix is equal to the given matrix, otherwise returns false.

fn op_mul<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MatrixOpMulArgs<'largs>

C++ method: QMatrix::operator*

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_mul(&self, (&::matrix::Matrix, cpp_utils::AsBox)) -> cpp_utils::CppBox<::matrix::Matrix>
2) fn op_mul(&self, (&::matrix::Matrix, cpp_utils::AsStruct)) -> ::matrix::Matrix
C++ method: QMatrix QMatrix::operator*(const QMatrix& o) const

C++ documentation:

Returns the result of multiplying this matrix by the given matrix.

Note that matrix multiplication is not commutative, i.e. a*b != b*a.

fn op_mul_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Matrix) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::operator*=(const QMatrix& arg1)

C++ documentation:

This is an overloaded function.

Returns the result of multiplying this matrix by the given matrix.

fn op_neq(&self, arg1: &Matrix) -> bool

C++ method: bool QMatrix::operator!=(const QMatrix& arg1) const

C++ documentation:

Returns true if this matrix is not equal to the given matrix, otherwise returns false.

fn reset(&mut self)

C++ method: void QMatrix::reset()

C++ documentation:

Resets the matrix to an identity matrix, i.e. all elements are set to zero, except m11 and m22 (specifying the scale) which are set to 1.

See also QMatrix(), isIdentity(), and Basic Matrix Operations.

fn rotate<'l0>(&'l0 mut self, a: c_double) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::rotate(double a)

C++ documentation:

Rotates the coordinate system the given degrees counterclockwise.

Note that if you apply a QMatrix to a point defined in widget coordinates, the direction of the rotation will be clockwise because the y-axis points downwards.

Returns a reference to the matrix.

See also setMatrix().

fn scale<'l0>(&'l0 mut self, sx: c_double, sy: c_double) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::scale(double sx, double sy)

C++ documentation:

Scales the coordinate system by sx horizontally and sy vertically, and returns a reference to the matrix.

See also setMatrix().

fn set_matrix(&mut self, m11: c_double, m12: c_double, m21: c_double, m22: c_double, dx: c_double, dy: c_double)

C++ method: void QMatrix::setMatrix(double m11, double m12, double m21, double m22, double dx, double dy)

C++ documentation:

Sets the matrix elements to the specified values, m11, m12, m21, m22, dx and dy.

Note that this function replaces the previous values. QMatrix provide the translate(), rotate(), scale() and shear() convenience functions to manipulate the various matrix elements based on the currently defined coordinate system.

See also QMatrix().

fn shear<'l0>(&'l0 mut self, sh: c_double, sv: c_double) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::shear(double sh, double sv)

C++ documentation:

Shears the coordinate system by sh horizontally and sv vertically, and returns a reference to the matrix.

See also setMatrix().

fn translate<'l0>(&'l0 mut self, dx: c_double, dy: c_double) -> &'l0 mut Matrix

C++ method: QMatrix& QMatrix::translate(double dx, double dy)

C++ documentation:

Moves the coordinate system dx along the x axis and dy along the y axis, and returns a reference to the matrix.

See also setMatrix().