Struct qt_gui::mouse_event::MouseEvent

pub struct MouseEvent {
    // some fields omitted
}

C++ type: QMouseEvent

C++ documentation:

The QMouseEvent class contains parameters that describe a mouse event.

Mouse events occur when a mouse button is pressed or released inside a widget, or when the mouse cursor is moved.

Mouse move events will occur only when a mouse button is pressed down, unless mouse tracking has been enabled with QWidget::setMouseTracking().

Qt automatically grabs the mouse when a mouse button is pressed inside a widget; the widget will continue to receive mouse events until the last mouse button is released.

A mouse event contains a special accept flag that indicates whether the receiver wants the event. You should call ignore() if the mouse event is not handled by your widget. A mouse event is propagated up the parent widget chain until a widget accepts it with accept(), or an event filter consumes it.

Note: If a mouse event is propagated to a widget for which Qt::WA_NoMousePropagation has been set, that mouse event will not be propagated further up the parent widget chain.

The state of the keyboard modifier keys can be found by calling the modifiers() function, inherited from QInputEvent.

The functions pos(), x(), and y() give the cursor position relative to the widget that receives the mouse event. If you move the widget as a result of the mouse event, use the global position returned by globalPos() to avoid a shaking motion.

The QWidget::setEnabled() function can be used to enable or disable mouse and keyboard events for a widget.

Reimplement the QWidget event handlers, QWidget::mousePressEvent(), QWidget::mouseReleaseEvent(), QWidget::mouseDoubleClickEvent(), and QWidget::mouseMoveEvent() to receive mouse events in your own widgets.

Methods

impl MouseEvent

fn accept(&mut self)

C++ method: void QMouseEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn button(&self) -> MouseButton

C++ method: Qt::MouseButton QMouseEvent::button() const

C++ documentation:

Returns the button that caused the event.

Note that the returned value is always Qt::NoButton for mouse move events.

See also buttons() and Qt::MouseButton.

fn buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QMouseEvent::buttons() const

C++ documentation:

Returns the button state when the event was generated. The button state is a combination of Qt::LeftButton, Qt::RightButton, Qt::MidButton using the OR operator. For mouse move events, this is all buttons that are pressed down. For mouse press and double click events this includes the button that caused the event. For mouse release events this excludes the button that caused the event.

See also button() and Qt::MouseButton.

fn flags(&self) -> Flags<MouseEventFlag>

C++ method: QFlags<Qt::MouseEventFlag> QMouseEvent::flags() const

C++ documentation:

Returns the mouse event flags.

The mouse event flags provide additional information about a mouse event.

This function was introduced in Qt 5.3.

See also Qt::MouseEventFlag and QGraphicsSceneMouseEvent::flags().

fn global_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MouseEventGlobalPosArgs<'largs>

C++ method: QMouseEvent::globalPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn global_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn global_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QMouseEvent::globalPos() const

C++ documentation:

Returns the global position of the mouse cursor at the time of the event. This is important on asynchronous window systems like X11. Whenever you move your widgets around in response to mouse events, globalPos() may differ a lot from the current pointer position QCursor::pos(), and from QWidget::mapToGlobal(pos()).

See also globalX() and globalY().

fn global_x(&self) -> c_int

C++ method: int QMouseEvent::globalX() const

C++ documentation:

Returns the global x position of the mouse cursor at the time of the event.

See also globalY() and globalPos().

fn global_y(&self) -> c_int

C++ method: int QMouseEvent::globalY() const

C++ documentation:

Returns the global y position of the mouse cursor at the time of the event.

See also globalX() and globalPos().

fn ignore(&mut self)

C++ method: void QMouseEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QMouseEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn local_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QMouseEvent::localPos() const

C++ documentation:

Returns the position of the mouse cursor as a QPointF, relative to the widget or item that received the event.

If you move the widget as a result of the mouse event, use the screen position returned by screenPos() to avoid a shaking motion.

This function was introduced in Qt 5.0.

See also x(), y(), windowPos(), and screenPos().

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QMouseEvent::modifiers() const

C++ documentation:

Returns the keyboard modifier flags that existed immediately before the event occurred.

See also QGuiApplication::keyboardModifiers().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MouseEventNewArgs

C++ method: QMouseEvent::QMouseEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((qt_core::event::Type, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::mouse_event::MouseEvent
2) fn new((qt_core::event::Type, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mouse_event::MouseEvent>
C++ method: [constructor] void QMouseEvent::QMouseEvent(QEvent::Type type, const QPointF& localPos, Qt::MouseButton button, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers)

Variant 2

Rust arguments:
1) fn new((qt_core::event::Type, &qt_core::point_f::PointF, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::mouse_event::MouseEvent
2) fn new((qt_core::event::Type, &qt_core::point_f::PointF, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mouse_event::MouseEvent>
C++ method: [constructor] void QMouseEvent::QMouseEvent(QEvent::Type type, const QPointF& localPos, const QPointF& screenPos, Qt::MouseButton button, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers)

Variant 3

Rust arguments:
1) fn new((qt_core::event::Type, &qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::mouse_event::MouseEvent
2) fn new((qt_core::event::Type, &qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point_f::PointF, qt_core::qt::MouseButton, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mouse_event::MouseEvent>
C++ method: [constructor] void QMouseEvent::QMouseEvent(QEvent::Type type, const QPointF& localPos, const QPointF& windowPos, const QPointF& screenPos, Qt::MouseButton button, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MouseEventPosArgs<'largs>

C++ method: QMouseEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QMouseEvent::pos() const

C++ documentation:

Returns the position of the mouse cursor, relative to the widget that received the event.

If you move the widget as a result of the mouse event, use the global position returned by globalPos() to avoid a shaking motion.

See also x(), y(), and globalPos().

fn register_event_type<Args>(args: Args) -> c_int where Args: MouseEventRegisterEventTypeArgs

C++ method: QMouseEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QMouseEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn screen_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QMouseEvent::screenPos() const

C++ documentation:

Returns the position of the mouse cursor as a QPointF, relative to the screen that received the event.

This function was introduced in Qt 5.0.

See also x(), y(), pos(), localPos(), and windowPos().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QMouseEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_modifiers(&mut self, amodifiers: Flags<KeyboardModifier>)

C++ method: void QMouseEvent::setModifiers(QFlags<Qt::KeyboardModifier> amodifiers)

fn set_timestamp(&mut self, atimestamp: c_ulong)

C++ method: void QMouseEvent::setTimestamp(unsigned long atimestamp)

fn source(&self) -> MouseEventSource

C++ method: Qt::MouseEventSource QMouseEvent::source() const

C++ documentation:

Returns information about the mouse event source.

The mouse event source can be used to distinguish between genuine and artificial mouse events. The latter are events that are synthesized from touch events by the operating system or Qt itself.

Note: Many platforms provide no such information. On such platforms Qt::MouseEventNotSynthesized is returned always.

This function was introduced in Qt 5.3.

See also Qt::MouseEventSource and QGraphicsSceneMouseEvent::source().

fn spontaneous(&self) -> bool

C++ method: bool QMouseEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn timestamp(&self) -> c_ulong

C++ method: unsigned long QMouseEvent::timestamp() const

C++ documentation:

Returns the window system's timestamp for this event.

fn type_(&self) -> Type

C++ method: QEvent::Type QMouseEvent::type() const

C++ documentation:

Returns the event type.

fn window_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QMouseEvent::windowPos() const

C++ documentation:

Returns the position of the mouse cursor as a QPointF, relative to the window that received the event.

If you move the widget as a result of the mouse event, use the global position returned by globalPos() to avoid a shaking motion.

This function was introduced in Qt 5.0.

See also x(), y(), pos(), localPos(), and screenPos().

fn x(&self) -> c_int

C++ method: int QMouseEvent::x() const

C++ documentation:

Returns the x position of the mouse cursor, relative to the widget that received the event.

See also y() and pos().

fn y(&self) -> c_int

C++ method: int QMouseEvent::y() const

C++ documentation:

Returns the y position of the mouse cursor, relative to the widget that received the event.

See also x() and pos().