Module qt_gui::movie::overloading

Traits

MovieBackgroundColorArgs
This trait represents a set of arguments accepted by Movie::background_color method.

MovieConnectArgs
This trait represents a set of arguments accepted by Movie::connect method.

MovieConnectStaticArgs
This trait represents a set of arguments accepted by Movie::connect_static method.

MovieCurrentImageArgs
This trait represents a set of arguments accepted by Movie::current_image method.

MovieCurrentPixmapArgs
This trait represents a set of arguments accepted by Movie::current_pixmap method.

MovieDestroyedArgs
This trait represents a set of arguments accepted by Movie::destroyed method.

MovieDisconnectArgs
This trait represents a set of arguments accepted by Movie::disconnect method.

MovieDisconnectStaticArgs
This trait represents a set of arguments accepted by Movie::disconnect_static method.

MovieDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Movie::dynamic_property_names method.

MovieFileNameArgs
This trait represents a set of arguments accepted by Movie::file_name method.

MovieFormatArgs
This trait represents a set of arguments accepted by Movie::format method.

MovieFrameRectArgs
This trait represents a set of arguments accepted by Movie::frame_rect method.

MovieNewArgs
This trait represents a set of arguments accepted by Movie::new method.

MovieObjectNameArgs
This trait represents a set of arguments accepted by Movie::object_name method.

MoviePropertyArgs
This trait represents a set of arguments accepted by Movie::property method.

MovieScaledSizeArgs
This trait represents a set of arguments accepted by Movie::scaled_size method.

MovieStartTimerArgs
This trait represents a set of arguments accepted by Movie::start_timer method.

MovieSupportedFormatsArgs
This trait represents a set of arguments accepted by Movie::supported_formats method.

MovieTrArgs
This trait represents a set of arguments accepted by Movie::tr method.

MovieTrUtf8Args
This trait represents a set of arguments accepted by Movie::tr_utf8 method.