Struct qt_gui::movie::Movie

pub struct Movie {
    // some fields omitted
}

C++ type: QMovie

C++ documentation:

The QMovie class is a convenience class for playing movies with QImageReader.

This class is used to show simple animations without sound. If you want to display video and media content, use the Qt Multimedia multimedia framework instead.

First, create a QMovie object by passing either the name of a file or a pointer to a QIODevice containing an animated image format to QMovie's constructor. You can call isValid() to check if the image data is valid, before starting the movie. To start the movie, call start(). QMovie will enter Running state, and emit started() and stateChanged(). To get the current state of the movie, call state().

To display the movie in your application, you can pass your QMovie object to QLabel::setMovie(). Example:

QLabel label;
QMovie *movie = new QMovie("animations/fire.gif");

label.setMovie(movie); movie->start();
Whenever a new frame is available in the movie, QMovie will emit updated(). If the size of the frame changes, resized() is emitted. You can call currentImage() or currentPixmap() to get a copy of the current frame. When the movie is done, QMovie emits finished(). If any error occurs during playback (i.e, the image file is corrupt), QMovie will emit error().

You can control the speed of the movie playback by calling setSpeed(), which takes the percentage of the original speed as an argument. Pause the movie by calling setPaused(true). QMovie will then enter Paused state and emit stateChanged(). If you call setPaused(false), QMovie will reenter Running state and start the movie again. To stop the movie, call stop().

Certain animation formats allow you to set the background color. You can call setBackgroundColor() to set the color, or backgroundColor() to retrieve the current background color.

currentFrameNumber() returns the sequence number of the current frame. The first frame in the animation has the sequence number 0. frameCount() returns the total number of frames in the animation, if the image format supports this. You can call loopCount() to get the number of times the movie should loop before finishing. nextFrameDelay() returns the number of milliseconds the current frame should be displayed.

QMovie can be instructed to cache frames of an animation by calling setCacheMode().

Call supportedFormats() for a list of formats that QMovie supports.

Methods

impl Movie

fn background_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieBackgroundColorArgs<'largs>

C++ method: QMovie::backgroundColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn background_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: QColor QMovie::backgroundColor() const

C++ documentation:

Returns the background color of the movie. If no background color has been assigned, an invalid QColor is returned.

See also setBackgroundColor().

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QMovie::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn cache_mode(&self) -> CacheMode

C++ method: QMovie::CacheMode QMovie::cacheMode() const

C++ documentation:

This property holds the movie's cache mode.

Caching frames can be useful when the underlying animation format handler that QMovie relies on to decode the animation data does not support jumping to particular frames in the animation, or even "rewinding" the animation to the beginning (for looping). Furthermore, if the image data comes from a sequential device, it is not possible for the underlying animation handler to seek back to frames whose data has already been read (making looping altogether impossible).

To aid in such situations, a QMovie object can be instructed to cache the frames, at the added memory cost of keeping the frames in memory for the lifetime of the object.

By default, this property is set to CacheNone.

Access functions:

CacheMode 	cacheMode() const

void 	setCacheMode(CacheMode mode)

See also QMovie::CacheMode.

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QMovie::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieConnectArgs<'largs>

C++ method: QMovie::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QMovie::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: MovieConnectStaticArgs

C++ method: QMovie::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QMovie::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QMovie::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn current_frame_number(&self) -> c_int

C++ method: int QMovie::currentFrameNumber() const

C++ documentation:

Returns the sequence number of the current frame. The number of the first frame in the movie is 0.

fn current_image<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieCurrentImageArgs<'largs>

C++ method: QMovie::currentImage

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_image(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::image::Image>
2) fn current_image(&self, cpp_utils::AsStruct) -> ::image::Image
C++ method: QImage QMovie::currentImage() const

C++ documentation:

Returns the current frame as a QImage.

See also currentPixmap() and updated().

fn current_pixmap<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieCurrentPixmapArgs<'largs>

C++ method: QMovie::currentPixmap

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_pixmap(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pixmap::Pixmap>
2) fn current_pixmap(&self, cpp_utils::AsStruct) -> ::pixmap::Pixmap
C++ method: QPixmap QMovie::currentPixmap() const

C++ documentation:

Returns the current frame as a QPixmap.

See also currentImage() and updated().

fn delete_later(&mut self)

C++ method: void QMovie::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: MovieDestroyedArgs<'largs>

C++ method: QMovie::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QMovie::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QMovie::device() const

C++ documentation:

Returns the device QMovie reads image data from. If no device has currently been assigned, 0 is returned.

See also setDevice() and fileName().

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: MovieDisconnectArgs<'largs>

C++ method: QMovie::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QMovie::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QMovie::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: MovieDisconnectStaticArgs

C++ method: QMovie::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QMovie::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QMovie::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QMovie::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QMovie::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QMovie::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieDynamicPropertyNamesArgs<'largs>

C++ method: QMovie::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QMovie::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn error(&mut self, error: ImageReaderError)

C++ method: void QMovie::error(QImageReader::ImageReaderError error)

C++ documentation:

This signal is emitted by QMovie when the error error occurred during playback. QMovie will stop the movie, and enter QMovie::NotRunning state.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QMovie::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QMovie::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieFileNameArgs<'largs>

C++ method: QMovie::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QMovie::fileName() const

C++ documentation:

Returns the name of the file that QMovie reads image data from. If no file name has been assigned, or if the assigned device is not a file, an empty QString is returned.

See also setFileName() and device().

fn finished(&mut self)

C++ method: void QMovie::finished()

C++ documentation:

This signal is emitted when the movie has finished.

See also QMovie::stop().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieFormatArgs<'largs>

C++ method: QMovie::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn format(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QMovie::format() const

C++ documentation:

Returns the format that QMovie uses when decoding image data. If no format has been assigned, an empty QByteArray() is returned.

See also setFormat().

fn frame_changed(&mut self, frame_number: c_int)

C++ method: void QMovie::frameChanged(int frameNumber)

C++ documentation:

This signal is emitted when the frame number has changed to frameNumber. You can call currentImage() or currentPixmap() to get a copy of the frame.

This function was introduced in Qt 4.1.

fn frame_count(&self) -> c_int

C++ method: int QMovie::frameCount() const

C++ documentation:

Returns the number of frames in the movie.

Certain animation formats do not support this feature, in which case 0 is returned.

fn frame_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieFrameRectArgs<'largs>

C++ method: QMovie::frameRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn frame_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn frame_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QMovie::frameRect() const

C++ documentation:

Returns the rect of the last frame. If no frame has yet been updated, an invalid QRect is returned.

See also currentImage() and currentPixmap().

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QMovie::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QMovie::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_valid(&self) -> bool

C++ method: bool QMovie::isValid() const

C++ documentation:

Returns true if the movie is valid (e.g., the image data is readable and the image format is supported); otherwise returns false.

fn is_widget_type(&self) -> bool

C++ method: bool QMovie::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QMovie::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn jump_to_frame(&mut self, frame_number: c_int) -> bool

C++ method: bool QMovie::jumpToFrame(int frameNumber)

C++ documentation:

Jumps to frame number frameNumber. Returns true on success; otherwise returns false.

fn jump_to_next_frame(&mut self) -> bool

C++ method: bool QMovie::jumpToNextFrame()

C++ documentation:

Jumps to the next frame. Returns true on success; otherwise returns false.

fn kill_timer(&mut self, id: c_int)

C++ method: void QMovie::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn loop_count(&self) -> c_int

C++ method: int QMovie::loopCount() const

C++ documentation:

Returns the number of times the movie will loop before it finishes. If the movie will only play once (no looping), loopCount returns 0. If the movie loops forever, loopCount returns -1.

Note that, if the image data comes from a sequential device (e.g. a socket), QMovie can only loop the movie if the cacheMode is set to QMovie::CacheAll.

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QMovie::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QMovie::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MovieNewArgs

C++ method: QMovie::QMovie

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::movie::Movie
2) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::movie::Movie
3) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, *mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::movie::Movie
4) fn new((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
5) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
6) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, *mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
C++ method: [constructor] void QMovie::QMovie(QIODevice* device, const QByteArray& format = ?, QObject* parent = ?)

Variant 2

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::movie::Movie
2) fn new((*mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::movie::Movie
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::movie::Movie>
4) fn new((*mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
C++ method: [constructor] void QMovie::QMovie(QObject* parent = ?)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::movie::Movie
2) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::movie::Movie
3) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, *mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::movie::Movie
4) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
5) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
6) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, *mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::movie::Movie>
C++ method: [constructor] void QMovie::QMovie(const QString& fileName, const QByteArray& format = ?, QObject* parent = ?)

fn next_frame_delay(&self) -> c_int

C++ method: int QMovie::nextFrameDelay() const

C++ documentation:

Returns the number of milliseconds QMovie will wait before updating the next frame in the animation.

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MovieObjectNameArgs<'largs>

C++ method: QMovie::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QMovie::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QMovie::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MoviePropertyArgs<'largs>

C++ method: QMovie::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QMovie::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QMovie::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QMovie::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QMovie::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QMovie::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn resized(&mut self, size: &Size)

C++ method: void QMovie::resized(const QSize& size)

C++ documentation:

This signal is emitted when the current frame has been resized to size. This effect is sometimes used in animations as an alternative to replacing the frame. You can call currentImage() or currentPixmap() to get a copy of the updated frame.

fn scaled_size<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: MovieScaledSizeArgs<'largs>

C++ method: QMovie::scaledSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn scaled_size(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn scaled_size(&mut self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QMovie::scaledSize()

C++ documentation:

Returns the scaled size of frames.

This function was introduced in Qt 4.1.

See also setScaledSize() and QImageReader::scaledSize().

fn set_background_color(&mut self, color: &Color)

C++ method: void QMovie::setBackgroundColor(const QColor& color)

C++ documentation:

For image formats that support it, this function sets the background color to color.

See also backgroundColor().

fn set_cache_mode(&mut self, mode: CacheMode)

C++ method: void QMovie::setCacheMode(QMovie::CacheMode mode)

C++ documentation:

This property holds the movie's cache mode.

Caching frames can be useful when the underlying animation format handler that QMovie relies on to decode the animation data does not support jumping to particular frames in the animation, or even "rewinding" the animation to the beginning (for looping). Furthermore, if the image data comes from a sequential device, it is not possible for the underlying animation handler to seek back to frames whose data has already been read (making looping altogether impossible).

To aid in such situations, a QMovie object can be instructed to cache the frames, at the added memory cost of keeping the frames in memory for the lifetime of the object.

By default, this property is set to CacheNone.

Access functions:

CacheMode 	cacheMode() const

void 	setCacheMode(CacheMode mode)

See also QMovie::CacheMode.

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QMovie::setDevice(QIODevice* device)

C++ documentation:

Sets the current device to device. QMovie will read image data from this device when the movie is running.

See also device() and setFormat().

fn set_file_name(&mut self, file_name: &String)

C++ method: void QMovie::setFileName(const QString& fileName)

C++ documentation:

Sets the name of the file that QMovie reads image data from, to fileName.

See also fileName(), setDevice(), and setFormat().

fn set_format(&mut self, format: &ByteArray)

C++ method: void QMovie::setFormat(const QByteArray& format)

C++ documentation:

Sets the format that QMovie will use when decoding image data, to format. By default, QMovie will attempt to guess the format of the image data.

You can call supportedFormats() for the full list of formats QMovie supports.

See also format() and QImageReader::supportedImageFormats().

fn set_object_name(&mut self, name: &String)

C++ method: void QMovie::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QMovie::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_paused(&mut self, paused: bool)

C++ method: void QMovie::setPaused(bool paused)

C++ documentation:

If paused is true, QMovie will enter Paused state and emit stateChanged(Paused); otherwise it will enter Running state and emit stateChanged(Running).

See also state().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QMovie::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_scaled_size(&mut self, size: &Size)

C++ method: void QMovie::setScaledSize(const QSize& size)

C++ documentation:

Sets the scaled frame size to size.

This function was introduced in Qt 4.1.

See also scaledSize() and QImageReader::setScaledSize().

fn set_speed(&mut self, percent_speed: c_int)

C++ method: void QMovie::setSpeed(int percentSpeed)

C++ documentation:

This property holds the movie's speed.

The speed is measured in percentage of the original movie speed. The default speed is 100%. Example:

QMovie movie("racecar.gif");
movie.setSpeed(200); // 2x speed

Access functions:

int 	speed() const

void 	setSpeed(int percentSpeed)

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QMovie::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QMovie::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn speed(&self) -> c_int

C++ method: int QMovie::speed() const

C++ documentation:

This property holds the movie's speed.

The speed is measured in percentage of the original movie speed. The default speed is 100%. Example:

QMovie movie("racecar.gif");
movie.setSpeed(200); // 2x speed

Access functions:

int 	speed() const

void 	setSpeed(int percentSpeed)

fn start(&mut self)

C++ method: void QMovie::start()

C++ documentation:

Starts the movie. QMovie will enter Running state, and start emitting updated() and resized() as the movie progresses.

If QMovie is in the Paused state, this function is equivalent to calling setPaused(false). If QMovie is already in the Running state, this function does nothing.

See also stop() and setPaused().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: MovieStartTimerArgs<'largs>

C++ method: QMovie::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QMovie::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn started(&mut self)

C++ method: void QMovie::started()

C++ documentation:

This signal is emitted after QMovie::start() has been called, and QMovie has entered QMovie::Running state.

fn state(&self) -> MovieState

C++ method: QMovie::MovieState QMovie::state() const

fn state_changed(&mut self, state: MovieState)

C++ method: void QMovie::stateChanged(QMovie::MovieState state)

C++ documentation:

This signal is emitted every time the state of the movie changes. The new state is specified by state.

See also QMovie::state().

fn stop(&mut self)

C++ method: void QMovie::stop()

C++ documentation:

Stops the movie. QMovie enters NotRunning state, and stops emitting updated() and resized(). If start() is called again, the movie will restart from the beginning.

If QMovie is already in the NotRunning state, this function does nothing.

See also start() and setPaused().

fn supported_formats<Args>(args: Args) -> Args::ReturnType where Args: MovieSupportedFormatsArgs

C++ method: QMovie::supportedFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QMovie::supportedFormats()

C++ documentation:

Returns the list of image formats supported by QMovie.

This function was introduced in Qt 4.1.

See also QImageReader::supportedImageFormats().

fn thread(&self) -> *mut Thread

C++ method: QThread* QMovie::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: MovieTrArgs

C++ method: QMovie::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QMovie::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: MovieTrUtf8Args

C++ method: QMovie::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QMovie::trUtf8(const char* s, const char* c, int n)

fn updated(&mut self, rect: &Rect)

C++ method: void QMovie::updated(const QRect& rect)

C++ documentation:

This signal is emitted when the rect rect in the current frame has been updated. You can call currentImage() or currentPixmap() to get a copy of the updated frame.

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QMovie::userData(unsigned int id) const