Struct qt_gui::native_gesture_event::NativeGestureEvent

pub struct NativeGestureEvent {
    // some fields omitted
}

C++ type: QNativeGestureEvent

C++ documentation:

The QNativeGestureEvent class contains parameters that describe a gesture event.

Native gesture events are generated by the operating system, typically by interpreting touch events. Gesture events are high-level events such as zoom or rotate.

Event Type	Description	Touch equence

Qt::ZoomNativeGesture	Magnification delta in percent.	OS X: Two-finger pinch.

Qt::SmartZoomNativeGesture	Boolean magnification state.	OS X: Two-finger douple tap (trackpad) / One-finger douple tap (magic mouse).

Qt::RotateNativeGesture	Rotation delta in degrees.	OS X: Two-finger rotate.

In addition, BeginNativeGesture and EndNativeGesture are sent before and after gesture event streams:

BeginNativeGesture ZoomNativeGesture ZoomNativeGesture ZoomNativeGesture EndNativeGesture

Methods

impl NativeGestureEvent

fn accept(&mut self)

C++ method: void QNativeGestureEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn gesture_type(&self) -> NativeGestureType

C++ method: Qt::NativeGestureType QNativeGestureEvent::gestureType() const

C++ documentation:

Returns the gesture type.

This function was introduced in Qt 5.2.

fn global_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: NativeGestureEventGlobalPosArgs<'largs>

C++ method: QNativeGestureEvent::globalPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn global_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn global_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: const QPoint QNativeGestureEvent::globalPos() const

C++ documentation:

Returns the position of the gesture as a QPointF in screen coordinates

This function was introduced in Qt 5.2.

fn ignore(&mut self)

C++ method: void QNativeGestureEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QNativeGestureEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn local_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QNativeGestureEvent::localPos() const

C++ documentation:

Returns the position of the gesture as a QPointF, relative to the widget or item that received the event.

This function was introduced in Qt 5.2.

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QNativeGestureEvent::modifiers() const

C++ documentation:

Returns the keyboard modifier flags that existed immediately before the event occurred.

See also QGuiApplication::keyboardModifiers().

fn new<Args>(args: Args) -> Args::ReturnType where Args: NativeGestureEventNewArgs

C++ method: QNativeGestureEvent::QNativeGestureEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((qt_core::qt::NativeGestureType, &qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_double, libc::c_ulong, u64, cpp_utils::AsStruct)) -> ::native_gesture_event::NativeGestureEvent
2) fn new((qt_core::qt::NativeGestureType, &qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_double, libc::c_ulong, u64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::native_gesture_event::NativeGestureEvent>
C++ method: [constructor] void QNativeGestureEvent::QNativeGestureEvent(Qt::NativeGestureType type, const QPointF& localPos, const QPointF& windowPos, const QPointF& screenPos, double value, unsigned long sequenceId, quint64 intArgument)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: NativeGestureEventPosArgs<'largs>

C++ method: QNativeGestureEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: const QPoint QNativeGestureEvent::pos() const

C++ documentation:

Returns the position of the mouse cursor, relative to the widget or item that received the event.

This function was introduced in Qt 5.2.

fn register_event_type<Args>(args: Args) -> c_int where Args: NativeGestureEventRegisterEventTypeArgs

C++ method: QNativeGestureEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QNativeGestureEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn screen_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QNativeGestureEvent::screenPos() const

C++ documentation:

Returns the position of the gesture as a QPointF in screen coordinates.

This function was introduced in Qt 5.2.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QNativeGestureEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_modifiers(&mut self, amodifiers: Flags<KeyboardModifier>)

C++ method: void QNativeGestureEvent::setModifiers(QFlags<Qt::KeyboardModifier> amodifiers)

fn set_timestamp(&mut self, atimestamp: c_ulong)

C++ method: void QNativeGestureEvent::setTimestamp(unsigned long atimestamp)

fn spontaneous(&self) -> bool

C++ method: bool QNativeGestureEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn timestamp(&self) -> c_ulong

C++ method: unsigned long QNativeGestureEvent::timestamp() const

C++ documentation:

Returns the window system's timestamp for this event.

fn type_(&self) -> Type

C++ method: QEvent::Type QNativeGestureEvent::type() const

C++ documentation:

Returns the event type.

fn value(&self) -> c_double

C++ method: double QNativeGestureEvent::value() const

C++ documentation:

Returns the gesture value. The value should be interpreted based on the gesture type. For example, a Zoom gesture provides a scale factor while a Rotate gesture provides a rotation delta.

This function was introduced in Qt 5.2.

See also QNativeGestureEvent and gestureType().

fn window_pos<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QNativeGestureEvent::windowPos() const

C++ documentation:

Returns the position of the gesture as a QPointF, relative to the window that received the event.

This function was introduced in Qt 5.2.