Module qt_gui::offscreen_surface::overloading

Traits

OffscreenSurfaceConnectArgs
This trait represents a set of arguments accepted by OffscreenSurface::connect method.

OffscreenSurfaceConnectStaticArgs
This trait represents a set of arguments accepted by OffscreenSurface::connect_static method.

OffscreenSurfaceDestroyedArgs
This trait represents a set of arguments accepted by OffscreenSurface::destroyed method.

OffscreenSurfaceDisconnectArgs
This trait represents a set of arguments accepted by OffscreenSurface::disconnect method.

OffscreenSurfaceDisconnectStaticArgs
This trait represents a set of arguments accepted by OffscreenSurface::disconnect_static method.

OffscreenSurfaceDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OffscreenSurface::dynamic_property_names method.

OffscreenSurfaceFormatArgs
This trait represents a set of arguments accepted by OffscreenSurface::format method.

OffscreenSurfaceNewArgs
This trait represents a set of arguments accepted by OffscreenSurface::new method.

OffscreenSurfaceObjectNameArgs
This trait represents a set of arguments accepted by OffscreenSurface::object_name method.

OffscreenSurfacePropertyArgs
This trait represents a set of arguments accepted by OffscreenSurface::property method.

OffscreenSurfaceRequestedFormatArgs
This trait represents a set of arguments accepted by OffscreenSurface::requested_format method.

OffscreenSurfaceSizeArgs
This trait represents a set of arguments accepted by OffscreenSurface::size method.

OffscreenSurfaceStartTimerArgs
This trait represents a set of arguments accepted by OffscreenSurface::start_timer method.

OffscreenSurfaceTrArgs
This trait represents a set of arguments accepted by OffscreenSurface::tr method.

OffscreenSurfaceTrUtf8Args
This trait represents a set of arguments accepted by OffscreenSurface::tr_utf8 method.