Enum qt_gui::open_g_l_buffer::UsagePattern

pub enum UsagePattern {
    StreamDraw,
    StreamRead,
    StreamCopy,
    StaticDraw,
    StaticRead,
    StaticCopy,
    DynamicDraw,
    DynamicRead,
    DynamicCopy,
}

C++ type: QOpenGLBuffer::UsagePattern

C++ documentation:

This enum defines the usage pattern of a QOpenGLBuffer object.

Variants
StreamDraw

The data will be set once and used a few times for drawing operations. Under OpenGL/ES 1.1 this is identical to StaticDraw. (C++ enum variant: StreamDraw = 35040)

StreamRead

The data will be set once and used a few times for reading data back from the OpenGL server. Not supported under OpenGL/ES. (C++ enum variant: StreamRead = 35041)

StreamCopy

The data will be set once and used a few times for reading data back from the OpenGL server for use in further drawing operations. Not supported under OpenGL/ES. (C++ enum variant: StreamCopy = 35042)

StaticDraw

The data will be set once and used many times for drawing operations. (C++ enum variant: StaticDraw = 35044)

StaticRead

The data will be set once and used many times for reading data back from the OpenGL server. Not supported under OpenGL/ES. (C++ enum variant: StaticRead = 35045)

StaticCopy

The data will be set once and used many times for reading data back from the OpenGL server for use in further drawing operations. Not supported under OpenGL/ES. (C++ enum variant: StaticCopy = 35046)

DynamicDraw

The data will be modified repeatedly and used many times for drawing operations. (C++ enum variant: DynamicDraw = 35048)

DynamicRead

The data will be modified repeatedly and used many times for reading data back from the OpenGL server. Not supported under OpenGL/ES. (C++ enum variant: DynamicRead = 35049)

DynamicCopy

The data will be modified repeatedly and used many times for reading data back from the OpenGL server for use in further drawing operations. Not supported under OpenGL/ES. (C++ enum variant: DynamicCopy = 35050)