Struct qt_gui::open_g_l_buffer::OpenGLBuffer

pub struct OpenGLBuffer {
    // some fields omitted
}

C++ type: QOpenGLBuffer

C++ documentation:

The QOpenGLBuffer class provides functions for creating and managing OpenGL buffer objects.

Buffer objects are created in the OpenGL server so that the client application can avoid uploading vertices, indices, texture image data, etc every time they are needed.

QOpenGLBuffer objects can be copied around as a reference to the underlying OpenGL buffer object:

QOpenGLBuffer buffer1(QOpenGLBuffer::IndexBuffer);
buffer1.create();

QOpenGLBuffer buffer2 = buffer1;
QOpenGLBuffer performs a shallow copy when objects are copied in this manner, but does not implement copy-on-write semantics. The original object will be affected whenever the copy is modified.

Methods

impl OpenGLBuffer

fn allocate<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLBufferAllocateArgs<'largs>

C++ method: QOpenGLBuffer::allocate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn allocate(&mut self, (*const libc::c_void, libc::c_int)) -> ()
C++ method: void QOpenGLBuffer::allocate(const void* data, int count)

C++ documentation:

Allocates count bytes of space to the buffer, initialized to the contents of data. Any previous contents will be removed.

It is assumed that create() has been called on this buffer and that it has been bound to the current context.

See also create(), read(), and write().

Variant 2

Rust arguments: fn allocate(&mut self, libc::c_int) -> ()
C++ method: void QOpenGLBuffer::allocate(int count)

C++ documentation:

This is an overloaded function.

Allocates count bytes of space to the buffer. Any previous contents will be removed.

It is assumed that create() has been called on this buffer and that it has been bound to the current context.

See also create() and write().

fn bind(&mut self) -> bool

C++ method: bool QOpenGLBuffer::bind()

C++ documentation:

Binds the buffer associated with this object to the current OpenGL context. Returns false if binding was not possible, usually because type() is not supported on this OpenGL implementation.

The buffer must be bound to the same QOpenGLContext current when create() was called, or to another QOpenGLContext that is sharing with it. Otherwise, false will be returned from this function.

See also release() and create().

fn buffer_id(&self) -> c_uint

C++ method: unsigned int QOpenGLBuffer::bufferId() const

C++ documentation:

Returns the OpenGL identifier associated with this buffer; zero if the buffer has not been created.

See also isCreated().

fn create(&mut self) -> bool

C++ method: bool QOpenGLBuffer::create()

C++ documentation:

Creates the buffer object in the OpenGL server. Returns true if the object was created; false otherwise.

This function must be called with a current QOpenGLContext. The buffer will be bound to and can only be used in that context (or any other context that is shared with it).

This function will return false if the OpenGL implementation does not support buffers, or there is no current QOpenGLContext.

See also isCreated(), allocate(), write(), and destroy().

fn destroy(&mut self)

C++ method: void QOpenGLBuffer::destroy()

C++ documentation:

Destroys this buffer object, including the storage being used in the OpenGL server. All references to the buffer will become invalid.

fn is_created(&self) -> bool

C++ method: bool QOpenGLBuffer::isCreated() const

C++ documentation:

Returns true if this buffer has been created; false otherwise.

See also create() and destroy().

fn map(&mut self, access: Access) -> *mut c_void

C++ method: void* QOpenGLBuffer::map(QOpenGLBuffer::Access access)

C++ documentation:

Maps the contents of this buffer into the application's memory space and returns a pointer to it. Returns null if memory mapping is not possible. The access parameter indicates the type of access to be performed.

It is assumed that create() has been called on this buffer and that it has been bound to the current context.

Note: This function is only supported under OpenGL ES 2.0 or earlier if the GL_OES_mapbuffer extension is present.

Note: On OpenGL ES 3.0 and newer, or, in case if desktop OpenGL, if GL_ARB_map_buffer_range is supported, this function uses glMapBufferRange instead of glMapBuffer.

See also unmap(), create(), bind(), and mapRange().

fn map_range(&mut self, offset: c_int, count: c_int, access: Flags<RangeAccessFlag>) -> *mut c_void

C++ method: void* QOpenGLBuffer::mapRange(int offset, int count, QFlags<QOpenGLBuffer::RangeAccessFlag> access)

C++ documentation:

Maps the range specified by offset and count of the contents of this buffer into the application's memory space and returns a pointer to it. Returns null if memory mapping is not possible. The access parameter specifies a combination of access flags.

It is assumed that create() has been called on this buffer and that it has been bound to the current context.

Note: This function is not available on OpenGL ES 2.0 and earlier.

See also unmap(), create(), and bind().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLBufferNewArgs

C++ method: QOpenGLBuffer::QOpenGLBuffer

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_buffer::OpenGLBuffer
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_buffer::OpenGLBuffer>
C++ method: [constructor] void QOpenGLBuffer::QOpenGLBuffer()

Variant 2

Rust arguments:
1) fn new((::open_g_l_buffer::Type, cpp_utils::AsStruct)) -> ::open_g_l_buffer::OpenGLBuffer
2) fn new((::open_g_l_buffer::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_buffer::OpenGLBuffer>
C++ method: [constructor] void QOpenGLBuffer::QOpenGLBuffer(QOpenGLBuffer::Type type)

Variant 3

Rust arguments:
1) fn new((&::open_g_l_buffer::OpenGLBuffer, cpp_utils::AsStruct)) -> ::open_g_l_buffer::OpenGLBuffer
2) fn new((&::open_g_l_buffer::OpenGLBuffer, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_buffer::OpenGLBuffer>
C++ method: [constructor] void QOpenGLBuffer::QOpenGLBuffer(const QOpenGLBuffer& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 OpenGLBuffer) -> &'l0 mut OpenGLBuffer

C++ method: QOpenGLBuffer& QOpenGLBuffer::operator=(const QOpenGLBuffer& other)

C++ documentation:

Assigns a shallow copy of other to this object.

Note: QOpenGLBuffer does not implement copy-on-write semantics, so other will be affected whenever the copy is modified.

fn read(&mut self, offset: c_int, data: *mut c_void, count: c_int) -> bool

C++ method: bool QOpenGLBuffer::read(int offset, void* data, int count)

C++ documentation:

Reads the count bytes in this buffer starting at offset into data. Returns true on success; false if reading from the buffer is not supported. Buffer reading is not supported under OpenGL/ES.

It is assumed that this buffer has been bound to the current context.

See also write() and bind().

fn release(&mut self)

C++ method: void QOpenGLBuffer::release()

C++ documentation:

Releases the buffer associated with this object from the current OpenGL context.

This function must be called with the same QOpenGLContext current as when bind() was called on the buffer.

See also bind().

fn release_static(type_: Type)

C++ method: static void QOpenGLBuffer::release(QOpenGLBuffer::Type type)

C++ documentation:

Releases the buffer associated with type in the current QOpenGLContext.

This function is a direct call to glBindBuffer(type, 0) for use when the caller does not know which QOpenGLBuffer has been bound to the context but wants to make sure that it is released.

QOpenGLBuffer::release(QOpenGLBuffer::VertexBuffer);

fn set_usage_pattern(&mut self, value: UsagePattern)

C++ method: void QOpenGLBuffer::setUsagePattern(QOpenGLBuffer::UsagePattern value)

C++ documentation:

Sets the usage pattern for this buffer object to value. This function must be called before allocate() or write().

See also usagePattern(), allocate(), and write().

fn size(&self) -> c_int

C++ method: int QOpenGLBuffer::size() const

C++ documentation:

Returns the size of the data in this buffer, for reading operations. Returns -1 if fetching the buffer size is not supported, or the buffer has not been created.

It is assumed that this buffer has been bound to the current context.

See also isCreated() and bind().

fn type_(&self) -> Type

C++ method: QOpenGLBuffer::Type QOpenGLBuffer::type() const

C++ documentation:

Returns the type of buffer represented by this object.

fn unmap(&mut self) -> bool

C++ method: bool QOpenGLBuffer::unmap()

C++ documentation:

Unmaps the buffer after it was mapped into the application's memory space with a previous call to map(). Returns true if the unmap succeeded; false otherwise.

It is assumed that this buffer has been bound to the current context, and that it was previously mapped with map().

Note: This function is only supported under OpenGL ES 2.0 and earlier if the GL_OES_mapbuffer extension is present.

See also map().

fn usage_pattern(&self) -> UsagePattern

C++ method: QOpenGLBuffer::UsagePattern QOpenGLBuffer::usagePattern() const

C++ documentation:

Returns the usage pattern for this buffer object. The default value is StaticDraw.

See also setUsagePattern().

fn write(&mut self, offset: c_int, data: *const c_void, count: c_int)

C++ method: void QOpenGLBuffer::write(int offset, const void* data, int count)

C++ documentation:

Replaces the count bytes of this buffer starting at offset with the contents of data. Any other bytes in the buffer will be left unmodified.

It is assumed that create() has been called on this buffer and that it has been bound to the current context.

See also create(), read(), and allocate().