Module qt_gui::open_g_l_context::overloading

Traits

HashArgs
This trait represents a set of arguments accepted by hash method.

OpenGLContextConnectArgs
This trait represents a set of arguments accepted by OpenGLContext::connect method.

OpenGLContextConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLContext::connect_static method.

OpenGLContextDestroyedArgs
This trait represents a set of arguments accepted by OpenGLContext::destroyed method.

OpenGLContextDisconnectArgs
This trait represents a set of arguments accepted by OpenGLContext::disconnect method.

OpenGLContextDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLContext::disconnect_static method.

OpenGLContextDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLContext::dynamic_property_names method.

OpenGLContextExtensionsArgs
This trait represents a set of arguments accepted by OpenGLContext::extensions method.

OpenGLContextFormatArgs
This trait represents a set of arguments accepted by OpenGLContext::format method.

OpenGLContextNativeHandleArgs
This trait represents a set of arguments accepted by OpenGLContext::native_handle method.

OpenGLContextNewArgs
This trait represents a set of arguments accepted by OpenGLContext::new method.

OpenGLContextObjectNameArgs
This trait represents a set of arguments accepted by OpenGLContext::object_name method.

OpenGLContextPropertyArgs
This trait represents a set of arguments accepted by OpenGLContext::property method.

OpenGLContextStartTimerArgs
This trait represents a set of arguments accepted by OpenGLContext::start_timer method.

OpenGLContextTrArgs
This trait represents a set of arguments accepted by OpenGLContext::tr method.

OpenGLContextTrUtf8Args
This trait represents a set of arguments accepted by OpenGLContext::tr_utf8 method.

OpenGLContextVersionFunctionsArgs
This trait represents a set of arguments accepted by OpenGLContext::version_functions method.