Struct qt_gui::open_g_l_context::OpenGLContext

pub struct OpenGLContext {
    // some fields omitted
}

C++ type: QOpenGLContext

C++ documentation:

The QOpenGLContext class represents a native OpenGL context, enabling OpenGL rendering on a QSurface.

QOpenGLContext represents the OpenGL state of an underlying OpenGL context. To set up a context, set its screen and format such that they match those of the surface or surfaces with which the context is meant to be used, if necessary make it share resources with other contexts with setShareContext(), and finally call create(). Use the return value or isValid() to check if the context was successfully initialized.

A context can be made current against a given surface by calling makeCurrent(). When OpenGL rendering is done, call swapBuffers() to swap the front and back buffers of the surface, so that the newly rendered content becomes visible. To be able to support certain platforms, QOpenGLContext requires that you call makeCurrent() again before starting rendering a new frame, after calling swapBuffers().

If the context is temporarily not needed, such as when the application is not rendering, it can be useful to delete it in order to free resources. You can connect to the aboutToBeDestroyed() signal to clean up any resources that have been allocated with different ownership from the QOpenGLContext itself.

Once a QOpenGLContext has been made current, you can render to it in a platform independent way by using Qt's OpenGL enablers such as QOpenGLFunctions, QOpenGLBuffer, QOpenGLShaderProgram, and QOpenGLFramebufferObject. It is also possible to use the platform's OpenGL API directly, without using the Qt enablers, although potentially at the cost of portability. The latter is necessary when wanting to use OpenGL 1.x or OpenGL ES 1.x.

For more information about the OpenGL API, refer to the official OpenGL documentation.

For an example of how to use QOpenGLContext see the OpenGL Window example.

Methods

impl OpenGLContext

fn about_to_be_destroyed(&mut self)

C++ method: void QOpenGLContext::aboutToBeDestroyed()

C++ documentation:

This signal is emitted before the underlying native OpenGL context is destroyed, such that users may clean up OpenGL resources that might otherwise be left dangling in the case of shared OpenGL contexts.

If you wish to make the context current in order to do clean-up, make sure to only connect to the signal using a direct connection.

fn are_sharing(first: *mut OpenGLContext, second: *mut OpenGLContext) -> bool

C++ method: static bool QOpenGLContext::areSharing(QOpenGLContext* first, QOpenGLContext* second)

C++ documentation:

Returns true if the first and second contexts are sharing OpenGL resources.

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QOpenGLContext::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QOpenGLContext::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextConnectArgs<'largs>

C++ method: QOpenGLContext::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect(&self, (*const qt_core::object::Object, *const libc::c_char, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: QMetaObject::Connection QOpenGLContext::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: OpenGLContextConnectStaticArgs

C++ method: QOpenGLContext::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QOpenGLContext::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
2) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::meta_object::Connection>
3) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
4) fn connect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char, qt_core::qt::ConnectionType, cpp_utils::AsStruct)) -> qt_core::meta_object::Connection
C++ method: static QMetaObject::Connection QOpenGLContext::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn create(&mut self) -> bool

C++ method: bool QOpenGLContext::create()

C++ documentation:

Attempts to create the OpenGL context with the current configuration.

The current configuration includes the format, the share context, and the screen.

If the OpenGL implementation on your system does not support the requested version of OpenGL context, then QOpenGLContext will try to create the closest matching version. The actual created context properties can be queried using the QSurfaceFormat returned by the format() function. For example, if you request a context that supports OpenGL 4.3 Core profile but the driver and/or hardware only supports version 3.2 Core profile contexts then you will get a 3.2 Core profile context.

Returns true if the native context was successfully created and is ready to be used with makeCurrent(), swapBuffers(), etc.

Note: If the context already exists, this function destroys the existing context first, and then creates a new one.

See also makeCurrent() and format().

fn current_context() -> *mut OpenGLContext

C++ method: static QOpenGLContext* QOpenGLContext::currentContext()

C++ documentation:

Returns the last context which called makeCurrent in the current thread, or 0, if no context is current.

fn default_framebuffer_object(&self) -> c_uint

C++ method: unsigned int QOpenGLContext::defaultFramebufferObject() const

C++ documentation:

Call this to get the default framebuffer object for the current surface.

On some platforms (for instance, iOS) the default framebuffer object depends on the surface being rendered to, and might be different from 0. Thus, instead of calling glBindFramebuffer(0), you should call glBindFramebuffer(ctx->defaultFramebufferObject()) if you want your application to work across different Qt platforms.

If you use the glBindFramebuffer() in QOpenGLFunctions you do not have to worry about this, as it automatically binds the current context's defaultFramebufferObject() when 0 is passed.

Note: Widgets that render via framebuffer objects, like QOpenGLWidget and QQuickWidget, will override the value returned from this function when painting is active, because at that time the correct "default" framebuffer is the widget's associated backing framebuffer, not the platform-specific one belonging to the top-level window's surface. This ensures the expected behavior for this function and other classes relying on it (for example, QOpenGLFramebufferObject::bindDefault() or QOpenGLFramebufferObject::release()).

See also QOpenGLFramebufferObject.

fn delete_later(&mut self)

C++ method: void QOpenGLContext::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLContextDestroyedArgs<'largs>

C++ method: QOpenGLContext::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut qt_core::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QOpenGLContext::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: OpenGLContextDisconnectArgs<'largs>

C++ method: QOpenGLContext::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const qt_core::object::Object) -> bool
2) fn disconnect(&self, (*const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: bool QOpenGLContext::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QOpenGLContext::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: OpenGLContextDisconnectStaticArgs

C++ method: QOpenGLContext::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&qt_core::meta_object::Connection) -> bool
C++ method: static bool QOpenGLContext::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const qt_core::object::Object, &qt_core::meta_method::MetaMethod, *const qt_core::object::Object, &qt_core::meta_method::MetaMethod)) -> bool
C++ method: static bool QOpenGLContext::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const qt_core::object::Object, *const libc::c_char, *const qt_core::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QOpenGLContext::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn done_current(&mut self)

C++ method: void QOpenGLContext::doneCurrent()

C++ documentation:

Convenience function for calling makeCurrent with a 0 surface.

This results in no context being current in the current thread.

See also makeCurrent() and currentContext().

fn dump_object_info(&mut self)

C++ method: void QOpenGLContext::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QOpenGLContext::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextDynamicPropertyNamesArgs<'largs>

C++ method: QOpenGLContext::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QOpenGLContext::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QOpenGLContext::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QOpenGLContext::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn extensions<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextExtensionsArgs<'largs>

C++ method: QOpenGLContext::extensions

This is an overloaded function. Available variants:

Rust arguments:
1) fn extensions(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn extensions(&self, cpp_utils::AsStruct) -> ::set::SetByteArrayRef
C++ method: QSet<QByteArray> QOpenGLContext::extensions() const

C++ documentation:

Returns the set of OpenGL extensions supported by this context.

The context or a sharing context must be current.

See also hasExtension().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextFormatArgs<'largs>

C++ method: QOpenGLContext::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
2) fn format(&self, cpp_utils::AsStruct) -> ::surface_format::SurfaceFormat
C++ method: QSurfaceFormat QOpenGLContext::format() const

C++ documentation:

Returns the format of the underlying platform context, if create() has been called.

Otherwise, returns the requested format.

The requested and the actual format may differ. Requesting a given OpenGL version does not mean the resulting context will target exactly the requested version. It is only guaranteed that the version/profile/options combination for the created context is compatible with the request, as long as the driver is able to provide such a context.

For example, requesting an OpenGL version 3.x core profile context may result in an OpenGL 4.x core profile context. Similarly, a request for OpenGL 2.1 may result in an OpenGL 3.0 context with deprecated functions enabled. Finally, depending on the driver, unsupported versions may result in either a context creation failure or in a context for the highest supported version.

Similar differences are possible in the buffer sizes, for example, the resulting context may have a larger depth buffer than requested. This is perfectly normal.

See also setFormat().

fn functions(&self) -> *mut OpenGLFunctions

C++ method: QOpenGLFunctions* QOpenGLContext::functions() const

C++ documentation:

Get the QOpenGLFunctions instance for this context.

QOpenGLContext offers this as a convenient way to access QOpenGLFunctions without having to manage it manually.

The context or a sharing context must be current.

The returned QOpenGLFunctions instance is ready to be used and it does not need initializeOpenGLFunctions() to be called.

fn get_proc_address(&self, proc_name: &ByteArray) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLContext::getProcAddress(const QByteArray& procName) const

C++ documentation:

Resolves the function pointer to an OpenGL extension function, identified by procName

Returns 0 if no such function can be found.

fn global_share_context() -> *mut OpenGLContext

C++ method: static QOpenGLContext* QOpenGLContext::globalShareContext()

C++ documentation:

Returns the application-wide shared OpenGL context, if present. Otherwise, returns a null pointer.

This is useful if you need to upload OpenGL objects (buffers, textures, etc.) before creating or showing a QOpenGLWidget or QQuickWidget.

Note: You must set the Qt::AA_ShareOpenGLContexts flag on QGuiApplication before creating the QGuiApplication object, otherwise Qt may not create a global shared context.

Warning: Do not attempt to make the context returned by this function current on any surface. Instead, you can create a new context which shares with the global one, and then make the new context current.

This function was introduced in Qt 5.5.

See also Qt::AA_ShareOpenGLContexts, setShareContext(), and makeCurrent().

fn has_extension(&self, extension: &ByteArray) -> bool

C++ method: bool QOpenGLContext::hasExtension(const QByteArray& extension) const

C++ documentation:

Returns true if this OpenGL context supports the specified OpenGL extension, false otherwise.

The context or a sharing context must be current.

See also extensions().

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QOpenGLContext::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QOpenGLContext::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_opengl_e_s(&self) -> bool

C++ method: bool QOpenGLContext::isOpenGLES() const

C++ documentation:

Returns true if the context is an OpenGL ES context.

If the context has not yet been created, the result is based on the requested format set via setFormat().

This function was introduced in Qt 5.3.

See also create(), format(), and setFormat().

fn is_valid(&self) -> bool

C++ method: bool QOpenGLContext::isValid() const

C++ documentation:

Returns if this context is valid, i.e. has been successfully created.

On some platforms the return value of false for a context that was successfully created previously indicates that the OpenGL context was lost.

The typical way to handle context loss scenarios in applications is to check via this function whenever makeCurrent() fails and returns false. If this function then returns false, recreate the underlying native OpenGL context by calling create(), call makeCurrent() again and then reinitialize all OpenGL resources.

See also create().

fn is_widget_type(&self) -> bool

C++ method: bool QOpenGLContext::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QOpenGLContext::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QOpenGLContext::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn make_current(&mut self, surface: *mut Surface) -> bool

C++ method: bool QOpenGLContext::makeCurrent(QSurface* surface)

C++ documentation:

Makes the context current in the current thread, against the given surface. Returns true if successful.

If surface is 0 this is equivalent to calling doneCurrent().

Do not call this function from a different thread than the one the QOpenGLContext instance lives in. If you wish to use QOpenGLContext from a different thread you should first call make sure it's not current in the current thread, by calling doneCurrent() if necessary. Then call moveToThread(otherThread) before using it in the other thread.

See also functions() and doneCurrent().

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QOpenGLContext::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QOpenGLContext::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn native_handle<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextNativeHandleArgs<'largs>

C++ method: QOpenGLContext::nativeHandle

This is an overloaded function. Available variants:

Rust arguments:
1) fn native_handle(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn native_handle(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QOpenGLContext::nativeHandle() const

C++ documentation:

Returns the native handle for the context.

This function provides access to the QOpenGLContext's underlying native context. The returned variant contains a platform-specific value type. These classes can be found in the module QtPlatformHeaders.

On platforms where retrieving the native handle is not supported, or if neither create() nor setNativeHandle() was called, a null variant is returned.

This function was introduced in Qt 5.4.

See also setNativeHandle().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLContextNewArgs

C++ method: QOpenGLContext::QOpenGLContext

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_context::OpenGLContext
2) fn new((*mut qt_core::object::Object, cpp_utils::AsStruct)) -> ::open_g_l_context::OpenGLContext
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_context::OpenGLContext>
4) fn new((*mut qt_core::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_context::OpenGLContext>
C++ method: [constructor] void QOpenGLContext::QOpenGLContext(QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextObjectNameArgs<'largs>

C++ method: QOpenGLContext::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QOpenGLContext::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn open_glm_odule_handle() -> *mut c_void

C++ method: static void* QOpenGLContext::openGLModuleHandle()

C++ documentation:

Returns the platform-specific handle for the OpenGL implementation that is currently in use. (for example, a HMODULE on Windows)

On platforms that do not use dynamic GL switch the return value is null.

The library might be GL-only, meaning that windowing system interface functions (for example EGL) may live in another, separate library.

Note: This function requires that the QGuiApplication instance is already created.

This function was introduced in Qt 5.3.

See also openGLModuleType().

fn open_glm_odule_type() -> OpenGLModuleType

C++ method: static QOpenGLContext::OpenGLModuleType QOpenGLContext::openGLModuleType()

C++ documentation:

Returns the underlying OpenGL implementation type.

On platforms where the OpenGL implementation is not dynamically loaded, the return value is determined during compile time and never changes.

Note: A desktop OpenGL implementation may be capable of creating ES-compatible contexts too. Therefore in most cases it is more appropriate to check QSurfaceFormat::renderableType() or using the the convenience function isOpenGLES().

Note: This function requires that the QGuiApplication instance is already created.

This function was introduced in Qt 5.3.

fn parent(&self) -> *mut Object

C++ method: QObject* QOpenGLContext::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLContextPropertyArgs<'largs>

C++ method: QOpenGLContext::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QOpenGLContext::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QOpenGLContext::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QOpenGLContext::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QOpenGLContext::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QOpenGLContext::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn screen(&self) -> *mut Screen

C++ method: QScreen* QOpenGLContext::screen() const

C++ documentation:

Returns the screen the context was created for.

See also setScreen().

fn set_format(&mut self, format: &SurfaceFormat)

C++ method: void QOpenGLContext::setFormat(const QSurfaceFormat& format)

C++ documentation:

Sets the format the OpenGL context should be compatible with. You need to call create() before it takes effect.

When the format is not explicitly set via this function, the format returned by QSurfaceFormat::defaultFormat() will be used. This means that when having multiple contexts, individual calls to this function can be replaced by one single call to QSurfaceFormat::setDefaultFormat() before creating the first context.

See also format().

fn set_native_handle(&mut self, handle: &Variant)

C++ method: void QOpenGLContext::setNativeHandle(const QVariant& handle)

C++ documentation:

Set the native handles for this context. When create() is called and a native handle is set, configuration settings, like format(), are ignored since this QOpenGLContext will wrap an already created native context instead of creating a new one from scratch.

On some platforms the native context handle is not sufficient and other related handles (for example, for a window or display) have to be provided in addition. Therefore handle is variant containing a platform-specific value type. These classes can be found in the QtPlatformHeaders module.

When create() is called with native handles set, QOpenGLContext does not take ownership of the handles, so destroying the QOpenGLContext does not destroy the native context.

Note: Some frameworks track the current context and surfaces internally. Making the adopted QOpenGLContext current via Qt will have no effect on such other frameworks' internal state. Therefore a subsequent makeCurrent done via the other framework may have no effect. It is therefore advisable to make explicit calls to make no context and surface current to reset the other frameworks' internal state after performing OpenGL operations via Qt.

Note: Using foreign contexts with Qt windows and Qt contexts with windows and surfaces created by other frameworks may give unexpected results, depending on the platform, due to potential mismatches in context and window pixel formats. To make sure this does not happen, avoid making contexts and surfaces from different frameworks current together. Instead, prefer approaches based on context sharing where OpenGL resources like textures are accessible both from Qt's and the foreign framework's contexts.

This function was introduced in Qt 5.4.

See also nativeHandle().

fn set_object_name(&mut self, name: &String)

C++ method: void QOpenGLContext::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QOpenGLContext::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QOpenGLContext::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_screen(&mut self, screen: *mut Screen)

C++ method: void QOpenGLContext::setScreen(QScreen* screen)

C++ documentation:

Sets the screen the OpenGL context should be valid for. You need to call create() before it takes effect.

See also screen().

fn set_share_context(&mut self, share_context: *mut OpenGLContext)

C++ method: void QOpenGLContext::setShareContext(QOpenGLContext* shareContext)

C++ documentation:

Makes this context share textures, shaders, and other OpenGL resources with shareContext. You need to call create() before it takes effect.

See also shareContext().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QOpenGLContext::setUserData(unsigned int id, QObjectUserData* data)

fn share_context(&self) -> *mut OpenGLContext

C++ method: QOpenGLContext* QOpenGLContext::shareContext() const

C++ documentation:

Returns the share context this context was created with.

If the underlying platform was not able to support the requested sharing, this will return 0.

See also setShareContext().

fn share_group(&self) -> *mut OpenGLContextGroup

C++ method: QOpenGLContextGroup* QOpenGLContext::shareGroup() const

C++ documentation:

Returns the share group this context belongs to.

fn signals_blocked(&self) -> bool

C++ method: bool QOpenGLContext::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: OpenGLContextStartTimerArgs<'largs>

C++ method: QOpenGLContext::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, qt_core::qt::TimerType)) -> libc::c_int
C++ method: int QOpenGLContext::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn supports_threaded_opengl() -> bool

C++ method: static bool QOpenGLContext::supportsThreadedOpenGL()

C++ documentation:

Returns true if the platform supports OpenGL rendering outside the main (gui) thread.

The value is controlled by the platform plugin in use and may also depend on the graphics drivers.

This function was introduced in Qt 5.5.

fn surface(&self) -> *mut Surface

C++ method: QSurface* QOpenGLContext::surface() const

C++ documentation:

Returns the surface the context has been made current with.

This is the surface passed as an argument to makeCurrent().

fn swap_buffers(&mut self, surface: *mut Surface)

C++ method: void QOpenGLContext::swapBuffers(QSurface* surface)

C++ documentation:

Swap the back and front buffers of surface.

Call this to finish a frame of OpenGL rendering, and make sure to call makeCurrent() again before you begin a new frame.

fn thread(&self) -> *mut Thread

C++ method: QThread* QOpenGLContext::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: OpenGLContextTrArgs

C++ method: QOpenGLContext::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QOpenGLContext::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: OpenGLContextTrUtf8Args

C++ method: QOpenGLContext::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QOpenGLContext::trUtf8(const char* s, const char* c, int n)

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QOpenGLContext::userData(unsigned int id) const

fn version_functions<'largs, Args>(&'largs self, args: Args) -> *mut AbstractOpenGLFunctions where Args: OpenGLContextVersionFunctionsArgs<'largs>

C++ method: QOpenGLContext::versionFunctions

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_functions(&self, ()) -> *mut ::open_g_l_version_functions::AbstractOpenGLFunctions
2) fn version_functions(&self, &::open_g_l_version_profile::OpenGLVersionProfile) -> *mut ::open_g_l_version_functions::AbstractOpenGLFunctions
C++ method: QAbstractOpenGLFunctions* QOpenGLContext::versionFunctions(const QOpenGLVersionProfile& versionProfile = ?) const

C++ documentation:

Returns a pointer to an object that provides access to all functions for the versionProfile of this context. There is no need to call QAbstractOpenGLFunctions::initializeOpenGLFunctions() as long as this context is current. It is also possible to call this function when the context is not current, but in that case it is the caller's responsibility to ensure proper initialization by calling QAbstractOpenGLFunctions::initializeOpenGLFunctions() afterwards.

Usually one would use the template version of this function to automatically have the result cast to the correct type.