Module qt_gui::open_g_l_debug_logger::overloading

Traits

OpenGLDebugLoggerConnectArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::connect method.

OpenGLDebugLoggerConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::connect_static method.

OpenGLDebugLoggerDestroyedArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::destroyed method.

OpenGLDebugLoggerDisableMessagesArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::disable_messages method.

OpenGLDebugLoggerDisconnectArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::disconnect method.

OpenGLDebugLoggerDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::disconnect_static method.

OpenGLDebugLoggerDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::dynamic_property_names method.

OpenGLDebugLoggerEnableMessagesArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::enable_messages method.

OpenGLDebugLoggerLoggedMessagesArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::logged_messages method.

OpenGLDebugLoggerNewArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::new method.

OpenGLDebugLoggerObjectNameArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::object_name method.

OpenGLDebugLoggerPropertyArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::property method.

OpenGLDebugLoggerPushGroupArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::push_group method.

OpenGLDebugLoggerStartLoggingArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::start_logging method.

OpenGLDebugLoggerStartTimerArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::start_timer method.

OpenGLDebugLoggerTrArgs
This trait represents a set of arguments accepted by OpenGLDebugLogger::tr method.

OpenGLDebugLoggerTrUtf8Args
This trait represents a set of arguments accepted by OpenGLDebugLogger::tr_utf8 method.