Enum qt_gui::open_g_l_debug_message::Severity

pub enum Severity {
    Any,
    Invalid,
    High,
    Medium,
    Low,
    Notification,
}

C++ type: QOpenGLDebugMessage::Severity

C++ documentation:

The Severity enum defines the severity of the debug message.

The Severities type is a typedef for QFlags<Severity>. It stores an OR combination of Severity values.

Variants
Any

This value corresponds to a mask of all possible message severities. (C++ enum variant: AnySeverity = -1)

Invalid

The severity of the message is invalid; this is the severity of a default-constructed QOpenGLDebugMessage object. (C++ enum variant: InvalidSeverity = 0)

High

The message has a high severity. (C++ enum variant: HighSeverity = 1)

Medium

The message has a medium severity. (C++ enum variant: MediumSeverity = 2)

Low

The message has a low severity. (C++ enum variant: LowSeverity = 4)

Notification

This variant corresponds to multiple C++ enum variants with the same value:

NotificationSeverity = 8: The message is a notification.

LastSeverity = 8