Enum qt_gui::open_g_l_debug_message::Source

pub enum Source {
    AnySource,
    InvalidSource,
    ApisOurce,
    WindowSystemSource,
    ShaderCompilerSource,
    ThirdPartySource,
    ApplicationSource,
    OtherSource,
}

C++ type: QOpenGLDebugMessage::Source

C++ documentation:

The Source enum defines the source of the debug message.

The Sources type is a typedef for QFlags<Source>. It stores an OR combination of Source values.

Variants
AnySource

This value corresponds to a mask of all possible message sources. (C++ enum variant: AnySource = -1)

InvalidSource

The source of the message is invalid; this is the source of a default-constructed QOpenGLDebugMessage object. (C++ enum variant: InvalidSource = 0)

ApisOurce

The message was generated in response to OpenGL API calls. (C++ enum variant: APISource = 1)

WindowSystemSource

The message was generated by the window system. (C++ enum variant: WindowSystemSource = 2)

ShaderCompilerSource

The message was generated by the shader compiler. (C++ enum variant: ShaderCompilerSource = 4)

ThirdPartySource

The message was generated by a third party, for instance an OpenGL framework a or debugging toolkit. (C++ enum variant: ThirdPartySource = 8)

ApplicationSource

The message was generated by the application itself. (C++ enum variant: ApplicationSource = 16)

OtherSource

This variant corresponds to multiple C++ enum variants with the same value:

OtherSource = 32: The message was generated by a source not included in this enumeration.

LastSource = 32