Struct qt_gui::open_g_l_debug_message::OpenGLDebugMessage

pub struct OpenGLDebugMessage {
    // some fields omitted
}

C++ type: QOpenGLDebugMessage

C++ documentation:

The QOpenGLDebugMessage class wraps an OpenGL debug message.

Debug messages are usually created by the OpenGL server and then read by OpenGL clients (either from the OpenGL internal debug log, or logged in real-time). A debug message has a textual representation, a vendor-specific numeric id, a source, a type and a severity.

It's also possible for applications or third-party libraries and toolkits to create and insert messages in the debug log. In order to do so, you can use the createApplicationMessage() or the createThirdPartyMessage() static functions.

Methods

impl OpenGLDebugMessage

fn create_application_message<Args>(args: Args) -> Args::ReturnType where Args: OpenGLDebugMessageCreateApplicationMessageArgs

C++ method: QOpenGLDebugMessage::createApplicationMessage

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_application_message((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
2) fn create_application_message((&qt_core::string::String, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
3) fn create_application_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
4) fn create_application_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, ::open_g_l_debug_message::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
5) fn create_application_message((&qt_core::string::String, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
6) fn create_application_message((&qt_core::string::String, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
7) fn create_application_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
8) fn create_application_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, ::open_g_l_debug_message::Type, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
C++ method: static QOpenGLDebugMessage QOpenGLDebugMessage::createApplicationMessage(const QString& text, unsigned int id = ?, QOpenGLDebugMessage::Severity severity = ?, QOpenGLDebugMessage::Type type = ?)

C++ documentation:

Constructs and returns a debug message with text as its text, id as id, severity as severity, and type as type. The message source will be set to ApplicationSource.

See also QOpenGLDebugLogger::logMessage() and createThirdPartyMessage().

fn create_third_party_message<Args>(args: Args) -> Args::ReturnType where Args: OpenGLDebugMessageCreateThirdPartyMessageArgs

C++ method: QOpenGLDebugMessage::createThirdPartyMessage

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_third_party_message((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
2) fn create_third_party_message((&qt_core::string::String, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
3) fn create_third_party_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
4) fn create_third_party_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, ::open_g_l_debug_message::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
5) fn create_third_party_message((&qt_core::string::String, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
6) fn create_third_party_message((&qt_core::string::String, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
7) fn create_third_party_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
8) fn create_third_party_message((&qt_core::string::String, libc::c_uint, ::open_g_l_debug_message::Severity, ::open_g_l_debug_message::Type, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
C++ method: static QOpenGLDebugMessage QOpenGLDebugMessage::createThirdPartyMessage(const QString& text, unsigned int id = ?, QOpenGLDebugMessage::Severity severity = ?, QOpenGLDebugMessage::Type type = ?)

C++ documentation:

Constructs and returns a debug message with text as its text, id as id, severity as severity, and type as type. The message source will be set to ThirdPartySource.

See also QOpenGLDebugLogger::logMessage() and createApplicationMessage().

fn id(&self) -> c_uint

C++ method: unsigned int QOpenGLDebugMessage::id() const

C++ documentation:

Returns the id of the debug message. Ids are generally vendor-specific.

fn message<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLDebugMessageMessageArgs<'largs>

C++ method: QOpenGLDebugMessage::message

This is an overloaded function. Available variants:

Rust arguments:
1) fn message(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn message(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QOpenGLDebugMessage::message() const

C++ documentation:

Returns the textual message contained by this debug message.

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLDebugMessageNewArgs

C++ method: QOpenGLDebugMessage::QOpenGLDebugMessage

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_debug_message::OpenGLDebugMessage
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
C++ method: [constructor] void QOpenGLDebugMessage::QOpenGLDebugMessage()

Variant 2

Rust arguments:
1) fn new((&::open_g_l_debug_message::OpenGLDebugMessage, cpp_utils::AsStruct)) -> ::open_g_l_debug_message::OpenGLDebugMessage
2) fn new((&::open_g_l_debug_message::OpenGLDebugMessage, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_debug_message::OpenGLDebugMessage>
C++ method: [constructor] void QOpenGLDebugMessage::QOpenGLDebugMessage(const QOpenGLDebugMessage& debugMessage)

fn op_assign<'l0, 'l1>(&'l0 mut self, debug_message: &'l1 OpenGLDebugMessage) -> &'l0 mut OpenGLDebugMessage

C++ method: QOpenGLDebugMessage& QOpenGLDebugMessage::operator=(const QOpenGLDebugMessage& debugMessage)

C++ documentation:

Assigns the message debugMessage to this object, and returns a reference to the copy.

fn op_eq(&self, debug_message: &OpenGLDebugMessage) -> bool

C++ method: bool QOpenGLDebugMessage::operator==(const QOpenGLDebugMessage& debugMessage) const

C++ documentation:

Returns true if this debug message is equal to debugMessage, or false otherwise. Two debugging messages are equal if they have the same textual message, the same id, the same source, the same type and the same severity.

See also operator!=().

fn op_neq(&self, debug_message: &OpenGLDebugMessage) -> bool

C++ method: bool QOpenGLDebugMessage::operator!=(const QOpenGLDebugMessage& debugMessage) const

C++ documentation:

Returns true if this message is different from debugMessage, or false otherwise.

See also operator==().

fn severity(&self) -> Severity

C++ method: QOpenGLDebugMessage::Severity QOpenGLDebugMessage::severity() const

C++ documentation:

Returns the severity of the debug message.

fn source(&self) -> Source

C++ method: QOpenGLDebugMessage::Source QOpenGLDebugMessage::source() const

C++ documentation:

Returns the source of the debug message.

fn swap(&mut self, debug_message: &mut OpenGLDebugMessage)

C++ method: void QOpenGLDebugMessage::swap(QOpenGLDebugMessage& debugMessage)

C++ documentation:

Swaps the message debugMessage with this message. This operation is very fast and never fails.

fn type_(&self) -> Type

C++ method: QOpenGLDebugMessage::Type QOpenGLDebugMessage::type() const

C++ documentation:

Returns the type of the debug message.