Struct qt_gui::open_g_l_framebuffer_object::OpenGLFramebufferObject

pub struct OpenGLFramebufferObject {
    // some fields omitted
}

C++ type: QOpenGLFramebufferObject

C++ documentation:

The QOpenGLFramebufferObject class encapsulates an OpenGL framebuffer object.

The QOpenGLFramebufferObject class encapsulates an OpenGL framebuffer object, defined by the GL_EXT_framebuffer_object extension. It provides a rendering surface that can be painted on with a QPainter with the help of QOpenGLPaintDevice, or rendered to using native OpenGL calls. This surface can be bound and used as a regular texture in your own OpenGL drawing code. By default, the QOpenGLFramebufferObject class generates a 2D OpenGL texture (using the GL_TEXTURE_2D target), which is used as the internal rendering target.

It is important to have a current OpenGL context when creating a QOpenGLFramebufferObject, otherwise initialization will fail.

Create the QOpenGLFrameBufferObject instance with the CombinedDepthStencil attachment if you want QPainter to render correctly. Note that you need to create a QOpenGLFramebufferObject with more than one sample per pixel for primitives to be antialiased when drawing using a QPainter. To create a multisample framebuffer object you should use one of the constructors that take a QOpenGLFramebufferObjectFormat parameter, and set the QOpenGLFramebufferObjectFormat::samples() property to a non-zero value.

For multisample framebuffer objects a color render buffer is created, otherwise a texture with the specified texture target is created. The color render buffer or texture will have the specified internal format, and will be bound to the GL_COLOR_ATTACHMENT0 attachment in the framebuffer object.

If you want to use a framebuffer object with multisampling enabled as a texture, you first need to copy from it to a regular framebuffer object using QOpenGLContext::blitFramebuffer().

It is possible to draw into a QOpenGLFramebufferObject using QPainter and QOpenGLPaintDevice in a separate thread.

Methods

impl OpenGLFramebufferObject

fn attachment(&self) -> Attachment

C++ method: QOpenGLFramebufferObject::Attachment QOpenGLFramebufferObject::attachment() const

C++ documentation:

Returns the status of the depth and stencil buffers attached to this framebuffer object.

See also setAttachment().

fn bind(&mut self) -> bool

C++ method: bool QOpenGLFramebufferObject::bind()

C++ documentation:

Switches rendering from the default, windowing system provided framebuffer to this framebuffer object. Returns true upon success, false otherwise.

Note: If takeTexture() was called, a new texture is created and associated with the framebuffer object. This is potentially expensive and changes the context state (the currently bound texture).

See also release().

fn bind_default() -> bool

C++ method: static bool QOpenGLFramebufferObject::bindDefault()

C++ documentation:

Switches rendering back to the default, windowing system provided framebuffer. Returns true upon success, false otherwise.

See also bind() and release().

fn blit_framebuffer<Args>(args: Args) where Args: OpenGLFramebufferObjectBlitFramebufferArgs

C++ method: QOpenGLFramebufferObject::blitFramebuffer

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject)) -> ()
2) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, libc::c_uint)) -> ()
3) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, libc::c_uint, libc::c_uint)) -> ()
C++ method: static void QOpenGLFramebufferObject::blitFramebuffer(QOpenGLFramebufferObject* target, QOpenGLFramebufferObject* source, unsigned int buffers = ?, unsigned int filter = ?)

C++ documentation:

This is an overloaded function.

Convenience overload to blit between two framebuffer objects.

Variant 2

Rust arguments:
1) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect)) -> ()
2) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect, libc::c_uint)) -> ()
3) fn blit_framebuffer((*mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect, *mut ::open_g_l_framebuffer_object::OpenGLFramebufferObject, &qt_core::rect::Rect, libc::c_uint, libc::c_uint)) -> ()
C++ method: static void QOpenGLFramebufferObject::blitFramebuffer(QOpenGLFramebufferObject* target, const QRect& targetRect, QOpenGLFramebufferObject* source, const QRect& sourceRect, unsigned int buffers = ?, unsigned int filter = ?)

C++ documentation:

Blits from the sourceRect rectangle in the source framebuffer object to the targetRect rectangle in the target framebuffer object.

If source or target is 0, the default framebuffer will be used instead of a framebuffer object as source or target respectively.

This function will have no effect unless hasOpenGLFramebufferBlit() returns true.

The buffers parameter should be a mask consisting of any combination of GL_COLOR_BUFFER_BIT, GL_DEPTH_BUFFER_BIT, and GL_STENCIL_BUFFER_BIT. Any buffer type that is not present both in the source and target buffers is ignored.

The sourceRect and targetRect rectangles may have different sizes; in this case buffers should not contain GL_DEPTH_BUFFER_BIT or GL_STENCIL_BUFFER_BIT. The filter parameter should be set to GL_LINEAR or GL_NEAREST, and specifies whether linear or nearest interpolation should be used when scaling is performed.

If source equals target a copy is performed within the same buffer. Results are undefined if the source and target rectangles overlap and have different sizes. The sizes must also be the same if any of the framebuffer objects are multisample framebuffers.

Note: The scissor test will restrict the blit area if enabled.

See also hasOpenGLFramebufferBlit().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLFramebufferObjectFormatArgs<'largs>

C++ method: QOpenGLFramebufferObject::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat>
2) fn format(&self, cpp_utils::AsStruct) -> ::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat
C++ method: QOpenGLFramebufferObjectFormat QOpenGLFramebufferObject::format() const

C++ documentation:

Returns the format of this framebuffer object.

fn handle(&self) -> c_uint

C++ method: unsigned int QOpenGLFramebufferObject::handle() const

C++ documentation:

Returns the OpenGL framebuffer object handle for this framebuffer object (returned by the glGenFrameBuffersEXT() function). This handle can be used to attach new images or buffers to the framebuffer. The user is responsible for cleaning up and destroying these objects.

fn has_opengl_framebuffer_blit() -> bool

C++ method: static bool QOpenGLFramebufferObject::hasOpenGLFramebufferBlit()

C++ documentation:

Returns true if the OpenGL GL_EXT_framebuffer_blit extension is present on this system; otherwise returns false.

See also blitFramebuffer().

fn has_opengl_framebuffer_objects() -> bool

C++ method: static bool QOpenGLFramebufferObject::hasOpenGLFramebufferObjects()

C++ documentation:

Returns true if the OpenGL GL_EXT_framebuffer_object extension is present on this system; otherwise returns false.

fn height(&self) -> c_int

C++ method: int QOpenGLFramebufferObject::height() const

C++ documentation:

Returns the height of the framebuffer object attachments.

fn is_bound(&self) -> bool

C++ method: bool QOpenGLFramebufferObject::isBound() const

C++ documentation:

Returns true if the framebuffer object is currently bound to the current context, otherwise false is returned.

fn is_valid(&self) -> bool

C++ method: bool QOpenGLFramebufferObject::isValid() const

C++ documentation:

Returns true if the framebuffer object is valid.

The framebuffer can become invalid if the initialization process fails, the user attaches an invalid buffer to the framebuffer object, or a non-power of two width/height is specified as the texture size if the texture target is GL_TEXTURE_2D. The non-power of two limitation does not apply if the OpenGL version is 2.0 or higher, or if the GL_ARB_texture_non_power_of_two extension is present.

The framebuffer can also become invalid if the QOpenGLContext that the framebuffer was created within is destroyed and there are no other shared contexts that can take over ownership of the framebuffer.

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFramebufferObjectNewArgs

C++ method: QOpenGLFramebufferObject::QOpenGLFramebufferObject

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
3) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
4) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
5) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
6) fn new((&qt_core::size::Size, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(const QSize& size, QOpenGLFramebufferObject::Attachment attachment, unsigned int target = ?, unsigned int internal_format = ?)

Variant 2

Rust arguments:
1) fn new((&qt_core::size::Size, &::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((&qt_core::size::Size, &::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(const QSize& size, const QOpenGLFramebufferObjectFormat& format)

Variant 3

Rust arguments:
1) fn new((&qt_core::size::Size, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((&qt_core::size::Size, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
3) fn new((&qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
4) fn new((&qt_core::size::Size, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(const QSize& size, unsigned int target = ?)

Variant 4

Rust arguments:
1) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
3) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
4) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
5) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
6) fn new((libc::c_int, libc::c_int, ::open_g_l_framebuffer_object::Attachment, libc::c_uint, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(int width, int height, QOpenGLFramebufferObject::Attachment attachment, unsigned int target = ?, unsigned int internal_format = ?)

Variant 5

Rust arguments:
1) fn new((libc::c_int, libc::c_int, &::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((libc::c_int, libc::c_int, &::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(int width, int height, const QOpenGLFramebufferObjectFormat& format)

Variant 6

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
2) fn new((libc::c_int, libc::c_int, libc::c_uint, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object::OpenGLFramebufferObject
3) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
4) fn new((libc::c_int, libc::c_int, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object::OpenGLFramebufferObject>
C++ method: [constructor] void QOpenGLFramebufferObject::QOpenGLFramebufferObject(int width, int height, unsigned int target = ?)

fn release(&mut self) -> bool

C++ method: bool QOpenGLFramebufferObject::release()

C++ documentation:

Switches rendering back to the default, windowing system provided framebuffer. Returns true upon success, false otherwise.

See also bind().

fn set_attachment(&mut self, attachment: Attachment)

C++ method: void QOpenGLFramebufferObject::setAttachment(QOpenGLFramebufferObject::Attachment attachment)

C++ documentation:

Sets the attachments of the framebuffer object to attachment.

This can be used to free or reattach the depth and stencil buffer attachments as needed.

Note: This function alters the current framebuffer binding.

See also attachment().

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLFramebufferObjectSizeArgs<'largs>

C++ method: QOpenGLFramebufferObject::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QOpenGLFramebufferObject::size() const

C++ documentation:

Returns the size of the texture attached to this framebuffer object.

fn take_texture(&mut self) -> c_uint

C++ method: unsigned int QOpenGLFramebufferObject::takeTexture()

C++ documentation:

Returns the texture id for the texture attached to this framebuffer object. The ownership of the texture is transferred to the caller.

If the framebuffer object is currently bound, an implicit release() will be done. During the next call to bind() a new texture will be created.

If a multisample framebuffer object is used, then there is no texture and the return value from this function will be invalid. Similarly, incomplete framebuffer objects will also return 0.

This function was introduced in Qt 5.3.

See also texture(), bind(), and release().

fn texture(&self) -> c_uint

C++ method: unsigned int QOpenGLFramebufferObject::texture() const

C++ documentation:

Returns the texture id for the texture attached as the default rendering target in this framebuffer object. This texture id can be bound as a normal texture in your own OpenGL code.

If a multisample framebuffer object is used then the value returned from this function will be invalid.

See also takeTexture().

fn to_image<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLFramebufferObjectToImageArgs<'largs>

C++ method: QOpenGLFramebufferObject::toImage

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_image(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::image::Image>
2) fn to_image(&self, cpp_utils::AsStruct) -> ::image::Image
C++ method: QImage QOpenGLFramebufferObject::toImage() const

C++ documentation:

This is an overloaded function.

Returns the contents of this framebuffer object as a QImage. This method flips the image from OpenGL coordinates to raster coordinates.

Variant 2

Rust arguments:
1) fn to_image(&self, (bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image::Image>
2) fn to_image(&self, (bool, cpp_utils::AsStruct)) -> ::image::Image
C++ method: QImage QOpenGLFramebufferObject::toImage(bool flipped) const

C++ documentation:

Returns the contents of this framebuffer object as a QImage.

If flipped is true the image is flipped from OpenGL coordinates to raster coordinates. If used together with QOpenGLPaintDevice, flipped should be the opposite of the value of QOpenGLPaintDevice::paintFlipped().

The returned image has a format of premultiplied ARGB32 or RGB32. The latter is used only when internalTextureFormat() is set to GL_RGB.

If the rendering in the framebuffer was not done with premultiplied alpha in mind, create a wrapper QImage with a non-premultiplied format. This is necessary before performing operations like QImage::save() because otherwise the image data would get unpremultiplied, even though it was not premultiplied in the first place. To create such a wrapper without performing a copy of the pixel data, do the following:

QImage fboImage(fbo.toImage());
QImage image(fboImage.constBits(), fboImage.width(), fboImage.height(), QImage::Format_ARGB32);

Since Qt 5.2 the function will fall back to premultiplied RGBA8888 or RGBx8888 when reading to (A)RGB32 is not supported. Since 5.4 an A2BGR30 image is returned if the internal format is RGB10_A2.

For multisampled framebuffer objects the samples are resolved using the GL_EXT_framebuffer_blit extension. If the extension is not available, the contents of the returned image is undefined.

For singlesampled framebuffers the contents is retrieved via glReadPixels. This is a potentially expensive and inefficient operation. Therefore it is recommended that this function is used as seldom as possible.

See also QOpenGLPaintDevice::paintFlipped().

fn width(&self) -> c_int

C++ method: int QOpenGLFramebufferObject::width() const

C++ documentation:

Returns the width of the framebuffer object attachments.