Struct qt_gui::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat

pub struct OpenGLFramebufferObjectFormat {
    // some fields omitted
}

C++ type: QOpenGLFramebufferObjectFormat

C++ documentation:

The QOpenGLFramebufferObjectFormat class specifies the format of an OpenGL framebuffer object.

A framebuffer object has several characteristics:

Number of samples per pixels.

Depth and/or stencil attachments.

Texture target.

Internal texture format.

Note that the desired attachments or number of samples per pixels might not be supported by the hardware driver. Call QOpenGLFramebufferObject::format() after creating a QOpenGLFramebufferObject to find the exact format that was used to create the frame buffer object.

Methods

impl OpenGLFramebufferObjectFormat

fn attachment(&self) -> Attachment

C++ method: QOpenGLFramebufferObject::Attachment QOpenGLFramebufferObjectFormat::attachment() const

C++ documentation:

Returns the configuration of the depth and stencil buffers attached to a framebuffer object. The default is QOpenGLFramebufferObject::NoAttachment.

See also setAttachment().

fn internal_texture_format(&self) -> c_uint

C++ method: unsigned int QOpenGLFramebufferObjectFormat::internalTextureFormat() const

C++ documentation:

Returns the internal format of a framebuffer object's texture or multisample framebuffer object's color buffer. The default is GL_RGBA8 on desktop OpenGL systems, and GL_RGBA on OpenGL/ES systems.

See also setInternalTextureFormat().

fn mipmap(&self) -> bool

C++ method: bool QOpenGLFramebufferObjectFormat::mipmap() const

C++ documentation:

Returns true if mipmapping is enabled.

See also setMipmap().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFramebufferObjectFormatNewArgs

C++ method: QOpenGLFramebufferObjectFormat::QOpenGLFramebufferObjectFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat>
C++ method: [constructor] void QOpenGLFramebufferObjectFormat::QOpenGLFramebufferObjectFormat()

Variant 2

Rust arguments:
1) fn new((&::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsStruct)) -> ::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat
2) fn new((&::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_framebuffer_object_format::OpenGLFramebufferObjectFormat>
C++ method: [constructor] void QOpenGLFramebufferObjectFormat::QOpenGLFramebufferObjectFormat(const QOpenGLFramebufferObjectFormat& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 OpenGLFramebufferObjectFormat) -> &'l0 mut OpenGLFramebufferObjectFormat

C++ method: QOpenGLFramebufferObjectFormat& QOpenGLFramebufferObjectFormat::operator=(const QOpenGLFramebufferObjectFormat& other)

C++ documentation:

Assigns other to this object.

fn op_eq(&self, other: &OpenGLFramebufferObjectFormat) -> bool

C++ method: bool QOpenGLFramebufferObjectFormat::operator==(const QOpenGLFramebufferObjectFormat& other) const

C++ documentation:

Returns true if all the options of this framebuffer object format are the same as other; otherwise returns false.

fn op_neq(&self, other: &OpenGLFramebufferObjectFormat) -> bool

C++ method: bool QOpenGLFramebufferObjectFormat::operator!=(const QOpenGLFramebufferObjectFormat& other) const

C++ documentation:

Returns false if all the options of this framebuffer object format are the same as other; otherwise returns true.

fn samples(&self) -> c_int

C++ method: int QOpenGLFramebufferObjectFormat::samples() const

C++ documentation:

Returns the number of samples per pixel if a framebuffer object is a multisample framebuffer object. Otherwise, returns 0. The default value is 0.

See also setSamples().

fn set_attachment(&mut self, attachment: Attachment)

C++ method: void QOpenGLFramebufferObjectFormat::setAttachment(QOpenGLFramebufferObject::Attachment attachment)

C++ documentation:

Sets the attachment configuration of a framebuffer object to attachment.

See also attachment().

fn set_internal_texture_format(&mut self, internal_texture_format: c_uint)

C++ method: void QOpenGLFramebufferObjectFormat::setInternalTextureFormat(unsigned int internalTextureFormat)

C++ documentation:

Sets the internal format of a framebuffer object's texture or multisample framebuffer object's color buffer to internalTextureFormat.

See also internalTextureFormat().

fn set_mipmap(&mut self, enabled: bool)

C++ method: void QOpenGLFramebufferObjectFormat::setMipmap(bool enabled)

C++ documentation:

Enables mipmapping if enabled is true; otherwise disables it.

Mipmapping is disabled by default.

If mipmapping is enabled, additional memory will be allocated for the mipmap levels. The mipmap levels can be updated by binding the texture and calling glGenerateMipmap(). Mipmapping cannot be enabled for multisampled framebuffer objects.

See also mipmap() and QOpenGLFramebufferObject::texture().

fn set_samples(&mut self, samples: c_int)

C++ method: void QOpenGLFramebufferObjectFormat::setSamples(int samples)

C++ documentation:

Sets the number of samples per pixel for a multisample framebuffer object to samples. The default sample count of 0 represents a regular non-multisample framebuffer object.

If the desired amount of samples per pixel is not supported by the hardware then the maximum number of samples per pixel will be used. Note that multisample framebuffer objects can not be bound as textures. Also, the GL_EXT_framebuffer_multisample extension is required to create a framebuffer with more than one sample per pixel.

See also samples().

fn set_texture_target(&mut self, target: c_uint)

C++ method: void QOpenGLFramebufferObjectFormat::setTextureTarget(unsigned int target)

C++ documentation:

Sets the texture target of the texture attached to a framebuffer object to target. Ignored for multisample framebuffer objects.

See also textureTarget() and samples().

fn texture_target(&self) -> c_uint

C++ method: unsigned int QOpenGLFramebufferObjectFormat::textureTarget() const

C++ documentation:

Returns the texture target of the texture attached to a framebuffer object. Ignored for multisample framebuffer objects. The default is GL_TEXTURE_2D.

See also setTextureTarget() and samples().