Struct qt_gui::open_g_l_paint_device::OpenGLPaintDevice

pub struct OpenGLPaintDevice {
    // some fields omitted
}

C++ type: QOpenGLPaintDevice

C++ documentation:

The QOpenGLPaintDevice class enables painting to an OpenGL context using QPainter.

The QOpenGLPaintDevice uses the current QOpenGL context to render QPainter draw commands. The context is captured upon construction. It requires support for OpenGL (ES) 2.0 or higher.

Methods

impl OpenGLPaintDevice

fn color_count(&self) -> c_int

C++ method: int QOpenGLPaintDevice::colorCount() const

C++ documentation:

Returns the number of different colors available for the paint device. If the number of colors available is too great to be represented by the int data type, then INT_MAX will be returned instead.

fn context(&self) -> *mut OpenGLContext

C++ method: QOpenGLContext* QOpenGLPaintDevice::context() const

C++ documentation:

Returns the OpenGL context associated with the paint device.

fn depth(&self) -> c_int

C++ method: int QOpenGLPaintDevice::depth() const

C++ documentation:

Returns the bit depth (number of bit planes) of the paint device.

fn dev_type(&self) -> c_int

C++ method: virtual int QOpenGLPaintDevice::devType() const

fn device_pixel_ratio(&self) -> c_int

C++ method: int QOpenGLPaintDevice::devicePixelRatio() const

C++ documentation:

Returns the device pixel ratio for device.

Common values are 1 for normal-dpi displays and 2 for high-dpi "retina" displays.

fn dots_per_meter_x(&self) -> c_double

C++ method: double QOpenGLPaintDevice::dotsPerMeterX() const

C++ documentation:

Returns the number of pixels per meter horizontally.

See also setDotsPerMeterX().

fn dots_per_meter_y(&self) -> c_double

C++ method: double QOpenGLPaintDevice::dotsPerMeterY() const

C++ documentation:

Returns the number of pixels per meter vertically.

See also setDotsPerMeterY().

fn ensure_active_target(&mut self)

C++ method: virtual void QOpenGLPaintDevice::ensureActiveTarget()

C++ documentation:

This virtual method is provided as a callback to allow re-binding a target frame buffer object or context when different QOpenGLPaintDevice instances are issuing draw calls alternately.

beginNativePainting() will also trigger this method.

The default implementation does nothing.

fn height(&self) -> c_int

C++ method: int QOpenGLPaintDevice::height() const

C++ documentation:

Returns the height of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also heightMM().

fn height_m_m(&self) -> c_int

C++ method: int QOpenGLPaintDevice::heightMM() const

C++ documentation:

Returns the height of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also height().

fn logical_dpi_x(&self) -> c_int

C++ method: int QOpenGLPaintDevice::logicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from widthMM().

Note that if the logicalDpiX() doesn't equal the physicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiY() and physicalDpiX().

fn logical_dpi_y(&self) -> c_int

C++ method: int QOpenGLPaintDevice::logicalDpiY() const

C++ documentation:

Returns the vertical resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from heightMM().

Note that if the logicalDpiY() doesn't equal the physicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiX() and physicalDpiY().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLPaintDeviceNewArgs

C++ method: QOpenGLPaintDevice::QOpenGLPaintDevice

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_paint_device::OpenGLPaintDevice
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_paint_device::OpenGLPaintDevice>
C++ method: [constructor] void QOpenGLPaintDevice::QOpenGLPaintDevice()

Variant 2

Rust arguments:
1) fn new((&qt_core::size::Size, cpp_utils::AsStruct)) -> ::open_g_l_paint_device::OpenGLPaintDevice
2) fn new((&qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_paint_device::OpenGLPaintDevice>
C++ method: [constructor] void QOpenGLPaintDevice::QOpenGLPaintDevice(const QSize& size)

Variant 3

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::open_g_l_paint_device::OpenGLPaintDevice
2) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_paint_device::OpenGLPaintDevice>
C++ method: [constructor] void QOpenGLPaintDevice::QOpenGLPaintDevice(int width, int height)

fn paint_engine(&self) -> *mut PaintEngine

C++ method: virtual QPaintEngine* QOpenGLPaintDevice::paintEngine() const

C++ documentation:

Reimplemented from QPaintDevice::paintEngine().

fn paint_flipped(&self) -> bool

C++ method: bool QOpenGLPaintDevice::paintFlipped() const

C++ documentation:

Returns true if painting is flipped around the Y-axis.

See also setPaintFlipped().

fn painting_active(&self) -> bool

C++ method: bool QOpenGLPaintDevice::paintingActive() const

C++ documentation:

Returns true if the device is currently being painted on, i.e. someone has called QPainter::begin() but not yet called QPainter::end() for this device; otherwise returns false.

See also QPainter::isActive().

fn physical_dpi_x(&self) -> c_int

C++ method: int QOpenGLPaintDevice::physicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiX() doesn't equal the logicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiY() and logicalDpiX().

fn physical_dpi_y(&self) -> c_int

C++ method: int QOpenGLPaintDevice::physicalDpiY() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiY() doesn't equal the logicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiX() and logicalDpiY().

fn set_device_pixel_ratio(&mut self, device_pixel_ratio: c_double)

C++ method: void QOpenGLPaintDevice::setDevicePixelRatio(double devicePixelRatio)

C++ documentation:

Sets the device pixel ratio for the paint device to devicePixelRatio.

fn set_dots_per_meter_x(&mut self, arg1: c_double)

C++ method: void QOpenGLPaintDevice::setDotsPerMeterX(double arg1)

C++ documentation:

Sets the number of pixels per meter horizontally to dpmx.

See also dotsPerMeterX().

fn set_dots_per_meter_y(&mut self, arg1: c_double)

C++ method: void QOpenGLPaintDevice::setDotsPerMeterY(double arg1)

C++ documentation:

Sets the number of pixels per meter vertically to dpmy.

See also dotsPerMeterY().

fn set_paint_flipped(&mut self, flipped: bool)

C++ method: void QOpenGLPaintDevice::setPaintFlipped(bool flipped)

C++ documentation:

Sets whether painting should be flipped around the Y-axis or not to flipped.

See also paintFlipped().

fn set_size(&mut self, size: &Size)

C++ method: void QOpenGLPaintDevice::setSize(const QSize& size)

C++ documentation:

Sets the pixel size of the paint device to size.

See also size().

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLPaintDeviceSizeArgs<'largs>

C++ method: QOpenGLPaintDevice::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QOpenGLPaintDevice::size() const

C++ documentation:

Returns the pixel size of the paint device.

See also setSize().

fn width(&self) -> c_int

C++ method: int QOpenGLPaintDevice::width() const

C++ documentation:

Returns the width of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also widthMM().

fn width_m_m(&self) -> c_int

C++ method: int QOpenGLPaintDevice::widthMM() const

C++ documentation:

Returns the width of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also width().