Struct qt_gui::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions

pub struct OpenGLPixelTransferOptions {
    // some fields omitted
}

C++ type: QOpenGLPixelTransferOptions

C++ documentation:

The QOpenGLPixelTransferOptions class describes the pixel storage modes that affect the unpacking of pixels during texture upload.

Methods

impl OpenGLPixelTransferOptions

fn alignment(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::alignment() const

C++ documentation:

Returns the current alignment requirement for each pixel row.

See also setAlignment().

fn image_height(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::imageHeight() const

C++ documentation:

Returns the currently set image height.

See also setImageHeight().

fn is_least_significant_bit_first(&self) -> bool

C++ method: bool QOpenGLPixelTransferOptions::isLeastSignificantBitFirst() const

C++ documentation:

Returns true if bits within a byte are ordered from least to most significant.

fn is_swap_bytes_enabled(&self) -> bool

C++ method: bool QOpenGLPixelTransferOptions::isSwapBytesEnabled() const

C++ documentation:

Returns true if the byte ordering for multibyte components is reversed.

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLPixelTransferOptionsNewArgs

C++ method: QOpenGLPixelTransferOptions::QOpenGLPixelTransferOptions

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions>
C++ method: [constructor] void QOpenGLPixelTransferOptions::QOpenGLPixelTransferOptions()

Variant 2

Rust arguments:
1) fn new((&::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions, cpp_utils::AsStruct)) -> ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions
2) fn new((&::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions>
C++ method: [constructor] void QOpenGLPixelTransferOptions::QOpenGLPixelTransferOptions(const QOpenGLPixelTransferOptions& arg1)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 OpenGLPixelTransferOptions) -> &'l0 mut OpenGLPixelTransferOptions

C++ method: QOpenGLPixelTransferOptions& QOpenGLPixelTransferOptions::operator=(const QOpenGLPixelTransferOptions& arg1)

fn row_length(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::rowLength() const

C++ documentation:

Returns the currently set row length.

See also setRowLength().

fn set_alignment(&mut self, alignment: c_int)

C++ method: void QOpenGLPixelTransferOptions::setAlignment(int alignment)

C++ documentation:

Sets the alignment requirements for each pixel row. Corresponds to GL_UNPACK_ALIGNMENT. The default value is 4, as specified by OpenGL.

See also alignment().

fn set_image_height(&mut self, image_height: c_int)

C++ method: void QOpenGLPixelTransferOptions::setImageHeight(int imageHeight)

C++ documentation:

Sets the image height for 3D textures to imageHeight. Corresponds to GL_UNPACK_IMAGE_HEIGHT. The default value is 0.

See also imageHeight().

fn set_least_significant_byte_first(&mut self, lsb_first: bool)

C++ method: void QOpenGLPixelTransferOptions::setLeastSignificantByteFirst(bool lsbFirst)

C++ documentation:

lsbFirst specifies if bits within a byte are ordered from least to most significat. The default value is false, meaning that the first bit in each byte is the most significant one. This is significant for bitmap data only. Corresponds to GL_UNPACK_LSB_FIRST.

fn set_row_length(&mut self, row_length: c_int)

C++ method: void QOpenGLPixelTransferOptions::setRowLength(int rowLength)

C++ documentation:

Sets the number of pixels in a row to rowLength. Corresponds to GL_UNPACK_ROW_LENGTH. The default value is 0.

See also rowLength().

fn set_skip_images(&mut self, skip_images: c_int)

C++ method: void QOpenGLPixelTransferOptions::setSkipImages(int skipImages)

C++ documentation:

Sets the number of images that are skipped to skipImages. Corresponds to GL_UNPACK_SKIP_IMAGES. Equivalent to incrementing the pointer passed to QOpenGLTexture::setData(). The default value is 0.

See also skipImages().

fn set_skip_pixels(&mut self, skip_pixels: c_int)

C++ method: void QOpenGLPixelTransferOptions::setSkipPixels(int skipPixels)

C++ documentation:

Sets the number of pixels that are skipped to skipPixels. Corresponds to GL_UNPACK_SKIP_PIXELS. Equivalent to incrementing the pointer passed to QOpenGLTexture::setData(). The default value is 0.

See also skipPixels().

fn set_skip_rows(&mut self, skip_rows: c_int)

C++ method: void QOpenGLPixelTransferOptions::setSkipRows(int skipRows)

C++ documentation:

Sets the number of rows that are skipped to skipRows. Corresponds to GL_UNPACK_SKIP_ROWS. Equivalent to incrementing the pointer passed to QOpenGLTexture::setData(). The default value is 0.

See also skipRows().

fn set_swap_bytes_enabled(&mut self, swap_bytes: bool)

C++ method: void QOpenGLPixelTransferOptions::setSwapBytesEnabled(bool swapBytes)

C++ documentation:

swapBytes specifies if the byte ordering for multibyte components is reversed. The default value is false. Corresponds to GL_UNPACK_SWAP_BYTES.

See also isSwapBytesEnabled().

fn skip_images(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::skipImages() const

C++ documentation:

Returns the number of images that are skipped.

See also setSkipImages().

fn skip_pixels(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::skipPixels() const

C++ documentation:

Returns the number of pixels that are skipped.

See also setSkipPixels().

fn skip_rows(&self) -> c_int

C++ method: int QOpenGLPixelTransferOptions::skipRows() const

C++ documentation:

Returns the number of rows that are skipped.

See also setSkipRows().

fn swap(&mut self, other: &mut OpenGLPixelTransferOptions)

C++ method: void QOpenGLPixelTransferOptions::swap(QOpenGLPixelTransferOptions& other)