Module qt_gui::open_g_l_shader::overloading

Traits

OpenGLShaderCompileSourceCodeArgs
This trait represents a set of arguments accepted by OpenGLShader::compile_source_code method.

OpenGLShaderConnectArgs
This trait represents a set of arguments accepted by OpenGLShader::connect method.

OpenGLShaderConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLShader::connect_static method.

OpenGLShaderDestroyedArgs
This trait represents a set of arguments accepted by OpenGLShader::destroyed method.

OpenGLShaderDisconnectArgs
This trait represents a set of arguments accepted by OpenGLShader::disconnect method.

OpenGLShaderDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLShader::disconnect_static method.

OpenGLShaderDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLShader::dynamic_property_names method.

OpenGLShaderHasOpenglShadersArgs
This trait represents a set of arguments accepted by OpenGLShader::has_opengl_shaders method.

OpenGLShaderLogArgs
This trait represents a set of arguments accepted by OpenGLShader::log method.

OpenGLShaderNewArgs
This trait represents a set of arguments accepted by OpenGLShader::new method.

OpenGLShaderObjectNameArgs
This trait represents a set of arguments accepted by OpenGLShader::object_name method.

OpenGLShaderPropertyArgs
This trait represents a set of arguments accepted by OpenGLShader::property method.

OpenGLShaderSourceCodeArgs
This trait represents a set of arguments accepted by OpenGLShader::source_code method.

OpenGLShaderStartTimerArgs
This trait represents a set of arguments accepted by OpenGLShader::start_timer method.

OpenGLShaderTrArgs
This trait represents a set of arguments accepted by OpenGLShader::tr method.

OpenGLShaderTrUtf8Args
This trait represents a set of arguments accepted by OpenGLShader::tr_utf8 method.