Module qt_gui::open_g_l_shader_program::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpenGLShaderProgramAddShaderFromSourceCodeArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::add_shader_from_source_code method.

OpenGLShaderProgramAttributeLocationArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::attribute_location method.

OpenGLShaderProgramBindAttributeLocationArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::bind_attribute_location method.

OpenGLShaderProgramConnectArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::connect method.

OpenGLShaderProgramConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::connect_static method.

OpenGLShaderProgramDefaultInnerTessellationLevelsArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::default_inner_tessellation_levels method.

OpenGLShaderProgramDefaultOuterTessellationLevelsArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::default_outer_tessellation_levels method.

OpenGLShaderProgramDestroyedArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::destroyed method.

OpenGLShaderProgramDisableAttributeArrayArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::disable_attribute_array method.

OpenGLShaderProgramDisconnectArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::disconnect method.

OpenGLShaderProgramDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::disconnect_static method.

OpenGLShaderProgramDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::dynamic_property_names method.

OpenGLShaderProgramEnableAttributeArrayArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::enable_attribute_array method.

OpenGLShaderProgramHasOpenglShaderProgramsArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::has_opengl_shader_programs method.

OpenGLShaderProgramLogArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::log method.

OpenGLShaderProgramNewArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::new method.

OpenGLShaderProgramObjectNameArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::object_name method.

OpenGLShaderProgramPropertyArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::property method.

OpenGLShaderProgramSetAttributeArrayArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::set_attribute_array method.

OpenGLShaderProgramSetAttributeBufferArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::set_attribute_buffer method.

OpenGLShaderProgramSetAttributeValueArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::set_attribute_value method.

OpenGLShaderProgramSetUniformValueArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::set_uniform_value method.

OpenGLShaderProgramSetUniformValueArrayArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::set_uniform_value_array method.

OpenGLShaderProgramShadersArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::shaders method.

OpenGLShaderProgramStartTimerArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::start_timer method.

OpenGLShaderProgramTrArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::tr method.

OpenGLShaderProgramTrUtf8Args
This trait represents a set of arguments accepted by OpenGLShaderProgram::tr_utf8 method.

OpenGLShaderProgramUniformLocationArgs
This trait represents a set of arguments accepted by OpenGLShaderProgram::uniform_location method.