Enum qt_gui::open_g_l_texture::ComparisonFunction

pub enum ComparisonFunction {
    CompareNever,
    CompareLess,
    CompareEqual,
    CompareLessEqual,
    CompareGreater,
    CommpareNotEqual,
    CompareGreaterEqual,
    CompareAlways,
}

C++ type: QOpenGLTexture::ComparisonFunction

C++ documentation:

This enum specifies which comparison operator is used when texture comparison is enabled on this texture.

This enum was introduced or modified in Qt 5.5.

Variants
CompareNever

Equivalent to GL_NEVER. (C++ enum variant: CompareNever = 512)

CompareLess

Equivalent to GL_LESS. (C++ enum variant: CompareLess = 513)

CompareEqual

Equivalent to GL_EQUAL. (C++ enum variant: CompareEqual = 514)

CompareLessEqual

Equivalent to GL_LEQUAL. (C++ enum variant: CompareLessEqual = 515)

CompareGreater

Equivalent to GL_GREATER. (C++ enum variant: CompareGreater = 516)

CommpareNotEqual

Equivalent to GL_NOTEQUAL. (C++ enum variant: CommpareNotEqual = 517)

CompareGreaterEqual

Equivalent to GL_GEQUAL. (C++ enum variant: CompareGreaterEqual = 518)

CompareAlways

Equivalent to GL_ALWAYS. (C++ enum variant: CompareAlways = 519)