Struct qt_gui::open_g_l_texture::OpenGLTexture

pub struct OpenGLTexture {
    // some fields omitted
}

C++ type: QOpenGLTexture

C++ documentation:

The QOpenGLTexture class encapsulates an OpenGL texture object.

QOpenGLTexture makes it easy to work with OpenGL textures and the myriad features and targets that they offer depending upon the capabilities of your OpenGL implementation.

The typical usage pattern for QOpenGLTexture is

Instantiate the object specifying the texture target type

Set properties that affect the storage requirements e.g. storage format, dimensions

Allocate the server-side storage

Optionally upload pixel data

Optionally set any additional properties e.g. filtering and border options

Render with texture or render to texture

In the common case of simply using a QImage as the source of texture pixel data most of the above steps are performed automatically.

// Prepare texture
QOpenGLTexture *texture = new QOpenGLTexture(QImage(fileName).mirrored());
texture->setMinificationFilter(QOpenGLTexture::LinearMipMapLinear);
texture->setMagnificationFilter(QOpenGLTexture::Linear);
...
// Render with texture
texture->bind();
glDrawArrays(...);

Note that the QImage is mirrored vertically to account for the fact that OpenGL and QImage use opposite directions for the y axis. Another option would be to transform your texture coordinates.

Methods

impl OpenGLTexture

fn allocate_storage<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureAllocateStorageArgs<'largs>

C++ method: QOpenGLTexture::allocateStorage

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn allocate_storage(&mut self, ()) -> ()
C++ method: void QOpenGLTexture::allocateStorage()

C++ documentation:

Allocates server-side storage for this texture object taking into account, the format, dimensions, mipmap levels, array layers and cubemap faces.

Once storage has been allocated it is no longer possible to change these properties.

If supported QOpenGLTexture makes use of immutable texture storage.

Once storage has been allocated for the texture then pixel data can be uploaded via one of the setData() overloads.

Note: If immutable texture storage is not available, then a default pixel format and pixel type will be used to create the mutable storage. You can use the other allocateStorage() overload to specify exactly the pixel format and the pixel type to use when allocating mutable storage; this is particulary useful under certain OpenGL ES implementations (notably, OpenGL ES 2), where the pixel format and the pixel type used at allocation time must perfectly match the format and the type passed to any subsequent setData() call.

See also isStorageAllocated() and setData().

Variant 2

Rust arguments: fn allocate_storage(&mut self, (::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType)) -> ()
C++ method: void QOpenGLTexture::allocateStorage(QOpenGLTexture::PixelFormat pixelFormat, QOpenGLTexture::PixelType pixelType)

C++ documentation:

Allocates server-side storage for this texture object taking into account, the format, dimensions, mipmap levels, array layers and cubemap faces.

Once storage has been allocated it is no longer possible to change these properties.

If supported QOpenGLTexture makes use of immutable texture storage. However, if immutable texture storage is not available, then the specified pixelFormat and pixelType will be used to allocate mutable storage; note that in certain OpenGL implementations (notably, OpenGL ES 2) they must perfectly match the format and the type passed to any subsequent setData() call.

Once storage has been allocated for the texture then pixel data can be uploaded via one of the setData() overloads.

This function was introduced in Qt 5.5.

See also isStorageAllocated() and setData().

fn bind<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureBindArgs<'largs>

C++ method: QOpenGLTexture::bind

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn bind(&mut self, ()) -> ()
C++ method: void QOpenGLTexture::bind()

C++ documentation:

Binds this texture to the currently active texture unit ready for rendering. Note that you do not need to bind QOpenGLTexture objects in order to modify them as the implementation makes use of the EXT_direct_state_access extension where available and simulates it where it is not.

See also release().

Variant 2

Rust arguments:
1) fn bind(&mut self, libc::c_uint) -> ()
2) fn bind(&mut self, (libc::c_uint, ::open_g_l_texture::TextureUnitReset)) -> ()
C++ method: void QOpenGLTexture::bind(unsigned int unit, QOpenGLTexture::TextureUnitReset reset = ?)

C++ documentation:

Binds this texture to texture unit unit ready for rendering. Note that you do not need to bind QOpenGLTexture objects in order to modify them as the implementation makes use of the EXT_direct_state_access extension where available and simulates it where it is not.

If parameter reset is true then this function will restore the active unit to the texture unit that was active upon entry.

See also release().

fn border_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLTextureBorderColorArgs<'largs>

C++ method: QOpenGLTexture::borderColor

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn border_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn border_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: QColor QOpenGLTexture::borderColor() const

C++ documentation:

Returns the borderColor of this texture.

See also setBorderColor().

Variant 2

Rust arguments: fn border_color(&self, *mut libc::c_float) -> ()
C++ method: void QOpenGLTexture::borderColor(float* border) const

C++ documentation:

Writes the texture border color into the first four elements of the array pointed to by border.

See also setBorderColor().

Variant 3

Rust arguments: fn border_color(&self, *mut libc::c_int) -> ()
C++ method: void QOpenGLTexture::borderColor(int* border) const

C++ documentation:

Writes the texture border color into the first four elements of the array pointed to by border.

This is an overloaded function.

Variant 4

Rust arguments: fn border_color(&self, *mut libc::c_uint) -> ()
C++ method: void QOpenGLTexture::borderColor(unsigned int* border) const

C++ documentation:

Writes the texture border color into the first four elements of the array pointed to by border.

This is an overloaded function.

fn bound_texture_id<Args>(args: Args) -> c_uint where Args: OpenGLTextureBoundTextureIdArgs

C++ method: QOpenGLTexture::boundTextureId

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn bound_texture_id(::open_g_l_texture::BindingTarget) -> libc::c_uint
C++ method: static unsigned int QOpenGLTexture::boundTextureId(QOpenGLTexture::BindingTarget target)

C++ documentation:

Returns the textureId of the texture that is bound to the target of the currently active texture unit.

Variant 2

Rust arguments: fn bound_texture_id((libc::c_uint, ::open_g_l_texture::BindingTarget)) -> libc::c_uint
C++ method: static unsigned int QOpenGLTexture::boundTextureId(unsigned int unit, QOpenGLTexture::BindingTarget target)

C++ documentation:

Returns the textureId of the texture that is bound to the target of the texture unit unit.

fn comparison_function(&self) -> ComparisonFunction

C++ method: QOpenGLTexture::ComparisonFunction QOpenGLTexture::comparisonFunction() const

C++ documentation:

Returns the texture comparison operator set on this texture. By default, a texture has a CompareLessEqual comparison function.

This function was introduced in Qt 5.5.

See also setComparisonFunction().

fn comparison_mode(&self) -> ComparisonMode

C++ method: QOpenGLTexture::ComparisonMode QOpenGLTexture::comparisonMode() const

C++ documentation:

Returns the texture comparison mode set on this texture. By default, a texture has a CompareNone comparison mode (i.e. comparisons are disabled).

This function was introduced in Qt 5.5.

See also setComparisonMode().

fn create(&mut self) -> bool

C++ method: bool QOpenGLTexture::create()

C++ documentation:

Creates the underlying OpenGL texture object. This requires a current valid OpenGL context. If the texture object already exists, this function does nothing.

Once the texture object is created you can obtain the object name from the textureId() function. This may be useful if you wish to make some raw OpenGL calls related to this texture.

Normally it should not be necessary to call this function directly as all functions that set properties of the texture object implicitly call create() on your behalf.

Returns true if the creation succeeded, otherwise returns false.

See also destroy(), isCreated(), and textureId().

fn create_texture_view(&self, target: Target, view_format: TextureFormat, minimum_mipmap_level: c_int, maximum_mipmap_level: c_int, minimum_layer: c_int, maximum_layer: c_int) -> *mut OpenGLTexture

C++ method: QOpenGLTexture* QOpenGLTexture::createTextureView(QOpenGLTexture::Target target, QOpenGLTexture::TextureFormat viewFormat, int minimumMipmapLevel, int maximumMipmapLevel, int minimumLayer, int maximumLayer) const

C++ documentation:

Attempts to create a texture view onto this texture. A texture view is somewhat analogous to a view in SQL in that it presents a restricted or reinterpreted view of the original data. Texture views do not allocate any more server-side storage, insted relying on the storage buffer of the source texture.

Texture views are only available when using immutable storage. For more information on texture views see http://www.opengl.org/wiki/Texture_Storage#Texture_views.

The target argument specifies the target to use for the view. Only some targets can be used depending upon the target of the original target. For e.g. a view onto a Target1DArray texture can specify either Target1DArray or Target1D but for the latter the number of array layers specified with minimumLayer and maximumLayer must be exactly 1.

Simpliar constraints apply for the viewFormat. See the above link and the specification for more details.

The minimumMipmapLevel, maximumMipmapLevel, minimumLayer, and maximumLayer arguments serve to restrict the parts of the texture accessible by the texture view.

If creation of the texture view fails this function will return 0. If the function succeeds it will return a pointer to a new QOpenGLTexture object that will return true from its isTextureView() function.

See also isTextureView().

fn depth(&self) -> c_int

C++ method: int QOpenGLTexture::depth() const

C++ documentation:

Returns the depth of a 3D texture.

See also width(), height(), and setSize().

fn depth_stencil_mode(&self) -> DepthStencilMode

C++ method: QOpenGLTexture::DepthStencilMode QOpenGLTexture::depthStencilMode() const

C++ documentation:

Returns the depth stencil mode for textures using a combined depth/stencil format.

This function was introduced in Qt 5.4.

See also setDepthStencilMode().

fn destroy(&mut self)

C++ method: void QOpenGLTexture::destroy()

C++ documentation:

Destroys the underlying OpenGL texture object. This requires a current valid OpenGL context.

See also create(), isCreated(), and textureId().

fn faces(&self) -> c_int

C++ method: int QOpenGLTexture::faces() const

C++ documentation:

Returns the number of faces for this texture. For cubemap and cubemap array type targets this will be 6.

For non-cubemap type targets this will return 1.

fn format(&self) -> TextureFormat

C++ method: QOpenGLTexture::TextureFormat QOpenGLTexture::format() const

C++ documentation:

Returns the format of this texture object.

See also setFormat().

fn generate_mip_maps<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureGenerateMipMapsArgs<'largs>

C++ method: QOpenGLTexture::generateMipMaps

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn generate_mip_maps(&mut self, ()) -> ()
C++ method: void QOpenGLTexture::generateMipMaps()

C++ documentation:

Generates mipmaps for this texture object from mipmap level 0. If you are using a texture target and filtering option that requires mipmaps and you have disabled automatic mipmap generation then you need to call this function or the overload to create the mipmap chain.

See also setAutoMipMapGenerationEnabled(), setMipLevels(), and mipLevels().

Variant 2

Rust arguments:
1) fn generate_mip_maps(&mut self, libc::c_int) -> ()
2) fn generate_mip_maps(&mut self, (libc::c_int, bool)) -> ()
C++ method: void QOpenGLTexture::generateMipMaps(int baseLevel, bool resetBaseLevel = ?)

C++ documentation:

Generates mipmaps for this texture object from mipmap level baseLevel. If you are using a texture target and filtering option that requires mipmaps and you have disabled automatic mipmap generation then you need to call this function or the overload to create the mipmap chain.

The generation of mipmaps to above baseLevel is achieved by setting the mipmap base level to baseLevel and then generating the mipmap chain. If resetBaseLevel is true, then the baseLevel of the texture will be reset to its previous value.

See also setAutoMipMapGenerationEnabled(), setMipLevels(), and mipLevels().

fn has_feature(feature: Feature) -> bool

C++ method: static bool QOpenGLTexture::hasFeature(QOpenGLTexture::Feature feature)

C++ documentation:

Returns true if your OpenGL implementation and version supports the texture feature feature.

fn height(&self) -> c_int

C++ method: int QOpenGLTexture::height() const

C++ documentation:

Returns the height of a 2D or 3D texture.

See also width(), depth(), and setSize().

fn is_auto_mip_map_generation_enabled(&self) -> bool

C++ method: bool QOpenGLTexture::isAutoMipMapGenerationEnabled() const

C++ documentation:

Returns whether auto mipmap generation is enabled for this texture object.

See also setAutoMipMapGenerationEnabled() and generateMipMaps().

fn is_bound(&self) -> bool

C++ method: bool QOpenGLTexture::isBound() const

C++ documentation:

Returns true if this texture is bound to the corresponding target of the currently active texture unit.

See also bind() and release().

fn is_bound_mut(&mut self, unit: c_uint) -> bool

C++ method: bool QOpenGLTexture::isBound(unsigned int unit)

C++ documentation:

Returns true if this texture is bound to the corresponding target of texture unit unit.

See also bind() and release().

fn is_created(&self) -> bool

C++ method: bool QOpenGLTexture::isCreated() const

C++ documentation:

Returns true if the underlying OpenGL texture object has been created.

See also create(), destroy(), and textureId().

fn is_fixed_sample_positions(&self) -> bool

C++ method: bool QOpenGLTexture::isFixedSamplePositions() const

C++ documentation:

Returns whether this texture uses a fixed pattern of multisample samples. If storage has not yet been allocated for this texture then this function returns the requested fixed sample position setting.

For texture targets that do not support multisampling this will return true.

See also setFixedSamplePositions() and isStorageAllocated().

fn is_storage_allocated(&self) -> bool

C++ method: bool QOpenGLTexture::isStorageAllocated() const

C++ documentation:

Returns true if server-side storage for this texture as been allocated.

The texture format, dimensions, mipmap levels and array layers cannot be altered once storage ihas been allocated.

See also allocateStorage(), setSize(), setMipLevels(), setLayers(), and setFormat().

fn is_texture_view(&self) -> bool

C++ method: bool QOpenGLTexture::isTextureView() const

C++ documentation:

Returns true if this texture object is actually a view onto another texture object.

See also createTextureView().

fn layers(&self) -> c_int

C++ method: int QOpenGLTexture::layers() const

C++ documentation:

Returns the number of array layers for this texture. If storage has not yet been allocated for this texture then this function returns the requested number of array layers.

For texture targets that do not support array layers this will return 1.

See also setLayers() and isStorageAllocated().

fn level_of_detail_range<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLTextureLevelOfDetailRangeArgs<'largs>

C++ method: QOpenGLTexture::levelOfDetailRange

This is an overloaded function. Available variants:

Rust arguments:
1) fn level_of_detail_range(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCFloatCFloat>
2) fn level_of_detail_range(&self, cpp_utils::AsStruct) -> ::pair::PairCFloatCFloat
C++ method: QPair<float, float> QOpenGLTexture::levelOfDetailRange() const

C++ documentation:

Returns the minimum and maximum level of detail parameters.

See also setLevelOfDetailRange(), minimumLevelOfDetail(), and maximumLevelOfDetail().

fn levelof_detail_bias(&self) -> c_float

C++ method: float QOpenGLTexture::levelofDetailBias() const

C++ documentation:

Returns the level of detail bias parameter.

See also setLevelofDetailBias().

fn magnification_filter(&self) -> Filter

C++ method: QOpenGLTexture::Filter QOpenGLTexture::magnificationFilter() const

C++ documentation:

Returns the magnification filter.

See also setMagnificationFilter().

fn maximum_anisotropy(&self) -> c_float

C++ method: float QOpenGLTexture::maximumAnisotropy() const

C++ documentation:

Returns the maximum level of anisotropy to be accounted for when performing texture lookups. This requires the GL_EXT_texture_filter_anisotropic extension.

See also setMaximumAnisotropy().

fn maximum_level_of_detail(&self) -> c_float

C++ method: float QOpenGLTexture::maximumLevelOfDetail() const

C++ documentation:

Returns the maximum level of detail parameter.

See also setMaximumLevelOfDetail(), minimumLevelOfDetail(), and levelOfDetailRange().

fn maximum_mip_levels(&self) -> c_int

C++ method: int QOpenGLTexture::maximumMipLevels() const

C++ documentation:

Returns the maximum number of mipmap levels that this texture can have given the current dimensions.

See also setMipLevels(), mipLevels(), and setSize().

fn min_mag_filters<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLTextureMinMagFiltersArgs<'largs>

C++ method: QOpenGLTexture::minMagFilters

This is an overloaded function. Available variants:

Rust arguments:
1) fn min_mag_filters(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairFilterFilter>
2) fn min_mag_filters(&self, cpp_utils::AsStruct) -> ::pair::PairFilterFilter
C++ method: QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter> QOpenGLTexture::minMagFilters() const

C++ documentation:

Returns the current minification and magnification filters.

See also setMinMagFilters().

fn minification_filter(&self) -> Filter

C++ method: QOpenGLTexture::Filter QOpenGLTexture::minificationFilter() const

C++ documentation:

Returns the minification filter.

See also setMinificationFilter().

fn minimum_level_of_detail(&self) -> c_float

C++ method: float QOpenGLTexture::minimumLevelOfDetail() const

C++ documentation:

Returns the minimum level of detail parameter.

See also setMinimumLevelOfDetail(), maximumLevelOfDetail(), and levelOfDetailRange().

fn mip_base_level(&self) -> c_int

C++ method: int QOpenGLTexture::mipBaseLevel() const

C++ documentation:

Returns the mipmap base level used for all texture lookups with this texture. The default is 0.

See also setMipBaseLevel(), mipMaxLevel(), and mipLevelRange().

fn mip_level_range<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLTextureMipLevelRangeArgs<'largs>

C++ method: QOpenGLTexture::mipLevelRange

This is an overloaded function. Available variants:

Rust arguments:
1) fn mip_level_range(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCIntCInt>
2) fn mip_level_range(&self, cpp_utils::AsStruct) -> ::pair::PairCIntCInt
C++ method: QPair<int, int> QOpenGLTexture::mipLevelRange() const

C++ documentation:

Returns the range of mipmap levels that can be used for texture lookups with this texture.

See also setMipLevelRange(), mipBaseLevel(), and mipMaxLevel().

fn mip_levels(&self) -> c_int

C++ method: int QOpenGLTexture::mipLevels() const

C++ documentation:

Returns the number of mipmap levels for this texture. If storage has not yet been allocated for this texture it returns the requested number of mipmap levels.

See also setMipLevels(), maximumMipLevels(), and isStorageAllocated().

fn mip_max_level(&self) -> c_int

C++ method: int QOpenGLTexture::mipMaxLevel() const

C++ documentation:

Returns the mipmap maximum level used for all texture lookups with this texture.

See also setMipMaxLevel(), mipBaseLevel(), and mipLevelRange().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLTextureNewArgs

C++ method: QOpenGLTexture::QOpenGLTexture

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((::open_g_l_texture::Target, cpp_utils::AsStruct)) -> ::open_g_l_texture::OpenGLTexture
2) fn new((::open_g_l_texture::Target, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_texture::OpenGLTexture>
C++ method: [constructor] void QOpenGLTexture::QOpenGLTexture(QOpenGLTexture::Target target)

Variant 2

Rust arguments:
1) fn new((&::image::Image, cpp_utils::AsStruct)) -> ::open_g_l_texture::OpenGLTexture
2) fn new((&::image::Image, ::open_g_l_texture::MipMapGeneration, cpp_utils::AsStruct)) -> ::open_g_l_texture::OpenGLTexture
3) fn new((&::image::Image, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_texture::OpenGLTexture>
4) fn new((&::image::Image, ::open_g_l_texture::MipMapGeneration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_texture::OpenGLTexture>
C++ method: [constructor] void QOpenGLTexture::QOpenGLTexture(const QImage& image, QOpenGLTexture::MipMapGeneration genMipMaps = ?)

fn release<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureReleaseArgs<'largs>

C++ method: QOpenGLTexture::release

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn release(&mut self, ()) -> ()
C++ method: void QOpenGLTexture::release()

C++ documentation:

Unbinds this texture from the currently active texture unit.

See also bind().

Variant 2

Rust arguments:
1) fn release(&mut self, libc::c_uint) -> ()
2) fn release(&mut self, (libc::c_uint, ::open_g_l_texture::TextureUnitReset)) -> ()
C++ method: void QOpenGLTexture::release(unsigned int unit, QOpenGLTexture::TextureUnitReset reset = ?)

C++ documentation:

Unbinds this texture from texture unit unit.

If parameter reset is true then this function will restore the active unit to the texture unit that was active upon entry.

fn samples(&self) -> c_int

C++ method: int QOpenGLTexture::samples() const

C++ documentation:

Returns the number of multisample sample points for this texture. If storage has not yet been allocated for this texture then this function returns the requested number of samples.

For texture targets that do not support multisampling this will return 0.

See also setSamples() and isStorageAllocated().

fn set_auto_mip_map_generation_enabled(&mut self, enabled: bool)

C++ method: void QOpenGLTexture::setAutoMipMapGenerationEnabled(bool enabled)

C++ documentation:

If enabled is true, enables automatic mipmap generation for this texture object to occur whenever the level 0 mipmap data is set via setData().

The automatic mipmap generation is enabled by default.

See also isAutoMipMapGenerationEnabled() and generateMipMaps().

fn set_border_color<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetBorderColorArgs<'largs>

C++ method: QOpenGLTexture::setBorderColor

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_border_color(&mut self, &::color::Color) -> ()
C++ method: void QOpenGLTexture::setBorderColor(QColor color)

C++ documentation:

Sets the border color of the texture to color.

Note: This function has no effect on Mac and Qt built for OpenGL ES 2.

See also borderColor().

Variant 2

Rust arguments: fn set_border_color(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QOpenGLTexture::setBorderColor(float r, float g, float b, float a)

C++ documentation:

Sets the color red to r, green to g, blue to b, and a to the alpha value.

This is an overloaded function.

Variant 3

Rust arguments: fn set_border_color(&mut self, (libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> ()
C++ method: void QOpenGLTexture::setBorderColor(int r, int g, int b, int a)

C++ documentation:

Sets the color red to r, green to g, blue to b, and the alpha value to a.

This is an overloaded function.

Variant 4

Rust arguments: fn set_border_color(&mut self, (libc::c_uint, libc::c_uint, libc::c_uint, libc::c_uint)) -> ()
C++ method: void QOpenGLTexture::setBorderColor(unsigned int r, unsigned int g, unsigned int b, unsigned int a)

C++ documentation:

Sets the color red to r, green to g, blue to b, and the alpha value to a.

This is an overloaded function.

fn set_comparison_function(&mut self, function: ComparisonFunction)

C++ method: void QOpenGLTexture::setComparisonFunction(QOpenGLTexture::ComparisonFunction function)

C++ documentation:

See also comparisonFunction().

fn set_comparison_mode(&mut self, mode: ComparisonMode)

C++ method: void QOpenGLTexture::setComparisonMode(QOpenGLTexture::ComparisonMode mode)

C++ documentation:

Sets the texture comparison mode on this texture to mode. The texture comparison mode is used by shadow samplers when sampling a depth texture.

This function was introduced in Qt 5.5.

See also comparisonMode().

fn set_compressed_data<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetCompressedDataArgs<'largs>

C++ method: QOpenGLTexture::setCompressedData

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, *const libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int dataSize, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

Variant 2

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, *mut libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int dataSize, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 3

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, *const libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int dataSize, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

Variant 4

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, *mut libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int dataSize, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 5

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, libc::c_int, *const libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, libc::c_int, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int layer, QOpenGLTexture::CubeMapFace cubeFace, int dataSize, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

Uploads compressed pixel data to mipLevel, array layer, and cubeFace. The pixel transfer can optionally be controlled with options. The dataSize argument should specify the size of the data pointed to by data.

If not using a compressed format() then you should use setData() instead of this function.

This function was introduced in Qt 5.3.

Variant 6

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, libc::c_int, *mut libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, libc::c_int, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int layer, QOpenGLTexture::CubeMapFace cubeFace, int dataSize, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 7

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, libc::c_int, *const libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, libc::c_int, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int layer, int dataSize, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

Variant 8

Rust arguments:
1) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, libc::c_int, *mut libc::c_void)) -> ()
2) fn set_compressed_data(&mut self, (libc::c_int, libc::c_int, libc::c_int, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setCompressedData(int mipLevel, int layer, int dataSize, void* data, const QOpenGLPixelTransferOptions* options = ?)

fn set_data<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetDataArgs<'largs>

C++ method: QOpenGLTexture::setData

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_data(&mut self, (::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void)) -> ()
2) fn set_data(&mut self, (::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.3.

Variant 2

Rust arguments:
1) fn set_data(&mut self, (::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void)) -> ()
2) fn set_data(&mut self, (::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 3

Rust arguments:
1) fn set_data(&mut self, &::image::Image) -> ()
2) fn set_data(&mut self, (&::image::Image, ::open_g_l_texture::MipMapGeneration)) -> ()
C++ method: void QOpenGLTexture::setData(const QImage& image, QOpenGLTexture::MipMapGeneration genMipMaps = ?)

C++ documentation:

This overload of setData() will allocate storage for you. The pixel data is contained in image. Mipmaps are generated by default. Set genMipMaps to DontGenerateMipMaps to turn off mipmap generation.

This is an overloaded function.

Variant 4

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.3.

Variant 5

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 6

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, int layer, QOpenGLTexture::CubeMapFace cubeFace, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

Uploads pixel data for this texture object mipLevel, array layer, and cubeFace. Storage must have been allocated before uploading pixel data. Some overloads of setData() will set appropriate dimensions, mipmap levels, and array layers and then allocate storage for you if they have enough information to do so. This will be noted in the function documentation.

The structure of the pixel data pointed to by data is specified by sourceFormat and sourceType. The pixel data upload can optionally be controlled by options.

If using a compressed format() then you should use setCompressedData() instead of this function.

This function was introduced in Qt 5.3.

See also setCompressedData().

Variant 7

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::CubeMapFace, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, int layer, QOpenGLTexture::CubeMapFace cubeFace, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, void* data, const QOpenGLPixelTransferOptions* options = ?)

Variant 8

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *const libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, int layer, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, const void* data, const QOpenGLPixelTransferOptions* options = ?)

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.3.

Variant 9

Rust arguments:
1) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void)) -> ()
2) fn set_data(&mut self, (libc::c_int, libc::c_int, ::open_g_l_texture::PixelFormat, ::open_g_l_texture::PixelType, *mut libc::c_void, *const ::open_g_l_pixel_transfer_options::OpenGLPixelTransferOptions)) -> ()
C++ method: void QOpenGLTexture::setData(int mipLevel, int layer, QOpenGLTexture::PixelFormat sourceFormat, QOpenGLTexture::PixelType sourceType, void* data, const QOpenGLPixelTransferOptions* options = ?)

fn set_depth_stencil_mode(&mut self, mode: DepthStencilMode)

C++ method: void QOpenGLTexture::setDepthStencilMode(QOpenGLTexture::DepthStencilMode mode)

C++ documentation:

If using a texture that has a combined depth/stencil format this function sets which component of the texture is accessed to mode.

When the parameter is set to DepthMode, then accessing it from the shader will access the depth component as a single float, as normal. But when the parameter is set to StencilMode, the shader will access the stencil component.

Note: This function has no effect on Mac and Qt built for OpenGL ES 2.

This function was introduced in Qt 5.4.

See also depthStencilMode().

fn set_fixed_sample_positions(&mut self, fixed: bool)

C++ method: void QOpenGLTexture::setFixedSamplePositions(bool fixed)

C++ documentation:

Sets whether the sample positions and number of samples used with a multisample capable texture target to fixed. If set to true the sample positions and number of samples used are the same for all texels in the image and will not depend upon the image size or internal format. This function should be called before storage is allocated for the texture.

For targets that do not support multisampling this function has no effect.

The default value is true.

See also isFixedSamplePositions() and isStorageAllocated().

fn set_format(&mut self, format: TextureFormat)

C++ method: void QOpenGLTexture::setFormat(QOpenGLTexture::TextureFormat format)

C++ documentation:

Sets the format of this texture object to format. This function must be called before texture storage is allocated.

Note that all formats may not be supported. The exact set of supported formats is dependent upon your OpenGL implementation and version.

See also format() and allocateStorage().

fn set_layers(&mut self, layers: c_int)

C++ method: void QOpenGLTexture::setLayers(int layers)

C++ documentation:

Sets the number of array layers to allocate storage for. This function should be called before storage is allocated for the texture.

For targets that do not support array layers this function has no effect.

See also layers() and isStorageAllocated().

fn set_level_of_detail_range(&mut self, min: c_float, max: c_float)

C++ method: void QOpenGLTexture::setLevelOfDetailRange(float min, float max)

C++ documentation:

Sets the minimum level of detail parameters to min and the maximum level to max.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also levelOfDetailRange(), setMinimumLevelOfDetail(), and setMaximumLevelOfDetail().

fn set_levelof_detail_bias(&mut self, bias: c_float)

C++ method: void QOpenGLTexture::setLevelofDetailBias(float bias)

C++ documentation:

Sets the level of detail bias to bias. Level of detail bias affects the point at which mipmapping levels change. Increasing values for level of detail bias makes the overall images blurrier or smoother. Decreasing values make the overall images sharper.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also levelofDetailBias().

fn set_magnification_filter(&mut self, filter: Filter)

C++ method: void QOpenGLTexture::setMagnificationFilter(QOpenGLTexture::Filter filter)

C++ documentation:

Sets the magnification filter to filter.

See also magnificationFilter(), setMinificationFilter(), and setMinMagFilters().

fn set_maximum_anisotropy(&mut self, anisotropy: c_float)

C++ method: void QOpenGLTexture::setMaximumAnisotropy(float anisotropy)

C++ documentation:

If your OpenGL implementation supports the GL_EXT_texture_filter_anisotropic extension this function sets the maximum anisotropy level to anisotropy.

See also maximumAnisotropy().

fn set_maximum_level_of_detail(&mut self, value: c_float)

C++ method: void QOpenGLTexture::setMaximumLevelOfDetail(float value)

C++ documentation:

Sets the maximum level of detail to value. This limits the selection of lowest resolution mipmap (highest mipmap level). The default value is 1000.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also maximumLevelOfDetail(), setMinimumLevelOfDetail(), and setLevelOfDetailRange().

fn set_min_mag_filters(&mut self, minification_filter: Filter, magnification_filter: Filter)

C++ method: void QOpenGLTexture::setMinMagFilters(QOpenGLTexture::Filter minificationFilter, QOpenGLTexture::Filter magnificationFilter)

C++ documentation:

Sets the minification filter to minificationFilter and the magnification filter to magnificationFilter.

See also minMagFilters(), setMinificationFilter(), and setMagnificationFilter().

fn set_minification_filter(&mut self, filter: Filter)

C++ method: void QOpenGLTexture::setMinificationFilter(QOpenGLTexture::Filter filter)

C++ documentation:

Sets the filter used for minification to filter.

See also minificationFilter(), setMagnificationFilter(), and setMinMagFilters().

fn set_minimum_level_of_detail(&mut self, value: c_float)

C++ method: void QOpenGLTexture::setMinimumLevelOfDetail(float value)

C++ documentation:

Sets the minimum level of detail to value. This limits the selection of highest resolution mipmap (lowest mipmap level). The default value is -1000.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also minimumLevelOfDetail(), setMaximumLevelOfDetail(), and setLevelOfDetailRange().

fn set_mip_base_level(&mut self, base_level: c_int)

C++ method: void QOpenGLTexture::setMipBaseLevel(int baseLevel)

C++ documentation:

Sets the base mipmap level used for all texture lookups with this texture to baseLevel.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also mipBaseLevel(), setMipMaxLevel(), and setMipLevelRange().

fn set_mip_level_range(&mut self, base_level: c_int, max_level: c_int)

C++ method: void QOpenGLTexture::setMipLevelRange(int baseLevel, int maxLevel)

C++ documentation:

Sets the range of mipmap levels that can be used for texture lookups with this texture to range from baseLevel to maxLevel.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also setMipBaseLevel(), setMipMaxLevel(), and mipLevelRange().

fn set_mip_levels(&mut self, levels: c_int)

C++ method: void QOpenGLTexture::setMipLevels(int levels)

C++ documentation:

For texture targets that support mipmaps, this function sets the requested number of mipmap levels to allocate storage for. This function should be called before storage is allocated for the texture.

If the texture target does not support mipmaps this function has no effect.

See also mipLevels(), maximumMipLevels(), and isStorageAllocated().

fn set_mip_max_level(&mut self, max_level: c_int)

C++ method: void QOpenGLTexture::setMipMaxLevel(int maxLevel)

C++ documentation:

Sets the maximum mipmap level used for all texture lookups with this texture to maxLevel.

Note: This function has no effect on Qt built for OpenGL ES 2.

See also mipMaxLevel(), setMipBaseLevel(), and setMipLevelRange().

fn set_samples(&mut self, samples: c_int)

C++ method: void QOpenGLTexture::setSamples(int samples)

C++ documentation:

Sets the number of samples to allocate storage for when rendering to a multisample capable texture target. This function should be called before storage is allocated for the texture.

For targets that do not support multisampling this function has no effect.

See also samples() and isStorageAllocated().

fn set_size<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetSizeArgs<'largs>

C++ method: QOpenGLTexture::setSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_size(&mut self, libc::c_int) -> ()
2) fn set_size(&mut self, (libc::c_int, libc::c_int)) -> ()
3) fn set_size(&mut self, (libc::c_int, libc::c_int, libc::c_int)) -> ()
C++ method: void QOpenGLTexture::setSize(int width, int height = ?, int depth = ?)

C++ documentation:

Sets the dimensions of this texture object to width, height, and depth. The default for each dimension is 1. The maximum allowable texture size is dependent upon your OpenGL implementation. Allocating storage for a texture less than the maximum size can still fail if your system is low on resources.

See also width(), height(), and depth().

fn set_swizzle_mask<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetSwizzleMaskArgs<'largs>

C++ method: QOpenGLTexture::setSwizzleMask

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_swizzle_mask(&mut self, (::open_g_l_texture::SwizzleComponent, ::open_g_l_texture::SwizzleValue)) -> ()
C++ method: void QOpenGLTexture::setSwizzleMask(QOpenGLTexture::SwizzleComponent component, QOpenGLTexture::SwizzleValue value)

C++ documentation:

GLSL shaders are able to reorder the components of the vec4 returned by texture functions. It is also desirable to be able to control this reordering from CPU side code. This is made possible by swizzle masks since OpenGL 3.3.

Each component of the texture can be mapped to one of the SwizzleValue options.

This function maps component to the output value.

Note: This function has no effect on Mac and Qt built for OpenGL ES 2.

See also swizzleMask().

Variant 2

Rust arguments: fn set_swizzle_mask(&mut self, (::open_g_l_texture::SwizzleValue, ::open_g_l_texture::SwizzleValue, ::open_g_l_texture::SwizzleValue, ::open_g_l_texture::SwizzleValue)) -> ()
C++ method: void QOpenGLTexture::setSwizzleMask(QOpenGLTexture::SwizzleValue r, QOpenGLTexture::SwizzleValue g, QOpenGLTexture::SwizzleValue b, QOpenGLTexture::SwizzleValue a)

C++ documentation:

Parameters r, g, b, and a are values used for setting the colors red, green, blue, and the alpha value.

This is an overloaded function.

fn set_wrap_mode<'largs, Args>(&'largs mut self, args: Args) where Args: OpenGLTextureSetWrapModeArgs<'largs>

C++ method: QOpenGLTexture::setWrapMode

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_wrap_mode(&mut self, (::open_g_l_texture::CoordinateDirection, ::open_g_l_texture::WrapMode)) -> ()
C++ method: void QOpenGLTexture::setWrapMode(QOpenGLTexture::CoordinateDirection direction, QOpenGLTexture::WrapMode mode)

C++ documentation:

Holds the texture dimension direction.

This is an overloaded function.

Variant 2

Rust arguments: fn set_wrap_mode(&mut self, ::open_g_l_texture::WrapMode) -> ()
C++ method: void QOpenGLTexture::setWrapMode(QOpenGLTexture::WrapMode mode)

C++ documentation:

Sets the wrap (or repeat mode) for all texture dimentions to mode.

See also wrapMode().

fn swizzle_mask(&self, component: SwizzleComponent) -> SwizzleValue

C++ method: QOpenGLTexture::SwizzleValue QOpenGLTexture::swizzleMask(QOpenGLTexture::SwizzleComponent component) const

C++ documentation:

Returns the swizzle mask for texture component.

See also setSwizzleMask().

fn target(&self) -> Target

C++ method: QOpenGLTexture::Target QOpenGLTexture::target() const

C++ documentation:

Returns the binding target of this texture.

This function was introduced in Qt 5.4.

fn texture_id(&self) -> c_uint

C++ method: unsigned int QOpenGLTexture::textureId() const

C++ documentation:

Returns the name of the underlying OpenGL texture object or 0 if it has not yet been created.

See also create(), destroy(), and isCreated().

fn width(&self) -> c_int

C++ method: int QOpenGLTexture::width() const

C++ documentation:

Returns the width of a 1D, 2D or 3D texture.

See also height(), depth(), and setSize().

fn wrap_mode(&self, direction: CoordinateDirection) -> WrapMode

C++ method: QOpenGLTexture::WrapMode QOpenGLTexture::wrapMode(QOpenGLTexture::CoordinateDirection direction) const

C++ documentation:

Returns the wrap mode for the texture dimension direction.

See also setWrapMode().