Module qt_gui::open_g_l_time_monitor::overloading

Traits

OpenGLTimeMonitorConnectArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::connect method.

OpenGLTimeMonitorConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::connect_static method.

OpenGLTimeMonitorDestroyedArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::destroyed method.

OpenGLTimeMonitorDisconnectArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::disconnect method.

OpenGLTimeMonitorDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::disconnect_static method.

OpenGLTimeMonitorDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::dynamic_property_names method.

OpenGLTimeMonitorNewArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::new method.

OpenGLTimeMonitorObjectIdsArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::object_ids method.

OpenGLTimeMonitorObjectNameArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::object_name method.

OpenGLTimeMonitorPropertyArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::property method.

OpenGLTimeMonitorStartTimerArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::start_timer method.

OpenGLTimeMonitorTrArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::tr method.

OpenGLTimeMonitorTrUtf8Args
This trait represents a set of arguments accepted by OpenGLTimeMonitor::tr_utf8 method.

OpenGLTimeMonitorWaitForIntervalsArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::wait_for_intervals method.

OpenGLTimeMonitorWaitForSamplesArgs
This trait represents a set of arguments accepted by OpenGLTimeMonitor::wait_for_samples method.