Module qt_gui::open_g_l_timer_query::overloading

Traits

OpenGLTimerQueryConnectArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::connect method.

OpenGLTimerQueryConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::connect_static method.

OpenGLTimerQueryDestroyedArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::destroyed method.

OpenGLTimerQueryDisconnectArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::disconnect method.

OpenGLTimerQueryDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::disconnect_static method.

OpenGLTimerQueryDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::dynamic_property_names method.

OpenGLTimerQueryNewArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::new method.

OpenGLTimerQueryObjectNameArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::object_name method.

OpenGLTimerQueryPropertyArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::property method.

OpenGLTimerQueryStartTimerArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::start_timer method.

OpenGLTimerQueryTrArgs
This trait represents a set of arguments accepted by OpenGLTimerQuery::tr method.

OpenGLTimerQueryTrUtf8Args
This trait represents a set of arguments accepted by OpenGLTimerQuery::tr_utf8 method.