Struct qt_gui::open_g_l_version_functions::OpenGLFunctions10CoreBackend

pub struct OpenGLFunctions10CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_1_0_CoreBackend

Methods

impl OpenGLFunctions10CoreBackend

fn blend_func(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_1_0_CoreBackend::BlendFunc() const

fn clear(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::Clear() const

fn clear_color(&self) -> extern fn(c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(float, float, float, float) QOpenGLFunctions_1_0_CoreBackend::ClearColor() const

fn clear_depth(&self) -> extern fn(c_double)

C++ method: void (*FN_PTR)(double) QOpenGLFunctions_1_0_CoreBackend::ClearDepth() const

fn clear_stencil(&self) -> extern fn(c_int)

C++ method: void (*FN_PTR)(int) QOpenGLFunctions_1_0_CoreBackend::ClearStencil() const

fn color_mask(&self) -> extern fn(c_uchar, c_uchar, c_uchar, c_uchar)

C++ method: void (*FN_PTR)(unsigned char, unsigned char, unsigned char, unsigned char) QOpenGLFunctions_1_0_CoreBackend::ColorMask() const

fn cull_face(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::CullFace() const

fn depth_func(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::DepthFunc() const

fn depth_mask(&self) -> extern fn(c_uchar)

C++ method: void (*FN_PTR)(unsigned char) QOpenGLFunctions_1_0_CoreBackend::DepthMask() const

fn depth_range(&self) -> extern fn(c_double, c_double)

C++ method: void (*FN_PTR)(double, double) QOpenGLFunctions_1_0_CoreBackend::DepthRange() const

fn disable(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::Disable() const

fn draw_buffer(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::DrawBuffer() const

fn enable(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::Enable() const

fn finish(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_1_0_CoreBackend::Finish() const

fn flush(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_1_0_CoreBackend::Flush() const

fn front_face(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::FrontFace() const

fn get_booleanv(&self) -> extern fn(c_uint, *mut c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned char*) QOpenGLFunctions_1_0_CoreBackend::GetBooleanv() const

fn get_doublev(&self) -> extern fn(c_uint, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, double*) QOpenGLFunctions_1_0_CoreBackend::GetDoublev() const

fn get_error(&self) -> extern fn() -> c_uint

C++ method: unsigned int (*FN_PTR)() QOpenGLFunctions_1_0_CoreBackend::GetError() const

fn get_floatv(&self) -> extern fn(c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, float*) QOpenGLFunctions_1_0_CoreBackend::GetFloatv() const

fn get_integerv(&self) -> extern fn(c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, int*) QOpenGLFunctions_1_0_CoreBackend::GetIntegerv() const

fn get_string(&self) -> extern fn(c_uint) -> *const c_uchar

C++ method: const unsigned char* (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::GetString() const

fn get_tex_image(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, void*) QOpenGLFunctions_1_0_CoreBackend::GetTexImage() const

fn get_tex_level_parameterfv(&self) -> extern fn(c_uint, c_int, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, float*) QOpenGLFunctions_1_0_CoreBackend::GetTexLevelParameterfv() const

fn get_tex_level_parameteriv(&self) -> extern fn(c_uint, c_int, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int*) QOpenGLFunctions_1_0_CoreBackend::GetTexLevelParameteriv() const

fn get_tex_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_1_0_CoreBackend::GetTexParameterfv() const

fn get_tex_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_1_0_CoreBackend::GetTexParameteriv() const

fn hint(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_1_0_CoreBackend::Hint() const

fn is_enabled(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::IsEnabled() const

fn line_width(&self) -> extern fn(c_float)

C++ method: void (*FN_PTR)(float) QOpenGLFunctions_1_0_CoreBackend::LineWidth() const

fn logic_op(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::LogicOp() const

fn pixel_storef(&self) -> extern fn(c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, float) QOpenGLFunctions_1_0_CoreBackend::PixelStoref() const

fn pixel_storei(&self) -> extern fn(c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int) QOpenGLFunctions_1_0_CoreBackend::PixelStorei() const

fn point_size(&self) -> extern fn(c_float)

C++ method: void (*FN_PTR)(float) QOpenGLFunctions_1_0_CoreBackend::PointSize() const

fn polygon_mode(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_1_0_CoreBackend::PolygonMode() const

fn read_buffer(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::ReadBuffer() const

fn read_pixels(&self) -> extern fn(c_int, c_int, c_int, c_int, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(int, int, int, int, unsigned int, unsigned int, void*) QOpenGLFunctions_1_0_CoreBackend::ReadPixels() const

fn scissor(&self) -> extern fn(c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(int, int, int, int) QOpenGLFunctions_1_0_CoreBackend::Scissor() const

fn set_blend_func(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_BlendFunc(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_clear(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Clear(void (*FN_PTR)(unsigned int) value)

fn set_clear_color(&mut self, value: extern fn(c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ClearColor(void (*FN_PTR)(float, float, float, float) value)

fn set_clear_depth(&mut self, value: extern fn(c_double))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ClearDepth(void (*FN_PTR)(double) value)

fn set_clear_stencil(&mut self, value: extern fn(c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ClearStencil(void (*FN_PTR)(int) value)

fn set_color_mask(&mut self, value: extern fn(c_uchar, c_uchar, c_uchar, c_uchar))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ColorMask(void (*FN_PTR)(unsigned char, unsigned char, unsigned char, unsigned char) value)

fn set_cull_face(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_CullFace(void (*FN_PTR)(unsigned int) value)

fn set_depth_func(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_DepthFunc(void (*FN_PTR)(unsigned int) value)

fn set_depth_mask(&mut self, value: extern fn(c_uchar))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_DepthMask(void (*FN_PTR)(unsigned char) value)

fn set_depth_range(&mut self, value: extern fn(c_double, c_double))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_DepthRange(void (*FN_PTR)(double, double) value)

fn set_disable(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Disable(void (*FN_PTR)(unsigned int) value)

fn set_draw_buffer(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_DrawBuffer(void (*FN_PTR)(unsigned int) value)

fn set_enable(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Enable(void (*FN_PTR)(unsigned int) value)

fn set_finish(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Finish(void (*FN_PTR)() value)

fn set_flush(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Flush(void (*FN_PTR)() value)

fn set_front_face(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_FrontFace(void (*FN_PTR)(unsigned int) value)

fn set_get_booleanv(&mut self, value: extern fn(c_uint, *mut c_uchar))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetBooleanv(void (*FN_PTR)(unsigned int, unsigned char*) value)

fn set_get_doublev(&mut self, value: extern fn(c_uint, *mut c_double))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetDoublev(void (*FN_PTR)(unsigned int, double*) value)

fn set_get_error(&mut self, value: extern fn() -> c_uint)

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetError(unsigned int (*FN_PTR)() value)

fn set_get_floatv(&mut self, value: extern fn(c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetFloatv(void (*FN_PTR)(unsigned int, float*) value)

fn set_get_integerv(&mut self, value: extern fn(c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetIntegerv(void (*FN_PTR)(unsigned int, int*) value)

fn set_get_string(&mut self, value: extern fn(c_uint) -> *const c_uchar)

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetString(const unsigned char* (*FN_PTR)(unsigned int) value)

fn set_get_tex_image(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetTexImage(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, void*) value)

fn set_get_tex_level_parameterfv(&mut self, value: extern fn(c_uint, c_int, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetTexLevelParameterfv(void (*FN_PTR)(unsigned int, int, unsigned int, float*) value)

fn set_get_tex_level_parameteriv(&mut self, value: extern fn(c_uint, c_int, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetTexLevelParameteriv(void (*FN_PTR)(unsigned int, int, unsigned int, int*) value)

fn set_get_tex_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetTexParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_tex_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_GetTexParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_hint(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Hint(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_is_enabled(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_IsEnabled(unsigned char (*FN_PTR)(unsigned int) value)

fn set_line_width(&mut self, value: extern fn(c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_LineWidth(void (*FN_PTR)(float) value)

fn set_logic_op(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_LogicOp(void (*FN_PTR)(unsigned int) value)

fn set_pixel_storef(&mut self, value: extern fn(c_uint, c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_PixelStoref(void (*FN_PTR)(unsigned int, float) value)

fn set_pixel_storei(&mut self, value: extern fn(c_uint, c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_PixelStorei(void (*FN_PTR)(unsigned int, int) value)

fn set_point_size(&mut self, value: extern fn(c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_PointSize(void (*FN_PTR)(float) value)

fn set_polygon_mode(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_PolygonMode(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_read_buffer(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ReadBuffer(void (*FN_PTR)(unsigned int) value)

fn set_read_pixels(&mut self, value: extern fn(c_int, c_int, c_int, c_int, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_ReadPixels(void (*FN_PTR)(int, int, int, int, unsigned int, unsigned int, void*) value)

fn set_scissor(&mut self, value: extern fn(c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Scissor(void (*FN_PTR)(int, int, int, int) value)

fn set_stencil_func(&mut self, value: extern fn(c_uint, c_int, c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_StencilFunc(void (*FN_PTR)(unsigned int, int, unsigned int) value)

fn set_stencil_mask(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_StencilMask(void (*FN_PTR)(unsigned int) value)

fn set_stencil_op(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_StencilOp(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_tex_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexImage1D(void (*FN_PTR)(unsigned int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_tex_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_tex_parameterf(&mut self, value: extern fn(c_uint, c_uint, c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexParameterf(void (*FN_PTR)(unsigned int, unsigned int, float) value)

fn set_tex_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *const c_float))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexParameterfv(void (*FN_PTR)(unsigned int, unsigned int, const float*) value)

fn set_tex_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_tex_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_TexParameteriv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_viewport(&mut self, value: extern fn(c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_0_CoreBackend::set_Viewport(void (*FN_PTR)(int, int, int, int) value)

fn stencil_func(&self) -> extern fn(c_uint, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int) QOpenGLFunctions_1_0_CoreBackend::StencilFunc() const

fn stencil_mask(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_0_CoreBackend::StencilMask() const

fn stencil_op(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_1_0_CoreBackend::StencilOp() const

fn tex_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_0_CoreBackend::TexImage1D() const

fn tex_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_0_CoreBackend::TexImage2D() const

fn tex_parameterf(&self) -> extern fn(c_uint, c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float) QOpenGLFunctions_1_0_CoreBackend::TexParameterf() const

fn tex_parameterfv(&self) -> extern fn(c_uint, c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const float*) QOpenGLFunctions_1_0_CoreBackend::TexParameterfv() const

fn tex_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_1_0_CoreBackend::TexParameteri() const

fn tex_parameteriv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_1_0_CoreBackend::TexParameteriv() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions10CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_1_0_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_1_0_CoreBackend::versionStatus()

fn viewport(&self) -> extern fn(c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(int, int, int, int) QOpenGLFunctions_1_0_CoreBackend::Viewport() const