Struct qt_gui::open_g_l_version_functions::OpenGLFunctions11CoreBackend

pub struct OpenGLFunctions11CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_1_1_CoreBackend

Methods

impl OpenGLFunctions11CoreBackend

fn bind_texture(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_1_1_CoreBackend::BindTexture() const

fn copy_tex_image1_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int) QOpenGLFunctions_1_1_CoreBackend::CopyTexImage1D() const

fn copy_tex_image2_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, int) QOpenGLFunctions_1_1_CoreBackend::CopyTexImage2D() const

fn copy_tex_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int) QOpenGLFunctions_1_1_CoreBackend::CopyTexSubImage1D() const

fn copy_tex_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) QOpenGLFunctions_1_1_CoreBackend::CopyTexSubImage2D() const

fn delete_textures(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_1_1_CoreBackend::DeleteTextures() const

fn draw_arrays(&self) -> extern fn(c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int) QOpenGLFunctions_1_1_CoreBackend::DrawArrays() const

fn draw_elements(&self) -> extern fn(c_uint, c_int, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, const void*) QOpenGLFunctions_1_1_CoreBackend::DrawElements() const

fn gen_textures(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_1_1_CoreBackend::GenTextures() const

fn get_pointerv(&self) -> extern fn(c_uint, *mut *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, void**) QOpenGLFunctions_1_1_CoreBackend::GetPointerv() const

fn indexub(&self) -> extern fn(c_uchar)

C++ method: void (*FN_PTR)(unsigned char) QOpenGLFunctions_1_1_CoreBackend::Indexub() const

fn indexubv(&self) -> extern fn(*const c_uchar)

C++ method: void (*FN_PTR)(const unsigned char*) QOpenGLFunctions_1_1_CoreBackend::Indexubv() const

fn is_texture(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_1_1_CoreBackend::IsTexture() const

fn polygon_offset(&self) -> extern fn(c_float, c_float)

C++ method: void (*FN_PTR)(float, float) QOpenGLFunctions_1_1_CoreBackend::PolygonOffset() const

fn set_bind_texture(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_BindTexture(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_copy_tex_image1_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_CopyTexImage1D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int) value)

fn set_copy_tex_image2_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_CopyTexImage2D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, int) value)

fn set_copy_tex_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_CopyTexSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, int, int) value)

fn set_copy_tex_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_CopyTexSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) value)

fn set_delete_textures(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_DeleteTextures(void (*FN_PTR)(int, const unsigned int*) value)

fn set_draw_arrays(&mut self, value: extern fn(c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_DrawArrays(void (*FN_PTR)(unsigned int, int, int) value)

fn set_draw_elements(&mut self, value: extern fn(c_uint, c_int, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_DrawElements(void (*FN_PTR)(unsigned int, int, unsigned int, const void*) value)

fn set_gen_textures(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_GenTextures(void (*FN_PTR)(int, unsigned int*) value)

fn set_get_pointerv(&mut self, value: extern fn(c_uint, *mut *mut c_void))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_GetPointerv(void (*FN_PTR)(unsigned int, void**) value)

fn set_indexub(&mut self, value: extern fn(c_uchar))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_Indexub(void (*FN_PTR)(unsigned char) value)

fn set_indexubv(&mut self, value: extern fn(*const c_uchar))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_Indexubv(void (*FN_PTR)(const unsigned char*) value)

fn set_is_texture(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_IsTexture(unsigned char (*FN_PTR)(unsigned int) value)

fn set_polygon_offset(&mut self, value: extern fn(c_float, c_float))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_PolygonOffset(void (*FN_PTR)(float, float) value)

fn set_tex_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_TexSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_tex_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_1_CoreBackend::set_TexSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn tex_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_1_CoreBackend::TexSubImage1D() const

fn tex_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_1_CoreBackend::TexSubImage2D() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions11CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_1_1_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_1_1_CoreBackend::versionStatus()