Struct qt_gui::open_g_l_version_functions::OpenGLFunctions12CoreBackend

pub struct OpenGLFunctions12CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_1_2_CoreBackend

Methods

impl OpenGLFunctions12CoreBackend

fn blend_color(&self) -> extern fn(c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(float, float, float, float) QOpenGLFunctions_1_2_CoreBackend::BlendColor() const

fn blend_equation(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_2_CoreBackend::BlendEquation() const

fn copy_tex_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int) QOpenGLFunctions_1_2_CoreBackend::CopyTexSubImage3D() const

fn draw_range_elements(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, unsigned int, const void*) QOpenGLFunctions_1_2_CoreBackend::DrawRangeElements() const

fn set_blend_color(&mut self, value: extern fn(c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_BlendColor(void (*FN_PTR)(float, float, float, float) value)

fn set_blend_equation(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_BlendEquation(void (*FN_PTR)(unsigned int) value)

fn set_copy_tex_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_CopyTexSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int) value)

fn set_draw_range_elements(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_DrawRangeElements(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, unsigned int, const void*) value)

fn set_tex_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_TexImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_tex_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_CoreBackend::set_TexSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn tex_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_CoreBackend::TexImage3D() const

fn tex_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_CoreBackend::TexSubImage3D() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions12CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_1_2_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_1_2_CoreBackend::versionStatus()