Struct qt_gui::open_g_l_version_functions::OpenGLFunctions12DeprecatedBackend

pub struct OpenGLFunctions12DeprecatedBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_1_2_DeprecatedBackend

Methods

impl OpenGLFunctions12DeprecatedBackend

fn color_sub_table(&self) -> extern fn(c_uint, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_DeprecatedBackend::ColorSubTable() const

fn color_table(&self) -> extern fn(c_uint, c_uint, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_DeprecatedBackend::ColorTable() const

fn color_table_parameterfv(&self) -> extern fn(c_uint, c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const float*) QOpenGLFunctions_1_2_DeprecatedBackend::ColorTableParameterfv() const

fn color_table_parameteriv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_1_2_DeprecatedBackend::ColorTableParameteriv() const

fn convolution_filter1_d(&self) -> extern fn(c_uint, c_uint, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionFilter1D() const

fn convolution_filter2_d(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionFilter2D() const

fn convolution_parameterf(&self) -> extern fn(c_uint, c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionParameterf() const

fn convolution_parameterfv(&self) -> extern fn(c_uint, c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const float*) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionParameterfv() const

fn convolution_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionParameteri() const

fn convolution_parameteriv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_1_2_DeprecatedBackend::ConvolutionParameteriv() const

fn copy_color_sub_table(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int) QOpenGLFunctions_1_2_DeprecatedBackend::CopyColorSubTable() const

fn copy_color_table(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, int) QOpenGLFunctions_1_2_DeprecatedBackend::CopyColorTable() const

fn copy_convolution_filter1_d(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, int) QOpenGLFunctions_1_2_DeprecatedBackend::CopyConvolutionFilter1D() const

fn copy_convolution_filter2_d(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int) QOpenGLFunctions_1_2_DeprecatedBackend::CopyConvolutionFilter2D() const

fn get_color_table(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*) QOpenGLFunctions_1_2_DeprecatedBackend::GetColorTable() const

fn get_color_table_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_1_2_DeprecatedBackend::GetColorTableParameterfv() const

fn get_color_table_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_1_2_DeprecatedBackend::GetColorTableParameteriv() const

fn get_convolution_filter(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*) QOpenGLFunctions_1_2_DeprecatedBackend::GetConvolutionFilter() const

fn get_convolution_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_1_2_DeprecatedBackend::GetConvolutionParameterfv() const

fn get_convolution_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_1_2_DeprecatedBackend::GetConvolutionParameteriv() const

fn get_histogram(&self) -> extern fn(c_uint, c_uchar, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, void*) QOpenGLFunctions_1_2_DeprecatedBackend::GetHistogram() const

fn get_histogram_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_1_2_DeprecatedBackend::GetHistogramParameterfv() const

fn get_histogram_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_1_2_DeprecatedBackend::GetHistogramParameteriv() const

fn get_minmax(&self) -> extern fn(c_uint, c_uchar, c_uint, c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, void*) QOpenGLFunctions_1_2_DeprecatedBackend::GetMinmax() const

fn get_minmax_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_1_2_DeprecatedBackend::GetMinmaxParameterfv() const

fn get_minmax_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_1_2_DeprecatedBackend::GetMinmaxParameteriv() const

fn get_separable_filter(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_void, *mut c_void, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*, void*, void*) QOpenGLFunctions_1_2_DeprecatedBackend::GetSeparableFilter() const

fn histogram(&self) -> extern fn(c_uint, c_int, c_uint, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char) QOpenGLFunctions_1_2_DeprecatedBackend::Histogram() const

fn minmax(&self) -> extern fn(c_uint, c_uint, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char) QOpenGLFunctions_1_2_DeprecatedBackend::Minmax() const

fn reset_histogram(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_2_DeprecatedBackend::ResetHistogram() const

fn reset_minmax(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_2_DeprecatedBackend::ResetMinmax() const

fn separable_filter2_d(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_uint, c_uint, *const c_void, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, unsigned int, unsigned int, const void*, const void*) QOpenGLFunctions_1_2_DeprecatedBackend::SeparableFilter2D() const

fn set_color_sub_table(&mut self, value: extern fn(c_uint, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ColorSubTable(void (*FN_PTR)(unsigned int, int, int, unsigned int, unsigned int, const void*) value)

fn set_color_table(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ColorTable(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int, const void*) value)

fn set_color_table_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *const c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ColorTableParameterfv(void (*FN_PTR)(unsigned int, unsigned int, const float*) value)

fn set_color_table_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ColorTableParameteriv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_convolution_filter1_d(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionFilter1D(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int, const void*) value)

fn set_convolution_filter2_d(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionFilter2D(void (*FN_PTR)(unsigned int, unsigned int, int, int, unsigned int, unsigned int, const void*) value)

fn set_convolution_parameterf(&mut self, value: extern fn(c_uint, c_uint, c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionParameterf(void (*FN_PTR)(unsigned int, unsigned int, float) value)

fn set_convolution_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *const c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionParameterfv(void (*FN_PTR)(unsigned int, unsigned int, const float*) value)

fn set_convolution_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_convolution_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ConvolutionParameteriv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_copy_color_sub_table(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_CopyColorSubTable(void (*FN_PTR)(unsigned int, int, int, int, int) value)

fn set_copy_color_table(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_CopyColorTable(void (*FN_PTR)(unsigned int, unsigned int, int, int, int) value)

fn set_copy_convolution_filter1_d(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_CopyConvolutionFilter1D(void (*FN_PTR)(unsigned int, unsigned int, int, int, int) value)

fn set_copy_convolution_filter2_d(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_CopyConvolutionFilter2D(void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int) value)

fn set_get_color_table(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetColorTable(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*) value)

fn set_get_color_table_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetColorTableParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_color_table_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetColorTableParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_convolution_filter(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetConvolutionFilter(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*) value)

fn set_get_convolution_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetConvolutionParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_convolution_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetConvolutionParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_histogram(&mut self, value: extern fn(c_uint, c_uchar, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetHistogram(void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, void*) value)

fn set_get_histogram_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetHistogramParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_histogram_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetHistogramParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_minmax(&mut self, value: extern fn(c_uint, c_uchar, c_uint, c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetMinmax(void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, void*) value)

fn set_get_minmax_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetMinmaxParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_minmax_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetMinmaxParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_separable_filter(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_void, *mut c_void, *mut c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_GetSeparableFilter(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, void*, void*, void*) value)

fn set_histogram(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uchar))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_Histogram(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char) value)

fn set_minmax(&mut self, value: extern fn(c_uint, c_uint, c_uchar))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_Minmax(void (*FN_PTR)(unsigned int, unsigned int, unsigned char) value)

fn set_reset_histogram(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ResetHistogram(void (*FN_PTR)(unsigned int) value)

fn set_reset_minmax(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_ResetMinmax(void (*FN_PTR)(unsigned int) value)

fn set_separable_filter2_d(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_uint, c_uint, *const c_void, *const c_void))

C++ method: void QOpenGLFunctions_1_2_DeprecatedBackend::set_SeparableFilter2D(void (*FN_PTR)(unsigned int, unsigned int, int, int, unsigned int, unsigned int, const void*, const void*) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions12DeprecatedBackendVersionStatusArgs

C++ method: QOpenGLFunctions_1_2_DeprecatedBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_1_2_DeprecatedBackend::versionStatus()