Struct qt_gui::open_g_l_version_functions::OpenGLFunctions13CoreBackend

pub struct OpenGLFunctions13CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_1_3_CoreBackend

Methods

impl OpenGLFunctions13CoreBackend

fn active_texture(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_1_3_CoreBackend::ActiveTexture() const

fn compressed_tex_image1_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexImage1D() const

fn compressed_tex_image2_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexImage2D() const

fn compressed_tex_image3_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexImage3D() const

fn compressed_tex_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexSubImage1D() const

fn compressed_tex_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexSubImage2D() const

fn compressed_tex_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_1_3_CoreBackend::CompressedTexSubImage3D() const

fn get_compressed_tex_image(&self) -> extern fn(c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, void*) QOpenGLFunctions_1_3_CoreBackend::GetCompressedTexImage() const

fn sample_coverage(&self) -> extern fn(c_float, c_uchar)

C++ method: void (*FN_PTR)(float, unsigned char) QOpenGLFunctions_1_3_CoreBackend::SampleCoverage() const

fn set_active_texture(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_ActiveTexture(void (*FN_PTR)(unsigned int) value)

fn set_compressed_tex_image1_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexImage1D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, const void*) value)

fn set_compressed_tex_image2_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexImage2D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, const void*) value)

fn set_compressed_tex_image3_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_int, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexImage3D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, int, int, const void*) value)

fn set_compressed_tex_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, unsigned int, int, const void*) value)

fn set_compressed_tex_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, int, const void*) value)

fn set_compressed_tex_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_CompressedTexSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, int, const void*) value)

fn set_get_compressed_tex_image(&mut self, value: extern fn(c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_GetCompressedTexImage(void (*FN_PTR)(unsigned int, int, void*) value)

fn set_sample_coverage(&mut self, value: extern fn(c_float, c_uchar))

C++ method: void QOpenGLFunctions_1_3_CoreBackend::set_SampleCoverage(void (*FN_PTR)(float, unsigned char) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions13CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_1_3_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_1_3_CoreBackend::versionStatus()