Struct qt_gui::open_g_l_version_functions::OpenGLFunctions20CoreBackend

pub struct OpenGLFunctions20CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_2_0_CoreBackend

Methods

impl OpenGLFunctions20CoreBackend

fn attach_shader(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::AttachShader() const

fn bind_attrib_location(&self) -> extern fn(c_uint, c_uint, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_2_0_CoreBackend::BindAttribLocation() const

fn blend_equation_separate(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::BlendEquationSeparate() const

fn compile_shader(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::CompileShader() const

fn create_program(&self) -> extern fn() -> c_uint

C++ method: unsigned int (*FN_PTR)() QOpenGLFunctions_2_0_CoreBackend::CreateProgram() const

fn create_shader(&self) -> extern fn(c_uint) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::CreateShader() const

fn delete_program(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::DeleteProgram() const

fn delete_shader(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::DeleteShader() const

fn detach_shader(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::DetachShader() const

fn disable_vertex_attrib_array(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::DisableVertexAttribArray() const

fn draw_buffers(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_2_0_CoreBackend::DrawBuffers() const

fn enable_vertex_attrib_array(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::EnableVertexAttribArray() const

fn get_active_attrib(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) QOpenGLFunctions_2_0_CoreBackend::GetActiveAttrib() const

fn get_active_uniform(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) QOpenGLFunctions_2_0_CoreBackend::GetActiveUniform() const

fn get_attached_shaders(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int*, unsigned int*) QOpenGLFunctions_2_0_CoreBackend::GetAttachedShaders() const

fn get_attrib_location(&self) -> extern fn(c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, const char*) QOpenGLFunctions_2_0_CoreBackend::GetAttribLocation() const

fn get_program_info_log(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, int, int*, char*) QOpenGLFunctions_2_0_CoreBackend::GetProgramInfoLog() const

fn get_programiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_2_0_CoreBackend::GetProgramiv() const

fn get_shader_info_log(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, int, int*, char*) QOpenGLFunctions_2_0_CoreBackend::GetShaderInfoLog() const

fn get_shader_source(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, int, int*, char*) QOpenGLFunctions_2_0_CoreBackend::GetShaderSource() const

fn get_shaderiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_2_0_CoreBackend::GetShaderiv() const

fn get_uniform_location(&self) -> extern fn(c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, const char*) QOpenGLFunctions_2_0_CoreBackend::GetUniformLocation() const

fn get_uniformfv(&self) -> extern fn(c_uint, c_int, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float*) QOpenGLFunctions_2_0_CoreBackend::GetUniformfv() const

fn get_uniformiv(&self) -> extern fn(c_uint, c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int*) QOpenGLFunctions_2_0_CoreBackend::GetUniformiv() const

fn get_vertex_attrib_pointerv(&self) -> extern fn(c_uint, c_uint, *mut *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, void**) QOpenGLFunctions_2_0_CoreBackend::GetVertexAttribPointerv() const

fn get_vertex_attribdv(&self) -> extern fn(c_uint, c_uint, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, double*) QOpenGLFunctions_2_0_CoreBackend::GetVertexAttribdv() const

fn get_vertex_attribfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_2_0_CoreBackend::GetVertexAttribfv() const

fn get_vertex_attribiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_2_0_CoreBackend::GetVertexAttribiv() const

fn is_program(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::IsProgram() const

fn is_shader(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::IsShader() const

fn link_program(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::LinkProgram() const

fn set_attach_shader(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_AttachShader(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_bind_attrib_location(&mut self, value: extern fn(c_uint, c_uint, *const c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_BindAttribLocation(void (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_blend_equation_separate(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_BlendEquationSeparate(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_compile_shader(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_CompileShader(void (*FN_PTR)(unsigned int) value)

fn set_create_program(&mut self, value: extern fn() -> c_uint)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_CreateProgram(unsigned int (*FN_PTR)() value)

fn set_create_shader(&mut self, value: extern fn(c_uint) -> c_uint)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_CreateShader(unsigned int (*FN_PTR)(unsigned int) value)

fn set_delete_program(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_DeleteProgram(void (*FN_PTR)(unsigned int) value)

fn set_delete_shader(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_DeleteShader(void (*FN_PTR)(unsigned int) value)

fn set_detach_shader(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_DetachShader(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_disable_vertex_attrib_array(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_DisableVertexAttribArray(void (*FN_PTR)(unsigned int) value)

fn set_draw_buffers(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_DrawBuffers(void (*FN_PTR)(int, const unsigned int*) value)

fn set_enable_vertex_attrib_array(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_EnableVertexAttribArray(void (*FN_PTR)(unsigned int) value)

fn set_get_active_attrib(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetActiveAttrib(void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) value)

fn set_get_active_uniform(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetActiveUniform(void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) value)

fn set_get_attached_shaders(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetAttachedShaders(void (*FN_PTR)(unsigned int, int, int*, unsigned int*) value)

fn set_get_attrib_location(&mut self, value: extern fn(c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetAttribLocation(int (*FN_PTR)(unsigned int, const char*) value)

fn set_get_program_info_log(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetProgramInfoLog(void (*FN_PTR)(unsigned int, int, int*, char*) value)

fn set_get_programiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetProgramiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_shader_info_log(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetShaderInfoLog(void (*FN_PTR)(unsigned int, int, int*, char*) value)

fn set_get_shader_source(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetShaderSource(void (*FN_PTR)(unsigned int, int, int*, char*) value)

fn set_get_shaderiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetShaderiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_uniform_location(&mut self, value: extern fn(c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetUniformLocation(int (*FN_PTR)(unsigned int, const char*) value)

fn set_get_uniformfv(&mut self, value: extern fn(c_uint, c_int, *mut c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetUniformfv(void (*FN_PTR)(unsigned int, int, float*) value)

fn set_get_uniformiv(&mut self, value: extern fn(c_uint, c_int, *mut c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetUniformiv(void (*FN_PTR)(unsigned int, int, int*) value)

fn set_get_vertex_attrib_pointerv(&mut self, value: extern fn(c_uint, c_uint, *mut *mut c_void))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetVertexAttribPointerv(void (*FN_PTR)(unsigned int, unsigned int, void**) value)

fn set_get_vertex_attribdv(&mut self, value: extern fn(c_uint, c_uint, *mut c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetVertexAttribdv(void (*FN_PTR)(unsigned int, unsigned int, double*) value)

fn set_get_vertex_attribfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetVertexAttribfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_vertex_attribiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_GetVertexAttribiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_is_program(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_IsProgram(unsigned char (*FN_PTR)(unsigned int) value)

fn set_is_shader(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_IsShader(unsigned char (*FN_PTR)(unsigned int) value)

fn set_link_program(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_LinkProgram(void (*FN_PTR)(unsigned int) value)

fn set_shader_source(&mut self, value: extern fn(c_uint, c_int, *const *const c_char, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_ShaderSource(void (*FN_PTR)(unsigned int, int, const char* const *, const int*) value)

fn set_stencil_func_separate(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_StencilFuncSeparate(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int) value)

fn set_stencil_mask_separate(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_StencilMaskSeparate(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_stencil_op_separate(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_StencilOpSeparate(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_uniform1f(&mut self, value: extern fn(c_int, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform1f(void (*FN_PTR)(int, float) value)

fn set_uniform1fv(&mut self, value: extern fn(c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform1fv(void (*FN_PTR)(int, int, const float*) value)

fn set_uniform1i(&mut self, value: extern fn(c_int, c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform1i(void (*FN_PTR)(int, int) value)

fn set_uniform1iv(&mut self, value: extern fn(c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform1iv(void (*FN_PTR)(int, int, const int*) value)

fn set_uniform2f(&mut self, value: extern fn(c_int, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform2f(void (*FN_PTR)(int, float, float) value)

fn set_uniform2fv(&mut self, value: extern fn(c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform2fv(void (*FN_PTR)(int, int, const float*) value)

fn set_uniform2i(&mut self, value: extern fn(c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform2i(void (*FN_PTR)(int, int, int) value)

fn set_uniform2iv(&mut self, value: extern fn(c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform2iv(void (*FN_PTR)(int, int, const int*) value)

fn set_uniform3f(&mut self, value: extern fn(c_int, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform3f(void (*FN_PTR)(int, float, float, float) value)

fn set_uniform3fv(&mut self, value: extern fn(c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform3fv(void (*FN_PTR)(int, int, const float*) value)

fn set_uniform3i(&mut self, value: extern fn(c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform3i(void (*FN_PTR)(int, int, int, int) value)

fn set_uniform3iv(&mut self, value: extern fn(c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform3iv(void (*FN_PTR)(int, int, const int*) value)

fn set_uniform4f(&mut self, value: extern fn(c_int, c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform4f(void (*FN_PTR)(int, float, float, float, float) value)

fn set_uniform4fv(&mut self, value: extern fn(c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform4fv(void (*FN_PTR)(int, int, const float*) value)

fn set_uniform4i(&mut self, value: extern fn(c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform4i(void (*FN_PTR)(int, int, int, int, int) value)

fn set_uniform4iv(&mut self, value: extern fn(c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_Uniform4iv(void (*FN_PTR)(int, int, const int*) value)

fn set_uniform_matrix2fv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_UniformMatrix2fv(void (*FN_PTR)(int, int, unsigned char, const float*) value)

fn set_uniform_matrix3fv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_UniformMatrix3fv(void (*FN_PTR)(int, int, unsigned char, const float*) value)

fn set_uniform_matrix4fv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_UniformMatrix4fv(void (*FN_PTR)(int, int, unsigned char, const float*) value)

fn set_use_program(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_UseProgram(void (*FN_PTR)(unsigned int) value)

fn set_validate_program(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_ValidateProgram(void (*FN_PTR)(unsigned int) value)

fn set_vertex_attrib1d(&mut self, value: extern fn(c_uint, c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1d(void (*FN_PTR)(unsigned int, double) value)

fn set_vertex_attrib1dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib1f(&mut self, value: extern fn(c_uint, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1f(void (*FN_PTR)(unsigned int, float) value)

fn set_vertex_attrib1fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib1s(&mut self, value: extern fn(c_uint, c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1s(void (*FN_PTR)(unsigned int, short) value)

fn set_vertex_attrib1sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib1sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib2d(&mut self, value: extern fn(c_uint, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2d(void (*FN_PTR)(unsigned int, double, double) value)

fn set_vertex_attrib2dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib2f(&mut self, value: extern fn(c_uint, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2f(void (*FN_PTR)(unsigned int, float, float) value)

fn set_vertex_attrib2fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib2s(&mut self, value: extern fn(c_uint, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2s(void (*FN_PTR)(unsigned int, short, short) value)

fn set_vertex_attrib2sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib2sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib3d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3d(void (*FN_PTR)(unsigned int, double, double, double) value)

fn set_vertex_attrib3dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib3f(&mut self, value: extern fn(c_uint, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3f(void (*FN_PTR)(unsigned int, float, float, float) value)

fn set_vertex_attrib3fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib3s(&mut self, value: extern fn(c_uint, c_short, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3s(void (*FN_PTR)(unsigned int, short, short, short) value)

fn set_vertex_attrib3sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib3sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4_nbv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nbv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib4_niv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Niv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib4_nsv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nsv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4_nub(&mut self, value: extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nub(void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) value)

fn set_vertex_attrib4_nubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib4_nuiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nuiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib4_nusv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4Nusv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn set_vertex_attrib4bv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4bv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib4d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4d(void (*FN_PTR)(unsigned int, double, double, double, double) value)

fn set_vertex_attrib4dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib4f(&mut self, value: extern fn(c_uint, c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4f(void (*FN_PTR)(unsigned int, float, float, float, float) value)

fn set_vertex_attrib4fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib4iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib4s(&mut self, value: extern fn(c_uint, c_short, c_short, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4s(void (*FN_PTR)(unsigned int, short, short, short, short) value)

fn set_vertex_attrib4sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4ubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4ubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib4usv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttrib4usv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn set_vertex_attrib_pointer(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uchar, c_int, *const c_void))

C++ method: void QOpenGLFunctions_2_0_CoreBackend::set_VertexAttribPointer(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char, int, const void*) value)

fn shader_source(&self) -> extern fn(c_uint, c_int, *const *const c_char, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const char* const *, const int*) QOpenGLFunctions_2_0_CoreBackend::ShaderSource() const

fn stencil_func_separate(&self) -> extern fn(c_uint, c_uint, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::StencilFuncSeparate() const

fn stencil_mask_separate(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::StencilMaskSeparate() const

fn stencil_op_separate(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_2_0_CoreBackend::StencilOpSeparate() const

fn uniform1f(&self) -> extern fn(c_int, c_float)

C++ method: void (*FN_PTR)(int, float) QOpenGLFunctions_2_0_CoreBackend::Uniform1f() const

fn uniform1fv(&self) -> extern fn(c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(int, int, const float*) QOpenGLFunctions_2_0_CoreBackend::Uniform1fv() const

fn uniform1i(&self) -> extern fn(c_int, c_int)

C++ method: void (*FN_PTR)(int, int) QOpenGLFunctions_2_0_CoreBackend::Uniform1i() const

fn uniform1iv(&self) -> extern fn(c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(int, int, const int*) QOpenGLFunctions_2_0_CoreBackend::Uniform1iv() const

fn uniform2f(&self) -> extern fn(c_int, c_float, c_float)

C++ method: void (*FN_PTR)(int, float, float) QOpenGLFunctions_2_0_CoreBackend::Uniform2f() const

fn uniform2fv(&self) -> extern fn(c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(int, int, const float*) QOpenGLFunctions_2_0_CoreBackend::Uniform2fv() const

fn uniform2i(&self) -> extern fn(c_int, c_int, c_int)

C++ method: void (*FN_PTR)(int, int, int) QOpenGLFunctions_2_0_CoreBackend::Uniform2i() const

fn uniform2iv(&self) -> extern fn(c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(int, int, const int*) QOpenGLFunctions_2_0_CoreBackend::Uniform2iv() const

fn uniform3f(&self) -> extern fn(c_int, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(int, float, float, float) QOpenGLFunctions_2_0_CoreBackend::Uniform3f() const

fn uniform3fv(&self) -> extern fn(c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(int, int, const float*) QOpenGLFunctions_2_0_CoreBackend::Uniform3fv() const

fn uniform3i(&self) -> extern fn(c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(int, int, int, int) QOpenGLFunctions_2_0_CoreBackend::Uniform3i() const

fn uniform3iv(&self) -> extern fn(c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(int, int, const int*) QOpenGLFunctions_2_0_CoreBackend::Uniform3iv() const

fn uniform4f(&self) -> extern fn(c_int, c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(int, float, float, float, float) QOpenGLFunctions_2_0_CoreBackend::Uniform4f() const

fn uniform4fv(&self) -> extern fn(c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(int, int, const float*) QOpenGLFunctions_2_0_CoreBackend::Uniform4fv() const

fn uniform4i(&self) -> extern fn(c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(int, int, int, int, int) QOpenGLFunctions_2_0_CoreBackend::Uniform4i() const

fn uniform4iv(&self) -> extern fn(c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(int, int, const int*) QOpenGLFunctions_2_0_CoreBackend::Uniform4iv() const

fn uniform_matrix2fv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(int, int, unsigned char, const float*) QOpenGLFunctions_2_0_CoreBackend::UniformMatrix2fv() const

fn uniform_matrix3fv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(int, int, unsigned char, const float*) QOpenGLFunctions_2_0_CoreBackend::UniformMatrix3fv() const

fn uniform_matrix4fv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(int, int, unsigned char, const float*) QOpenGLFunctions_2_0_CoreBackend::UniformMatrix4fv() const

fn use_program(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::UseProgram() const

fn validate_program(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_2_0_CoreBackend::ValidateProgram() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions20CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_2_0_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_2_0_CoreBackend::versionStatus()

fn vertex_attrib1d(&self) -> extern fn(c_uint, c_double)

C++ method: void (*FN_PTR)(unsigned int, double) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1d() const

fn vertex_attrib1dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1dv() const

fn vertex_attrib1f(&self) -> extern fn(c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, float) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1f() const

fn vertex_attrib1fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1fv() const

fn vertex_attrib1s(&self) -> extern fn(c_uint, c_short)

C++ method: void (*FN_PTR)(unsigned int, short) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1s() const

fn vertex_attrib1sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib1sv() const

fn vertex_attrib2d(&self) -> extern fn(c_uint, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2d() const

fn vertex_attrib2dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2dv() const

fn vertex_attrib2f(&self) -> extern fn(c_uint, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2f() const

fn vertex_attrib2fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2fv() const

fn vertex_attrib2s(&self) -> extern fn(c_uint, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2s() const

fn vertex_attrib2sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib2sv() const

fn vertex_attrib3d(&self) -> extern fn(c_uint, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3d() const

fn vertex_attrib3dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3dv() const

fn vertex_attrib3f(&self) -> extern fn(c_uint, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float, float) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3f() const

fn vertex_attrib3fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3fv() const

fn vertex_attrib3s(&self) -> extern fn(c_uint, c_short, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short, short) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3s() const

fn vertex_attrib3sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib3sv() const

fn vertex_attrib4_nbv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nbv() const

fn vertex_attrib4_niv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Niv() const

fn vertex_attrib4_nsv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nsv() const

fn vertex_attrib4_nub(&self) -> extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nub() const

fn vertex_attrib4_nubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nubv() const

fn vertex_attrib4_nuiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nuiv() const

fn vertex_attrib4_nusv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4Nusv() const

fn vertex_attrib4bv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4bv() const

fn vertex_attrib4d(&self) -> extern fn(c_uint, c_double, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double, double) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4d() const

fn vertex_attrib4dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4dv() const

fn vertex_attrib4f(&self) -> extern fn(c_uint, c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float, float, float) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4f() const

fn vertex_attrib4fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4fv() const

fn vertex_attrib4iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4iv() const

fn vertex_attrib4s(&self) -> extern fn(c_uint, c_short, c_short, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short, short, short) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4s() const

fn vertex_attrib4sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4sv() const

fn vertex_attrib4ubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4ubv() const

fn vertex_attrib4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4uiv() const

fn vertex_attrib4usv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_2_0_CoreBackend::VertexAttrib4usv() const

fn vertex_attrib_pointer(&self) -> extern fn(c_uint, c_int, c_uint, c_uchar, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char, int, const void*) QOpenGLFunctions_2_0_CoreBackend::VertexAttribPointer() const