Struct qt_gui::open_g_l_version_functions::OpenGLFunctions20DeprecatedBackend

pub struct OpenGLFunctions20DeprecatedBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_2_0_DeprecatedBackend

Methods

impl OpenGLFunctions20DeprecatedBackend

fn set_vertex_attrib1d(&mut self, value: extern fn(c_uint, c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1d(void (*FN_PTR)(unsigned int, double) value)

fn set_vertex_attrib1dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib1f(&mut self, value: extern fn(c_uint, c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1f(void (*FN_PTR)(unsigned int, float) value)

fn set_vertex_attrib1fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib1s(&mut self, value: extern fn(c_uint, c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1s(void (*FN_PTR)(unsigned int, short) value)

fn set_vertex_attrib1sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib1sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib2d(&mut self, value: extern fn(c_uint, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2d(void (*FN_PTR)(unsigned int, double, double) value)

fn set_vertex_attrib2dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib2f(&mut self, value: extern fn(c_uint, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2f(void (*FN_PTR)(unsigned int, float, float) value)

fn set_vertex_attrib2fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib2s(&mut self, value: extern fn(c_uint, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2s(void (*FN_PTR)(unsigned int, short, short) value)

fn set_vertex_attrib2sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib2sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib3d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3d(void (*FN_PTR)(unsigned int, double, double, double) value)

fn set_vertex_attrib3dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib3f(&mut self, value: extern fn(c_uint, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3f(void (*FN_PTR)(unsigned int, float, float, float) value)

fn set_vertex_attrib3fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib3s(&mut self, value: extern fn(c_uint, c_short, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3s(void (*FN_PTR)(unsigned int, short, short, short) value)

fn set_vertex_attrib3sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib3sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4_nbv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nbv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib4_niv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Niv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib4_nsv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nsv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4_nub(&mut self, value: extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nub(void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) value)

fn set_vertex_attrib4_nubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib4_nuiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nuiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib4_nusv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4Nusv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn set_vertex_attrib4bv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4bv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib4d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4d(void (*FN_PTR)(unsigned int, double, double, double, double) value)

fn set_vertex_attrib4dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib4f(&mut self, value: extern fn(c_uint, c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4f(void (*FN_PTR)(unsigned int, float, float, float, float) value)

fn set_vertex_attrib4fv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4fv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_vertex_attrib4iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib4s(&mut self, value: extern fn(c_uint, c_short, c_short, c_short, c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4s(void (*FN_PTR)(unsigned int, short, short, short, short) value)

fn set_vertex_attrib4sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib4ubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4ubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib4usv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_2_0_DeprecatedBackend::set_VertexAttrib4usv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions20DeprecatedBackendVersionStatusArgs

C++ method: QOpenGLFunctions_2_0_DeprecatedBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_2_0_DeprecatedBackend::versionStatus()

fn vertex_attrib1d(&self) -> extern fn(c_uint, c_double)

C++ method: void (*FN_PTR)(unsigned int, double) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1d() const

fn vertex_attrib1dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1dv() const

fn vertex_attrib1f(&self) -> extern fn(c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, float) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1f() const

fn vertex_attrib1fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1fv() const

fn vertex_attrib1s(&self) -> extern fn(c_uint, c_short)

C++ method: void (*FN_PTR)(unsigned int, short) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1s() const

fn vertex_attrib1sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib1sv() const

fn vertex_attrib2d(&self) -> extern fn(c_uint, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2d() const

fn vertex_attrib2dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2dv() const

fn vertex_attrib2f(&self) -> extern fn(c_uint, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2f() const

fn vertex_attrib2fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2fv() const

fn vertex_attrib2s(&self) -> extern fn(c_uint, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2s() const

fn vertex_attrib2sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib2sv() const

fn vertex_attrib3d(&self) -> extern fn(c_uint, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3d() const

fn vertex_attrib3dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3dv() const

fn vertex_attrib3f(&self) -> extern fn(c_uint, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float, float) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3f() const

fn vertex_attrib3fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3fv() const

fn vertex_attrib3s(&self) -> extern fn(c_uint, c_short, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short, short) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3s() const

fn vertex_attrib3sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib3sv() const

fn vertex_attrib4_nbv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nbv() const

fn vertex_attrib4_niv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Niv() const

fn vertex_attrib4_nsv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nsv() const

fn vertex_attrib4_nub(&self) -> extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nub() const

fn vertex_attrib4_nubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nubv() const

fn vertex_attrib4_nuiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nuiv() const

fn vertex_attrib4_nusv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4Nusv() const

fn vertex_attrib4bv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4bv() const

fn vertex_attrib4d(&self) -> extern fn(c_uint, c_double, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double, double) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4d() const

fn vertex_attrib4dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4dv() const

fn vertex_attrib4f(&self) -> extern fn(c_uint, c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float, float, float) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4f() const

fn vertex_attrib4fv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4fv() const

fn vertex_attrib4iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4iv() const

fn vertex_attrib4s(&self) -> extern fn(c_uint, c_short, c_short, c_short, c_short)

C++ method: void (*FN_PTR)(unsigned int, short, short, short, short) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4s() const

fn vertex_attrib4sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4sv() const

fn vertex_attrib4ubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4ubv() const

fn vertex_attrib4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4uiv() const

fn vertex_attrib4usv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_2_0_DeprecatedBackend::VertexAttrib4usv() const