Struct qt_gui::open_g_l_version_functions::OpenGLFunctions30CoreBackend

pub struct OpenGLFunctions30CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_3_0_CoreBackend

Methods

impl OpenGLFunctions30CoreBackend

fn begin_conditional_render(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::BeginConditionalRender() const

fn begin_transform_feedback(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::BeginTransformFeedback() const

fn bind_buffer_base(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::BindBufferBase() const

fn bind_buffer_range(&self) -> extern fn(c_uint, c_uint, c_uint, c_long, c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, long) QOpenGLFunctions_3_0_CoreBackend::BindBufferRange() const

fn bind_frag_data_location(&self) -> extern fn(c_uint, c_uint, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_3_0_CoreBackend::BindFragDataLocation() const

fn bind_framebuffer(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::BindFramebuffer() const

fn bind_renderbuffer(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::BindRenderbuffer() const

fn bind_vertex_array(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::BindVertexArray() const

fn blit_framebuffer(&self) -> extern fn(c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(int, int, int, int, int, int, int, int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::BlitFramebuffer() const

fn check_framebuffer_status(&self) -> extern fn(c_uint) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::CheckFramebufferStatus() const

fn clamp_color(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::ClampColor() const

fn clear_bufferfi(&self) -> extern fn(c_uint, c_int, c_float, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, float, int) QOpenGLFunctions_3_0_CoreBackend::ClearBufferfi() const

fn clear_bufferfv(&self) -> extern fn(c_uint, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, const float*) QOpenGLFunctions_3_0_CoreBackend::ClearBufferfv() const

fn clear_bufferiv(&self) -> extern fn(c_uint, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const int*) QOpenGLFunctions_3_0_CoreBackend::ClearBufferiv() const

fn clear_bufferuiv(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::ClearBufferuiv() const

fn color_maski(&self) -> extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) QOpenGLFunctions_3_0_CoreBackend::ColorMaski() const

fn delete_framebuffers(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::DeleteFramebuffers() const

fn delete_renderbuffers(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::DeleteRenderbuffers() const

fn delete_vertex_arrays(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::DeleteVertexArrays() const

fn disablei(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Disablei() const

fn enablei(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Enablei() const

fn end_conditional_render(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_3_0_CoreBackend::EndConditionalRender() const

fn end_transform_feedback(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_3_0_CoreBackend::EndTransformFeedback() const

fn flush_mapped_buffer_range(&self) -> extern fn(c_uint, c_long, c_long)

C++ method: void (*FN_PTR)(unsigned int, long, long) QOpenGLFunctions_3_0_CoreBackend::FlushMappedBufferRange() const

fn framebuffer_renderbuffer(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::FramebufferRenderbuffer() const

fn framebuffer_texture1_d(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int) QOpenGLFunctions_3_0_CoreBackend::FramebufferTexture1D() const

fn framebuffer_texture2_d(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int) QOpenGLFunctions_3_0_CoreBackend::FramebufferTexture2D() const

fn framebuffer_texture3_d(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int, int) QOpenGLFunctions_3_0_CoreBackend::FramebufferTexture3D() const

fn framebuffer_texture_layer(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int) QOpenGLFunctions_3_0_CoreBackend::FramebufferTextureLayer() const

fn gen_framebuffers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GenFramebuffers() const

fn gen_renderbuffers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GenRenderbuffers() const

fn gen_vertex_arrays(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GenVertexArrays() const

fn generate_mipmap(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::GenerateMipmap() const

fn get_booleani_v(&self) -> extern fn(c_uint, c_uint, *mut c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char*) QOpenGLFunctions_3_0_CoreBackend::GetBooleani_v() const

fn get_frag_data_location(&self) -> extern fn(c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, const char*) QOpenGLFunctions_3_0_CoreBackend::GetFragDataLocation() const

fn get_framebuffer_attachment_parameteriv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_3_0_CoreBackend::GetFramebufferAttachmentParameteriv() const

fn get_integeri_v(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_0_CoreBackend::GetIntegeri_v() const

fn get_renderbuffer_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_0_CoreBackend::GetRenderbufferParameteriv() const

fn get_stringi(&self) -> extern fn(c_uint, c_uint) -> *const c_uchar

C++ method: const unsigned char* (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::GetStringi() const

fn get_tex_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_0_CoreBackend::GetTexParameterIiv() const

fn get_tex_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GetTexParameterIuiv() const

fn get_transform_feedback_varying(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) QOpenGLFunctions_3_0_CoreBackend::GetTransformFeedbackVarying() const

fn get_uniformuiv(&self) -> extern fn(c_uint, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GetUniformuiv() const

fn get_vertex_attrib_iiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_0_CoreBackend::GetVertexAttribIiv() const

fn get_vertex_attrib_iuiv(&self) -> extern fn(c_uint, c_uint, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) QOpenGLFunctions_3_0_CoreBackend::GetVertexAttribIuiv() const

fn is_enabledi(&self) -> extern fn(c_uint, c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::IsEnabledi() const

fn is_framebuffer(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::IsFramebuffer() const

fn is_renderbuffer(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::IsRenderbuffer() const

fn is_vertex_array(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_3_0_CoreBackend::IsVertexArray() const

fn map_buffer_range(&self) -> extern fn(c_uint, c_long, c_long, c_uint) -> *mut c_void

C++ method: void* (*FN_PTR)(unsigned int, long, long, unsigned int) QOpenGLFunctions_3_0_CoreBackend::MapBufferRange() const

fn renderbuffer_storage(&self) -> extern fn(c_uint, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int) QOpenGLFunctions_3_0_CoreBackend::RenderbufferStorage() const

fn renderbuffer_storage_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int) QOpenGLFunctions_3_0_CoreBackend::RenderbufferStorageMultisample() const

fn set_begin_conditional_render(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BeginConditionalRender(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_begin_transform_feedback(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BeginTransformFeedback(void (*FN_PTR)(unsigned int) value)

fn set_bind_buffer_base(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindBufferBase(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_bind_buffer_range(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_long, c_long))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindBufferRange(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, long) value)

fn set_bind_frag_data_location(&mut self, value: extern fn(c_uint, c_uint, *const c_char))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindFragDataLocation(void (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_bind_framebuffer(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindFramebuffer(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_bind_renderbuffer(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindRenderbuffer(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_bind_vertex_array(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BindVertexArray(void (*FN_PTR)(unsigned int) value)

fn set_blit_framebuffer(&mut self, value: extern fn(c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_BlitFramebuffer(void (*FN_PTR)(int, int, int, int, int, int, int, int, unsigned int, unsigned int) value)

fn set_check_framebuffer_status(&mut self, value: extern fn(c_uint) -> c_uint)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_CheckFramebufferStatus(unsigned int (*FN_PTR)(unsigned int) value)

fn set_clamp_color(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ClampColor(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_clear_bufferfi(&mut self, value: extern fn(c_uint, c_int, c_float, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ClearBufferfi(void (*FN_PTR)(unsigned int, int, float, int) value)

fn set_clear_bufferfv(&mut self, value: extern fn(c_uint, c_int, *const c_float))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ClearBufferfv(void (*FN_PTR)(unsigned int, int, const float*) value)

fn set_clear_bufferiv(&mut self, value: extern fn(c_uint, c_int, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ClearBufferiv(void (*FN_PTR)(unsigned int, int, const int*) value)

fn set_clear_bufferuiv(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ClearBufferuiv(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_color_maski(&mut self, value: extern fn(c_uint, c_uchar, c_uchar, c_uchar, c_uchar))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_ColorMaski(void (*FN_PTR)(unsigned int, unsigned char, unsigned char, unsigned char, unsigned char) value)

fn set_delete_framebuffers(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_DeleteFramebuffers(void (*FN_PTR)(int, const unsigned int*) value)

fn set_delete_renderbuffers(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_DeleteRenderbuffers(void (*FN_PTR)(int, const unsigned int*) value)

fn set_delete_vertex_arrays(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_DeleteVertexArrays(void (*FN_PTR)(int, const unsigned int*) value)

fn set_disablei(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Disablei(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_enablei(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Enablei(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_end_conditional_render(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_EndConditionalRender(void (*FN_PTR)() value)

fn set_end_transform_feedback(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_EndTransformFeedback(void (*FN_PTR)() value)

fn set_flush_mapped_buffer_range(&mut self, value: extern fn(c_uint, c_long, c_long))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FlushMappedBufferRange(void (*FN_PTR)(unsigned int, long, long) value)

fn set_framebuffer_renderbuffer(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FramebufferRenderbuffer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_framebuffer_texture1_d(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FramebufferTexture1D(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int) value)

fn set_framebuffer_texture2_d(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FramebufferTexture2D(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int) value)

fn set_framebuffer_texture3_d(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FramebufferTexture3D(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int, int) value)

fn set_framebuffer_texture_layer(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_FramebufferTextureLayer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int) value)

fn set_gen_framebuffers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GenFramebuffers(void (*FN_PTR)(int, unsigned int*) value)

fn set_gen_renderbuffers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GenRenderbuffers(void (*FN_PTR)(int, unsigned int*) value)

fn set_gen_vertex_arrays(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GenVertexArrays(void (*FN_PTR)(int, unsigned int*) value)

fn set_generate_mipmap(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GenerateMipmap(void (*FN_PTR)(unsigned int) value)

fn set_get_booleani_v(&mut self, value: extern fn(c_uint, c_uint, *mut c_uchar))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetBooleani_v(void (*FN_PTR)(unsigned int, unsigned int, unsigned char*) value)

fn set_get_frag_data_location(&mut self, value: extern fn(c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetFragDataLocation(int (*FN_PTR)(unsigned int, const char*) value)

fn set_get_framebuffer_attachment_parameteriv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetFramebufferAttachmentParameteriv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_integeri_v(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetIntegeri_v(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_renderbuffer_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetRenderbufferParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_stringi(&mut self, value: extern fn(c_uint, c_uint) -> *const c_uchar)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetStringi(const unsigned char* (*FN_PTR)(unsigned int, unsigned int) value)

fn set_get_tex_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetTexParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_tex_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetTexParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) value)

fn set_get_transform_feedback_varying(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_int, *mut c_uint, *mut c_char))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetTransformFeedbackVarying(void (*FN_PTR)(unsigned int, unsigned int, int, int*, int*, unsigned int*, char*) value)

fn set_get_uniformuiv(&mut self, value: extern fn(c_uint, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetUniformuiv(void (*FN_PTR)(unsigned int, int, unsigned int*) value)

fn set_get_vertex_attrib_iiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetVertexAttribIiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_vertex_attrib_iuiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_GetVertexAttribIuiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) value)

fn set_is_enabledi(&mut self, value: extern fn(c_uint, c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_IsEnabledi(unsigned char (*FN_PTR)(unsigned int, unsigned int) value)

fn set_is_framebuffer(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_IsFramebuffer(unsigned char (*FN_PTR)(unsigned int) value)

fn set_is_renderbuffer(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_IsRenderbuffer(unsigned char (*FN_PTR)(unsigned int) value)

fn set_is_vertex_array(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_IsVertexArray(unsigned char (*FN_PTR)(unsigned int) value)

fn set_map_buffer_range(&mut self, value: extern fn(c_uint, c_long, c_long, c_uint) -> *mut c_void)

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_MapBufferRange(void* (*FN_PTR)(unsigned int, long, long, unsigned int) value)

fn set_renderbuffer_storage(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_RenderbufferStorage(void (*FN_PTR)(unsigned int, unsigned int, int, int) value)

fn set_renderbuffer_storage_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_RenderbufferStorageMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int) value)

fn set_tex_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_TexParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_tex_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_TexParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_transform_feedback_varyings(&mut self, value: extern fn(c_uint, c_int, *const *const c_char, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_TransformFeedbackVaryings(void (*FN_PTR)(unsigned int, int, const char* const *, unsigned int) value)

fn set_uniform1ui(&mut self, value: extern fn(c_int, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform1ui(void (*FN_PTR)(int, unsigned int) value)

fn set_uniform1uiv(&mut self, value: extern fn(c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform1uiv(void (*FN_PTR)(int, int, const unsigned int*) value)

fn set_uniform2ui(&mut self, value: extern fn(c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform2ui(void (*FN_PTR)(int, unsigned int, unsigned int) value)

fn set_uniform2uiv(&mut self, value: extern fn(c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform2uiv(void (*FN_PTR)(int, int, const unsigned int*) value)

fn set_uniform3ui(&mut self, value: extern fn(c_int, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform3ui(void (*FN_PTR)(int, unsigned int, unsigned int, unsigned int) value)

fn set_uniform3uiv(&mut self, value: extern fn(c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform3uiv(void (*FN_PTR)(int, int, const unsigned int*) value)

fn set_uniform4ui(&mut self, value: extern fn(c_int, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform4ui(void (*FN_PTR)(int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_uniform4uiv(&mut self, value: extern fn(c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_Uniform4uiv(void (*FN_PTR)(int, int, const unsigned int*) value)

fn set_vertex_attrib_i1i(&mut self, value: extern fn(c_uint, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI1i(void (*FN_PTR)(unsigned int, int) value)

fn set_vertex_attrib_i1iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI1iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i1ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI1ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_attrib_i1uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI1uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i2i(&mut self, value: extern fn(c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI2i(void (*FN_PTR)(unsigned int, int, int) value)

fn set_vertex_attrib_i2iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI2iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i2ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI2ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i2uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI2uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i3i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI3i(void (*FN_PTR)(unsigned int, int, int, int) value)

fn set_vertex_attrib_i3iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI3iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i3ui(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI3ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i4bv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4bv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib_i4i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4i(void (*FN_PTR)(unsigned int, int, int, int, int) value)

fn set_vertex_attrib_i4iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i4sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib_i4ubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4ubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib_i4ui(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i4usv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribI4usv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn set_vertex_attrib_i_pointer(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_3_0_CoreBackend::set_VertexAttribIPointer(void (*FN_PTR)(unsigned int, int, unsigned int, int, const void*) value)

fn tex_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_3_0_CoreBackend::TexParameterIiv() const

fn tex_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::TexParameterIuiv() const

fn transform_feedback_varyings(&self) -> extern fn(c_uint, c_int, *const *const c_char, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const char* const *, unsigned int) QOpenGLFunctions_3_0_CoreBackend::TransformFeedbackVaryings() const

fn uniform1ui(&self) -> extern fn(c_int, c_uint)

C++ method: void (*FN_PTR)(int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Uniform1ui() const

fn uniform1uiv(&self) -> extern fn(c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::Uniform1uiv() const

fn uniform2ui(&self) -> extern fn(c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Uniform2ui() const

fn uniform2uiv(&self) -> extern fn(c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::Uniform2uiv() const

fn uniform3ui(&self) -> extern fn(c_int, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Uniform3ui() const

fn uniform3uiv(&self) -> extern fn(c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::Uniform3uiv() const

fn uniform4ui(&self) -> extern fn(c_int, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::Uniform4ui() const

fn uniform4uiv(&self) -> extern fn(c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::Uniform4uiv() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions30CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_3_0_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_3_0_CoreBackend::versionStatus()

fn vertex_attrib_i1i(&self) -> extern fn(c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI1i() const

fn vertex_attrib_i1iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI1iv() const

fn vertex_attrib_i1ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI1ui() const

fn vertex_attrib_i1uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI1uiv() const

fn vertex_attrib_i2i(&self) -> extern fn(c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI2i() const

fn vertex_attrib_i2iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI2iv() const

fn vertex_attrib_i2ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI2ui() const

fn vertex_attrib_i2uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI2uiv() const

fn vertex_attrib_i3i(&self) -> extern fn(c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI3i() const

fn vertex_attrib_i3iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI3iv() const

fn vertex_attrib_i3ui(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI3ui() const

fn vertex_attrib_i3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI3uiv() const

fn vertex_attrib_i4bv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4bv() const

fn vertex_attrib_i4i(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4i() const

fn vertex_attrib_i4iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4iv() const

fn vertex_attrib_i4sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4sv() const

fn vertex_attrib_i4ubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4ubv() const

fn vertex_attrib_i4ui(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4ui() const

fn vertex_attrib_i4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4uiv() const

fn vertex_attrib_i4usv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribI4usv() const

fn vertex_attrib_i_pointer(&self) -> extern fn(c_uint, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, const void*) QOpenGLFunctions_3_0_CoreBackend::VertexAttribIPointer() const