Struct qt_gui::open_g_l_version_functions::OpenGLFunctions30DeprecatedBackend

pub struct OpenGLFunctions30DeprecatedBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_3_0_DeprecatedBackend

Methods

impl OpenGLFunctions30DeprecatedBackend

fn set_vertex_attrib_i1i(&mut self, value: extern fn(c_uint, c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI1i(void (*FN_PTR)(unsigned int, int) value)

fn set_vertex_attrib_i1iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI1iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i1ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI1ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_attrib_i1uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI1uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i2i(&mut self, value: extern fn(c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI2i(void (*FN_PTR)(unsigned int, int, int) value)

fn set_vertex_attrib_i2iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI2iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i2ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI2ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i2uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI2uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i3i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI3i(void (*FN_PTR)(unsigned int, int, int, int) value)

fn set_vertex_attrib_i3iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI3iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i3ui(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI3ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i4bv(&mut self, value: extern fn(c_uint, *const c_schar))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4bv(void (*FN_PTR)(unsigned int, const signed char*) value)

fn set_vertex_attrib_i4i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4i(void (*FN_PTR)(unsigned int, int, int, int, int) value)

fn set_vertex_attrib_i4iv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4iv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_vertex_attrib_i4sv(&mut self, value: extern fn(c_uint, *const c_short))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4sv(void (*FN_PTR)(unsigned int, const short*) value)

fn set_vertex_attrib_i4ubv(&mut self, value: extern fn(c_uint, *const c_uchar))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4ubv(void (*FN_PTR)(unsigned int, const unsigned char*) value)

fn set_vertex_attrib_i4ui(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_i4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_i4usv(&mut self, value: extern fn(c_uint, *const c_ushort))

C++ method: void QOpenGLFunctions_3_0_DeprecatedBackend::set_VertexAttribI4usv(void (*FN_PTR)(unsigned int, const unsigned short*) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions30DeprecatedBackendVersionStatusArgs

C++ method: QOpenGLFunctions_3_0_DeprecatedBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_3_0_DeprecatedBackend::versionStatus()

fn vertex_attrib_i1i(&self) -> extern fn(c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI1i() const

fn vertex_attrib_i1iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI1iv() const

fn vertex_attrib_i1ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI1ui() const

fn vertex_attrib_i1uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI1uiv() const

fn vertex_attrib_i2i(&self) -> extern fn(c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI2i() const

fn vertex_attrib_i2iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI2iv() const

fn vertex_attrib_i2ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI2ui() const

fn vertex_attrib_i2uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI2uiv() const

fn vertex_attrib_i3i(&self) -> extern fn(c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI3i() const

fn vertex_attrib_i3iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI3iv() const

fn vertex_attrib_i3ui(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI3ui() const

fn vertex_attrib_i3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI3uiv() const

fn vertex_attrib_i4bv(&self) -> extern fn(c_uint, *const c_schar)

C++ method: void (*FN_PTR)(unsigned int, const signed char*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4bv() const

fn vertex_attrib_i4i(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4i() const

fn vertex_attrib_i4iv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4iv() const

fn vertex_attrib_i4sv(&self) -> extern fn(c_uint, *const c_short)

C++ method: void (*FN_PTR)(unsigned int, const short*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4sv() const

fn vertex_attrib_i4ubv(&self) -> extern fn(c_uint, *const c_uchar)

C++ method: void (*FN_PTR)(unsigned int, const unsigned char*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4ubv() const

fn vertex_attrib_i4ui(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4ui() const

fn vertex_attrib_i4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4uiv() const

fn vertex_attrib_i4usv(&self) -> extern fn(c_uint, *const c_ushort)

C++ method: void (*FN_PTR)(unsigned int, const unsigned short*) QOpenGLFunctions_3_0_DeprecatedBackend::VertexAttribI4usv() const