Struct qt_gui::open_g_l_version_functions::OpenGLFunctions32CoreBackend

pub struct OpenGLFunctions32CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_3_2_CoreBackend

Methods

impl OpenGLFunctions32CoreBackend

fn draw_elements_base_vertex(&self) -> extern fn(c_uint, c_int, c_uint, *const c_void, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int) QOpenGLFunctions_3_2_CoreBackend::DrawElementsBaseVertex() const

fn draw_elements_instanced_base_vertex(&self) -> extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, int) QOpenGLFunctions_3_2_CoreBackend::DrawElementsInstancedBaseVertex() const

fn draw_range_elements_base_vertex(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, c_uint, *const c_void, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, unsigned int, const void*, int) QOpenGLFunctions_3_2_CoreBackend::DrawRangeElementsBaseVertex() const

fn framebuffer_texture(&self) -> extern fn(c_uint, c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) QOpenGLFunctions_3_2_CoreBackend::FramebufferTexture() const

fn get_buffer_parameteri64v(&self) -> extern fn(c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long*) QOpenGLFunctions_3_2_CoreBackend::GetBufferParameteri64v() const

fn get_integer64i_v(&self) -> extern fn(c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long*) QOpenGLFunctions_3_2_CoreBackend::GetInteger64i_v() const

fn get_integer64v(&self) -> extern fn(c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, long*) QOpenGLFunctions_3_2_CoreBackend::GetInteger64v() const

fn get_multisamplefv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_3_2_CoreBackend::GetMultisamplefv() const

fn multi_draw_elements_base_vertex(&self) -> extern fn(c_uint, *const c_int, c_uint, *const *const c_void, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*, unsigned int, const void* const *, int, const int*) QOpenGLFunctions_3_2_CoreBackend::MultiDrawElementsBaseVertex() const

fn provoking_vertex(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_3_2_CoreBackend::ProvokingVertex() const

fn sample_maski(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_2_CoreBackend::SampleMaski() const

fn set_draw_elements_base_vertex(&mut self, value: extern fn(c_uint, c_int, c_uint, *const c_void, c_int))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_DrawElementsBaseVertex(void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int) value)

fn set_draw_elements_instanced_base_vertex(&mut self, value: extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_int))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_DrawElementsInstancedBaseVertex(void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, int) value)

fn set_draw_range_elements_base_vertex(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, c_uint, *const c_void, c_int))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_DrawRangeElementsBaseVertex(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, unsigned int, const void*, int) value)

fn set_framebuffer_texture(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_FramebufferTexture(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) value)

fn set_get_buffer_parameteri64v(&mut self, value: extern fn(c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_GetBufferParameteri64v(void (*FN_PTR)(unsigned int, unsigned int, long*) value)

fn set_get_integer64i_v(&mut self, value: extern fn(c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_GetInteger64i_v(void (*FN_PTR)(unsigned int, unsigned int, long*) value)

fn set_get_integer64v(&mut self, value: extern fn(c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_GetInteger64v(void (*FN_PTR)(unsigned int, long*) value)

fn set_get_multisamplefv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_GetMultisamplefv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_multi_draw_elements_base_vertex(&mut self, value: extern fn(c_uint, *const c_int, c_uint, *const *const c_void, c_int, *const c_int))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_MultiDrawElementsBaseVertex(void (*FN_PTR)(unsigned int, const int*, unsigned int, const void* const *, int, const int*) value)

fn set_provoking_vertex(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_ProvokingVertex(void (*FN_PTR)(unsigned int) value)

fn set_sample_maski(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_SampleMaski(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_tex_image2_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_TexImage2DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) value)

fn set_tex_image3_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_3_2_CoreBackend::set_TexImage3DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) value)

fn tex_image2_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) QOpenGLFunctions_3_2_CoreBackend::TexImage2DMultisample() const

fn tex_image3_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) QOpenGLFunctions_3_2_CoreBackend::TexImage3DMultisample() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions32CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_3_2_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_3_2_CoreBackend::versionStatus()