Struct qt_gui::open_g_l_version_functions::OpenGLFunctions33CoreBackend

pub struct OpenGLFunctions33CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_3_3_CoreBackend

Methods

impl OpenGLFunctions33CoreBackend

fn bind_frag_data_location_indexed(&self) -> extern fn(c_uint, c_uint, c_uint, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, const char*) QOpenGLFunctions_3_3_CoreBackend::BindFragDataLocationIndexed() const

fn bind_sampler(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::BindSampler() const

fn color_p3ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::ColorP3ui() const

fn color_p3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::ColorP3uiv() const

fn color_p4ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::ColorP4ui() const

fn color_p4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::ColorP4uiv() const

fn delete_samplers(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::DeleteSamplers() const

fn gen_samplers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_3_3_CoreBackend::GenSamplers() const

fn get_frag_data_index(&self) -> extern fn(c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, const char*) QOpenGLFunctions_3_3_CoreBackend::GetFragDataIndex() const

fn get_query_objecti64v(&self) -> extern fn(c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long*) QOpenGLFunctions_3_3_CoreBackend::GetQueryObjecti64v() const

fn get_query_objectui64v(&self) -> extern fn(c_uint, c_uint, *mut c_ulong)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned long*) QOpenGLFunctions_3_3_CoreBackend::GetQueryObjectui64v() const

fn get_sampler_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_3_CoreBackend::GetSamplerParameterIiv() const

fn get_sampler_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) QOpenGLFunctions_3_3_CoreBackend::GetSamplerParameterIuiv() const

fn get_sampler_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_3_3_CoreBackend::GetSamplerParameterfv() const

fn get_sampler_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_3_3_CoreBackend::GetSamplerParameteriv() const

fn is_sampler(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_3_3_CoreBackend::IsSampler() const

fn multi_tex_coord_p1ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP1ui() const

fn multi_tex_coord_p1uiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP1uiv() const

fn multi_tex_coord_p2ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP2ui() const

fn multi_tex_coord_p2uiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP2uiv() const

fn multi_tex_coord_p3ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP3ui() const

fn multi_tex_coord_p3uiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP3uiv() const

fn multi_tex_coord_p4ui(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP4ui() const

fn multi_tex_coord_p4uiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::MultiTexCoordP4uiv() const

fn normal_p3ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::NormalP3ui() const

fn normal_p3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::NormalP3uiv() const

fn query_counter(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::QueryCounter() const

fn sampler_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_3_3_CoreBackend::SamplerParameterIiv() const

fn sampler_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::SamplerParameterIuiv() const

fn sampler_parameterf(&self) -> extern fn(c_uint, c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float) QOpenGLFunctions_3_3_CoreBackend::SamplerParameterf() const

fn sampler_parameterfv(&self) -> extern fn(c_uint, c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const float*) QOpenGLFunctions_3_3_CoreBackend::SamplerParameterfv() const

fn sampler_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_3_3_CoreBackend::SamplerParameteri() const

fn sampler_parameteriv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_3_3_CoreBackend::SamplerParameteriv() const

fn secondary_color_p3ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::SecondaryColorP3ui() const

fn secondary_color_p3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::SecondaryColorP3uiv() const

fn set_bind_frag_data_location_indexed(&mut self, value: extern fn(c_uint, c_uint, c_uint, *const c_char))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_BindFragDataLocationIndexed(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, const char*) value)

fn set_bind_sampler(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_BindSampler(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_color_p3ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_ColorP3ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_color_p3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_ColorP3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_color_p4ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_ColorP4ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_color_p4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_ColorP4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_delete_samplers(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_DeleteSamplers(void (*FN_PTR)(int, const unsigned int*) value)

fn set_gen_samplers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GenSamplers(void (*FN_PTR)(int, unsigned int*) value)

fn set_get_frag_data_index(&mut self, value: extern fn(c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetFragDataIndex(int (*FN_PTR)(unsigned int, const char*) value)

fn set_get_query_objecti64v(&mut self, value: extern fn(c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetQueryObjecti64v(void (*FN_PTR)(unsigned int, unsigned int, long*) value)

fn set_get_query_objectui64v(&mut self, value: extern fn(c_uint, c_uint, *mut c_ulong))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetQueryObjectui64v(void (*FN_PTR)(unsigned int, unsigned int, unsigned long*) value)

fn set_get_sampler_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetSamplerParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_sampler_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetSamplerParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) value)

fn set_get_sampler_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetSamplerParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_sampler_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_GetSamplerParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_is_sampler(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_IsSampler(unsigned char (*FN_PTR)(unsigned int) value)

fn set_multi_tex_coord_p1ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP1ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_multi_tex_coord_p1uiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP1uiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_multi_tex_coord_p2ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP2ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_multi_tex_coord_p2uiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP2uiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_multi_tex_coord_p3ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP3ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_multi_tex_coord_p3uiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP3uiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_multi_tex_coord_p4ui(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP4ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_multi_tex_coord_p4uiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_MultiTexCoordP4uiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_normal_p3ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_NormalP3ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_normal_p3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_NormalP3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_query_counter(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_QueryCounter(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_sampler_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_sampler_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_sampler_parameterf(&mut self, value: extern fn(c_uint, c_uint, c_float))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameterf(void (*FN_PTR)(unsigned int, unsigned int, float) value)

fn set_sampler_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *const c_float))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameterfv(void (*FN_PTR)(unsigned int, unsigned int, const float*) value)

fn set_sampler_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_sampler_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SamplerParameteriv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_secondary_color_p3ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SecondaryColorP3ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_secondary_color_p3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_SecondaryColorP3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_tex_coord_p1ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP1ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_tex_coord_p1uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP1uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_tex_coord_p2ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP2ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_tex_coord_p2uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP2uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_tex_coord_p3ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP3ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_tex_coord_p3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_tex_coord_p4ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP4ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_tex_coord_p4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_TexCoordP4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_attrib_divisor(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribDivisor(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_attrib_p1ui(&mut self, value: extern fn(c_uint, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP1ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_attrib_p1uiv(&mut self, value: extern fn(c_uint, c_uint, c_uchar, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP1uiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) value)

fn set_vertex_attrib_p2ui(&mut self, value: extern fn(c_uint, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP2ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_attrib_p2uiv(&mut self, value: extern fn(c_uint, c_uint, c_uchar, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP2uiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) value)

fn set_vertex_attrib_p3ui(&mut self, value: extern fn(c_uint, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP3ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_attrib_p3uiv(&mut self, value: extern fn(c_uint, c_uint, c_uchar, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP3uiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) value)

fn set_vertex_attrib_p4ui(&mut self, value: extern fn(c_uint, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP4ui(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_attrib_p4uiv(&mut self, value: extern fn(c_uint, c_uint, c_uchar, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexAttribP4uiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) value)

fn set_vertex_p2ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP2ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_p2uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP2uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_p3ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP3ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_p3uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP3uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn set_vertex_p4ui(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP4ui(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_p4uiv(&mut self, value: extern fn(c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_3_3_CoreBackend::set_VertexP4uiv(void (*FN_PTR)(unsigned int, const unsigned int*) value)

fn tex_coord_p1ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::TexCoordP1ui() const

fn tex_coord_p1uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::TexCoordP1uiv() const

fn tex_coord_p2ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::TexCoordP2ui() const

fn tex_coord_p2uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::TexCoordP2uiv() const

fn tex_coord_p3ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::TexCoordP3ui() const

fn tex_coord_p3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::TexCoordP3uiv() const

fn tex_coord_p4ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::TexCoordP4ui() const

fn tex_coord_p4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::TexCoordP4uiv() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions33CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_3_3_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_3_3_CoreBackend::versionStatus()

fn vertex_attrib_divisor(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexAttribDivisor() const

fn vertex_attrib_p1ui(&self) -> extern fn(c_uint, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP1ui() const

fn vertex_attrib_p1uiv(&self) -> extern fn(c_uint, c_uint, c_uchar, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP1uiv() const

fn vertex_attrib_p2ui(&self) -> extern fn(c_uint, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP2ui() const

fn vertex_attrib_p2uiv(&self) -> extern fn(c_uint, c_uint, c_uchar, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP2uiv() const

fn vertex_attrib_p3ui(&self) -> extern fn(c_uint, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP3ui() const

fn vertex_attrib_p3uiv(&self) -> extern fn(c_uint, c_uint, c_uchar, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP3uiv() const

fn vertex_attrib_p4ui(&self) -> extern fn(c_uint, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP4ui() const

fn vertex_attrib_p4uiv(&self) -> extern fn(c_uint, c_uint, c_uchar, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned char, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexAttribP4uiv() const

fn vertex_p2ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexP2ui() const

fn vertex_p2uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexP2uiv() const

fn vertex_p3ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexP3ui() const

fn vertex_p3uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexP3uiv() const

fn vertex_p4ui(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_3_3_CoreBackend::VertexP4ui() const

fn vertex_p4uiv(&self) -> extern fn(c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, const unsigned int*) QOpenGLFunctions_3_3_CoreBackend::VertexP4uiv() const