Struct qt_gui::open_g_l_version_functions::OpenGLFunctions40CoreBackend

pub struct OpenGLFunctions40CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_0_CoreBackend

Methods

impl OpenGLFunctions40CoreBackend

fn begin_query_indexed(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BeginQueryIndexed() const

fn bind_transform_feedback(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BindTransformFeedback() const

fn blend_equation_separatei(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BlendEquationSeparatei() const

fn blend_equationi(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BlendEquationi() const

fn blend_func_separatei(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BlendFuncSeparatei() const

fn blend_funci(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::BlendFunci() const

fn delete_transform_feedbacks(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_4_0_CoreBackend::DeleteTransformFeedbacks() const

fn draw_arrays_indirect(&self) -> extern fn(c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, const void*) QOpenGLFunctions_4_0_CoreBackend::DrawArraysIndirect() const

fn draw_elements_indirect(&self) -> extern fn(c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const void*) QOpenGLFunctions_4_0_CoreBackend::DrawElementsIndirect() const

fn draw_transform_feedback(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::DrawTransformFeedback() const

fn draw_transform_feedback_stream(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::DrawTransformFeedbackStream() const

fn end_query_indexed(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_0_CoreBackend::EndQueryIndexed() const

fn gen_transform_feedbacks(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_0_CoreBackend::GenTransformFeedbacks() const

fn get_active_subroutine_name(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) QOpenGLFunctions_4_0_CoreBackend::GetActiveSubroutineName() const

fn get_active_subroutine_uniform_name(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) QOpenGLFunctions_4_0_CoreBackend::GetActiveSubroutineUniformName() const

fn get_active_subroutine_uniformiv(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_0_CoreBackend::GetActiveSubroutineUniformiv() const

fn get_program_stageiv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_0_CoreBackend::GetProgramStageiv() const

fn get_query_indexediv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_0_CoreBackend::GetQueryIndexediv() const

fn get_subroutine_index(&self) -> extern fn(c_uint, c_uint, *const c_char) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_4_0_CoreBackend::GetSubroutineIndex() const

fn get_subroutine_uniform_location(&self) -> extern fn(c_uint, c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_4_0_CoreBackend::GetSubroutineUniformLocation() const

fn get_uniform_subroutineuiv(&self) -> extern fn(c_uint, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int*) QOpenGLFunctions_4_0_CoreBackend::GetUniformSubroutineuiv() const

fn get_uniformdv(&self) -> extern fn(c_uint, c_int, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, int, double*) QOpenGLFunctions_4_0_CoreBackend::GetUniformdv() const

fn is_transform_feedback(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_4_0_CoreBackend::IsTransformFeedback() const

fn min_sample_shading(&self) -> extern fn(c_float)

C++ method: void (*FN_PTR)(float) QOpenGLFunctions_4_0_CoreBackend::MinSampleShading() const

fn patch_parameterfv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_4_0_CoreBackend::PatchParameterfv() const

fn patch_parameteri(&self) -> extern fn(c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int) QOpenGLFunctions_4_0_CoreBackend::PatchParameteri() const

fn pause_transform_feedback(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_4_0_CoreBackend::PauseTransformFeedback() const

fn resume_transform_feedback(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_4_0_CoreBackend::ResumeTransformFeedback() const

fn set_begin_query_indexed(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BeginQueryIndexed(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_bind_transform_feedback(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BindTransformFeedback(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_blend_equation_separatei(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BlendEquationSeparatei(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_blend_equationi(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BlendEquationi(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_blend_func_separatei(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BlendFuncSeparatei(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_blend_funci(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_BlendFunci(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_delete_transform_feedbacks(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_DeleteTransformFeedbacks(void (*FN_PTR)(int, const unsigned int*) value)

fn set_draw_arrays_indirect(&mut self, value: extern fn(c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_DrawArraysIndirect(void (*FN_PTR)(unsigned int, const void*) value)

fn set_draw_elements_indirect(&mut self, value: extern fn(c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_DrawElementsIndirect(void (*FN_PTR)(unsigned int, unsigned int, const void*) value)

fn set_draw_transform_feedback(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_DrawTransformFeedback(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_draw_transform_feedback_stream(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_DrawTransformFeedbackStream(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_end_query_indexed(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_EndQueryIndexed(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_gen_transform_feedbacks(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GenTransformFeedbacks(void (*FN_PTR)(int, unsigned int*) value)

fn set_get_active_subroutine_name(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetActiveSubroutineName(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) value)

fn set_get_active_subroutine_uniform_name(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetActiveSubroutineUniformName(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) value)

fn set_get_active_subroutine_uniformiv(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetActiveSubroutineUniformiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_program_stageiv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetProgramStageiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_query_indexediv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetQueryIndexediv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_subroutine_index(&mut self, value: extern fn(c_uint, c_uint, *const c_char) -> c_uint)

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetSubroutineIndex(unsigned int (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_get_subroutine_uniform_location(&mut self, value: extern fn(c_uint, c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetSubroutineUniformLocation(int (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_get_uniform_subroutineuiv(&mut self, value: extern fn(c_uint, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetUniformSubroutineuiv(void (*FN_PTR)(unsigned int, int, unsigned int*) value)

fn set_get_uniformdv(&mut self, value: extern fn(c_uint, c_int, *mut c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_GetUniformdv(void (*FN_PTR)(unsigned int, int, double*) value)

fn set_is_transform_feedback(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_IsTransformFeedback(unsigned char (*FN_PTR)(unsigned int) value)

fn set_min_sample_shading(&mut self, value: extern fn(c_float))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_MinSampleShading(void (*FN_PTR)(float) value)

fn set_patch_parameterfv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_PatchParameterfv(void (*FN_PTR)(unsigned int, const float*) value)

fn set_patch_parameteri(&mut self, value: extern fn(c_uint, c_int))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_PatchParameteri(void (*FN_PTR)(unsigned int, int) value)

fn set_pause_transform_feedback(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_PauseTransformFeedback(void (*FN_PTR)() value)

fn set_resume_transform_feedback(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_ResumeTransformFeedback(void (*FN_PTR)() value)

fn set_uniform1d(&mut self, value: extern fn(c_int, c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform1d(void (*FN_PTR)(int, double) value)

fn set_uniform1dv(&mut self, value: extern fn(c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform1dv(void (*FN_PTR)(int, int, const double*) value)

fn set_uniform2d(&mut self, value: extern fn(c_int, c_double, c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform2d(void (*FN_PTR)(int, double, double) value)

fn set_uniform2dv(&mut self, value: extern fn(c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform2dv(void (*FN_PTR)(int, int, const double*) value)

fn set_uniform3d(&mut self, value: extern fn(c_int, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform3d(void (*FN_PTR)(int, double, double, double) value)

fn set_uniform3dv(&mut self, value: extern fn(c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform3dv(void (*FN_PTR)(int, int, const double*) value)

fn set_uniform4d(&mut self, value: extern fn(c_int, c_double, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform4d(void (*FN_PTR)(int, double, double, double, double) value)

fn set_uniform4dv(&mut self, value: extern fn(c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_Uniform4dv(void (*FN_PTR)(int, int, const double*) value)

fn set_uniform_matrix2dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix2dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix2x3dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix2x3dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix2x4dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix2x4dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix3dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix3dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix3x2dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix3x2dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix3x4dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix3x4dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix4dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix4dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix4x2dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix4x2dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_matrix4x3dv(&mut self, value: extern fn(c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformMatrix4x3dv(void (*FN_PTR)(int, int, unsigned char, const double*) value)

fn set_uniform_subroutinesuiv(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_0_CoreBackend::set_UniformSubroutinesuiv(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn uniform1d(&self) -> extern fn(c_int, c_double)

C++ method: void (*FN_PTR)(int, double) QOpenGLFunctions_4_0_CoreBackend::Uniform1d() const

fn uniform1dv(&self) -> extern fn(c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(int, int, const double*) QOpenGLFunctions_4_0_CoreBackend::Uniform1dv() const

fn uniform2d(&self) -> extern fn(c_int, c_double, c_double)

C++ method: void (*FN_PTR)(int, double, double) QOpenGLFunctions_4_0_CoreBackend::Uniform2d() const

fn uniform2dv(&self) -> extern fn(c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(int, int, const double*) QOpenGLFunctions_4_0_CoreBackend::Uniform2dv() const

fn uniform3d(&self) -> extern fn(c_int, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(int, double, double, double) QOpenGLFunctions_4_0_CoreBackend::Uniform3d() const

fn uniform3dv(&self) -> extern fn(c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(int, int, const double*) QOpenGLFunctions_4_0_CoreBackend::Uniform3dv() const

fn uniform4d(&self) -> extern fn(c_int, c_double, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(int, double, double, double, double) QOpenGLFunctions_4_0_CoreBackend::Uniform4d() const

fn uniform4dv(&self) -> extern fn(c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(int, int, const double*) QOpenGLFunctions_4_0_CoreBackend::Uniform4dv() const

fn uniform_matrix2dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix2dv() const

fn uniform_matrix2x3dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix2x3dv() const

fn uniform_matrix2x4dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix2x4dv() const

fn uniform_matrix3dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix3dv() const

fn uniform_matrix3x2dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix3x2dv() const

fn uniform_matrix3x4dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix3x4dv() const

fn uniform_matrix4dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix4dv() const

fn uniform_matrix4x2dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix4x2dv() const

fn uniform_matrix4x3dv(&self) -> extern fn(c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(int, int, unsigned char, const double*) QOpenGLFunctions_4_0_CoreBackend::UniformMatrix4x3dv() const

fn uniform_subroutinesuiv(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_0_CoreBackend::UniformSubroutinesuiv() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions40CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_0_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_0_CoreBackend::versionStatus()