Struct qt_gui::open_g_l_version_functions::OpenGLFunctions41CoreBackend

pub struct OpenGLFunctions41CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_1_CoreBackend

Methods

impl OpenGLFunctions41CoreBackend

fn active_shader_program(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::ActiveShaderProgram() const

fn bind_program_pipeline(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_1_CoreBackend::BindProgramPipeline() const

fn clear_depthf(&self) -> extern fn(c_float)

C++ method: void (*FN_PTR)(float) QOpenGLFunctions_4_1_CoreBackend::ClearDepthf() const

fn create_shader_programv(&self) -> extern fn(c_uint, c_int, *const *const c_char) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int, int, const char* const *) QOpenGLFunctions_4_1_CoreBackend::CreateShaderProgramv() const

fn delete_program_pipelines(&self) -> extern fn(c_int, *const c_uint)

C++ method: void (*FN_PTR)(int, const unsigned int*) QOpenGLFunctions_4_1_CoreBackend::DeleteProgramPipelines() const

fn depth_range_arrayv(&self) -> extern fn(c_uint, c_int, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, const double*) QOpenGLFunctions_4_1_CoreBackend::DepthRangeArrayv() const

fn depth_range_indexed(&self) -> extern fn(c_uint, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double) QOpenGLFunctions_4_1_CoreBackend::DepthRangeIndexed() const

fn depth_rangef(&self) -> extern fn(c_float, c_float)

C++ method: void (*FN_PTR)(float, float) QOpenGLFunctions_4_1_CoreBackend::DepthRangef() const

fn gen_program_pipelines(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_1_CoreBackend::GenProgramPipelines() const

fn get_doublei_v(&self) -> extern fn(c_uint, c_uint, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, double*) QOpenGLFunctions_4_1_CoreBackend::GetDoublei_v() const

fn get_floati_v(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_4_1_CoreBackend::GetFloati_v() const

fn get_program_binary(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_uint, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int*, unsigned int*, void*) QOpenGLFunctions_4_1_CoreBackend::GetProgramBinary() const

fn get_program_pipeline_info_log(&self) -> extern fn(c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, int, int*, char*) QOpenGLFunctions_4_1_CoreBackend::GetProgramPipelineInfoLog() const

fn get_program_pipelineiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_1_CoreBackend::GetProgramPipelineiv() const

fn get_shader_precision_format(&self) -> extern fn(c_uint, c_uint, *mut c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*, int*) QOpenGLFunctions_4_1_CoreBackend::GetShaderPrecisionFormat() const

fn get_vertex_attrib_ldv(&self) -> extern fn(c_uint, c_uint, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, double*) QOpenGLFunctions_4_1_CoreBackend::GetVertexAttribLdv() const

fn is_program_pipeline(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_4_1_CoreBackend::IsProgramPipeline() const

fn program_binary(&self) -> extern fn(c_uint, c_uint, *const c_void, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const void*, int) QOpenGLFunctions_4_1_CoreBackend::ProgramBinary() const

fn program_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_4_1_CoreBackend::ProgramParameteri() const

fn program_uniform1d(&self) -> extern fn(c_uint, c_int, c_double)

C++ method: void (*FN_PTR)(unsigned int, int, double) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1d() const

fn program_uniform1dv(&self) -> extern fn(c_uint, c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1dv() const

fn program_uniform1f(&self) -> extern fn(c_uint, c_int, c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1f() const

fn program_uniform1fv(&self) -> extern fn(c_uint, c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1fv() const

fn program_uniform1i(&self) -> extern fn(c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1i() const

fn program_uniform1iv(&self) -> extern fn(c_uint, c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, const int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1iv() const

fn program_uniform1ui(&self) -> extern fn(c_uint, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1ui() const

fn program_uniform1uiv(&self) -> extern fn(c_uint, c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, const unsigned int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform1uiv() const

fn program_uniform2d(&self) -> extern fn(c_uint, c_int, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, int, double, double) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2d() const

fn program_uniform2dv(&self) -> extern fn(c_uint, c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2dv() const

fn program_uniform2f(&self) -> extern fn(c_uint, c_int, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float, float) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2f() const

fn program_uniform2fv(&self) -> extern fn(c_uint, c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2fv() const

fn program_uniform2i(&self) -> extern fn(c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2i() const

fn program_uniform2iv(&self) -> extern fn(c_uint, c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, const int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2iv() const

fn program_uniform2ui(&self) -> extern fn(c_uint, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2ui() const

fn program_uniform2uiv(&self) -> extern fn(c_uint, c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, const unsigned int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform2uiv() const

fn program_uniform3d(&self) -> extern fn(c_uint, c_int, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, int, double, double, double) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3d() const

fn program_uniform3dv(&self) -> extern fn(c_uint, c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3dv() const

fn program_uniform3f(&self) -> extern fn(c_uint, c_int, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float, float, float) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3f() const

fn program_uniform3fv(&self) -> extern fn(c_uint, c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3fv() const

fn program_uniform3i(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3i() const

fn program_uniform3iv(&self) -> extern fn(c_uint, c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, const int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3iv() const

fn program_uniform3ui(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3ui() const

fn program_uniform3uiv(&self) -> extern fn(c_uint, c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, const unsigned int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform3uiv() const

fn program_uniform4d(&self) -> extern fn(c_uint, c_int, c_double, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, int, double, double, double, double) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4d() const

fn program_uniform4dv(&self) -> extern fn(c_uint, c_int, c_int, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4dv() const

fn program_uniform4f(&self) -> extern fn(c_uint, c_int, c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float, float, float, float) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4f() const

fn program_uniform4fv(&self) -> extern fn(c_uint, c_int, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4fv() const

fn program_uniform4i(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4i() const

fn program_uniform4iv(&self) -> extern fn(c_uint, c_int, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, const int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4iv() const

fn program_uniform4ui(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4ui() const

fn program_uniform4uiv(&self) -> extern fn(c_uint, c_int, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, const unsigned int*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniform4uiv() const

fn program_uniform_matrix2dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2dv() const

fn program_uniform_matrix2fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2fv() const

fn program_uniform_matrix2x3dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2x3dv() const

fn program_uniform_matrix2x3fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2x3fv() const

fn program_uniform_matrix2x4dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2x4dv() const

fn program_uniform_matrix2x4fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix2x4fv() const

fn program_uniform_matrix3dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3dv() const

fn program_uniform_matrix3fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3fv() const

fn program_uniform_matrix3x2dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3x2dv() const

fn program_uniform_matrix3x2fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3x2fv() const

fn program_uniform_matrix3x4dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3x4dv() const

fn program_uniform_matrix3x4fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix3x4fv() const

fn program_uniform_matrix4dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4dv() const

fn program_uniform_matrix4fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4fv() const

fn program_uniform_matrix4x2dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4x2dv() const

fn program_uniform_matrix4x2fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4x2fv() const

fn program_uniform_matrix4x3dv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4x3dv() const

fn program_uniform_matrix4x3fv(&self) -> extern fn(c_uint, c_int, c_int, c_uchar, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) QOpenGLFunctions_4_1_CoreBackend::ProgramUniformMatrix4x3fv() const

fn release_shader_compiler(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_4_1_CoreBackend::ReleaseShaderCompiler() const

fn scissor_arrayv(&self) -> extern fn(c_uint, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const int*) QOpenGLFunctions_4_1_CoreBackend::ScissorArrayv() const

fn scissor_indexed(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int) QOpenGLFunctions_4_1_CoreBackend::ScissorIndexed() const

fn scissor_indexedv(&self) -> extern fn(c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, const int*) QOpenGLFunctions_4_1_CoreBackend::ScissorIndexedv() const

fn set_active_shader_program(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ActiveShaderProgram(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_bind_program_pipeline(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_BindProgramPipeline(void (*FN_PTR)(unsigned int) value)

fn set_clear_depthf(&mut self, value: extern fn(c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ClearDepthf(void (*FN_PTR)(float) value)

fn set_create_shader_programv(&mut self, value: extern fn(c_uint, c_int, *const *const c_char) -> c_uint)

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_CreateShaderProgramv(unsigned int (*FN_PTR)(unsigned int, int, const char* const *) value)

fn set_delete_program_pipelines(&mut self, value: extern fn(c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_DeleteProgramPipelines(void (*FN_PTR)(int, const unsigned int*) value)

fn set_depth_range_arrayv(&mut self, value: extern fn(c_uint, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_DepthRangeArrayv(void (*FN_PTR)(unsigned int, int, const double*) value)

fn set_depth_range_indexed(&mut self, value: extern fn(c_uint, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_DepthRangeIndexed(void (*FN_PTR)(unsigned int, double, double) value)

fn set_depth_rangef(&mut self, value: extern fn(c_float, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_DepthRangef(void (*FN_PTR)(float, float) value)

fn set_gen_program_pipelines(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GenProgramPipelines(void (*FN_PTR)(int, unsigned int*) value)

fn set_get_doublei_v(&mut self, value: extern fn(c_uint, c_uint, *mut c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetDoublei_v(void (*FN_PTR)(unsigned int, unsigned int, double*) value)

fn set_get_floati_v(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetFloati_v(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_program_binary(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_uint, *mut c_void))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetProgramBinary(void (*FN_PTR)(unsigned int, int, int*, unsigned int*, void*) value)

fn set_get_program_pipeline_info_log(&mut self, value: extern fn(c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetProgramPipelineInfoLog(void (*FN_PTR)(unsigned int, int, int*, char*) value)

fn set_get_program_pipelineiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetProgramPipelineiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_shader_precision_format(&mut self, value: extern fn(c_uint, c_uint, *mut c_int, *mut c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetShaderPrecisionFormat(void (*FN_PTR)(unsigned int, unsigned int, int*, int*) value)

fn set_get_vertex_attrib_ldv(&mut self, value: extern fn(c_uint, c_uint, *mut c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_GetVertexAttribLdv(void (*FN_PTR)(unsigned int, unsigned int, double*) value)

fn set_is_program_pipeline(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_IsProgramPipeline(unsigned char (*FN_PTR)(unsigned int) value)

fn set_program_binary(&mut self, value: extern fn(c_uint, c_uint, *const c_void, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramBinary(void (*FN_PTR)(unsigned int, unsigned int, const void*, int) value)

fn set_program_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_program_uniform1d(&mut self, value: extern fn(c_uint, c_int, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1d(void (*FN_PTR)(unsigned int, int, double) value)

fn set_program_uniform1dv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1dv(void (*FN_PTR)(unsigned int, int, int, const double*) value)

fn set_program_uniform1f(&mut self, value: extern fn(c_uint, c_int, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1f(void (*FN_PTR)(unsigned int, int, float) value)

fn set_program_uniform1fv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1fv(void (*FN_PTR)(unsigned int, int, int, const float*) value)

fn set_program_uniform1i(&mut self, value: extern fn(c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1i(void (*FN_PTR)(unsigned int, int, int) value)

fn set_program_uniform1iv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1iv(void (*FN_PTR)(unsigned int, int, int, const int*) value)

fn set_program_uniform1ui(&mut self, value: extern fn(c_uint, c_int, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1ui(void (*FN_PTR)(unsigned int, int, unsigned int) value)

fn set_program_uniform1uiv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform1uiv(void (*FN_PTR)(unsigned int, int, int, const unsigned int*) value)

fn set_program_uniform2d(&mut self, value: extern fn(c_uint, c_int, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2d(void (*FN_PTR)(unsigned int, int, double, double) value)

fn set_program_uniform2dv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2dv(void (*FN_PTR)(unsigned int, int, int, const double*) value)

fn set_program_uniform2f(&mut self, value: extern fn(c_uint, c_int, c_float, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2f(void (*FN_PTR)(unsigned int, int, float, float) value)

fn set_program_uniform2fv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2fv(void (*FN_PTR)(unsigned int, int, int, const float*) value)

fn set_program_uniform2i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2i(void (*FN_PTR)(unsigned int, int, int, int) value)

fn set_program_uniform2iv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2iv(void (*FN_PTR)(unsigned int, int, int, const int*) value)

fn set_program_uniform2ui(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2ui(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) value)

fn set_program_uniform2uiv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform2uiv(void (*FN_PTR)(unsigned int, int, int, const unsigned int*) value)

fn set_program_uniform3d(&mut self, value: extern fn(c_uint, c_int, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3d(void (*FN_PTR)(unsigned int, int, double, double, double) value)

fn set_program_uniform3dv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3dv(void (*FN_PTR)(unsigned int, int, int, const double*) value)

fn set_program_uniform3f(&mut self, value: extern fn(c_uint, c_int, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3f(void (*FN_PTR)(unsigned int, int, float, float, float) value)

fn set_program_uniform3fv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3fv(void (*FN_PTR)(unsigned int, int, int, const float*) value)

fn set_program_uniform3i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3i(void (*FN_PTR)(unsigned int, int, int, int, int) value)

fn set_program_uniform3iv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3iv(void (*FN_PTR)(unsigned int, int, int, const int*) value)

fn set_program_uniform3ui(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3ui(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, unsigned int) value)

fn set_program_uniform3uiv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform3uiv(void (*FN_PTR)(unsigned int, int, int, const unsigned int*) value)

fn set_program_uniform4d(&mut self, value: extern fn(c_uint, c_int, c_double, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4d(void (*FN_PTR)(unsigned int, int, double, double, double, double) value)

fn set_program_uniform4dv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4dv(void (*FN_PTR)(unsigned int, int, int, const double*) value)

fn set_program_uniform4f(&mut self, value: extern fn(c_uint, c_int, c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4f(void (*FN_PTR)(unsigned int, int, float, float, float, float) value)

fn set_program_uniform4fv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4fv(void (*FN_PTR)(unsigned int, int, int, const float*) value)

fn set_program_uniform4i(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4i(void (*FN_PTR)(unsigned int, int, int, int, int, int) value)

fn set_program_uniform4iv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4iv(void (*FN_PTR)(unsigned int, int, int, const int*) value)

fn set_program_uniform4ui(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4ui(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_program_uniform4uiv(&mut self, value: extern fn(c_uint, c_int, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniform4uiv(void (*FN_PTR)(unsigned int, int, int, const unsigned int*) value)

fn set_program_uniform_matrix2dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix2fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix2x3dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2x3dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix2x3fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2x3fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix2x4dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2x4dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix2x4fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix2x4fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix3dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix3fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix3x2dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3x2dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix3x2fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3x2fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix3x4dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3x4dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix3x4fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix3x4fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix4dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix4fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix4x2dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4x2dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix4x2fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4x2fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_program_uniform_matrix4x3dv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4x3dv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const double*) value)

fn set_program_uniform_matrix4x3fv(&mut self, value: extern fn(c_uint, c_int, c_int, c_uchar, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ProgramUniformMatrix4x3fv(void (*FN_PTR)(unsigned int, int, int, unsigned char, const float*) value)

fn set_release_shader_compiler(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ReleaseShaderCompiler(void (*FN_PTR)() value)

fn set_scissor_arrayv(&mut self, value: extern fn(c_uint, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ScissorArrayv(void (*FN_PTR)(unsigned int, int, const int*) value)

fn set_scissor_indexed(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ScissorIndexed(void (*FN_PTR)(unsigned int, int, int, int, int) value)

fn set_scissor_indexedv(&mut self, value: extern fn(c_uint, *const c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ScissorIndexedv(void (*FN_PTR)(unsigned int, const int*) value)

fn set_shader_binary(&mut self, value: extern fn(c_int, *const c_uint, c_uint, *const c_void, c_int))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ShaderBinary(void (*FN_PTR)(int, const unsigned int*, unsigned int, const void*, int) value)

fn set_use_program_stages(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_UseProgramStages(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_validate_program_pipeline(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ValidateProgramPipeline(void (*FN_PTR)(unsigned int) value)

fn set_vertex_attrib_l1d(&mut self, value: extern fn(c_uint, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL1d(void (*FN_PTR)(unsigned int, double) value)

fn set_vertex_attrib_l1dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL1dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib_l2d(&mut self, value: extern fn(c_uint, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL2d(void (*FN_PTR)(unsigned int, double, double) value)

fn set_vertex_attrib_l2dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL2dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib_l3d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL3d(void (*FN_PTR)(unsigned int, double, double, double) value)

fn set_vertex_attrib_l3dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL3dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib_l4d(&mut self, value: extern fn(c_uint, c_double, c_double, c_double, c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL4d(void (*FN_PTR)(unsigned int, double, double, double, double) value)

fn set_vertex_attrib_l4dv(&mut self, value: extern fn(c_uint, *const c_double))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribL4dv(void (*FN_PTR)(unsigned int, const double*) value)

fn set_vertex_attrib_l_pointer(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_VertexAttribLPointer(void (*FN_PTR)(unsigned int, int, unsigned int, int, const void*) value)

fn set_viewport_arrayv(&mut self, value: extern fn(c_uint, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ViewportArrayv(void (*FN_PTR)(unsigned int, int, const float*) value)

fn set_viewport_indexedf(&mut self, value: extern fn(c_uint, c_float, c_float, c_float, c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ViewportIndexedf(void (*FN_PTR)(unsigned int, float, float, float, float) value)

fn set_viewport_indexedfv(&mut self, value: extern fn(c_uint, *const c_float))

C++ method: void QOpenGLFunctions_4_1_CoreBackend::set_ViewportIndexedfv(void (*FN_PTR)(unsigned int, const float*) value)

fn shader_binary(&self) -> extern fn(c_int, *const c_uint, c_uint, *const c_void, c_int)

C++ method: void (*FN_PTR)(int, const unsigned int*, unsigned int, const void*, int) QOpenGLFunctions_4_1_CoreBackend::ShaderBinary() const

fn use_program_stages(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_1_CoreBackend::UseProgramStages() const

fn validate_program_pipeline(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_1_CoreBackend::ValidateProgramPipeline() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions41CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_1_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_1_CoreBackend::versionStatus()

fn vertex_attrib_l1d(&self) -> extern fn(c_uint, c_double)

C++ method: void (*FN_PTR)(unsigned int, double) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL1d() const

fn vertex_attrib_l1dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL1dv() const

fn vertex_attrib_l2d(&self) -> extern fn(c_uint, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL2d() const

fn vertex_attrib_l2dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL2dv() const

fn vertex_attrib_l3d(&self) -> extern fn(c_uint, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL3d() const

fn vertex_attrib_l3dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL3dv() const

fn vertex_attrib_l4d(&self) -> extern fn(c_uint, c_double, c_double, c_double, c_double)

C++ method: void (*FN_PTR)(unsigned int, double, double, double, double) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL4d() const

fn vertex_attrib_l4dv(&self) -> extern fn(c_uint, *const c_double)

C++ method: void (*FN_PTR)(unsigned int, const double*) QOpenGLFunctions_4_1_CoreBackend::VertexAttribL4dv() const

fn vertex_attrib_l_pointer(&self) -> extern fn(c_uint, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, const void*) QOpenGLFunctions_4_1_CoreBackend::VertexAttribLPointer() const

fn viewport_arrayv(&self) -> extern fn(c_uint, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, int, const float*) QOpenGLFunctions_4_1_CoreBackend::ViewportArrayv() const

fn viewport_indexedf(&self) -> extern fn(c_uint, c_float, c_float, c_float, c_float)

C++ method: void (*FN_PTR)(unsigned int, float, float, float, float) QOpenGLFunctions_4_1_CoreBackend::ViewportIndexedf() const

fn viewport_indexedfv(&self) -> extern fn(c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, const float*) QOpenGLFunctions_4_1_CoreBackend::ViewportIndexedfv() const