Struct qt_gui::open_g_l_version_functions::OpenGLFunctions42CoreBackend

pub struct OpenGLFunctions42CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_2_CoreBackend

Methods

impl OpenGLFunctions42CoreBackend

fn bind_image_texture(&self) -> extern fn(c_uint, c_uint, c_int, c_uchar, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned char, int, unsigned int, unsigned int) QOpenGLFunctions_4_2_CoreBackend::BindImageTexture() const

fn draw_arrays_instanced_base_instance(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, unsigned int) QOpenGLFunctions_4_2_CoreBackend::DrawArraysInstancedBaseInstance() const

fn draw_elements_instanced_base_instance(&self) -> extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, unsigned int) QOpenGLFunctions_4_2_CoreBackend::DrawElementsInstancedBaseInstance() const

fn draw_elements_instanced_base_vertex_base_instance(&self) -> extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_int, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, int, unsigned int) QOpenGLFunctions_4_2_CoreBackend::DrawElementsInstancedBaseVertexBaseInstance() const

fn draw_transform_feedback_instanced(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_4_2_CoreBackend::DrawTransformFeedbackInstanced() const

fn draw_transform_feedback_stream_instanced(&self) -> extern fn(c_uint, c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) QOpenGLFunctions_4_2_CoreBackend::DrawTransformFeedbackStreamInstanced() const

fn get_active_atomic_counter_bufferiv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_2_CoreBackend::GetActiveAtomicCounterBufferiv() const

fn get_internalformativ(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*) QOpenGLFunctions_4_2_CoreBackend::GetInternalformativ() const

fn memory_barrier(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_2_CoreBackend::MemoryBarrier() const

fn set_bind_image_texture(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uchar, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_BindImageTexture(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned char, int, unsigned int, unsigned int) value)

fn set_draw_arrays_instanced_base_instance(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_uint))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_DrawArraysInstancedBaseInstance(void (*FN_PTR)(unsigned int, int, int, int, unsigned int) value)

fn set_draw_elements_instanced_base_instance(&mut self, value: extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_uint))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_DrawElementsInstancedBaseInstance(void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, unsigned int) value)

fn set_draw_elements_instanced_base_vertex_base_instance(&mut self, value: extern fn(c_uint, c_int, c_uint, *const c_void, c_int, c_int, c_uint))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_DrawElementsInstancedBaseVertexBaseInstance(void (*FN_PTR)(unsigned int, int, unsigned int, const void*, int, int, unsigned int) value)

fn set_draw_transform_feedback_instanced(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_DrawTransformFeedbackInstanced(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_draw_transform_feedback_stream_instanced(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_DrawTransformFeedbackStreamInstanced(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) value)

fn set_get_active_atomic_counter_bufferiv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_GetActiveAtomicCounterBufferiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_internalformativ(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_GetInternalformativ(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*) value)

fn set_memory_barrier(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_MemoryBarrier(void (*FN_PTR)(unsigned int) value)

fn set_tex_storage1_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_TexStorage1D(void (*FN_PTR)(unsigned int, int, unsigned int, int) value)

fn set_tex_storage2_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_TexStorage2D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int) value)

fn set_tex_storage3_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_2_CoreBackend::set_TexStorage3D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int) value)

fn tex_storage1_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int) QOpenGLFunctions_4_2_CoreBackend::TexStorage1D() const

fn tex_storage2_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int) QOpenGLFunctions_4_2_CoreBackend::TexStorage2D() const

fn tex_storage3_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int) QOpenGLFunctions_4_2_CoreBackend::TexStorage3D() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions42CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_2_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_2_CoreBackend::versionStatus()