Struct qt_gui::open_g_l_version_functions::OpenGLFunctions43CoreBackend

pub struct OpenGLFunctions43CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_3_CoreBackend

Methods

impl OpenGLFunctions43CoreBackend

fn bind_vertex_buffer(&self) -> extern fn(c_uint, c_uint, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long, int) QOpenGLFunctions_4_3_CoreBackend::BindVertexBuffer() const

fn clear_buffer_data(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_3_CoreBackend::ClearBufferData() const

fn clear_buffer_sub_data(&self) -> extern fn(c_uint, c_uint, c_long, c_long, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long, long, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_3_CoreBackend::ClearBufferSubData() const

fn copy_image_sub_data(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int, unsigned int, unsigned int, int, int, int, int, int, int, int) QOpenGLFunctions_4_3_CoreBackend::CopyImageSubData() const

fn debug_message_control(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *const c_uint, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, const unsigned int*, unsigned char) QOpenGLFunctions_4_3_CoreBackend::DebugMessageControl() const

fn debug_message_insert(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_int, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int, const char*) QOpenGLFunctions_4_3_CoreBackend::DebugMessageInsert() const

fn dispatch_compute(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::DispatchCompute() const

fn dispatch_compute_indirect(&self) -> extern fn(c_long)

C++ method: void (*FN_PTR)(long) QOpenGLFunctions_4_3_CoreBackend::DispatchComputeIndirect() const

fn framebuffer_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_4_3_CoreBackend::FramebufferParameteri() const

fn get_debug_message_log(&self) -> extern fn(c_uint, c_int, *mut c_uint, *mut c_uint, *mut c_uint, *mut c_uint, *mut c_int, *mut c_char) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int, int, unsigned int*, unsigned int*, unsigned int*, unsigned int*, int*, char*) QOpenGLFunctions_4_3_CoreBackend::GetDebugMessageLog() const

fn get_framebuffer_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_3_CoreBackend::GetFramebufferParameteriv() const

fn get_internalformati64v(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, long*) QOpenGLFunctions_4_3_CoreBackend::GetInternalformati64v() const

fn get_object_label(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int*, char*) QOpenGLFunctions_4_3_CoreBackend::GetObjectLabel() const

fn get_object_ptr_label(&self) -> extern fn(*const c_void, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(const void*, int, int*, char*) QOpenGLFunctions_4_3_CoreBackend::GetObjectPtrLabel() const

fn get_program_interfaceiv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_3_CoreBackend::GetProgramInterfaceiv() const

fn get_program_resource_index(&self) -> extern fn(c_uint, c_uint, *const c_char) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_4_3_CoreBackend::GetProgramResourceIndex() const

fn get_program_resource_location(&self) -> extern fn(c_uint, c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_4_3_CoreBackend::GetProgramResourceLocation() const

fn get_program_resource_location_index(&self) -> extern fn(c_uint, c_uint, *const c_char) -> c_int

C++ method: int (*FN_PTR)(unsigned int, unsigned int, const char*) QOpenGLFunctions_4_3_CoreBackend::GetProgramResourceLocationIndex() const

fn get_program_resource_name(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) QOpenGLFunctions_4_3_CoreBackend::GetProgramResourceName() const

fn get_program_resourceiv(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *const c_uint, c_int, *mut c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, const unsigned int*, int, int*, int*) QOpenGLFunctions_4_3_CoreBackend::GetProgramResourceiv() const

fn invalidate_buffer_data(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_3_CoreBackend::InvalidateBufferData() const

fn invalidate_buffer_sub_data(&self) -> extern fn(c_uint, c_long, c_long)

C++ method: void (*FN_PTR)(unsigned int, long, long) QOpenGLFunctions_4_3_CoreBackend::InvalidateBufferSubData() const

fn invalidate_framebuffer(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_3_CoreBackend::InvalidateFramebuffer() const

fn invalidate_sub_framebuffer(&self) -> extern fn(c_uint, c_int, *const c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*, int, int, int, int) QOpenGLFunctions_4_3_CoreBackend::InvalidateSubFramebuffer() const

fn invalidate_tex_image(&self) -> extern fn(c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int) QOpenGLFunctions_4_3_CoreBackend::InvalidateTexImage() const

fn invalidate_tex_sub_image(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) QOpenGLFunctions_4_3_CoreBackend::InvalidateTexSubImage() const

fn multi_draw_arrays_indirect(&self) -> extern fn(c_uint, *const c_void, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, const void*, int, int) QOpenGLFunctions_4_3_CoreBackend::MultiDrawArraysIndirect() const

fn multi_draw_elements_indirect(&self) -> extern fn(c_uint, c_uint, *const c_void, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const void*, int, int) QOpenGLFunctions_4_3_CoreBackend::MultiDrawElementsIndirect() const

fn object_label(&self) -> extern fn(c_uint, c_uint, c_int, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const char*) QOpenGLFunctions_4_3_CoreBackend::ObjectLabel() const

fn object_ptr_label(&self) -> extern fn(*const c_void, c_int, *const c_char)

C++ method: void (*FN_PTR)(const void*, int, const char*) QOpenGLFunctions_4_3_CoreBackend::ObjectPtrLabel() const

fn pop_debug_group(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_4_3_CoreBackend::PopDebugGroup() const

fn push_debug_group(&self) -> extern fn(c_uint, c_uint, c_int, *const c_char)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const char*) QOpenGLFunctions_4_3_CoreBackend::PushDebugGroup() const

fn set_bind_vertex_buffer(&mut self, value: extern fn(c_uint, c_uint, c_long, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_BindVertexBuffer(void (*FN_PTR)(unsigned int, unsigned int, long, int) value)

fn set_clear_buffer_data(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_ClearBufferData(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, const void*) value)

fn set_clear_buffer_sub_data(&mut self, value: extern fn(c_uint, c_uint, c_long, c_long, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_ClearBufferSubData(void (*FN_PTR)(unsigned int, unsigned int, long, long, unsigned int, unsigned int, const void*) value)

fn set_copy_image_sub_data(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_CopyImageSubData(void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int, unsigned int, unsigned int, int, int, int, int, int, int, int) value)

fn set_debug_message_control(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *const c_uint, c_uchar))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_DebugMessageControl(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, const unsigned int*, unsigned char) value)

fn set_debug_message_insert(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_int, *const c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_DebugMessageInsert(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, int, const char*) value)

fn set_dispatch_compute(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_DispatchCompute(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_dispatch_compute_indirect(&mut self, value: extern fn(c_long))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_DispatchComputeIndirect(void (*FN_PTR)(long) value)

fn set_framebuffer_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_FramebufferParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_get_debug_message_log(&mut self, value: extern fn(c_uint, c_int, *mut c_uint, *mut c_uint, *mut c_uint, *mut c_uint, *mut c_int, *mut c_char) -> c_uint)

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetDebugMessageLog(unsigned int (*FN_PTR)(unsigned int, int, unsigned int*, unsigned int*, unsigned int*, unsigned int*, int*, char*) value)

fn set_get_framebuffer_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetFramebufferParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_internalformati64v(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_long))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetInternalformati64v(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, long*) value)

fn set_get_object_label(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetObjectLabel(void (*FN_PTR)(unsigned int, unsigned int, int, int*, char*) value)

fn set_get_object_ptr_label(&mut self, value: extern fn(*const c_void, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetObjectPtrLabel(void (*FN_PTR)(const void*, int, int*, char*) value)

fn set_get_program_interfaceiv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramInterfaceiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_program_resource_index(&mut self, value: extern fn(c_uint, c_uint, *const c_char) -> c_uint)

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramResourceIndex(unsigned int (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_get_program_resource_location(&mut self, value: extern fn(c_uint, c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramResourceLocation(int (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_get_program_resource_location_index(&mut self, value: extern fn(c_uint, c_uint, *const c_char) -> c_int)

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramResourceLocationIndex(int (*FN_PTR)(unsigned int, unsigned int, const char*) value)

fn set_get_program_resource_name(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_int, *mut c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramResourceName(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int*, char*) value)

fn set_get_program_resourceiv(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *const c_uint, c_int, *mut c_int, *mut c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_GetProgramResourceiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, const unsigned int*, int, int*, int*) value)

fn set_invalidate_buffer_data(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateBufferData(void (*FN_PTR)(unsigned int) value)

fn set_invalidate_buffer_sub_data(&mut self, value: extern fn(c_uint, c_long, c_long))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateBufferSubData(void (*FN_PTR)(unsigned int, long, long) value)

fn set_invalidate_framebuffer(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateFramebuffer(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_invalidate_sub_framebuffer(&mut self, value: extern fn(c_uint, c_int, *const c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateSubFramebuffer(void (*FN_PTR)(unsigned int, int, const unsigned int*, int, int, int, int) value)

fn set_invalidate_tex_image(&mut self, value: extern fn(c_uint, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateTexImage(void (*FN_PTR)(unsigned int, int) value)

fn set_invalidate_tex_sub_image(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_InvalidateTexSubImage(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) value)

fn set_multi_draw_arrays_indirect(&mut self, value: extern fn(c_uint, *const c_void, c_int, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_MultiDrawArraysIndirect(void (*FN_PTR)(unsigned int, const void*, int, int) value)

fn set_multi_draw_elements_indirect(&mut self, value: extern fn(c_uint, c_uint, *const c_void, c_int, c_int))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_MultiDrawElementsIndirect(void (*FN_PTR)(unsigned int, unsigned int, const void*, int, int) value)

fn set_object_label(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_ObjectLabel(void (*FN_PTR)(unsigned int, unsigned int, int, const char*) value)

fn set_object_ptr_label(&mut self, value: extern fn(*const c_void, c_int, *const c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_ObjectPtrLabel(void (*FN_PTR)(const void*, int, const char*) value)

fn set_pop_debug_group(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_PopDebugGroup(void (*FN_PTR)() value)

fn set_push_debug_group(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_char))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_PushDebugGroup(void (*FN_PTR)(unsigned int, unsigned int, int, const char*) value)

fn set_shader_storage_block_binding(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_ShaderStorageBlockBinding(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_tex_buffer_range(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_long, c_long))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_TexBufferRange(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, long) value)

fn set_tex_storage2_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_TexStorage2DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) value)

fn set_tex_storage3_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_TexStorage3DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) value)

fn set_texture_view(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_TextureView(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_vertex_attrib_binding(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_VertexAttribBinding(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_attrib_format(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_VertexAttribFormat(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_attrib_i_format(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_VertexAttribIFormat(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) value)

fn set_vertex_attrib_l_format(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_VertexAttribLFormat(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) value)

fn set_vertex_binding_divisor(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_3_CoreBackend::set_VertexBindingDivisor(void (*FN_PTR)(unsigned int, unsigned int) value)

fn shader_storage_block_binding(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::ShaderStorageBlockBinding() const

fn tex_buffer_range(&self) -> extern fn(c_uint, c_uint, c_uint, c_long, c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, long) QOpenGLFunctions_4_3_CoreBackend::TexBufferRange() const

fn tex_storage2_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) QOpenGLFunctions_4_3_CoreBackend::TexStorage2DMultisample() const

fn tex_storage3_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) QOpenGLFunctions_4_3_CoreBackend::TexStorage3DMultisample() const

fn texture_view(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::TextureView() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions43CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_3_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_3_CoreBackend::versionStatus()

fn vertex_attrib_binding(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::VertexAttribBinding() const

fn vertex_attrib_format(&self) -> extern fn(c_uint, c_int, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_4_3_CoreBackend::VertexAttribFormat() const

fn vertex_attrib_i_format(&self) -> extern fn(c_uint, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::VertexAttribIFormat() const

fn vertex_attrib_l_format(&self) -> extern fn(c_uint, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::VertexAttribLFormat() const

fn vertex_binding_divisor(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_3_CoreBackend::VertexBindingDivisor() const