Struct qt_gui::open_g_l_version_functions::OpenGLFunctions44CoreBackend

pub struct OpenGLFunctions44CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_4_CoreBackend

Methods

impl OpenGLFunctions44CoreBackend

fn bind_buffers_base(&self) -> extern fn(c_uint, c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*) QOpenGLFunctions_4_4_CoreBackend::BindBuffersBase() const

fn bind_buffers_range(&self) -> extern fn(c_uint, c_uint, c_int, *const c_uint, *const c_long, *const c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*, const long*, const long*) QOpenGLFunctions_4_4_CoreBackend::BindBuffersRange() const

fn bind_image_textures(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_4_CoreBackend::BindImageTextures() const

fn bind_samplers(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_4_CoreBackend::BindSamplers() const

fn bind_textures(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_4_CoreBackend::BindTextures() const

fn bind_vertex_buffers(&self) -> extern fn(c_uint, c_int, *const c_uint, *const c_long, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*, const long*, const int*) QOpenGLFunctions_4_4_CoreBackend::BindVertexBuffers() const

fn buffer_storage(&self) -> extern fn(c_uint, c_long, *const c_void, c_uint)

C++ method: void (*FN_PTR)(unsigned int, long, const void*, unsigned int) QOpenGLFunctions_4_4_CoreBackend::BufferStorage() const

fn clear_tex_image(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_4_CoreBackend::ClearTexImage() const

fn clear_tex_sub_image(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_4_CoreBackend::ClearTexSubImage() const

fn set_bind_buffers_base(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindBuffersBase(void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*) value)

fn set_bind_buffers_range(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_uint, *const c_long, *const c_long))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindBuffersRange(void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*, const long*, const long*) value)

fn set_bind_image_textures(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindImageTextures(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_bind_samplers(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindSamplers(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_bind_textures(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindTextures(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_bind_vertex_buffers(&mut self, value: extern fn(c_uint, c_int, *const c_uint, *const c_long, *const c_int))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BindVertexBuffers(void (*FN_PTR)(unsigned int, int, const unsigned int*, const long*, const int*) value)

fn set_buffer_storage(&mut self, value: extern fn(c_uint, c_long, *const c_void, c_uint))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_BufferStorage(void (*FN_PTR)(unsigned int, long, const void*, unsigned int) value)

fn set_clear_tex_image(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_ClearTexImage(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, const void*) value)

fn set_clear_tex_sub_image(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_4_CoreBackend::set_ClearTexSubImage(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions44CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_4_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_4_CoreBackend::versionStatus()