Struct qt_gui::open_g_l_version_functions::OpenGLFunctions45CoreBackend

pub struct OpenGLFunctions45CoreBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_5_CoreBackend

Methods

impl OpenGLFunctions45CoreBackend

fn bind_texture_unit(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::BindTextureUnit() const

fn blit_named_framebuffer(&self) -> extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int, int, int, int, int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::BlitNamedFramebuffer() const

fn check_named_framebuffer_status(&self) -> extern fn(c_uint, c_uint) -> c_uint

C++ method: unsigned int (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::CheckNamedFramebufferStatus() const

fn clear_named_buffer_data(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_5_CoreBackend::ClearNamedBufferData() const

fn clear_named_buffer_sub_data(&self) -> extern fn(c_uint, c_uint, c_long, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_5_CoreBackend::ClearNamedBufferSubData() const

fn clear_named_framebufferfi(&self) -> extern fn(c_uint, c_uint, c_float, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float, int) QOpenGLFunctions_4_5_CoreBackend::ClearNamedFramebufferfi() const

fn clear_named_framebufferfv(&self) -> extern fn(c_uint, c_uint, c_int, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const float*) QOpenGLFunctions_4_5_CoreBackend::ClearNamedFramebufferfv() const

fn clear_named_framebufferiv(&self) -> extern fn(c_uint, c_uint, c_int, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const int*) QOpenGLFunctions_4_5_CoreBackend::ClearNamedFramebufferiv() const

fn clear_named_framebufferuiv(&self) -> extern fn(c_uint, c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*) QOpenGLFunctions_4_5_CoreBackend::ClearNamedFramebufferuiv() const

fn clip_control(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::ClipControl() const

fn compressed_texture_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_4_5_CoreBackend::CompressedTextureSubImage1D() const

fn compressed_texture_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_4_5_CoreBackend::CompressedTextureSubImage2D() const

fn compressed_texture_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, int, const void*) QOpenGLFunctions_4_5_CoreBackend::CompressedTextureSubImage3D() const

fn copy_named_buffer_sub_data(&self) -> extern fn(c_uint, c_uint, c_long, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long, long, int) QOpenGLFunctions_4_5_CoreBackend::CopyNamedBufferSubData() const

fn copy_texture_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int) QOpenGLFunctions_4_5_CoreBackend::CopyTextureSubImage1D() const

fn copy_texture_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) QOpenGLFunctions_4_5_CoreBackend::CopyTextureSubImage2D() const

fn copy_texture_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int) QOpenGLFunctions_4_5_CoreBackend::CopyTextureSubImage3D() const

fn create_buffers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateBuffers() const

fn create_framebuffers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateFramebuffers() const

fn create_program_pipelines(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateProgramPipelines() const

fn create_queries(&self) -> extern fn(c_uint, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateQueries() const

fn create_renderbuffers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateRenderbuffers() const

fn create_samplers(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateSamplers() const

fn create_textures(&self) -> extern fn(c_uint, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateTextures() const

fn create_transform_feedbacks(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateTransformFeedbacks() const

fn create_vertex_arrays(&self) -> extern fn(c_int, *mut c_uint)

C++ method: void (*FN_PTR)(int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::CreateVertexArrays() const

fn disable_vertex_array_attrib(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::DisableVertexArrayAttrib() const

fn enable_vertex_array_attrib(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::EnableVertexArrayAttrib() const

fn flush_mapped_named_buffer_range(&self) -> extern fn(c_uint, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, long, int) QOpenGLFunctions_4_5_CoreBackend::FlushMappedNamedBufferRange() const

fn generate_texture_mipmap(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_5_CoreBackend::GenerateTextureMipmap() const

fn get_compressed_texture_image(&self) -> extern fn(c_uint, c_int, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetCompressedTextureImage() const

fn get_compressed_texture_sub_image(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetCompressedTextureSubImage() const

fn get_graphics_reset_status(&self) -> extern fn() -> c_uint

C++ method: unsigned int (*FN_PTR)() QOpenGLFunctions_4_5_CoreBackend::GetGraphicsResetStatus() const

fn get_named_buffer_parameteri64v(&self) -> extern fn(c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, long*) QOpenGLFunctions_4_5_CoreBackend::GetNamedBufferParameteri64v() const

fn get_named_buffer_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetNamedBufferParameteriv() const

fn get_named_buffer_pointerv(&self) -> extern fn(c_uint, c_uint, *mut *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, void**) QOpenGLFunctions_4_5_CoreBackend::GetNamedBufferPointerv() const

fn get_named_buffer_sub_data(&self) -> extern fn(c_uint, c_long, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, long, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetNamedBufferSubData() const

fn get_named_framebuffer_attachment_parameteriv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetNamedFramebufferAttachmentParameteriv() const

fn get_named_framebuffer_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetNamedFramebufferParameteriv() const

fn get_named_renderbuffer_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetNamedRenderbufferParameteriv() const

fn get_texture_image(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetTextureImage() const

fn get_texture_level_parameterfv(&self) -> extern fn(c_uint, c_int, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, float*) QOpenGLFunctions_4_5_CoreBackend::GetTextureLevelParameterfv() const

fn get_texture_level_parameteriv(&self) -> extern fn(c_uint, c_int, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetTextureLevelParameteriv() const

fn get_texture_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetTextureParameterIiv() const

fn get_texture_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::GetTextureParameterIuiv() const

fn get_texture_parameterfv(&self) -> extern fn(c_uint, c_uint, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float*) QOpenGLFunctions_4_5_CoreBackend::GetTextureParameterfv() const

fn get_texture_parameteriv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetTextureParameteriv() const

fn get_texture_sub_image(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetTextureSubImage() const

fn get_transform_feedbacki64_v(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long*) QOpenGLFunctions_4_5_CoreBackend::GetTransformFeedbacki64_v() const

fn get_transform_feedbacki_v(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetTransformFeedbacki_v() const

fn get_transform_feedbackiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetTransformFeedbackiv() const

fn get_vertex_array_indexed64iv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_long)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long*) QOpenGLFunctions_4_5_CoreBackend::GetVertexArrayIndexed64iv() const

fn get_vertex_array_indexediv(&self) -> extern fn(c_uint, c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetVertexArrayIndexediv() const

fn get_vertex_arrayiv(&self) -> extern fn(c_uint, c_uint, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int*) QOpenGLFunctions_4_5_CoreBackend::GetVertexArrayiv() const

fn getn_compressed_tex_image(&self) -> extern fn(c_uint, c_int, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetnCompressedTexImage() const

fn getn_tex_image(&self) -> extern fn(c_uint, c_int, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_CoreBackend::GetnTexImage() const

fn getn_uniformdv(&self) -> extern fn(c_uint, c_int, c_int, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, int, int, double*) QOpenGLFunctions_4_5_CoreBackend::GetnUniformdv() const

fn getn_uniformfv(&self) -> extern fn(c_uint, c_int, c_int, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, int, int, float*) QOpenGLFunctions_4_5_CoreBackend::GetnUniformfv() const

fn getn_uniformiv(&self) -> extern fn(c_uint, c_int, c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, int, int, int*) QOpenGLFunctions_4_5_CoreBackend::GetnUniformiv() const

fn getn_uniformuiv(&self) -> extern fn(c_uint, c_int, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, int, unsigned int*) QOpenGLFunctions_4_5_CoreBackend::GetnUniformuiv() const

fn invalidate_named_framebuffer_data(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_5_CoreBackend::InvalidateNamedFramebufferData() const

fn invalidate_named_framebuffer_sub_data(&self) -> extern fn(c_uint, c_int, *const c_uint, c_int, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*, int, int, int, int) QOpenGLFunctions_4_5_CoreBackend::InvalidateNamedFramebufferSubData() const

fn map_named_buffer(&self) -> extern fn(c_uint, c_uint) -> *mut c_void

C++ method: void* (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::MapNamedBuffer() const

fn map_named_buffer_range(&self) -> extern fn(c_uint, c_long, c_int, c_uint) -> *mut c_void

C++ method: void* (*FN_PTR)(unsigned int, long, int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::MapNamedBufferRange() const

fn memory_barrier_by_region(&self) -> extern fn(c_uint)

C++ method: void (*FN_PTR)(unsigned int) QOpenGLFunctions_4_5_CoreBackend::MemoryBarrierByRegion() const

fn named_buffer_data(&self) -> extern fn(c_uint, c_int, *const c_void, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const void*, unsigned int) QOpenGLFunctions_4_5_CoreBackend::NamedBufferData() const

fn named_buffer_storage(&self) -> extern fn(c_uint, c_int, *const c_void, c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const void*, unsigned int) QOpenGLFunctions_4_5_CoreBackend::NamedBufferStorage() const

fn named_buffer_sub_data(&self) -> extern fn(c_uint, c_long, c_int, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, long, int, const void*) QOpenGLFunctions_4_5_CoreBackend::NamedBufferSubData() const

fn named_framebuffer_draw_buffer(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferDrawBuffer() const

fn named_framebuffer_draw_buffers(&self) -> extern fn(c_uint, c_int, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, const unsigned int*) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferDrawBuffers() const

fn named_framebuffer_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferParameteri() const

fn named_framebuffer_read_buffer(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferReadBuffer() const

fn named_framebuffer_renderbuffer(&self) -> extern fn(c_uint, c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferRenderbuffer() const

fn named_framebuffer_texture(&self) -> extern fn(c_uint, c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferTexture() const

fn named_framebuffer_texture_layer(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int) QOpenGLFunctions_4_5_CoreBackend::NamedFramebufferTextureLayer() const

fn named_renderbuffer_storage(&self) -> extern fn(c_uint, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int) QOpenGLFunctions_4_5_CoreBackend::NamedRenderbufferStorage() const

fn named_renderbuffer_storage_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int) QOpenGLFunctions_4_5_CoreBackend::NamedRenderbufferStorageMultisample() const

fn readn_pixels(&self) -> extern fn(c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(int, int, int, int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_CoreBackend::ReadnPixels() const

fn set_bind_texture_unit(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_BindTextureUnit(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_blit_named_framebuffer(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_BlitNamedFramebuffer(void (*FN_PTR)(unsigned int, unsigned int, int, int, int, int, int, int, int, int, unsigned int, unsigned int) value)

fn set_check_named_framebuffer_status(&mut self, value: extern fn(c_uint, c_uint) -> c_uint)

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CheckNamedFramebufferStatus(unsigned int (*FN_PTR)(unsigned int, unsigned int) value)

fn set_clear_named_buffer_data(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedBufferData(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int, const void*) value)

fn set_clear_named_buffer_sub_data(&mut self, value: extern fn(c_uint, c_uint, c_long, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedBufferSubData(void (*FN_PTR)(unsigned int, unsigned int, long, int, unsigned int, unsigned int, const void*) value)

fn set_clear_named_framebufferfi(&mut self, value: extern fn(c_uint, c_uint, c_float, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedFramebufferfi(void (*FN_PTR)(unsigned int, unsigned int, float, int) value)

fn set_clear_named_framebufferfv(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedFramebufferfv(void (*FN_PTR)(unsigned int, unsigned int, int, const float*) value)

fn set_clear_named_framebufferiv(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedFramebufferiv(void (*FN_PTR)(unsigned int, unsigned int, int, const int*) value)

fn set_clear_named_framebufferuiv(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClearNamedFramebufferuiv(void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*) value)

fn set_clip_control(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ClipControl(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_compressed_texture_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CompressedTextureSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, unsigned int, int, const void*) value)

fn set_compressed_texture_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CompressedTextureSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, int, const void*) value)

fn set_compressed_texture_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_int, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CompressedTextureSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, int, const void*) value)

fn set_copy_named_buffer_sub_data(&mut self, value: extern fn(c_uint, c_uint, c_long, c_long, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CopyNamedBufferSubData(void (*FN_PTR)(unsigned int, unsigned int, long, long, int) value)

fn set_copy_texture_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CopyTextureSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, int, int) value)

fn set_copy_texture_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CopyTextureSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int) value)

fn set_copy_texture_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CopyTextureSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int) value)

fn set_create_buffers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateBuffers(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_framebuffers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateFramebuffers(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_program_pipelines(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateProgramPipelines(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_queries(&mut self, value: extern fn(c_uint, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateQueries(void (*FN_PTR)(unsigned int, int, unsigned int*) value)

fn set_create_renderbuffers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateRenderbuffers(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_samplers(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateSamplers(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_textures(&mut self, value: extern fn(c_uint, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateTextures(void (*FN_PTR)(unsigned int, int, unsigned int*) value)

fn set_create_transform_feedbacks(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateTransformFeedbacks(void (*FN_PTR)(int, unsigned int*) value)

fn set_create_vertex_arrays(&mut self, value: extern fn(c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_CreateVertexArrays(void (*FN_PTR)(int, unsigned int*) value)

fn set_disable_vertex_array_attrib(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_DisableVertexArrayAttrib(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_enable_vertex_array_attrib(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_EnableVertexArrayAttrib(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_flush_mapped_named_buffer_range(&mut self, value: extern fn(c_uint, c_long, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_FlushMappedNamedBufferRange(void (*FN_PTR)(unsigned int, long, int) value)

fn set_generate_texture_mipmap(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GenerateTextureMipmap(void (*FN_PTR)(unsigned int) value)

fn set_get_compressed_texture_image(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetCompressedTextureImage(void (*FN_PTR)(unsigned int, int, int, void*) value)

fn set_get_compressed_texture_sub_image(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetCompressedTextureSubImage(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, int, void*) value)

fn set_get_graphics_reset_status(&mut self, value: extern fn() -> c_uint)

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetGraphicsResetStatus(unsigned int (*FN_PTR)() value)

fn set_get_named_buffer_parameteri64v(&mut self, value: extern fn(c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedBufferParameteri64v(void (*FN_PTR)(unsigned int, unsigned int, long*) value)

fn set_get_named_buffer_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedBufferParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_named_buffer_pointerv(&mut self, value: extern fn(c_uint, c_uint, *mut *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedBufferPointerv(void (*FN_PTR)(unsigned int, unsigned int, void**) value)

fn set_get_named_buffer_sub_data(&mut self, value: extern fn(c_uint, c_long, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedBufferSubData(void (*FN_PTR)(unsigned int, long, int, void*) value)

fn set_get_named_framebuffer_attachment_parameteriv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedFramebufferAttachmentParameteriv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_named_framebuffer_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedFramebufferParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_named_renderbuffer_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetNamedRenderbufferParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_texture_image(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureImage(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, int, void*) value)

fn set_get_texture_level_parameterfv(&mut self, value: extern fn(c_uint, c_int, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureLevelParameterfv(void (*FN_PTR)(unsigned int, int, unsigned int, float*) value)

fn set_get_texture_level_parameteriv(&mut self, value: extern fn(c_uint, c_int, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureLevelParameteriv(void (*FN_PTR)(unsigned int, int, unsigned int, int*) value)

fn set_get_texture_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_texture_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int*) value)

fn set_get_texture_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *mut c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureParameterfv(void (*FN_PTR)(unsigned int, unsigned int, float*) value)

fn set_get_texture_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureParameteriv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_texture_sub_image(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTextureSubImage(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, int, void*) value)

fn set_get_transform_feedbacki64_v(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTransformFeedbacki64_v(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long*) value)

fn set_get_transform_feedbacki_v(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTransformFeedbacki_v(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_transform_feedbackiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetTransformFeedbackiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_get_vertex_array_indexed64iv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_long))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetVertexArrayIndexed64iv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long*) value)

fn set_get_vertex_array_indexediv(&mut self, value: extern fn(c_uint, c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetVertexArrayIndexediv(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int*) value)

fn set_get_vertex_arrayiv(&mut self, value: extern fn(c_uint, c_uint, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetVertexArrayiv(void (*FN_PTR)(unsigned int, unsigned int, int*) value)

fn set_getn_compressed_tex_image(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnCompressedTexImage(void (*FN_PTR)(unsigned int, int, int, void*) value)

fn set_getn_tex_image(&mut self, value: extern fn(c_uint, c_int, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnTexImage(void (*FN_PTR)(unsigned int, int, unsigned int, unsigned int, int, void*) value)

fn set_getn_uniformdv(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_double))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnUniformdv(void (*FN_PTR)(unsigned int, int, int, double*) value)

fn set_getn_uniformfv(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnUniformfv(void (*FN_PTR)(unsigned int, int, int, float*) value)

fn set_getn_uniformiv(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnUniformiv(void (*FN_PTR)(unsigned int, int, int, int*) value)

fn set_getn_uniformuiv(&mut self, value: extern fn(c_uint, c_int, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_GetnUniformuiv(void (*FN_PTR)(unsigned int, int, int, unsigned int*) value)

fn set_invalidate_named_framebuffer_data(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_InvalidateNamedFramebufferData(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_invalidate_named_framebuffer_sub_data(&mut self, value: extern fn(c_uint, c_int, *const c_uint, c_int, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_InvalidateNamedFramebufferSubData(void (*FN_PTR)(unsigned int, int, const unsigned int*, int, int, int, int) value)

fn set_map_named_buffer(&mut self, value: extern fn(c_uint, c_uint) -> *mut c_void)

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_MapNamedBuffer(void* (*FN_PTR)(unsigned int, unsigned int) value)

fn set_map_named_buffer_range(&mut self, value: extern fn(c_uint, c_long, c_int, c_uint) -> *mut c_void)

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_MapNamedBufferRange(void* (*FN_PTR)(unsigned int, long, int, unsigned int) value)

fn set_memory_barrier_by_region(&mut self, value: extern fn(c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_MemoryBarrierByRegion(void (*FN_PTR)(unsigned int) value)

fn set_named_buffer_data(&mut self, value: extern fn(c_uint, c_int, *const c_void, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedBufferData(void (*FN_PTR)(unsigned int, int, const void*, unsigned int) value)

fn set_named_buffer_storage(&mut self, value: extern fn(c_uint, c_int, *const c_void, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedBufferStorage(void (*FN_PTR)(unsigned int, int, const void*, unsigned int) value)

fn set_named_buffer_sub_data(&mut self, value: extern fn(c_uint, c_long, c_int, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedBufferSubData(void (*FN_PTR)(unsigned int, long, int, const void*) value)

fn set_named_framebuffer_draw_buffer(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferDrawBuffer(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_named_framebuffer_draw_buffers(&mut self, value: extern fn(c_uint, c_int, *const c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferDrawBuffers(void (*FN_PTR)(unsigned int, int, const unsigned int*) value)

fn set_named_framebuffer_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_named_framebuffer_read_buffer(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferReadBuffer(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_named_framebuffer_renderbuffer(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferRenderbuffer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, unsigned int) value)

fn set_named_framebuffer_texture(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferTexture(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int) value)

fn set_named_framebuffer_texture_layer(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedFramebufferTextureLayer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, int) value)

fn set_named_renderbuffer_storage(&mut self, value: extern fn(c_uint, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedRenderbufferStorage(void (*FN_PTR)(unsigned int, unsigned int, int, int) value)

fn set_named_renderbuffer_storage_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_NamedRenderbufferStorageMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int) value)

fn set_readn_pixels(&mut self, value: extern fn(c_int, c_int, c_int, c_int, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_ReadnPixels(void (*FN_PTR)(int, int, int, int, unsigned int, unsigned int, int, void*) value)

fn set_texture_barrier(&mut self, value: extern fn())

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureBarrier(void (*FN_PTR)() value)

fn set_texture_buffer(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureBuffer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_texture_buffer_range(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_long, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureBufferRange(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) value)

fn set_texture_parameter_iiv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameterIiv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_texture_parameter_iuiv(&mut self, value: extern fn(c_uint, c_uint, *const c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameterIuiv(void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) value)

fn set_texture_parameterf(&mut self, value: extern fn(c_uint, c_uint, c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameterf(void (*FN_PTR)(unsigned int, unsigned int, float) value)

fn set_texture_parameterfv(&mut self, value: extern fn(c_uint, c_uint, *const c_float))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameterfv(void (*FN_PTR)(unsigned int, unsigned int, const float*) value)

fn set_texture_parameteri(&mut self, value: extern fn(c_uint, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameteri(void (*FN_PTR)(unsigned int, unsigned int, int) value)

fn set_texture_parameteriv(&mut self, value: extern fn(c_uint, c_uint, *const c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureParameteriv(void (*FN_PTR)(unsigned int, unsigned int, const int*) value)

fn set_texture_storage1_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureStorage1D(void (*FN_PTR)(unsigned int, int, unsigned int, int) value)

fn set_texture_storage2_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureStorage2D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int) value)

fn set_texture_storage2_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureStorage2DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) value)

fn set_texture_storage3_d(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureStorage3D(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int) value)

fn set_texture_storage3_d_multisample(&mut self, value: extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureStorage3DMultisample(void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) value)

fn set_texture_sub_image1_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureSubImage1D(void (*FN_PTR)(unsigned int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_texture_sub_image2_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureSubImage2D(void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_texture_sub_image3_d(&mut self, value: extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TextureSubImage3D(void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) value)

fn set_transform_feedback_buffer_base(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TransformFeedbackBufferBase(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_transform_feedback_buffer_range(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_long, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_TransformFeedbackBufferRange(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) value)

fn set_unmap_named_buffer(&mut self, value: extern fn(c_uint) -> c_uchar)

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_UnmapNamedBuffer(unsigned char (*FN_PTR)(unsigned int) value)

fn set_vertex_array_attrib_binding(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayAttribBinding(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_vertex_array_attrib_format(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint, c_uchar, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayAttribFormat(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned char, unsigned int) value)

fn set_vertex_array_attrib_i_format(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayAttribIFormat(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int) value)

fn set_vertex_array_attrib_l_format(&mut self, value: extern fn(c_uint, c_uint, c_int, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayAttribLFormat(void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int) value)

fn set_vertex_array_binding_divisor(&mut self, value: extern fn(c_uint, c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayBindingDivisor(void (*FN_PTR)(unsigned int, unsigned int, unsigned int) value)

fn set_vertex_array_element_buffer(&mut self, value: extern fn(c_uint, c_uint))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayElementBuffer(void (*FN_PTR)(unsigned int, unsigned int) value)

fn set_vertex_array_vertex_buffer(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_long, c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayVertexBuffer(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) value)

fn set_vertex_array_vertex_buffers(&mut self, value: extern fn(c_uint, c_uint, c_int, *const c_uint, *const c_long, *const c_int))

C++ method: void QOpenGLFunctions_4_5_CoreBackend::set_VertexArrayVertexBuffers(void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*, const long*, const int*) value)

fn texture_barrier(&self) -> extern fn()

C++ method: void (*FN_PTR)() QOpenGLFunctions_4_5_CoreBackend::TextureBarrier() const

fn texture_buffer(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::TextureBuffer() const

fn texture_buffer_range(&self) -> extern fn(c_uint, c_uint, c_uint, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) QOpenGLFunctions_4_5_CoreBackend::TextureBufferRange() const

fn texture_parameter_iiv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_4_5_CoreBackend::TextureParameterIiv() const

fn texture_parameter_iuiv(&self) -> extern fn(c_uint, c_uint, *const c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const unsigned int*) QOpenGLFunctions_4_5_CoreBackend::TextureParameterIuiv() const

fn texture_parameterf(&self) -> extern fn(c_uint, c_uint, c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, float) QOpenGLFunctions_4_5_CoreBackend::TextureParameterf() const

fn texture_parameterfv(&self) -> extern fn(c_uint, c_uint, *const c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const float*) QOpenGLFunctions_4_5_CoreBackend::TextureParameterfv() const

fn texture_parameteri(&self) -> extern fn(c_uint, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int) QOpenGLFunctions_4_5_CoreBackend::TextureParameteri() const

fn texture_parameteriv(&self) -> extern fn(c_uint, c_uint, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, const int*) QOpenGLFunctions_4_5_CoreBackend::TextureParameteriv() const

fn texture_storage1_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int) QOpenGLFunctions_4_5_CoreBackend::TextureStorage1D() const

fn texture_storage2_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int) QOpenGLFunctions_4_5_CoreBackend::TextureStorage2D() const

fn texture_storage2_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, unsigned char) QOpenGLFunctions_4_5_CoreBackend::TextureStorage2DMultisample() const

fn texture_storage3_d(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int) QOpenGLFunctions_4_5_CoreBackend::TextureStorage3D() const

fn texture_storage3_d_multisample(&self) -> extern fn(c_uint, c_int, c_uint, c_int, c_int, c_int, c_uchar)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int, int, int, int, unsigned char) QOpenGLFunctions_4_5_CoreBackend::TextureStorage3DMultisample() const

fn texture_sub_image1_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_5_CoreBackend::TextureSubImage1D() const

fn texture_sub_image2_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_5_CoreBackend::TextureSubImage2D() const

fn texture_sub_image3_d(&self) -> extern fn(c_uint, c_int, c_int, c_int, c_int, c_int, c_int, c_int, c_uint, c_uint, *const c_void)

C++ method: void (*FN_PTR)(unsigned int, int, int, int, int, int, int, int, unsigned int, unsigned int, const void*) QOpenGLFunctions_4_5_CoreBackend::TextureSubImage3D() const

fn transform_feedback_buffer_base(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::TransformFeedbackBufferBase() const

fn transform_feedback_buffer_range(&self) -> extern fn(c_uint, c_uint, c_uint, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) QOpenGLFunctions_4_5_CoreBackend::TransformFeedbackBufferRange() const

fn unmap_named_buffer(&self) -> extern fn(c_uint) -> c_uchar

C++ method: unsigned char (*FN_PTR)(unsigned int) QOpenGLFunctions_4_5_CoreBackend::UnmapNamedBuffer() const

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions45CoreBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_5_CoreBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_5_CoreBackend::versionStatus()

fn vertex_array_attrib_binding(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayAttribBinding() const

fn vertex_array_attrib_format(&self) -> extern fn(c_uint, c_uint, c_int, c_uint, c_uchar, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned char, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayAttribFormat() const

fn vertex_array_attrib_i_format(&self) -> extern fn(c_uint, c_uint, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayAttribIFormat() const

fn vertex_array_attrib_l_format(&self) -> extern fn(c_uint, c_uint, c_int, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayAttribLFormat() const

fn vertex_array_binding_divisor(&self) -> extern fn(c_uint, c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayBindingDivisor() const

fn vertex_array_element_buffer(&self) -> extern fn(c_uint, c_uint)

C++ method: void (*FN_PTR)(unsigned int, unsigned int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayElementBuffer() const

fn vertex_array_vertex_buffer(&self) -> extern fn(c_uint, c_uint, c_uint, c_long, c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, long, int) QOpenGLFunctions_4_5_CoreBackend::VertexArrayVertexBuffer() const

fn vertex_array_vertex_buffers(&self) -> extern fn(c_uint, c_uint, c_int, *const c_uint, *const c_long, *const c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, const unsigned int*, const long*, const int*) QOpenGLFunctions_4_5_CoreBackend::VertexArrayVertexBuffers() const