Struct qt_gui::open_g_l_version_functions::OpenGLFunctions45DeprecatedBackend

pub struct OpenGLFunctions45DeprecatedBackend {
    // some fields omitted
}

C++ type: QOpenGLFunctions_4_5_DeprecatedBackend

Methods

impl OpenGLFunctions45DeprecatedBackend

fn getn_color_table(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnColorTable() const

fn getn_convolution_filter(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnConvolutionFilter() const

fn getn_histogram(&self) -> extern fn(c_uint, c_uchar, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnHistogram() const

fn getn_mapdv(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_double)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, double*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnMapdv() const

fn getn_mapfv(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, float*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnMapfv() const

fn getn_mapiv(&self) -> extern fn(c_uint, c_uint, c_int, *mut c_int)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, int, int*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnMapiv() const

fn getn_minmax(&self) -> extern fn(c_uint, c_uchar, c_uint, c_uint, c_int, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, int, void*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnMinmax() const

fn getn_pixel_mapfv(&self) -> extern fn(c_uint, c_int, *mut c_float)

C++ method: void (*FN_PTR)(unsigned int, int, float*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnPixelMapfv() const

fn getn_pixel_mapuiv(&self) -> extern fn(c_uint, c_int, *mut c_uint)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned int*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnPixelMapuiv() const

fn getn_pixel_mapusv(&self) -> extern fn(c_uint, c_int, *mut c_ushort)

C++ method: void (*FN_PTR)(unsigned int, int, unsigned short*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnPixelMapusv() const

fn getn_polygon_stipple(&self) -> extern fn(c_int, *mut c_uchar)

C++ method: void (*FN_PTR)(int, unsigned char*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnPolygonStipple() const

fn getn_separable_filter(&self) -> extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void, c_int, *mut c_void, *mut c_void)

C++ method: void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*, int, void*, void*) QOpenGLFunctions_4_5_DeprecatedBackend::GetnSeparableFilter() const

fn set_getn_color_table(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnColorTable(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*) value)

fn set_getn_convolution_filter(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnConvolutionFilter(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*) value)

fn set_getn_histogram(&mut self, value: extern fn(c_uint, c_uchar, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnHistogram(void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, int, void*) value)

fn set_getn_mapdv(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_double))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnMapdv(void (*FN_PTR)(unsigned int, unsigned int, int, double*) value)

fn set_getn_mapfv(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_float))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnMapfv(void (*FN_PTR)(unsigned int, unsigned int, int, float*) value)

fn set_getn_mapiv(&mut self, value: extern fn(c_uint, c_uint, c_int, *mut c_int))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnMapiv(void (*FN_PTR)(unsigned int, unsigned int, int, int*) value)

fn set_getn_minmax(&mut self, value: extern fn(c_uint, c_uchar, c_uint, c_uint, c_int, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnMinmax(void (*FN_PTR)(unsigned int, unsigned char, unsigned int, unsigned int, int, void*) value)

fn set_getn_pixel_mapfv(&mut self, value: extern fn(c_uint, c_int, *mut c_float))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnPixelMapfv(void (*FN_PTR)(unsigned int, int, float*) value)

fn set_getn_pixel_mapuiv(&mut self, value: extern fn(c_uint, c_int, *mut c_uint))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnPixelMapuiv(void (*FN_PTR)(unsigned int, int, unsigned int*) value)

fn set_getn_pixel_mapusv(&mut self, value: extern fn(c_uint, c_int, *mut c_ushort))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnPixelMapusv(void (*FN_PTR)(unsigned int, int, unsigned short*) value)

fn set_getn_polygon_stipple(&mut self, value: extern fn(c_int, *mut c_uchar))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnPolygonStipple(void (*FN_PTR)(int, unsigned char*) value)

fn set_getn_separable_filter(&mut self, value: extern fn(c_uint, c_uint, c_uint, c_int, *mut c_void, c_int, *mut c_void, *mut c_void))

C++ method: void QOpenGLFunctions_4_5_DeprecatedBackend::set_GetnSeparableFilter(void (*FN_PTR)(unsigned int, unsigned int, unsigned int, int, void*, int, void*, void*) value)

fn version_status<Args>(args: Args) -> Args::ReturnType where Args: OpenGLFunctions45DeprecatedBackendVersionStatusArgs

C++ method: QOpenGLFunctions_4_5_DeprecatedBackend::versionStatus

This is an overloaded function. Available variants:

Rust arguments:
1) fn version_status(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_functions::OpenGLVersionStatus>
2) fn version_status(cpp_utils::AsStruct) -> ::open_g_l_version_functions::OpenGLVersionStatus
C++ method: static QOpenGLVersionStatus QOpenGLFunctions_4_5_DeprecatedBackend::versionStatus()