Struct qt_gui::open_g_l_version_profile::OpenGLVersionProfile

pub struct OpenGLVersionProfile {
    // some fields omitted
}

C++ type: QOpenGLVersionProfile

C++ documentation:

The QOpenGLVersionProfile class represents the version and if applicable the profile of an OpenGL context.

An object of this class can be passed to QOpenGLContext::versionFunctions() to request a functions object for a specific version and profile of OpenGL.

It also contains some helper functions to check if a version supports profiles or is a legacy version.

Methods

impl OpenGLVersionProfile

fn has_profiles(&self) -> bool

C++ method: bool QOpenGLVersionProfile::hasProfiles() const

C++ documentation:

Returns true if profiles are supported by the OpenGL version returned by version(). Only OpenGL versions >= 3.2 support profiles.

See also profile() and version().

fn is_legacy_version(&self) -> bool

C++ method: bool QOpenGLVersionProfile::isLegacyVersion() const

C++ documentation:

Returns true is the OpenGL version returned by version() contains deprecated functions and does not support profiles i.e. if the OpenGL version is <= 3.1.

fn is_valid(&self) -> bool

C++ method: bool QOpenGLVersionProfile::isValid() const

C++ documentation:

Returns true if the version number is valid. Note that for a default constructed QOpenGLVersionProfile object this function will return false.

See also setVersion() and version().

fn new<Args>(args: Args) -> Args::ReturnType where Args: OpenGLVersionProfileNewArgs

C++ method: QOpenGLVersionProfile::QOpenGLVersionProfile

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::open_g_l_version_profile::OpenGLVersionProfile
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::open_g_l_version_profile::OpenGLVersionProfile>
C++ method: [constructor] void QOpenGLVersionProfile::QOpenGLVersionProfile()

Variant 2

Rust arguments:
1) fn new((&::open_g_l_version_profile::OpenGLVersionProfile, cpp_utils::AsStruct)) -> ::open_g_l_version_profile::OpenGLVersionProfile
2) fn new((&::open_g_l_version_profile::OpenGLVersionProfile, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_version_profile::OpenGLVersionProfile>
C++ method: [constructor] void QOpenGLVersionProfile::QOpenGLVersionProfile(const QOpenGLVersionProfile& other)

Variant 3

Rust arguments:
1) fn new((&::surface_format::SurfaceFormat, cpp_utils::AsStruct)) -> ::open_g_l_version_profile::OpenGLVersionProfile
2) fn new((&::surface_format::SurfaceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_version_profile::OpenGLVersionProfile>
C++ method: [constructor] void QOpenGLVersionProfile::QOpenGLVersionProfile(const QSurfaceFormat& format)

fn op_assign<'l0, 'l1>(&'l0 mut self, rhs: &'l1 OpenGLVersionProfile) -> &'l0 mut OpenGLVersionProfile

C++ method: QOpenGLVersionProfile& QOpenGLVersionProfile::operator=(const QOpenGLVersionProfile& rhs)

C++ documentation:

Assigns the version and profile of rhs to this QOpenGLVersionProfile object.

fn profile(&self) -> OpenGLContextProfile

C++ method: QSurfaceFormat::OpenGLContextProfile QOpenGLVersionProfile::profile() const

C++ documentation:

Returns the OpenGL profile. Only makes sense if profiles are supported by this version.

See also setProfile().

fn set_profile(&mut self, profile: OpenGLContextProfile)

C++ method: void QOpenGLVersionProfile::setProfile(QSurfaceFormat::OpenGLContextProfile profile)

C++ documentation:

Sets the OpenGL profile profile. Only makes sense if profiles are supported by this version.

See also profile().

fn set_version(&mut self, major_version: c_int, minor_version: c_int)

C++ method: void QOpenGLVersionProfile::setVersion(int majorVersion, int minorVersion)

C++ documentation:

Sets the major and minor version numbers to majorVersion and minorVersion respectively.

See also version().

fn version<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: OpenGLVersionProfileVersionArgs<'largs>

C++ method: QOpenGLVersionProfile::version

This is an overloaded function. Available variants:

Rust arguments:
1) fn version(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCIntCInt>
2) fn version(&self, cpp_utils::AsStruct) -> ::pair::PairCIntCInt
C++ method: QPair<int, int> QOpenGLVersionProfile::version() const

C++ documentation:

Returns a QPair<int,int> where the components represent the major and minor OpenGL version numbers respectively.

See also setVersion().