Module qt_gui::open_g_l_vertex_array_object::overloading

Traits

BinderNewArgs
This trait represents a set of arguments accepted by Binder::new method.

OpenGLVertexArrayObjectConnectArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::connect method.

OpenGLVertexArrayObjectConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::connect_static method.

OpenGLVertexArrayObjectDestroyedArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::destroyed method.

OpenGLVertexArrayObjectDisconnectArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::disconnect method.

OpenGLVertexArrayObjectDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::disconnect_static method.

OpenGLVertexArrayObjectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::dynamic_property_names method.

OpenGLVertexArrayObjectNewArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::new method.

OpenGLVertexArrayObjectObjectNameArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::object_name method.

OpenGLVertexArrayObjectPropertyArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::property method.

OpenGLVertexArrayObjectStartTimerArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::start_timer method.

OpenGLVertexArrayObjectTrArgs
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::tr method.

OpenGLVertexArrayObjectTrUtf8Args
This trait represents a set of arguments accepted by OpenGLVertexArrayObject::tr_utf8 method.