Struct qt_gui::open_g_l_vertex_array_object::Binder

pub struct Binder {
    // some fields omitted
}

C++ type: QOpenGLVertexArrayObject::Binder

C++ documentation:

The QOpenGLVertexArrayObject::Binder class is a convenience class to help with the binding and releasing of OpenGL Vertex Array Objects.

QOpenGLVertexArrayObject::Binder is a simple convenience class that can be used to assist with the binding and releasing of QOpenGLVertexArrayObject instances. This class is to QOpenGLVertexArrayObject as QMutexLocker is to QMutex.

This class implements the RAII principle which helps to ensure behavior in complex code or in the presence of exceptions.

The constructor of this class accepts a QOpenGLVertexArrayObject (VAO) as an argument and attempts to bind the VAO, calling QOpenGLVertexArrayObject::create() if necessary. The destructor of this class calls QOpenGLVertexArrayObject::release() which unbinds the VAO.

If needed the VAO can be temporarily unbound with the release() function and bound once more with rebind().

Methods

impl Binder

fn new<Args>(args: Args) -> Args::ReturnType where Args: BinderNewArgs

C++ method: QOpenGLVertexArrayObject::Binder::Binder

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::open_g_l_vertex_array_object::OpenGLVertexArrayObject, cpp_utils::AsStruct)) -> ::open_g_l_vertex_array_object::Binder
2) fn new((*mut ::open_g_l_vertex_array_object::OpenGLVertexArrayObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::open_g_l_vertex_array_object::Binder>
C++ method: [constructor] void QOpenGLVertexArrayObject::Binder::Binder(QOpenGLVertexArrayObject* v)

fn rebind(&mut self)

C++ method: void QOpenGLVertexArrayObject::Binder::rebind()

C++ documentation:

Can be used to rebind the associated vertex array object.

See also release().

fn release(&mut self)

C++ method: void QOpenGLVertexArrayObject::Binder::release()

C++ documentation:

Can be used to temporarily release the associated vertex array object.

See also rebind().