Enum qt_gui::open_g_l_window::UpdateBehavior

pub enum UpdateBehavior {
    NoPartialUpdate,
    PartialUpdateBlit,
    PartialUpdateBlend,
}

C++ type: QOpenGLWindow::UpdateBehavior

C++ documentation:

This enum describes the update strategy of the QOpenGLWindow.

Variants
NoPartialUpdate

Indicates that the entire window surface will redrawn on each update and so no additional framebuffers are needed. This is the setting used in most cases and is equivalent to how drawing directly via QWindow would function. (C++ enum variant: NoPartialUpdate = 0)

PartialUpdateBlit

Indicates that the drawing performed in paintGL() does not cover the entire window. In this case an extra framebuffer object is created under the hood, and rendering performed in paintGL() will target this framebuffer. This framebuffer is then blitted onto the window surface's default framebuffer after each paint. This allows having QPainter-based drawing code in paintGL() which only repaints a smaller area at a time, because, unlike NoPartialUpdate, the previous content is preserved. (C++ enum variant: PartialUpdateBlit = 1)

PartialUpdateBlend

Similar to PartialUpdateBlit, but instead of using framebuffer blits, the contents of the extra framebuffer is rendered by drawing a textured quad with blending enabled. This, unlike PartialUpdateBlit, allows alpha blended content and works even when the glBlitFramebuffer is not available. Performance-wise this setting is likely to be somewhat slower than PartialUpdateBlit. (C++ enum variant: PartialUpdateBlend = 2)