Module qt_gui::open_g_l_window::overloading

Traits

OpenGLWindowBaseSizeArgs
This trait represents a set of arguments accepted by OpenGLWindow::base_size method.

OpenGLWindowConnectArgs
This trait represents a set of arguments accepted by OpenGLWindow::connect method.

OpenGLWindowConnectStaticArgs
This trait represents a set of arguments accepted by OpenGLWindow::connect_static method.

OpenGLWindowCursorArgs
This trait represents a set of arguments accepted by OpenGLWindow::cursor method.

OpenGLWindowDestroyedArgs
This trait represents a set of arguments accepted by OpenGLWindow::destroyed method.

OpenGLWindowDisconnectArgs
This trait represents a set of arguments accepted by OpenGLWindow::disconnect method.

OpenGLWindowDisconnectStaticArgs
This trait represents a set of arguments accepted by OpenGLWindow::disconnect_static method.

OpenGLWindowDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by OpenGLWindow::dynamic_property_names method.

OpenGLWindowFilePathArgs
This trait represents a set of arguments accepted by OpenGLWindow::file_path method.

OpenGLWindowFormatArgs
This trait represents a set of arguments accepted by OpenGLWindow::format method.

OpenGLWindowFrameGeometryArgs
This trait represents a set of arguments accepted by OpenGLWindow::frame_geometry method.

OpenGLWindowFrameMarginsArgs
This trait represents a set of arguments accepted by OpenGLWindow::frame_margins method.

OpenGLWindowFramePositionArgs
This trait represents a set of arguments accepted by OpenGLWindow::frame_position method.

OpenGLWindowGeometryArgs
This trait represents a set of arguments accepted by OpenGLWindow::geometry method.

OpenGLWindowGrabFramebufferArgs
This trait represents a set of arguments accepted by OpenGLWindow::grab_framebuffer method.

OpenGLWindowIconArgs
This trait represents a set of arguments accepted by OpenGLWindow::icon method.

OpenGLWindowIsAncestorOfArgs
This trait represents a set of arguments accepted by OpenGLWindow::is_ancestor_of method.

OpenGLWindowMapFromGlobalArgs
This trait represents a set of arguments accepted by OpenGLWindow::map_from_global method.

OpenGLWindowMapToGlobalArgs
This trait represents a set of arguments accepted by OpenGLWindow::map_to_global method.

OpenGLWindowMaskArgs
This trait represents a set of arguments accepted by OpenGLWindow::mask method.

OpenGLWindowMaximumSizeArgs
This trait represents a set of arguments accepted by OpenGLWindow::maximum_size method.

OpenGLWindowMinimumSizeArgs
This trait represents a set of arguments accepted by OpenGLWindow::minimum_size method.

OpenGLWindowNewArgs
This trait represents a set of arguments accepted by OpenGLWindow::new method.

OpenGLWindowObjectNameArgs
This trait represents a set of arguments accepted by OpenGLWindow::object_name method.

OpenGLWindowPositionArgs
This trait represents a set of arguments accepted by OpenGLWindow::position method.

OpenGLWindowPropertyArgs
This trait represents a set of arguments accepted by OpenGLWindow::property method.

OpenGLWindowRequestedFormatArgs
This trait represents a set of arguments accepted by OpenGLWindow::requested_format method.

OpenGLWindowResizeArgs
This trait represents a set of arguments accepted by OpenGLWindow::resize method.

OpenGLWindowSetGeometryArgs
This trait represents a set of arguments accepted by OpenGLWindow::set_geometry method.

OpenGLWindowSetPositionArgs
This trait represents a set of arguments accepted by OpenGLWindow::set_position method.

OpenGLWindowSizeArgs
This trait represents a set of arguments accepted by OpenGLWindow::size method.

OpenGLWindowSizeIncrementArgs
This trait represents a set of arguments accepted by OpenGLWindow::size_increment method.

OpenGLWindowStartTimerArgs
This trait represents a set of arguments accepted by OpenGLWindow::start_timer method.

OpenGLWindowTitleArgs
This trait represents a set of arguments accepted by OpenGLWindow::title method.

OpenGLWindowTrArgs
This trait represents a set of arguments accepted by OpenGLWindow::tr method.

OpenGLWindowTrUtf8Args
This trait represents a set of arguments accepted by OpenGLWindow::tr_utf8 method.

OpenGLWindowUpdateArgs
This trait represents a set of arguments accepted by OpenGLWindow::update method.