Module qt_gui::page_layout::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

PageLayoutFullRectArgs
This trait represents a set of arguments accepted by PageLayout::full_rect method.

PageLayoutFullRectPixelsArgs
This trait represents a set of arguments accepted by PageLayout::full_rect_pixels method.

PageLayoutFullRectPointsArgs
This trait represents a set of arguments accepted by PageLayout::full_rect_points method.

PageLayoutMarginsArgs
This trait represents a set of arguments accepted by PageLayout::margins method.

PageLayoutMarginsPixelsArgs
This trait represents a set of arguments accepted by PageLayout::margins_pixels method.

PageLayoutMarginsPointsArgs
This trait represents a set of arguments accepted by PageLayout::margins_points method.

PageLayoutMaximumMarginsArgs
This trait represents a set of arguments accepted by PageLayout::maximum_margins method.

PageLayoutMinimumMarginsArgs
This trait represents a set of arguments accepted by PageLayout::minimum_margins method.

PageLayoutNewArgs
This trait represents a set of arguments accepted by PageLayout::new method.

PageLayoutPageSizeArgs
This trait represents a set of arguments accepted by PageLayout::page_size method.

PageLayoutPaintRectArgs
This trait represents a set of arguments accepted by PageLayout::paint_rect method.

PageLayoutPaintRectPixelsArgs
This trait represents a set of arguments accepted by PageLayout::paint_rect_pixels method.

PageLayoutPaintRectPointsArgs
This trait represents a set of arguments accepted by PageLayout::paint_rect_points method.

PageLayoutSetPageSizeArgs
This trait represents a set of arguments accepted by PageLayout::set_page_size method.