Struct qt_gui::page_layout::PageLayout

pub struct PageLayout {
    // some fields omitted
}

C++ type: QPageLayout

C++ documentation:

Describes the size, orientation and margins of a page.

The QPageLayout class defines the layout of a page in a paged document, with the page size, orientation and margins able to be set and the full page and paintable page rectangles defined by those attributes able to be queried in a variety of units.

The page size is defined by the QPageSize class which can be queried for page size attributes. Note that the QPageSize itself is always defined in a Portrait orientation.

The minimum margins can be defined for the layout but normally default to 0. When used in conjunction with Qt's printing support the minimum margins will reflect the minimum printable area defined by the printer.

In the default StandardMode the current margins and minimum margins are always taken into account. The paintable rectangle is the full page rectangle less the current margins, and the current margins can only be set to values between the minimum margins and the maximum margins allowed by the full page size.

In FullPageMode the current margins and minimum margins are not taken into account. The paintable rectangle is the full page rectangle, and the current margins can be set to any values regardless of the minimum margins and page size.

Methods

impl PageLayout

fn full_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutFullRectArgs<'largs>

C++ method: QPageLayout::fullRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn full_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn full_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QPageLayout::fullRect() const

C++ documentation:

Returns the full page rectangle in the current layout units.

The page rectangle takes into account the page size and page orientation, but not the page margins.

See also paintRect() and units().

Variant 2

Rust arguments:
1) fn full_rect(&self, (::page_layout::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn full_rect(&self, (::page_layout::Unit, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QPageLayout::fullRect(QPageLayout::Unit units) const

C++ documentation:

Returns the full page rectangle in the required units.

The page rectangle takes into account the page size and page orientation, but not the page margins.

See also paintRect().

fn full_rect_pixels<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutFullRectPixelsArgs<'largs>

C++ method: QPageLayout::fullRectPixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn full_rect_pixels(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn full_rect_pixels(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QPageLayout::fullRectPixels(int resolution) const

C++ documentation:

Returns the full page rectangle in device pixels for the given resolution.

The page rectangle takes into account the page size and page orientation, but not the page margins.

See also paintRect().

fn full_rect_points<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutFullRectPointsArgs<'largs>

C++ method: QPageLayout::fullRectPoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn full_rect_points(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn full_rect_points(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPageLayout::fullRectPoints() const

C++ documentation:

Returns the full page rectangle in Postscript Points (1/72 of an inch).

The page rectangle takes into account the page size and page orientation, but not the page margins.

See also paintRect().

fn is_equivalent_to(&self, other: &PageLayout) -> bool

C++ method: bool QPageLayout::isEquivalentTo(const QPageLayout& other) const

C++ documentation:

Returns true if this page layout is equivalent to the other page layout, i.e. if the page has the same size, margins and orientation.

fn is_valid(&self) -> bool

C++ method: bool QPageLayout::isValid() const

C++ documentation:

Returns true if this page layout is valid.

fn margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutMarginsArgs<'largs>

C++ method: QPageLayout::margins

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::margins_f::MarginsF>
2) fn margins(&self, cpp_utils::AsStruct) -> qt_core::margins_f::MarginsF
C++ method: QMarginsF QPageLayout::margins() const

C++ documentation:

Returns the margins of the page layout using the currently set units.

See also setMargins() and units().

Variant 2

Rust arguments:
1) fn margins(&self, (::page_layout::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::margins_f::MarginsF>
2) fn margins(&self, (::page_layout::Unit, cpp_utils::AsStruct)) -> qt_core::margins_f::MarginsF
C++ method: QMarginsF QPageLayout::margins(QPageLayout::Unit units) const

C++ documentation:

Returns the margins of the page layout using the requested units.

See also setMargins() and margins().

fn margins_pixels<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutMarginsPixelsArgs<'largs>

C++ method: QPageLayout::marginsPixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn margins_pixels(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::margins::Margins>
2) fn margins_pixels(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::margins::Margins
C++ method: QMargins QPageLayout::marginsPixels(int resolution) const

C++ documentation:

Returns the margins of the page layout in device pixels for the given resolution.

See also setMargins().

fn margins_points<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutMarginsPointsArgs<'largs>

C++ method: QPageLayout::marginsPoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn margins_points(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::margins::Margins>
2) fn margins_points(&self, cpp_utils::AsStruct) -> qt_core::margins::Margins
C++ method: QMargins QPageLayout::marginsPoints() const

C++ documentation:

Returns the margins of the page layout in Postscript Points (1/72 of an inch).

See also setMargins() and margins().

fn maximum_margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutMaximumMarginsArgs<'largs>

C++ method: QPageLayout::maximumMargins

This is an overloaded function. Available variants:

Rust arguments:
1) fn maximum_margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::margins_f::MarginsF>
2) fn maximum_margins(&self, cpp_utils::AsStruct) -> qt_core::margins_f::MarginsF
C++ method: QMarginsF QPageLayout::maximumMargins() const

C++ documentation:

Returns the maximum margins that would be applied if the page layout was in StandardMode.

The maximum margins allowed are calculated as the full size of the page minus the minimum margins set. For example, if the page width is 100 points and the minimum right margin is 10 points, then the maximum left margin will be 90 points.

See also setMinimumMargins() and minimumMargins().

fn minimum_margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutMinimumMarginsArgs<'largs>

C++ method: QPageLayout::minimumMargins

This is an overloaded function. Available variants:

Rust arguments:
1) fn minimum_margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::margins_f::MarginsF>
2) fn minimum_margins(&self, cpp_utils::AsStruct) -> qt_core::margins_f::MarginsF
C++ method: QMarginsF QPageLayout::minimumMargins() const

C++ documentation:

Returns the minimum margins of the page layout.

See also setMinimumMargins() and maximumMargins().

fn mode(&self) -> Mode

C++ method: QPageLayout::Mode QPageLayout::mode() const

C++ documentation:

Returns the page layout mode.

See also setMode().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PageLayoutNewArgs

C++ method: QPageLayout::QPageLayout

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::page_layout::PageLayout
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::page_layout::PageLayout>
C++ method: [constructor] void QPageLayout::QPageLayout()

Variant 2

Rust arguments:
1) fn new((&::page_layout::PageLayout, cpp_utils::AsStruct)) -> ::page_layout::PageLayout
2) fn new((&::page_layout::PageLayout, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_layout::PageLayout>
C++ method: [constructor] void QPageLayout::QPageLayout(const QPageLayout& other)

Variant 3

Rust arguments:
1) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, cpp_utils::AsStruct)) -> ::page_layout::PageLayout
2) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, ::page_layout::Unit, cpp_utils::AsStruct)) -> ::page_layout::PageLayout
3) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, ::page_layout::Unit, &qt_core::margins_f::MarginsF, cpp_utils::AsStruct)) -> ::page_layout::PageLayout
4) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_layout::PageLayout>
5) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, ::page_layout::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_layout::PageLayout>
6) fn new((&::page_size::PageSize, ::page_layout::Orientation, &qt_core::margins_f::MarginsF, ::page_layout::Unit, &qt_core::margins_f::MarginsF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_layout::PageLayout>
C++ method: [constructor] void QPageLayout::QPageLayout(const QPageSize& pageSize, QPageLayout::Orientation orientation, const QMarginsF& margins, QPageLayout::Unit units = ?, const QMarginsF& minMargins = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 PageLayout) -> &'l0 mut PageLayout

C++ method: QPageLayout& QPageLayout::operator=(const QPageLayout& other)

C++ documentation:

Assignment operator, assigns other to this.

fn orientation(&self) -> Orientation

C++ method: QPageLayout::Orientation QPageLayout::orientation() const

C++ documentation:

Returns the page orientation of the page layout.

See also setOrientation().

fn page_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutPageSizeArgs<'largs>

C++ method: QPageLayout::pageSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn page_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::page_size::PageSize>
2) fn page_size(&self, cpp_utils::AsStruct) -> ::page_size::PageSize
C++ method: QPageSize QPageLayout::pageSize() const

C++ documentation:

Returns the page size of the page layout.

Note that the QPageSize is always defined in a Portrait orientation. To obtain a size that takes the set orientation into account you must use fullRect().

See also setPageSize().

fn paint_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutPaintRectArgs<'largs>

C++ method: QPageLayout::paintRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn paint_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn paint_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QPageLayout::paintRect() const

C++ documentation:

Returns the page rectangle in the current layout units.

The paintable rectangle takes into account the page size, orientation and margins.

If the FullPageMode mode is set then the fullRect() is returned and the margins must be manually managed.

Variant 2

Rust arguments:
1) fn paint_rect(&self, (::page_layout::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn paint_rect(&self, (::page_layout::Unit, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QPageLayout::paintRect(QPageLayout::Unit units) const

C++ documentation:

Returns the page rectangle in the required units.

The paintable rectangle takes into account the page size, orientation and margins.

If the FullPageMode mode is set then the fullRect() is returned and the margins must be manually managed.

fn paint_rect_pixels<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutPaintRectPixelsArgs<'largs>

C++ method: QPageLayout::paintRectPixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn paint_rect_pixels(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn paint_rect_pixels(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QPageLayout::paintRectPixels(int resolution) const

C++ documentation:

Returns the paintable rectangle in rounded device pixels for the given resolution.

The paintable rectangle takes into account the page size, orientation and margins.

If the FullPageMode mode is set then the fullRect() is returned and the margins must be manually managed.

fn paint_rect_points<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageLayoutPaintRectPointsArgs<'largs>

C++ method: QPageLayout::paintRectPoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn paint_rect_points(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn paint_rect_points(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPageLayout::paintRectPoints() const

C++ documentation:

Returns the paintable rectangle in rounded Postscript Points (1/72 of an inch).

The paintable rectangle takes into account the page size, orientation and margins.

If the FullPageMode mode is set then the fullRect() is returned and the margins must be manually managed.

fn set_bottom_margin(&mut self, bottom_margin: c_double) -> bool

C++ method: bool QPageLayout::setBottomMargin(double bottomMargin)

C++ documentation:

Sets the bottom page margin of the page layout to bottomMargin. Returns true if the margin was successfully set.

The units used are those currently defined for the layout. To use different units call setUnits() first.

If in the default StandardMode then the new margin must fall between the minimum margin set and the maximum margin allowed by the page size, otherwise the margin will not be set.

If in FullPageMode then any margin values will be accepted.

See also setMargins() and margins().

fn set_left_margin(&mut self, left_margin: c_double) -> bool

C++ method: bool QPageLayout::setLeftMargin(double leftMargin)

C++ documentation:

Sets the left page margin of the page layout to leftMargin. Returns true if the margin was successfully set.

The units used are those currently defined for the layout. To use different units call setUnits() first.

If in the default StandardMode then the new margin must fall between the minimum margin set and the maximum margin allowed by the page size, otherwise the margin will not be set.

If in FullPageMode then any margin values will be accepted.

See also setMargins() and margins().

fn set_margins(&mut self, margins: &MarginsF) -> bool

C++ method: bool QPageLayout::setMargins(const QMarginsF& margins)

C++ documentation:

Sets the page margins of the page layout to margins Returns true if the margins were successfully set.

The units used are those currently defined for the layout. To use different units then call setUnits() first.

If in the default StandardMode then all the new margins must fall between the minimum margins set and the maximum margins allowed by the page size, otherwise the margins will not be set.

If in FullPageMode then any margin values will be accepted.

See also margins() and units().

fn set_minimum_margins(&mut self, min_margins: &MarginsF)

C++ method: void QPageLayout::setMinimumMargins(const QMarginsF& minMargins)

C++ documentation:

Sets the minimum page margins of the page layout to minMargins.

It is not recommended to override the default values set for a page size as this may be the minimum printable area for a physical print device.

If the StandardMode mode is set then the existing margins will be clamped to the new minMargins and the maximum allowed by the page size. If the FullPageMode is set then the existing margins will be unchanged.

See also minimumMargins() and setMargins().

fn set_mode(&mut self, mode: Mode)

C++ method: void QPageLayout::setMode(QPageLayout::Mode mode)

C++ documentation:

Sets a page layout mode to mode.

See also mode().

fn set_orientation(&mut self, orientation: Orientation)

C++ method: void QPageLayout::setOrientation(QPageLayout::Orientation orientation)

C++ documentation:

Sets the page orientation of the page layout to orientation.

Changing the orientation does not affect the current margins or the minimum margins.

See also orientation().

fn set_page_size<'largs, Args>(&'largs mut self, args: Args) where Args: PageLayoutSetPageSizeArgs<'largs>

C++ method: QPageLayout::setPageSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_page_size(&mut self, &::page_size::PageSize) -> ()
2) fn set_page_size(&mut self, (&::page_size::PageSize, &qt_core::margins_f::MarginsF)) -> ()
C++ method: void QPageLayout::setPageSize(const QPageSize& pageSize, const QMarginsF& minMargins = ?)

C++ documentation:

Sets the page size of the page layout to pageSize.

Optionally define the minimum allowed margins minMargins, e.g. the minimum margins able to be printed by a physical print device, otherwise the minimum margins will default to 0.

If StandardMode is set then the existing margins will be clamped to the new minimum margins and the maximum margins allowed by the page size. If FullPageMode is set then the existing margins will be unchanged.

See also pageSize().

fn set_right_margin(&mut self, right_margin: c_double) -> bool

C++ method: bool QPageLayout::setRightMargin(double rightMargin)

C++ documentation:

Sets the right page margin of the page layout to rightMargin. Returns true if the margin was successfully set.

The units used are those currently defined for the layout. To use different units call setUnits() first.

If in the default StandardMode then the new margin must fall between the minimum margin set and the maximum margin allowed by the page size, otherwise the margin will not be set.

If in FullPageMode then any margin values will be accepted.

See also setMargins() and margins().

fn set_top_margin(&mut self, top_margin: c_double) -> bool

C++ method: bool QPageLayout::setTopMargin(double topMargin)

C++ documentation:

Sets the top page margin of the page layout to topMargin. Returns true if the margin was successfully set.

The units used are those currently defined for the layout. To use different units call setUnits() first.

If in the default StandardMode then the new margin must fall between the minimum margin set and the maximum margin allowed by the page size, otherwise the margin will not be set.

If in FullPageMode then any margin values will be accepted.

See also setMargins() and margins().

fn set_units(&mut self, units: Unit)

C++ method: void QPageLayout::setUnits(QPageLayout::Unit units)

C++ documentation:

Sets the units used to define the page layout.

See also units().

fn swap(&mut self, other: &mut PageLayout)

C++ method: void QPageLayout::swap(QPageLayout& other)

C++ documentation:

Swaps this page layout with other. This function is very fast and never fails.

fn units(&self) -> Unit

C++ method: QPageLayout::Unit QPageLayout::units() const

C++ documentation:

Returns the units the page layout is currently defined in.

See also setUnits().