Enum qt_gui::page_size::PageSizeId

pub enum PageSizeId {
    A4,
    B5,
    Letter,
    Legal,
    Executive,
    A0,
    A1,
    A2,
    A3,
    A5,
    A6,
    A7,
    A8,
    A9,
    B0,
    B1,
    B10,
    B2,
    B3,
    B4,
    B6,
    B7,
    B8,
    B9,
    C5E,
    Comm10E,
    Dle,
    Folio,
    Ledger,
    Tabloid,
    Custom,
    A10,
    A3Extra,
    A4Extra,
    A4Plus,
    A4Small,
    A5Extra,
    B5Extra,
    JisB0,
    JisB1,
    JisB2,
    JisB3,
    JisB4,
    JisB5,
    JisB6,
    JisB7,
    JisB8,
    JisB9,
    JisB10,
    AnsiC,
    AnsiD,
    AnsiE,
    LegalExtra,
    LetterExtra,
    LetterPlus,
    LetterSmall,
    TabloidExtra,
    ArchA,
    ArchB,
    ArchC,
    ArchD,
    ArchE,
    Imperial7x9,
    Imperial8x10,
    Imperial9x11,
    Imperial9x12,
    Imperial10x11,
    Imperial10x13,
    Imperial10x14,
    Imperial12x11,
    Imperial15x11,
    ExecutiveStandard,
    Note,
    Quarto,
    Statement,
    SuperA,
    SuperB,
    Postcard,
    DoublePostcard,
    Prc16K,
    Prc32K,
    Prc32KBig,
    FanFoldUS,
    FanFoldGerman,
    FanFoldGermanLegal,
    EnvelopeB4,
    EnvelopeB5,
    EnvelopeB6,
    EnvelopeC0,
    EnvelopeC1,
    EnvelopeC2,
    EnvelopeC3,
    EnvelopeC4,
    EnvelopeC6,
    EnvelopeC65,
    EnvelopeC7,
    Envelope9,
    Envelope11,
    Envelope12,
    Envelope14,
    EnvelopeMonarch,
    EnvelopePersonal,
    EnvelopeChou3,
    EnvelopeChou4,
    EnvelopeInvite,
    EnvelopeItalian,
    EnvelopeKaku2,
    EnvelopeKaku3,
    EnvelopePrc1,
    EnvelopePrc2,
    EnvelopePrc3,
    EnvelopePrc4,
    EnvelopePrc5,
    EnvelopePrc6,
    EnvelopePrc7,
    EnvelopePrc8,
    EnvelopePrc9,
    EnvelopePrc10,
    EnvelopeYou4,
}

C++ type: QPageSize::PageSizeId

C++ documentation:

This enum type lists the available page sizes as defined in the Postscript PPD standard. These values are duplicated in QPagedPaintDevice and QPrinter.

The defined sizes are:

Due to historic reasons QPageSize::Executive is not the same as the standard Postscript and Windows Executive size, use QPageSize::ExecutiveStandard instead.

The Postscript standard size QPageSize::Folio is different to the Windows DMPAPER_FOLIO size, use the Postscript standard size QPageSize::FanFoldGermanLegal if needed.

Variants
A4

210 x 297 mm, 8.26 x 11.69 inches (C++ enum variant: A4 = 0)

B5

176 x 250 mm, 6.93 x 9.84 inches (C++ enum variant: B5 = 1)

Letter

This variant corresponds to multiple C++ enum variants with the same value:

Letter = 2: 8.5 x 11 inches, 215.9 x 279.4 mm

AnsiA = 2: = Letter

Legal

8.5 x 14 inches, 215.9 x 355.6 mm (C++ enum variant: Legal = 3)

Executive

7.5 x 10 inches, 190.5 x 254 mm (C++ enum variant: Executive = 4)

A0

841 x 1189 mm (C++ enum variant: A0 = 5)

A1

594 x 841 mm (C++ enum variant: A1 = 6)

A2

420 x 594 mm (C++ enum variant: A2 = 7)

A3

297 x 420 mm (C++ enum variant: A3 = 8)

A5

148 x 210 mm (C++ enum variant: A5 = 9)

A6

105 x 148 mm (C++ enum variant: A6 = 10)

A7

74 x 105 mm (C++ enum variant: A7 = 11)

A8

52 x 74 mm (C++ enum variant: A8 = 12)

A9

37 x 52 mm (C++ enum variant: A9 = 13)

B0

1000 x 1414 mm (C++ enum variant: B0 = 14)

B1

707 x 1000 mm (C++ enum variant: B1 = 15)

B10

31 x 44 mm (C++ enum variant: B10 = 16)

B2

500 x 707 mm (C++ enum variant: B2 = 17)

B3

353 x 500 mm (C++ enum variant: B3 = 18)

B4

250 x 353 mm (C++ enum variant: B4 = 19)

B6

125 x 176 mm (C++ enum variant: B6 = 20)

B7

88 x 125 mm (C++ enum variant: B7 = 21)

B8

62 x 88 mm (C++ enum variant: B8 = 22)

B9

44 x 62 mm (C++ enum variant: B9 = 23)

C5E

This variant corresponds to multiple C++ enum variants with the same value:

C5E = 24: 163 x 229 mm

EnvelopeC5 = 24: = C5E

Comm10E

This variant corresponds to multiple C++ enum variants with the same value:

Comm10E = 25: 105 x 241 mm, U.S. Common 10 Envelope

Envelope10 = 25: = Comm10E

Dle

This variant corresponds to multiple C++ enum variants with the same value:

DLE = 26: 110 x 220 mm

EnvelopeDL = 26: = DLE

Folio

210 x 330 mm (C++ enum variant: Folio = 27)

Ledger

This variant corresponds to multiple C++ enum variants with the same value:

Ledger = 28: 431.8 x 279.4 mm

AnsiB = 28: = Ledger

Tabloid

279.4 x 431.8 mm (C++ enum variant: Tabloid = 29)

Custom

Unknown, or a user defined size. (C++ enum variant: Custom = 30)

A10

(C++ enum variant: A10 = 31)

A3Extra

(C++ enum variant: A3Extra = 32)

A4Extra

(C++ enum variant: A4Extra = 33)

A4Plus

(C++ enum variant: A4Plus = 34)

A4Small

(C++ enum variant: A4Small = 35)

A5Extra

(C++ enum variant: A5Extra = 36)

B5Extra

(C++ enum variant: B5Extra = 37)

JisB0

(C++ enum variant: JisB0 = 38)

JisB1

(C++ enum variant: JisB1 = 39)

JisB2

(C++ enum variant: JisB2 = 40)

JisB3

(C++ enum variant: JisB3 = 41)

JisB4

(C++ enum variant: JisB4 = 42)

JisB5

(C++ enum variant: JisB5 = 43)

JisB6

, (C++ enum variant: JisB6 = 44)

JisB7

(C++ enum variant: JisB7 = 45)

JisB8

(C++ enum variant: JisB8 = 46)

JisB9

(C++ enum variant: JisB9 = 47)

JisB10

(C++ enum variant: JisB10 = 48)

AnsiC

(C++ enum variant: AnsiC = 49)

AnsiD

(C++ enum variant: AnsiD = 50)

AnsiE

(C++ enum variant: AnsiE = 51)

LegalExtra

(C++ enum variant: LegalExtra = 52)

LetterExtra

(C++ enum variant: LetterExtra = 53)

LetterPlus

(C++ enum variant: LetterPlus = 54)

LetterSmall

(C++ enum variant: LetterSmall = 55)

TabloidExtra

(C++ enum variant: TabloidExtra = 56)

ArchA

(C++ enum variant: ArchA = 57)

ArchB

(C++ enum variant: ArchB = 58)

ArchC

(C++ enum variant: ArchC = 59)

ArchD

(C++ enum variant: ArchD = 60)

ArchE

(C++ enum variant: ArchE = 61)

Imperial7x9

(C++ enum variant: Imperial7x9 = 62)

Imperial8x10

(C++ enum variant: Imperial8x10 = 63)

Imperial9x11

(C++ enum variant: Imperial9x11 = 64)

Imperial9x12

(C++ enum variant: Imperial9x12 = 65)

Imperial10x11

(C++ enum variant: Imperial10x11 = 66)

Imperial10x13

(C++ enum variant: Imperial10x13 = 67)

Imperial10x14

(C++ enum variant: Imperial10x14 = 68)

Imperial12x11

(C++ enum variant: Imperial12x11 = 69)

Imperial15x11

(C++ enum variant: Imperial15x11 = 70)

ExecutiveStandard

(C++ enum variant: ExecutiveStandard = 71)

Note

(C++ enum variant: Note = 72)

Quarto

(C++ enum variant: Quarto = 73)

Statement

(C++ enum variant: Statement = 74)

SuperA

(C++ enum variant: SuperA = 75)

SuperB

(C++ enum variant: SuperB = 76)

Postcard

(C++ enum variant: Postcard = 77)

DoublePostcard

(C++ enum variant: DoublePostcard = 78)

Prc16K

(C++ enum variant: Prc16K = 79)

Prc32K

(C++ enum variant: Prc32K = 80)

Prc32KBig

(C++ enum variant: Prc32KBig = 81)

FanFoldUS

(C++ enum variant: FanFoldUS = 82)

FanFoldGerman

(C++ enum variant: FanFoldGerman = 83)

FanFoldGermanLegal

(C++ enum variant: FanFoldGermanLegal = 84)

EnvelopeB4

(C++ enum variant: EnvelopeB4 = 85)

EnvelopeB5

(C++ enum variant: EnvelopeB5 = 86)

EnvelopeB6

(C++ enum variant: EnvelopeB6 = 87)

EnvelopeC0

(C++ enum variant: EnvelopeC0 = 88)

EnvelopeC1

(C++ enum variant: EnvelopeC1 = 89)

EnvelopeC2

(C++ enum variant: EnvelopeC2 = 90)

EnvelopeC3

(C++ enum variant: EnvelopeC3 = 91)

EnvelopeC4

(C++ enum variant: EnvelopeC4 = 92)

EnvelopeC6

(C++ enum variant: EnvelopeC6 = 93)

EnvelopeC65

(C++ enum variant: EnvelopeC65 = 94)

EnvelopeC7

(C++ enum variant: EnvelopeC7 = 95)

Envelope9

(C++ enum variant: Envelope9 = 96)

Envelope11

(C++ enum variant: Envelope11 = 97)

Envelope12

(C++ enum variant: Envelope12 = 98)

Envelope14

(C++ enum variant: Envelope14 = 99)

EnvelopeMonarch

(C++ enum variant: EnvelopeMonarch = 100)

EnvelopePersonal

(C++ enum variant: EnvelopePersonal = 101)

EnvelopeChou3

(C++ enum variant: EnvelopeChou3 = 102)

EnvelopeChou4

(C++ enum variant: EnvelopeChou4 = 103)

EnvelopeInvite

(C++ enum variant: EnvelopeInvite = 104)

EnvelopeItalian

(C++ enum variant: EnvelopeItalian = 105)

EnvelopeKaku2

(C++ enum variant: EnvelopeKaku2 = 106)

EnvelopeKaku3

(C++ enum variant: EnvelopeKaku3 = 107)

EnvelopePrc1

(C++ enum variant: EnvelopePrc1 = 108)

EnvelopePrc2

(C++ enum variant: EnvelopePrc2 = 109)

EnvelopePrc3

(C++ enum variant: EnvelopePrc3 = 110)

EnvelopePrc4

(C++ enum variant: EnvelopePrc4 = 111)

EnvelopePrc5

(C++ enum variant: EnvelopePrc5 = 112)

EnvelopePrc6

(C++ enum variant: EnvelopePrc6 = 113)

EnvelopePrc7

(C++ enum variant: EnvelopePrc7 = 114)

EnvelopePrc8

(C++ enum variant: EnvelopePrc8 = 115)

EnvelopePrc9

(C++ enum variant: EnvelopePrc9 = 116)

EnvelopePrc10

(C++ enum variant: EnvelopePrc10 = 117)

EnvelopeYou4

This variant corresponds to multiple C++ enum variants with the same value:

EnvelopeYou4 = 118:

LastPageSize = 118: = EnvelopeYou4

NPageSize = 118

NPaperSize = 118