Module qt_gui::page_size::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

PageSizeDefinitionSizeArgs
This trait represents a set of arguments accepted by PageSize::definition_size method.

PageSizeDefinitionSizeStaticArgs
This trait represents a set of arguments accepted by PageSize::definition_size_static method.

PageSizeIdStaticArgs
This trait represents a set of arguments accepted by PageSize::id_static method.

PageSizeKeyArgs
This trait represents a set of arguments accepted by PageSize::key method.

PageSizeKeyStaticArgs
This trait represents a set of arguments accepted by PageSize::key_static method.

PageSizeNameArgs
This trait represents a set of arguments accepted by PageSize::name method.

PageSizeNameStaticArgs
This trait represents a set of arguments accepted by PageSize::name_static method.

PageSizeNewArgs
This trait represents a set of arguments accepted by PageSize::new method.

PageSizeRectArgs
This trait represents a set of arguments accepted by PageSize::rect method.

PageSizeRectPixelsArgs
This trait represents a set of arguments accepted by PageSize::rect_pixels method.

PageSizeRectPointsArgs
This trait represents a set of arguments accepted by PageSize::rect_points method.

PageSizeSizeArgs
This trait represents a set of arguments accepted by PageSize::size method.

PageSizeSizePixelsArgs
This trait represents a set of arguments accepted by PageSize::size_pixels method.

PageSizeSizePixelsStaticArgs
This trait represents a set of arguments accepted by PageSize::size_pixels_static method.

PageSizeSizePointsArgs
This trait represents a set of arguments accepted by PageSize::size_points method.

PageSizeSizePointsStaticArgs
This trait represents a set of arguments accepted by PageSize::size_points_static method.

PageSizeSizeStaticArgs
This trait represents a set of arguments accepted by PageSize::size_static method.