Struct qt_gui::page_size::PageSize

pub struct PageSize {
    // some fields omitted
}

C++ type: QPageSize

C++ documentation:

The QPageSize class describes the size and name of a defined page size

This class implements support for the set of standard page sizes as defined in the Adobe Postscript PPD Standard v4.3. It defines the standard set of page sizes in points, millimeters and inches and ensures these sizes are consistently used. Other size units can be used but will be calculated results and so may not always be consistent. The defined point sizes are always a integer, all other sizes can be fractions of a unit.

The defined size is always in width x height order with no implied page orientation. Note that it is possible for page sizes to be defined where the width is greater than the height, such as QPageSize::Ledger, so you cannot rely on comparing the width and height values to determine page orientation.

For example, A4 is defined by the standard as 210mm x 297mm, 8.27in x 11.69in, or 595pt x 842pt.

You can also define custom page sizes with custom names in any units you want and this unit size will be preserved and used as the base for all other unit size calculations.

When creating a QPageSize using a custom QSize you can choose if you want QPageSize to try match the size to a standard page size. By default QPaperSize uses a FuzzyMatch mode where it will match a given page size to a standard page size if it falls within 3 postscript points of a defined standard size. You can override this to request only an exact match but this is not recommended as conversions between units can easily lose 3 points and result in incorrect page sizes.

A QPageSize instance may also be obtained by querying the supported page sizes for a print device. In this case the localized name returned is that defined by the printer itself. Note that the print device may not support the current default locale language.

The class also provides convenience methods for converting page size IDs to and from various unit sizes.

Methods

impl PageSize

fn definition_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeDefinitionSizeArgs<'largs>

C++ method: QPageSize::definitionSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn definition_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn definition_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QPageSize::definitionSize() const

C++ documentation:

Returns the definition size of the page size.

For a standard page size this will be the size as defined in the relevant standard, i.e. ISO A4 will be defined in millimeters while ANSI Letter will be defined in inches.

For a custom page size this will be the original size used to create the page size object.

If the QPageSize is invalid then the QSizeF will be invalid.

See also definitionUnits().

fn definition_size_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeDefinitionSizeStaticArgs

C++ method: QPageSize::definitionSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn definition_size_static((::page_size::PageSizeId, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn definition_size_static((::page_size::PageSizeId, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: static QSizeF QPageSize::definitionSize(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the definition size of the standard pageSizeId.

To obtain the definition units, call QPageSize::definitionUnits().

fn definition_units(&self) -> Unit

C++ method: QPageSize::Unit QPageSize::definitionUnits() const

C++ documentation:

Returns the definition units of the page size.

For a standard page size this will be the units as defined in the relevant standard, i.e. ISO A4 will be defined in millimeters while ANSI Letter will be defined in inches.

For a custom page size this will be the original units used to create the page size object.

If the QPageSize is invalid then the QPageSize::Unit will be invalid.

See also definitionSize().

fn definition_units_static(page_size_id: PageSizeId) -> Unit

C++ method: static QPageSize::Unit QPageSize::definitionUnits(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the definition units of the standard pageSizeId.

To obtain the definition size, call QPageSize::definitionSize().

fn id(&self) -> PageSizeId

C++ method: QPageSize::PageSizeId QPageSize::id() const

C++ documentation:

Returns the standard QPageSize::PageSizeId of the page, or QPageSize::Custom.

If the QPageSize is invalid then the ID will be QPageSize::Custom.

fn id_static<Args>(args: Args) -> PageSizeId where Args: PageSizeIdStaticArgs

C++ method: QPageSize::id

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn id_static(&qt_core::size::Size) -> ::page_size::PageSizeId
2) fn id_static((&qt_core::size::Size, ::page_size::SizeMatchPolicy)) -> ::page_size::PageSizeId
C++ method: static QPageSize::PageSizeId QPageSize::id(const QSize& pointSize, QPageSize::SizeMatchPolicy matchPolicy = ?)

C++ documentation:

Returns the standard QPageSize::PageSizeId of the given pointSize in points using the given matchPolicy.

If using FuzzyMatch then the point size of the PageSizeId returned may not exactly match the pointSize you passed in. You should call QPageSize::sizePoints() using the returned PageSizeId to find out the actual point size of the PageSizeId before using it in any calculations.

Variant 2

Rust arguments:
1) fn id_static((&qt_core::size_f::SizeF, ::page_size::Unit)) -> ::page_size::PageSizeId
2) fn id_static((&qt_core::size_f::SizeF, ::page_size::Unit, ::page_size::SizeMatchPolicy)) -> ::page_size::PageSizeId
C++ method: static QPageSize::PageSizeId QPageSize::id(const QSizeF& size, QPageSize::Unit units, QPageSize::SizeMatchPolicy matchPolicy = ?)

C++ documentation:

Returns the standard QPageSize::PageSizeId of the given size in units using the given matchPolicy.

If using FuzzyMatch then the unit size of the PageSizeId returned may not exactly match the size you passed in. You should call QPageSize::size() using the returned PageSizeId to find out the actual unit size of the PageSizeId before using it in any calculations.

Variant 3

Rust arguments: fn id_static(libc::c_int) -> ::page_size::PageSizeId
C++ method: static QPageSize::PageSizeId QPageSize::id(int windowsId)

C++ documentation:

Returns the PageSizeId for the given Windows DMPAPER enum value windowsId.

If there is no matching PageSizeId then QPageSize::Custom is returned.

fn is_equivalent_to(&self, other: &PageSize) -> bool

C++ method: bool QPageSize::isEquivalentTo(const QPageSize& other) const

C++ documentation:

Returns true if this page is equivalent to the other page, i.e. if the page has the same size regardless of other attributes like name.

fn is_valid(&self) -> bool

C++ method: bool QPageSize::isValid() const

C++ documentation:

Returns true if this page size is valid.

The page size may be invalid if created with an invalid PageSizeId, or a negative or invalid QSize or QSizeF, or the null constructor.

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeKeyArgs<'largs>

C++ method: QPageSize::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn key(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QPageSize::key() const

C++ documentation:

Returns the unique key of the page size.

By default this is the PPD standard mediaOption keyword for the page size, or the PPD custom format key. If the QPageSize instance was obtained from a print device then this will be the key provided by the print device and may differ from the standard key.

If the QPageSize is invalid then the key will be an empty string.

This key should never be shown to end users, it is an internal key only. For a human-readable name use name().

See also name().

fn key_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeKeyStaticArgs

C++ method: QPageSize::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_static((::page_size::PageSizeId, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn key_static((::page_size::PageSizeId, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QPageSize::key(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the PPD mediaOption keyword of the standard pageSizeId.

If the QPageSize is invalid then the key will be empty.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeNameArgs<'largs>

C++ method: QPageSize::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QPageSize::name() const

C++ documentation:

Returns a localized human-readable name for the page size.

If the QPageSize instance was obtained from a print device then the name used is that provided by the print device. Note that a print device may not support the current default locale language.

If the QPageSize is invalid then the name will be an empty string.

fn name_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeNameStaticArgs

C++ method: QPageSize::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name_static((::page_size::PageSizeId, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn name_static((::page_size::PageSizeId, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: static QString QPageSize::name(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the localized name of the standard pageSizeId.

If the QPageSize is invalid then the name will be empty.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PageSizeNewArgs

C++ method: QPageSize::QPageSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::page_size::PageSize
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::page_size::PageSize>
C++ method: [constructor] void QPageSize::QPageSize()

Variant 2

Rust arguments:
1) fn new((::page_size::PageSizeId, cpp_utils::AsStruct)) -> ::page_size::PageSize
2) fn new((::page_size::PageSizeId, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
C++ method: [constructor] void QPageSize::QPageSize(QPageSize::PageSizeId pageSizeId)

Variant 3

Rust arguments:
1) fn new((&::page_size::PageSize, cpp_utils::AsStruct)) -> ::page_size::PageSize
2) fn new((&::page_size::PageSize, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
C++ method: [constructor] void QPageSize::QPageSize(const QPageSize& other)

Variant 4

Rust arguments:
1) fn new((&qt_core::size::Size, cpp_utils::AsStruct)) -> ::page_size::PageSize
2) fn new((&qt_core::size::Size, &qt_core::string::String, cpp_utils::AsStruct)) -> ::page_size::PageSize
3) fn new((&qt_core::size::Size, &qt_core::string::String, ::page_size::SizeMatchPolicy, cpp_utils::AsStruct)) -> ::page_size::PageSize
4) fn new((&qt_core::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
5) fn new((&qt_core::size::Size, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
6) fn new((&qt_core::size::Size, &qt_core::string::String, ::page_size::SizeMatchPolicy, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
C++ method: [constructor] void QPageSize::QPageSize(const QSize& pointSize, const QString& name = ?, QPageSize::SizeMatchPolicy matchPolicy = ?)

Variant 5

Rust arguments:
1) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, cpp_utils::AsStruct)) -> ::page_size::PageSize
2) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, &qt_core::string::String, cpp_utils::AsStruct)) -> ::page_size::PageSize
3) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, &qt_core::string::String, ::page_size::SizeMatchPolicy, cpp_utils::AsStruct)) -> ::page_size::PageSize
4) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
5) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
6) fn new((&qt_core::size_f::SizeF, ::page_size::Unit, &qt_core::string::String, ::page_size::SizeMatchPolicy, cpp_utils::AsBox)) -> cpp_utils::CppBox<::page_size::PageSize>
C++ method: [constructor] void QPageSize::QPageSize(const QSizeF& size, QPageSize::Unit units, const QString& name = ?, QPageSize::SizeMatchPolicy matchPolicy = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 PageSize) -> &'l0 mut PageSize

C++ method: QPageSize& QPageSize::operator=(const QPageSize& other)

C++ documentation:

Assignment operator, assigns other to this.

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeRectArgs<'largs>

C++ method: QPageSize::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, (::page_size::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn rect(&self, (::page_size::Unit, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QPageSize::rect(QPageSize::Unit units) const

C++ documentation:

Returns the page rectangle in the required units.

If the QPageSize is invalid then the QRect will be invalid.

fn rect_pixels<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeRectPixelsArgs<'largs>

C++ method: QPageSize::rectPixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect_pixels(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect_pixels(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QPageSize::rectPixels(int resolution) const

C++ documentation:

Returns the page rectangle in Device Pixels at the given resolution.

If the QPageSize is invalid then the QRect will be invalid.

fn rect_points<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeRectPointsArgs<'largs>

C++ method: QPageSize::rectPoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect_points(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn rect_points(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPageSize::rectPoints() const

C++ documentation:

Returns the page rectangle in Postscript Points (1/72 of an inch).

If the QPageSize is invalid then the QRect will be invalid.

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeSizeArgs<'largs>

C++ method: QPageSize::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, (::page_size::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn size(&self, (::page_size::Unit, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: QSizeF QPageSize::size(QPageSize::Unit units) const

C++ documentation:

Returns the size of the page in the required units.

If the QPageSize is invalid then the QSizeF will be invalid.

fn size_pixels<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeSizePixelsArgs<'largs>

C++ method: QPageSize::sizePixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_pixels(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_pixels(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: QSize QPageSize::sizePixels(int resolution) const

C++ documentation:

Returns the size of the page in Device Pixels at the given resolution.

If the QPageSize is invalid then the QSize will be invalid.

fn size_pixels_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeSizePixelsStaticArgs

C++ method: QPageSize::sizePixels

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_pixels_static((::page_size::PageSizeId, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_pixels_static((::page_size::PageSizeId, libc::c_int, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: static QSize QPageSize::sizePixels(QPageSize::PageSizeId pageSizeId, int resolution)

C++ documentation:

Returns the size of the standard pageSizeId in Device Pixels for the given resolution.

fn size_points<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PageSizeSizePointsArgs<'largs>

C++ method: QPageSize::sizePoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_points(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_points(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QPageSize::sizePoints() const

C++ documentation:

Returns the size of the page in Postscript Points (1/72 of an inch).

If the QPageSize is invalid then the QSize will be invalid.

fn size_points_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeSizePointsStaticArgs

C++ method: QPageSize::sizePoints

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_points_static((::page_size::PageSizeId, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_points_static((::page_size::PageSizeId, cpp_utils::AsStruct)) -> qt_core::size::Size
C++ method: static QSize QPageSize::sizePoints(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the size of the standard pageSizeId in Points.

fn size_static<Args>(args: Args) -> Args::ReturnType where Args: PageSizeSizeStaticArgs

C++ method: QPageSize::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_static((::page_size::PageSizeId, ::page_size::Unit, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn size_static((::page_size::PageSizeId, ::page_size::Unit, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: static QSizeF QPageSize::size(QPageSize::PageSizeId pageSizeId, QPageSize::Unit units)

C++ documentation:

Returns the size of the standard pageSizeId in the requested units.

fn swap(&mut self, other: &mut PageSize)

C++ method: void QPageSize::swap(QPageSize& other)

C++ documentation:

Swaps this QPageSize with other. This function is very fast and never fails.

fn windows_id(&self) -> c_int

C++ method: int QPageSize::windowsId() const

C++ documentation:

Returns the Windows DMPAPER enum value for the page size.

Not all valid PPD page sizes have a Windows equivalent, in which case 0 will be returned.

If the QPageSize is invalid then the Windows ID will be 0.

See also id().

fn windows_id_static(page_size_id: PageSizeId) -> c_int

C++ method: static int QPageSize::windowsId(QPageSize::PageSizeId pageSizeId)

C++ documentation:

Returns the Windows DMPAPER enum value of the standard pageSizeId.

Not all valid PPD page sizes have a Windows equivalent, in which case 0 will be returned.