Struct qt_gui::paged_paint_device::PagedPaintDevice

pub struct PagedPaintDevice {
    // some fields omitted
}

C++ type: QPagedPaintDevice

C++ documentation:

The QPagedPaintDevice class is a represents a paintdevice that supports multiple pages.

Paged paint devices are used to generate output for printing or for formats like PDF. QPdfWriter and QPrinter inherit from it.

Methods

impl PagedPaintDevice

fn color_count(&self) -> c_int

C++ method: int QPagedPaintDevice::colorCount() const

C++ documentation:

Returns the number of different colors available for the paint device. If the number of colors available is too great to be represented by the int data type, then INT_MAX will be returned instead.

fn depth(&self) -> c_int

C++ method: int QPagedPaintDevice::depth() const

C++ documentation:

Returns the bit depth (number of bit planes) of the paint device.

fn dev_type(&self) -> c_int

C++ method: virtual int QPagedPaintDevice::devType() const

fn device_pixel_ratio(&self) -> c_int

C++ method: int QPagedPaintDevice::devicePixelRatio() const

C++ documentation:

Returns the device pixel ratio for device.

Common values are 1 for normal-dpi displays and 2 for high-dpi "retina" displays.

fn height(&self) -> c_int

C++ method: int QPagedPaintDevice::height() const

C++ documentation:

Returns the height of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also heightMM().

fn height_m_m(&self) -> c_int

C++ method: int QPagedPaintDevice::heightMM() const

C++ documentation:

Returns the height of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also height().

fn logical_dpi_x(&self) -> c_int

C++ method: int QPagedPaintDevice::logicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from widthMM().

Note that if the logicalDpiX() doesn't equal the physicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiY() and physicalDpiX().

fn logical_dpi_y(&self) -> c_int

C++ method: int QPagedPaintDevice::logicalDpiY() const

C++ documentation:

Returns the vertical resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from heightMM().

Note that if the logicalDpiY() doesn't equal the physicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiX() and physicalDpiY().

fn margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PagedPaintDeviceMarginsArgs<'largs>

C++ method: QPagedPaintDevice::margins

This is an overloaded function. Available variants:

Rust arguments:
1) fn margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::paged_paint_device::Margins>
2) fn margins(&self, cpp_utils::AsStruct) -> ::paged_paint_device::Margins
C++ method: QPagedPaintDevice::Margins QPagedPaintDevice::margins() const

C++ documentation:

Returns the current margins of the paint device. The default is 0.

Margins are specified in millimeters.

See also setMargins().

fn new_page(&mut self) -> bool

C++ method: pure virtual bool QPagedPaintDevice::newPage()

C++ documentation:

Starts a new page. Returns true on success.

fn page_layout<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PagedPaintDevicePageLayoutArgs<'largs>

C++ method: QPagedPaintDevice::pageLayout

This is an overloaded function. Available variants:

Rust arguments:
1) fn page_layout(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::page_layout::PageLayout>
2) fn page_layout(&self, cpp_utils::AsStruct) -> ::page_layout::PageLayout
C++ method: QPageLayout QPagedPaintDevice::pageLayout() const

C++ documentation:

Returns the current page layout. Use this method to access the current QPageSize, QPageLayout::Orientation, QMarginsF, fullRect() and paintRect().

Note that you cannot use the setters on the returned object, you must either call the individual QPagedPaintDevice setters or use setPageLayout().

This function was introduced in Qt 5.3.

See also setPageLayout(), setPageSize(), setPageOrientation(), and setPageMargins().

fn page_size(&self) -> PageSize

C++ method: QPagedPaintDevice::PageSize QPagedPaintDevice::pageSize() const

C++ documentation:

Returns the currently used page size.

See also setPageSize().

fn page_size_m_m<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PagedPaintDevicePageSizeMMArgs<'largs>

C++ method: QPagedPaintDevice::pageSizeMM

This is an overloaded function. Available variants:

Rust arguments:
1) fn page_size_m_m(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn page_size_m_m(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QPagedPaintDevice::pageSizeMM() const

C++ documentation:

Returns the page size in millimeters.

See also setPageSizeMM().

fn paint_engine(&self) -> *mut PaintEngine

C++ method: pure virtual QPaintEngine* QPagedPaintDevice::paintEngine() const

C++ documentation:

Returns a pointer to the paint engine used for drawing on the device.

fn painting_active(&self) -> bool

C++ method: bool QPagedPaintDevice::paintingActive() const

C++ documentation:

Returns true if the device is currently being painted on, i.e. someone has called QPainter::begin() but not yet called QPainter::end() for this device; otherwise returns false.

See also QPainter::isActive().

fn physical_dpi_x(&self) -> c_int

C++ method: int QPagedPaintDevice::physicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiX() doesn't equal the logicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiY() and logicalDpiX().

fn physical_dpi_y(&self) -> c_int

C++ method: int QPagedPaintDevice::physicalDpiY() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiY() doesn't equal the logicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiX() and logicalDpiY().

fn set_margins(&mut self, margins: &Margins)

C++ method: virtual void QPagedPaintDevice::setMargins(const QPagedPaintDevice::Margins& margins)

C++ documentation:

Sets the margins to be used to margins.

Margins are specified in millimeters.

The margins are purely a hint to the drawing method. They don't affect the coordinate system or clipping.

See also margins().

fn set_page_layout(&mut self, page_layout: &PageLayout) -> bool

C++ method: bool QPagedPaintDevice::setPageLayout(const QPageLayout& pageLayout)

C++ documentation:

Sets the page layout to newPageLayout.

You should call this before calling QPainter::begin(), or immediately before calling newPage() to apply the new page layout to a new page. You should not call any painting methods between a call to setPageLayout() and newPage() as the wrong paint metrics may be used.

Returns true if the page layout was successfully set to newPageLayout.

This function was introduced in Qt 5.3.

See also pageLayout().

fn set_page_margins<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: PagedPaintDeviceSetPageMarginsArgs<'largs>

C++ method: QPagedPaintDevice::setPageMargins

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_page_margins(&mut self, &qt_core::margins_f::MarginsF) -> bool
C++ method: bool QPagedPaintDevice::setPageMargins(const QMarginsF& margins)

C++ documentation:

Set the page margins in the current page layout units.

You should call this before calling QPainter::begin(), or immediately before calling newPage() to apply the new margins to a new page. You should not call any painting methods between a call to setPageMargins() and newPage() as the wrong paint metrics may be used.

To get the current page margins use pageLayout().pageMargins().

Returns true if the page margins were successfully set to margins.

This function was introduced in Qt 5.3.

See also pageLayout().

Variant 2

Rust arguments: fn set_page_margins(&mut self, (&qt_core::margins_f::MarginsF, ::page_layout::Unit)) -> bool
C++ method: bool QPagedPaintDevice::setPageMargins(const QMarginsF& margins, QPageLayout::Unit units)

C++ documentation:

Set the page margins defined in the given units.

You should call this before calling QPainter::begin(), or immediately before calling newPage() to apply the new margins to a new page. You should not call any painting methods between a call to setPageMargins() and newPage() as the wrong paint metrics may be used.

To get the current page margins use pageLayout().pageMargins().

Returns true if the page margins were successfully set to margins.

This function was introduced in Qt 5.3.

See also pageLayout().

fn set_page_orientation(&mut self, orientation: Orientation) -> bool

C++ method: bool QPagedPaintDevice::setPageOrientation(QPageLayout::Orientation orientation)

C++ documentation:

Sets the page orientation.

The page orientation is used to define the orientation of the page size when obtaining the page rect.

You should call this before calling QPainter::begin(), or immediately before calling newPage() to apply the new orientation to a new page. You should not call any painting methods between a call to setPageOrientation() and newPage() as the wrong paint metrics may be used.

To get the current QPageLayout::Orientation use pageLayout().pageOrientation().

Returns true if the page orientation was successfully set to orientation.

This function was introduced in Qt 5.3.

See also pageLayout().

fn set_page_size<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PagedPaintDeviceSetPageSizeArgs<'largs>

C++ method: QPagedPaintDevice::setPageSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_page_size(&mut self, &::page_size::PageSize) -> bool
C++ method: bool QPagedPaintDevice::setPageSize(const QPageSize& pageSize)

C++ documentation:

Sets the page size to pageSize.

To get the current QPageSize use pageLayout().pageSize().

You should call this before calling QPainter::begin(), or immediately before calling newPage() to apply the new page size to a new page. You should not call any painting methods between a call to setPageSize() and newPage() as the wrong paint metrics may be used.

Returns true if the page size was successfully set to pageSize.

This function was introduced in Qt 5.3.

See also pageSize() and pageLayout().

Variant 2

Rust arguments: fn set_page_size(&mut self, ::paged_paint_device::PageSize) -> ()
C++ method: virtual void QPagedPaintDevice::setPageSize(QPagedPaintDevice::PageSize size)

C++ documentation:

Sets the size of the a page to size.

See also setPageSizeMM().

fn set_page_size_m_m(&mut self, size: &SizeF)

C++ method: virtual void QPagedPaintDevice::setPageSizeMM(const QSizeF& size)

C++ documentation:

Sets the page size to size. size is specified in millimeters.

If the size matches a standard QPagedPaintDevice::PageSize then that page size will be used, otherwise QPagedPaintDevice::Custom will be set.

See also pageSizeMM().

fn width(&self) -> c_int

C++ method: int QPagedPaintDevice::width() const

C++ documentation:

Returns the width of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also widthMM().

fn width_m_m(&self) -> c_int

C++ method: int QPagedPaintDevice::widthMM() const

C++ documentation:

Returns the width of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also width().