Enum qt_gui::paint_engine::DirtyFlag

pub enum DirtyFlag {
    DirtyPen,
    DirtyBrush,
    DirtyBrushOrigin,
    DirtyFont,
    DirtyBackground,
    DirtyBackgroundMode,
    DirtyTransform,
    DirtyClipRegion,
    DirtyClipPath,
    DirtyHints,
    DirtyCompositionMode,
    DirtyClipEnabled,
    DirtyOpacity,
    AllDirty,
}

C++ type: QPaintEngine::DirtyFlag

C++ documentation:

These types are used by QPainter to trigger lazy updates of the various states in the QPaintEngine using QPaintEngine::updateState().

A paint engine must update every dirty state.

The DirtyFlags type is a typedef for QFlags<DirtyFlag>. It stores an OR combination of DirtyFlag values.

Variants
DirtyPen

The pen is dirty and needs to be updated. (C++ enum variant: DirtyPen = 1)

DirtyBrush

The brush is dirty and needs to be updated. (C++ enum variant: DirtyBrush = 2)

DirtyBrushOrigin

The brush origin is dirty and needs to updated. (C++ enum variant: DirtyBrushOrigin = 4)

DirtyFont

The font is dirty and needs to be updated. (C++ enum variant: DirtyFont = 8)

DirtyBackground

The background is dirty and needs to be updated. (C++ enum variant: DirtyBackground = 16)

DirtyBackgroundMode

The background mode is dirty and needs to be updated. (C++ enum variant: DirtyBackgroundMode = 32)

DirtyTransform

The transform is dirty and needs to be updated. (C++ enum variant: DirtyTransform = 64)

DirtyClipRegion

The clip region is dirty and needs to be updated. (C++ enum variant: DirtyClipRegion = 128)

DirtyClipPath

The clip path is dirty and needs to be updated. (C++ enum variant: DirtyClipPath = 256)

DirtyHints

The render hints is dirty and needs to be updated. (C++ enum variant: DirtyHints = 512)

DirtyCompositionMode

The composition mode is dirty and needs to be updated. (C++ enum variant: DirtyCompositionMode = 1024)

DirtyClipEnabled

Whether clipping is enabled or not is dirty and needs to be updated. (C++ enum variant: DirtyClipEnabled = 2048)

DirtyOpacity

The constant opacity has changed and needs to be updated as part of the state change in QPaintEngine::updateState(). (C++ enum variant: DirtyOpacity = 4096)

AllDirty

Convenience enum used internally. (C++ enum variant: AllDirty = 65535)