Enum qt_gui::paint_engine::PaintEngineFeature

pub enum PaintEngineFeature {
    AllFeatures,
    PrimitiveTransform,
    PatternTransform,
    PixmapTransform,
    PatternBrush,
    LinearGradientFill,
    RadialGradientFill,
    ConicalGradientFill,
    AlphaBlend,
    PorterDuff,
    PainterPaths,
    Antialiasing,
    BrushStroke,
    ConstantOpacity,
    MaskedBrush,
    PerspectiveTransform,
    BlendModes,
    ObjectBoundingModeGradients,
    RasterOpModes,
    PaintOutsidePaintEvent,
}

C++ type: QPaintEngine::PaintEngineFeature

C++ documentation:

This enum is used to describe the features or capabilities that the paint engine has. If a feature is not supported by the engine, QPainter will do a best effort to emulate that feature through other means and pass on an alpha blended QImage to the engine with the emulated results. Some features cannot be emulated: AlphaBlend and PorterDuff.

The PaintEngineFeatures type is a typedef for QFlags<PaintEngineFeature>. It stores an OR combination of PaintEngineFeature values.

Variants
AllFeatures

All of the above features. This enum value is usually used as a bit mask. (C++ enum variant: AllFeatures = -1)

PrimitiveTransform

The engine has support for transforming drawing primitives. (C++ enum variant: PrimitiveTransform = 1)

PatternTransform

The engine has support for transforming brush patterns. (C++ enum variant: PatternTransform = 2)

PixmapTransform

The engine can transform pixmaps, including rotation and shearing. (C++ enum variant: PixmapTransform = 4)

PatternBrush

The engine is capable of rendering brushes with the brush patterns specified in Qt::BrushStyle. (C++ enum variant: PatternBrush = 8)

LinearGradientFill

The engine supports linear gradient fills. (C++ enum variant: LinearGradientFill = 16)

RadialGradientFill

The engine supports radial gradient fills. (C++ enum variant: RadialGradientFill = 32)

ConicalGradientFill

The engine supports conical gradient fills. (C++ enum variant: ConicalGradientFill = 64)

AlphaBlend

The engine can alpha blend primitives. (C++ enum variant: AlphaBlend = 128)

PorterDuff

The engine supports Porter-Duff operations (C++ enum variant: PorterDuff = 256)

PainterPaths

The engine has path support. (C++ enum variant: PainterPaths = 512)

Antialiasing

The engine can use antialising to improve the appearance of rendered primitives. (C++ enum variant: Antialiasing = 1024)

BrushStroke

The engine supports drawing strokes that contain brushes as fills, not just solid colors (e.g. a dashed gradient line of width 2). (C++ enum variant: BrushStroke = 2048)

ConstantOpacity

The engine supports the feature provided by QPainter::setOpacity(). (C++ enum variant: ConstantOpacity = 4096)

MaskedBrush

The engine is capable of rendering brushes that has a texture with an alpha channel or a mask. (C++ enum variant: MaskedBrush = 8192)

PerspectiveTransform

The engine has support for performing perspective transformations on primitives. (C++ enum variant: PerspectiveTransform = 16384)

BlendModes

The engine supports blending modes. (C++ enum variant: BlendModes = 32768)

ObjectBoundingModeGradients

The engine has native support for gradients with coordinate mode QGradient::ObjectBoundingMode. Otherwise, if QPaintEngine::PatternTransform is supported, object bounding mode gradients are converted to gradients with coordinate mode QGradient::LogicalMode and a brush transform for the coordinate mapping. (C++ enum variant: ObjectBoundingModeGradients = 65536)

RasterOpModes

The engine supports bitwise raster operations. (C++ enum variant: RasterOpModes = 131072)

PaintOutsidePaintEvent

The engine is capable of painting outside of paint events. (C++ enum variant: PaintOutsidePaintEvent = 536870912)