Module qt_gui::paint_engine::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

PaintEngineCoordinateOffsetArgs
This trait represents a set of arguments accepted by PaintEngine::coordinate_offset method.

PaintEngineDrawEllipseArgs
This trait represents a set of arguments accepted by PaintEngine::draw_ellipse method.

PaintEngineDrawImageArgs
This trait represents a set of arguments accepted by PaintEngine::draw_image method.

PaintEngineDrawLinesArgs
This trait represents a set of arguments accepted by PaintEngine::draw_lines method.

PaintEngineDrawPointsArgs
This trait represents a set of arguments accepted by PaintEngine::draw_points method.

PaintEngineDrawPolygonArgs
This trait represents a set of arguments accepted by PaintEngine::draw_polygon method.

PaintEngineDrawRectsArgs
This trait represents a set of arguments accepted by PaintEngine::draw_rects method.

PaintEngineSystemClipArgs
This trait represents a set of arguments accepted by PaintEngine::system_clip method.

PaintEngineSystemRectArgs
This trait represents a set of arguments accepted by PaintEngine::system_rect method.