Struct qt_gui::paint_engine::PaintEngine

pub struct PaintEngine {
    // some fields omitted
}

C++ type: QPaintEngine

C++ documentation:

The QPaintEngine class provides an abstract definition of how QPainter draws to a given device on a given platform.

Qt provides several premade implementations of QPaintEngine for the different painter backends we support. The primary paint engine provided is the raster paint engine, which contains a software rasterizer which supports the full feature set on all supported platforms. This is the default for painting on QWidget-based classes in e.g. on Windows, X11 and Mac OS X, it is the backend for painting on QImage and it is used as a fallback for paint engines that do not support a certain capability. In addition we provide QPaintEngine implementations for OpenGL (accessible through QGLWidget) and printing (which allows using QPainter to draw on a QPrinter object).

If one wants to use QPainter to draw to a different backend, one must subclass QPaintEngine and reimplement all its virtual functions. The QPaintEngine implementation is then made available by subclassing QPaintDevice and reimplementing the virtual function QPaintDevice::paintEngine().

QPaintEngine is created and owned by the QPaintDevice that created it.

Methods

impl PaintEngine

fn begin(&mut self, pdev: *mut PaintDevice) -> bool

C++ method: pure virtual bool QPaintEngine::begin(QPaintDevice* pdev)

C++ documentation:

Reimplement this function to initialise your paint engine when painting is to start on the paint device pdev. Return true if the initialization was successful; otherwise return false.

See also end() and isActive().

fn clear_dirty(&mut self, df: Flags<DirtyFlag>)

C++ method: void QPaintEngine::clearDirty(QFlags<QPaintEngine::DirtyFlag> df)

fn coordinate_offset<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineCoordinateOffsetArgs<'largs>

C++ method: QPaintEngine::coordinateOffset

This is an overloaded function. Available variants:

Rust arguments:
1) fn coordinate_offset(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn coordinate_offset(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: virtual QPoint QPaintEngine::coordinateOffset() const

fn draw_ellipse<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawEllipseArgs<'largs>

C++ method: QPaintEngine::drawEllipse

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_ellipse(&mut self, &qt_core::rect::Rect) -> ()
C++ method: virtual void QPaintEngine::drawEllipse(const QRect& r)

C++ documentation:

The default implementation of this function calls the floating point version of this function

Variant 2

Rust arguments: fn draw_ellipse(&mut self, &qt_core::rect_f::RectF) -> ()
C++ method: virtual void QPaintEngine::drawEllipse(const QRectF& r)

C++ documentation:

Reimplement this function to draw the largest ellipse that can be contained within rectangle rect.

The default implementation calls drawPolygon().

fn draw_image<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawImageArgs<'largs>

C++ method: QPaintEngine::drawImage

This is an overloaded function. Available variants:

Rust arguments:
1) fn draw_image(&mut self, (&qt_core::rect_f::RectF, &::image::Image, &qt_core::rect_f::RectF)) -> ()
2) fn draw_image(&mut self, (&qt_core::rect_f::RectF, &::image::Image, &qt_core::rect_f::RectF, qt_core::flags::Flags<qt_core::qt::ImageConversionFlag>)) -> ()
C++ method: virtual void QPaintEngine::drawImage(const QRectF& r, const QImage& pm, const QRectF& sr, QFlags<Qt::ImageConversionFlag> flags = ?)

C++ documentation:

Reimplement this function to draw the part of the image specified by the sr rectangle in the given rectangle using the given conversion flags flags, to convert it to a pixmap.

fn draw_lines<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawLinesArgs<'largs>

C++ method: QPaintEngine::drawLines

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_lines(&mut self, (*const qt_core::line::Line, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawLines(const QLine* lines, int lineCount)

C++ documentation:

This is an overloaded function.

The default implementation converts the first lineCount lines in lines to a QLineF and calls the floating point version of this function.

Variant 2

Rust arguments: fn draw_lines(&mut self, (*const qt_core::line_f::LineF, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawLines(const QLineF* lines, int lineCount)

C++ documentation:

The default implementation splits the list of lines in lines into lineCount separate calls to drawPath() or drawPolygon() depending on the feature set of the paint engine.

fn draw_path(&mut self, path: &PainterPath)

C++ method: virtual void QPaintEngine::drawPath(const QPainterPath& path)

C++ documentation:

The default implementation ignores the path and does nothing.

fn draw_pixmap(&mut self, r: &RectF, pm: &Pixmap, sr: &RectF)

C++ method: pure virtual void QPaintEngine::drawPixmap(const QRectF& r, const QPixmap& pm, const QRectF& sr)

C++ documentation:

Reimplement this function to draw the part of the pm specified by the sr rectangle in the given r.

fn draw_points<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawPointsArgs<'largs>

C++ method: QPaintEngine::drawPoints

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_points(&mut self, (*const qt_core::point::Point, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawPoints(const QPoint* points, int pointCount)

C++ documentation:

Draws the first pointCount points in the buffer points

The default implementation converts the first pointCount QPoints in points to QPointFs and calls the floating point version of drawPoints.

Variant 2

Rust arguments: fn draw_points(&mut self, (*const qt_core::point_f::PointF, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawPoints(const QPointF* points, int pointCount)

C++ documentation:

Draws the first pointCount points in the buffer points

fn draw_polygon<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawPolygonArgs<'largs>

C++ method: QPaintEngine::drawPolygon

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_polygon(&mut self, (*const qt_core::point::Point, libc::c_int, ::paint_engine::PolygonDrawMode)) -> ()
C++ method: virtual void QPaintEngine::drawPolygon(const QPoint* points, int pointCount, QPaintEngine::PolygonDrawMode mode)

C++ documentation:

This is an overloaded function.

Reimplement this virtual function to draw the polygon defined by the pointCount first points in points, using mode mode.

Note: At least one of the drawPolygon() functions must be reimplemented.

Variant 2

Rust arguments: fn draw_polygon(&mut self, (*const qt_core::point_f::PointF, libc::c_int, ::paint_engine::PolygonDrawMode)) -> ()
C++ method: virtual void QPaintEngine::drawPolygon(const QPointF* points, int pointCount, QPaintEngine::PolygonDrawMode mode)

C++ documentation:

Reimplement this virtual function to draw the polygon defined by the pointCount first points in points, using mode mode.

Note: At least one of the drawPolygon() functions must be reimplemented.

fn draw_rects<'largs, Args>(&'largs mut self, args: Args) where Args: PaintEngineDrawRectsArgs<'largs>

C++ method: QPaintEngine::drawRects

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_rects(&mut self, (*const qt_core::rect::Rect, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawRects(const QRect* rects, int rectCount)

C++ documentation:

This is an overloaded function.

The default implementation converts the first rectCount rectangles in the buffer rects to a QRectF and calls the floating point version of this function.

Variant 2

Rust arguments: fn draw_rects(&mut self, (*const qt_core::rect_f::RectF, libc::c_int)) -> ()
C++ method: virtual void QPaintEngine::drawRects(const QRectF* rects, int rectCount)

C++ documentation:

Draws the first rectCount rectangles in the buffer rects. The default implementation of this function calls drawPath() or drawPolygon() depending on the feature set of the paint engine.

fn draw_text_item(&mut self, p: &PointF, text_item: &TextItem)

C++ method: virtual void QPaintEngine::drawTextItem(const QPointF& p, const QTextItem& textItem)

C++ documentation:

This function draws the text item textItem at position p. The default implementation of this function converts the text to a QPainterPath and paints the resulting path.

fn draw_tiled_pixmap(&mut self, r: &RectF, pixmap: &Pixmap, s: &PointF)

C++ method: virtual void QPaintEngine::drawTiledPixmap(const QRectF& r, const QPixmap& pixmap, const QPointF& s)

C++ documentation:

Reimplement this function to draw the pixmap in the given rect, starting at the given p. The pixmap will be drawn repeatedly until the rect is filled.

fn end(&mut self) -> bool

C++ method: pure virtual bool QPaintEngine::end()

C++ documentation:

Reimplement this function to finish painting on the current paint device. Return true if painting was finished successfully; otherwise return false.

See also begin() and isActive().

fn fix_neg_rect(&mut self, x: *mut c_int, y: *mut c_int, w: *mut c_int, h: *mut c_int)

C++ method: void QPaintEngine::fix_neg_rect(int* x, int* y, int* w, int* h)

fn has_feature(&self, feature: Flags<PaintEngineFeature>) -> bool

C++ method: bool QPaintEngine::hasFeature(QFlags<QPaintEngine::PaintEngineFeature> feature) const

C++ documentation:

Returns true if the paint engine supports the specified feature; otherwise returns false.

fn is_active(&self) -> bool

C++ method: bool QPaintEngine::isActive() const

C++ documentation:

Returns true if the paint engine is actively drawing; otherwise returns false.

See also setActive().

fn is_extended(&self) -> bool

C++ method: bool QPaintEngine::isExtended() const

fn paint_device(&self) -> *mut PaintDevice

C++ method: QPaintDevice* QPaintEngine::paintDevice() const

C++ documentation:

Returns the device that this engine is painting on, if painting is active; otherwise returns 0.

fn painter(&self) -> *mut Painter

C++ method: QPainter* QPaintEngine::painter() const

C++ documentation:

Returns the paint engine's painter.

fn set_active(&mut self, new_state: bool)

C++ method: void QPaintEngine::setActive(bool newState)

C++ documentation:

Sets the active state of the paint engine to state.

See also isActive().

fn set_dirty(&mut self, df: Flags<DirtyFlag>)

C++ method: void QPaintEngine::setDirty(QFlags<QPaintEngine::DirtyFlag> df)

fn set_paint_device(&mut self, device: *mut PaintDevice)

C++ method: void QPaintEngine::setPaintDevice(QPaintDevice* device)

fn set_system_clip(&mut self, base_clip: &Region)

C++ method: void QPaintEngine::setSystemClip(const QRegion& baseClip)

fn set_system_rect(&mut self, rect: &Rect)

C++ method: void QPaintEngine::setSystemRect(const QRect& rect)

fn sync_state(&mut self)

C++ method: void QPaintEngine::syncState()

fn system_clip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineSystemClipArgs<'largs>

C++ method: QPaintEngine::systemClip

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_clip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::region::Region>
2) fn system_clip(&self, cpp_utils::AsStruct) -> ::region::Region
C++ method: QRegion QPaintEngine::systemClip() const

fn system_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineSystemRectArgs<'largs>

C++ method: QPaintEngine::systemRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn system_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPaintEngine::systemRect() const

fn test_dirty(&mut self, df: Flags<DirtyFlag>) -> bool

C++ method: bool QPaintEngine::testDirty(QFlags<QPaintEngine::DirtyFlag> df)

fn type_(&self) -> Type

C++ method: pure virtual QPaintEngine::Type QPaintEngine::type() const

C++ documentation:

Reimplement this function to return the paint engine Type.

fn update_state(&mut self, state: &PaintEngineState)

C++ method: pure virtual void QPaintEngine::updateState(const QPaintEngineState& state)

C++ documentation:

Reimplement this function to update the state of a paint engine.

When implemented, this function is responsible for checking the paint engine's current state and update the properties that are changed. Use the QPaintEngineState::state() function to find out which properties that must be updated, then use the corresponding get function to retrieve the current values for the given properties.

See also QPaintEngineState.