Struct qt_gui::paint_engine_state::PaintEngineState

pub struct PaintEngineState {
    // some fields omitted
}

C++ type: QPaintEngineState

C++ documentation:

The QPaintEngineState class provides information about the active paint engine's current state.

QPaintEngineState records which properties that have changed since the last time the paint engine was updated, as well as their current value.

Which properties that have changed can at any time be retrieved using the state() function. This function returns an instance of the QPaintEngine::DirtyFlags type which stores an OR combination of QPaintEngine::DirtyFlag values. The QPaintEngine::DirtyFlag enum defines whether a property has changed since the last update or not.

If a property is marked with a dirty flag, its current value can be retrieved using the corresponding get function:

Property Flag	Current Property Value

QPaintEngine::DirtyBackground	backgroundBrush()

QPaintEngine::DirtyBackgroundMode	backgroundMode()

QPaintEngine::DirtyBrush	brush()

QPaintEngine::DirtyBrushOrigin	brushOrigin()

QPaintEngine::DirtyClipRegion or QPaintEngine::DirtyClipPath	clipOperation()

QPaintEngine::DirtyClipPath	clipPath()

QPaintEngine::DirtyClipRegion	clipRegion()

QPaintEngine::DirtyCompositionMode	compositionMode()

QPaintEngine::DirtyFont	font()

QPaintEngine::DirtyTransform	transform()

QPaintEngine::DirtyClipEnabled	isClipEnabled()

QPaintEngine::DirtyPen	pen()

QPaintEngine::DirtyHints	renderHints()

The QPaintEngineState class also provide the painter() function which returns a pointer to the painter that is currently updating the paint engine.

An instance of this class, representing the current state of the active paint engine, is passed as argument to the QPaintEngine::updateState() function. The only situation in which you will have to use this class directly is when implementing your own paint engine.

Methods

impl PaintEngineState

fn background_brush<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateBackgroundBrushArgs<'largs>

C++ method: QPaintEngineState::backgroundBrush

This is an overloaded function. Available variants:

Rust arguments:
1) fn background_brush(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background_brush(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QPaintEngineState::backgroundBrush() const

C++ documentation:

Returns the background brush in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyBackground flag.

See also state() and QPaintEngine::updateState().

fn background_mode(&self) -> BgmOde

C++ method: Qt::BGMode QPaintEngineState::backgroundMode() const

C++ documentation:

Returns the background mode in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyBackgroundMode flag.

See also state() and QPaintEngine::updateState().

fn brush<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateBrushArgs<'largs>

C++ method: QPaintEngineState::brush

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn brush(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QPaintEngineState::brush() const

C++ documentation:

Returns the brush in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyBrush flag.

See also state() and QPaintEngine::updateState().

fn brush_needs_resolving(&self) -> bool

C++ method: bool QPaintEngineState::brushNeedsResolving() const

C++ documentation:

Returns whether the coordinate of the fill have been specified as bounded by the current rendering operation and have to be resolved (about the currently rendered primitive).

This function was introduced in Qt 4.3.

fn brush_origin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateBrushOriginArgs<'largs>

C++ method: QPaintEngineState::brushOrigin

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush_origin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn brush_origin(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QPaintEngineState::brushOrigin() const

C++ documentation:

Returns the brush origin in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyBrushOrigin flag.

See also state() and QPaintEngine::updateState().

fn clip_operation(&self) -> ClipOperation

C++ method: Qt::ClipOperation QPaintEngineState::clipOperation() const

C++ documentation:

Returns the clip operation in the current paint engine state.

This variable should only be used when the state() returns a combination which includes either the QPaintEngine::DirtyClipPath or the QPaintEngine::DirtyClipRegion flag.

See also state() and QPaintEngine::updateState().

fn clip_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateClipPathArgs<'largs>

C++ method: QPaintEngineState::clipPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn clip_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::painter_path::PainterPath>
2) fn clip_path(&self, cpp_utils::AsStruct) -> ::painter_path::PainterPath
C++ method: QPainterPath QPaintEngineState::clipPath() const

C++ documentation:

Returns the clip path in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyClipPath flag.

See also state() and QPaintEngine::updateState().

fn clip_region<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateClipRegionArgs<'largs>

C++ method: QPaintEngineState::clipRegion

This is an overloaded function. Available variants:

Rust arguments:
1) fn clip_region(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::region::Region>
2) fn clip_region(&self, cpp_utils::AsStruct) -> ::region::Region
C++ method: QRegion QPaintEngineState::clipRegion() const

C++ documentation:

Returns the clip region in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyClipRegion flag.

See also state() and QPaintEngine::updateState().

fn composition_mode(&self) -> CompositionMode

C++ method: QPainter::CompositionMode QPaintEngineState::compositionMode() const

C++ documentation:

Returns the composition mode in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyCompositionMode flag.

See also state() and QPaintEngine::updateState().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateFontArgs<'largs>

C++ method: QPaintEngineState::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> ::font::Font
C++ method: QFont QPaintEngineState::font() const

C++ documentation:

Returns the font in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyFont flag.

See also state() and QPaintEngine::updateState().

fn is_clip_enabled(&self) -> bool

C++ method: bool QPaintEngineState::isClipEnabled() const

C++ documentation:

Returns whether clipping is enabled or not in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyClipEnabled flag.

See also state() and QPaintEngine::updateState().

fn matrix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateMatrixArgs<'largs>

C++ method: QPaintEngineState::matrix

This is an overloaded function. Available variants:

Rust arguments:
1) fn matrix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::matrix::Matrix>
2) fn matrix(&self, cpp_utils::AsStruct) -> ::matrix::Matrix
C++ method: QMatrix QPaintEngineState::matrix() const

C++ documentation:

Returns the matrix in the current paint engine state.

Note: It is advisable to use transform() instead of this function to preserve the properties of perspective transformations.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyTransform flag.

This function was introduced in Qt 4.2.

See also state() and QPaintEngine::updateState().

fn opacity(&self) -> c_double

C++ method: double QPaintEngineState::opacity() const

C++ documentation:

Returns the opacity in the current paint engine state.

This function was introduced in Qt 4.2.

fn painter(&self) -> *mut Painter

C++ method: QPainter* QPaintEngineState::painter() const

C++ documentation:

Returns a pointer to the painter currently updating the paint engine.

fn pen<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStatePenArgs<'largs>

C++ method: QPaintEngineState::pen

This is an overloaded function. Available variants:

Rust arguments:
1) fn pen(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pen::Pen>
2) fn pen(&self, cpp_utils::AsStruct) -> ::pen::Pen
C++ method: QPen QPaintEngineState::pen() const

C++ documentation:

Returns the pen in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyPen flag.

See also state() and QPaintEngine::updateState().

fn pen_needs_resolving(&self) -> bool

C++ method: bool QPaintEngineState::penNeedsResolving() const

C++ documentation:

Returns whether the coordinate of the stroke have been specified as bounded by the current rendering operation and have to be resolved (about the currently rendered primitive).

This function was introduced in Qt 4.3.

fn render_hints(&self) -> Flags<RenderHint>

C++ method: QFlags<QPainter::RenderHint> QPaintEngineState::renderHints() const

C++ documentation:

Returns the render hints in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyHints flag.

See also state() and QPaintEngine::updateState().

fn state(&self) -> Flags<DirtyFlag>

C++ method: QFlags<QPaintEngine::DirtyFlag> QPaintEngineState::state() const

fn transform<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaintEngineStateTransformArgs<'largs>

C++ method: QPaintEngineState::transform

This is an overloaded function. Available variants:

Rust arguments:
1) fn transform(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
2) fn transform(&self, cpp_utils::AsStruct) -> ::transform::Transform
C++ method: QTransform QPaintEngineState::transform() const

C++ documentation:

Returns the matrix in the current paint engine state.

This variable should only be used when the state() returns a combination which includes the QPaintEngine::DirtyTransform flag.

This function was introduced in Qt 4.3.

See also state() and QPaintEngine::updateState().