Enum qt_gui::painter::PixmapFragmentHint

pub enum PixmapFragmentHint {
    _Invalid,
    OpaqueHint,
}

C++ type: QPainter::PixmapFragmentHint

C++ documentation:

This enum was introduced or modified in Qt 4.7.

The PixmapFragmentHints type is a typedef for QFlags<PixmapFragmentHint>. It stores an OR combination of PixmapFragmentHint values.

See also QPainter::drawPixmapFragments() and QPainter::PixmapFragment.

Variants
_Invalid

This variant is added in Rust because enums with one variant and C representation are not supported.

OpaqueHint

Indicates that the pixmap fragments to be drawn are opaque. Opaque fragments are potentially faster to draw. (C++ enum variant: OpaqueHint = 1)