Enum qt_gui::painter::RenderHint

pub enum RenderHint {
    Antialiasing,
    TextAntialiasing,
    SmoothPixmapTransform,
    HighQualityAntialiasing,
    NonCosmeticDefaultPen,
    Qt4CompatiblePainting,
}

C++ type: QPainter::RenderHint

C++ documentation:

Renderhints are used to specify flags to QPainter that may or may not be respected by any given engine.

The RenderHints type is a typedef for QFlags<RenderHint>. It stores an OR combination of RenderHint values.

See also renderHints(), setRenderHint(), Rendering Quality, and Concentric Circles Example.

Variants
Antialiasing

Indicates that the engine should antialias edges of primitives if possible. (C++ enum variant: Antialiasing = 1)

TextAntialiasing

Indicates that the engine should antialias text if possible. To forcibly disable antialiasing for text, do not use this hint. Instead, set QFont::NoAntialias on your font's style strategy. (C++ enum variant: TextAntialiasing = 2)

SmoothPixmapTransform

Indicates that the engine should use a smooth pixmap transformation algorithm (such as bilinear) rather than nearest neighbor. (C++ enum variant: SmoothPixmapTransform = 4)

HighQualityAntialiasing

This value is obsolete and will be ignored, use the Antialiasing render hint instead. (C++ enum variant: HighQualityAntialiasing = 8)

NonCosmeticDefaultPen

This value is obsolete, the default for QPen is now non-cosmetic. (C++ enum variant: NonCosmeticDefaultPen = 16)

Qt4CompatiblePainting

Compatibility hint telling the engine to use the same X11 based fill rules as in Qt 4, where aliased rendering is offset by slightly less than half a pixel. Also will treat default constructed pens as cosmetic. Potentially useful when porting a Qt 4 application to Qt 5. (C++ enum variant: Qt4CompatiblePainting = 32)