Module qt_gui::painter::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

PainterBoundingRectArgs
This trait represents a set of arguments accepted by Painter::bounding_rect method.

PainterBrushOriginArgs
This trait represents a set of arguments accepted by Painter::brush_origin method.

PainterClipBoundingRectArgs
This trait represents a set of arguments accepted by Painter::clip_bounding_rect method.

PainterClipPathArgs
This trait represents a set of arguments accepted by Painter::clip_path method.

PainterClipRegionArgs
This trait represents a set of arguments accepted by Painter::clip_region method.

PainterCombinedMatrixArgs
This trait represents a set of arguments accepted by Painter::combined_matrix method.

PainterCombinedTransformArgs
This trait represents a set of arguments accepted by Painter::combined_transform method.

PainterDrawArcArgs
This trait represents a set of arguments accepted by Painter::draw_arc method.

PainterDrawChordArgs
This trait represents a set of arguments accepted by Painter::draw_chord method.

PainterDrawConvexPolygonArgs
This trait represents a set of arguments accepted by Painter::draw_convex_polygon method.

PainterDrawEllipseArgs
This trait represents a set of arguments accepted by Painter::draw_ellipse method.

PainterDrawImageArgs
This trait represents a set of arguments accepted by Painter::draw_image method.

PainterDrawLineArgs
This trait represents a set of arguments accepted by Painter::draw_line method.

PainterDrawLinesArgs
This trait represents a set of arguments accepted by Painter::draw_lines method.

PainterDrawPictureArgs
This trait represents a set of arguments accepted by Painter::draw_picture method.

PainterDrawPieArgs
This trait represents a set of arguments accepted by Painter::draw_pie method.

PainterDrawPixmapArgs
This trait represents a set of arguments accepted by Painter::draw_pixmap method.

PainterDrawPixmapFragmentsArgs
This trait represents a set of arguments accepted by Painter::draw_pixmap_fragments method.

PainterDrawPointArgs
This trait represents a set of arguments accepted by Painter::draw_point method.

PainterDrawPointsArgs
This trait represents a set of arguments accepted by Painter::draw_points method.

PainterDrawPolygonArgs
This trait represents a set of arguments accepted by Painter::draw_polygon method.

PainterDrawPolylineArgs
This trait represents a set of arguments accepted by Painter::draw_polyline method.

PainterDrawRectArgs
This trait represents a set of arguments accepted by Painter::draw_rect method.

PainterDrawRectsArgs
This trait represents a set of arguments accepted by Painter::draw_rects method.

PainterDrawRoundRectArgs
This trait represents a set of arguments accepted by Painter::draw_round_rect method.

PainterDrawRoundedRectArgs
This trait represents a set of arguments accepted by Painter::draw_rounded_rect method.

PainterDrawStaticTextArgs
This trait represents a set of arguments accepted by Painter::draw_static_text method.

PainterDrawTextArgs
This trait represents a set of arguments accepted by Painter::draw_text method.

PainterDrawTextItemArgs
This trait represents a set of arguments accepted by Painter::draw_text_item method.

PainterDrawTiledPixmapArgs
This trait represents a set of arguments accepted by Painter::draw_tiled_pixmap method.

PainterEraseRectArgs
This trait represents a set of arguments accepted by Painter::erase_rect method.

PainterFillRectArgs
This trait represents a set of arguments accepted by Painter::fill_rect method.

PainterFontInfoArgs
This trait represents a set of arguments accepted by Painter::font_info method.

PainterFontMetricsArgs
This trait represents a set of arguments accepted by Painter::font_metrics method.

PainterNewArgs
This trait represents a set of arguments accepted by Painter::new method.

PainterRedirectedArgs
This trait represents a set of arguments accepted by Painter::redirected method.

PainterSetBrushArgs
This trait represents a set of arguments accepted by Painter::set_brush method.

PainterSetBrushOriginArgs
This trait represents a set of arguments accepted by Painter::set_brush_origin method.

PainterSetClipPathArgs
This trait represents a set of arguments accepted by Painter::set_clip_path method.

PainterSetClipRectArgs
This trait represents a set of arguments accepted by Painter::set_clip_rect method.

PainterSetClipRegionArgs
This trait represents a set of arguments accepted by Painter::set_clip_region method.

PainterSetMatrixArgs
This trait represents a set of arguments accepted by Painter::set_matrix method.

PainterSetPenArgs
This trait represents a set of arguments accepted by Painter::set_pen method.

PainterSetRedirectedArgs
This trait represents a set of arguments accepted by Painter::set_redirected method.

PainterSetRenderHintArgs
This trait represents a set of arguments accepted by Painter::set_render_hint method.

PainterSetRenderHintsArgs
This trait represents a set of arguments accepted by Painter::set_render_hints method.

PainterSetTransformArgs
This trait represents a set of arguments accepted by Painter::set_transform method.

PainterSetViewportArgs
This trait represents a set of arguments accepted by Painter::set_viewport method.

PainterSetWindowArgs
This trait represents a set of arguments accepted by Painter::set_window method.

PainterSetWorldMatrixArgs
This trait represents a set of arguments accepted by Painter::set_world_matrix method.

PainterSetWorldTransformArgs
This trait represents a set of arguments accepted by Painter::set_world_transform method.

PainterTranslateArgs
This trait represents a set of arguments accepted by Painter::translate method.

PainterViewportArgs
This trait represents a set of arguments accepted by Painter::viewport method.

PainterWindowArgs
This trait represents a set of arguments accepted by Painter::window method.

PixmapFragmentCreateArgs
This trait represents a set of arguments accepted by PixmapFragment::create method.