Struct qt_gui::painter::PixmapFragment

pub struct PixmapFragment {
    // some fields omitted
}

C++ type: QPainter::PixmapFragment

C++ documentation:

This class is used in conjunction with the QPainter::drawPixmapFragments() function to specify how a pixmap, or sub-rect of a pixmap, is drawn.

The sourceLeft, sourceTop, width and height variables are used as a source rectangle within the pixmap passed into the QPainter::drawPixmapFragments() function. The variables x, y, width and height are used to calculate the target rectangle that is drawn. x and y denotes the center of the target rectangle. The width and height in the target rectangle is scaled by the scaleX and scaleY values. The resulting target rectangle is then rotated rotation degrees around the x, y center point.

Methods

impl PixmapFragment

fn create<Args>(args: Args) -> Args::ReturnType where Args: PixmapFragmentCreateArgs

C++ method: QPainter::PixmapFragment::create

This is an overloaded function. Available variants:

Rust arguments:
1) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter::PixmapFragment>
2) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter::PixmapFragment>
3) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter::PixmapFragment>
4) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter::PixmapFragment>
5) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter::PixmapFragment>
6) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, cpp_utils::AsStruct)) -> ::painter::PixmapFragment
7) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, cpp_utils::AsStruct)) -> ::painter::PixmapFragment
8) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::painter::PixmapFragment
9) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::painter::PixmapFragment
10) fn create((&qt_core::point_f::PointF, &qt_core::rect_f::RectF, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::painter::PixmapFragment
C++ method: static QPainter::PixmapFragment QPainter::PixmapFragment::create(const QPointF& pos, const QRectF& sourceRect, double scaleX = ?, double scaleY = ?, double rotation = ?, double opacity = ?)

C++ documentation:

This is a convenience function that returns a QPainter::PixmapFragment that is initialized with the pos, sourceRect, scaleX, scaleY, rotation, opacity parameters.

This function was introduced in Qt 4.7.

fn height(&self) -> c_double

C++ method: double QPainter::PixmapFragment::height() const

fn opacity(&self) -> c_double

C++ method: double QPainter::PixmapFragment::opacity() const

fn rotation(&self) -> c_double

C++ method: double QPainter::PixmapFragment::rotation() const

fn scale_x(&self) -> c_double

C++ method: double QPainter::PixmapFragment::scaleX() const

fn scale_y(&self) -> c_double

C++ method: double QPainter::PixmapFragment::scaleY() const

fn set_height(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_height(double value)

fn set_opacity(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_opacity(double value)

fn set_rotation(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_rotation(double value)

fn set_scale_x(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_scaleX(double value)

fn set_scale_y(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_scaleY(double value)

fn set_source_left(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_sourceLeft(double value)

fn set_source_top(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_sourceTop(double value)

fn set_width(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_width(double value)

fn set_x(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_x(double value)

fn set_y(&mut self, value: c_double)

C++ method: void QPainter::PixmapFragment::set_y(double value)

fn source_left(&self) -> c_double

C++ method: double QPainter::PixmapFragment::sourceLeft() const

fn source_top(&self) -> c_double

C++ method: double QPainter::PixmapFragment::sourceTop() const

fn width(&self) -> c_double

C++ method: double QPainter::PixmapFragment::width() const

fn x(&self) -> c_double

C++ method: double QPainter::PixmapFragment::x() const

fn y(&self) -> c_double

C++ method: double QPainter::PixmapFragment::y() const