Function qt_gui::painter_path::op_shl

pub fn op_shl<Args>(args: Args) -> Args::ReturnType where Args: OpShlArgs

C++ method: operator<<

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_shl((&qt_core::debug::Debug, &::painter_path::PainterPath, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::debug::Debug>
2) fn op_shl((&qt_core::debug::Debug, &::painter_path::PainterPath, cpp_utils::AsStruct)) -> qt_core::debug::Debug
C++ method: QDebug operator<<(QDebug arg1, const QPainterPath& arg2)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for QDataStream & operator<<(QDataStream & stream, const QIcon & icon):

Writes the given icon to the given stream as a PNG image. If the icon contains more than one image, all images will be written to the stream. Note that writing the stream to a file will not produce a valid image file.

This function was introduced in Qt 4.2.