Struct qt_gui::painter_path::Element

pub struct Element {
    // some fields omitted
}

C++ type: QPainterPath::Element

C++ documentation:

The QPainterPath::Element class specifies the position and type of a subpath.

Once a QPainterPath object is constructed, subpaths like lines and curves can be added to the path (creating QPainterPath::LineToElement and QPainterPath::CurveToElement components).

The lines and curves stretch from the currentPosition() to the position passed as argument. The currentPosition() of the QPainterPath object is always the end position of the last subpath that was added (or the initial start point). The moveTo() function can be used to move the currentPosition() without adding a line or curve, creating a QPainterPath::MoveToElement component.

Methods

impl Element

fn as_q_point_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ElementAsQPointFArgs<'largs>

C++ method: QPainterPath::Element::operator QPointF

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_point_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn as_q_point_f(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QPainterPath::Element::operator QPointF() const

C++ documentation:

Returns the element's position.

See also x and y.

fn is_curve_to(&self) -> bool

C++ method: bool QPainterPath::Element::isCurveTo() const

C++ documentation:

Returns true if the element is a curve, otherwise returns false.

See also type and QPainterPath::CurveToElement.

fn is_line_to(&self) -> bool

C++ method: bool QPainterPath::Element::isLineTo() const

C++ documentation:

Returns true if the element is a line, otherwise returns false.

See also type and QPainterPath::LineToElement.

fn is_move_to(&self) -> bool

C++ method: bool QPainterPath::Element::isMoveTo() const

C++ documentation:

Returns true if the element is moving the current position, otherwise returns false.

See also type and QPainterPath::MoveToElement.

fn op_eq(&self, e: &Element) -> bool

C++ method: bool QPainterPath::Element::operator==(const QPainterPath::Element& e) const

C++ documentation:

Returns true if this element is equal to other; otherwise returns false.

This function was introduced in Qt 4.2.

See also operator!=().

fn op_neq(&self, e: &Element) -> bool

C++ method: bool QPainterPath::Element::operator!=(const QPainterPath::Element& e) const

C++ documentation:

Returns true if this element is not equal to other; otherwise returns false.

This function was introduced in Qt 4.2.

See also operator==().

fn set_type(&mut self, value: ElementType)

C++ method: void QPainterPath::Element::set_type(QPainterPath::ElementType value)

fn set_x(&mut self, value: c_double)

C++ method: void QPainterPath::Element::set_x(double value)

fn set_y(&mut self, value: c_double)

C++ method: void QPainterPath::Element::set_y(double value)

fn type_(&self) -> ElementType

C++ method: QPainterPath::ElementType QPainterPath::Element::type() const

fn x(&self) -> c_double

C++ method: double QPainterPath::Element::x() const

fn y(&self) -> c_double

C++ method: double QPainterPath::Element::y() const