Struct qt_gui::painter_path_stroker::PainterPathStroker

pub struct PainterPathStroker {
    // some fields omitted
}

C++ type: QPainterPathStroker

C++ documentation:

The QPainterPathStroker class is used to generate fillable outlines for a given painter path.

By calling the createStroke() function, passing a given QPainterPath as argument, a new painter path representing the outline of the given path is created. The newly created painter path can then be filled to draw the original painter path's outline.

You can control the various design aspects (width, cap styles, join styles and dash pattern) of the outlining using the following functions:

setWidth()

setCapStyle()

setJoinStyle()

setDashPattern()

The setDashPattern() function accepts both a Qt::PenStyle object and a vector representation of the pattern as argument.

In addition you can specify a curve's threshold, controlling the granularity with which a curve is drawn, using the setCurveThreshold() function. The default threshold is a well adjusted value (0.25), and normally you should not need to modify it. However, you can make the curve's appearance smoother by decreasing its value.

You can also control the miter limit for the generated outline using the setMiterLimit() function. The miter limit describes how far from each join the miter join can extend. The limit is specified in the units of width so the pixelwise miter limit will be miterlimit * width. This value is only used if the join style is Qt::MiterJoin.

The painter path generated by the createStroke() function should only be used for outlining the given painter path. Otherwise it may cause unexpected behavior. Generated outlines also require the Qt::WindingFill rule which is set by default.

Methods

impl PainterPathStroker

fn cap_style(&self) -> PenCapStyle

C++ method: Qt::PenCapStyle QPainterPathStroker::capStyle() const

C++ documentation:

Returns the cap style of the generated outlines.

See also setCapStyle().

fn create_stroke<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PainterPathStrokerCreateStrokeArgs<'largs>

C++ method: QPainterPathStroker::createStroke

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_stroke(&self, (&::painter_path::PainterPath, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter_path::PainterPath>
2) fn create_stroke(&self, (&::painter_path::PainterPath, cpp_utils::AsStruct)) -> ::painter_path::PainterPath
C++ method: QPainterPath QPainterPathStroker::createStroke(const QPainterPath& path) const

C++ documentation:

Generates a new path that is a fillable area representing the outline of the given path.

The various design aspects of the outline are based on the stroker's properties: width(), capStyle(), joinStyle(), dashPattern(), curveThreshold() and miterLimit().

The generated path should only be used for outlining the given painter path. Otherwise it may cause unexpected behavior. Generated outlines also require the Qt::WindingFill rule which is set by default.

fn curve_threshold(&self) -> c_double

C++ method: double QPainterPathStroker::curveThreshold() const

C++ documentation:

Returns the curve flattening threshold for the generated outlines.

See also setCurveThreshold().

fn dash_offset(&self) -> c_double

C++ method: double QPainterPathStroker::dashOffset() const

C++ documentation:

Returns the dash offset for the generated outlines.

See also setDashOffset().

fn dash_pattern<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PainterPathStrokerDashPatternArgs<'largs>

C++ method: QPainterPathStroker::dashPattern

This is an overloaded function. Available variants:

Rust arguments:
1) fn dash_pattern(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorCDouble>
2) fn dash_pattern(&self, cpp_utils::AsStruct) -> ::vector::VectorCDouble
C++ method: QVector<double> QPainterPathStroker::dashPattern() const

C++ documentation:

Returns the dash pattern for the generated outlines.

See also setDashPattern().

fn join_style(&self) -> PenJoinStyle

C++ method: Qt::PenJoinStyle QPainterPathStroker::joinStyle() const

C++ documentation:

Returns the join style of the generated outlines.

See also setJoinStyle().

fn miter_limit(&self) -> c_double

C++ method: double QPainterPathStroker::miterLimit() const

C++ documentation:

Returns the miter limit for the generated outlines.

See also setMiterLimit().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PainterPathStrokerNewArgs

C++ method: QPainterPathStroker::QPainterPathStroker

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::painter_path_stroker::PainterPathStroker
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::painter_path_stroker::PainterPathStroker>
C++ method: [constructor] void QPainterPathStroker::QPainterPathStroker()

Variant 2

Rust arguments:
1) fn new((&::pen::Pen, cpp_utils::AsStruct)) -> ::painter_path_stroker::PainterPathStroker
2) fn new((&::pen::Pen, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter_path_stroker::PainterPathStroker>
C++ method: [constructor] void QPainterPathStroker::QPainterPathStroker(const QPen& pen)

fn set_cap_style(&mut self, style: PenCapStyle)

C++ method: void QPainterPathStroker::setCapStyle(Qt::PenCapStyle style)

C++ documentation:

Sets the cap style of the generated outlines to style. If a dash pattern is set, each segment of the pattern is subject to the cap style.

See also capStyle().

fn set_curve_threshold(&mut self, threshold: c_double)

C++ method: void QPainterPathStroker::setCurveThreshold(double threshold)

C++ documentation:

Specifies the curve flattening threshold, controlling the granularity with which the generated outlines' curve is drawn.

The default threshold is a well adjusted value (0.25), and normally you should not need to modify it. However, you can make the curve's appearance smoother by decreasing its value.

See also curveThreshold().

fn set_dash_offset(&mut self, offset: c_double)

C++ method: void QPainterPathStroker::setDashOffset(double offset)

C++ documentation:

Sets the dash offset for the generated outlines to offset.

See the documentation for QPen::setDashOffset() for a description of the dash offset.

See also dashOffset().

fn set_dash_pattern<'largs, Args>(&'largs mut self, args: Args) where Args: PainterPathStrokerSetDashPatternArgs<'largs>

C++ method: QPainterPathStroker::setDashPattern

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_dash_pattern(&mut self, qt_core::qt::PenStyle) -> ()
C++ method: void QPainterPathStroker::setDashPattern(Qt::PenStyle arg1)

C++ documentation:

Sets the dash pattern for the generated outlines to style.

See also dashPattern().

Variant 2

Rust arguments: fn set_dash_pattern(&mut self, &::vector::VectorCDouble) -> ()
C++ method: void QPainterPathStroker::setDashPattern(const QVector<double>& dashPattern)

C++ documentation:

This is an overloaded function.

Sets the dash pattern for the generated outlines to dashPattern. This function makes it possible to specify custom dash patterns.

Each element in the vector contains the lengths of the dashes and spaces in the stroke, beginning with the first dash in the first element, the first space in the second element, and alternating between dashes and spaces for each following pair of elements.

The vector can contain an odd number of elements, in which case the last element will be extended by the length of the first element when the pattern repeats.

fn set_join_style(&mut self, style: PenJoinStyle)

C++ method: void QPainterPathStroker::setJoinStyle(Qt::PenJoinStyle style)

C++ documentation:

Sets the join style of the generated outlines to style.

See also joinStyle().

fn set_miter_limit(&mut self, length: c_double)

C++ method: void QPainterPathStroker::setMiterLimit(double length)

C++ documentation:

Sets the miter limit of the generated outlines to limit.

The miter limit describes how far from each join the miter join can extend. The limit is specified in units of the currently set width. So the pixelwise miter limit will be miterlimit * width.

This value is only used if the join style is Qt::MiterJoin.

See also miterLimit().

fn set_width(&mut self, width: c_double)

C++ method: void QPainterPathStroker::setWidth(double width)

C++ documentation:

Sets the width of the generated outline painter path to width.

The generated outlines will extend approximately 50% of width to each side of the given input path's original outline.

See also width().

fn width(&self) -> c_double

C++ method: double QPainterPathStroker::width() const

C++ documentation:

Returns the width of the generated outlines.

See also setWidth().