Struct qt_gui::pair::PairAccessibleInterfaceMutPtrFlagsRelationFlag

pub struct PairAccessibleInterfaceMutPtrFlagsRelationFlag {
    // some fields omitted
}

C++ type: QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>

C++ documentation:

The QPair class is a template class that stores a pair of items.

QPair<T1, T2> can be used in your application if the STL pair type is not available. It stores one value of type T1 and one value of type T2. It can be used as a return value for a function that needs to return two values, or as the value type of a generic container.

Here's an example of a QPair that stores one QString and one double value:

QPair<QString, double> pair;

The components are accessible as public data members called first and second. For example:

pair.first = "pi";
pair.second = 3.14159265358979323846;

QPair's template data types (T1 and T2) must be assignable data types. You cannot, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; these requirements are documented on a per-function basis.

Methods

impl PairAccessibleInterfaceMutPtrFlagsRelationFlag

fn first(&self) -> *mut AccessibleInterface

C++ method: QAccessibleInterface* QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::first() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: PairAccessibleInterfaceMutPtrFlagsRelationFlagNewArgs

C++ method: QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::QPair

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pair::PairAccessibleInterfaceMutPtrFlagsRelationFlag
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairAccessibleInterfaceMutPtrFlagsRelationFlag>
C++ method: [constructor] void QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::QPair()

Variant 2

Rust arguments:
1) fn new((&*mut ::accessible_interface::AccessibleInterface, qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsStruct)) -> ::pair::PairAccessibleInterfaceMutPtrFlagsRelationFlag
2) fn new((&*mut ::accessible_interface::AccessibleInterface, qt_core::flags::Flags<::accessible::RelationFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pair::PairAccessibleInterfaceMutPtrFlagsRelationFlag>
C++ method: [constructor] void QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::QPair(QAccessibleInterface* const & t1, const QFlags<QAccessible::RelationFlag>& t2)

fn second(&self) -> Flags<RelationFlag>

C++ method: QFlags<QAccessible::RelationFlag> QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::second() const

fn set_first(&mut self, value: *mut AccessibleInterface)

C++ method: void QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::set_first(QAccessibleInterface* value)

fn set_second(&mut self, value: Flags<RelationFlag>)

C++ method: void QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::set_second(QFlags<QAccessible::RelationFlag> value)

fn swap(&mut self, other: &mut PairAccessibleInterfaceMutPtrFlagsRelationFlag)

C++ method: void QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>::swap(QPair<QAccessibleInterface*, QFlags<QAccessible::RelationFlag>>& other)