Struct qt_gui::pair::PairCDoubleColorRef

pub struct PairCDoubleColorRef {
    // some fields omitted
}

C++ type: QPair<double, QColor>

C++ documentation:

The QPair class is a template class that stores a pair of items.

QPair<T1, T2> can be used in your application if the STL pair type is not available. It stores one value of type T1 and one value of type T2. It can be used as a return value for a function that needs to return two values, or as the value type of a generic container.

Here's an example of a QPair that stores one QString and one double value:

QPair<QString, double> pair;

The components are accessible as public data members called first and second. For example:

pair.first = "pi";
pair.second = 3.14159265358979323846;

QPair's template data types (T1 and T2) must be assignable data types. You cannot, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; these requirements are documented on a per-function basis.

Methods

impl PairCDoubleColorRef

fn first(&self) -> c_double

C++ method: double QPair<double, QColor>::first() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: PairCDoubleColorRefNewArgs

C++ method: QPair<double, QColor>::QPair

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pair::PairCDoubleColorRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCDoubleColorRef>
C++ method: [constructor] void QPair<double, QColor>::QPair()

Variant 2

Rust arguments:
1) fn new((&libc::c_double, &::color::Color, cpp_utils::AsStruct)) -> ::pair::PairCDoubleColorRef
2) fn new((&libc::c_double, &::color::Color, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pair::PairCDoubleColorRef>
C++ method: [constructor] void QPair<double, QColor>::QPair(const double& t1, const QColor& t2)

fn second<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PairCDoubleColorRefSecondArgs<'largs>

C++ method: QPair<double, QColor>::second

This is an overloaded function. Available variants:

Rust arguments:
1) fn second(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn second(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: QColor QPair<double, QColor>::second() const

fn set_first(&mut self, value: c_double)

C++ method: void QPair<double, QColor>::set_first(double value)

fn set_second(&mut self, value: &Color)

C++ method: void QPair<double, QColor>::set_second(QColor value)

fn swap(&mut self, other: &mut PairCDoubleColorRef)

C++ method: void QPair<double, QColor>::swap(QPair<double, QColor>& other)