Struct qt_gui::pair::PairCFloatCFloat

pub struct PairCFloatCFloat {
    // some fields omitted
}

C++ type: QPair<float, float>

C++ documentation:

The QPair class is a template class that stores a pair of items.

QPair<T1, T2> can be used in your application if the STL pair type is not available. It stores one value of type T1 and one value of type T2. It can be used as a return value for a function that needs to return two values, or as the value type of a generic container.

Here's an example of a QPair that stores one QString and one double value:

QPair<QString, double> pair;

The components are accessible as public data members called first and second. For example:

pair.first = "pi";
pair.second = 3.14159265358979323846;

QPair's template data types (T1 and T2) must be assignable data types. You cannot, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; these requirements are documented on a per-function basis.

Methods

impl PairCFloatCFloat

fn first(&self) -> c_float

C++ method: float QPair<float, float>::first() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: PairCFloatCFloatNewArgs

C++ method: QPair<float, float>::QPair

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pair::PairCFloatCFloat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCFloatCFloat>
C++ method: [constructor] void QPair<float, float>::QPair()

Variant 2

Rust arguments:
1) fn new((&libc::c_float, &libc::c_float, cpp_utils::AsStruct)) -> ::pair::PairCFloatCFloat
2) fn new((&libc::c_float, &libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pair::PairCFloatCFloat>
C++ method: [constructor] void QPair<float, float>::QPair(const float& t1, const float& t2)

fn second(&self) -> c_float

C++ method: float QPair<float, float>::second() const

fn set_first(&mut self, value: c_float)

C++ method: void QPair<float, float>::set_first(float value)

fn set_second(&mut self, value: c_float)

C++ method: void QPair<float, float>::set_second(float value)

fn swap(&mut self, other: &mut PairCFloatCFloat)

C++ method: void QPair<float, float>::swap(QPair<float, float>& other)