Struct qt_gui::pair::PairFilterFilter

pub struct PairFilterFilter {
    // some fields omitted
}

C++ type: QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>

C++ documentation:

The QPair class is a template class that stores a pair of items.

QPair<T1, T2> can be used in your application if the STL pair type is not available. It stores one value of type T1 and one value of type T2. It can be used as a return value for a function that needs to return two values, or as the value type of a generic container.

Here's an example of a QPair that stores one QString and one double value:

QPair<QString, double> pair;

The components are accessible as public data members called first and second. For example:

pair.first = "pi";
pair.second = 3.14159265358979323846;

QPair's template data types (T1 and T2) must be assignable data types. You cannot, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; these requirements are documented on a per-function basis.

Methods

impl PairFilterFilter

fn first(&self) -> Filter

C++ method: QOpenGLTexture::Filter QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::first() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: PairFilterFilterNewArgs

C++ method: QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::QPair

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pair::PairFilterFilter
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairFilterFilter>
C++ method: [constructor] void QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::QPair()

Variant 2

Rust arguments:
1) fn new((&::open_g_l_texture::Filter, &::open_g_l_texture::Filter, cpp_utils::AsStruct)) -> ::pair::PairFilterFilter
2) fn new((&::open_g_l_texture::Filter, &::open_g_l_texture::Filter, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pair::PairFilterFilter>
C++ method: [constructor] void QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::QPair(const QOpenGLTexture::Filter& t1, const QOpenGLTexture::Filter& t2)

fn second(&self) -> Filter

C++ method: QOpenGLTexture::Filter QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::second() const

fn set_first(&mut self, value: Filter)

C++ method: void QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::set_first(QOpenGLTexture::Filter value)

fn set_second(&mut self, value: Filter)

C++ method: void QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::set_second(QOpenGLTexture::Filter value)

fn swap(&mut self, other: &mut PairFilterFilter)

C++ method: void QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>::swap(QPair<QOpenGLTexture::Filter, QOpenGLTexture::Filter>& other)