Struct qt_gui::palette::Palette

pub struct Palette {
    // some fields omitted
}

C++ type: QPalette

C++ documentation:

The QPalette class contains color groups for each widget state.

A palette consists of three color groups: Active, Disabled, and Inactive. All widgets in Qt contain a palette and use their palette to draw themselves. This makes the user interface easily configurable and easier to keep consistent.

If you create a new widget we strongly recommend that you use the colors in the palette rather than hard-coding specific colors.

The color groups:

The Active group is used for the window that has keyboard focus.

The Inactive group is used for other windows.

The Disabled group is used for widgets (not windows) that are disabled for some reason.

Both active and inactive windows can contain disabled widgets. (Disabled widgets are often called inaccessible or grayed out.)

In most styles, Active and Inactive look the same.

Colors and brushes can be set for particular roles in any of a palette's color groups with setColor() and setBrush(). A color group contains a group of colors used by widgets for drawing themselves. We recommend that widgets use color group roles from the palette such as "foreground" and "base" rather than literal colors like "red" or "turquoise". The color roles are enumerated and defined in the ColorRole documentation.

We strongly recommend that you use the default palette of the current style (returned by QGuiApplication::palette()) and modify that as necessary. This is done by Qt's widgets when they are drawn.

To modify a color group you call the functions setColor() and setBrush(), depending on whether you want a pure color or a pixmap pattern.

There are also corresponding color() and brush() getters, and a commonly used convenience function to get the ColorRole for the current ColorGroup: window(), windowText(), base(), etc.

You can copy a palette using the copy constructor and test to see if two palettes are identical using isCopyOf().

QPalette is optimized by the use of implicit sharing, so it is very efficient to pass QPalette objects as arguments.

Warning: Some styles do not use the palette for all drawing, for instance, if they make use of native theme engines. This is the case for both the Windows XP, Windows Vista, and the Mac OS X styles.

Methods

impl Palette

fn alternate_base<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::alternateBase() const

C++ documentation:

Returns the alternate base brush of the current color group.

See also ColorRole and brush().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaletteAsQVariantArgs<'largs>

C++ method: QPalette::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QPalette::operator QVariant() const

C++ documentation:

Returns the palette as a QVariant

fn background<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::background() const

C++ documentation:

Use window() instead.

fn base<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::base() const

C++ documentation:

Returns the base brush of the current color group.

See also ColorRole and brush().

fn bright_text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::brightText() const

C++ documentation:

Returns the bright text foreground brush of the current color group.

See also ColorRole and brush().

fn brush<'largs, Args>(&'largs self, args: Args) -> &'largs Brush where Args: PaletteBrushArgs<'largs>

C++ method: QPalette::brush

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn brush(&self, (::palette::ColorGroup, ::palette::ColorRole)) -> &'l0 ::brush::Brush
C++ method: const QBrush& QPalette::brush(QPalette::ColorGroup cg, QPalette::ColorRole cr) const

C++ documentation:

Returns the brush in the specified color group, used for the given color role.

See also color(), setBrush(), and ColorRole.

Variant 2

Rust arguments: fn brush(&self, ::palette::ColorRole) -> &'l0 ::brush::Brush
C++ method: const QBrush& QPalette::brush(QPalette::ColorRole cr) const

C++ documentation:

This is an overloaded function.

Returns the brush that has been set for the given color role in the current ColorGroup.

See also color(), setBrush(), and ColorRole.

fn button<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::button() const

C++ documentation:

Returns the button brush of the current color group.

See also ColorRole and brush().

fn button_text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::buttonText() const

C++ documentation:

Returns the button text foreground brush of the current color group.

See also ColorRole and brush().

fn cache_key(&self) -> i64

C++ method: qint64 QPalette::cacheKey() const

C++ documentation:

Returns a number that identifies the contents of this QPalette object. Distinct QPalette objects can have the same key if they refer to the same contents.

The cacheKey() will change when the palette is altered.

fn color<'largs, Args>(&'largs self, args: Args) -> &'largs Color where Args: PaletteColorArgs<'largs>

C++ method: QPalette::color

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn color(&self, (::palette::ColorGroup, ::palette::ColorRole)) -> &'l0 ::color::Color
C++ method: const QColor& QPalette::color(QPalette::ColorGroup cg, QPalette::ColorRole cr) const

C++ documentation:

Returns the color in the specified color group, used for the given color role.

See also brush(), setColor(), and ColorRole.

Variant 2

Rust arguments: fn color(&self, ::palette::ColorRole) -> &'l0 ::color::Color
C++ method: const QColor& QPalette::color(QPalette::ColorRole cr) const

C++ documentation:

This is an overloaded function.

Returns the color that has been set for the given color role in the current ColorGroup.

See also brush() and ColorRole.

fn current_color_group(&self) -> ColorGroup

C++ method: QPalette::ColorGroup QPalette::currentColorGroup() const

C++ documentation:

Returns the palette's current color group.

See also setCurrentColorGroup().

fn dark<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::dark() const

C++ documentation:

Returns the dark brush of the current color group.

See also ColorRole and brush().

fn foreground<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::foreground() const

C++ documentation:

Use windowText() instead.

fn highlight<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::highlight() const

C++ documentation:

Returns the highlight brush of the current color group.

See also ColorRole and brush().

fn highlighted_text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::highlightedText() const

C++ documentation:

Returns the highlighted text brush of the current color group.

See also ColorRole and brush().

fn is_brush_set(&self, cg: ColorGroup, cr: ColorRole) -> bool

C++ method: bool QPalette::isBrushSet(QPalette::ColorGroup cg, QPalette::ColorRole cr) const

C++ documentation:

Returns true if the ColorGroup cg and ColorRole cr has been set previously on this palette; otherwise returns false.

This function was introduced in Qt 4.2.

See also setBrush().

fn is_copy_of(&self, p: &Palette) -> bool

C++ method: bool QPalette::isCopyOf(const QPalette& p) const

C++ documentation:

Returns true if this palette and p are copies of each other, i.e. one of them was created as a copy of the other and neither was subsequently modified; otherwise returns false. This is much stricter than equality.

See also operator=() and operator==().

fn is_equal(&self, cr1: ColorGroup, cr2: ColorGroup) -> bool

C++ method: bool QPalette::isEqual(QPalette::ColorGroup cr1, QPalette::ColorGroup cr2) const

C++ documentation:

Returns true (usually quickly) if color group cg1 is equal to cg2; otherwise returns false.

fn light<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::light() const

C++ documentation:

Returns the light brush of the current color group.

See also ColorRole and brush().

fn link<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::link() const

C++ documentation:

Returns the unvisited link text brush of the current color group.

See also ColorRole and brush().

fn link_visited<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::linkVisited() const

C++ documentation:

Returns the visited link text brush of the current color group.

See also ColorRole and brush().

fn mid<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::mid() const

C++ documentation:

Returns the mid brush of the current color group.

See also ColorRole and brush().

fn midlight<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::midlight() const

C++ documentation:

Returns the midlight brush of the current color group.

See also ColorRole and brush().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PaletteNewArgs

C++ method: QPalette::QPalette

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::palette::Palette
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::GlobalColor, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((qt_core::qt::GlobalColor, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(Qt::GlobalColor button)

Variant 3

Rust arguments:
1) fn new((&::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((&::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, &::brush::Brush, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(const QBrush& windowText, const QBrush& button, const QBrush& light, const QBrush& dark, const QBrush& mid, const QBrush& text, const QBrush& bright_text, const QBrush& base, const QBrush& window)

Variant 4

Rust arguments:
1) fn new((&::color::Color, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((&::color::Color, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(const QColor& button)

Variant 5

Rust arguments:
1) fn new((&::color::Color, &::color::Color, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((&::color::Color, &::color::Color, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(const QColor& button, const QColor& window)

Variant 6

Rust arguments:
1) fn new((&::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((&::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, &::color::Color, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(const QColor& windowText, const QColor& window, const QColor& light, const QColor& dark, const QColor& mid, const QColor& text, const QColor& base)

Variant 7

Rust arguments:
1) fn new((&::palette::Palette, cpp_utils::AsStruct)) -> ::palette::Palette
2) fn new((&::palette::Palette, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
C++ method: [constructor] void QPalette::QPalette(const QPalette& palette)

fn op_assign<'l0, 'l1>(&'l0 mut self, palette: &'l1 Palette) -> &'l0 mut Palette

C++ method: QPalette& QPalette::operator=(const QPalette& palette)

C++ documentation:

Assigns p to this palette and returns a reference to this palette.

This operation is fast thanks to implicit sharing.

fn op_eq(&self, p: &Palette) -> bool

C++ method: bool QPalette::operator==(const QPalette& p) const

C++ documentation:

Returns true (usually quickly) if this palette is equal to p; otherwise returns false (slowly).

Note: The current ColorGroup is not taken into account when comparing palettes

See also operator!=().

fn op_neq(&self, p: &Palette) -> bool

C++ method: bool QPalette::operator!=(const QPalette& p) const

C++ documentation:

Returns true (slowly) if this palette is different from p; otherwise returns false (usually quickly).

Note: The current ColorGroup is not taken into account when comparing palettes

See also operator==().

fn resolve<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PaletteResolveArgs<'largs>

C++ method: QPalette::resolve

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn resolve(&self, (&::palette::Palette, cpp_utils::AsBox)) -> cpp_utils::CppBox<::palette::Palette>
2) fn resolve(&self, (&::palette::Palette, cpp_utils::AsStruct)) -> ::palette::Palette
C++ method: QPalette QPalette::resolve(const QPalette& arg1) const

C++ documentation:

Returns a new QPalette that has attributes copied from other.

Variant 2

Rust arguments: fn resolve(&self, ()) -> libc::c_uint
C++ method: unsigned int QPalette::resolve() const

fn resolve_mut(&mut self, mask: c_uint)

C++ method: void QPalette::resolve(unsigned int mask)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for QPalette QPalette::resolve(const QPalette & other) const:

Returns a new QPalette that has attributes copied from other.

fn set_brush<'largs, Args>(&'largs mut self, args: Args) where Args: PaletteSetBrushArgs<'largs>

C++ method: QPalette::setBrush

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_brush(&mut self, (::palette::ColorGroup, ::palette::ColorRole, &::brush::Brush)) -> ()
C++ method: void QPalette::setBrush(QPalette::ColorGroup cg, QPalette::ColorRole cr, const QBrush& brush)

C++ documentation:

This is an overloaded function.

Sets the brush in the specified color group, used for the given color role, to brush.

See also brush(), setColor(), and ColorRole.

Variant 2

Rust arguments: fn set_brush(&mut self, (::palette::ColorRole, &::brush::Brush)) -> ()
C++ method: void QPalette::setBrush(QPalette::ColorRole cr, const QBrush& brush)

C++ documentation:

Sets the brush for the given color role to the specified brush for all groups in the palette.

See also brush(), setColor(), and ColorRole.

fn set_color<'largs, Args>(&'largs mut self, args: Args) where Args: PaletteSetColorArgs<'largs>

C++ method: QPalette::setColor

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_color(&mut self, (::palette::ColorGroup, ::palette::ColorRole, &::color::Color)) -> ()
C++ method: void QPalette::setColor(QPalette::ColorGroup cg, QPalette::ColorRole cr, const QColor& color)

C++ documentation:

Sets the color in the specified color group, used for the given color role, to the specified solid color.

See also setBrush(), color(), and ColorRole.

Variant 2

Rust arguments: fn set_color(&mut self, (::palette::ColorRole, &::color::Color)) -> ()
C++ method: void QPalette::setColor(QPalette::ColorRole cr, const QColor& color)

C++ documentation:

This is an overloaded function.

Sets the color used for the given color role, in all color groups, to the specified solid color.

See also brush(), setColor(), and ColorRole.

fn set_color_group(&mut self, cr: ColorGroup, window_text: &Brush, button: &Brush, light: &Brush, dark: &Brush, mid: &Brush, text: &Brush, bright_text: &Brush, base: &Brush, window: &Brush)

C++ method: void QPalette::setColorGroup(QPalette::ColorGroup cr, const QBrush& windowText, const QBrush& button, const QBrush& light, const QBrush& dark, const QBrush& mid, const QBrush& text, const QBrush& bright_text, const QBrush& base, const QBrush& window)

C++ documentation:

Sets a the group at cg. You can pass either brushes, pixmaps or plain colors for windowText, button, light, dark, mid, text, bright_text, base and window.

See also QBrush.

fn set_current_color_group(&mut self, cg: ColorGroup)

C++ method: void QPalette::setCurrentColorGroup(QPalette::ColorGroup cg)

C++ documentation:

Set the palette's current color group to cg.

See also currentColorGroup().

fn shadow<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::shadow() const

C++ documentation:

Returns the shadow brush of the current color group.

See also ColorRole and brush().

fn swap(&mut self, other: &mut Palette)

C++ method: void QPalette::swap(QPalette& other)

C++ documentation:

Swaps this palette instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::text() const

C++ documentation:

Returns the text foreground brush of the current color group.

See also ColorRole and brush().

fn tool_tip_base<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::toolTipBase() const

C++ documentation:

Returns the tool tip base brush of the current color group. This brush is used by QToolTip and QWhatsThis.

Note: Tool tips use the Inactive color group of QPalette, because tool tips are not active windows.

This function was introduced in Qt 4.4.

See also ColorRole and brush().

fn tool_tip_text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::toolTipText() const

C++ documentation:

Returns the tool tip text brush of the current color group. This brush is used by QToolTip and QWhatsThis.

Note: Tool tips use the Inactive color group of QPalette, because tool tips are not active windows.

This function was introduced in Qt 4.4.

See also ColorRole and brush().

fn window<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::window() const

C++ documentation:

Returns the window (general background) brush of the current color group.

See also ColorRole and brush().

fn window_text<'l0>(&'l0 self) -> &'l0 Brush

C++ method: const QBrush& QPalette::windowText() const

C++ documentation:

Returns the window text (general foreground) brush of the current color group.

See also ColorRole and brush().