Module qt_gui::pdf_writer::overloading

Traits

PdfWriterConnectArgs
This trait represents a set of arguments accepted by PdfWriter::connect method.

PdfWriterConnectStaticArgs
This trait represents a set of arguments accepted by PdfWriter::connect_static method.

PdfWriterCreatorArgs
This trait represents a set of arguments accepted by PdfWriter::creator method.

PdfWriterDestroyedArgs
This trait represents a set of arguments accepted by PdfWriter::destroyed method.

PdfWriterDisconnectArgs
This trait represents a set of arguments accepted by PdfWriter::disconnect method.

PdfWriterDisconnectStaticArgs
This trait represents a set of arguments accepted by PdfWriter::disconnect_static method.

PdfWriterDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PdfWriter::dynamic_property_names method.

PdfWriterMarginsArgs
This trait represents a set of arguments accepted by PdfWriter::margins method.

PdfWriterNewArgs
This trait represents a set of arguments accepted by PdfWriter::new method.

PdfWriterObjectNameArgs
This trait represents a set of arguments accepted by PdfWriter::object_name method.

PdfWriterPageLayoutArgs
This trait represents a set of arguments accepted by PdfWriter::page_layout method.

PdfWriterPageSizeMMArgs
This trait represents a set of arguments accepted by PdfWriter::page_size_m_m method.

PdfWriterPropertyArgs
This trait represents a set of arguments accepted by PdfWriter::property method.

PdfWriterSetPageMarginsArgs
This trait represents a set of arguments accepted by PdfWriter::set_page_margins method.

PdfWriterSetPageSizeArgs
This trait represents a set of arguments accepted by PdfWriter::set_page_size method.

PdfWriterStartTimerArgs
This trait represents a set of arguments accepted by PdfWriter::start_timer method.

PdfWriterTitleArgs
This trait represents a set of arguments accepted by PdfWriter::title method.

PdfWriterTrArgs
This trait represents a set of arguments accepted by PdfWriter::tr method.

PdfWriterTrUtf8Args
This trait represents a set of arguments accepted by PdfWriter::tr_utf8 method.