Struct qt_gui::picture::Picture

pub struct Picture {
    // some fields omitted
}

C++ type: QPicture

C++ documentation:

The QPicture class is a paint device that records and replays QPainter commands.

A picture serializes painter commands to an IO device in a platform-independent format. They are sometimes referred to as meta-files.

Qt pictures use a proprietary binary format. Unlike native picture (meta-file) formats on many window systems, Qt pictures have no limitations regarding their contents. Everything that can be painted on a widget or pixmap (e.g., fonts, pixmaps, regions, transformed graphics, etc.) can also be stored in a picture.

QPicture is resolution independent, i.e. a QPicture can be displayed on different devices (for example svg, pdf, ps, printer and screen) looking the same. This is, for instance, needed for WYSIWYG print preview. QPicture runs in the default system dpi, and scales the painter to match differences in resolution depending on the window system.

Example of how to record a picture:

        QPicture picture;
        QPainter painter;
        painter.begin(&picture);           // paint in picture
        painter.drawEllipse(10,20, 80,70); // draw an ellipse
        painter.end();                     // painting done
        picture.save("drawing.pic");       // save picture

Note that the list of painter commands is reset on each call to the QPainter::begin() function.

Example of how to replay a picture:

        QPicture picture;
        picture.load("drawing.pic");           // load picture
        QPainter painter;
        painter.begin(&myImage);               // paint in myImage
        painter.drawPicture(0, 0, picture);    // draw the picture at (0,0)
        painter.end();                         // painting done

Pictures can also be drawn using play(). Some basic data about a picture is available, for example, size(), isNull() and boundingRect().

Methods

impl Picture

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PictureBoundingRectArgs<'largs>

C++ method: QPicture::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn bounding_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPicture::boundingRect() const

C++ documentation:

Returns the picture's bounding rectangle or an invalid rectangle if the picture contains no data.

See also setBoundingRect().

fn color_count(&self) -> c_int

C++ method: int QPicture::colorCount() const

C++ documentation:

Returns the number of different colors available for the paint device. If the number of colors available is too great to be represented by the int data type, then INT_MAX will be returned instead.

fn data(&self) -> *const c_char

C++ method: const char* QPicture::data() const

C++ documentation:

Returns a pointer to the picture data. The pointer is only valid until the next non-const function is called on this picture. The returned pointer is 0 if the picture contains no data.

See also setData(), size(), and isNull().

fn depth(&self) -> c_int

C++ method: int QPicture::depth() const

C++ documentation:

Returns the bit depth (number of bit planes) of the paint device.

fn detach(&mut self)

C++ method: void QPicture::detach()

fn dev_type(&self) -> c_int

C++ method: virtual int QPicture::devType() const

fn device_pixel_ratio(&self) -> c_int

C++ method: int QPicture::devicePixelRatio() const

C++ documentation:

Returns the device pixel ratio for device.

Common values are 1 for normal-dpi displays and 2 for high-dpi "retina" displays.

fn height(&self) -> c_int

C++ method: int QPicture::height() const

C++ documentation:

Returns the height of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also heightMM().

fn height_m_m(&self) -> c_int

C++ method: int QPicture::heightMM() const

C++ documentation:

Returns the height of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also height().

fn input_format_list<Args>(args: Args) -> Args::ReturnType where Args: PictureInputFormatListArgs

C++ method: QPicture::inputFormatList

This is an overloaded function. Available variants:

Rust arguments:
1) fn input_format_list(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn input_format_list(cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: static QStringList QPicture::inputFormatList()

C++ documentation:

Returns a list of picture formats that are supported for picture input.

Note that if you want to iterate over the list, you should iterate over a copy, e.g.

        QStringList list = QPicture::inputFormatList();
        foreach (const QString &string, list)
            myProcessing(string);

See also outputFormatList(), inputFormats(), and QPictureIO.

fn input_formats<Args>(args: Args) -> Args::ReturnType where Args: PictureInputFormatsArgs

C++ method: QPicture::inputFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn input_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn input_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QPicture::inputFormats()

C++ documentation:

Returns a list of picture formats that are supported for picture input.

See also outputFormats(), inputFormatList(), and QPictureIO.

fn is_detached(&self) -> bool

C++ method: bool QPicture::isDetached() const

fn is_null(&self) -> bool

C++ method: bool QPicture::isNull() const

C++ documentation:

Returns true if the picture contains no data; otherwise returns false.

fn load<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: PictureLoadArgs<'largs>

C++ method: QPicture::load

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn load(&mut self, *mut qt_core::qiod_evice::QiodEvice) -> bool
2) fn load(&mut self, (*mut qt_core::qiod_evice::QiodEvice, *const libc::c_char)) -> bool
C++ method: bool QPicture::load(QIODevice* dev, const char* format = ?)

C++ documentation:

This is an overloaded function.

dev is the device to use for loading.

Variant 2

Rust arguments:
1) fn load(&mut self, &qt_core::string::String) -> bool
2) fn load(&mut self, (&qt_core::string::String, *const libc::c_char)) -> bool
C++ method: bool QPicture::load(const QString& fileName, const char* format = ?)

C++ documentation:

Loads a picture from the file specified by fileName and returns true if successful; otherwise invalidates the picture and returns false.

Please note that the format parameter has been deprecated and will have no effect.

See also save().

fn logical_dpi_x(&self) -> c_int

C++ method: int QPicture::logicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from widthMM().

Note that if the logicalDpiX() doesn't equal the physicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiY() and physicalDpiX().

fn logical_dpi_y(&self) -> c_int

C++ method: int QPicture::logicalDpiY() const

C++ documentation:

Returns the vertical resolution of the device in dots per inch, which is used when computing font sizes. For X11, this is usually the same as could be computed from heightMM().

Note that if the logicalDpiY() doesn't equal the physicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also logicalDpiX() and physicalDpiY().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PictureNewArgs

C++ method: QPicture::QPicture

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::picture::Picture, cpp_utils::AsStruct)) -> ::picture::Picture
2) fn new((&::picture::Picture, cpp_utils::AsBox)) -> cpp_utils::CppBox<::picture::Picture>
C++ method: [constructor] void QPicture::QPicture(const QPicture& arg1)

Variant 2

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::picture::Picture
2) fn new(cpp_utils::AsStruct) -> ::picture::Picture
3) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::picture::Picture>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::picture::Picture>
C++ method: [constructor] void QPicture::QPicture(int formatVersion = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, p: &'l1 Picture) -> &'l0 mut Picture

C++ method: QPicture& QPicture::operator=(const QPicture& p)

C++ documentation:

Assigns picture p to this picture and returns a reference to this picture.

fn output_format_list<Args>(args: Args) -> Args::ReturnType where Args: PictureOutputFormatListArgs

C++ method: QPicture::outputFormatList

This is an overloaded function. Available variants:

Rust arguments:
1) fn output_format_list(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn output_format_list(cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: static QStringList QPicture::outputFormatList()

C++ documentation:

Returns a list of picture formats that are supported for picture output.

Note that if you want to iterate over the list, you should iterate over a copy, e.g.

        QStringList list = QPicture::outputFormatList();
        foreach (const QString &string, list)
            myProcessing(string);

See also inputFormatList(), outputFormats(), and QPictureIO.

fn output_formats<Args>(args: Args) -> Args::ReturnType where Args: PictureOutputFormatsArgs

C++ method: QPicture::outputFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn output_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn output_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QPicture::outputFormats()

C++ documentation:

Returns a list of picture formats that are supported for picture output.

See also inputFormats(), outputFormatList(), and QPictureIO.

fn paint_engine(&self) -> *mut PaintEngine

C++ method: virtual QPaintEngine* QPicture::paintEngine() const

fn painting_active(&self) -> bool

C++ method: bool QPicture::paintingActive() const

C++ documentation:

Returns true if the device is currently being painted on, i.e. someone has called QPainter::begin() but not yet called QPainter::end() for this device; otherwise returns false.

See also QPainter::isActive().

fn physical_dpi_x(&self) -> c_int

C++ method: int QPicture::physicalDpiX() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiX() doesn't equal the logicalDpiX(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiY() and logicalDpiX().

fn physical_dpi_y(&self) -> c_int

C++ method: int QPicture::physicalDpiY() const

C++ documentation:

Returns the horizontal resolution of the device in dots per inch. For example, when printing, this resolution refers to the physical printer's resolution. The logical DPI on the other hand, refers to the resolution used by the actual paint engine.

Note that if the physicalDpiY() doesn't equal the logicalDpiY(), the corresponding QPaintEngine must handle the resolution mapping.

See also physicalDpiX() and logicalDpiY().

fn picture_format(file_name: &String) -> *const c_char

C++ method: static const char* QPicture::pictureFormat(const QString& fileName)

C++ documentation:

Returns a string that specifies the picture format of the file fileName, or 0 if the file cannot be read or if the format is not recognized.

See also load() and save().

fn play(&mut self, p: *mut Painter) -> bool

C++ method: bool QPicture::play(QPainter* p)

C++ documentation:

Replays the picture using painter, and returns true if successful; otherwise returns false.

This function does exactly the same as QPainter::drawPicture() with (x, y) = (0, 0).

fn save<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: PictureSaveArgs<'largs>

C++ method: QPicture::save

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn save(&mut self, *mut qt_core::qiod_evice::QiodEvice) -> bool
2) fn save(&mut self, (*mut qt_core::qiod_evice::QiodEvice, *const libc::c_char)) -> bool
C++ method: bool QPicture::save(QIODevice* dev, const char* format = ?)

C++ documentation:

This is an overloaded function.

dev is the device to use for saving.

Variant 2

Rust arguments:
1) fn save(&mut self, &qt_core::string::String) -> bool
2) fn save(&mut self, (&qt_core::string::String, *const libc::c_char)) -> bool
C++ method: bool QPicture::save(const QString& fileName, const char* format = ?)

C++ documentation:

Saves a picture to the file specified by fileName and returns true if successful; otherwise returns false.

Please note that the format parameter has been deprecated and will have no effect.

See also load().

fn set_bounding_rect(&mut self, r: &Rect)

C++ method: void QPicture::setBoundingRect(const QRect& r)

C++ documentation:

Sets the picture's bounding rectangle to r. The automatically calculated value is overridden.

See also boundingRect().

fn set_data(&mut self, data: *const c_char, size: c_uint)

C++ method: virtual void QPicture::setData(const char* data, unsigned int size)

C++ documentation:

Sets the picture data directly from data and size. This function copies the input data.

See also data() and size().

fn size(&self) -> c_uint

C++ method: unsigned int QPicture::size() const

C++ documentation:

Returns the size of the picture data.

See also data().

fn swap(&mut self, other: &mut Picture)

C++ method: void QPicture::swap(QPicture& other)

C++ documentation:

Swaps picture other with this picture. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn width(&self) -> c_int

C++ method: int QPicture::width() const

C++ documentation:

Returns the width of the paint device in default coordinate system units (e.g. pixels for QPixmap and QWidget).

See also widthMM().

fn width_m_m(&self) -> c_int

C++ method: int QPicture::widthMM() const

C++ documentation:

Returns the width of the paint device in millimeters. Due to platform limitations it may not be possible to use this function to determine the actual physical size of a widget on the screen.

See also width().