Module qt_gui::picture_format_plugin::overloading

Traits

PictureFormatPluginConnectArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::connect method.

PictureFormatPluginConnectStaticArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::connect_static method.

PictureFormatPluginDestroyedArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::destroyed method.

PictureFormatPluginDisconnectArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::disconnect method.

PictureFormatPluginDisconnectStaticArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::disconnect_static method.

PictureFormatPluginDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::dynamic_property_names method.

PictureFormatPluginObjectNameArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::object_name method.

PictureFormatPluginPropertyArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::property method.

PictureFormatPluginStartTimerArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::start_timer method.

PictureFormatPluginTrArgs
This trait represents a set of arguments accepted by PictureFormatPlugin::tr method.

PictureFormatPluginTrUtf8Args
This trait represents a set of arguments accepted by PictureFormatPlugin::tr_utf8 method.