Struct qt_gui::picture_i_o::PictureIO

pub struct PictureIO {
    // some fields omitted
}

C++ type: QPictureIO

C++ documentation:

The QPictureIO class contains parameters for loading and saving pictures.

QPictureIO contains a QIODevice object that is used for picture data I/O. The programmer can install new picture file formats in addition to those that Qt provides.

You don't normally need to use this class; QPicture::load(), QPicture::save().

Methods

impl PictureIO

fn define_ioh_andler(format: *const c_char, header: *const c_char, flags: *const c_char, read_picture: extern fn(*mut PictureIO), write_picture: extern fn(*mut PictureIO))

C++ method: static void QPictureIO::defineIOHandler(const char* format, const char* header, const char* flags, void (*FN_PTR)(QPictureIO*) read_picture, void (*FN_PTR)(QPictureIO*) write_picture)

C++ documentation:

Defines a picture I/O handler for the picture format called format, which is recognized using the regular expression defined in header, read using readPicture and written using writePicture.

flags is a string of single-character flags for this format. The only flag defined currently is T (upper case), so the only legal value for flags are "T" and the empty string. The "T" flag means that the picture file is a text file, and Qt should treat all newline conventions as equivalent. (XPM files and some PPM files are text files for example.)

format is used to select a handler to write a QPicture; header is used to select a handler to read an picture file.

If readPicture is a null pointer, the QPictureIO will not be able to read pictures in format. If writePicture is a null pointer, the QPictureIO will not be able to write pictures in format. If both are null, the QPictureIO object is valid but useless.

Example:

void readSVG(QPictureIO *picture)
{
    // read the picture using the picture->ioDevice()
}

void writeSVG(QPictureIO *picture) { // write the picture using the picture->ioDevice() }

// add the SVG picture handler // ...
Before the regular expression test, all the 0 bytes in the file header are converted to 1 bytes. This is done because when Qt was ASCII-based, QRegExp could not handle 0 bytes in strings.

The regexp is only applied on the first 14 bytes of the file.

(Note that if one handlerIO supports writing a format and another supports reading it, Qt supports both reading and writing. If two handlers support the same operation, Qt chooses one arbitrarily.)

fn description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PictureIODescriptionArgs<'largs>

C++ method: QPictureIO::description

This is an overloaded function. Available variants:

Rust arguments:
1) fn description(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn description(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QPictureIO::description() const

C++ documentation:

Returns the picture description string.

See also setDescription().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PictureIOFileNameArgs<'largs>

C++ method: QPictureIO::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QPictureIO::fileName() const

C++ documentation:

Returns the file name currently set.

See also setFileName().

fn format(&self) -> *const c_char

C++ method: const char* QPictureIO::format() const

C++ documentation:

Returns the picture format string or 0 if no format has been explicitly set.

See also setFormat().

fn gamma(&self) -> c_float

C++ method: float QPictureIO::gamma() const

C++ documentation:

Returns the gamma value at which the picture will be viewed.

See also setGamma().

fn input_formats<Args>(args: Args) -> Args::ReturnType where Args: PictureIOInputFormatsArgs

C++ method: QPictureIO::inputFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn input_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn input_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QPictureIO::inputFormats()

C++ documentation:

Returns a sorted list of picture formats that are supported for picture input.

fn io_device(&self) -> *mut QiodEvice

C++ method: QIODevice* QPictureIO::ioDevice() const

C++ documentation:

Returns the IO device currently set.

See also setIODevice().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PictureIONewArgs

C++ method: QPictureIO::QPictureIO

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::picture_i_o::PictureIO
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::picture_i_o::PictureIO>
C++ method: [constructor] void QPictureIO::QPictureIO()

Variant 2

Rust arguments:
1) fn new((*mut qt_core::qiod_evice::QiodEvice, *const libc::c_char, cpp_utils::AsStruct)) -> ::picture_i_o::PictureIO
2) fn new((*mut qt_core::qiod_evice::QiodEvice, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::picture_i_o::PictureIO>
C++ method: [constructor] void QPictureIO::QPictureIO(QIODevice* ioDevice, const char* format)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, *const libc::c_char, cpp_utils::AsStruct)) -> ::picture_i_o::PictureIO
2) fn new((&qt_core::string::String, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::picture_i_o::PictureIO>
C++ method: [constructor] void QPictureIO::QPictureIO(const QString& fileName, const char* format)

fn output_formats<Args>(args: Args) -> Args::ReturnType where Args: PictureIOOutputFormatsArgs

C++ method: QPictureIO::outputFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn output_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn output_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QPictureIO::outputFormats()

C++ documentation:

Returns a sorted list of picture formats that are supported for picture output.

fn parameters(&self) -> *const c_char

C++ method: const char* QPictureIO::parameters() const

C++ documentation:

Returns the picture's parameters string.

See also setParameters().

fn picture<'l0>(&'l0 self) -> &'l0 Picture

C++ method: const QPicture& QPictureIO::picture() const

C++ documentation:

Returns the picture currently set.

See also setPicture().

fn picture_format<Args>(args: Args) -> Args::ReturnType where Args: PictureIOPictureFormatArgs

C++ method: QPictureIO::pictureFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn picture_format((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn picture_format((*mut qt_core::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: static QByteArray QPictureIO::pictureFormat(QIODevice* arg1)

C++ documentation:

This is an overloaded function.

Returns a string that specifies the picture format of the picture read from IO device d, or 0 if the device cannot be read or if the format is not recognized.

Make sure that d is at the right position in the device (for example, at the beginning of the file).

See also QIODevice::pos().

Variant 2

Rust arguments:
1) fn picture_format((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn picture_format((&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: static QByteArray QPictureIO::pictureFormat(const QString& fileName)

C++ documentation:

Returns a string that specifies the picture format of the file fileName, or null if the file cannot be read or if the format is not recognized.

fn quality(&self) -> c_int

C++ method: int QPictureIO::quality() const

C++ documentation:

Returns the quality of the written picture, related to the compression ratio.

See also setQuality() and QPicture::save().

fn read(&mut self) -> bool

C++ method: bool QPictureIO::read()

C++ documentation:

Reads an picture into memory and returns true if the picture was successfully read; otherwise returns false.

Before reading an picture you must set an IO device or a file name. If both an IO device and a file name have been set, the IO device will be used.

Setting the picture file format string is optional.

Note that this function does not set the format used to read the picture. If you need that information, use the pictureFormat() static functions.

Example:

        QPictureIO iio;
        QPixmap  pixmap;
        iio.setFileName("vegeburger.pic");
        if (iio.read()) {        // OK
            QPicture picture = iio.picture();
            QPainter painter(&pixmap);
            painter.drawPicture(0, 0, picture);
        }

See also setIODevice(), setFileName(), setFormat(), write(), and QPixmap::load().

fn set_description(&mut self, arg1: &String)

C++ method: void QPictureIO::setDescription(const QString& arg1)

C++ documentation:

Sets the picture description string for picture handlers that support picture descriptions to description.

Currently, no picture format supported by Qt uses the description string.

See also description().

fn set_file_name(&mut self, arg1: &String)

C++ method: void QPictureIO::setFileName(const QString& arg1)

C++ documentation:

Sets the name of the file to read or write an picture from to fileName.

See also fileName() and setIODevice().

fn set_format(&mut self, arg1: *const c_char)

C++ method: void QPictureIO::setFormat(const char* arg1)

C++ documentation:

Sets the picture format to format for the picture to be read or written.

It is necessary to specify a format before writing an picture, but it is not necessary to specify a format before reading an picture.

If no format has been set, Qt guesses the picture format before reading it. If a format is set the picture will only be read if it has that format.

See also read(), write(), and format().

fn set_gamma(&mut self, arg1: c_float)

C++ method: void QPictureIO::setGamma(float arg1)

C++ documentation:

Sets the gamma value at which the picture will be viewed to gamma. If the picture format stores a gamma value for which the picture is intended to be used, then this setting will be used to modify the picture. Setting to 0.0 will disable gamma correction (i.e. any specification in the file will be ignored).

The default value is 0.0.

See also gamma().

fn set_iod_evice(&mut self, arg1: *mut QiodEvice)

C++ method: void QPictureIO::setIODevice(QIODevice* arg1)

C++ documentation:

Sets the IO device to be used for reading or writing an picture.

Setting the IO device allows pictures to be read/written to any block-oriented QIODevice.

If ioDevice is not null, this IO device will override file name settings.

See also setFileName().

fn set_parameters(&mut self, arg1: *const c_char)

C++ method: void QPictureIO::setParameters(const char* arg1)

C++ documentation:

Sets the picture's parameter string to parameters. This is for picture handlers that require special parameters.

Although the current picture formats supported by Qt ignore the parameters string, it may be used in future extensions or by contributions (for example, JPEG).

See also parameters().

fn set_picture(&mut self, arg1: &Picture)

C++ method: void QPictureIO::setPicture(const QPicture& arg1)

C++ documentation:

Sets the picture to picture.

See also picture().

fn set_quality(&mut self, arg1: c_int)

C++ method: void QPictureIO::setQuality(int arg1)

C++ documentation:

Sets the quality of the written picture to q, related to the compression ratio.

q must be in the range -1..100. Specify 0 to obtain small compressed files, 100 for large uncompressed files. (-1 signifies the default compression.)

See also quality() and QPicture::save().

fn set_status(&mut self, arg1: c_int)

C++ method: void QPictureIO::setStatus(int arg1)

C++ documentation:

Sets the picture IO status to status. A non-zero value indicates an error, whereas 0 means that the IO operation was successful.

See also status().

fn status(&self) -> c_int

C++ method: int QPictureIO::status() const

C++ documentation:

Returns the picture's IO status. A non-zero value indicates an error, whereas 0 means that the IO operation was successful.

See also setStatus().

fn write(&mut self) -> bool

C++ method: bool QPictureIO::write()

C++ documentation:

Writes an picture to an IO device and returns true if the picture was successfully written; otherwise returns false.

Before writing an picture you must set an IO device or a file name. If both an IO device and a file name have been set, the IO device will be used.

The picture will be written using the specified picture format.

Example:

        QPictureIO iio;
        QPicture   picture;
        QPainter painter(&picture);
        painter.drawPixmap(0, 0, pixmap);
        iio.setPicture(picture);
        iio.setFileName("vegeburger.pic");
        iio.setFormat("PIC");
        if (iio.write())
            return true; // returned true if written successfully

See also setIODevice(), setFileName(), setFormat(), read(), and QPixmap::save().