Function qt_gui::pixel_format::pixel_format_alpha

pub fn pixel_format_alpha<Args>(args: Args) -> Args::ReturnType where Args: PixelFormatAlphaArgs

C++ method: qPixelFormatAlpha

This is an overloaded function. Available variants:

Rust arguments:
1) fn pixel_format_alpha((libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
2) fn pixel_format_alpha((libc::c_uchar, ::pixel_format::TypeInterpretation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
3) fn pixel_format_alpha((libc::c_uchar, cpp_utils::AsStruct)) -> ::pixel_format::PixelFormat
4) fn pixel_format_alpha((libc::c_uchar, ::pixel_format::TypeInterpretation, cpp_utils::AsStruct)) -> ::pixel_format::PixelFormat
C++ method: QPixelFormat qPixelFormatAlpha(unsigned char channelSize, QPixelFormat::TypeInterpretation typeInt = ?)

C++ documentation:

Constructor function for creating an Alpha format. A mask format can be described by passing 1 to channelSize. Its also possible to define very accurate alpha formats using doubles to describe each pixel by passing 8 as channelSize and FloatingPoint as typeInterpretation.

This function was introduced in Qt 5.5.

See also QPixelFormat::TypeInterpretation.