Module qt_gui::pixel_format::overloading

Traits

PixelFormatAlphaArgs
This trait represents a set of arguments accepted by pixel_format_alpha method.

PixelFormatCmykArgs
This trait represents a set of arguments accepted by pixel_format_cmyk method.

PixelFormatGrayscaleArgs
This trait represents a set of arguments accepted by pixel_format_grayscale method.

PixelFormatHslArgs
This trait represents a set of arguments accepted by pixel_format_hsl method.

PixelFormatHsvArgs
This trait represents a set of arguments accepted by pixel_format_hsv method.

PixelFormatNewArgs
This trait represents a set of arguments accepted by PixelFormat::new method.

PixelFormatRgbaArgs
This trait represents a set of arguments accepted by pixel_format_rgba method.

PixelFormatYuvArgs
This trait represents a set of arguments accepted by pixel_format_yuv method.