Struct qt_gui::pixel_format::PixelFormat

pub struct PixelFormat {
    // some fields omitted
}

C++ type: QPixelFormat

C++ documentation:

QPixelFormat is a class for describing different pixel layouts in graphics buffers

In Qt there is a often a need to represent the layout of the pixels in a graphics buffer. Internally QPixelFormat stores everything in a 64 bit datastructure. This gives performance but also some limitations.

QPixelFormat can describe 5 color channels and 1 alpha channel, each can use 6 bits to describe the size of the color channel.

The position of the alpha channel is described with a separate enum. This is to make it possible to describe QImage formats like ARGB32, and also describe typical OpenGL formats like RBGA8888.

How pixels are suppose to be read is determined by the TypeInterpretation enum. It describes if color values are suppose to be read byte per byte, or if a pixel is suppose to be read as a complete int and then masked.

There is no support for describing YUV's macro pixels. Instead a list of YUV formats has been made. When a QPixelFormat is describing a YUV format, the bitsPerPixel value has been deduced by the YUV Layout enum. Also, the color channels should all be set to zero except the fifth color channel that should store the bitsPerPixel value.

Methods

impl PixelFormat

fn alpha_position(&self) -> AlphaPosition

C++ method: QPixelFormat::AlphaPosition QPixelFormat::alphaPosition() const

C++ documentation:

Accessor function for alphaPosition.

fn alpha_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::alphaSize() const

C++ documentation:

Accessor function for the alpha channel size.

fn alpha_usage(&self) -> AlphaUsage

C++ method: QPixelFormat::AlphaUsage QPixelFormat::alphaUsage() const

C++ documentation:

Accessor function for alphaUsage.

fn bits_per_pixel(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::bitsPerPixel() const

C++ documentation:

Accessor function for the bits used per pixel. This function returns the sum of the color channels + the size of the alpha channel.

fn black_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::blackSize() const

C++ documentation:

Accessor function for the black/key color channel.

fn blue_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::blueSize() const

C++ documentation:

Accessor function for the size of the blue color channel.

fn brightness_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::brightnessSize() const

C++ documentation:

Accessor function for the brightness channel size.

fn byte_order(&self) -> ByteOrder

C++ method: QPixelFormat::ByteOrder QPixelFormat::byteOrder() const

C++ documentation:

The byte order is almost always set the the byte order of the current system. However, it can be useful to describe some YUV formats. This function should never return QPixelFormat::CurrentSystemEndian as this value is translated to a endian value in the constructor.

fn channel_count(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::channelCount() const

C++ documentation:

Accessor function for getting the channelCount. Channel Count is deduced by color channels with a size > 0 and if the size of the alpha channel is > 0.

fn color_model(&self) -> ColorModel

C++ method: QPixelFormat::ColorModel QPixelFormat::colorModel() const

C++ documentation:

Accessor function for getting the colorModel.

fn cyan_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::cyanSize() const

C++ documentation:

Accessor function for the cyan color channel.

fn green_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::greenSize() const

C++ documentation:

Accessor function for the size of the green color channel.

fn hue_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::hueSize() const

C++ documentation:

Accessor function for the hue channel size.

fn lightness_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::lightnessSize() const

C++ documentation:

Accessor function for the lightness channel size.

fn magenta_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::magentaSize() const

C++ documentation:

Accessor function for the megenta color channel.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PixelFormatNewArgs

C++ method: QPixelFormat::QPixelFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pixel_format::PixelFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
C++ method: [constructor] void QPixelFormat::QPixelFormat()

Variant 2

Rust arguments:
1) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, cpp_utils::AsStruct)) -> ::pixel_format::PixelFormat
2) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, ::pixel_format::ByteOrder, cpp_utils::AsStruct)) -> ::pixel_format::PixelFormat
3) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, ::pixel_format::ByteOrder, libc::c_uchar, cpp_utils::AsStruct)) -> ::pixel_format::PixelFormat
4) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
5) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, ::pixel_format::ByteOrder, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
6) fn new((::pixel_format::ColorModel, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, ::pixel_format::AlphaUsage, ::pixel_format::AlphaPosition, ::pixel_format::AlphaPremultiplied, ::pixel_format::TypeInterpretation, ::pixel_format::ByteOrder, libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixel_format::PixelFormat>
C++ method: [constructor] void QPixelFormat::QPixelFormat(QPixelFormat::ColorModel colorModel, unsigned char firstSize, unsigned char secondSize, unsigned char thirdSize, unsigned char fourthSize, unsigned char fifthSize, unsigned char alphaSize, QPixelFormat::AlphaUsage alphaUsage, QPixelFormat::AlphaPosition alphaPosition, QPixelFormat::AlphaPremultiplied premultiplied, QPixelFormat::TypeInterpretation typeInterpretation, QPixelFormat::ByteOrder byteOrder = ?, unsigned char subEnum = ?)

fn premultiplied(&self) -> AlphaPremultiplied

C++ method: QPixelFormat::AlphaPremultiplied QPixelFormat::premultiplied() const

C++ documentation:

Accessor function for the AlphaPremultiplied enum. This indicates if the alpha channel is multiplied in to the color channels.

fn red_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::redSize() const

C++ documentation:

Accessor function for the size of the red color channel.

fn saturation_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::saturationSize() const

C++ documentation:

Accessor function for the saturation channel size.

fn sub_enum(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::subEnum() const

fn type_interpretation(&self) -> TypeInterpretation

C++ method: QPixelFormat::TypeInterpretation QPixelFormat::typeInterpretation() const

C++ documentation:

Accessor function for the type representation of a color channel or a pixel.

See also TypeInterpretation.

fn yellow_size(&self) -> c_uchar

C++ method: unsigned char QPixelFormat::yellowSize() const

C++ documentation:

Accessor function for the yellow color channel.

fn yuv_layout(&self) -> YuvlAyout

C++ method: QPixelFormat::YUVLayout QPixelFormat::yuvLayout() const

C++ documentation:

Accessor function for the YUVLayout. It is difficult to describe the color channels of a YUV pixel format since YUV color model uses macro pixels. Instead the layout of the pixels are stored as an enum.