Module qt_gui::pixmap::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

PixmapAsQVariantArgs
This trait represents a set of arguments accepted by Pixmap::as_q_variant method.

PixmapConvertFromImageArgs
This trait represents a set of arguments accepted by Pixmap::convert_from_image method.

PixmapCopyArgs
This trait represents a set of arguments accepted by Pixmap::copy method.

PixmapCreateHeuristicMaskArgs
This trait represents a set of arguments accepted by Pixmap::create_heuristic_mask method.

PixmapCreateMaskFromColorArgs
This trait represents a set of arguments accepted by Pixmap::create_mask_from_color method.

PixmapFillArgs
This trait represents a set of arguments accepted by Pixmap::fill method.

PixmapFromImageArgs
This trait represents a set of arguments accepted by Pixmap::from_image method.

PixmapFromImageReaderArgs
This trait represents a set of arguments accepted by Pixmap::from_image_reader method.

PixmapGrabWidgetArgs
This trait represents a set of arguments accepted by Pixmap::grab_widget method.

PixmapGrabWindowArgs
This trait represents a set of arguments accepted by Pixmap::grab_window method.

PixmapLoadArgs
This trait represents a set of arguments accepted by Pixmap::load method.

PixmapLoadFromDataArgs
This trait represents a set of arguments accepted by Pixmap::load_from_data method.

PixmapMaskArgs
This trait represents a set of arguments accepted by Pixmap::mask method.

PixmapNewArgs
This trait represents a set of arguments accepted by Pixmap::new method.

PixmapRectArgs
This trait represents a set of arguments accepted by Pixmap::rect method.

PixmapSaveArgs
This trait represents a set of arguments accepted by Pixmap::save method.

PixmapScaledArgs
This trait represents a set of arguments accepted by Pixmap::scaled method.

PixmapScaledToHeightArgs
This trait represents a set of arguments accepted by Pixmap::scaled_to_height method.

PixmapScaledToWidthArgs
This trait represents a set of arguments accepted by Pixmap::scaled_to_width method.

PixmapScrollArgs
This trait represents a set of arguments accepted by Pixmap::scroll method.

PixmapSizeArgs
This trait represents a set of arguments accepted by Pixmap::size method.

PixmapToImageArgs
This trait represents a set of arguments accepted by Pixmap::to_image method.

PixmapTransformedArgs
This trait represents a set of arguments accepted by Pixmap::transformed method.

PixmapTrueMatrixArgs
This trait represents a set of arguments accepted by Pixmap::true_matrix method.