Struct qt_gui::pixmap_cache::Key

pub struct Key {
    // some fields omitted
}

C++ type: QPixmapCache::Key

C++ documentation:

The QPixmapCache::Key class can be used for efficient access to the QPixmapCache.

Use QPixmapCache::insert() to receive an instance of Key generated by the pixmap cache. You can store the key in your own objects for a very efficient one-to-one object-to-pixmap mapping.

Methods

impl Key

fn new<Args>(args: Args) -> Args::ReturnType where Args: KeyNewArgs

C++ method: QPixmapCache::Key::Key

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::pixmap_cache::Key
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::pixmap_cache::Key>
C++ method: [constructor] void QPixmapCache::Key::Key()

Variant 2

Rust arguments:
1) fn new((&::pixmap_cache::Key, cpp_utils::AsStruct)) -> ::pixmap_cache::Key
2) fn new((&::pixmap_cache::Key, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap_cache::Key>
C++ method: [constructor] void QPixmapCache::Key::Key(const QPixmapCache::Key& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Key) -> &'l0 mut Key

C++ method: QPixmapCache::Key& QPixmapCache::Key::operator=(const QPixmapCache::Key& other)

fn op_eq(&self, key: &Key) -> bool

C++ method: bool QPixmapCache::Key::operator==(const QPixmapCache::Key& key) const

fn op_neq(&self, key: &Key) -> bool

C++ method: bool QPixmapCache::Key::operator!=(const QPixmapCache::Key& key) const