Struct qt_gui::pixmap_cache::PixmapCache

pub struct PixmapCache {
    // some fields omitted
}

C++ type: QPixmapCache

C++ documentation:

The QPixmapCache class provides an application-wide cache for pixmaps.

This class is a tool for optimized drawing with QPixmap. You can use it to store temporary pixmaps that are expensive to generate without using more storage space than cacheLimit(). Use insert() to insert pixmaps, find() to find them, and clear() to empty the cache.

QPixmapCache contains no member data, only static functions to access the global pixmap cache. It creates an internal QCache object for caching the pixmaps.

The cache associates a pixmap with a user-provided string as a key, or with a QPixmapCache::Key that the cache generates. Using QPixmapCache::Key for keys is faster than using strings. The string API is very convenient for complex keys but the QPixmapCache::Key API will be very efficient and convenient for a one-to-one object-to-pixmap mapping - in this case, you can store the keys as members of an object.

If two pixmaps are inserted into the cache using equal keys then the last pixmap will replace the first pixmap in the cache. This follows the behavior of the QHash and QCache classes.

The cache becomes full when the total size of all pixmaps in the cache exceeds cacheLimit(). The initial cache limit is 10240 KB (10 MB); you can change this by calling setCacheLimit() with the required value. A pixmap takes roughly (width * height * depth)/8 bytes of memory.

The Qt Quarterly article Optimizing with QPixmapCache explains how to use QPixmapCache to speed up applications by caching the results of painting.

Methods

impl PixmapCache

fn cache_limit() -> c_int

C++ method: static int QPixmapCache::cacheLimit()

C++ documentation:

Returns the cache limit (in kilobytes).

The default cache limit is 10240 KB.

See also setCacheLimit().

fn clear()

C++ method: static void QPixmapCache::clear()

C++ documentation:

Removes all pixmaps from the cache.

fn find<Args>(args: Args) -> Args::ReturnType where Args: PixmapCacheFindArgs

C++ method: QPixmapCache::find

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn find(&qt_core::string::String) -> *mut ::pixmap::Pixmap
C++ method: static QPixmap* QPixmapCache::find(const QString& key)

Variant 2

Rust arguments: fn find((&::pixmap_cache::Key, *mut ::pixmap::Pixmap)) -> bool
C++ method: static bool QPixmapCache::find(const QPixmapCache::Key& key, QPixmap* pixmap)

C++ documentation:

Looks for a cached pixmap associated with the given key in the cache. If the pixmap is found, the function sets pixmap to that pixmap and returns true; otherwise it leaves pixmap alone and returns false. If the pixmap is not found, it means that the key is no longer valid, so it will be released for the next insertion.

This function was introduced in Qt 4.6.

Variant 3

Rust arguments: fn find((&qt_core::string::String, &mut ::pixmap::Pixmap)) -> bool
C++ method: static bool QPixmapCache::find(const QString& key, QPixmap& pixmap)

Variant 4

Rust arguments: fn find((&qt_core::string::String, *mut ::pixmap::Pixmap)) -> bool
C++ method: static bool QPixmapCache::find(const QString& key, QPixmap* pixmap)

C++ documentation:

Looks for a cached pixmap associated with the given key in the cache. If the pixmap is found, the function sets pixmap to that pixmap and returns true; otherwise it leaves pixmap alone and returns false.

Example:

QPixmap pm;
if (!QPixmapCache::find("my_big_image", &pm)) {
    pm.load("bigimage.png");
    QPixmapCache::insert("my_big_image", pm);
}
painter->drawPixmap(0, 0, pm);

This function was introduced in Qt 4.6.

fn insert<Args>(args: Args) -> Args::ReturnType where Args: PixmapCacheInsertArgs

C++ method: QPixmapCache::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn insert((&::pixmap::Pixmap, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pixmap_cache::Key>
2) fn insert((&::pixmap::Pixmap, cpp_utils::AsStruct)) -> ::pixmap_cache::Key
C++ method: static QPixmapCache::Key QPixmapCache::insert(const QPixmap& pixmap)

C++ documentation:

Inserts a copy of the given pixmap into the cache and returns a key that can be used to retrieve it.

When a pixmap is inserted and the cache is about to exceed its limit, it removes pixmaps until there is enough room for the pixmap to be inserted.

The oldest pixmaps (least recently accessed in the cache) are deleted when more space is needed.

This function was introduced in Qt 4.6.

See also setCacheLimit() and replace().

Variant 2

Rust arguments: fn insert((&qt_core::string::String, &::pixmap::Pixmap)) -> bool
C++ method: static bool QPixmapCache::insert(const QString& key, const QPixmap& pixmap)

C++ documentation:

Inserts a copy of the pixmap pixmap associated with the key into the cache.

All pixmaps inserted by the Qt library have a key starting with "$qt", so your own pixmap keys should never begin "$qt".

When a pixmap is inserted and the cache is about to exceed its limit, it removes pixmaps until there is enough room for the pixmap to be inserted.

The oldest pixmaps (least recently accessed in the cache) are deleted when more space is needed.

The function returns true if the object was inserted into the cache; otherwise it returns false.

See also setCacheLimit().

fn remove<Args>(args: Args) where Args: PixmapCacheRemoveArgs

C++ method: QPixmapCache::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&::pixmap_cache::Key) -> ()
C++ method: static void QPixmapCache::remove(const QPixmapCache::Key& key)

C++ documentation:

Removes the pixmap associated with key from the cache and releases the key for a future insertion.

This function was introduced in Qt 4.6.

Variant 2

Rust arguments: fn remove(&qt_core::string::String) -> ()
C++ method: static void QPixmapCache::remove(const QString& key)

C++ documentation:

Removes the pixmap associated with key from the cache.

fn replace(key: &Key, pixmap: &Pixmap) -> bool

C++ method: static bool QPixmapCache::replace(const QPixmapCache::Key& key, const QPixmap& pixmap)

C++ documentation:

Replaces the pixmap associated with the given key with the pixmap specified. Returns true if the pixmap has been correctly inserted into the cache; otherwise returns false.

This function was introduced in Qt 4.6.

See also setCacheLimit() and insert().

fn set_cache_limit(arg1: c_int)

C++ method: static void QPixmapCache::setCacheLimit(int arg1)

C++ documentation:

Sets the cache limit to n kilobytes.

The default setting is 10240 KB.

See also cacheLimit().