Module qt_gui::polygon::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

PolygonAppendArgs
This trait represents a set of arguments accepted by Polygon::append method.

PolygonAsQVariantArgs
This trait represents a set of arguments accepted by Polygon::as_q_variant method.

PolygonBoundingRectArgs
This trait represents a set of arguments accepted by Polygon::bounding_rect method.

PolygonCountArgs
This trait represents a set of arguments accepted by Polygon::count method.

PolygonFillArgs
This trait represents a set of arguments accepted by Polygon::fill method.

PolygonIndexOfArgs
This trait represents a set of arguments accepted by Polygon::index_of method.

PolygonInsertArgs
This trait represents a set of arguments accepted by Polygon::insert method.

PolygonIntersectedArgs
This trait represents a set of arguments accepted by Polygon::intersected method.

PolygonLastIndexOfArgs
This trait represents a set of arguments accepted by Polygon::last_index_of method.

PolygonMidArgs
This trait represents a set of arguments accepted by Polygon::mid method.

PolygonNewArgs
This trait represents a set of arguments accepted by Polygon::new method.

PolygonOpAddArgs
This trait represents a set of arguments accepted by Polygon::op_add method.

PolygonOpAddAssignArgs
This trait represents a set of arguments accepted by Polygon::op_add_assign method.

PolygonOpShlArgs
This trait represents a set of arguments accepted by Polygon::op_shl method.

PolygonPointArgs
This trait represents a set of arguments accepted by Polygon::point method.

PolygonPutPointsArgs
This trait represents a set of arguments accepted by Polygon::put_points method.

PolygonRemoveArgs
This trait represents a set of arguments accepted by Polygon::remove method.

PolygonSetPointArgs
This trait represents a set of arguments accepted by Polygon::set_point method.

PolygonSubtractedArgs
This trait represents a set of arguments accepted by Polygon::subtracted method.

PolygonTakeAtArgs
This trait represents a set of arguments accepted by Polygon::take_at method.

PolygonTakeFirstArgs
This trait represents a set of arguments accepted by Polygon::take_first method.

PolygonTakeLastArgs
This trait represents a set of arguments accepted by Polygon::take_last method.

PolygonTranslateArgs
This trait represents a set of arguments accepted by Polygon::translate method.

PolygonTranslatedArgs
This trait represents a set of arguments accepted by Polygon::translated method.

PolygonUnitedArgs
This trait represents a set of arguments accepted by Polygon::united method.

PolygonValueArgs
This trait represents a set of arguments accepted by Polygon::value method.