Struct qt_gui::polygon::Polygon

pub struct Polygon {
    // some fields omitted
}

C++ type: QPolygon

C++ documentation:

The QPolygon class provides a vector of points using integer precision.

A QPolygon object is a QVector<QPoint>. The easiest way to add points to a QPolygon is to use QVector's streaming operator, as illustrated below:

        QPolygon polygon;
        polygon << QPoint(10, 20) << QPoint(20, 30);

In addition to the functions provided by QVector, QPolygon provides some point-specific functions.

Each point in a polygon can be retrieved by passing its index to the point() function. To populate the polygon, QPolygon provides the setPoint() function to set the point at a given index, the setPoints() function to set all the points in the polygon (resizing it to the given number of points), and the putPoints() function which copies a number of given points into the polygon from a specified index (resizing the polygon if necessary).

QPolygon provides the boundingRect() and translate() functions for geometry functions. Use the QMatrix::map() function for more general transformations of QPolygons.

The QPolygon class is implicitly shared.

Methods

impl Polygon

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonAppendArgs<'largs>

C++ method: QPolygon::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &qt_core::point::Point) -> ()
C++ method: void QPolygon::append(const QPoint& t)

C++ documentation:

Inserts value at the end of the vector.

Example:

QVector<QString> vector(0);
vector.append("one");
vector.append("two");
vector.append("three");
// vector: ["one", "two", "three"]

This is the same as calling resize(size() + 1) and assigning value to the new last element in the vector.

This operation is relatively fast, because QVector typically allocates more memory than necessary, so it can grow without reallocating the entire vector each time.

See also operator<<(), prepend(), and insert().

Variant 2

Rust arguments: fn append(&mut self, &::vector::VectorPointRef) -> ()
C++ method: void QPolygon::append(const QVector<QPoint>& l)

C++ documentation:

This is an overloaded function.

Appends the items of the value vector to this vector.

This function was introduced in Qt 5.5.

See also operator<<() and operator+=().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonAsQVariantArgs<'largs>

C++ method: QPolygon::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QPolygon::operator QVariant() const

C++ documentation:

Returns the polygon as a QVariant

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 Point

C++ method: const QPoint& QPolygon::at(int i) const

C++ documentation:

Returns the item at index position i in the vector.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 Point

C++ method: const QPoint& QPolygon::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut Point

C++ method: QPoint& QPolygon::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last().

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonBoundingRectArgs<'largs>

C++ method: QPolygon::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn bounding_rect(&self, cpp_utils::AsStruct) -> qt_core::rect::Rect
C++ method: QRect QPolygon::boundingRect() const

C++ documentation:

Returns the bounding rectangle of the polygon, or QRect(0, 0, 0, 0) if the polygon is empty.

See also QVector::isEmpty().

fn capacity(&self) -> c_int

C++ method: int QPolygon::capacity() const

C++ documentation:

Returns the maximum number of items that can be stored in the vector without forcing a reallocation.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the vector, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QPolygon::clear()

C++ documentation:

Removes all the elements from the vector and releases the memory used by the vector.

fn const_data(&self) -> *const Point

C++ method: const QPoint* QPolygon::constData() const

C++ documentation:

Returns a const pointer to the data stored in the vector. The pointer can be used to access the items in the vector. The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also data() and operator[]().

fn contains(&self, t: &Point) -> bool

C++ method: bool QPolygon::contains(const QPoint& t) const

C++ documentation:

Returns true if the vector contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn contains_point(&self, pt: &Point, fill_rule: FillRule) -> bool

C++ method: bool QPolygon::containsPoint(const QPoint& pt, Qt::FillRule fillRule) const

C++ documentation:

Returns true if the given point is inside the polygon according to the specified fillRule; otherwise returns false.

This function was introduced in Qt 4.3.

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonCountArgs<'largs>

C++ method: QPolygon::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QPolygon::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &qt_core::point::Point) -> libc::c_int
C++ method: int QPolygon::count(const QPoint& t) const

C++ documentation:

Returns the number of occurrences of value in the vector.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn data(&self) -> *const Point

C++ method: const QPoint* QPolygon::data() const

C++ documentation:

This is an overloaded function.

fn data_mut(&mut self) -> *mut Point

C++ method: QPoint* QPolygon::data()

C++ documentation:

Returns a pointer to the data stored in the vector. The pointer can be used to access and modify the items in the vector.

Example:

QVector<int> vector(10);
int data = vector.data();
for (int i = 0; i < 10; ++i)
    data[i] = 2  i;

The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also constData() and operator[]().

fn detach(&mut self)

C++ method: void QPolygon::detach()

fn empty(&self) -> bool

C++ method: bool QPolygon::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the vector is empty; otherwise returns false.

fn ends_with(&self, t: &Point) -> bool

C++ method: bool QPolygon::endsWith(const QPoint& t) const

C++ documentation:

Returns true if this vector is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and last().

fn fill<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointRef where Args: PolygonFillArgs<'largs>

C++ method: QPolygon::fill

This is an overloaded function. Available variants:

Rust arguments:
1) fn fill(&mut self, &'l1 qt_core::point::Point) -> &'l0 mut ::vector::VectorPointRef
2) fn fill(&mut self, (&'l1 qt_core::point::Point, libc::c_int)) -> &'l0 mut ::vector::VectorPointRef
C++ method: QVector<QPoint>& QPolygon::fill(const QPoint& t, int size = ?)

C++ documentation:

Assigns value to all items in the vector. If size is different from -1 (the default), the vector is resized to size size beforehand.

Example:

QVector<QString> vector(3);
vector.fill("Yes");
// vector: ["Yes", "Yes", "Yes"]

vector.fill("oh", 5); // vector: ["oh", "oh", "oh", "oh", "oh"]
See also resize().

fn first<'l0>(&'l0 self) -> &'l0 Point

C++ method: const QPoint& QPolygon::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut Point

C++ method: QPoint& QPolygon::first()

C++ documentation:

Returns a reference to the first item in the vector. This function assumes that the vector isn't empty.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 Point

C++ method: const QPoint& QPolygon::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut Point

C++ method: QPoint& QPolygon::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first().

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonIndexOfArgs<'largs>

C++ method: QPolygon::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &qt_core::point::Point) -> libc::c_int
2) fn index_of(&self, (&qt_core::point::Point, libc::c_int)) -> libc::c_int
C++ method: int QPolygon::indexOf(const QPoint& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the vector, searching forward from index position from. Returns -1 if no item matched.

Example:

QVector<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.indexOf("B");            // returns 1
vector.indexOf("B", 1);         // returns 1
vector.indexOf("B", 2);         // returns 3
vector.indexOf("X");            // returns -1

This function requires the value type to have an implementation of operator==().

See also lastIndexOf() and contains().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonInsertArgs<'largs>

C++ method: QPolygon::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, (libc::c_int, &qt_core::point::Point)) -> ()
C++ method: void QPolygon::insert(int i, const QPoint& t)

C++ documentation:

Inserts value at index position i in the vector. If i is 0, the value is prepended to the vector. If i is size(), the value is appended to the vector.

Example:

QVector<QString> vector;
vector << "alpha" << "beta" << "delta";
vector.insert(2, "gamma");
// vector: ["alpha", "beta", "gamma", "delta"]

For large vectors, this operation can be slow (linear time), because it requires moving all the items at indexes i and above by one position further in memory. If you want a container class that provides a fast insert() function, use QLinkedList instead.

See also append(), prepend(), and remove().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, libc::c_int, &qt_core::point::Point)) -> ()
C++ method: void QPolygon::insert(int i, int n, const QPoint& t)

C++ documentation:

This is an overloaded function.

Inserts count copies of value at index position i in the vector.

Example:

QVector<double> vector;
vector << 2.718 << 1.442 << 0.4342;
vector.insert(1, 3, 9.9);
// vector: [2.718, 9.9, 9.9, 9.9, 1.442, 0.4342]

fn intersected<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonIntersectedArgs<'largs>

C++ method: QPolygon::intersected

This is an overloaded function. Available variants:

Rust arguments:
1) fn intersected(&self, (&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn intersected(&self, (&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QPolygon::intersected(const QPolygon& r) const

C++ documentation:

Returns a polygon which is the intersection of this polygon and r.

Set operations on polygons will treat the polygons as areas. Non-closed polygons will be treated as implicitly closed.

This function was introduced in Qt 4.3.

fn is_detached(&self) -> bool

C++ method: bool QPolygon::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QPolygon::isEmpty() const

C++ documentation:

Returns true if the vector has size 0; otherwise returns false.

See also size() and resize().

fn is_shared_with(&self, other: &VectorPointRef) -> bool

C++ method: bool QPolygon::isSharedWith(const QVector<QPoint>& other) const

fn last<'l0>(&'l0 self) -> &'l0 Point

C++ method: const QPoint& QPolygon::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonLastIndexOfArgs<'largs>

C++ method: QPolygon::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &qt_core::point::Point) -> libc::c_int
2) fn last_index_of(&self, (&qt_core::point::Point, libc::c_int)) -> libc::c_int
C++ method: int QPolygon::lastIndexOf(const QPoint& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of the value value in the vector, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.lastIndexOf("B");        // returns 3
vector.lastIndexOf("B", 3);     // returns 3
vector.lastIndexOf("B", 2);     // returns 1
vector.lastIndexOf("X");        // returns -1

This function requires the value type to have an implementation of operator==().

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut Point

C++ method: QPoint& QPolygon::last()

C++ documentation:

Returns a reference to the last item in the vector. This function assumes that the vector isn't empty.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QPolygon::length() const

C++ documentation:

Same as size() and count().

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also size(), count(), and QList::length().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonMidArgs<'largs>

C++ method: QPolygon::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorPointRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorPointRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorPointRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorPointRef
C++ method: QVector<QPoint> QPolygon::mid(int pos, int len = ?) const

C++ documentation:

Returns a sub-vector which contains elements from this vector, starting at position pos. If length is -1 (the default), all elements after pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PolygonNewArgs

C++ method: QPolygon::QPolygon

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::polygon::Polygon
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon()

Variant 2

Rust arguments:
1) fn new((&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
2) fn new((&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon(const QPolygon& a)

Variant 3

Rust arguments:
1) fn new((&qt_core::rect::Rect, cpp_utils::AsStruct)) -> ::polygon::Polygon
2) fn new((&qt_core::rect::Rect, bool, cpp_utils::AsStruct)) -> ::polygon::Polygon
3) fn new((&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
4) fn new((&qt_core::rect::Rect, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon(const QRect& r, bool closed = ?)

Variant 4

Rust arguments:
1) fn new((&::vector::VectorPointRef, cpp_utils::AsStruct)) -> ::polygon::Polygon
2) fn new((&::vector::VectorPointRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon(const QVector<QPoint>& v)

Variant 5

Rust arguments:
1) fn new((libc::c_int, *const libc::c_int, cpp_utils::AsStruct)) -> ::polygon::Polygon
2) fn new((libc::c_int, *const libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon(int nPoints, const int* points)

Variant 6

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::polygon::Polygon
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
C++ method: [constructor] void QPolygon::QPolygon(int size)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonOpAddArgs<'largs>

C++ method: QPolygon::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::vector::VectorPointRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorPointRef>
2) fn op_add(&self, (&::vector::VectorPointRef, cpp_utils::AsStruct)) -> ::vector::VectorPointRef
C++ method: QVector<QPoint> QPolygon::operator+(const QVector<QPoint>& l) const

C++ documentation:

Returns a vector that contains all the items in this vector followed by all the items in the other vector.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointRef where Args: PolygonOpAddAssignArgs<'largs>

C++ method: QPolygon::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 qt_core::point::Point) -> &'l0 mut ::vector::VectorPointRef
C++ method: QVector<QPoint>& QPolygon::operator+=(const QPoint& t)

C++ documentation:

This is an overloaded function.

Appends value to the vector.

See also append() and operator<<().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::vector::VectorPointRef) -> &'l0 mut ::vector::VectorPointRef
C++ method: QVector<QPoint>& QPolygon::operator+=(const QVector<QPoint>& l)

C++ documentation:

Appends the items of the other vector to this vector and returns a reference to this vector.

See also operator+() and append().

fn op_eq(&self, v: &VectorPointRef) -> bool

C++ method: bool QPolygon::operator==(const QVector<QPoint>& v) const

C++ documentation:

Returns true if other is equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 Point

C++ method: const QPoint& QPolygon::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(i).

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut Point

C++ method: QPoint& QPolygon::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

Note that using non-const operators can cause QVector to do a deep copy.

See also at() and value().

fn op_neq(&self, v: &VectorPointRef) -> bool

C++ method: bool QPolygon::operator!=(const QVector<QPoint>& v) const

C++ documentation:

Returns true if other is not equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointRef where Args: PolygonOpShlArgs<'largs>

C++ method: QPolygon::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 qt_core::point::Point) -> &'l0 mut ::vector::VectorPointRef
C++ method: QVector<QPoint>& QPolygon::operator<<(const QPoint& t)

C++ documentation:

Appends value to the vector and returns a reference to this vector.

See also append() and operator+=().

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::vector::VectorPointRef) -> &'l0 mut ::vector::VectorPointRef
C++ method: QVector<QPoint>& QPolygon::operator<<(const QVector<QPoint>& l)

C++ documentation:

Appends other to the vector and returns a reference to the vector.

fn point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonPointArgs<'largs>

C++ method: QPolygon::point

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn point(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn point(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QPolygon::point(int i) const

C++ documentation:

This is an overloaded function.

Returns the point at the given index.

Variant 2

Rust arguments: fn point(&self, (libc::c_int, *mut libc::c_int, *mut libc::c_int)) -> ()
C++ method: void QPolygon::point(int i, int* x, int* y) const

C++ documentation:

Extracts the coordinates of the point at the given index to *x and *y (if they are valid pointers).

See also setPoint().

fn pop_back(&mut self)

C++ method: void QPolygon::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast().

fn pop_front(&mut self)

C++ method: void QPolygon::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst().

fn prepend(&mut self, t: &Point)

C++ method: void QPolygon::prepend(const QPoint& t)

C++ documentation:

Inserts value at the beginning of the vector.

Example:

QVector<QString> vector;
vector.prepend("one");
vector.prepend("two");
vector.prepend("three");
// vector: ["three", "two", "one"]

This is the same as vector.insert(0, value).

For large vectors, this operation can be slow (linear time), because it requires moving all the items in the vector by one position further in memory. If you want a container class that provides a fast prepend() function, use QList or QLinkedList instead.

See also append() and insert().

fn push_back(&mut self, t: &Point)

C++ method: void QPolygon::push_back(const QPoint& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &Point)

C++ method: void QPolygon::push_front(const QPoint& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn put_points<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonPutPointsArgs<'largs>

C++ method: QPolygon::putPoints

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn put_points(&mut self, (libc::c_int, libc::c_int, &::polygon::Polygon)) -> ()
2) fn put_points(&mut self, (libc::c_int, libc::c_int, &::polygon::Polygon, libc::c_int)) -> ()
C++ method: void QPolygon::putPoints(int index, int nPoints, const QPolygon& from, int fromIndex = ?)

C++ documentation:

This is an overloaded function.

Copies nPoints points from the given fromIndex ( 0 by default) in fromPolygon into this polygon, starting at the specified index. For example:

        QPolygon polygon1;
        polygon1.putPoints(0, 3, 1,2, 0,0, 5,6);
        // polygon1 is now the three-point polygon(1,2, 0,0, 5,6);

QPolygon polygon2; polygon2.putPoints(0, 3, 4,4, 5,5, 6,6); // polygon2 is now (4,4, 5,5, 6,6);

polygon1.putPoints(2, 3, polygon2); // polygon1 is now the five-point polygon(1,2, 0,0, 4,4, 5,5, 6,6);

Variant 2

Rust arguments: fn put_points(&mut self, (libc::c_int, libc::c_int, *const libc::c_int)) -> ()
C++ method: void QPolygon::putPoints(int index, int nPoints, const int* points)

fn remove<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonRemoveArgs<'largs>

C++ method: QPolygon::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&mut self, libc::c_int) -> ()
C++ method: void QPolygon::remove(int i)

C++ documentation:

This is an overloaded function.

Removes the element at index position i.

See also insert(), replace(), and fill().

Variant 2

Rust arguments: fn remove(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QPolygon::remove(int i, int n)

C++ documentation:

This is an overloaded function.

Removes count elements from the middle of the vector, starting at index position i.

See also insert(), replace(), and fill().

fn remove_all(&mut self, t: &Point) -> c_int

C++ method: int QPolygon::removeAll(const QPoint& t)

C++ documentation:

Removes all elements that compare equal to t from the vector. Returns the number of elements removed, if any.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeOne() and QList::removeAll().

fn remove_at(&mut self, i: c_int)

C++ method: void QPolygon::removeAt(int i)

C++ documentation:

Removes the element at index position i. Equivalent to

remove(i);

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also remove() and QList::removeAt().

fn remove_first(&mut self)

C++ method: void QPolygon::removeFirst()

C++ documentation:

Removes the first item in the vector. Calling this function is equivalent to calling remove(0). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeFirst(), and isEmpty().

fn remove_last(&mut self)

C++ method: void QPolygon::removeLast()

C++ documentation:

Removes the last item in the vector. Calling this function is equivalent to calling remove(size() - 1). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeLast(), removeFirst(), and isEmpty().

fn remove_one(&mut self, t: &Point) -> bool

C++ method: bool QPolygon::removeOne(const QPoint& t)

C++ documentation:

Removes the first element that compares equal to t from the vector. Returns whether an element was, in fact, removed.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeAll() and QList::removeOne().

fn replace(&mut self, i: c_int, t: &Point)

C++ method: void QPolygon::replace(int i, const QPoint& t)

C++ documentation:

Replaces the item at index position i with value.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also operator[]() and remove().

fn reserve(&mut self, size: c_int)

C++ method: void QPolygon::reserve(int size)

C++ documentation:

Attempts to allocate memory for at least size elements. If you know in advance how large the vector will be, you can call this function, and if you call resize() often you are likely to get better performance. If size is an underestimate, the worst that will happen is that the QVector will be a bit slower.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to change the size of the vector, call resize().

See also squeeze() and capacity().

fn resize(&mut self, size: c_int)

C++ method: void QPolygon::resize(int size)

C++ documentation:

Sets the size of the vector to size. If size is greater than the current size, elements are added to the end; the new elements are initialized with a default-constructed value. If size is less than the current size, elements are removed from the end.

See also size().

fn set_point<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonSetPointArgs<'largs>

C++ method: QPolygon::setPoint

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_point(&mut self, (libc::c_int, &qt_core::point::Point)) -> ()
C++ method: void QPolygon::setPoint(int index, const QPoint& p)

C++ documentation:

This is an overloaded function.

Sets the point at the given index to the given point.

Variant 2

Rust arguments: fn set_point(&mut self, (libc::c_int, libc::c_int, libc::c_int)) -> ()
C++ method: void QPolygon::setPoint(int index, int x, int y)

C++ documentation:

Sets the point at the given index to the point specified by (x, y).

See also point(), putPoints(), and setPoints().

fn set_points(&mut self, n_points: c_int, points: *const c_int)

C++ method: void QPolygon::setPoints(int nPoints, const int* points)

C++ documentation:

Resizes the polygon to nPoints and populates it with the given points.

The example code creates a polygon with two points (10, 20) and (30, 40):

        static const int points[] = { 10, 20, 30, 40 };
        QPolygon polygon;
        polygon.setPoints(2, points);

See also setPoint() and putPoints().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QPolygon::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QPolygon::size() const

C++ documentation:

Returns the number of items in the vector.

See also isEmpty() and resize().

fn squeeze(&mut self)

C++ method: void QPolygon::squeeze()

C++ documentation:

Releases any memory not required to store the items.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn starts_with(&self, t: &Point) -> bool

C++ method: bool QPolygon::startsWith(const QPoint& t) const

C++ documentation:

Returns true if this vector is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and first().

fn subtracted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonSubtractedArgs<'largs>

C++ method: QPolygon::subtracted

This is an overloaded function. Available variants:

Rust arguments:
1) fn subtracted(&self, (&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn subtracted(&self, (&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QPolygon::subtracted(const QPolygon& r) const

C++ documentation:

Returns a polygon which is r subtracted from this polygon.

Set operations on polygons will treat the polygons as areas. Non-closed polygons will be treated as implicitly closed.

This function was introduced in Qt 4.3.

fn swap(&mut self, other: &mut Polygon)

C++ method: void QPolygon::swap(QPolygon& other)

C++ documentation:

Swaps polygon other with this polygon. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonTakeAtArgs<'largs>

C++ method: QPolygon::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QPolygon::takeAt(int i)

C++ documentation:

Removes the element at index position i and returns it.

Equivalent to

T t = at(i);
remove(i);
return t;

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also takeFirst(), takeLast(), and QList::takeAt().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonTakeFirstArgs<'largs>

C++ method: QPolygon::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QPolygon::takeFirst()

C++ documentation:

Removes the first item in the vector and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also takeLast() and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonTakeLastArgs<'largs>

C++ method: QPolygon::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QPolygon::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

If you don't use the return value, removeLast() is more efficient.

This function was introduced in Qt 5.1.

See also takeFirst() and removeLast().

fn translate<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonTranslateArgs<'largs>

C++ method: QPolygon::translate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn translate(&mut self, &qt_core::point::Point) -> ()
C++ method: void QPolygon::translate(const QPoint& offset)

C++ documentation:

This is an overloaded function.

Translates all points in the polygon by the given offset.

See also translated().

Variant 2

Rust arguments: fn translate(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QPolygon::translate(int dx, int dy)

C++ documentation:

Translates all points in the polygon by (dx, dy).

See also translated().

fn translated<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonTranslatedArgs<'largs>

C++ method: QPolygon::translated

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn translated(&self, (&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn translated(&self, (&qt_core::point::Point, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QPolygon::translated(const QPoint& offset) const

C++ documentation:

This is an overloaded function.

Returns a copy of the polygon that is translated by the given offset.

This function was introduced in Qt 4.6.

See also translate().

Variant 2

Rust arguments:
1) fn translated(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn translated(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QPolygon::translated(int dx, int dy) const

C++ documentation:

Returns a copy of the polygon that is translated by (dx, dy).

This function was introduced in Qt 4.6.

See also translate().

fn united<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonUnitedArgs<'largs>

C++ method: QPolygon::united

This is an overloaded function. Available variants:

Rust arguments:
1) fn united(&self, (&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn united(&self, (&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QPolygon::united(const QPolygon& r) const

C++ documentation:

Returns a polygon which is the union of this polygon and r.

Set operations on polygons, will treat the polygons as areas, and implicitly close the polygon.

This function was introduced in Qt 4.3.

See also intersected() and subtracted().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonValueArgs<'largs>

C++ method: QPolygon::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QPolygon::value(int i) const

C++ documentation:

Returns the value at index position i in the vector.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that i is within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn value(&self, (libc::c_int, &qt_core::point::Point, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QPolygon::value(int i, const QPoint& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.