Struct qt_gui::polygon_f::PolygonF

pub struct PolygonF {
    // some fields omitted
}

C++ type: QPolygonF

C++ documentation:

The QPolygonF class provides a vector of points using floating point precision.

A QPolygonF is a QVector<QPointF>. The easiest way to add points to a QPolygonF is to use its streaming operator, as illustrated below:

        QPolygonF polygon;
        polygon << QPointF(10.4, 20.5) << QPointF(20.2, 30.2);

In addition to the functions provided by QVector, QPolygonF provides the boundingRect() and translate() functions for geometry operations. Use the QMatrix::map() function for more general transformations of QPolygonFs.

QPolygonF also provides the isClosed() function to determine whether a polygon's start and end points are the same, and the toPolygon() function returning an integer precision copy of this polygon.

The QPolygonF class is implicitly shared.

Methods

impl PolygonF

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonFAppendArgs<'largs>

C++ method: QPolygonF::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &qt_core::point_f::PointF) -> ()
C++ method: void QPolygonF::append(const QPointF& t)

C++ documentation:

Inserts value at the end of the vector.

Example:

QVector<QString> vector(0);
vector.append("one");
vector.append("two");
vector.append("three");
// vector: ["one", "two", "three"]

This is the same as calling resize(size() + 1) and assigning value to the new last element in the vector.

This operation is relatively fast, because QVector typically allocates more memory than necessary, so it can grow without reallocating the entire vector each time.

See also operator<<(), prepend(), and insert().

Variant 2

Rust arguments: fn append(&mut self, &qt_core::vector::VectorPointFRef) -> ()
C++ method: void QPolygonF::append(const QVector<QPointF>& l)

C++ documentation:

This is an overloaded function.

Appends the items of the value vector to this vector.

This function was introduced in Qt 5.5.

See also operator<<() and operator+=().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFAsQVariantArgs<'largs>

C++ method: QPolygonF::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QPolygonF::operator QVariant() const

C++ documentation:

Returns the polygon as a QVariant.

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::at(int i) const

C++ documentation:

Returns the item at index position i in the vector.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut PointF

C++ method: QPointF& QPolygonF::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last().

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFBoundingRectArgs<'largs>

C++ method: QPolygonF::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn bounding_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QPolygonF::boundingRect() const

C++ documentation:

Returns the bounding rectangle of the polygon, or QRectF(0,0,0,0) if the polygon is empty.

See also QVector::isEmpty().

fn capacity(&self) -> c_int

C++ method: int QPolygonF::capacity() const

C++ documentation:

Returns the maximum number of items that can be stored in the vector without forcing a reallocation.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the vector, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QPolygonF::clear()

C++ documentation:

Removes all the elements from the vector and releases the memory used by the vector.

fn const_data(&self) -> *const PointF

C++ method: const QPointF* QPolygonF::constData() const

C++ documentation:

Returns a const pointer to the data stored in the vector. The pointer can be used to access the items in the vector. The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also data() and operator[]().

fn contains(&self, t: &PointF) -> bool

C++ method: bool QPolygonF::contains(const QPointF& t) const

C++ documentation:

Returns true if the vector contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn contains_point(&self, pt: &PointF, fill_rule: FillRule) -> bool

C++ method: bool QPolygonF::containsPoint(const QPointF& pt, Qt::FillRule fillRule) const

C++ documentation:

Returns true if the given point is inside the polygon according to the specified fillRule; otherwise returns false.

This function was introduced in Qt 4.3.

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonFCountArgs<'largs>

C++ method: QPolygonF::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QPolygonF::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &qt_core::point_f::PointF) -> libc::c_int
C++ method: int QPolygonF::count(const QPointF& t) const

C++ documentation:

Returns the number of occurrences of value in the vector.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn data(&self) -> *const PointF

C++ method: const QPointF* QPolygonF::data() const

C++ documentation:

This is an overloaded function.

fn data_mut(&mut self) -> *mut PointF

C++ method: QPointF* QPolygonF::data()

C++ documentation:

Returns a pointer to the data stored in the vector. The pointer can be used to access and modify the items in the vector.

Example:

QVector<int> vector(10);
int data = vector.data();
for (int i = 0; i < 10; ++i)
    data[i] = 2  i;

The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also constData() and operator[]().

fn detach(&mut self)

C++ method: void QPolygonF::detach()

fn empty(&self) -> bool

C++ method: bool QPolygonF::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the vector is empty; otherwise returns false.

fn ends_with(&self, t: &PointF) -> bool

C++ method: bool QPolygonF::endsWith(const QPointF& t) const

C++ documentation:

Returns true if this vector is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and last().

fn fill<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointFRef where Args: PolygonFFillArgs<'largs>

C++ method: QPolygonF::fill

This is an overloaded function. Available variants:

Rust arguments:
1) fn fill(&mut self, &'l1 qt_core::point_f::PointF) -> &'l0 mut qt_core::vector::VectorPointFRef
2) fn fill(&mut self, (&'l1 qt_core::point_f::PointF, libc::c_int)) -> &'l0 mut qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF>& QPolygonF::fill(const QPointF& t, int size = ?)

C++ documentation:

Assigns value to all items in the vector. If size is different from -1 (the default), the vector is resized to size size beforehand.

Example:

QVector<QString> vector(3);
vector.fill("Yes");
// vector: ["Yes", "Yes", "Yes"]

vector.fill("oh", 5); // vector: ["oh", "oh", "oh", "oh", "oh"]
See also resize().

fn first<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut PointF

C++ method: QPointF& QPolygonF::first()

C++ documentation:

Returns a reference to the first item in the vector. This function assumes that the vector isn't empty.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut PointF

C++ method: QPointF& QPolygonF::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first().

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonFIndexOfArgs<'largs>

C++ method: QPolygonF::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &qt_core::point_f::PointF) -> libc::c_int
2) fn index_of(&self, (&qt_core::point_f::PointF, libc::c_int)) -> libc::c_int
C++ method: int QPolygonF::indexOf(const QPointF& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the vector, searching forward from index position from. Returns -1 if no item matched.

Example:

QVector<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.indexOf("B");            // returns 1
vector.indexOf("B", 1);         // returns 1
vector.indexOf("B", 2);         // returns 3
vector.indexOf("X");            // returns -1

This function requires the value type to have an implementation of operator==().

See also lastIndexOf() and contains().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonFInsertArgs<'largs>

C++ method: QPolygonF::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, (libc::c_int, &qt_core::point_f::PointF)) -> ()
C++ method: void QPolygonF::insert(int i, const QPointF& t)

C++ documentation:

Inserts value at index position i in the vector. If i is 0, the value is prepended to the vector. If i is size(), the value is appended to the vector.

Example:

QVector<QString> vector;
vector << "alpha" << "beta" << "delta";
vector.insert(2, "gamma");
// vector: ["alpha", "beta", "gamma", "delta"]

For large vectors, this operation can be slow (linear time), because it requires moving all the items at indexes i and above by one position further in memory. If you want a container class that provides a fast insert() function, use QLinkedList instead.

See also append(), prepend(), and remove().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, libc::c_int, &qt_core::point_f::PointF)) -> ()
C++ method: void QPolygonF::insert(int i, int n, const QPointF& t)

C++ documentation:

This is an overloaded function.

Inserts count copies of value at index position i in the vector.

Example:

QVector<double> vector;
vector << 2.718 << 1.442 << 0.4342;
vector.insert(1, 3, 9.9);
// vector: [2.718, 9.9, 9.9, 9.9, 1.442, 0.4342]

fn intersected<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFIntersectedArgs<'largs>

C++ method: QPolygonF::intersected

This is an overloaded function. Available variants:

Rust arguments:
1) fn intersected(&self, (&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn intersected(&self, (&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QPolygonF::intersected(const QPolygonF& r) const

C++ documentation:

Returns a polygon which is the intersection of this polygon and r.

Set operations on polygons will treat the polygons as areas. Non-closed polygons will be treated as implicitly closed.

This function was introduced in Qt 4.3.

fn is_closed(&self) -> bool

C++ method: bool QPolygonF::isClosed() const

C++ documentation:

Returns true if the polygon is closed; otherwise returns false.

A polygon is said to be closed if its start point and end point are equal.

See also QVector::first() and QVector::last().

fn is_detached(&self) -> bool

C++ method: bool QPolygonF::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QPolygonF::isEmpty() const

C++ documentation:

Returns true if the vector has size 0; otherwise returns false.

See also size() and resize().

fn is_shared_with(&self, other: &VectorPointFRef) -> bool

C++ method: bool QPolygonF::isSharedWith(const QVector<QPointF>& other) const

fn last<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: PolygonFLastIndexOfArgs<'largs>

C++ method: QPolygonF::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &qt_core::point_f::PointF) -> libc::c_int
2) fn last_index_of(&self, (&qt_core::point_f::PointF, libc::c_int)) -> libc::c_int
C++ method: int QPolygonF::lastIndexOf(const QPointF& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of the value value in the vector, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.lastIndexOf("B");        // returns 3
vector.lastIndexOf("B", 3);     // returns 3
vector.lastIndexOf("B", 2);     // returns 1
vector.lastIndexOf("X");        // returns -1

This function requires the value type to have an implementation of operator==().

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut PointF

C++ method: QPointF& QPolygonF::last()

C++ documentation:

Returns a reference to the last item in the vector. This function assumes that the vector isn't empty.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QPolygonF::length() const

C++ documentation:

Same as size() and count().

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also size(), count(), and QList::length().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFMidArgs<'largs>

C++ method: QPolygonF::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::vector::VectorPointFRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QPolygonF::mid(int pos, int len = ?) const

C++ documentation:

Returns a sub-vector which contains elements from this vector, starting at position pos. If length is -1 (the default), all elements after pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PolygonFNewArgs

C++ method: QPolygonF::QPolygonF

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::polygon_f::PolygonF
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF()

Variant 2

Rust arguments:
1) fn new((&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
2) fn new((&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF(const QPolygon& a)

Variant 3

Rust arguments:
1) fn new((&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
2) fn new((&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF(const QPolygonF& a)

Variant 4

Rust arguments:
1) fn new((&qt_core::rect_f::RectF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
2) fn new((&qt_core::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF(const QRectF& r)

Variant 5

Rust arguments:
1) fn new((&qt_core::vector::VectorPointFRef, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
2) fn new((&qt_core::vector::VectorPointFRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF(const QVector<QPointF>& v)

Variant 6

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
C++ method: [constructor] void QPolygonF::QPolygonF(int size)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFOpAddArgs<'largs>

C++ method: QPolygonF::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&qt_core::vector::VectorPointFRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn op_add(&self, (&qt_core::vector::VectorPointFRef, cpp_utils::AsStruct)) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QPolygonF::operator+(const QVector<QPointF>& l) const

C++ documentation:

Returns a vector that contains all the items in this vector followed by all the items in the other vector.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointFRef where Args: PolygonFOpAddAssignArgs<'largs>

C++ method: QPolygonF::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 qt_core::point_f::PointF) -> &'l0 mut qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF>& QPolygonF::operator+=(const QPointF& t)

C++ documentation:

This is an overloaded function.

Appends value to the vector.

See also append() and operator<<().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 qt_core::vector::VectorPointFRef) -> &'l0 mut qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF>& QPolygonF::operator+=(const QVector<QPointF>& l)

C++ documentation:

Appends the items of the other vector to this vector and returns a reference to this vector.

See also operator+() and append().

fn op_eq(&self, v: &VectorPointFRef) -> bool

C++ method: bool QPolygonF::operator==(const QVector<QPointF>& v) const

C++ documentation:

Returns true if other is equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 PointF

C++ method: const QPointF& QPolygonF::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(i).

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut PointF

C++ method: QPointF& QPolygonF::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

Note that using non-const operators can cause QVector to do a deep copy.

See also at() and value().

fn op_neq(&self, v: &VectorPointFRef) -> bool

C++ method: bool QPolygonF::operator!=(const QVector<QPointF>& v) const

C++ documentation:

Returns true if other is not equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorPointFRef where Args: PolygonFOpShlArgs<'largs>

C++ method: QPolygonF::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 qt_core::point_f::PointF) -> &'l0 mut qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF>& QPolygonF::operator<<(const QPointF& t)

C++ documentation:

Appends value to the vector and returns a reference to this vector.

See also append() and operator+=().

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 qt_core::vector::VectorPointFRef) -> &'l0 mut qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF>& QPolygonF::operator<<(const QVector<QPointF>& l)

C++ documentation:

Appends other to the vector and returns a reference to the vector.

fn pop_back(&mut self)

C++ method: void QPolygonF::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast().

fn pop_front(&mut self)

C++ method: void QPolygonF::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst().

fn prepend(&mut self, t: &PointF)

C++ method: void QPolygonF::prepend(const QPointF& t)

C++ documentation:

Inserts value at the beginning of the vector.

Example:

QVector<QString> vector;
vector.prepend("one");
vector.prepend("two");
vector.prepend("three");
// vector: ["three", "two", "one"]

This is the same as vector.insert(0, value).

For large vectors, this operation can be slow (linear time), because it requires moving all the items in the vector by one position further in memory. If you want a container class that provides a fast prepend() function, use QList or QLinkedList instead.

See also append() and insert().

fn push_back(&mut self, t: &PointF)

C++ method: void QPolygonF::push_back(const QPointF& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &PointF)

C++ method: void QPolygonF::push_front(const QPointF& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonFRemoveArgs<'largs>

C++ method: QPolygonF::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&mut self, libc::c_int) -> ()
C++ method: void QPolygonF::remove(int i)

C++ documentation:

This is an overloaded function.

Removes the element at index position i.

See also insert(), replace(), and fill().

Variant 2

Rust arguments: fn remove(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QPolygonF::remove(int i, int n)

C++ documentation:

This is an overloaded function.

Removes count elements from the middle of the vector, starting at index position i.

See also insert(), replace(), and fill().

fn remove_all(&mut self, t: &PointF) -> c_int

C++ method: int QPolygonF::removeAll(const QPointF& t)

C++ documentation:

Removes all elements that compare equal to t from the vector. Returns the number of elements removed, if any.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeOne() and QList::removeAll().

fn remove_at(&mut self, i: c_int)

C++ method: void QPolygonF::removeAt(int i)

C++ documentation:

Removes the element at index position i. Equivalent to

remove(i);

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also remove() and QList::removeAt().

fn remove_first(&mut self)

C++ method: void QPolygonF::removeFirst()

C++ documentation:

Removes the first item in the vector. Calling this function is equivalent to calling remove(0). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeFirst(), and isEmpty().

fn remove_last(&mut self)

C++ method: void QPolygonF::removeLast()

C++ documentation:

Removes the last item in the vector. Calling this function is equivalent to calling remove(size() - 1). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeLast(), removeFirst(), and isEmpty().

fn remove_one(&mut self, t: &PointF) -> bool

C++ method: bool QPolygonF::removeOne(const QPointF& t)

C++ documentation:

Removes the first element that compares equal to t from the vector. Returns whether an element was, in fact, removed.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeAll() and QList::removeOne().

fn replace(&mut self, i: c_int, t: &PointF)

C++ method: void QPolygonF::replace(int i, const QPointF& t)

C++ documentation:

Replaces the item at index position i with value.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also operator[]() and remove().

fn reserve(&mut self, size: c_int)

C++ method: void QPolygonF::reserve(int size)

C++ documentation:

Attempts to allocate memory for at least size elements. If you know in advance how large the vector will be, you can call this function, and if you call resize() often you are likely to get better performance. If size is an underestimate, the worst that will happen is that the QVector will be a bit slower.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to change the size of the vector, call resize().

See also squeeze() and capacity().

fn resize(&mut self, size: c_int)

C++ method: void QPolygonF::resize(int size)

C++ documentation:

Sets the size of the vector to size. If size is greater than the current size, elements are added to the end; the new elements are initialized with a default-constructed value. If size is less than the current size, elements are removed from the end.

See also size().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QPolygonF::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QPolygonF::size() const

C++ documentation:

Returns the number of items in the vector.

See also isEmpty() and resize().

fn squeeze(&mut self)

C++ method: void QPolygonF::squeeze()

C++ documentation:

Releases any memory not required to store the items.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn starts_with(&self, t: &PointF) -> bool

C++ method: bool QPolygonF::startsWith(const QPointF& t) const

C++ documentation:

Returns true if this vector is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and first().

fn subtracted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFSubtractedArgs<'largs>

C++ method: QPolygonF::subtracted

This is an overloaded function. Available variants:

Rust arguments:
1) fn subtracted(&self, (&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn subtracted(&self, (&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QPolygonF::subtracted(const QPolygonF& r) const

C++ documentation:

Returns a polygon which is r subtracted from this polygon.

Set operations on polygons will treat the polygons as areas. Non-closed polygons will be treated as implicitly closed.

This function was introduced in Qt 4.3.

fn swap(&mut self, other: &mut PolygonF)

C++ method: void QPolygonF::swap(QPolygonF& other)

C++ documentation:

Swaps polygon other with this polygon. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonFTakeAtArgs<'largs>

C++ method: QPolygonF::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QPolygonF::takeAt(int i)

C++ documentation:

Removes the element at index position i and returns it.

Equivalent to

T t = at(i);
remove(i);
return t;

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also takeFirst(), takeLast(), and QList::takeAt().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonFTakeFirstArgs<'largs>

C++ method: QPolygonF::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QPolygonF::takeFirst()

C++ documentation:

Removes the first item in the vector and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also takeLast() and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: PolygonFTakeLastArgs<'largs>

C++ method: QPolygonF::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QPolygonF::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

If you don't use the return value, removeLast() is more efficient.

This function was introduced in Qt 5.1.

See also takeFirst() and removeLast().

fn to_polygon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFToPolygonArgs<'largs>

C++ method: QPolygonF::toPolygon

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_polygon(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn to_polygon(&self, cpp_utils::AsStruct) -> ::polygon::Polygon
C++ method: QPolygon QPolygonF::toPolygon() const

C++ documentation:

Creates and returns a QPolygon by converting each QPointF to a QPoint.

See also QPointF::toPoint().

fn translate<'largs, Args>(&'largs mut self, args: Args) where Args: PolygonFTranslateArgs<'largs>

C++ method: QPolygonF::translate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn translate(&mut self, &qt_core::point_f::PointF) -> ()
C++ method: void QPolygonF::translate(const QPointF& offset)

C++ documentation:

Translate all points in the polygon by the given offset.

See also translated().

Variant 2

Rust arguments: fn translate(&mut self, (libc::c_double, libc::c_double)) -> ()
C++ method: void QPolygonF::translate(double dx, double dy)

C++ documentation:

This is an overloaded function.

Translates all points in the polygon by (dx, dy).

See also translated().

fn translated<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFTranslatedArgs<'largs>

C++ method: QPolygonF::translated

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn translated(&self, (&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn translated(&self, (&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QPolygonF::translated(const QPointF& offset) const

C++ documentation:

Returns a copy of the polygon that is translated by the given offset.

This function was introduced in Qt 4.6.

See also translate().

Variant 2

Rust arguments:
1) fn translated(&self, (libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn translated(&self, (libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QPolygonF::translated(double dx, double dy) const

C++ documentation:

This is an overloaded function.

Returns a copy of the polygon that is translated by (dx, dy).

This function was introduced in Qt 4.6.

See also translate().

fn united<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFUnitedArgs<'largs>

C++ method: QPolygonF::united

This is an overloaded function. Available variants:

Rust arguments:
1) fn united(&self, (&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn united(&self, (&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QPolygonF::united(const QPolygonF& r) const

C++ documentation:

Returns a polygon which is the union of this polygon and r.

Set operations on polygons will treat the polygons as areas. Non-closed polygons will be treated as implicitly closed.

This function was introduced in Qt 4.3.

See also intersected() and subtracted().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PolygonFValueArgs<'largs>

C++ method: QPolygonF::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QPolygonF::value(int i) const

C++ documentation:

Returns the value at index position i in the vector.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that i is within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn value(&self, (libc::c_int, &qt_core::point_f::PointF, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QPolygonF::value(int i, const QPointF& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.