Module qt_gui::qmatrix4x4::overloading

Traits

Matrix4x4AsQVariantArgs
This trait represents a set of arguments accepted by Matrix4x4::as_q_variant method.

Matrix4x4ColumnArgs
This trait represents a set of arguments accepted by Matrix4x4::column method.

Matrix4x4InvertedArgs
This trait represents a set of arguments accepted by Matrix4x4::inverted method.

Matrix4x4MapArgs
This trait represents a set of arguments accepted by Matrix4x4::map method.

Matrix4x4MapRectArgs
This trait represents a set of arguments accepted by Matrix4x4::map_rect method.

Matrix4x4MapVectorArgs
This trait represents a set of arguments accepted by Matrix4x4::map_vector method.

Matrix4x4NewArgs
This trait represents a set of arguments accepted by Matrix4x4::new method.

Matrix4x4OpMulAssignArgs
This trait represents a set of arguments accepted by Matrix4x4::op_mul_assign method.

Matrix4x4OrthoArgs
This trait represents a set of arguments accepted by Matrix4x4::ortho method.

Matrix4x4RotateArgs
This trait represents a set of arguments accepted by Matrix4x4::rotate method.

Matrix4x4RowArgs
This trait represents a set of arguments accepted by Matrix4x4::row method.

Matrix4x4ScaleArgs
This trait represents a set of arguments accepted by Matrix4x4::scale method.

Matrix4x4ToAffineArgs
This trait represents a set of arguments accepted by Matrix4x4::to_affine method.

Matrix4x4ToTransformArgs
This trait represents a set of arguments accepted by Matrix4x4::to_transform method.

Matrix4x4TranslateArgs
This trait represents a set of arguments accepted by Matrix4x4::translate method.

Matrix4x4TransposedArgs
This trait represents a set of arguments accepted by Matrix4x4::transposed method.

Matrix4x4ViewportArgs
This trait represents a set of arguments accepted by Matrix4x4::viewport method.