Struct qt_gui::qmatrix4x4::Matrix4x4

pub struct Matrix4x4 {
    // some fields omitted
}

C++ type: QMatrix4x4

C++ documentation:

The QMatrix4x4 class represents a 4x4 transformation matrix in 3D space.

The QMatrix4x4 class in general is treated as a row-major matrix, in that the constructors and operator() functions take data in row-major format, as is familiar in C-style usage.

Internally the data is stored as column-major format, so as to be optimal for passing to OpenGL functions, which expect column-major data.

When using these functions be aware that they return data in column-major format:

data()

constData()

Methods

impl Matrix4x4

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4AsQVariantArgs<'largs>

C++ method: QMatrix4x4::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QMatrix4x4::operator QVariant() const

C++ documentation:

Returns the matrix as a QVariant.

fn column<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4ColumnArgs<'largs>

C++ method: QMatrix4x4::column

This is an overloaded function. Available variants:

Rust arguments:
1) fn column(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn column(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
C++ method: QVector4D QMatrix4x4::column(int index) const

C++ documentation:

Returns the elements of column index as a 4D vector.

See also setColumn() and row().

fn const_data(&self) -> *const c_float

C++ method: const float* QMatrix4x4::constData() const

C++ documentation:

Returns a constant pointer to the raw data of this matrix. This raw data is stored in column-major format.

See also data().

fn copy_data_to(&self, values: *mut c_float)

C++ method: void QMatrix4x4::copyDataTo(float* values) const

C++ documentation:

Retrieves the 16 items in this matrix and copies them to values in row-major order.

fn data(&self) -> *const c_float

C++ method: const float* QMatrix4x4::data() const

C++ documentation:

Returns a constant pointer to the raw data of this matrix. This raw data is stored in column-major format.

See also constData().

fn data_mut(&mut self) -> *mut c_float

C++ method: float* QMatrix4x4::data()

C++ documentation:

Returns a pointer to the raw data of this matrix.

See also constData() and optimize().

fn determinant(&self) -> c_double

C++ method: double QMatrix4x4::determinant() const

C++ documentation:

Returns the determinant of this matrix.

fn fill(&mut self, value: c_float)

C++ method: void QMatrix4x4::fill(float value)

C++ documentation:

Fills all elements of this matrx with value.

fn flip_coordinates(&mut self)

C++ method: void QMatrix4x4::flipCoordinates()

C++ documentation:

Flips between right-handed and left-handed coordinate systems by multiplying the y and z co-ordinates by -1. This is normally used to create a left-handed orthographic view without scaling the viewport as ortho() does.

See also ortho().

Related Non-Members

fn frustum(&mut self, left: c_float, right: c_float, bottom: c_float, top: c_float, near_plane: c_float, far_plane: c_float)

C++ method: void QMatrix4x4::frustum(float left, float right, float bottom, float top, float nearPlane, float farPlane)

C++ documentation:

Multiplies this matrix by another that applies a perspective frustum projection for a window with lower-left corner (left, bottom), upper-right corner (right, top), and the specified nearPlane and farPlane clipping planes.

See also ortho() and perspective().

fn inverted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4InvertedArgs<'largs>

C++ method: QMatrix4x4::inverted

This is an overloaded function. Available variants:

Rust arguments:
1) fn inverted(&self, (*mut bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
2) fn inverted(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
3) fn inverted(&self, (*mut bool, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
4) fn inverted(&self, cpp_utils::AsStruct) -> ::qmatrix4x4::Matrix4x4
C++ method: QMatrix4x4 QMatrix4x4::inverted(bool* invertible = ?) const

C++ documentation:

Returns the inverse of this matrix. Returns the identity if this matrix cannot be inverted; i.e. determinant() is zero. If invertible is not null, then true will be written to that location if the matrix can be inverted; false otherwise.

If the matrix is recognized as the identity or an orthonormal matrix, then this function will quickly invert the matrix using optimized routines.

See also determinant() and normalMatrix().

fn is_affine(&self) -> bool

C++ method: bool QMatrix4x4::isAffine() const

C++ documentation:

Returns true if this matrix is affine matrix; false otherwise.

An affine matrix is a 4x4 matrix with row 3 equal to (0, 0, 0, 1), e.g. no projective coefficients.

This function was introduced in Qt 5.5.

See also isIdentity().

fn is_identity(&self) -> bool

C++ method: bool QMatrix4x4::isIdentity() const

C++ documentation:

Returns true if this matrix is the identity; false otherwise.

See also setToIdentity().

fn look_at(&mut self, eye: &Vector3D, center: &Vector3D, up: &Vector3D)

C++ method: void QMatrix4x4::lookAt(const QVector3D& eye, const QVector3D& center, const QVector3D& up)

C++ documentation:

Multiplies this matrix by a viewing matrix derived from an eye point. The center value indicates the center of the view that the eye is looking at. The up value indicates which direction should be considered up with respect to the eye.

Note: The up vector must not be parallel to the line of sight from eye to center.

fn map<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4MapArgs<'largs>

C++ method: QMatrix4x4::map

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map(&self, (&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn map(&self, (&qt_core::point::Point, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QMatrix4x4::map(const QPoint& point) const

C++ documentation:

Maps point by multiplying this matrix by point.

See also mapRect().

Variant 2

Rust arguments:
1) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QMatrix4x4::map(const QPointF& point) const

C++ documentation:

Maps point by multiplying this matrix by point.

See also mapRect().

Variant 3

Rust arguments:
1) fn map(&self, (&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn map(&self, (&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: QVector3D QMatrix4x4::map(const QVector3D& point) const

C++ documentation:

Maps point by multiplying this matrix by point.

See also mapRect() and mapVector().

Variant 4

Rust arguments:
1) fn map(&self, (&::qvector4d::Vector4D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn map(&self, (&::qvector4d::Vector4D, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
C++ method: QVector4D QMatrix4x4::map(const QVector4D& point) const

C++ documentation:

Maps point by multiplying this matrix by point.

See also mapRect().

fn map_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4MapRectArgs<'largs>

C++ method: QMatrix4x4::mapRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QMatrix4x4::mapRect(const QRect& rect) const

C++ documentation:

Maps rect by multiplying this matrix by the corners of rect and then forming a new rectangle from the results. The returned rectangle will be an ordinary 2D rectangle with sides parallel to the horizontal and vertical axes.

See also map().

Variant 2

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QMatrix4x4::mapRect(const QRectF& rect) const

C++ documentation:

Maps rect by multiplying this matrix by the corners of rect and then forming a new rectangle from the results. The returned rectangle will be an ordinary 2D rectangle with sides parallel to the horizontal and vertical axes.

See also map().

fn map_vector<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4MapVectorArgs<'largs>

C++ method: QMatrix4x4::mapVector

This is an overloaded function. Available variants:

Rust arguments:
1) fn map_vector(&self, (&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn map_vector(&self, (&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: QVector3D QMatrix4x4::mapVector(const QVector3D& vector) const

C++ documentation:

Maps vector by multiplying the top 3x3 portion of this matrix by vector. The translation and projection components of this matrix are ignored.

See also map().

fn new<Args>(args: Args) -> Args::ReturnType where Args: Matrix4x4NewArgs

C++ method: QMatrix4x4::QMatrix4x4

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::qmatrix4x4::Matrix4x4
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&::matrix::Matrix, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((&::matrix::Matrix, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(const QMatrix& matrix)

Variant 4

Rust arguments:
1) fn new((&::transform::Transform, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((&::transform::Transform, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(const QTransform& transform)

Variant 5

Rust arguments:
1) fn new((*const libc::c_float, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((*const libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(const float* values)

Variant 6

Rust arguments:
1) fn new((*const libc::c_float, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((*const libc::c_float, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(const float* values, int cols, int rows)

Variant 7

Rust arguments:
1) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::qmatrix4x4::Matrix4x4
2) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
C++ method: [constructor] void QMatrix4x4::QMatrix4x4(float m11, float m12, float m13, float m14, float m21, float m22, float m23, float m24, float m31, float m32, float m33, float m34, float m41, float m42, float m43, float m44)

fn op_add_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Matrix4x4) -> &'l0 mut Matrix4x4

C++ method: QMatrix4x4& QMatrix4x4::operator+=(const QMatrix4x4& other)

C++ documentation:

Adds the contents of other to this matrix.

fn op_call<'l0>(&'l0 self, row: c_int, column: c_int) -> &'l0 c_float

C++ method: const float& QMatrix4x4::operator()(int row, int column) const

C++ documentation:

Returns a constant reference to the element at position (row, column) in this matrix.

See also column() and row().

fn op_call_mut<'l0>(&'l0 mut self, row: c_int, column: c_int) -> &'l0 mut c_float

C++ method: float& QMatrix4x4::operator()(int row, int column)

C++ documentation:

Returns a reference to the element at position (row, column) in this matrix so that the element can be assigned to.

See also optimize(), setColumn(), and setRow().

fn op_div_assign<'l0>(&'l0 mut self, divisor: c_float) -> &'l0 mut Matrix4x4

C++ method: QMatrix4x4& QMatrix4x4::operator/=(float divisor)

C++ documentation:

This is an overloaded function.

Divides all elements of this matrix by divisor.

fn op_eq(&self, other: &Matrix4x4) -> bool

C++ method: bool QMatrix4x4::operator==(const QMatrix4x4& other) const

C++ documentation:

Returns true if this matrix is identical to other; false otherwise. This operator uses an exact floating-point comparison.

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Matrix4x4 where Args: Matrix4x4OpMulAssignArgs<'largs>

C++ method: QMatrix4x4::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::qmatrix4x4::Matrix4x4) -> &'l0 mut ::qmatrix4x4::Matrix4x4
C++ method: QMatrix4x4& QMatrix4x4::operator*=(const QMatrix4x4& other)

C++ documentation:

Multiplies the contents of other by this matrix.

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::qmatrix4x4::Matrix4x4
C++ method: QMatrix4x4& QMatrix4x4::operator*=(float factor)

C++ documentation:

This is an overloaded function.

Multiplies all elements of this matrix by factor.

fn op_neq(&self, other: &Matrix4x4) -> bool

C++ method: bool QMatrix4x4::operator!=(const QMatrix4x4& other) const

C++ documentation:

Returns true if this matrix is not identical to other; false otherwise. This operator uses an exact floating-point comparison.

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Matrix4x4) -> &'l0 mut Matrix4x4

C++ method: QMatrix4x4& QMatrix4x4::operator-=(const QMatrix4x4& other)

C++ documentation:

Subtracts the contents of other from this matrix.

fn optimize(&mut self)

C++ method: void QMatrix4x4::optimize()

C++ documentation:

Optimize the usage of this matrix from its current elements.

Some operations such as translate(), scale(), and rotate() can be performed more efficiently if the matrix being modified is already known to be the identity, a previous translate(), a previous scale(), etc.

Normally the QMatrix4x4 class keeps track of this special type internally as operations are performed. However, if the matrix is modified directly with operator()() or data(), then QMatrix4x4 will lose track of the special type and will revert to the safest but least efficient operations thereafter.

By calling optimize() after directly modifying the matrix, the programmer can force QMatrix4x4 to recover the special type if the elements appear to conform to one of the known optimized types.

See also operator()(), data(), and translate().

fn ortho<'largs, Args>(&'largs mut self, args: Args) where Args: Matrix4x4OrthoArgs<'largs>

C++ method: QMatrix4x4::ortho

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn ortho(&mut self, &qt_core::rect::Rect) -> ()
C++ method: void QMatrix4x4::ortho(const QRect& rect)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that applies an orthographic projection for a window with boundaries specified by rect. The near and far clipping planes will be -1 and 1 respectively.

See also frustum() and perspective().

Variant 2

Rust arguments: fn ortho(&mut self, &qt_core::rect_f::RectF) -> ()
C++ method: void QMatrix4x4::ortho(const QRectF& rect)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that applies an orthographic projection for a window with boundaries specified by rect. The near and far clipping planes will be -1 and 1 respectively.

See also frustum() and perspective().

Variant 3

Rust arguments: fn ortho(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::ortho(float left, float right, float bottom, float top, float nearPlane, float farPlane)

C++ documentation:

Multiplies this matrix by another that applies an orthographic projection for a window with lower-left corner (left, bottom), upper-right corner (right, top), and the specified nearPlane and farPlane clipping planes.

See also frustum() and perspective().

fn perspective(&mut self, vertical_angle: c_float, aspect_ratio: c_float, near_plane: c_float, far_plane: c_float)

C++ method: void QMatrix4x4::perspective(float verticalAngle, float aspectRatio, float nearPlane, float farPlane)

C++ documentation:

Multiplies this matrix by another that applies a perspective projection. The vertical field of view will be verticalAngle degrees within a window with a given aspectRatio that determines the horizontal field of view. The projection will have the specified nearPlane and farPlane clipping planes which are the distances from the viewer to the corresponding planes.

See also ortho() and frustum().

fn rotate<'largs, Args>(&'largs mut self, args: Args) where Args: Matrix4x4RotateArgs<'largs>

C++ method: QMatrix4x4::rotate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn rotate(&mut self, &::quaternion::Quaternion) -> ()
C++ method: void QMatrix4x4::rotate(const QQuaternion& quaternion)

C++ documentation:

Multiples this matrix by another that rotates coordinates according to a specified quaternion. The quaternion is assumed to have been normalized.

See also scale(), translate(), and QQuaternion.

Variant 2

Rust arguments: fn rotate(&mut self, (libc::c_float, &::qvector3d::Vector3D)) -> ()
C++ method: void QMatrix4x4::rotate(float angle, const QVector3D& vector)

C++ documentation:

Multiples this matrix by another that rotates coordinates through angle degrees about vector.

See also scale() and translate().

Variant 3

Rust arguments:
1) fn rotate(&mut self, (libc::c_float, libc::c_float, libc::c_float)) -> ()
2) fn rotate(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::rotate(float angle, float x, float y, float z = ?)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that rotates coordinates through angle degrees about the vector (x, y, z).

See also scale() and translate().

fn row<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4RowArgs<'largs>

C++ method: QMatrix4x4::row

This is an overloaded function. Available variants:

Rust arguments:
1) fn row(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn row(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
C++ method: QVector4D QMatrix4x4::row(int index) const

C++ documentation:

Returns the elements of row index as a 4D vector.

See also setRow() and column().

fn scale<'largs, Args>(&'largs mut self, args: Args) where Args: Matrix4x4ScaleArgs<'largs>

C++ method: QMatrix4x4::scale

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn scale(&mut self, &::qvector3d::Vector3D) -> ()
C++ method: void QMatrix4x4::scale(const QVector3D& vector)

C++ documentation:

Multiplies this matrix by another that scales coordinates by the components of vector.

See also translate() and rotate().

Variant 2

Rust arguments: fn scale(&mut self, libc::c_float) -> ()
C++ method: void QMatrix4x4::scale(float factor)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that scales coordinates by the given factor.

See also translate() and rotate().

Variant 3

Rust arguments: fn scale(&mut self, (libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::scale(float x, float y)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that scales coordinates by the components x, and y.

See also translate() and rotate().

Variant 4

Rust arguments: fn scale(&mut self, (libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::scale(float x, float y, float z)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that scales coordinates by the components x, y, and z.

See also translate() and rotate().

fn set_column(&mut self, index: c_int, value: &Vector4D)

C++ method: void QMatrix4x4::setColumn(int index, const QVector4D& value)

C++ documentation:

Sets the elements of column index to the components of value.

See also column() and setRow().

fn set_row(&mut self, index: c_int, value: &Vector4D)

C++ method: void QMatrix4x4::setRow(int index, const QVector4D& value)

C++ documentation:

Sets the elements of row index to the components of value.

See also row() and setColumn().

fn set_to_identity(&mut self)

C++ method: void QMatrix4x4::setToIdentity()

C++ documentation:

Sets this matrix to the identity.

See also isIdentity().

fn to_affine<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4ToAffineArgs<'largs>

C++ method: QMatrix4x4::toAffine

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_affine(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::matrix::Matrix>
2) fn to_affine(&self, cpp_utils::AsStruct) -> ::matrix::Matrix
C++ method: QMatrix QMatrix4x4::toAffine() const

C++ documentation:

Returns the conventional Qt 2D affine transformation matrix that corresponds to this matrix. It is assumed that this matrix only contains 2D affine transformation elements.

See also toTransform().

fn to_transform<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4ToTransformArgs<'largs>

C++ method: QMatrix4x4::toTransform

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_transform(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
2) fn to_transform(&self, cpp_utils::AsStruct) -> ::transform::Transform
C++ method: QTransform QMatrix4x4::toTransform() const

C++ documentation:

Returns the conventional Qt 2D transformation matrix that corresponds to this matrix.

The returned QTransform is formed by simply dropping the third row and third column of the QMatrix4x4. This is suitable for implementing orthographic projections where the z co-ordinate should be dropped rather than projected.

See also toAffine().

Variant 2

Rust arguments:
1) fn to_transform(&self, (libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
2) fn to_transform(&self, (libc::c_float, cpp_utils::AsStruct)) -> ::transform::Transform
C++ method: QTransform QMatrix4x4::toTransform(float distanceToPlane) const

C++ documentation:

Returns the conventional Qt 2D transformation matrix that corresponds to this matrix.

If distanceToPlane is non-zero, it indicates a projection factor to use to adjust for the z co-ordinate. The value of 1024 corresponds to the projection factor used by QTransform::rotate() for the x and y axes.

If distanceToPlane is zero, then the returned QTransform is formed by simply dropping the third row and third column of the QMatrix4x4. This is suitable for implementing orthographic projections where the z co-ordinate should be dropped rather than projected.

See also toAffine().

fn translate<'largs, Args>(&'largs mut self, args: Args) where Args: Matrix4x4TranslateArgs<'largs>

C++ method: QMatrix4x4::translate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn translate(&mut self, &::qvector3d::Vector3D) -> ()
C++ method: void QMatrix4x4::translate(const QVector3D& vector)

C++ documentation:

Multiplies this matrix by another that translates coordinates by the components of vector.

See also scale() and rotate().

Variant 2

Rust arguments: fn translate(&mut self, (libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::translate(float x, float y)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that translates coordinates by the components x, and y.

See also scale() and rotate().

Variant 3

Rust arguments: fn translate(&mut self, (libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::translate(float x, float y, float z)

C++ documentation:

This is an overloaded function.

Multiplies this matrix by another that translates coordinates by the components x, y, and z.

See also scale() and rotate().

fn transposed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Matrix4x4TransposedArgs<'largs>

C++ method: QMatrix4x4::transposed

This is an overloaded function. Available variants:

Rust arguments:
1) fn transposed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qmatrix4x4::Matrix4x4>
2) fn transposed(&self, cpp_utils::AsStruct) -> ::qmatrix4x4::Matrix4x4
C++ method: QMatrix4x4 QMatrix4x4::transposed() const

C++ documentation:

Returns this matrix, transposed about its diagonal.

fn viewport<'largs, Args>(&'largs mut self, args: Args) where Args: Matrix4x4ViewportArgs<'largs>

C++ method: QMatrix4x4::viewport

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn viewport(&mut self, &qt_core::rect_f::RectF) -> ()
C++ method: void QMatrix4x4::viewport(const QRectF& rect)

C++ documentation:

This is an overloaded function.

Sets up viewport transform for viewport bounded by rect and with near and far set to 0 and 1 respectively.

Variant 2

Rust arguments:
1) fn viewport(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
2) fn viewport(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
3) fn viewport(&mut self, (libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QMatrix4x4::viewport(float left, float bottom, float width, float height, float nearPlane = ?, float farPlane = ?)

C++ documentation:

Multiplies this matrix by another that performs the scale and bias transformation used by OpenGL to transform from normalized device coordinates (NDC) to viewport (window) coordinates. That is it maps points from the cube ranging over [-1, 1] in each dimension to the viewport with it's near-lower-left corner at (left, bottom, nearPlane) and with size (width, height, farPlane - nearPlane).

This matches the transform used by the fixed function OpenGL viewport transform controlled by the functions glViewport() and glDepthRange().