Module qt_gui::quaternion::overloading

Traits

OpMulArgs
This trait represents a set of arguments accepted by op_mul method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

QuaternionAsQVariantArgs
This trait represents a set of arguments accepted by Quaternion::as_q_variant method.

QuaternionConjugateArgs
This trait represents a set of arguments accepted by Quaternion::conjugate method.

QuaternionConjugatedArgs
This trait represents a set of arguments accepted by Quaternion::conjugated method.

QuaternionFromAxesArgs
This trait represents a set of arguments accepted by Quaternion::from_axes method.

QuaternionFromAxisAndAngleArgs
This trait represents a set of arguments accepted by Quaternion::from_axis_and_angle method.

QuaternionFromDirectionArgs
This trait represents a set of arguments accepted by Quaternion::from_direction method.

QuaternionFromEulerAnglesArgs
This trait represents a set of arguments accepted by Quaternion::from_euler_angles method.

QuaternionGetAxisAndAngleArgs
This trait represents a set of arguments accepted by Quaternion::get_axis_and_angle method.

QuaternionInvertedArgs
This trait represents a set of arguments accepted by Quaternion::inverted method.

QuaternionNewArgs
This trait represents a set of arguments accepted by Quaternion::new method.

QuaternionNlerpArgs
This trait represents a set of arguments accepted by Quaternion::nlerp method.

QuaternionNormalizedArgs
This trait represents a set of arguments accepted by Quaternion::normalized method.

QuaternionOpMulAssignArgs
This trait represents a set of arguments accepted by Quaternion::op_mul_assign method.

QuaternionRotatedVectorArgs
This trait represents a set of arguments accepted by Quaternion::rotated_vector method.

QuaternionRotationToArgs
This trait represents a set of arguments accepted by Quaternion::rotation_to method.

QuaternionSetVectorArgs
This trait represents a set of arguments accepted by Quaternion::set_vector method.

QuaternionSlerpArgs
This trait represents a set of arguments accepted by Quaternion::slerp method.

QuaternionToEulerAnglesArgs
This trait represents a set of arguments accepted by Quaternion::to_euler_angles method.

QuaternionToVector4DArgs
This trait represents a set of arguments accepted by Quaternion::to_vector4_d method.

QuaternionVectorArgs
This trait represents a set of arguments accepted by Quaternion::vector method.