Struct qt_gui::quaternion::Quaternion

pub struct Quaternion {
    // some fields omitted
}

C++ type: QQuaternion

C++ documentation:

The QQuaternion class represents a quaternion consisting of a vector and scalar.

Quaternions are used to represent rotations in 3D space, and consist of a 3D rotation axis specified by the x, y, and z coordinates, and a scalar representing the rotation angle.

Methods

impl Quaternion

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionAsQVariantArgs<'largs>

C++ method: QQuaternion::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QQuaternion::operator QVariant() const

C++ documentation:

Returns the quaternion as a QVariant.

fn conjugate<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionConjugateArgs<'largs>

C++ method: QQuaternion::conjugate

This is an overloaded function. Available variants:

Rust arguments:
1) fn conjugate(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn conjugate(&self, cpp_utils::AsStruct) -> ::quaternion::Quaternion
C++ method: QQuaternion QQuaternion::conjugate() const

C++ documentation:

Use conjugated() instead.

fn conjugated<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionConjugatedArgs<'largs>

C++ method: QQuaternion::conjugated

This is an overloaded function. Available variants:

Rust arguments:
1) fn conjugated(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn conjugated(&self, cpp_utils::AsStruct) -> ::quaternion::Quaternion
C++ method: QQuaternion QQuaternion::conjugated() const

C++ documentation:

Returns the conjugate of this quaternion, which is (-x, -y, -z, scalar).

This function was introduced in Qt 5.5.

fn dot_product(q1: &Quaternion, q2: &Quaternion) -> c_float

C++ method: static float QQuaternion::dotProduct(const QQuaternion& q1, const QQuaternion& q2)

C++ documentation:

Returns the dot product of q1 and q2.

This function was introduced in Qt 5.5.

See also length().

fn from_axes<Args>(args: Args) -> Args::ReturnType where Args: QuaternionFromAxesArgs

C++ method: QQuaternion::fromAxes

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_axes((&::qvector3d::Vector3D, &::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_axes((&::qvector3d::Vector3D, &::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromAxes(const QVector3D& xAxis, const QVector3D& yAxis, const QVector3D& zAxis)

C++ documentation:

Constructs the quaternion using 3 axes (xAxis, yAxis, zAxis).

Note: The axes are assumed to be orthonormal.

This function was introduced in Qt 5.5.

See also getAxes() and fromRotationMatrix().

fn from_axis_and_angle<Args>(args: Args) -> Args::ReturnType where Args: QuaternionFromAxisAndAngleArgs

C++ method: QQuaternion::fromAxisAndAngle

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_axis_and_angle((&::qvector3d::Vector3D, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_axis_and_angle((&::qvector3d::Vector3D, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromAxisAndAngle(const QVector3D& axis, float angle)

C++ documentation:

Creates a normalized quaternion that corresponds to rotating through angle degrees about the specified 3D axis.

See also getAxisAndAngle().

Variant 2

Rust arguments:
1) fn from_axis_and_angle((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_axis_and_angle((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromAxisAndAngle(float x, float y, float z, float angle)

C++ documentation:

Creates a normalized quaternion that corresponds to rotating through angle degrees about the 3D axis (x, y, z).

See also getAxisAndAngle().

fn from_direction<Args>(args: Args) -> Args::ReturnType where Args: QuaternionFromDirectionArgs

C++ method: QQuaternion::fromDirection

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_direction((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_direction((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromDirection(const QVector3D& direction, const QVector3D& up)

C++ documentation:

Constructs the quaternion using specified forward direction direction and upward direction up. If the upward direction was not specified or the forward and upward vectors are collinear, a new orthonormal upward direction will be generated.

This function was introduced in Qt 5.5.

See also fromAxes() and rotationTo().

fn from_euler_angles<Args>(args: Args) -> Args::ReturnType where Args: QuaternionFromEulerAnglesArgs

C++ method: QQuaternion::fromEulerAngles

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_euler_angles((&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_euler_angles((&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromEulerAngles(const QVector3D& eulerAngles)

C++ documentation:

This is an overloaded function.

Creates a quaternion that corresponds to a rotation of eulerAngles: eulerAngles.z() degrees around the z axis, eulerAngles.x() degrees around the x axis, and eulerAngles.y() degrees around the y axis (in that order).

This function was introduced in Qt 5.5.

See also toEulerAngles().

Variant 2

Rust arguments:
1) fn from_euler_angles((libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn from_euler_angles((libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::fromEulerAngles(float pitch, float yaw, float roll)

C++ documentation:

Creates a quaternion that corresponds to a rotation of roll degrees around the z axis, pitch degrees around the x axis, and yaw degrees around the y axis (in that order).

This function was introduced in Qt 5.5.

See also getEulerAngles().

fn get_axes(&self, x_axis: *mut Vector3D, y_axis: *mut Vector3D, z_axis: *mut Vector3D)

C++ method: void QQuaternion::getAxes(QVector3D* xAxis, QVector3D* yAxis, QVector3D* zAxis) const

C++ documentation:

Returns the 3 orthonormal axes (xAxis, yAxis, zAxis) defining the quaternion.

This function was introduced in Qt 5.5.

See also fromAxes() and toRotationMatrix().

fn get_axis_and_angle<'largs, Args>(&'largs self, args: Args) where Args: QuaternionGetAxisAndAngleArgs<'largs>

C++ method: QQuaternion::getAxisAndAngle

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn get_axis_and_angle(&self, (*mut ::qvector3d::Vector3D, *mut libc::c_float)) -> ()
C++ method: void QQuaternion::getAxisAndAngle(QVector3D* axis, float* angle) const

C++ documentation:

This is an overloaded function.

Extracts a 3D axis axis and a rotating angle angle (in degrees) that corresponds to this quaternion.

This function was introduced in Qt 5.5.

See also fromAxisAndAngle().

Variant 2

Rust arguments: fn get_axis_and_angle(&self, (*mut libc::c_float, *mut libc::c_float, *mut libc::c_float, *mut libc::c_float)) -> ()
C++ method: void QQuaternion::getAxisAndAngle(float* x, float* y, float* z, float* angle) const

C++ documentation:

Extracts a 3D axis (x, y, z) and a rotating angle angle (in degrees) that corresponds to this quaternion.

This function was introduced in Qt 5.5.

See also fromAxisAndAngle().

fn get_euler_angles(&self, pitch: *mut c_float, yaw: *mut c_float, roll: *mut c_float)

C++ method: void QQuaternion::getEulerAngles(float* pitch, float* yaw, float* roll) const

C++ documentation:

Calculates roll, pitch, and yaw Euler angles (in degrees) that corresponds to this quaternion.

This function was introduced in Qt 5.5.

See also fromEulerAngles().

fn inverted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionInvertedArgs<'largs>

C++ method: QQuaternion::inverted

This is an overloaded function. Available variants:

Rust arguments:
1) fn inverted(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn inverted(&self, cpp_utils::AsStruct) -> ::quaternion::Quaternion
C++ method: QQuaternion QQuaternion::inverted() const

C++ documentation:

Returns the inverse of this quaternion. If this quaternion is null, then a null quaternion is returned.

This function was introduced in Qt 5.5.

See also isNull() and length().

fn is_identity(&self) -> bool

C++ method: bool QQuaternion::isIdentity() const

C++ documentation:

Returns true if the x, y, and z components of this quaternion are set to 0.0, and the scalar component is set to 1.0; otherwise returns false.

fn is_null(&self) -> bool

C++ method: bool QQuaternion::isNull() const

C++ documentation:

Returns true if the x, y, z, and scalar components of this quaternion are set to 0.0; otherwise returns false.

fn length(&self) -> c_float

C++ method: float QQuaternion::length() const

C++ documentation:

Returns the length of the quaternion. This is also called the "norm".

See also lengthSquared(), normalized(), and dotProduct().

fn length_squared(&self) -> c_float

C++ method: float QQuaternion::lengthSquared() const

C++ documentation:

Returns the squared length of the quaternion.

See also length() and dotProduct().

fn new<Args>(args: Args) -> Args::ReturnType where Args: QuaternionNewArgs

C++ method: QQuaternion::QQuaternion

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::quaternion::Quaternion
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::quaternion::Quaternion>
C++ method: [constructor] void QQuaternion::QQuaternion()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
C++ method: [constructor] void QQuaternion::QQuaternion(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&::qvector4d::Vector4D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
2) fn new((&::qvector4d::Vector4D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
C++ method: [constructor] void QQuaternion::QQuaternion(const QVector4D& vector)

Variant 4

Rust arguments:
1) fn new((libc::c_float, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
2) fn new((libc::c_float, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
C++ method: [constructor] void QQuaternion::QQuaternion(float scalar, const QVector3D& vector)

Variant 5

Rust arguments:
1) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
2) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
C++ method: [constructor] void QQuaternion::QQuaternion(float scalar, float xpos, float ypos, float zpos)

fn nlerp<Args>(args: Args) -> Args::ReturnType where Args: QuaternionNlerpArgs

C++ method: QQuaternion::nlerp

This is an overloaded function. Available variants:

Rust arguments:
1) fn nlerp((&::quaternion::Quaternion, &::quaternion::Quaternion, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn nlerp((&::quaternion::Quaternion, &::quaternion::Quaternion, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::nlerp(const QQuaternion& q1, const QQuaternion& q2, float t)

C++ documentation:

Interpolates along the shortest linear path between the rotational positions q1 and q2. The value t should be between 0 and 1, indicating the distance to travel between q1 and q2. The result will be normalized().

If t is less than or equal to 0, then q1 will be returned. If t is greater than or equal to 1, then q2 will be returned.

The nlerp() function is typically faster than slerp() and will give approximate results to spherical interpolation that are good enough for some applications.

See also slerp().

fn normalize(&mut self)

C++ method: void QQuaternion::normalize()

C++ documentation:

Normalizes the current quaternion in place. Nothing happens if this is a null quaternion or the length of the quaternion is very close to 1.

See also length() and normalized().

fn normalized<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionNormalizedArgs<'largs>

C++ method: QQuaternion::normalized

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn normalized(&self, cpp_utils::AsStruct) -> ::quaternion::Quaternion
C++ method: QQuaternion QQuaternion::normalized() const

C++ documentation:

Returns the normalized unit form of this quaternion.

If this quaternion is null, then a null quaternion is returned. If the length of the quaternion is very close to 1, then the quaternion will be returned as-is. Otherwise the normalized form of the quaternion of length 1 will be returned.

See also normalize(), length(), and dotProduct().

fn op_add_assign<'l0, 'l1>(&'l0 mut self, quaternion: &'l1 Quaternion) -> &'l0 mut Quaternion

C++ method: QQuaternion& QQuaternion::operator+=(const QQuaternion& quaternion)

C++ documentation:

Adds the given quaternion to this quaternion and returns a reference to this quaternion.

See also operator-=().

fn op_div_assign<'l0>(&'l0 mut self, divisor: c_float) -> &'l0 mut Quaternion

C++ method: QQuaternion& QQuaternion::operator/=(float divisor)

C++ documentation:

Divides this quaternion's components by the given divisor, and returns a reference to this quaternion.

See also operator*=().

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Quaternion where Args: QuaternionOpMulAssignArgs<'largs>

C++ method: QQuaternion::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::quaternion::Quaternion) -> &'l0 mut ::quaternion::Quaternion
C++ method: QQuaternion& QQuaternion::operator*=(const QQuaternion& quaternion)

C++ documentation:

Multiplies this quaternion by quaternion and returns a reference to this quaternion.

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::quaternion::Quaternion
C++ method: QQuaternion& QQuaternion::operator*=(float factor)

C++ documentation:

Multiplies this quaternion's components by the given factor, and returns a reference to this quaternion.

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, quaternion: &'l1 Quaternion) -> &'l0 mut Quaternion

C++ method: QQuaternion& QQuaternion::operator-=(const QQuaternion& quaternion)

C++ documentation:

Subtracts the given quaternion from this quaternion and returns a reference to this quaternion.

See also operator+=().

fn rotated_vector<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionRotatedVectorArgs<'largs>

C++ method: QQuaternion::rotatedVector

This is an overloaded function. Available variants:

Rust arguments:
1) fn rotated_vector(&self, (&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn rotated_vector(&self, (&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: QVector3D QQuaternion::rotatedVector(const QVector3D& vector) const

C++ documentation:

Rotates vector with this quaternion to produce a new vector in 3D space. The following code:

QVector3D result = q.rotatedVector(vector);

is equivalent to the following:

QVector3D result = (q  QQuaternion(0, vector)  q.conjugated()).vector();

fn rotation_to<Args>(args: Args) -> Args::ReturnType where Args: QuaternionRotationToArgs

C++ method: QQuaternion::rotationTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn rotation_to((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn rotation_to((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::rotationTo(const QVector3D& from, const QVector3D& to)

C++ documentation:

Returns the shortest arc quaternion to rotate from the direction described by the vector from to the direction described by the vector to.

This function was introduced in Qt 5.5.

See also fromDirection().

fn scalar(&self) -> c_float

C++ method: float QQuaternion::scalar() const

C++ documentation:

Returns the scalar component of this quaternion.

See also setScalar(), x(), y(), and z().

fn set_scalar(&mut self, scalar: c_float)

C++ method: void QQuaternion::setScalar(float scalar)

C++ documentation:

Sets the scalar component of this quaternion to scalar.

See also scalar(), setX(), setY(), and setZ().

fn set_vector<'largs, Args>(&'largs mut self, args: Args) where Args: QuaternionSetVectorArgs<'largs>

C++ method: QQuaternion::setVector

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_vector(&mut self, &::qvector3d::Vector3D) -> ()
C++ method: void QQuaternion::setVector(const QVector3D& vector)

C++ documentation:

Sets the vector component of this quaternion to vector.

See also vector() and setScalar().

Variant 2

Rust arguments: fn set_vector(&mut self, (libc::c_float, libc::c_float, libc::c_float)) -> ()
C++ method: void QQuaternion::setVector(float x, float y, float z)

C++ documentation:

Sets the vector component of this quaternion to (x, y, z).

See also vector() and setScalar().

fn set_x(&mut self, x: c_float)

C++ method: void QQuaternion::setX(float x)

C++ documentation:

Sets the x coordinate of this quaternion's vector to the given x coordinate.

See also x(), setY(), setZ(), and setScalar().

fn set_y(&mut self, y: c_float)

C++ method: void QQuaternion::setY(float y)

C++ documentation:

Sets the y coordinate of this quaternion's vector to the given y coordinate.

See also y(), setX(), setZ(), and setScalar().

fn set_z(&mut self, z: c_float)

C++ method: void QQuaternion::setZ(float z)

C++ documentation:

Sets the z coordinate of this quaternion's vector to the given z coordinate.

See also z(), setX(), setY(), and setScalar().

fn slerp<Args>(args: Args) -> Args::ReturnType where Args: QuaternionSlerpArgs

C++ method: QQuaternion::slerp

This is an overloaded function. Available variants:

Rust arguments:
1) fn slerp((&::quaternion::Quaternion, &::quaternion::Quaternion, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::quaternion::Quaternion>
2) fn slerp((&::quaternion::Quaternion, &::quaternion::Quaternion, libc::c_float, cpp_utils::AsStruct)) -> ::quaternion::Quaternion
C++ method: static QQuaternion QQuaternion::slerp(const QQuaternion& q1, const QQuaternion& q2, float t)

C++ documentation:

Interpolates along the shortest spherical path between the rotational positions q1 and q2. The value t should be between 0 and 1, indicating the spherical distance to travel between q1 and q2.

If t is less than or equal to 0, then q1 will be returned. If t is greater than or equal to 1, then q2 will be returned.

See also nlerp().

fn to_euler_angles<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionToEulerAnglesArgs<'largs>

C++ method: QQuaternion::toEulerAngles

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_euler_angles(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn to_euler_angles(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QQuaternion::toEulerAngles() const

C++ documentation:

This is an overloaded function.

Calculates roll, pitch, and yaw Euler angles (in degrees) that corresponds to this quaternion.

This function was introduced in Qt 5.5.

See also fromEulerAngles().

fn to_vector4_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionToVector4DArgs<'largs>

C++ method: QQuaternion::toVector4D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector4_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn to_vector4_d(&self, cpp_utils::AsStruct) -> ::qvector4d::Vector4D
C++ method: QVector4D QQuaternion::toVector4D() const

C++ documentation:

Returns this quaternion as a 4D vector.

fn vector<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: QuaternionVectorArgs<'largs>

C++ method: QQuaternion::vector

This is an overloaded function. Available variants:

Rust arguments:
1) fn vector(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn vector(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QQuaternion::vector() const

C++ documentation:

Returns the vector component of this quaternion.

See also setVector() and scalar().

fn x(&self) -> c_float

C++ method: float QQuaternion::x() const

C++ documentation:

Returns the x coordinate of this quaternion's vector.

See also setX(), y(), z(), and scalar().

fn y(&self) -> c_float

C++ method: float QQuaternion::y() const

C++ documentation:

Returns the y coordinate of this quaternion's vector.

See also setY(), x(), z(), and scalar().

fn z(&self) -> c_float

C++ method: float QQuaternion::z() const

C++ documentation:

Returns the z coordinate of this quaternion's vector.

See also setZ(), x(), y(), and scalar().