Module qt_gui::qvector2d::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

Vector2DAsQVariantArgs
This trait represents a set of arguments accepted by Vector2D::as_q_variant method.

Vector2DNewArgs
This trait represents a set of arguments accepted by Vector2D::new method.

Vector2DNormalizedArgs
This trait represents a set of arguments accepted by Vector2D::normalized method.

Vector2DOpDivAssignArgs
This trait represents a set of arguments accepted by Vector2D::op_div_assign method.

Vector2DOpMulAssignArgs
This trait represents a set of arguments accepted by Vector2D::op_mul_assign method.

Vector2DToPointArgs
This trait represents a set of arguments accepted by Vector2D::to_point method.

Vector2DToPointFArgs
This trait represents a set of arguments accepted by Vector2D::to_point_f method.

Vector2DToVector3DArgs
This trait represents a set of arguments accepted by Vector2D::to_vector3_d method.

Vector2DToVector4DArgs
This trait represents a set of arguments accepted by Vector2D::to_vector4_d method.