Struct qt_gui::qvector2d::Vector2D

pub struct Vector2D {
    // some fields omitted
}

C++ type: QVector2D

C++ documentation:

The QVector2D class represents a vector or vertex in 2D space.

The QVector2D class can also be used to represent vertices in 2D space. We therefore do not need to provide a separate vertex class.

Methods

impl Vector2D

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DAsQVariantArgs<'largs>

C++ method: QVector2D::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QVector2D::operator QVariant() const

C++ documentation:

Returns the 2D vector as a QVariant.

fn distance_to_line(&self, point: &Vector2D, direction: &Vector2D) -> c_float

C++ method: float QVector2D::distanceToLine(const QVector2D& point, const QVector2D& direction) const

C++ documentation:

Returns the distance that this vertex is from a line defined by point and the unit vector direction.

If direction is a null vector, then it does not define a line. In that case, the distance from point to this vertex is returned.

This function was introduced in Qt 5.1.

See also distanceToPoint().

fn distance_to_point(&self, point: &Vector2D) -> c_float

C++ method: float QVector2D::distanceToPoint(const QVector2D& point) const

C++ documentation:

Returns the distance from this vertex to a point defined by the vertex point.

This function was introduced in Qt 5.1.

See also distanceToLine().

fn dot_product(v1: &Vector2D, v2: &Vector2D) -> c_float

C++ method: static float QVector2D::dotProduct(const QVector2D& v1, const QVector2D& v2)

C++ documentation:

Returns the dot product of v1 and v2.

fn is_null(&self) -> bool

C++ method: bool QVector2D::isNull() const

C++ documentation:

Returns true if the x and y coordinates are set to 0.0, otherwise returns false.

fn length(&self) -> c_float

C++ method: float QVector2D::length() const

C++ documentation:

Returns the length of the vector from the origin.

See also lengthSquared() and normalized().

fn length_squared(&self) -> c_float

C++ method: float QVector2D::lengthSquared() const

C++ documentation:

Returns the squared length of the vector from the origin. This is equivalent to the dot product of the vector with itself.

See also length() and dotProduct().

fn new<Args>(args: Args) -> Args::ReturnType where Args: Vector2DNewArgs

C++ method: QVector2D::QVector2D

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::qvector2d::Vector2D
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&qt_core::point::Point, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(const QPoint& point)

Variant 4

Rust arguments:
1) fn new((&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(const QPointF& point)

Variant 5

Rust arguments:
1) fn new((&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(const QVector3D& vector)

Variant 6

Rust arguments:
1) fn new((&::qvector4d::Vector4D, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((&::qvector4d::Vector4D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(const QVector4D& vector)

Variant 7

Rust arguments:
1) fn new((libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::qvector2d::Vector2D
2) fn new((libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector2d::Vector2D>
C++ method: [constructor] void QVector2D::QVector2D(float xpos, float ypos)

fn normalize(&mut self)

C++ method: void QVector2D::normalize()

C++ documentation:

Normalizes the currect vector in place. Nothing happens if this vector is a null vector or the length of the vector is very close to 1.

See also length() and normalized().

fn normalized<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DNormalizedArgs<'largs>

C++ method: QVector2D::normalized

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
2) fn normalized(&self, cpp_utils::AsStruct) -> ::qvector2d::Vector2D
C++ method: QVector2D QVector2D::normalized() const

C++ documentation:

Returns the normalized unit vector form of this vector.

If this vector is null, then a null vector is returned. If the length of the vector is very close to 1, then the vector will be returned as-is. Otherwise the normalized form of the vector of length 1 will be returned.

See also length() and normalize().

fn op_add_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector2D) -> &'l0 mut Vector2D

C++ method: QVector2D& QVector2D::operator+=(const QVector2D& vector)

C++ documentation:

Adds the given vector to this vector and returns a reference to this vector.

See also operator-=().

fn op_div_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector2D where Args: Vector2DOpDivAssignArgs<'largs>

C++ method: QVector2D::operator/=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_div_assign(&mut self, &'l1 ::qvector2d::Vector2D) -> &'l0 mut ::qvector2d::Vector2D
C++ method: QVector2D& QVector2D::operator/=(const QVector2D& vector)

C++ documentation:

Divides the components of this vector by the corresponding components in vector.

This function was introduced in Qt 5.5.

See also operator*=().

Variant 2

Rust arguments: fn op_div_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector2d::Vector2D
C++ method: QVector2D& QVector2D::operator/=(float divisor)

C++ documentation:

Divides this vector's coordinates by the given divisor, and returns a reference to this vector.

See also operator*=().

fn op_index(&self, i: c_int) -> c_float

C++ method: float QVector2D::operator[](int i) const

C++ documentation:

Returns the component of the vector at index position i.

i must be a valid index position in the vector (i.e., 0 <= i < 2).

This function was introduced in Qt 5.2.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut c_float

C++ method: float& QVector2D::operator[](int i)

C++ documentation:

Returns the component of the vector at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < 2).

This function was introduced in Qt 5.2.

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector2D where Args: Vector2DOpMulAssignArgs<'largs>

C++ method: QVector2D::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::qvector2d::Vector2D) -> &'l0 mut ::qvector2d::Vector2D
C++ method: QVector2D& QVector2D::operator*=(const QVector2D& vector)

C++ documentation:

Multiplies the components of this vector by the corresponding components in vector.

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector2d::Vector2D
C++ method: QVector2D& QVector2D::operator*=(float factor)

C++ documentation:

Multiplies this vector's coordinates by the given factor, and returns a reference to this vector.

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector2D) -> &'l0 mut Vector2D

C++ method: QVector2D& QVector2D::operator-=(const QVector2D& vector)

C++ documentation:

Subtracts the given vector from this vector and returns a reference to this vector.

See also operator+=().

fn set_x(&mut self, x: c_float)

C++ method: void QVector2D::setX(float x)

C++ documentation:

Sets the x coordinate of this point to the given x coordinate.

See also x() and setY().

fn set_y(&mut self, y: c_float)

C++ method: void QVector2D::setY(float y)

C++ documentation:

Sets the y coordinate of this point to the given y coordinate.

See also y() and setX().

fn to_point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DToPointArgs<'largs>

C++ method: QVector2D::toPoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn to_point(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QVector2D::toPoint() const

C++ documentation:

Returns the QPoint form of this 2D vector.

See also toPointF() and toVector3D().

fn to_point_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DToPointFArgs<'largs>

C++ method: QVector2D::toPointF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn to_point_f(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QVector2D::toPointF() const

C++ documentation:

Returns the QPointF form of this 2D vector.

See also toPoint() and toVector3D().

fn to_vector3_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DToVector3DArgs<'largs>

C++ method: QVector2D::toVector3D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector3_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn to_vector3_d(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector2D::toVector3D() const

C++ documentation:

Returns the 3D form of this 2D vector, with the z coordinate set to zero.

See also toVector4D() and toPoint().

fn to_vector4_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector2DToVector4DArgs<'largs>

C++ method: QVector2D::toVector4D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector4_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn to_vector4_d(&self, cpp_utils::AsStruct) -> ::qvector4d::Vector4D
C++ method: QVector4D QVector2D::toVector4D() const

C++ documentation:

Returns the 4D form of this 2D vector, with the z and w coordinates set to zero.

See also toVector3D() and toPoint().

fn x(&self) -> c_float

C++ method: float QVector2D::x() const

C++ documentation:

Returns the x coordinate of this point.

See also setX() and y().

fn y(&self) -> c_float

C++ method: float QVector2D::y() const

C++ documentation:

Returns the y coordinate of this point.

See also setY() and x().