Module qt_gui::qvector3d::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

Vector3DAsQVariantArgs
This trait represents a set of arguments accepted by Vector3D::as_q_variant method.

Vector3DCrossProductArgs
This trait represents a set of arguments accepted by Vector3D::cross_product method.

Vector3DDistanceToPlaneArgs
This trait represents a set of arguments accepted by Vector3D::distance_to_plane method.

Vector3DNewArgs
This trait represents a set of arguments accepted by Vector3D::new method.

Vector3DNormalArgs
This trait represents a set of arguments accepted by Vector3D::normal method.

Vector3DNormalizedArgs
This trait represents a set of arguments accepted by Vector3D::normalized method.

Vector3DOpDivAssignArgs
This trait represents a set of arguments accepted by Vector3D::op_div_assign method.

Vector3DOpMulAssignArgs
This trait represents a set of arguments accepted by Vector3D::op_mul_assign method.

Vector3DProjectArgs
This trait represents a set of arguments accepted by Vector3D::project method.

Vector3DToPointArgs
This trait represents a set of arguments accepted by Vector3D::to_point method.

Vector3DToPointFArgs
This trait represents a set of arguments accepted by Vector3D::to_point_f method.

Vector3DToVector2DArgs
This trait represents a set of arguments accepted by Vector3D::to_vector2_d method.

Vector3DToVector4DArgs
This trait represents a set of arguments accepted by Vector3D::to_vector4_d method.

Vector3DUnprojectArgs
This trait represents a set of arguments accepted by Vector3D::unproject method.