Struct qt_gui::qvector3d::Vector3D

pub struct Vector3D {
    // some fields omitted
}

C++ type: QVector3D

C++ documentation:

The QVector3D class represents a vector or vertex in 3D space.

Vectors are one of the main building blocks of 3D representation and drawing. They consist of three coordinates, traditionally called x, y, and z.

The QVector3D class can also be used to represent vertices in 3D space. We therefore do not need to provide a separate vertex class.

Methods

impl Vector3D

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DAsQVariantArgs<'largs>

C++ method: QVector3D::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QVector3D::operator QVariant() const

C++ documentation:

Returns the 3D vector as a QVariant.

fn cross_product<Args>(args: Args) -> Args::ReturnType where Args: Vector3DCrossProductArgs

C++ method: QVector3D::crossProduct

This is an overloaded function. Available variants:

Rust arguments:
1) fn cross_product((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn cross_product((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: static QVector3D QVector3D::crossProduct(const QVector3D& v1, const QVector3D& v2)

C++ documentation:

Returns the cross-product of vectors v1 and v2, which corresponds to the normal vector of a plane defined by v1 and v2.

See also normal().

fn distance_to_line(&self, point: &Vector3D, direction: &Vector3D) -> c_float

C++ method: float QVector3D::distanceToLine(const QVector3D& point, const QVector3D& direction) const

C++ documentation:

Returns the distance that this vertex is from a line defined by point and the unit vector direction.

If direction is a null vector, then it does not define a line. In that case, the distance from point to this vertex is returned.

See also distanceToPlane().

fn distance_to_plane<'largs, Args>(&'largs self, args: Args) -> c_float where Args: Vector3DDistanceToPlaneArgs<'largs>

C++ method: QVector3D::distanceToPlane

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn distance_to_plane(&self, (&::qvector3d::Vector3D, &::qvector3d::Vector3D)) -> libc::c_float
C++ method: float QVector3D::distanceToPlane(const QVector3D& plane, const QVector3D& normal) const

C++ documentation:

Returns the distance from this vertex to a plane defined by the vertex plane and a normal unit vector. The normal parameter is assumed to have been normalized to a unit vector.

The return value will be negative if the vertex is below the plane, or zero if it is on the plane.

See also normal() and distanceToLine().

Variant 2

Rust arguments: fn distance_to_plane(&self, (&::qvector3d::Vector3D, &::qvector3d::Vector3D, &::qvector3d::Vector3D)) -> libc::c_float
C++ method: float QVector3D::distanceToPlane(const QVector3D& plane1, const QVector3D& plane2, const QVector3D& plane3) const

C++ documentation:

This is an overloaded function.

Returns the distance from this vertex a plane defined by the vertices plane1, plane2 and plane3.

The return value will be negative if the vertex is below the plane, or zero if it is on the plane.

The two vectors that define the plane are plane2 - plane1 and plane3 - plane1.

See also normal() and distanceToLine().

fn distance_to_point(&self, point: &Vector3D) -> c_float

C++ method: float QVector3D::distanceToPoint(const QVector3D& point) const

C++ documentation:

Returns the distance from this vertex to a point defined by the vertex point.

This function was introduced in Qt 5.1.

See also distanceToPlane() and distanceToLine().

fn dot_product(v1: &Vector3D, v2: &Vector3D) -> c_float

C++ method: static float QVector3D::dotProduct(const QVector3D& v1, const QVector3D& v2)

C++ documentation:

Returns the dot product of v1 and v2.

fn is_null(&self) -> bool

C++ method: bool QVector3D::isNull() const

C++ documentation:

Returns true if the x, y, and z coordinates are set to 0.0, otherwise returns false.

fn length(&self) -> c_float

C++ method: float QVector3D::length() const

C++ documentation:

Returns the length of the vector from the origin.

See also lengthSquared() and normalized().

fn length_squared(&self) -> c_float

C++ method: float QVector3D::lengthSquared() const

C++ documentation:

Returns the squared length of the vector from the origin. This is equivalent to the dot product of the vector with itself.

See also length() and dotProduct().

fn new<Args>(args: Args) -> Args::ReturnType where Args: Vector3DNewArgs

C++ method: QVector3D::QVector3D

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::qvector3d::Vector3D
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&qt_core::point::Point, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(const QPoint& point)

Variant 4

Rust arguments:
1) fn new((&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(const QPointF& point)

Variant 5

Rust arguments:
1) fn new((&::qvector2d::Vector2D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((&::qvector2d::Vector2D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(const QVector2D& vector)

Variant 6

Rust arguments:
1) fn new((&::qvector2d::Vector2D, libc::c_float, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((&::qvector2d::Vector2D, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(const QVector2D& vector, float zpos)

Variant 7

Rust arguments:
1) fn new((&::qvector4d::Vector4D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((&::qvector4d::Vector4D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(const QVector4D& vector)

Variant 8

Rust arguments:
1) fn new((libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
2) fn new((libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
C++ method: [constructor] void QVector3D::QVector3D(float xpos, float ypos, float zpos)

fn normal<Args>(args: Args) -> Args::ReturnType where Args: Vector3DNormalArgs

C++ method: QVector3D::normal

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn normal((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn normal((&::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: static QVector3D QVector3D::normal(const QVector3D& v1, const QVector3D& v2)

C++ documentation:

Returns the normal vector of a plane defined by vectors v1 and v2, normalized to be a unit vector.

Use crossProduct() to compute the cross-product of v1 and v2 if you do not need the result to be normalized to a unit vector.

See also crossProduct() and distanceToPlane().

Variant 2

Rust arguments:
1) fn normal((&::qvector3d::Vector3D, &::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn normal((&::qvector3d::Vector3D, &::qvector3d::Vector3D, &::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: static QVector3D QVector3D::normal(const QVector3D& v1, const QVector3D& v2, const QVector3D& v3)

C++ documentation:

This is an overloaded function.

Returns the normal vector of a plane defined by vectors v2 - v1 and v3 - v1, normalized to be a unit vector.

Use crossProduct() to compute the cross-product of v2 - v1 and v3 - v1 if you do not need the result to be normalized to a unit vector.

See also crossProduct() and distanceToPlane().

fn normalize(&mut self)

C++ method: void QVector3D::normalize()

C++ documentation:

Normalizes the currect vector in place. Nothing happens if this vector is a null vector or the length of the vector is very close to 1.

See also length() and normalized().

fn normalized<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DNormalizedArgs<'largs>

C++ method: QVector3D::normalized

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn normalized(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector3D::normalized() const

C++ documentation:

Returns the normalized unit vector form of this vector.

If this vector is null, then a null vector is returned. If the length of the vector is very close to 1, then the vector will be returned as-is. Otherwise the normalized form of the vector of length 1 will be returned.

See also length() and normalize().

fn op_add_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector3D) -> &'l0 mut Vector3D

C++ method: QVector3D& QVector3D::operator+=(const QVector3D& vector)

C++ documentation:

Adds the given vector to this vector and returns a reference to this vector.

See also operator-=().

fn op_div_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector3D where Args: Vector3DOpDivAssignArgs<'largs>

C++ method: QVector3D::operator/=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_div_assign(&mut self, &'l1 ::qvector3d::Vector3D) -> &'l0 mut ::qvector3d::Vector3D
C++ method: QVector3D& QVector3D::operator/=(const QVector3D& vector)

C++ documentation:

Divides the components of this vector by the corresponding components in vector.

This function was introduced in Qt 5.5.

See also operator*=().

Variant 2

Rust arguments: fn op_div_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector3d::Vector3D
C++ method: QVector3D& QVector3D::operator/=(float divisor)

C++ documentation:

Divides this vector's coordinates by the given divisor, and returns a reference to this vector.

See also operator*=().

fn op_index(&self, i: c_int) -> c_float

C++ method: float QVector3D::operator[](int i) const

C++ documentation:

Returns the component of the vector at index position i.

i must be a valid index position in the vector (i.e., 0 <= i < 3).

This function was introduced in Qt 5.2.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut c_float

C++ method: float& QVector3D::operator[](int i)

C++ documentation:

Returns the component of the vector at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < 3).

This function was introduced in Qt 5.2.

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector3D where Args: Vector3DOpMulAssignArgs<'largs>

C++ method: QVector3D::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::qvector3d::Vector3D) -> &'l0 mut ::qvector3d::Vector3D
C++ method: QVector3D& QVector3D::operator*=(const QVector3D& vector)

C++ documentation:

This is an overloaded function.

Multiplies the components of this vector by the corresponding components in vector.

Note: this is not the same as the crossProduct() of this vector and vector.

See also crossProduct().

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector3d::Vector3D
C++ method: QVector3D& QVector3D::operator*=(float factor)

C++ documentation:

Multiplies this vector's coordinates by the given factor, and returns a reference to this vector.

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector3D) -> &'l0 mut Vector3D

C++ method: QVector3D& QVector3D::operator-=(const QVector3D& vector)

C++ documentation:

Subtracts the given vector from this vector and returns a reference to this vector.

See also operator+=().

fn project<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DProjectArgs<'largs>

C++ method: QVector3D::project

This is an overloaded function. Available variants:

Rust arguments:
1) fn project(&self, (&::qmatrix4x4::Matrix4x4, &::qmatrix4x4::Matrix4x4, &qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn project(&self, (&::qmatrix4x4::Matrix4x4, &::qmatrix4x4::Matrix4x4, &qt_core::rect::Rect, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector3D::project(const QMatrix4x4& modelView, const QMatrix4x4& projection, const QRect& viewport) const

C++ documentation:

Returns the window coordinates of this vector initially in object/model coordinates using the model view matrix modelView, the projection matrix projection and the viewport dimensions viewport.

When transforming from clip to normalized space, a division by the w component on the vector components takes place. To prevent dividing by 0 if w equals to 0, it is set to 1.

Note: the returned y coordinates are in OpenGL orientation. OpenGL expects the bottom to be 0 whereas for Qt top is 0.

This function was introduced in Qt 5.5.

See also unproject().

fn set_x(&mut self, x: c_float)

C++ method: void QVector3D::setX(float x)

C++ documentation:

Sets the x coordinate of this point to the given x coordinate.

See also x(), setY(), and setZ().

fn set_y(&mut self, y: c_float)

C++ method: void QVector3D::setY(float y)

C++ documentation:

Sets the y coordinate of this point to the given y coordinate.

See also y(), setX(), and setZ().

fn set_z(&mut self, z: c_float)

C++ method: void QVector3D::setZ(float z)

C++ documentation:

Sets the z coordinate of this point to the given z coordinate.

See also z(), setX(), and setY().

fn to_point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DToPointArgs<'largs>

C++ method: QVector3D::toPoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn to_point(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QVector3D::toPoint() const

C++ documentation:

Returns the QPoint form of this 3D vector. The z coordinate is dropped.

See also toPointF() and toVector2D().

fn to_point_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DToPointFArgs<'largs>

C++ method: QVector3D::toPointF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn to_point_f(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QVector3D::toPointF() const

C++ documentation:

Returns the QPointF form of this 3D vector. The z coordinate is dropped.

See also toPoint() and toVector2D().

fn to_vector2_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DToVector2DArgs<'largs>

C++ method: QVector3D::toVector2D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector2_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
2) fn to_vector2_d(&self, cpp_utils::AsStruct) -> ::qvector2d::Vector2D
C++ method: QVector2D QVector3D::toVector2D() const

C++ documentation:

Returns the 2D vector form of this 3D vector, dropping the z coordinate.

See also toVector4D() and toPoint().

fn to_vector4_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DToVector4DArgs<'largs>

C++ method: QVector3D::toVector4D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector4_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn to_vector4_d(&self, cpp_utils::AsStruct) -> ::qvector4d::Vector4D
C++ method: QVector4D QVector3D::toVector4D() const

C++ documentation:

Returns the 4D form of this 3D vector, with the w coordinate set to zero.

See also toVector2D() and toPoint().

fn unproject<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector3DUnprojectArgs<'largs>

C++ method: QVector3D::unproject

This is an overloaded function. Available variants:

Rust arguments:
1) fn unproject(&self, (&::qmatrix4x4::Matrix4x4, &::qmatrix4x4::Matrix4x4, &qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn unproject(&self, (&::qmatrix4x4::Matrix4x4, &::qmatrix4x4::Matrix4x4, &qt_core::rect::Rect, cpp_utils::AsStruct)) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector3D::unproject(const QMatrix4x4& modelView, const QMatrix4x4& projection, const QRect& viewport) const

C++ documentation:

Returns the object/model coordinates of this vector initially in window coordinates using the model view matrix modelView, the projection matrix projection and the viewport dimensions viewport.

When transforming from clip to normalized space, a division by the w component of the vector components takes place. To prevent dividing by 0 if w equals to 0, it is set to 1.

Note: y coordinates in viewport should use OpenGL orientation. OpenGL expects the bottom to be 0 whereas for Qt top is 0.

This function was introduced in Qt 5.5.

See also project().

fn x(&self) -> c_float

C++ method: float QVector3D::x() const

C++ documentation:

Returns the x coordinate of this point.

See also setX(), y(), and z().

fn y(&self) -> c_float

C++ method: float QVector3D::y() const

C++ documentation:

Returns the y coordinate of this point.

See also setY(), x(), and z().

fn z(&self) -> c_float

C++ method: float QVector3D::z() const

C++ documentation:

Returns the z coordinate of this point.

See also setZ(), x(), and y().