Module qt_gui::qvector4d::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

Vector4DAsQVariantArgs
This trait represents a set of arguments accepted by Vector4D::as_q_variant method.

Vector4DNewArgs
This trait represents a set of arguments accepted by Vector4D::new method.

Vector4DNormalizedArgs
This trait represents a set of arguments accepted by Vector4D::normalized method.

Vector4DOpDivAssignArgs
This trait represents a set of arguments accepted by Vector4D::op_div_assign method.

Vector4DOpMulAssignArgs
This trait represents a set of arguments accepted by Vector4D::op_mul_assign method.

Vector4DToPointArgs
This trait represents a set of arguments accepted by Vector4D::to_point method.

Vector4DToPointFArgs
This trait represents a set of arguments accepted by Vector4D::to_point_f method.

Vector4DToVector2DAffineArgs
This trait represents a set of arguments accepted by Vector4D::to_vector2_d_affine method.

Vector4DToVector2DArgs
This trait represents a set of arguments accepted by Vector4D::to_vector2_d method.

Vector4DToVector3DAffineArgs
This trait represents a set of arguments accepted by Vector4D::to_vector3_d_affine method.

Vector4DToVector3DArgs
This trait represents a set of arguments accepted by Vector4D::to_vector3_d method.