Struct qt_gui::qvector4d::Vector4D

pub struct Vector4D {
    // some fields omitted
}

C++ type: QVector4D

C++ documentation:

The QVector4D class represents a vector or vertex in 4D space.

The QVector4D class can also be used to represent vertices in 4D space. We therefore do not need to provide a separate vertex class.

Methods

impl Vector4D

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DAsQVariantArgs<'largs>

C++ method: QVector4D::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QVector4D::operator QVariant() const

C++ documentation:

Returns the 4D vector as a QVariant.

fn dot_product(v1: &Vector4D, v2: &Vector4D) -> c_float

C++ method: static float QVector4D::dotProduct(const QVector4D& v1, const QVector4D& v2)

C++ documentation:

Returns the dot product of v1 and v2.

fn is_null(&self) -> bool

C++ method: bool QVector4D::isNull() const

C++ documentation:

Returns true if the x, y, z, and w coordinates are set to 0.0, otherwise returns false.

fn length(&self) -> c_float

C++ method: float QVector4D::length() const

C++ documentation:

Returns the length of the vector from the origin.

See also lengthSquared() and normalized().

fn length_squared(&self) -> c_float

C++ method: float QVector4D::lengthSquared() const

C++ documentation:

Returns the squared length of the vector from the origin. This is equivalent to the dot product of the vector with itself.

See also length() and dotProduct().

fn new<Args>(args: Args) -> Args::ReturnType where Args: Vector4DNewArgs

C++ method: QVector4D::QVector4D

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::qvector4d::Vector4D
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&qt_core::point::Point, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QPoint& point)

Variant 4

Rust arguments:
1) fn new((&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QPointF& point)

Variant 5

Rust arguments:
1) fn new((&::qvector2d::Vector2D, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&::qvector2d::Vector2D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QVector2D& vector)

Variant 6

Rust arguments:
1) fn new((&::qvector2d::Vector2D, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&::qvector2d::Vector2D, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QVector2D& vector, float zpos, float wpos)

Variant 7

Rust arguments:
1) fn new((&::qvector3d::Vector3D, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&::qvector3d::Vector3D, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QVector3D& vector)

Variant 8

Rust arguments:
1) fn new((&::qvector3d::Vector3D, libc::c_float, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((&::qvector3d::Vector3D, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(const QVector3D& vector, float wpos)

Variant 9

Rust arguments:
1) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsStruct)) -> ::qvector4d::Vector4D
2) fn new((libc::c_float, libc::c_float, libc::c_float, libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::qvector4d::Vector4D>
C++ method: [constructor] void QVector4D::QVector4D(float xpos, float ypos, float zpos, float wpos)

fn normalize(&mut self)

C++ method: void QVector4D::normalize()

C++ documentation:

Normalizes the currect vector in place. Nothing happens if this vector is a null vector or the length of the vector is very close to 1.

See also length() and normalized().

fn normalized<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DNormalizedArgs<'largs>

C++ method: QVector4D::normalized

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector4d::Vector4D>
2) fn normalized(&self, cpp_utils::AsStruct) -> ::qvector4d::Vector4D
C++ method: QVector4D QVector4D::normalized() const

C++ documentation:

Returns the normalized unit vector form of this vector.

If this vector is null, then a null vector is returned. If the length of the vector is very close to 1, then the vector will be returned as-is. Otherwise the normalized form of the vector of length 1 will be returned.

See also length() and normalize().

fn op_add_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector4D) -> &'l0 mut Vector4D

C++ method: QVector4D& QVector4D::operator+=(const QVector4D& vector)

C++ documentation:

Adds the given vector to this vector and returns a reference to this vector.

See also operator-=().

fn op_div_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector4D where Args: Vector4DOpDivAssignArgs<'largs>

C++ method: QVector4D::operator/=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_div_assign(&mut self, &'l1 ::qvector4d::Vector4D) -> &'l0 mut ::qvector4d::Vector4D
C++ method: QVector4D& QVector4D::operator/=(const QVector4D& vector)

C++ documentation:

Divides the components of this vector by the corresponding components in vector.

This function was introduced in Qt 5.5.

See also operator*=().

Variant 2

Rust arguments: fn op_div_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector4d::Vector4D
C++ method: QVector4D& QVector4D::operator/=(float divisor)

C++ documentation:

Divides this vector's coordinates by the given divisor, and returns a reference to this vector.

See also operator*=().

fn op_index(&self, i: c_int) -> c_float

C++ method: float QVector4D::operator[](int i) const

C++ documentation:

Returns the component of the vector at index position i.

i must be a valid index position in the vector (i.e., 0 <= i < 4).

This function was introduced in Qt 5.2.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut c_float

C++ method: float& QVector4D::operator[](int i)

C++ documentation:

Returns the component of the vector at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < 4).

This function was introduced in Qt 5.2.

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Vector4D where Args: Vector4DOpMulAssignArgs<'largs>

C++ method: QVector4D::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::qvector4d::Vector4D) -> &'l0 mut ::qvector4d::Vector4D
C++ method: QVector4D& QVector4D::operator*=(const QVector4D& vector)

C++ documentation:

Multiplies the components of this vector by the corresponding components in vector.

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::qvector4d::Vector4D
C++ method: QVector4D& QVector4D::operator*=(float factor)

C++ documentation:

Multiplies this vector's coordinates by the given factor, and returns a reference to this vector.

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, vector: &'l1 Vector4D) -> &'l0 mut Vector4D

C++ method: QVector4D& QVector4D::operator-=(const QVector4D& vector)

C++ documentation:

Subtracts the given vector from this vector and returns a reference to this vector.

See also operator+=().

fn set_w(&mut self, w: c_float)

C++ method: void QVector4D::setW(float w)

C++ documentation:

Sets the w coordinate of this point to the given w coordinate.

See also w(), setX(), setY(), and setZ().

fn set_x(&mut self, x: c_float)

C++ method: void QVector4D::setX(float x)

C++ documentation:

Sets the x coordinate of this point to the given x coordinate.

See also x(), setY(), setZ(), and setW().

fn set_y(&mut self, y: c_float)

C++ method: void QVector4D::setY(float y)

C++ documentation:

Sets the y coordinate of this point to the given y coordinate.

See also y(), setX(), setZ(), and setW().

fn set_z(&mut self, z: c_float)

C++ method: void QVector4D::setZ(float z)

C++ documentation:

Sets the z coordinate of this point to the given z coordinate.

See also z(), setX(), setY(), and setW().

fn to_point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToPointArgs<'largs>

C++ method: QVector4D::toPoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn to_point(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QVector4D::toPoint() const

C++ documentation:

Returns the QPoint form of this 4D vector. The z and w coordinates are dropped.

See also toPointF() and toVector2D().

fn to_point_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToPointFArgs<'largs>

C++ method: QVector4D::toPointF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn to_point_f(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QVector4D::toPointF() const

C++ documentation:

Returns the QPointF form of this 4D vector. The z and w coordinates are dropped.

See also toPoint() and toVector2D().

fn to_vector2_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToVector2DArgs<'largs>

C++ method: QVector4D::toVector2D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector2_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
2) fn to_vector2_d(&self, cpp_utils::AsStruct) -> ::qvector2d::Vector2D
C++ method: QVector2D QVector4D::toVector2D() const

C++ documentation:

Returns the 2D vector form of this 4D vector, dropping the z and w coordinates.

See also toVector2DAffine(), toVector3D(), and toPoint().

fn to_vector2_d_affine<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToVector2DAffineArgs<'largs>

C++ method: QVector4D::toVector2DAffine

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector2_d_affine(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
2) fn to_vector2_d_affine(&self, cpp_utils::AsStruct) -> ::qvector2d::Vector2D
C++ method: QVector2D QVector4D::toVector2DAffine() const

C++ documentation:

Returns the 2D vector form of this 4D vector, dividing the x and y coordinates by the w coordinate and dropping the z coordinate. Returns a null vector if w is zero.

See also toVector2D(), toVector3DAffine(), and toPoint().

fn to_vector3_d<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToVector3DArgs<'largs>

C++ method: QVector4D::toVector3D

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector3_d(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn to_vector3_d(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector4D::toVector3D() const

C++ documentation:

Returns the 3D vector form of this 4D vector, dropping the w coordinate.

See also toVector3DAffine(), toVector2D(), and toPoint().

fn to_vector3_d_affine<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: Vector4DToVector3DAffineArgs<'largs>

C++ method: QVector4D::toVector3DAffine

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_vector3_d_affine(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector3d::Vector3D>
2) fn to_vector3_d_affine(&self, cpp_utils::AsStruct) -> ::qvector3d::Vector3D
C++ method: QVector3D QVector4D::toVector3DAffine() const

C++ documentation:

Returns the 3D vector form of this 4D vector, dividing the x, y, and z coordinates by the w coordinate. Returns a null vector if w is zero.

See also toVector3D(), toVector2DAffine(), and toPoint().

fn w(&self) -> c_float

C++ method: float QVector4D::w() const

C++ documentation:

Returns the w coordinate of this point.

See also setW(), x(), y(), and z().

fn x(&self) -> c_float

C++ method: float QVector4D::x() const

C++ documentation:

Returns the x coordinate of this point.

See also setX(), y(), z(), and w().

fn y(&self) -> c_float

C++ method: float QVector4D::y() const

C++ documentation:

Returns the y coordinate of this point.

See also setY(), x(), z(), and w().

fn z(&self) -> c_float

C++ method: float QVector4D::z() const

C++ documentation:

Returns the z coordinate of this point.

See also setZ(), x(), y(), and w().