Module qt_gui::raw_font::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

RawFontAdvancesForGlyphIndexesArgs
This trait represents a set of arguments accepted by RawFont::advances_for_glyph_indexes method.

RawFontAlphaMapForGlyphArgs
This trait represents a set of arguments accepted by RawFont::alpha_map_for_glyph method.

RawFontBoundingRectArgs
This trait represents a set of arguments accepted by RawFont::bounding_rect method.

RawFontFamilyNameArgs
This trait represents a set of arguments accepted by RawFont::family_name method.

RawFontFontTableArgs
This trait represents a set of arguments accepted by RawFont::font_table method.

RawFontFromFontArgs
This trait represents a set of arguments accepted by RawFont::from_font method.

RawFontGlyphIndexesForStringArgs
This trait represents a set of arguments accepted by RawFont::glyph_indexes_for_string method.

RawFontNewArgs
This trait represents a set of arguments accepted by RawFont::new method.

RawFontPathForGlyphArgs
This trait represents a set of arguments accepted by RawFont::path_for_glyph method.

RawFontStyleNameArgs
This trait represents a set of arguments accepted by RawFont::style_name method.

RawFontSupportedWritingSystemsArgs
This trait represents a set of arguments accepted by RawFont::supported_writing_systems method.

RawFontSupportsCharacterArgs
This trait represents a set of arguments accepted by RawFont::supports_character method.