Struct qt_gui::raw_font::RawFont

pub struct RawFont {
    // some fields omitted
}

C++ type: QRawFont

C++ documentation:

The QRawFont class provides access to a single physical instance of a font.

Note: QRawFont is a low level class. For most purposes QFont is a more appropriate class.

Most commonly, when presenting text in a user interface, the exact fonts used to render the characters is to some extent unknown. This can be the case for several reasons: For instance, the actual, physical fonts present on the target system could be unexpected to the developers, or the text could contain user selected styles, sizes or writing systems that are not supported by font chosen in the code.

Therefore, Qt's QFont class really represents a query for fonts. When text is interpreted, Qt will do its best to match the text to the query, but depending on the support, different fonts can be used behind the scenes.

For most use cases, this is both expected and necessary, as it minimizes the possibility of text in the user interface being undisplayable. In some cases, however, more direct control over the process might be useful. It is for these use cases the QRawFont class exists.

A QRawFont object represents a single, physical instance of a given font in a given pixel size. I.e. in the typical case it represents a set of TrueType or OpenType font tables and uses a user specified pixel size to convert metrics into logical pixel units. It can be used in combination with the QGlyphRun class to draw specific glyph indexes at specific positions, and also have accessors to some relevant data in the physical font.

QRawFont only provides support for the main font technologies: GDI and DirectWrite on Windows platforms, FreeType on Linux platforms and CoreText on Mac OS X. For other font back-ends, the APIs will be disabled.

QRawFont can be constructed in a number of ways:

It can be constructed by calling QTextLayout::glyphs() or QTextFragment::glyphs(). The returned QGlyphs objects will contain QRawFont objects which represent the actual fonts used to render each portion of the text.

It can be constructed by passing a QFont object to QRawFont::fromFont(). The function will return a QRawFont object representing the font that will be selected as response to the QFont query and the selected writing system.

It can be constructed by passing a file name or QByteArray directly to the QRawFont constructor, or by calling loadFromFile() or loadFromData(). In this case, the font will not be registered in QFontDatabase, and it will not be available as part of regular font selection.

QRawFont is considered local to the thread in which it is constructed (either using a constructor, or by calling loadFromData() or loadFromFile()). The QRawFont cannot be moved to a different thread, but will have to be recreated in the thread in question.

Note: For the requirement of caching glyph indexes and font selections for static text to avoid reshaping and relayouting in the inner loop of an application, a better choice is the QStaticText class, since it optimizes the memory cost of the cache and also provides the possibility of paint engine specific caches for an additional speed-up.

Methods

impl RawFont

fn advances_for_glyph_indexes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontAdvancesForGlyphIndexesArgs<'largs>

C++ method: QRawFont::advancesForGlyphIndexes

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn advances_for_glyph_indexes(&self, (&qt_core::vector::VectorCUint, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn advances_for_glyph_indexes(&self, (&qt_core::vector::VectorCUint, cpp_utils::AsStruct)) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QRawFont::advancesForGlyphIndexes(const QVector<unsigned int>& glyphIndexes) const

C++ documentation:

This is an overloaded function.

Returns the QRawFont's advances for each of the glyphIndexes in pixel units. The advances give the distance from the position of a given glyph to where the next glyph should be drawn to make it appear as if the two glyphs are unspaced. The advance of each glyph is calculated separately.

See also QTextLine::horizontalAdvance() and QFontMetricsF::width().

Variant 2

Rust arguments:
1) fn advances_for_glyph_indexes(&self, (&qt_core::vector::VectorCUint, qt_core::flags::Flags<::raw_font::LayoutFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn advances_for_glyph_indexes(&self, (&qt_core::vector::VectorCUint, qt_core::flags::Flags<::raw_font::LayoutFlag>, cpp_utils::AsStruct)) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QRawFont::advancesForGlyphIndexes(const QVector<unsigned int>& glyphIndexes, QFlags<QRawFont::LayoutFlag> layoutFlags) const

C++ documentation:

Returns the QRawFont's advances for each of the glyphIndexes in pixel units. The advances give the distance from the position of a given glyph to where the next glyph should be drawn to make it appear as if the two glyphs are unspaced. How the advances are calculated is controlled by layoutFlags.

This function was introduced in Qt 5.1.

See also QTextLine::horizontalAdvance() and QFontMetricsF::width().

Variant 3

Rust arguments: fn advances_for_glyph_indexes(&self, (*const libc::c_uint, *mut qt_core::point_f::PointF, libc::c_int)) -> bool
C++ method: bool QRawFont::advancesForGlyphIndexes(const unsigned int* glyphIndexes, QPointF* advances, int numGlyphs) const

C++ documentation:

This is an overloaded function.

Returns the QRawFont's advances for each of the glyphIndexes in pixel units. The advances give the distance from the position of a given glyph to where the next glyph should be drawn to make it appear as if the two glyphs are unspaced. The glyph indexes are given with the array glyphIndexes while the results are returned through advances, both of them must have numGlyphs elements. The advance of each glyph is calculated separately

See also QTextLine::horizontalAdvance() and QFontMetricsF::width().

Variant 4

Rust arguments: fn advances_for_glyph_indexes(&self, (*const libc::c_uint, *mut qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<::raw_font::LayoutFlag>)) -> bool
C++ method: bool QRawFont::advancesForGlyphIndexes(const unsigned int* glyphIndexes, QPointF* advances, int numGlyphs, QFlags<QRawFont::LayoutFlag> layoutFlags) const

C++ documentation:

Returns the QRawFont's advances for each of the glyphIndexes in pixel units. The advances give the distance from the position of a given glyph to where the next glyph should be drawn to make it appear as if the two glyphs are unspaced. The glyph indexes are given with the array glyphIndexes while the results are returned through advances, both of them must have numGlyphs elements. How the advances are calculated is controlled by layoutFlags.

This function was introduced in Qt 5.1.

See also QTextLine::horizontalAdvance() and QFontMetricsF::width().

fn alpha_map_for_glyph<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontAlphaMapForGlyphArgs<'largs>

C++ method: QRawFont::alphaMapForGlyph

This is an overloaded function. Available variants:

Rust arguments:
1) fn alpha_map_for_glyph(&self, (u32, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image::Image>
2) fn alpha_map_for_glyph(&self, (u32, ::raw_font::AntialiasingType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image::Image>
3) fn alpha_map_for_glyph(&self, (u32, ::raw_font::AntialiasingType, &::transform::Transform, cpp_utils::AsBox)) -> cpp_utils::CppBox<::image::Image>
4) fn alpha_map_for_glyph(&self, (u32, cpp_utils::AsStruct)) -> ::image::Image
5) fn alpha_map_for_glyph(&self, (u32, ::raw_font::AntialiasingType, cpp_utils::AsStruct)) -> ::image::Image
6) fn alpha_map_for_glyph(&self, (u32, ::raw_font::AntialiasingType, &::transform::Transform, cpp_utils::AsStruct)) -> ::image::Image
C++ method: QImage QRawFont::alphaMapForGlyph(quint32 glyphIndex, QRawFont::AntialiasingType antialiasingType = ?, const QTransform& transform = ?) const

C++ documentation:

This function returns a rasterized image of the glyph at the given glyphIndex in the underlying font, using the transform specified. If the QRawFont is not valid, this function will return an invalid QImage.

If antialiasingType is set to QRawFont::SubPixelAntialiasing, then the resulting image will be in QImage::Format_RGB32 and the RGB values of each pixel will represent the subpixel opacities of the pixel in the rasterization of the glyph. Otherwise, the image will be in the format of QImage::Format_Indexed8 and each pixel will contain the opacity of the pixel in the rasterization.

See also pathForGlyph() and QPainter::drawGlyphRun().

fn ascent(&self) -> c_double

C++ method: double QRawFont::ascent() const

C++ documentation:

Returns the ascent of this QRawFont in pixel units.

See also QFontMetricsF::ascent().

fn average_char_width(&self) -> c_double

C++ method: double QRawFont::averageCharWidth() const

C++ documentation:

Returns the average character width of this QRawFont in pixel units.

See also QFontMetricsF::averageCharWidth().

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontBoundingRectArgs<'largs>

C++ method: QRawFont::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, (u32, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn bounding_rect(&self, (u32, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QRawFont::boundingRect(quint32 glyphIndex) const

C++ documentation:

Returns the smallest rectangle containing the glyph with the given glyphIndex.

This function was introduced in Qt 5.0.

fn descent(&self) -> c_double

C++ method: double QRawFont::descent() const

C++ documentation:

Returns the descent of this QRawFont in pixel units.

See also QFontMetricsF::descent().

fn family_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontFamilyNameArgs<'largs>

C++ method: QRawFont::familyName

This is an overloaded function. Available variants:

Rust arguments:
1) fn family_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn family_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QRawFont::familyName() const

C++ documentation:

Returns the family name of this QRawFont.

fn font_table<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontFontTableArgs<'largs>

C++ method: QRawFont::fontTable

This is an overloaded function. Available variants:

Rust arguments:
1) fn font_table(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn font_table(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QRawFont::fontTable(const char* tagName) const

C++ documentation:

Retrieves the sfnt table named tagName from the underlying physical font, or an empty byte array if no such table was found. The returned font table's byte order is Big Endian, like the sfnt format specifies. The tagName must be four characters long and should be formatted in the default endianness of the current platform.

fn from_font<Args>(args: Args) -> Args::ReturnType where Args: RawFontFromFontArgs

C++ method: QRawFont::fromFont

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_font((&::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
2) fn from_font((&::font::Font, ::font_database::WritingSystem, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
3) fn from_font((&::font::Font, cpp_utils::AsStruct)) -> ::raw_font::RawFont
4) fn from_font((&::font::Font, ::font_database::WritingSystem, cpp_utils::AsStruct)) -> ::raw_font::RawFont
C++ method: static QRawFont QRawFont::fromFont(const QFont& font, QFontDatabase::WritingSystem writingSystem = ?)

C++ documentation:

Fetches the physical representation based on a font query. The physical font returned is the font that will be preferred by Qt in order to display text in the selected writingSystem.

Warning: This function is potentially expensive and should not be called in performance sensitive code.

fn glyph_indexes_for_chars(&self, chars: *const Char, num_chars: c_int, glyph_indexes: *mut c_uint, num_glyphs: *mut c_int) -> bool

C++ method: bool QRawFont::glyphIndexesForChars(const QChar* chars, int numChars, unsigned int* glyphIndexes, int* numGlyphs) const

C++ documentation:

Converts a string of unicode points to glyph indexes using the CMAP table in the underlying font. The function works like glyphIndexesForString() except it take an array (chars), the results will be returned though glyphIndexes array and number of glyphs will be set in numGlyphs. The size of glyphIndexes array must be at least numChars, if that's still not enough, this function will return false, then you can resize glyphIndexes from the size returned in numGlyphs.

See also glyphIndexesForString(), advancesForGlyphIndexes(), QGlyphRun, QTextLayout::glyphRuns(), and QTextFragment::glyphRuns().

fn glyph_indexes_for_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontGlyphIndexesForStringArgs<'largs>

C++ method: QRawFont::glyphIndexesForString

This is an overloaded function. Available variants:

Rust arguments:
1) fn glyph_indexes_for_string(&self, (&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::vector::VectorCUint>
2) fn glyph_indexes_for_string(&self, (&qt_core::string::String, cpp_utils::AsStruct)) -> qt_core::vector::VectorCUint
C++ method: QVector<unsigned int> QRawFont::glyphIndexesForString(const QString& text) const

C++ documentation:

Converts the string of unicode points given by text to glyph indexes using the CMAP table in the underlying font, and returns a vector containing the result.

Note that, in cases where there are other tables in the font that affect the shaping of the text, the returned glyph indexes will not correctly represent the rendering of the text. To get the correctly shaped text, you can use QTextLayout to lay out and shape the text, then call QTextLayout::glyphs() to get the set of glyph index list and QRawFont pairs.

See also advancesForGlyphIndexes(), glyphIndexesForChars(), QGlyphRun, QTextLayout::glyphRuns(), and QTextFragment::glyphRuns().

fn hinting_preference(&self) -> HintingPreference

C++ method: QFont::HintingPreference QRawFont::hintingPreference() const

C++ documentation:

Returns the hinting preference used to construct this QRawFont.

See also QFont::hintingPreference().

fn is_valid(&self) -> bool

C++ method: bool QRawFont::isValid() const

C++ documentation:

Returns true if the QRawFont is valid and false otherwise.

fn leading(&self) -> c_double

C++ method: double QRawFont::leading() const

C++ documentation:

Returns the leading of this QRawFont in pixel units.

See also QFontMetricsF::leading().

fn line_thickness(&self) -> c_double

C++ method: double QRawFont::lineThickness() const

C++ documentation:

Returns the thickness for drawing lines (underline, overline, etc.) along with text drawn in this font.

fn load_from_data(&mut self, font_data: &ByteArray, pixel_size: c_double, hinting_preference: HintingPreference)

C++ method: void QRawFont::loadFromData(const QByteArray& fontData, double pixelSize, QFont::HintingPreference hintingPreference)

C++ documentation:

Replaces the current QRawFont with the font contained in the supplied fontData for the size (in pixels) given by pixelSize, and using the hinting preference specified by hintingPreference.

The fontData must contain a TrueType or OpenType font.

See also loadFromFile().

fn load_from_file(&mut self, file_name: &String, pixel_size: c_double, hinting_preference: HintingPreference)

C++ method: void QRawFont::loadFromFile(const QString& fileName, double pixelSize, QFont::HintingPreference hintingPreference)

C++ documentation:

Replaces the current QRawFont with the contents of the file referenced by fileName for the size (in pixels) given by pixelSize, and using the hinting preference specified by hintingPreference.

The file must reference a TrueType or OpenType font.

See also loadFromData().

fn max_char_width(&self) -> c_double

C++ method: double QRawFont::maxCharWidth() const

C++ documentation:

Returns the width of the widest character in the font.

See also QFontMetricsF::maxWidth().

fn new<Args>(args: Args) -> Args::ReturnType where Args: RawFontNewArgs

C++ method: QRawFont::QRawFont

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::raw_font::RawFont
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::raw_font::RawFont>
C++ method: [constructor] void QRawFont::QRawFont()

Variant 2

Rust arguments:
1) fn new((&qt_core::byte_array::ByteArray, libc::c_double, cpp_utils::AsStruct)) -> ::raw_font::RawFont
2) fn new((&qt_core::byte_array::ByteArray, libc::c_double, ::font::HintingPreference, cpp_utils::AsStruct)) -> ::raw_font::RawFont
3) fn new((&qt_core::byte_array::ByteArray, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
4) fn new((&qt_core::byte_array::ByteArray, libc::c_double, ::font::HintingPreference, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
C++ method: [constructor] void QRawFont::QRawFont(const QByteArray& fontData, double pixelSize, QFont::HintingPreference hintingPreference = ?)

Variant 3

Rust arguments:
1) fn new((&::raw_font::RawFont, cpp_utils::AsStruct)) -> ::raw_font::RawFont
2) fn new((&::raw_font::RawFont, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
C++ method: [constructor] void QRawFont::QRawFont(const QRawFont& other)

Variant 4

Rust arguments:
1) fn new((&qt_core::string::String, libc::c_double, cpp_utils::AsStruct)) -> ::raw_font::RawFont
2) fn new((&qt_core::string::String, libc::c_double, ::font::HintingPreference, cpp_utils::AsStruct)) -> ::raw_font::RawFont
3) fn new((&qt_core::string::String, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
4) fn new((&qt_core::string::String, libc::c_double, ::font::HintingPreference, cpp_utils::AsBox)) -> cpp_utils::CppBox<::raw_font::RawFont>
C++ method: [constructor] void QRawFont::QRawFont(const QString& fileName, double pixelSize, QFont::HintingPreference hintingPreference = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 RawFont) -> &'l0 mut RawFont

C++ method: QRawFont& QRawFont::operator=(const QRawFont& other)

C++ documentation:

Assigns other to this QRawFont.

fn op_eq(&self, other: &RawFont) -> bool

C++ method: bool QRawFont::operator==(const QRawFont& other) const

C++ documentation:

Returns true if this QRawFont is equal to other. Otherwise, returns false.

fn op_neq(&self, other: &RawFont) -> bool

C++ method: bool QRawFont::operator!=(const QRawFont& other) const

C++ documentation:

Returns true if this QRawFont is not equal to other. Otherwise, returns false.

fn path_for_glyph<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontPathForGlyphArgs<'largs>

C++ method: QRawFont::pathForGlyph

This is an overloaded function. Available variants:

Rust arguments:
1) fn path_for_glyph(&self, (u32, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter_path::PainterPath>
2) fn path_for_glyph(&self, (u32, cpp_utils::AsStruct)) -> ::painter_path::PainterPath
C++ method: QPainterPath QRawFont::pathForGlyph(quint32 glyphIndex) const

C++ documentation:

This function returns the shape of the glyph at a given glyphIndex in the underlying font if the QRawFont is valid. Otherwise, it returns an empty QPainterPath.

The returned glyph will always be unhinted.

See also alphaMapForGlyph() and QPainterPath::addText().

fn pixel_size(&self) -> c_double

C++ method: double QRawFont::pixelSize() const

C++ documentation:

Returns the pixel size set for this QRawFont. The pixel size affects how glyphs are rasterized, the size of glyphs returned by pathForGlyph(), and is used to convert internal metrics from design units to logical pixel units.

See also setPixelSize().

fn set_pixel_size(&mut self, pixel_size: c_double)

C++ method: void QRawFont::setPixelSize(double pixelSize)

C++ documentation:

Sets the pixel size with which this font should be rendered to pixelSize.

See also pixelSize().

fn style(&self) -> Style

C++ method: QFont::Style QRawFont::style() const

C++ documentation:

Returns the style of this QRawFont.

See also QFont::style().

fn style_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontStyleNameArgs<'largs>

C++ method: QRawFont::styleName

This is an overloaded function. Available variants:

Rust arguments:
1) fn style_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn style_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QRawFont::styleName() const

C++ documentation:

Returns the style name of this QRawFont.

See also QFont::styleName().

fn supported_writing_systems<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RawFontSupportedWritingSystemsArgs<'largs>

C++ method: QRawFont::supportedWritingSystems

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_writing_systems(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListWritingSystem>
2) fn supported_writing_systems(&self, cpp_utils::AsStruct) -> ::list::ListWritingSystem
C++ method: QList<QFontDatabase::WritingSystem> QRawFont::supportedWritingSystems() const

C++ documentation:

Returns a list of writing systems supported by the font according to designer supplied information in the font file. Please note that this does not guarantee support for a specific unicode point in the font. You can use the supportsCharacter() to check support for a single, specific character.

Note: The list is determined based on the unicode ranges and codepage ranges set in the font's OS/2 table and requires such a table to be present in the underlying font file.

See also supportsCharacter().

fn supports_character<'largs, Args>(&'largs self, args: Args) -> bool where Args: RawFontSupportsCharacterArgs<'largs>

C++ method: QRawFont::supportsCharacter

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn supports_character(&self, &qt_core::char::Char) -> bool
C++ method: bool QRawFont::supportsCharacter(QChar character) const

C++ documentation:

Returns true if the font has a glyph that corresponds to the given character.

See also supportedWritingSystems().

Variant 2

Rust arguments: fn supports_character(&self, libc::c_uint) -> bool
C++ method: bool QRawFont::supportsCharacter(unsigned int ucs4) const

C++ documentation:

This is an overloaded function.

Returns true if the font has a glyph that corresponds to the UCS-4 encoded character ucs4.

See also supportedWritingSystems().

fn swap(&mut self, other: &mut RawFont)

C++ method: void QRawFont::swap(QRawFont& other)

C++ documentation:

Swaps this raw font with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn underline_position(&self) -> c_double

C++ method: double QRawFont::underlinePosition() const

C++ documentation:

Returns the position from baseline for drawing underlines below the text rendered with this font.

fn units_per_em(&self) -> c_double

C++ method: double QRawFont::unitsPerEm() const

C++ documentation:

Returns the number of design units define the width and height of the em square for this QRawFont. This value is used together with the pixel size when converting design metrics to pixel units, as the internal metrics are specified in design units and the pixel size gives the size of 1 em in pixels.

See also pixelSize() and setPixelSize().

fn weight(&self) -> c_int

C++ method: int QRawFont::weight() const

C++ documentation:

Returns the weight of this QRawFont.

See also QFont::weight().

fn x_height(&self) -> c_double

C++ method: double QRawFont::xHeight() const

C++ documentation:

Returns the xHeight of this QRawFont in pixel units.

See also QFontMetricsF::xHeight().