Module qt_gui::reg_exp_validator::overloading

Traits

RegExpValidatorConnectArgs
This trait represents a set of arguments accepted by RegExpValidator::connect method.

RegExpValidatorConnectStaticArgs
This trait represents a set of arguments accepted by RegExpValidator::connect_static method.

RegExpValidatorDestroyedArgs
This trait represents a set of arguments accepted by RegExpValidator::destroyed method.

RegExpValidatorDisconnectArgs
This trait represents a set of arguments accepted by RegExpValidator::disconnect method.

RegExpValidatorDisconnectStaticArgs
This trait represents a set of arguments accepted by RegExpValidator::disconnect_static method.

RegExpValidatorDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by RegExpValidator::dynamic_property_names method.

RegExpValidatorLocaleArgs
This trait represents a set of arguments accepted by RegExpValidator::locale method.

RegExpValidatorNewArgs
This trait represents a set of arguments accepted by RegExpValidator::new method.

RegExpValidatorObjectNameArgs
This trait represents a set of arguments accepted by RegExpValidator::object_name method.

RegExpValidatorPropertyArgs
This trait represents a set of arguments accepted by RegExpValidator::property method.

RegExpValidatorStartTimerArgs
This trait represents a set of arguments accepted by RegExpValidator::start_timer method.

RegExpValidatorTrArgs
This trait represents a set of arguments accepted by RegExpValidator::tr method.

RegExpValidatorTrUtf8Args
This trait represents a set of arguments accepted by RegExpValidator::tr_utf8 method.