Module qt_gui::region::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

RegionAsQVariantArgs
This trait represents a set of arguments accepted by Region::as_q_variant method.

RegionBoundingRectArgs
This trait represents a set of arguments accepted by Region::bounding_rect method.

RegionContainsArgs
This trait represents a set of arguments accepted by Region::contains method.

RegionIntersectedArgs
This trait represents a set of arguments accepted by Region::intersected method.

RegionIntersectsArgs
This trait represents a set of arguments accepted by Region::intersects method.

RegionNewArgs
This trait represents a set of arguments accepted by Region::new method.

RegionOpAddArgs
This trait represents a set of arguments accepted by Region::op_add method.

RegionOpAddAssignArgs
This trait represents a set of arguments accepted by Region::op_add_assign method.

RegionOpBitAndArgs
This trait represents a set of arguments accepted by Region::op_bit_and method.

RegionOpBitAndAssignArgs
This trait represents a set of arguments accepted by Region::op_bit_and_assign method.

RegionOpBitOrArgs
This trait represents a set of arguments accepted by Region::op_bit_or method.

RegionOpBitXorArgs
This trait represents a set of arguments accepted by Region::op_bit_xor method.

RegionOpSubArgs
This trait represents a set of arguments accepted by Region::op_sub method.

RegionRectsArgs
This trait represents a set of arguments accepted by Region::rects method.

RegionSubtractedArgs
This trait represents a set of arguments accepted by Region::subtracted method.

RegionTranslateArgs
This trait represents a set of arguments accepted by Region::translate method.

RegionTranslatedArgs
This trait represents a set of arguments accepted by Region::translated method.

RegionUnitedArgs
This trait represents a set of arguments accepted by Region::united method.

RegionXoredArgs
This trait represents a set of arguments accepted by Region::xored method.