Module qt_gui::screen::overloading

Traits

ScreenAvailableGeometryArgs
This trait represents a set of arguments accepted by Screen::available_geometry method.

ScreenAvailableSizeArgs
This trait represents a set of arguments accepted by Screen::available_size method.

ScreenAvailableVirtualGeometryArgs
This trait represents a set of arguments accepted by Screen::available_virtual_geometry method.

ScreenAvailableVirtualSizeArgs
This trait represents a set of arguments accepted by Screen::available_virtual_size method.

ScreenConnectArgs
This trait represents a set of arguments accepted by Screen::connect method.

ScreenConnectStaticArgs
This trait represents a set of arguments accepted by Screen::connect_static method.

ScreenDestroyedArgs
This trait represents a set of arguments accepted by Screen::destroyed method.

ScreenDisconnectArgs
This trait represents a set of arguments accepted by Screen::disconnect method.

ScreenDisconnectStaticArgs
This trait represents a set of arguments accepted by Screen::disconnect_static method.

ScreenDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Screen::dynamic_property_names method.

ScreenGeometryArgs
This trait represents a set of arguments accepted by Screen::geometry method.

ScreenGrabWindowArgs
This trait represents a set of arguments accepted by Screen::grab_window method.

ScreenMapBetweenArgs
This trait represents a set of arguments accepted by Screen::map_between method.

ScreenNameArgs
This trait represents a set of arguments accepted by Screen::name method.

ScreenObjectNameArgs
This trait represents a set of arguments accepted by Screen::object_name method.

ScreenPhysicalSizeArgs
This trait represents a set of arguments accepted by Screen::physical_size method.

ScreenPropertyArgs
This trait represents a set of arguments accepted by Screen::property method.

ScreenSizeArgs
This trait represents a set of arguments accepted by Screen::size method.

ScreenStartTimerArgs
This trait represents a set of arguments accepted by Screen::start_timer method.

ScreenTrArgs
This trait represents a set of arguments accepted by Screen::tr method.

ScreenTrUtf8Args
This trait represents a set of arguments accepted by Screen::tr_utf8 method.

ScreenTransformBetweenArgs
This trait represents a set of arguments accepted by Screen::transform_between method.

ScreenVirtualGeometryArgs
This trait represents a set of arguments accepted by Screen::virtual_geometry method.

ScreenVirtualSiblingsArgs
This trait represents a set of arguments accepted by Screen::virtual_siblings method.

ScreenVirtualSizeArgs
This trait represents a set of arguments accepted by Screen::virtual_size method.