Struct qt_gui::scroll_prepare_event::ScrollPrepareEvent

pub struct ScrollPrepareEvent {
    // some fields omitted
}

C++ type: QScrollPrepareEvent

C++ documentation:

The QScrollPrepareEvent class is sent in preparation of scrolling.

The scroll prepare event is sent before scrolling (usually by QScroller) is started. The object receiving this event should set viewportSize, maxContentPos and contentPos. It also should accept this event to indicate that scrolling should be started.

It is not guaranteed that a QScrollEvent will be sent after an acceepted QScrollPrepareEvent, e.g. in a case where the maximum content position is (0,0).

Methods

impl ScrollPrepareEvent

fn accept(&mut self)

C++ method: void QScrollPrepareEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn content_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollPrepareEventContentPosArgs<'largs>

C++ method: QScrollPrepareEvent::contentPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn content_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn content_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QScrollPrepareEvent::contentPos() const

C++ documentation:

Returns the current position of the content as set by setContentPos.

See also setContentPos().

fn content_pos_range<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollPrepareEventContentPosRangeArgs<'largs>

C++ method: QScrollPrepareEvent::contentPosRange

This is an overloaded function. Available variants:

Rust arguments:
1) fn content_pos_range(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn content_pos_range(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QScrollPrepareEvent::contentPosRange() const

C++ documentation:

Returns the range of coordinates for the content as set by setContentPosRange().

See also setContentPosRange().

fn ignore(&mut self)

C++ method: void QScrollPrepareEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QScrollPrepareEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: ScrollPrepareEventNewArgs

C++ method: QScrollPrepareEvent::QScrollPrepareEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> ::scroll_prepare_event::ScrollPrepareEvent
2) fn new((&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::scroll_prepare_event::ScrollPrepareEvent>
C++ method: [constructor] void QScrollPrepareEvent::QScrollPrepareEvent(const QPointF& startPos)

fn register_event_type<Args>(args: Args) -> c_int where Args: ScrollPrepareEventRegisterEventTypeArgs

C++ method: QScrollPrepareEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QScrollPrepareEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QScrollPrepareEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_content_pos(&mut self, pos: &PointF)

C++ method: void QScrollPrepareEvent::setContentPos(const QPointF& pos)

C++ documentation:

Sets the current content position to pos.

See also contentPos().

fn set_content_pos_range(&mut self, rect: &RectF)

C++ method: void QScrollPrepareEvent::setContentPosRange(const QRectF& rect)

C++ documentation:

Sets the range of content coordinates to rect.

See also contentPosRange().

fn set_viewport_size(&mut self, size: &SizeF)

C++ method: void QScrollPrepareEvent::setViewportSize(const QSizeF& size)

C++ documentation:

Sets the size of the area that is to be scrolled to size.

See also viewportSize().

fn spontaneous(&self) -> bool

C++ method: bool QScrollPrepareEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn start_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollPrepareEventStartPosArgs<'largs>

C++ method: QScrollPrepareEvent::startPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn start_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QScrollPrepareEvent::startPos() const

C++ documentation:

Returns the position of the touch or mouse event that started the scrolling.

fn type_(&self) -> Type

C++ method: QEvent::Type QScrollPrepareEvent::type() const

C++ documentation:

Returns the event type.

fn viewport_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ScrollPrepareEventViewportSizeArgs<'largs>

C++ method: QScrollPrepareEvent::viewportSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn viewport_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn viewport_size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QScrollPrepareEvent::viewportSize() const

C++ documentation:

Returns size of the area that is to be scrolled as set by setViewportSize

See also setViewportSize().