Struct qt_gui::set::SetByteArrayRef

pub struct SetByteArrayRef {
    // some fields omitted
}

C++ type: QSet<QByteArray>

C++ documentation:

The QSet class is a template class that provides a hash-table-based set.

QSet<T> is one of Qt's generic container classes. It stores values in an unspecified order and provides very fast lookup of the values. Internally, QSet<T> is implemented as a QHash.

Here's an example QSet with QString values:

QSet<QString> set;

To insert a value into the set, use insert():

set.insert("one");
set.insert("three");
set.insert("seven");

Another way to insert items into the set is to use operator<<():

set << "twelve" << "fifteen" << "nineteen";

To test whether an item belongs to the set or not, use contains():

if (!set.contains("ninety-nine"))
    ...

If you want to navigate through all the values stored in a QSet, you can use an iterator. QSet supports both Java-style iterators (QSetIterator and QMutableSetIterator) and STL-style iterators (QSet::iterator and QSet::const_iterator). Here's how to iterate over a QSet<QWidget > using a Java-style iterator:

QSetIterator<QWidget > i(set);
while (i.hasNext())
    qDebug() << i.next();

Here's the same code, but using an STL-style iterator:

QSet<QWidget >::const_iterator i = set.constBegin();
while (i != set.constEnd()) {
    qDebug() << i;
    ++i;
}

QSet is unordered, so an iterator's sequence cannot be assumed to be predictable. If ordering by key is required, use a QMap.

To navigate through a QSet, you can also use foreach:

QSet<QString> set;
...
foreach (const QString &value, set)
    qDebug() << value;

Items can be removed from the set using remove(). There is also a clear() function that removes all items.

QSet's value data type must be an assignable data type. You cannot, for example, store a QWidget as a value; instead, store a QWidget *. In addition, the type must provide operator==(), and there must also be a global qHash() function that returns a hash value for an argument of the key's type. See the QHash documentation for a list of types supported by qHash().

Internally, QSet uses a hash table to perform lookups. The hash table automatically grows and shrinks to provide fast lookups without wasting memory. You can still control the size of the hash table by calling reserve(), if you already know approximately how many elements the QSet will contain, but this isn't necessary to obtain good performance. You can also call capacity() to retrieve the hash table's size.

Methods

impl SetByteArrayRef

fn capacity(&self) -> c_int

C++ method: int QSet<QByteArray>::capacity() const

C++ documentation:

Returns the number of buckets in the set's internal hash table.

The sole purpose of this function is to provide a means of fine tuning QSet's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the set, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QSet<QByteArray>::clear()

C++ documentation:

Removes all elements from the set.

See also remove().

fn contains<'largs, Args>(&'largs self, args: Args) -> bool where Args: SetByteArrayRefContainsArgs<'largs>

C++ method: QSet<QByteArray>::contains

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn contains(&self, &qt_core::byte_array::ByteArray) -> bool
C++ method: bool QSet<QByteArray>::contains(const QByteArray& value) const

C++ documentation:

Returns true if the set contains item value; otherwise returns false.

See also insert(), remove(), and find().

Variant 2

Rust arguments: fn contains(&self, &::set::SetByteArrayRef) -> bool
C++ method: bool QSet<QByteArray>::contains(const QSet<QByteArray>& set) const

C++ documentation:

Returns true if the set contains all items from the other set; otherwise returns false.

This function was introduced in Qt 4.6.

See also insert(), remove(), and find().

fn count(&self) -> c_int

C++ method: int QSet<QByteArray>::count() const

C++ documentation:

Same as size().

fn detach(&mut self)

C++ method: void QSet<QByteArray>::detach()

fn empty(&self) -> bool

C++ method: bool QSet<QByteArray>::empty() const

C++ documentation:

Returns true if the set is empty. This function is provided for STL compatibility. It is equivalent to isEmpty().

fn from_list<Args>(args: Args) -> Args::ReturnType where Args: SetByteArrayRefFromListArgs

C++ method: QSet<QByteArray>::fromList

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_list((&qt_core::list::ListByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn from_list((&qt_core::list::ListByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
C++ method: static QSet<QByteArray> QSet<QByteArray>::fromList(const QList<QByteArray>& list)

C++ documentation:

Returns a new QSet object containing the data contained in list. Since QSet doesn't allow duplicates, the resulting QSet might be smaller than the list, because QList can contain duplicates.

Example:

QStringList list;
list << "Julia" << "Mike" << "Mike" << "Julia" << "Julia";

QSet<QString> set = QSet<QString>::fromList(list); set.contains("Julia"); // returns true set.contains("Mike"); // returns true set.size(); // returns 2
See also toList() and QList::toSet().

fn intersect<'l0, 'l1>(&'l0 mut self, other: &'l1 SetByteArrayRef) -> &'l0 mut SetByteArrayRef

C++ method: QSet<QByteArray>& QSet<QByteArray>::intersect(const QSet<QByteArray>& other)

C++ documentation:

Removes all items from this set that are not contained in the other set. A reference to this set is returned.

See also operator&=(), unite(), and subtract().

fn is_detached(&self) -> bool

C++ method: bool QSet<QByteArray>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QSet<QByteArray>::isEmpty() const

C++ documentation:

Returns true if the set contains no elements; otherwise returns false.

See also size().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SetByteArrayRefNewArgs

C++ method: QSet<QByteArray>::QSet

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::set::SetByteArrayRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::set::SetByteArrayRef>
C++ method: [constructor] void QSet<QByteArray>::QSet()

Variant 2

Rust arguments:
1) fn new((&::set::SetByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
2) fn new((&::set::SetByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
C++ method: [constructor] void QSet<QByteArray>::QSet(const QSet<QByteArray>& other)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefOpAddArgs<'largs>

C++ method: QSet<QByteArray>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::set::SetByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn op_add(&self, (&::set::SetByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
C++ method: QSet<QByteArray> QSet<QByteArray>::operator+(const QSet<QByteArray>& other) const

C++ documentation:

Returns a new QSet that is the union of this set and the other set.

See also unite(), operator|=(), operator&(), and operator-().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetByteArrayRef where Args: SetByteArrayRefOpAddAssignArgs<'largs>

C++ method: QSet<QByteArray>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 qt_core::byte_array::ByteArray) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator+=(const QByteArray& value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::set::SetByteArrayRef) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator+=(const QSet<QByteArray>& other)

C++ documentation:

Same as unite(other).

See also operator|(), operator&=(), and operator-=().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 SetByteArrayRef) -> &'l0 mut SetByteArrayRef

C++ method: QSet<QByteArray>& QSet<QByteArray>::operator=(const QSet<QByteArray>& other)

C++ documentation:

Assigns the other set to this set and returns a reference to this set.

fn op_bit_and<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefOpBitAndArgs<'largs>

C++ method: QSet<QByteArray>::operator&

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_bit_and(&self, (&::set::SetByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn op_bit_and(&self, (&::set::SetByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
C++ method: QSet<QByteArray> QSet<QByteArray>::operator&(const QSet<QByteArray>& other) const

C++ documentation:

Returns a new QSet that is the intersection of this set and the other set.

See also intersect(), operator&=(), operator|(), and operator-().

fn op_bit_and_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetByteArrayRef where Args: SetByteArrayRefOpBitAndAssignArgs<'largs>

C++ method: QSet<QByteArray>::operator&=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_bit_and_assign(&mut self, &'l1 qt_core::byte_array::ByteArray) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator&=(const QByteArray& value)

C++ documentation:

This is an overloaded function.

Same as intersect(other), if we consider other to be a set that contains the singleton value.

Variant 2

Rust arguments: fn op_bit_and_assign(&mut self, &'l1 ::set::SetByteArrayRef) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator&=(const QSet<QByteArray>& other)

C++ documentation:

Same as intersect(other).

See also operator&(), operator|=(), and operator-=().

fn op_bit_or<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefOpBitOrArgs<'largs>

C++ method: QSet<QByteArray>::operator|

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_bit_or(&self, (&::set::SetByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn op_bit_or(&self, (&::set::SetByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
C++ method: QSet<QByteArray> QSet<QByteArray>::operator|(const QSet<QByteArray>& other) const

C++ documentation:

Returns a new QSet that is the union of this set and the other set.

See also unite(), operator|=(), operator&(), and operator-().

fn op_bit_or_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetByteArrayRef where Args: SetByteArrayRefOpBitOrAssignArgs<'largs>

C++ method: QSet<QByteArray>::operator|=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_bit_or_assign(&mut self, &'l1 qt_core::byte_array::ByteArray) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator|=(const QByteArray& value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

Variant 2

Rust arguments: fn op_bit_or_assign(&mut self, &'l1 ::set::SetByteArrayRef) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator|=(const QSet<QByteArray>& other)

C++ documentation:

Same as unite(other).

See also operator|(), operator&=(), and operator-=().

fn op_eq(&self, other: &SetByteArrayRef) -> bool

C++ method: bool QSet<QByteArray>::operator==(const QSet<QByteArray>& other) const

C++ documentation:

Returns true if the other set is equal to this set; otherwise returns false.

Two sets are considered equal if they contain the same elements.

This function requires the value type to implement operator==().

See also operator!=().

fn op_neq(&self, other: &SetByteArrayRef) -> bool

C++ method: bool QSet<QByteArray>::operator!=(const QSet<QByteArray>& other) const

C++ documentation:

Returns true if the other set is not equal to this set; otherwise returns false.

Two sets are considered equal if they contain the same elements.

This function requires the value type to implement operator==().

See also operator==().

fn op_shl<'l0, 'l1>(&'l0 mut self, value: &'l1 ByteArray) -> &'l0 mut SetByteArrayRef

C++ method: QSet<QByteArray>& QSet<QByteArray>::operator<<(const QByteArray& value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

fn op_sub<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefOpSubArgs<'largs>

C++ method: QSet<QByteArray>::operator-

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_sub(&self, (&::set::SetByteArrayRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetByteArrayRef>
2) fn op_sub(&self, (&::set::SetByteArrayRef, cpp_utils::AsStruct)) -> ::set::SetByteArrayRef
C++ method: QSet<QByteArray> QSet<QByteArray>::operator-(const QSet<QByteArray>& other) const

C++ documentation:

Returns a new QSet that is the set difference of this set and the other set, i.e., this set - other set.

See also subtract(), operator-=(), operator|(), and operator&().

fn op_sub_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetByteArrayRef where Args: SetByteArrayRefOpSubAssignArgs<'largs>

C++ method: QSet<QByteArray>::operator-=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_sub_assign(&mut self, &'l1 qt_core::byte_array::ByteArray) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator-=(const QByteArray& value)

C++ documentation:

Removes the occurrence of item value from the set, if it is found, and returns a reference to the set. If the value is not contained the set, nothing is removed.

See also remove().

Variant 2

Rust arguments: fn op_sub_assign(&mut self, &'l1 ::set::SetByteArrayRef) -> &'l0 mut ::set::SetByteArrayRef
C++ method: QSet<QByteArray>& QSet<QByteArray>::operator-=(const QSet<QByteArray>& other)

C++ documentation:

Same as subtract(other).

See also operator-(), operator|=(), and operator&=().

fn remove(&mut self, value: &ByteArray) -> bool

C++ method: bool QSet<QByteArray>::remove(const QByteArray& value)

C++ documentation:

Removes any occurrence of item value from the set. Returns true if an item was actually removed; otherwise returns false.

See also contains() and insert().

fn reserve(&mut self, size: c_int)

C++ method: void QSet<QByteArray>::reserve(int size)

C++ documentation:

Ensures that the set's internal hash table consists of at least size buckets.

This function is useful for code that needs to build a huge set and wants to avoid repeated reallocation. For example:

QSet<QString> set;
set.reserve(20000);
for (int i = 0; i < 20000; ++i)
    set.insert(values[i]);

Ideally, size should be slightly more than the maximum number of elements expected in the set. size doesn't have to be prime, because QSet will use a prime number internally anyway. If size is an underestimate, the worst that will happen is that the QSet will be a bit slower.

In general, you will rarely ever need to call this function. QSet's internal hash table automatically shrinks or grows to provide good performance without wasting too much memory.

See also squeeze() and capacity().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QSet<QByteArray>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QSet<QByteArray>::size() const

C++ documentation:

Returns the number of items in the set.

See also isEmpty() and count().

fn squeeze(&mut self)

C++ method: void QSet<QByteArray>::squeeze()

C++ documentation:

Reduces the size of the set's internal hash table to save memory.

The sole purpose of this function is to provide a means of fine tuning QSet's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn subtract<'l0, 'l1>(&'l0 mut self, other: &'l1 SetByteArrayRef) -> &'l0 mut SetByteArrayRef

C++ method: QSet<QByteArray>& QSet<QByteArray>::subtract(const QSet<QByteArray>& other)

C++ documentation:

Removes all items from this set that are contained in the other set. Returns a reference to this set.

See also operator-=(), unite(), and intersect().

fn swap(&mut self, other: &mut SetByteArrayRef)

C++ method: void QSet<QByteArray>::swap(QSet<QByteArray>& other)

C++ documentation:

Swaps set other with this set. This operation is very fast and never fails.

fn to_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefToListArgs<'largs>

C++ method: QSet<QByteArray>::toList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn to_list(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QSet<QByteArray>::toList() const

C++ documentation:

Returns a new QList containing the elements in the set. The order of the elements in the QList is undefined.

Example:

QSet<QString> set;
set << "red" << "green" << "blue" << ... << "black";

QList<QString> list = set.toList(); qSort(list);
See also fromList() and QList::fromSet().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 SetByteArrayRef) -> &'l0 mut SetByteArrayRef

C++ method: QSet<QByteArray>& QSet<QByteArray>::unite(const QSet<QByteArray>& other)

C++ documentation:

Each item in the other set that isn't already in this set is inserted into this set. A reference to this set is returned.

See also operator|=(), intersect(), and subtract().

fn values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetByteArrayRefValuesArgs<'largs>

C++ method: QSet<QByteArray>::values

This is an overloaded function. Available variants:

Rust arguments:
1) fn values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn values(&self, cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: QList<QByteArray> QSet<QByteArray>::values() const

C++ documentation:

Returns a new QList containing the elements in the set. The order of the elements in the QList is undefined.

This is the same as toList().

See also fromList() and QList::fromSet().