Struct qt_gui::shortcut_event::ShortcutEvent

pub struct ShortcutEvent {
    // some fields omitted
}

C++ type: QShortcutEvent

C++ documentation:

The QShortcutEvent class provides an event which is generated when the user presses a key combination.

Normally you do not need to use this class directly; QShortcut provides a higher-level interface to handle shortcut keys.

Methods

impl ShortcutEvent

fn accept(&mut self)

C++ method: void QShortcutEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn ignore(&mut self)

C++ method: void QShortcutEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QShortcutEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn is_ambiguous(&self) -> bool

C++ method: bool QShortcutEvent::isAmbiguous() const

C++ documentation:

Returns true if the key sequence that triggered the event is ambiguous.

See also QShortcut::activatedAmbiguously().

fn key<'l0>(&'l0 self) -> &'l0 KeySequence

C++ method: const QKeySequence& QShortcutEvent::key() const

C++ documentation:

Returns the key sequence that triggered the event.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ShortcutEventNewArgs

C++ method: QShortcutEvent::QShortcutEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::key_sequence::KeySequence, libc::c_int, cpp_utils::AsStruct)) -> ::shortcut_event::ShortcutEvent
2) fn new((&::key_sequence::KeySequence, libc::c_int, bool, cpp_utils::AsStruct)) -> ::shortcut_event::ShortcutEvent
3) fn new((&::key_sequence::KeySequence, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::shortcut_event::ShortcutEvent>
4) fn new((&::key_sequence::KeySequence, libc::c_int, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::shortcut_event::ShortcutEvent>
C++ method: [constructor] void QShortcutEvent::QShortcutEvent(const QKeySequence& key, int id, bool ambiguous = ?)

fn register_event_type<Args>(args: Args) -> c_int where Args: ShortcutEventRegisterEventTypeArgs

C++ method: QShortcutEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QShortcutEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QShortcutEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn shortcut_id(&self) -> c_int

C++ method: int QShortcutEvent::shortcutId() const

C++ documentation:

Returns the ID of the QShortcut object for which this event was generated.

See also QShortcut::id().

fn spontaneous(&self) -> bool

C++ method: bool QShortcutEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QShortcutEvent::type() const

C++ documentation:

Returns the event type.