Struct qt_gui::standard_item::StandardItem

pub struct StandardItem {
    // some fields omitted
}

C++ type: QStandardItem

C++ documentation:

The QStandardItem class provides an item for use with the QStandardItemModel class.

Items usually contain text, icons, or checkboxes.

Each item can have its own background brush which is set with the setBackground() function. The current background brush can be found with background(). The text label for each item can be rendered with its own font and brush. These are specified with the setFont() and setForeground() functions, and read with font() and foreground().

By default, items are enabled, editable, selectable, checkable, and can be used both as the source of a drag and drop operation and as a drop target. Each item's flags can be changed by calling setFlags(). Checkable items can be checked and unchecked with the setCheckState() function. The corresponding checkState() function indicates whether the item is currently checked.

You can store application-specific data in an item by calling setData().

Each item can have a two-dimensional table of child items. This makes it possible to build hierarchies of items. The typical hierarchy is the tree, in which case the child table is a table with a single column (a list).

The dimensions of the child table can be set with setRowCount() and setColumnCount(). Items can be positioned in the child table with setChild(). Get a pointer to a child item with child(). New rows and columns of children can also be inserted with insertRow() and insertColumn(), or appended with appendRow() and appendColumn(). When using the append and insert functions, the dimensions of the child table will grow as needed.

An existing row of children can be removed with removeRow() or takeRow(); correspondingly, a column can be removed with removeColumn() or takeColumn().

An item's children can be sorted by calling sortChildren().

Methods

impl StandardItem

fn accessible_description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemAccessibleDescriptionArgs<'largs>

C++ method: QStandardItem::accessibleDescription

This is an overloaded function. Available variants:

Rust arguments:
1) fn accessible_description(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn accessible_description(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::accessibleDescription() const

C++ documentation:

Returns the item's accessible description.

The accessible description is used by assistive technologies (i.e. for users who cannot use conventional means of interaction).

See also setAccessibleDescription() and accessibleText().

fn accessible_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemAccessibleTextArgs<'largs>

C++ method: QStandardItem::accessibleText

This is an overloaded function. Available variants:

Rust arguments:
1) fn accessible_text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn accessible_text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::accessibleText() const

C++ documentation:

Returns the item's accessible text.

The accessible text is used by assistive technologies (i.e. for users who cannot use conventional means of interaction).

See also setAccessibleText() and accessibleDescription().

fn append_column(&mut self, items: &ListStandardItemMutPtr)

C++ method: void QStandardItem::appendColumn(const QList<QStandardItem*>& items)

C++ documentation:

Appends a column containing items. If necessary, the row count is increased to the size of items.

See also insertColumn().

fn append_row<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemAppendRowArgs<'largs>

C++ method: QStandardItem::appendRow

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append_row(&mut self, *mut ::standard_item::StandardItem) -> ()
C++ method: void QStandardItem::appendRow(QStandardItem* item)

C++ documentation:

This is an overloaded function.

Appends a row containing item.

When building a list or a tree that has only one column, this function provides a convenient way to append a single new item.

Variant 2

Rust arguments: fn append_row(&mut self, &::list::ListStandardItemMutPtr) -> ()
C++ method: void QStandardItem::appendRow(const QList<QStandardItem*>& items)

C++ documentation:

Appends a row containing items. If necessary, the column count is increased to the size of items.

See also insertRow().

fn append_rows(&mut self, items: &ListStandardItemMutPtr)

C++ method: void QStandardItem::appendRows(const QList<QStandardItem*>& items)

C++ documentation:

Appends rows containing items. The column count will not change.

See also insertRow().

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemBackgroundArgs<'largs>

C++ method: QStandardItem::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QStandardItem::background() const

C++ documentation:

Returns the brush used to render the item's background.

See also foreground() and setBackground().

fn check_state(&self) -> CheckState

C++ method: Qt::CheckState QStandardItem::checkState() const

C++ documentation:

Returns the checked state of the item.

See also setCheckState() and isCheckable().

fn child<'largs, Args>(&'largs self, args: Args) -> *mut StandardItem where Args: StandardItemChildArgs<'largs>

C++ method: QStandardItem::child

This is an overloaded function. Available variants:

Rust arguments:
1) fn child(&self, libc::c_int) -> *mut ::standard_item::StandardItem
2) fn child(&self, (libc::c_int, libc::c_int)) -> *mut ::standard_item::StandardItem
C++ method: QStandardItem* QStandardItem::child(int row, int column = ?) const

C++ documentation:

Returns the child item at (row, column) if one has been set; otherwise returns 0.

See also setChild(), takeChild(), and parent().

fn clone(&self) -> *mut StandardItem

C++ method: virtual QStandardItem* QStandardItem::clone() const

C++ documentation:

Returns a copy of this item. The item's children are not copied.

When subclassing QStandardItem, you can reimplement this function to provide QStandardItemModel with a factory that it can use to create new items on demand.

See also QStandardItemModel::setItemPrototype() and operator=().

fn column(&self) -> c_int

C++ method: int QStandardItem::column() const

C++ documentation:

Returns the column where the item is located in its parent's child table, or -1 if the item has no parent.

See also row() and parent().

fn column_count(&self) -> c_int

C++ method: int QStandardItem::columnCount() const

C++ documentation:

Returns the number of child item columns that the item has.

See also setColumnCount() and rowCount().

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemDataArgs<'largs>

C++ method: QStandardItem::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn data(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
3) fn data(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
4) fn data(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: virtual QVariant QStandardItem::data(int role = ?) const

C++ documentation:

Returns the item's data for the given role, or an invalid QVariant if there is no data for the role.

Note: The default implementation treats Qt::EditRole and Qt::DisplayRole as referring to the same data.

See also setData().

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QStandardItem::flags() const

C++ documentation:

Returns the item flags for the item.

The item flags determine how the user can interact with the item.

By default, items are enabled, editable, selectable, checkable, and can be used both as the source of a drag and drop operation and as a drop target.

See also setFlags().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemFontArgs<'largs>

C++ method: QStandardItem::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> ::font::Font
C++ method: QFont QStandardItem::font() const

C++ documentation:

Returns the font used to render the item's text.

See also setFont().

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemForegroundArgs<'largs>

C++ method: QStandardItem::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QStandardItem::foreground() const

C++ documentation:

Returns the brush used to render the item's foreground (e.g. text).

See also setForeground() and background().

fn has_children(&self) -> bool

C++ method: bool QStandardItem::hasChildren() const

C++ documentation:

Returns true if this item has any children; otherwise returns false.

See also rowCount(), columnCount(), and child().

fn icon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemIconArgs<'largs>

C++ method: QStandardItem::icon

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::icon::Icon>
2) fn icon(&self, cpp_utils::AsStruct) -> ::icon::Icon
C++ method: QIcon QStandardItem::icon() const

C++ documentation:

Returns the item's icon.

See also setIcon() and iconSize.

fn index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemIndexArgs<'largs>

C++ method: QStandardItem::index

This is an overloaded function. Available variants:

Rust arguments:
1) fn index(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::model_index::ModelIndex>
2) fn index(&self, cpp_utils::AsStruct) -> qt_core::model_index::ModelIndex
C++ method: QModelIndex QStandardItem::index() const

C++ documentation:

Returns the QModelIndex associated with this item.

When you need to invoke item functionality in a QModelIndex-based API (e.g. QAbstractItemView), you can call this function to obtain an index that corresponds to the item's location in the model.

If the item is not associated with a model, an invalid QModelIndex is returned.

See also model() and QStandardItemModel::itemFromIndex().

fn insert_column(&mut self, column: c_int, items: &ListStandardItemMutPtr)

C++ method: void QStandardItem::insertColumn(int column, const QList<QStandardItem*>& items)

C++ documentation:

Inserts a column at column containing items. If necessary, the row count is increased to the size of items.

See also insertColumns() and insertRow().

fn insert_columns(&mut self, column: c_int, count: c_int)

C++ method: void QStandardItem::insertColumns(int column, int count)

C++ documentation:

Inserts count columns of child items at column column.

See also insertColumn() and insertRows().

fn insert_row<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemInsertRowArgs<'largs>

C++ method: QStandardItem::insertRow

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_row(&mut self, (libc::c_int, *mut ::standard_item::StandardItem)) -> ()
C++ method: void QStandardItem::insertRow(int row, QStandardItem* item)

C++ documentation:

This is an overloaded function.

Inserts a row at row containing item.

When building a list or a tree that has only one column, this function provides a convenient way to insert a single new item.

Variant 2

Rust arguments: fn insert_row(&mut self, (libc::c_int, &::list::ListStandardItemMutPtr)) -> ()
C++ method: void QStandardItem::insertRow(int row, const QList<QStandardItem*>& items)

C++ documentation:

Inserts a row at row containing items. If necessary, the column count is increased to the size of items.

See also insertRows() and insertColumn().

fn insert_rows<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemInsertRowsArgs<'largs>

C++ method: QStandardItem::insertRows

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_rows(&mut self, (libc::c_int, &::list::ListStandardItemMutPtr)) -> ()
C++ method: void QStandardItem::insertRows(int row, const QList<QStandardItem*>& items)

C++ documentation:

Inserts items at row. The column count won't be changed.

See also insertRow() and insertColumn().

Variant 2

Rust arguments: fn insert_rows(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QStandardItem::insertRows(int row, int count)

C++ documentation:

Inserts count rows of child items at row row.

See also insertRow() and insertColumns().

fn is_checkable(&self) -> bool

C++ method: bool QStandardItem::isCheckable() const

C++ documentation:

Returns whether the item is user-checkable.

The default value is false.

See also setCheckable(), checkState(), and isTristate().

fn is_drag_enabled(&self) -> bool

C++ method: bool QStandardItem::isDragEnabled() const

C++ documentation:

Returns whether the item is drag enabled. An item that is drag enabled can be dragged by the user.

The default value is true.

Note that item dragging must be enabled in the view for dragging to work; see QAbstractItemView::dragEnabled.

See also setDragEnabled(), isDropEnabled(), and flags().

fn is_drop_enabled(&self) -> bool

C++ method: bool QStandardItem::isDropEnabled() const

C++ documentation:

Returns whether the item is drop enabled. When an item is drop enabled, it can be used as a drop target.

The default value is true.

See also setDropEnabled(), isDragEnabled(), and flags().

fn is_editable(&self) -> bool

C++ method: bool QStandardItem::isEditable() const

C++ documentation:

Returns whether the item can be edited by the user.

When an item is editable (and enabled), the user can edit the item by invoking one of the view's edit triggers; see QAbstractItemView::editTriggers.

The default value is true.

See also setEditable() and flags().

fn is_enabled(&self) -> bool

C++ method: bool QStandardItem::isEnabled() const

C++ documentation:

Returns whether the item is enabled.

When an item is enabled, the user can interact with it. The possible types of interaction are specified by the other item flags, such as isEditable() and isSelectable().

The default value is true.

See also setEnabled() and flags().

fn is_selectable(&self) -> bool

C++ method: bool QStandardItem::isSelectable() const

C++ documentation:

Returns whether the item is selectable by the user.

The default value is true.

See also setSelectable() and flags().

fn is_tristate(&self) -> bool

C++ method: bool QStandardItem::isTristate() const

C++ documentation:

Returns whether the item is tristate; that is, if it's checkable with three separate states.

The default value is false.

See also setTristate(), isCheckable(), and checkState().

fn model(&self) -> *mut StandardItemModel

C++ method: QStandardItemModel* QStandardItem::model() const

C++ documentation:

Returns the QStandardItemModel that this item belongs to.

If the item is not a child of another item that belongs to the model, this function returns 0.

See also index().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StandardItemNewArgs

C++ method: QStandardItem::QStandardItem

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::standard_item::StandardItem
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::standard_item::StandardItem>
C++ method: [constructor] void QStandardItem::QStandardItem()

Variant 2

Rust arguments:
1) fn new((&::icon::Icon, &qt_core::string::String, cpp_utils::AsStruct)) -> ::standard_item::StandardItem
2) fn new((&::icon::Icon, &qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::standard_item::StandardItem>
C++ method: [constructor] void QStandardItem::QStandardItem(const QIcon& icon, const QString& text)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::standard_item::StandardItem
2) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::standard_item::StandardItem>
C++ method: [constructor] void QStandardItem::QStandardItem(const QString& text)

Variant 4

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::standard_item::StandardItem
2) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::standard_item::StandardItem
3) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::standard_item::StandardItem>
4) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::standard_item::StandardItem>
C++ method: [constructor] void QStandardItem::QStandardItem(int rows, int columns = ?)

fn op_lt(&self, other: &StandardItem) -> bool

C++ method: virtual bool QStandardItem::operator<(const QStandardItem& other) const

C++ documentation:

Returns true if this item is less than other; otherwise returns false.

The default implementation uses the data for the item's sort role (see QStandardItemModel::sortRole) to perform the comparison if the item belongs to a model; otherwise, the data for the item's Qt::DisplayRole (text()) is used to perform the comparison.

sortChildren() and QStandardItemModel::sort() use this function when sorting items. If you want custom sorting, you can subclass QStandardItem and reimplement this function.

fn parent(&self) -> *mut StandardItem

C++ method: QStandardItem* QStandardItem::parent() const

C++ documentation:

Returns the item's parent item, or 0 if the item has no parent.

Note: For toplevel items parent() returns 0. To receive toplevel item's parent use QStandardItemModel::invisibleRootItem() instead.

See also child() and QStandardItemModel::invisibleRootItem().

fn read(&mut self, in_: &mut DataStream)

C++ method: virtual void QStandardItem::read(QDataStream& in)

C++ documentation:

Reads the item from stream in. Only the data and flags of the item are read, not the child items.

See also write().

fn remove_column(&mut self, column: c_int)

C++ method: void QStandardItem::removeColumn(int column)

C++ documentation:

Removes the given column. The items that were in the column are deleted.

See also takeColumn(), removeColumns(), and removeRow().

fn remove_columns(&mut self, column: c_int, count: c_int)

C++ method: void QStandardItem::removeColumns(int column, int count)

C++ documentation:

Removes count columns at column column. The items that were in those columns are deleted.

See also removeColumn() and removeRows().

fn remove_row(&mut self, row: c_int)

C++ method: void QStandardItem::removeRow(int row)

C++ documentation:

Removes the given row. The items that were in the row are deleted.

See also takeRow(), removeRows(), and removeColumn().

fn remove_rows(&mut self, row: c_int, count: c_int)

C++ method: void QStandardItem::removeRows(int row, int count)

C++ documentation:

Removes count rows at row row. The items that were in those rows are deleted.

See also removeRow() and removeColumn().

fn row(&self) -> c_int

C++ method: int QStandardItem::row() const

C++ documentation:

Returns the row where the item is located in its parent's child table, or -1 if the item has no parent.

See also column() and parent().

fn row_count(&self) -> c_int

C++ method: int QStandardItem::rowCount() const

C++ documentation:

Returns the number of child item rows that the item has.

See also setRowCount() and columnCount().

fn set_accessible_description(&mut self, accessible_description: &String)

C++ method: void QStandardItem::setAccessibleDescription(const QString& accessibleDescription)

C++ documentation:

Sets the item's accessible description to the string specified by accessibleDescription.

The accessible description is used by assistive technologies (i.e. for users who cannot use conventional means of interaction).

See also accessibleDescription() and setAccessibleText().

fn set_accessible_text(&mut self, accessible_text: &String)

C++ method: void QStandardItem::setAccessibleText(const QString& accessibleText)

C++ documentation:

Sets the item's accessible text to the string specified by accessibleText.

The accessible text is used by assistive technologies (i.e. for users who cannot use conventional means of interaction).

See also accessibleText() and setAccessibleDescription().

fn set_background(&mut self, brush: &Brush)

C++ method: void QStandardItem::setBackground(const QBrush& brush)

C++ documentation:

Sets the item's background brush to the specified brush.

See also background() and setForeground().

fn set_check_state(&mut self, check_state: CheckState)

C++ method: void QStandardItem::setCheckState(Qt::CheckState checkState)

C++ documentation:

Sets the check state of the item to be state.

See also checkState() and setCheckable().

fn set_checkable(&mut self, checkable: bool)

C++ method: void QStandardItem::setCheckable(bool checkable)

C++ documentation:

Sets whether the item is user-checkable. If checkable is true, the item can be checked by the user; otherwise, the user cannot check the item.

The item delegate will render a checkable item with a check box next to the item's text.

See also isCheckable(), setCheckState(), and setTristate().

fn set_child<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemSetChildArgs<'largs>

C++ method: QStandardItem::setChild

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_child(&mut self, (libc::c_int, *mut ::standard_item::StandardItem)) -> ()
C++ method: void QStandardItem::setChild(int row, QStandardItem* item)

C++ documentation:

This is an overloaded function.

Sets the child at row to item.

Variant 2

Rust arguments: fn set_child(&mut self, (libc::c_int, libc::c_int, *mut ::standard_item::StandardItem)) -> ()
C++ method: void QStandardItem::setChild(int row, int column, QStandardItem* item)

C++ documentation:

Sets the child item at (row, column) to item. This item (the parent item) takes ownership of item. If necessary, the row count and column count are increased to fit the item.

See also child().

fn set_column_count(&mut self, columns: c_int)

C++ method: void QStandardItem::setColumnCount(int columns)

C++ documentation:

Sets the number of child item columns to columns. If this is less than columnCount(), the data in the unwanted columns is discarded.

See also columnCount() and setRowCount().

fn set_data<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemSetDataArgs<'largs>

C++ method: QStandardItem::setData

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_data(&mut self, &qt_core::variant::Variant) -> ()
2) fn set_data(&mut self, (&qt_core::variant::Variant, libc::c_int)) -> ()
C++ method: virtual void QStandardItem::setData(const QVariant& value, int role = ?)

C++ documentation:

Sets the item's data for the given role to the specified value.

If you subclass QStandardItem and reimplement this function, your reimplementation should call emitDataChanged() if you do not call the base implementation of setData(). This will ensure that e.g. views using the model are notified of the changes.

Note: The default implementation treats Qt::EditRole and Qt::DisplayRole as referring to the same data.

See also Qt::ItemDataRole, data(), and setFlags().

fn set_drag_enabled(&mut self, drag_enabled: bool)

C++ method: void QStandardItem::setDragEnabled(bool dragEnabled)

C++ documentation:

Sets whether the item is drag enabled. If dragEnabled is true, the item can be dragged by the user; otherwise, the user cannot drag the item.

Note that you also need to ensure that item dragging is enabled in the view; see QAbstractItemView::dragEnabled.

See also isDragEnabled(), setDropEnabled(), and setFlags().

fn set_drop_enabled(&mut self, drop_enabled: bool)

C++ method: void QStandardItem::setDropEnabled(bool dropEnabled)

C++ documentation:

Sets whether the item is drop enabled. If dropEnabled is true, the item can be used as a drop target; otherwise, it cannot.

Note that you also need to ensure that drops are enabled in the view; see QWidget::acceptDrops(); and that the model supports the desired drop actions; see QAbstractItemModel::supportedDropActions().

See also isDropEnabled(), setDragEnabled(), and setFlags().

fn set_editable(&mut self, editable: bool)

C++ method: void QStandardItem::setEditable(bool editable)

C++ documentation:

Sets whether the item is editable. If editable is true, the item can be edited by the user; otherwise, the user cannot edit the item.

How the user can edit items in a view is determined by the view's edit triggers; see QAbstractItemView::editTriggers.

See also isEditable() and setFlags().

fn set_enabled(&mut self, enabled: bool)

C++ method: void QStandardItem::setEnabled(bool enabled)

C++ documentation:

Sets whether the item is enabled. If enabled is true, the item is enabled, meaning that the user can interact with the item; if enabled is false, the user cannot interact with the item.

This flag takes precedence over the other item flags; e.g. if an item is not enabled, it cannot be selected by the user, even if the Qt::ItemIsSelectable flag has been set.

See also isEnabled(), Qt::ItemIsEnabled, and setFlags().

fn set_flags(&mut self, flags: Flags<ItemFlag>)

C++ method: void QStandardItem::setFlags(QFlags<Qt::ItemFlag> flags)

C++ documentation:

Sets the item flags for the item to flags.

The item flags determine how the user can interact with the item. This is often used to disable an item.

See also flags() and setData().

fn set_font(&mut self, font: &Font)

C++ method: void QStandardItem::setFont(const QFont& font)

C++ documentation:

Sets the font used to display the item's text to the given font.

See also font(), setText(), and setForeground().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QStandardItem::setForeground(const QBrush& brush)

C++ documentation:

Sets the brush used to display the item's foreground (e.g. text) to the given brush.

See also foreground(), setBackground(), and setFont().

fn set_icon(&mut self, icon: &Icon)

C++ method: void QStandardItem::setIcon(const QIcon& icon)

C++ documentation:

Sets the item's icon to the icon specified.

See also icon().

fn set_row_count(&mut self, rows: c_int)

C++ method: void QStandardItem::setRowCount(int rows)

C++ documentation:

Sets the number of child item rows to rows. If this is less than rowCount(), the data in the unwanted rows is discarded.

See also rowCount() and setColumnCount().

fn set_selectable(&mut self, selectable: bool)

C++ method: void QStandardItem::setSelectable(bool selectable)

C++ documentation:

Sets whether the item is selectable. If selectable is true, the item can be selected by the user; otherwise, the user cannot select the item.

You can control the selection behavior and mode by manipulating their view properties; see QAbstractItemView::selectionMode and QAbstractItemView::selectionBehavior.

See also isSelectable() and setFlags().

fn set_size_hint(&mut self, size_hint: &Size)

C++ method: void QStandardItem::setSizeHint(const QSize& sizeHint)

C++ documentation:

Sets the size hint for the item to be size. If no size hint is set, the item delegate will compute the size hint based on the item data.

See also sizeHint().

fn set_status_tip(&mut self, status_tip: &String)

C++ method: void QStandardItem::setStatusTip(const QString& statusTip)

C++ documentation:

Sets the item's status tip to the string specified by statusTip.

See also statusTip(), setToolTip(), and setWhatsThis().

fn set_text(&mut self, text: &String)

C++ method: void QStandardItem::setText(const QString& text)

C++ documentation:

Sets the item's text to the text specified.

See also text(), setFont(), and setForeground().

fn set_text_alignment(&mut self, text_alignment: Flags<AlignmentFlag>)

C++ method: void QStandardItem::setTextAlignment(QFlags<Qt::AlignmentFlag> textAlignment)

C++ documentation:

Sets the text alignment for the item's text to the alignment specified.

See also textAlignment().

fn set_tool_tip(&mut self, tool_tip: &String)

C++ method: void QStandardItem::setToolTip(const QString& toolTip)

C++ documentation:

Sets the item's tooltip to the string specified by toolTip.

See also toolTip(), setStatusTip(), and setWhatsThis().

fn set_tristate(&mut self, tristate: bool)

C++ method: void QStandardItem::setTristate(bool tristate)

C++ documentation:

Sets whether the item is tristate. If tristate is true, the item is checkable with three separate states; otherwise, the item is checkable with two states. (Note that this also requires that the item is checkable; see isCheckable().)

See also isTristate(), setCheckable(), and setCheckState().

fn set_whats_this(&mut self, whats_this: &String)

C++ method: void QStandardItem::setWhatsThis(const QString& whatsThis)

C++ documentation:

Sets the item's "What's This?" help to the string specified by whatsThis.

See also whatsThis(), setStatusTip(), and setToolTip().

fn size_hint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemSizeHintArgs<'largs>

C++ method: QStandardItem::sizeHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn size_hint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size_hint(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: QSize QStandardItem::sizeHint() const

C++ documentation:

Returns the size hint set for the item, or an invalid QSize if no size hint has been set.

If no size hint has been set, the item delegate will compute the size hint based on the item data.

See also setSizeHint().

fn sort_children<'largs, Args>(&'largs mut self, args: Args) where Args: StandardItemSortChildrenArgs<'largs>

C++ method: QStandardItem::sortChildren

This is an overloaded function. Available variants:

Rust arguments:
1) fn sort_children(&mut self, libc::c_int) -> ()
2) fn sort_children(&mut self, (libc::c_int, qt_core::qt::SortOrder)) -> ()
C++ method: void QStandardItem::sortChildren(int column, Qt::SortOrder order = ?)

C++ documentation:

Sorts the children of the item using the given order, by the values in the given column.

Note: This function is recursive, therefore it sorts the children of the item, its grandchildren, etc.

See also operator<().

fn status_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemStatusTipArgs<'largs>

C++ method: QStandardItem::statusTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn status_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn status_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::statusTip() const

C++ documentation:

Returns the item's status tip.

See also setStatusTip(), toolTip(), and whatsThis().

fn take_child<'largs, Args>(&'largs mut self, args: Args) -> *mut StandardItem where Args: StandardItemTakeChildArgs<'largs>

C++ method: QStandardItem::takeChild

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_child(&mut self, libc::c_int) -> *mut ::standard_item::StandardItem
2) fn take_child(&mut self, (libc::c_int, libc::c_int)) -> *mut ::standard_item::StandardItem
C++ method: QStandardItem* QStandardItem::takeChild(int row, int column = ?)

C++ documentation:

Removes the child item at (row, column) without deleting it, and returns a pointer to the item. If there was no child at the given location, then this function returns 0.

Note that this function, unlike takeRow() and takeColumn(), does not affect the dimensions of the child table.

See also child(), takeRow(), and takeColumn().

fn take_column<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: StandardItemTakeColumnArgs<'largs>

C++ method: QStandardItem::takeColumn

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_column(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStandardItemMutPtr>
2) fn take_column(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListStandardItemMutPtr
C++ method: QList<QStandardItem*> QStandardItem::takeColumn(int column)

C++ documentation:

Removes column without deleting the column items, and returns a list of pointers to the removed items. For items in the column that have not been set, the corresponding pointers in the list will be 0.

See also removeColumn(), insertColumn(), and takeRow().

fn take_row<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: StandardItemTakeRowArgs<'largs>

C++ method: QStandardItem::takeRow

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_row(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStandardItemMutPtr>
2) fn take_row(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListStandardItemMutPtr
C++ method: QList<QStandardItem*> QStandardItem::takeRow(int row)

C++ documentation:

Removes row without deleting the row items, and returns a list of pointers to the removed items. For items in the row that have not been set, the corresponding pointers in the list will be 0.

See also removeRow(), insertRow(), and takeColumn().

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemTextArgs<'largs>

C++ method: QStandardItem::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::text() const

C++ documentation:

Returns the item's text. This is the text that's presented to the user in a view.

See also setText().

fn text_alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QStandardItem::textAlignment() const

C++ documentation:

Returns the text alignment for the item's text.

See also setTextAlignment().

fn tool_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemToolTipArgs<'largs>

C++ method: QStandardItem::toolTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn tool_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tool_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::toolTip() const

C++ documentation:

Returns the item's tooltip.

See also setToolTip(), statusTip(), and whatsThis().

fn type_(&self) -> c_int

C++ method: virtual int QStandardItem::type() const

C++ documentation:

Returns the type of this item. The type is used to distinguish custom items from the base class. When subclassing QStandardItem, you should reimplement this function and return a new value greater than or equal to UserType.

See also QStandardItem::Type.

fn whats_this<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StandardItemWhatsThisArgs<'largs>

C++ method: QStandardItem::whatsThis

This is an overloaded function. Available variants:

Rust arguments:
1) fn whats_this(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn whats_this(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStandardItem::whatsThis() const

C++ documentation:

Returns the item's "What's This?" help.

See also setWhatsThis(), toolTip(), and statusTip().

fn write(&self, out: &mut DataStream)

C++ method: virtual void QStandardItem::write(QDataStream& out) const

C++ documentation:

Writes the item to stream out. Only the data and flags of the item are written, not the child items.

See also read().