Struct qt_gui::static_text::StaticText

pub struct StaticText {
    // some fields omitted
}

C++ type: QStaticText

C++ documentation:

The QStaticText class enables optimized drawing of text when the text and its layout is updated rarely.

QStaticText provides a way to cache layout data for a block of text so that it can be drawn more efficiently than by using QPainter::drawText() in which the layout information is recalculated with every call.

The class primarily provides an optimization for cases where the text, its font and the transformations on the painter are static over several paint events. If the text or its layout is changed for every iteration, QPainter::drawText() is the more efficient alternative, since the static text's layout would have to be recalculated to take the new state into consideration.

Translating the painter will not cause the layout of the text to be recalculated, but will cause a very small performance impact on drawStaticText(). Altering any other parts of the painter's transformation or the painter's font will cause the layout of the static text to be recalculated. This should be avoided as often as possible to maximize the performance benefit of using QStaticText.

In addition, only affine transformations are supported by drawStaticText(). Calling drawStaticText() on a projected painter will perform slightly worse than using the regular drawText() call, so this should be avoided.

class MyWidget: public QWidget
{
public:
    MyWidget(QWidget *parent = 0) : QWidget(parent), m_staticText("This is static text")

protected: void paintEvent(QPaintEvent *) { QPainter painter(this); painter.drawStaticText(0, 0, m_staticText); }

private: QStaticText m_staticText; };
The QStaticText class can be used to mimic the behavior of QPainter::drawText() to a specific point with no boundaries, and also when QPainter::drawText() is called with a bounding rectangle.

If a bounding rectangle is not required, create a QStaticText object without setting a preferred text width. The text will then occupy a single line.

If you set a text width on the QStaticText object, this will bound the text. The text will be formatted so that no line exceeds the given width. The text width set for QStaticText will not automatically be used for clipping. To achieve clipping in addition to line breaks, use QPainter::setClipRect(). The position of the text is decided by the argument passed to QPainter::drawStaticText() and can change from call to call with a minimal impact on performance.

For extra convenience, it is possible to apply formatting to the text using the HTML subset supported by QTextDocument. QStaticText will attempt to guess the format of the input text using Qt::mightBeRichText(), and interpret it as rich text if this function returns true. To force QStaticText to display its contents as either plain text or rich text, use the function QStaticText::setTextFormat() and pass in, respectively, Qt::PlainText and Qt::RichText.

QStaticText can only represent text, so only HTML tags which alter the layout or appearance of the text will be respected. Adding an image to the input HTML, for instance, will cause the image to be included as part of the layout, affecting the positions of the text glyphs, but it will not be displayed. The result will be an empty area the size of the image in the output. Similarly, using tables will cause the text to be laid out in table format, but the borders will not be drawn.

If it's the first time the static text is drawn, or if the static text, or the painter's font has been altered since the last time it was drawn, the text's layout has to be recalculated. On some paint engines, changing the matrix of the painter will also cause the layout to be recalculated. In particular, this will happen for any engine except for the OpenGL2 paint engine. Recalculating the layout will impose an overhead on the QPainter::drawStaticText() call where it occurs. To avoid this overhead in the paint event, you can call prepare() ahead of time to ensure that the layout is calculated.

Methods

impl StaticText

fn new<Args>(args: Args) -> Args::ReturnType where Args: StaticTextNewArgs

C++ method: QStaticText::QStaticText

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::static_text::StaticText
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::static_text::StaticText>
C++ method: [constructor] void QStaticText::QStaticText()

Variant 2

Rust arguments:
1) fn new((&::static_text::StaticText, cpp_utils::AsStruct)) -> ::static_text::StaticText
2) fn new((&::static_text::StaticText, cpp_utils::AsBox)) -> cpp_utils::CppBox<::static_text::StaticText>
C++ method: [constructor] void QStaticText::QStaticText(const QStaticText& other)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::static_text::StaticText
2) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::static_text::StaticText>
C++ method: [constructor] void QStaticText::QStaticText(const QString& text)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 StaticText) -> &'l0 mut StaticText

C++ method: QStaticText& QStaticText::operator=(const QStaticText& arg1)

C++ documentation:

Assigns other to this QStaticText.

fn op_eq(&self, arg1: &StaticText) -> bool

C++ method: bool QStaticText::operator==(const QStaticText& arg1) const

C++ documentation:

Compares other to this QStaticText. Returns true if the texts, fonts and text widths are equal.

fn op_neq(&self, arg1: &StaticText) -> bool

C++ method: bool QStaticText::operator!=(const QStaticText& arg1) const

C++ documentation:

Compares other to this QStaticText. Returns true if the texts, fonts or maximum sizes are different.

fn performance_hint(&self) -> PerformanceHint

C++ method: QStaticText::PerformanceHint QStaticText::performanceHint() const

C++ documentation:

Returns which performance hint is set for the QStaticText.

See also setPerformanceHint().

fn prepare<'largs, Args>(&'largs mut self, args: Args) where Args: StaticTextPrepareArgs<'largs>

C++ method: QStaticText::prepare

This is an overloaded function. Available variants:

Rust arguments:
1) fn prepare(&mut self, &::transform::Transform) -> ()
2) fn prepare(&mut self, (&::transform::Transform, &::font::Font)) -> ()
3) fn prepare(&mut self, ()) -> ()
C++ method: void QStaticText::prepare(const QTransform& matrix = ?, const QFont& font = ?)

C++ documentation:

Prepares the QStaticText object for being painted with the given matrix and the given font to avoid overhead when the actual drawStaticText() call is made.

When drawStaticText() is called, the layout of the QStaticText will be recalculated if any part of the QStaticText object has changed since the last time it was drawn. It will also be recalculated if the painter's font is not the same as when the QStaticText was last drawn, or, on any other paint engine than the OpenGL2 engine, if the painter's matrix has been altered since the static text was last drawn.

To avoid the overhead of creating the layout the first time you draw the QStaticText after making changes, you can use the prepare() function and pass in the matrix and font you expect to use when drawing the text.

See also QPainter::setFont() and QPainter::setMatrix().

fn set_performance_hint(&mut self, performance_hint: PerformanceHint)

C++ method: void QStaticText::setPerformanceHint(QStaticText::PerformanceHint performanceHint)

C++ documentation:

Sets the performance hint of the QStaticText according to the performanceHint provided. The performanceHint is used to customize how much caching is done internally to improve performance.

The default is QStaticText::ModerateCaching.

Note: This function will cause the layout of the text to require recalculation.

See also performanceHint().

fn set_text(&mut self, text: &String)

C++ method: void QStaticText::setText(const QString& text)

C++ documentation:

Sets the text of the QStaticText to text.

Note: This function will cause the layout of the text to require recalculation.

See also text().

fn set_text_format(&mut self, text_format: TextFormat)

C++ method: void QStaticText::setTextFormat(Qt::TextFormat textFormat)

C++ documentation:

Sets the text format of the QStaticText to textFormat. If textFormat is set to Qt::AutoText (the default), the format of the text will try to be determined using the function Qt::mightBeRichText(). If the text format is Qt::PlainText, then the text will be displayed as is, whereas it will be interpreted as HTML if the format is Qt::RichText. HTML tags that alter the font of the text, its color, or its layout are supported by QStaticText.

Note: This function will cause the layout of the text to require recalculation.

See also textFormat(), setText(), and text().

fn set_text_option(&mut self, text_option: &TextOption)

C++ method: void QStaticText::setTextOption(const QTextOption& textOption)

C++ documentation:

Sets the text option structure that controls the layout process to the given textOption.

See also textOption().

fn set_text_width(&mut self, text_width: c_double)

C++ method: void QStaticText::setTextWidth(double textWidth)

C++ documentation:

Sets the preferred width for this QStaticText. If the text is wider than the specified width, it will be broken into multiple lines and grow vertically. If the text cannot be split into multiple lines, it will be larger than the specified textWidth.

Setting the preferred text width to a negative number will cause the text to be unbounded.

Use size() to get the actual size of the text.

Note: This function will cause the layout of the text to require recalculation.

See also textWidth() and size().

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StaticTextSizeArgs<'largs>

C++ method: QStaticText::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size_f::SizeF
C++ method: QSizeF QStaticText::size() const

C++ documentation:

Returns the size of the bounding rect for this QStaticText.

See also textWidth().

fn swap(&mut self, other: &mut StaticText)

C++ method: void QStaticText::swap(QStaticText& other)

C++ documentation:

Swaps this static text instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StaticTextTextArgs<'largs>

C++ method: QStaticText::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QStaticText::text() const

C++ documentation:

Returns the text of the QStaticText.

See also setText().

fn text_format(&self) -> TextFormat

C++ method: Qt::TextFormat QStaticText::textFormat() const

C++ documentation:

Returns the text format of the QStaticText.

See also setTextFormat(), setText(), and text().

fn text_option<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StaticTextTextOptionArgs<'largs>

C++ method: QStaticText::textOption

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_option(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_option::TextOption>
2) fn text_option(&self, cpp_utils::AsStruct) -> ::text_option::TextOption
C++ method: QTextOption QStaticText::textOption() const

C++ documentation:

Returns the current text option used to control the layout process.

See also setTextOption().

fn text_width(&self) -> c_double

C++ method: double QStaticText::textWidth() const

C++ documentation:

Returns the preferred width for this QStaticText.

See also setTextWidth().