Struct qt_gui::surface::Surface

pub struct Surface {
    // some fields omitted
}

C++ type: QSurface

C++ documentation:

The QSurface class is an abstraction of renderable surfaces in Qt.

The size of the surface is accessible with the size() function. The rendering specific attributes of the surface are accessible through the format() function.

Methods

impl Surface

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SurfaceFormatArgs<'largs>

C++ method: QSurface::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
2) fn format(&self, cpp_utils::AsStruct) -> ::surface_format::SurfaceFormat
C++ method: pure virtual QSurfaceFormat QSurface::format() const

C++ documentation:

Returns the format of the surface.

fn size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SurfaceSizeArgs<'largs>

C++ method: QSurface::size

This is an overloaded function. Available variants:

Rust arguments:
1) fn size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::size::Size>
2) fn size(&self, cpp_utils::AsStruct) -> qt_core::size::Size
C++ method: pure virtual QSize QSurface::size() const

C++ documentation:

Returns the size of the surface in pixels.

fn supports_opengl(&self) -> bool

C++ method: bool QSurface::supportsOpenGL() const

C++ documentation:

Returns true if the surface is OpenGL compatible and can be used in conjunction with QOpenGLContext; otherwise returns false.

This function was introduced in Qt 5.3.

fn surface_class(&self) -> SurfaceClass

C++ method: QSurface::SurfaceClass QSurface::surfaceClass() const

C++ documentation:

Returns the surface class of this surface.

fn surface_type(&self) -> SurfaceType

C++ method: pure virtual QSurface::SurfaceType QSurface::surfaceType() const

C++ documentation:

Returns the type of the surface.