Struct qt_gui::surface_format::SurfaceFormat

pub struct SurfaceFormat {
    // some fields omitted
}

C++ type: QSurfaceFormat

C++ documentation:

The QSurfaceFormat class represents the format of a QSurface.

The format includes the size of the color buffers, red, green, and blue; the size of the alpha buffer; the size of the depth and stencil buffers; and number of samples per pixel for multisampling. In addition, the format contains surface configuration parameters such as OpenGL profile and version for rendering, whether or not to enable stereo buffers, and swap behaviour.

Methods

impl SurfaceFormat

fn alpha_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::alphaBufferSize() const

C++ documentation:

Get the size in bits of the alpha channel of the color buffer.

See also setAlphaBufferSize().

fn blue_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::blueBufferSize() const

C++ documentation:

Get the size in bits of the blue channel of the color buffer.

See also setBlueBufferSize().

fn default_format<Args>(args: Args) -> Args::ReturnType where Args: SurfaceFormatDefaultFormatArgs

C++ method: QSurfaceFormat::defaultFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn default_format(cpp_utils::AsBox) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
2) fn default_format(cpp_utils::AsStruct) -> ::surface_format::SurfaceFormat
C++ method: static QSurfaceFormat QSurfaceFormat::defaultFormat()

C++ documentation:

Returns the global default surface format.

When setDefaultFormat() is not called, this is a default-constructed QSurfaceFormat.

This function was introduced in Qt 5.4.

See also setDefaultFormat().

fn depth_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::depthBufferSize() const

C++ documentation:

Returns the depth buffer size.

See also setDepthBufferSize().

fn green_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::greenBufferSize() const

C++ documentation:

Get the size in bits of the green channel of the color buffer.

See also setGreenBufferSize().

fn has_alpha(&self) -> bool

C++ method: bool QSurfaceFormat::hasAlpha() const

C++ documentation:

Returns true if the alpha buffer size is greater than zero.

This means that the surface might be used with per pixel translucency effects.

fn major_version(&self) -> c_int

C++ method: int QSurfaceFormat::majorVersion() const

C++ documentation:

Returns the major OpenGL version.

The default version is 2.0.

See also setMajorVersion().

fn minor_version(&self) -> c_int

C++ method: int QSurfaceFormat::minorVersion() const

C++ documentation:

Returns the minor OpenGL version.

See also setMinorVersion().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SurfaceFormatNewArgs

C++ method: QSurfaceFormat::QSurfaceFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::surface_format::SurfaceFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
C++ method: [constructor] void QSurfaceFormat::QSurfaceFormat()

Variant 2

Rust arguments:
1) fn new((qt_core::flags::Flags<::surface_format::FormatOption>, cpp_utils::AsStruct)) -> ::surface_format::SurfaceFormat
2) fn new((qt_core::flags::Flags<::surface_format::FormatOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
C++ method: [constructor] void QSurfaceFormat::QSurfaceFormat(QFlags<QSurfaceFormat::FormatOption> options)

Variant 3

Rust arguments:
1) fn new((&::surface_format::SurfaceFormat, cpp_utils::AsStruct)) -> ::surface_format::SurfaceFormat
2) fn new((&::surface_format::SurfaceFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::surface_format::SurfaceFormat>
C++ method: [constructor] void QSurfaceFormat::QSurfaceFormat(const QSurfaceFormat& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 SurfaceFormat) -> &'l0 mut SurfaceFormat

C++ method: QSurfaceFormat& QSurfaceFormat::operator=(const QSurfaceFormat& other)

C++ documentation:

Assigns other to this object.

fn options(&self) -> Flags<FormatOption>

C++ method: QFlags<QSurfaceFormat::FormatOption> QSurfaceFormat::options() const

C++ documentation:

Returns the currently set format options.

This function was introduced in Qt 5.3.

See also setOption(), setOptions(), and testOption().

fn profile(&self) -> OpenGLContextProfile

C++ method: QSurfaceFormat::OpenGLContextProfile QSurfaceFormat::profile() const

C++ documentation:

Get the configured OpenGL context profile.

This setting is ignored if the requested OpenGL version is less than 3.2.

See also setProfile().

fn red_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::redBufferSize() const

C++ documentation:

Get the size in bits of the red channel of the color buffer.

See also setRedBufferSize().

fn renderable_type(&self) -> RenderableType

C++ method: QSurfaceFormat::RenderableType QSurfaceFormat::renderableType() const

C++ documentation:

Gets the renderable type.

Chooses between desktop OpenGL, OpenGL ES, and OpenVG.

See also setRenderableType().

fn samples(&self) -> c_int

C++ method: int QSurfaceFormat::samples() const

C++ documentation:

Returns the number of samples per pixel when multisampling is enabled. By default, multisampling is disabled.

See also setSamples().

fn set_alpha_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setAlphaBufferSize(int size)

C++ documentation:

Set the desired size in bits of the alpha channel of the color buffer.

See also alphaBufferSize().

fn set_blue_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setBlueBufferSize(int size)

C++ documentation:

Set the desired size in bits of the blue channel of the color buffer.

Note: On Mac OSX, be sure to set the buffer size of all color channels, otherwise this setting will have no effect. If one of the buffer sizes is not set, the current bit-depth of the screen is used.

See also blueBufferSize().

fn set_default_format(format: &SurfaceFormat)

C++ method: static void QSurfaceFormat::setDefaultFormat(const QSurfaceFormat& format)

C++ documentation:

Sets the global default surface format.

This format is used by default in QOpenGLContext, QWindow, QOpenGLWidget and similar classes.

It can always be overridden on a per-instance basis by using the class in question's own setFormat() function. However, it is often more convenient to set the format for all windows once at the start of the application. It also guarantees proper behavior in cases where shared contexts are required, because settings the format via this function guarantees that all contexts and surfaces, even the ones created internally by Qt, will use the same format.

Note: When setting Qt::AA_ShareOpenGLContexts, it is strongly recommended to place the call to this function before the construction of the QGuiApplication or QApplication. Otherwise format will not be applied to the global share context and therefore issues may arise with context sharing afterwards.

This function was introduced in Qt 5.4.

See also defaultFormat().

fn set_depth_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setDepthBufferSize(int size)

C++ documentation:

Set the minimum depth buffer size to size.

See also depthBufferSize().

fn set_green_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setGreenBufferSize(int size)

C++ documentation:

Set the desired size in bits of the green channel of the color buffer.

Note: On Mac OSX, be sure to set the buffer size of all color channels, otherwise this setting will have no effect. If one of the buffer sizes is not set, the current bit-depth of the screen is used.

See also greenBufferSize().

fn set_major_version(&mut self, major_version: c_int)

C++ method: void QSurfaceFormat::setMajorVersion(int majorVersion)

C++ documentation:

Sets the desired major OpenGL version.

See also majorVersion().

fn set_minor_version(&mut self, minor_version: c_int)

C++ method: void QSurfaceFormat::setMinorVersion(int minorVersion)

C++ documentation:

Sets the desired minor OpenGL version.

The default version is 2.0.

See also minorVersion().

fn set_option<'largs, Args>(&'largs mut self, args: Args) where Args: SurfaceFormatSetOptionArgs<'largs>

C++ method: QSurfaceFormat::setOption

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn set_option(&mut self, ::surface_format::FormatOption) -> ()
2) fn set_option(&mut self, (::surface_format::FormatOption, bool)) -> ()
C++ method: void QSurfaceFormat::setOption(QSurfaceFormat::FormatOption option, bool on = ?)

C++ documentation:

Sets the format option option if on is true; otherwise, clears the option.

This function was introduced in Qt 5.3.

See also setOptions(), options(), and testOption().

Variant 2

Rust arguments: fn set_option(&mut self, qt_core::flags::Flags<::surface_format::FormatOption>) -> ()
C++ method: void QSurfaceFormat::setOption(QFlags<QSurfaceFormat::FormatOption> opt)

fn set_options(&mut self, options: Flags<FormatOption>)

C++ method: void QSurfaceFormat::setOptions(QFlags<QSurfaceFormat::FormatOption> options)

C++ documentation:

Sets the format options to options.

This function was introduced in Qt 5.3.

See also options() and testOption().

fn set_profile(&mut self, profile: OpenGLContextProfile)

C++ method: void QSurfaceFormat::setProfile(QSurfaceFormat::OpenGLContextProfile profile)

C++ documentation:

Sets the desired OpenGL context profile.

This setting is ignored if the requested OpenGL version is less than 3.2.

See also profile().

fn set_red_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setRedBufferSize(int size)

C++ documentation:

Set the desired size in bits of the red channel of the color buffer.

Note: On Mac OSX, be sure to set the buffer size of all color channels, otherwise this setting will have no effect. If one of the buffer sizes is not set, the current bit-depth of the screen is used.

See also redBufferSize().

fn set_renderable_type(&mut self, type_: RenderableType)

C++ method: void QSurfaceFormat::setRenderableType(QSurfaceFormat::RenderableType type)

C++ documentation:

Sets the desired renderable type.

Chooses between desktop OpenGL, OpenGL ES, and OpenVG.

See also renderableType().

fn set_samples(&mut self, num_samples: c_int)

C++ method: void QSurfaceFormat::setSamples(int numSamples)

C++ documentation:

Set the preferred number of samples per pixel when multisampling is enabled to numSamples. By default, multisampling is disabled.

See also samples().

fn set_stencil_buffer_size(&mut self, size: c_int)

C++ method: void QSurfaceFormat::setStencilBufferSize(int size)

C++ documentation:

Set the preferred stencil buffer size to size bits.

See also stencilBufferSize().

fn set_stereo(&mut self, enable: bool)

C++ method: void QSurfaceFormat::setStereo(bool enable)

C++ documentation:

If enable is true enables stereo buffering; otherwise disables stereo buffering.

Stereo buffering is disabled by default.

Stereo buffering provides extra color buffers to generate left-eye and right-eye images.

See also stereo().

fn set_swap_behavior(&mut self, behavior: SwapBehavior)

C++ method: void QSurfaceFormat::setSwapBehavior(QSurfaceFormat::SwapBehavior behavior)

C++ documentation:

Set the swap behavior of the surface.

The swap behavior specifies whether single, double, or triple buffering is desired. The default, DefaultSwapBehavior, gives the default swap behavior of the platform.

See also swapBehavior().

fn set_swap_interval(&mut self, interval: c_int)

C++ method: void QSurfaceFormat::setSwapInterval(int interval)

C++ documentation:

Sets the preferred swap interval. The swap interval specifies the minimum number of video frames that are displayed before a buffer swap occurs. This can be used to sync the GL drawing into a window to the vertical refresh of the screen.

Setting an interval value of 0 will turn the vertical refresh syncing off, any value higher than 0 will turn the vertical syncing on. Setting interval to a higher value, for example 10, results in having 10 vertical retraces between every buffer swap.

The default interval is 1.

Changing the swap interval may not be supported by the underlying platform. In this case, the request will be silently ignored.

This function was introduced in Qt 5.3.

See also swapInterval().

fn set_version(&mut self, major: c_int, minor: c_int)

C++ method: void QSurfaceFormat::setVersion(int major, int minor)

C++ documentation:

Sets the desired major and minor OpenGL versions.

The default version is 2.0.

See also version().

fn stencil_buffer_size(&self) -> c_int

C++ method: int QSurfaceFormat::stencilBufferSize() const

C++ documentation:

Returns the stencil buffer size in bits.

See also setStencilBufferSize().

fn stereo(&self) -> bool

C++ method: bool QSurfaceFormat::stereo() const

C++ documentation:

Returns true if stereo buffering is enabled; otherwise returns false. Stereo buffering is disabled by default.

See also setStereo().

fn swap_behavior(&self) -> SwapBehavior

C++ method: QSurfaceFormat::SwapBehavior QSurfaceFormat::swapBehavior() const

C++ documentation:

Returns the configured swap behaviour.

See also setSwapBehavior().

fn swap_interval(&self) -> c_int

C++ method: int QSurfaceFormat::swapInterval() const

C++ documentation:

Returns the swap interval.

This function was introduced in Qt 5.3.

See also setSwapInterval().

fn test_option<'largs, Args>(&'largs self, args: Args) -> bool where Args: SurfaceFormatTestOptionArgs<'largs>

C++ method: QSurfaceFormat::testOption

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn test_option(&self, ::surface_format::FormatOption) -> bool
C++ method: bool QSurfaceFormat::testOption(QSurfaceFormat::FormatOption option) const

C++ documentation:

Returns true if the format option option is set; otherwise returns false.

This function was introduced in Qt 5.3.

See also options().

Variant 2

Rust arguments: fn test_option(&self, qt_core::flags::Flags<::surface_format::FormatOption>) -> bool
C++ method: bool QSurfaceFormat::testOption(QFlags<QSurfaceFormat::FormatOption> opt) const

fn version<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SurfaceFormatVersionArgs<'largs>

C++ method: QSurfaceFormat::version

This is an overloaded function. Available variants:

Rust arguments:
1) fn version(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pair::PairCIntCInt>
2) fn version(&self, cpp_utils::AsStruct) -> ::pair::PairCIntCInt
C++ method: QPair<int, int> QSurfaceFormat::version() const

C++ documentation:

Returns a QPair<int, int> representing the OpenGL version.

Useful for version checks, for example format.version() >= qMakePair(3, 2)

See also setVersion().