Module qt_gui::syntax_highlighter::overloading

Traits

SyntaxHighlighterConnectArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::connect method.

SyntaxHighlighterConnectStaticArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::connect_static method.

SyntaxHighlighterDestroyedArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::destroyed method.

SyntaxHighlighterDisconnectArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::disconnect method.

SyntaxHighlighterDisconnectStaticArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::disconnect_static method.

SyntaxHighlighterDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::dynamic_property_names method.

SyntaxHighlighterObjectNameArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::object_name method.

SyntaxHighlighterPropertyArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::property method.

SyntaxHighlighterStartTimerArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::start_timer method.

SyntaxHighlighterTrArgs
This trait represents a set of arguments accepted by SyntaxHighlighter::tr method.

SyntaxHighlighterTrUtf8Args
This trait represents a set of arguments accepted by SyntaxHighlighter::tr_utf8 method.