Struct qt_gui::text_block::Iterator

pub struct Iterator {
    // some fields omitted
}

C++ type: QTextBlock::iterator

C++ documentation:

The QTextBlock::iterator class provides an iterator for reading the contents of a QTextBlock.

A block consists of a sequence of text fragments. This class provides a way to iterate over these, and read their contents. It does not provide a way to modify the internal structure or contents of the block.

An iterator can be constructed and used to access the fragments within a text block in the following way:

    QTextBlock::iterator it;
    for (it = currentBlock.begin(); !(it.atEnd()); ++it) {
        QTextFragment currentFragment = it.fragment();
        if (currentFragment.isValid())
            processFragment(currentFragment);
    }

Methods

impl Iterator

fn at_end(&self) -> bool

C++ method: bool QTextBlock::iterator::atEnd() const

C++ documentation:

Returns true if the current item is the last item in the text block.

fn fragment<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorFragmentArgs<'largs>

C++ method: QTextBlock::iterator::fragment

This is an overloaded function. Available variants:

Rust arguments:
1) fn fragment(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_fragment::TextFragment>
2) fn fragment(&self, cpp_utils::AsStruct) -> ::text_fragment::TextFragment
C++ method: QTextFragment QTextBlock::iterator::fragment() const

C++ documentation:

Returns the text fragment the iterator currently points to.

fn new<Args>(args: Args) -> Args::ReturnType where Args: IteratorNewArgs

C++ method: QTextBlock::iterator::iterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_block::Iterator
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block::Iterator>
C++ method: [constructor] void QTextBlock::iterator::iterator()

Variant 2

Rust arguments:
1) fn new((&::text_block::Iterator, cpp_utils::AsStruct)) -> ::text_block::Iterator
2) fn new((&::text_block::Iterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_block::Iterator>
C++ method: [constructor] void QTextBlock::iterator::iterator(const QTextBlock::iterator& o)

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QTextBlock::iterator& QTextBlock::iterator::operator--()

C++ documentation:

The prefix -- operator (--i) makes the preceding item current and returns an iterator pointing to the new current item.

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpDecPostfixArgs<'largs>

C++ method: QTextBlock::iterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_block::Iterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_block::Iterator
C++ method: QTextBlock::iterator QTextBlock::iterator::operator--(int arg1)

C++ documentation:

The postfix -- operator (i--) makes the preceding item current and returns an iterator to the old current item.

fn op_eq(&self, o: &Iterator) -> bool

C++ method: bool QTextBlock::iterator::operator==(const QTextBlock::iterator& o) const

C++ documentation:

Retuns true if this iterator is the same as the other iterator; otherwise returns false.

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QTextBlock::iterator& QTextBlock::iterator::operator++()

C++ documentation:

The prefix ++ operator (++i) advances the iterator to the next item in the hash and returns an iterator to the new current item.

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpIncPostfixArgs<'largs>

C++ method: QTextBlock::iterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_block::Iterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_block::Iterator
C++ method: QTextBlock::iterator QTextBlock::iterator::operator++(int arg1)

C++ documentation:

The postfix ++ operator (i++) advances the iterator to the next item in the text block and returns an iterator to the old current item.

fn op_neq(&self, o: &Iterator) -> bool

C++ method: bool QTextBlock::iterator::operator!=(const QTextBlock::iterator& o) const

C++ documentation:

Retuns true if this iterator is different from the other iterator; otherwise returns false.