Module qt_gui::text_block_format::overloading

Traits

TextBlockFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextBlockFormat::as_q_variant method.

TextBlockFormatBackgroundArgs
This trait represents a set of arguments accepted by TextBlockFormat::background method.

TextBlockFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::brush_property method.

TextBlockFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::color_property method.

TextBlockFormatForegroundArgs
This trait represents a set of arguments accepted by TextBlockFormat::foreground method.

TextBlockFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::length_property method.

TextBlockFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::length_vector_property method.

TextBlockFormatLineHeightArgs
This trait represents a set of arguments accepted by TextBlockFormat::line_height method.

TextBlockFormatNewArgs
This trait represents a set of arguments accepted by TextBlockFormat::new method.

TextBlockFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::pen_property method.

TextBlockFormatPropertiesArgs
This trait represents a set of arguments accepted by TextBlockFormat::properties method.

TextBlockFormatPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::property method.

TextBlockFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::set_property method.

TextBlockFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextBlockFormat::string_property method.

TextBlockFormatTabPositionsArgs
This trait represents a set of arguments accepted by TextBlockFormat::tab_positions method.

TextBlockFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_block_format method.

TextBlockFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_char_format method.

TextBlockFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_frame_format method.

TextBlockFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_image_format method.

TextBlockFormatToListFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_list_format method.

TextBlockFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_table_cell_format method.

TextBlockFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextBlockFormat::to_table_format method.