Struct qt_gui::text_block_format::TextBlockFormat

pub struct TextBlockFormat {
    // some fields omitted
}

C++ type: QTextBlockFormat

C++ documentation:

The QTextBlockFormat class provides formatting information for blocks of text in a QTextDocument.

A document is composed of a list of blocks, represented by QTextBlock objects. Each block can contain an item of some kind, such as a paragraph of text, a table, a list, or an image. Every block has an associated QTextBlockFormat that specifies its characteristics.

To cater for left-to-right and right-to-left languages you can set a block's direction with setDirection(). Paragraph alignment is set with setAlignment(). Margins are controlled by setTopMargin(), setBottomMargin(), setLeftMargin(), setRightMargin(). Overall indentation is set with setIndent(), the indentation of the first line with setTextIndent().

Line spacing is set with setLineHeight() and retrieved via lineHeight() and lineHeightType(). The types of line spacing available are in the LineHeightTypes enum.

Line breaking can be enabled and disabled with setNonBreakableLines().

The brush used to paint the paragraph's background is set with setBackground(), and other aspects of the text's appearance can be customized by using the setProperty() function with the OutlinePen, ForegroundBrush, and BackgroundBrush QTextFormat::Property values.

If a text block is part of a list, it can also have a list format that is accessible with the listFormat() function.

Methods

impl TextBlockFormat

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QTextBlockFormat::alignment() const

C++ documentation:

Returns the paragraph's alignment.

See also setAlignment().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatAsQVariantArgs<'largs>

C++ method: QTextBlockFormat::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTextBlockFormat::operator QVariant() const

C++ documentation:

Returns the text format as a QVariant

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatBackgroundArgs<'largs>

C++ method: QTextBlockFormat::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextBlockFormat::background() const

C++ documentation:

Returns the brush used to paint the document's background.

See also setBackground(), clearBackground(), and foreground().

fn bool_property(&self, property_id: c_int) -> bool

C++ method: bool QTextBlockFormat::boolProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QTextFormat::Bool type, false is returned instead.

See also setProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn bottom_margin(&self) -> c_double

C++ method: double QTextBlockFormat::bottomMargin() const

C++ documentation:

Returns the paragraph's bottom margin.

See also setBottomMargin() and topMargin().

fn brush_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatBrushPropertyArgs<'largs>

C++ method: QTextBlockFormat::brushProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::brush::Brush>
2) fn brush_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::brush::Brush
C++ method: QBrush QTextBlockFormat::brushProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Brush type, Qt::NoBrush is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn clear_background(&mut self)

C++ method: void QTextBlockFormat::clearBackground()

C++ documentation:

Clears the brush used to paint the document's background. The default brush will be used.

See also background(), setBackground(), and clearForeground().

fn clear_foreground(&mut self)

C++ method: void QTextBlockFormat::clearForeground()

C++ documentation:

Clears the brush used to paint the document's foreground. The default brush will be used.

See also foreground(), setForeground(), and clearBackground().

fn clear_property(&mut self, property_id: c_int)

C++ method: void QTextBlockFormat::clearProperty(int propertyId)

C++ documentation:

Clears the value of the property given by propertyId

See also Property.

fn color_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatColorPropertyArgs<'largs>

C++ method: QTextBlockFormat::colorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn color_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::color::Color>
2) fn color_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::color::Color
C++ method: QColor QTextBlockFormat::colorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Color type, an invalid color is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn double_property(&self, property_id: c_int) -> c_double

C++ method: double QTextBlockFormat::doubleProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QVariant::Double or QMetaType::Float type, 0 is returned instead.

See also setProperty(), boolProperty(), intProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatForegroundArgs<'largs>

C++ method: QTextBlockFormat::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextBlockFormat::foreground() const

C++ documentation:

Returns the brush used to render foreground details, such as text, frame outlines, and table borders.

See also setForeground(), clearForeground(), and background().

fn has_property(&self, property_id: c_int) -> bool

C++ method: bool QTextBlockFormat::hasProperty(int propertyId) const

C++ documentation:

Returns true if the text format has a property with the given propertyId; otherwise returns false.

See also properties() and Property.

fn indent(&self) -> c_int

C++ method: int QTextBlockFormat::indent() const

C++ documentation:

Returns the paragraph's indent.

See also setIndent().

fn int_property(&self, property_id: c_int) -> c_int

C++ method: int QTextBlockFormat::intProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property is not of QTextFormat::Integer type, 0 is returned instead.

See also setProperty(), boolProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn is_block_format(&self) -> bool

C++ method: bool QTextBlockFormat::isBlockFormat() const

C++ documentation:

Returns true if this text format is a BlockFormat; otherwise returns false.

fn is_char_format(&self) -> bool

C++ method: bool QTextBlockFormat::isCharFormat() const

C++ documentation:

Returns true if this text format is a CharFormat; otherwise returns false.

fn is_empty(&self) -> bool

C++ method: bool QTextBlockFormat::isEmpty() const

C++ documentation:

Returns true if the format does not store any properties; false otherwise.

This function was introduced in Qt 5.3.

See also propertyCount() and properties().

fn is_frame_format(&self) -> bool

C++ method: bool QTextBlockFormat::isFrameFormat() const

C++ documentation:

Returns true if this text format is a FrameFormat; otherwise returns false.

fn is_image_format(&self) -> bool

C++ method: bool QTextBlockFormat::isImageFormat() const

C++ documentation:

Returns true if this text format is an image format; otherwise returns false.

fn is_list_format(&self) -> bool

C++ method: bool QTextBlockFormat::isListFormat() const

C++ documentation:

Returns true if this text format is a ListFormat; otherwise returns false.

fn is_table_cell_format(&self) -> bool

C++ method: bool QTextBlockFormat::isTableCellFormat() const

C++ documentation:

Returns true if this text format is a TableCellFormat; otherwise returns false.

This function was introduced in Qt 4.4.

fn is_table_format(&self) -> bool

C++ method: bool QTextBlockFormat::isTableFormat() const

C++ documentation:

Returns true if this text format is a TableFormat; otherwise returns false.

fn is_valid(&self) -> bool

C++ method: bool QTextBlockFormat::isValid() const

C++ documentation:

Returns true if this block format is valid; otherwise returns false.

fn layout_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextBlockFormat::layoutDirection() const

C++ documentation:

Returns the document's layout direction.

See also setLayoutDirection().

fn left_margin(&self) -> c_double

C++ method: double QTextBlockFormat::leftMargin() const

C++ documentation:

Returns the paragraph's left margin.

See also setLeftMargin(), rightMargin(), and indent().

fn length_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatLengthPropertyArgs<'largs>

C++ method: QTextBlockFormat::lengthProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_length::TextLength>
2) fn length_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_length::TextLength
C++ method: QTextLength QTextBlockFormat::lengthProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthVectorProperty(), and Property.

fn length_vector_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatLengthVectorPropertyArgs<'largs>

C++ method: QTextBlockFormat::lengthVectorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorTextLengthRef>
2) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorTextLengthRef
C++ method: QVector<QTextLength> QTextBlockFormat::lengthVectorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId. If the property isn't of QTextFormat::LengthVector type, an empty length vector is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), and Property.

fn line_height<'largs, Args>(&'largs self, args: Args) -> c_double where Args: TextBlockFormatLineHeightArgs<'largs>

C++ method: QTextBlockFormat::lineHeight

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn line_height(&self, ()) -> libc::c_double
C++ method: double QTextBlockFormat::lineHeight() const

C++ documentation:

This returns the LineHeight property for the paragraph.

This function was introduced in Qt 4.8.

See also LineHeightTypes, setLineHeight(), and lineHeightType().

Variant 2

Rust arguments: fn line_height(&self, (libc::c_double, libc::c_double)) -> libc::c_double
C++ method: double QTextBlockFormat::lineHeight(double scriptLineHeight, double scaling) const

C++ documentation:

Returns the height of the lines in the paragraph based on the height of the script line given by scriptLineHeight and the specified scaling factor.

The value that is returned is also dependent on the given LineHeightType of the paragraph as well as the LineHeight setting that has been set for the paragraph.

The scaling is needed for heights that include a fixed number of pixels, to scale them appropriately for printing.

This function was introduced in Qt 4.8.

See also LineHeightTypes, setLineHeight(), and lineHeightType().

fn line_height_type(&self) -> c_int

C++ method: int QTextBlockFormat::lineHeightType() const

C++ documentation:

This returns the LineHeightType property of the paragraph.

This function was introduced in Qt 4.8.

See also LineHeightTypes, setLineHeight(), and lineHeight().

fn merge(&mut self, other: &TextFormat)

C++ method: void QTextBlockFormat::merge(const QTextFormat& other)

C++ documentation:

Merges the other format with this format; where there are conflicts the other format takes precedence.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextBlockFormatNewArgs

C++ method: QTextBlockFormat::QTextBlockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
C++ method: [constructor] void QTextBlockFormat::QTextBlockFormat()

fn non_breakable_lines(&self) -> bool

C++ method: bool QTextBlockFormat::nonBreakableLines() const

C++ documentation:

Returns true if the lines in the paragraph are non-breakable; otherwise returns false.

See also setNonBreakableLines().

fn object_index(&self) -> c_int

C++ method: int QTextBlockFormat::objectIndex() const

C++ documentation:

Returns the index of the format object, or -1 if the format object is invalid.

See also setObjectIndex().

fn object_type(&self) -> c_int

C++ method: int QTextBlockFormat::objectType() const

C++ documentation:

Returns the text format's object type.

See also ObjectTypes and setObjectType().

fn op_eq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextBlockFormat::operator==(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is the same as the other text format.

fn op_neq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextBlockFormat::operator!=(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is different from the other text format.

fn page_break_policy(&self) -> Flags<PageBreakFlag>

C++ method: QFlags<QTextFormat::PageBreakFlag> QTextBlockFormat::pageBreakPolicy() const

C++ documentation:

Returns the currently set page break policy for the paragraph. The default is QTextFormat::PageBreak_Auto.

This function was introduced in Qt 4.2.

See also setPageBreakPolicy().

fn pen_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatPenPropertyArgs<'largs>

C++ method: QTextBlockFormat::penProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn pen_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pen::Pen>
2) fn pen_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::pen::Pen
C++ method: QPen QTextBlockFormat::penProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Pen type, Qt::NoPen is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn properties<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatPropertiesArgs<'largs>

C++ method: QTextBlockFormat::properties

This is an overloaded function. Available variants:

Rust arguments:
1) fn properties(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::map::MapCIntVariantRef>
2) fn properties(&self, cpp_utils::AsStruct) -> qt_core::map::MapCIntVariantRef
C++ method: QMap<int, QVariant> QTextBlockFormat::properties() const

C++ documentation:

Returns a map with all properties of this text format.

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatPropertyArgs<'largs>

C++ method: QTextBlockFormat::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QTextBlockFormat::property(int propertyId) const

C++ documentation:

Returns the property specified by the given propertyId.

See also setProperty() and Property.

fn property_count(&self) -> c_int

C++ method: int QTextBlockFormat::propertyCount() const

C++ documentation:

Returns the number of properties stored in the format.

This function was introduced in Qt 4.3.

fn right_margin(&self) -> c_double

C++ method: double QTextBlockFormat::rightMargin() const

C++ documentation:

Returns the paragraph's right margin.

See also setRightMargin() and leftMargin().

fn set_alignment(&mut self, alignment: Flags<AlignmentFlag>)

C++ method: void QTextBlockFormat::setAlignment(QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the paragraph's alignment.

See also alignment().

fn set_background(&mut self, brush: &Brush)

C++ method: void QTextBlockFormat::setBackground(const QBrush& brush)

C++ documentation:

Sets the brush use to paint the document's background to the brush specified.

See also background(), clearBackground(), and setForeground().

fn set_bottom_margin(&mut self, margin: c_double)

C++ method: void QTextBlockFormat::setBottomMargin(double margin)

C++ documentation:

Sets the paragraph's bottom margin.

See also bottomMargin(), setTopMargin(), setLeftMargin(), and setRightMargin().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QTextBlockFormat::setForeground(const QBrush& brush)

C++ documentation:

Sets the foreground brush to the specified brush. The foreground brush is mostly used to render text.

See also foreground(), clearForeground(), and setBackground().

fn set_indent(&mut self, indent: c_int)

C++ method: void QTextBlockFormat::setIndent(int indent)

C++ documentation:

Sets the paragraph's indentation. Margins are set independently of indentation with setLeftMargin() and setTextIndent(). The indentation is an integer that is multiplied with the document-wide standard indent, resulting in the actual indent of the paragraph.

See also indent() and QTextDocument::indentWidth().

fn set_layout_direction(&mut self, direction: LayoutDirection)

C++ method: void QTextBlockFormat::setLayoutDirection(Qt::LayoutDirection direction)

C++ documentation:

Sets the document's layout direction to the specified direction.

See also layoutDirection().

fn set_left_margin(&mut self, margin: c_double)

C++ method: void QTextBlockFormat::setLeftMargin(double margin)

C++ documentation:

Sets the paragraph's left margin. Indentation can be applied separately with setIndent().

See also leftMargin(), setRightMargin(), setTopMargin(), and setBottomMargin().

fn set_line_height(&mut self, height: c_double, height_type: c_int)

C++ method: void QTextBlockFormat::setLineHeight(double height, int heightType)

C++ documentation:

Sets the line height for the paragraph to the value given by height which is dependent on heightType in the way described by the LineHeightTypes enum.

This function was introduced in Qt 4.8.

See also LineHeightTypes, lineHeight(), and lineHeightType().

fn set_non_breakable_lines(&mut self, b: bool)

C++ method: void QTextBlockFormat::setNonBreakableLines(bool b)

C++ documentation:

If b is true, the lines in the paragraph are treated as non-breakable; otherwise they are breakable.

See also nonBreakableLines().

fn set_object_index(&mut self, object: c_int)

C++ method: void QTextBlockFormat::setObjectIndex(int object)

C++ documentation:

Sets the format object's object index.

See also objectIndex().

fn set_object_type(&mut self, type_: c_int)

C++ method: void QTextBlockFormat::setObjectType(int type)

C++ documentation:

Sets the text format's object type to type.

See also ObjectTypes and objectType().

fn set_page_break_policy(&mut self, flags: Flags<PageBreakFlag>)

C++ method: void QTextBlockFormat::setPageBreakPolicy(QFlags<QTextFormat::PageBreakFlag> flags)

C++ documentation:

Sets the page break policy for the paragraph to policy.

This function was introduced in Qt 4.2.

See also pageBreakPolicy().

fn set_property<'largs, Args>(&'largs mut self, args: Args) where Args: TextBlockFormatSetPropertyArgs<'largs>

C++ method: QTextBlockFormat::setProperty

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_property(&mut self, (libc::c_int, &qt_core::variant::Variant)) -> ()
C++ method: void QTextBlockFormat::setProperty(int propertyId, const QVariant& value)

C++ documentation:

Sets the property specified by the propertyId to the given value.

See also property() and Property.

Variant 2

Rust arguments: fn set_property(&mut self, (libc::c_int, &::vector::VectorTextLengthRef)) -> ()
C++ method: void QTextBlockFormat::setProperty(int propertyId, const QVector<QTextLength>& lengths)

C++ documentation:

Sets the value of the property given by propertyId to value.

See also lengthVectorProperty() and Property.

fn set_right_margin(&mut self, margin: c_double)

C++ method: void QTextBlockFormat::setRightMargin(double margin)

C++ documentation:

Sets the paragraph's right margin.

See also rightMargin(), setLeftMargin(), setTopMargin(), and setBottomMargin().

fn set_tab_positions(&mut self, tabs: &ListTabRef)

C++ method: void QTextBlockFormat::setTabPositions(const QList<QTextOption::Tab>& tabs)

C++ documentation:

Sets the tab positions for the text block to those specified by tabs.

This function was introduced in Qt 4.4.

See also tabPositions().

fn set_text_indent(&mut self, aindent: c_double)

C++ method: void QTextBlockFormat::setTextIndent(double aindent)

C++ documentation:

Sets the indent for the first line in the block. This allows the first line of a paragraph to be indented differently to the other lines, enhancing the readability of the text.

See also textIndent(), setLeftMargin(), setRightMargin(), setTopMargin(), and setBottomMargin().

fn set_top_margin(&mut self, margin: c_double)

C++ method: void QTextBlockFormat::setTopMargin(double margin)

C++ documentation:

Sets the paragraph's top margin.

See also topMargin(), setBottomMargin(), setLeftMargin(), and setRightMargin().

fn string_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatStringPropertyArgs<'largs>

C++ method: QTextBlockFormat::stringProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn string_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn string_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTextBlockFormat::stringProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::String type, an empty string is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn swap(&mut self, other: &mut TextFormat)

C++ method: void QTextBlockFormat::swap(QTextFormat& other)

C++ documentation:

Swaps this text format with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn tab_positions<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatTabPositionsArgs<'largs>

C++ method: QTextBlockFormat::tabPositions

This is an overloaded function. Available variants:

Rust arguments:
1) fn tab_positions(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListTabRef>
2) fn tab_positions(&self, cpp_utils::AsStruct) -> ::list::ListTabRef
C++ method: QList<QTextOption::Tab> QTextBlockFormat::tabPositions() const

C++ documentation:

Returns a list of tab positions defined for the text block.

This function was introduced in Qt 4.4.

See also setTabPositions().

fn text_indent(&self) -> c_double

C++ method: double QTextBlockFormat::textIndent() const

C++ documentation:

Returns the paragraph's text indent.

See also setTextIndent().

fn to_block_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToBlockFormatArgs<'largs>

C++ method: QTextBlockFormat::toBlockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_block_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
2) fn to_block_format(&self, cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
C++ method: QTextBlockFormat QTextBlockFormat::toBlockFormat() const

C++ documentation:

Returns this format as a block format.

fn to_char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToCharFormatArgs<'largs>

C++ method: QTextBlockFormat::toCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn to_char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextBlockFormat::toCharFormat() const

C++ documentation:

Returns this format as a character format.

fn to_frame_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToFrameFormatArgs<'largs>

C++ method: QTextBlockFormat::toFrameFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_frame_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame_format::TextFrameFormat>
2) fn to_frame_format(&self, cpp_utils::AsStruct) -> ::text_frame_format::TextFrameFormat
C++ method: QTextFrameFormat QTextBlockFormat::toFrameFormat() const

C++ documentation:

Returns this format as a frame format.

fn to_image_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToImageFormatArgs<'largs>

C++ method: QTextBlockFormat::toImageFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_image_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_image_format::TextImageFormat>
2) fn to_image_format(&self, cpp_utils::AsStruct) -> ::text_image_format::TextImageFormat
C++ method: QTextImageFormat QTextBlockFormat::toImageFormat() const

C++ documentation:

Returns this format as an image format.

fn to_list_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToListFormatArgs<'largs>

C++ method: QTextBlockFormat::toListFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_list_format::TextListFormat>
2) fn to_list_format(&self, cpp_utils::AsStruct) -> ::text_list_format::TextListFormat
C++ method: QTextListFormat QTextBlockFormat::toListFormat() const

C++ documentation:

Returns this format as a list format.

fn to_table_cell_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToTableCellFormatArgs<'largs>

C++ method: QTextBlockFormat::toTableCellFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_cell_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_cell_format::TextTableCellFormat>
2) fn to_table_cell_format(&self, cpp_utils::AsStruct) -> ::text_table_cell_format::TextTableCellFormat
C++ method: QTextTableCellFormat QTextBlockFormat::toTableCellFormat() const

C++ documentation:

Returns this format as a table cell format.

This function was introduced in Qt 4.4.

fn to_table_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBlockFormatToTableFormatArgs<'largs>

C++ method: QTextBlockFormat::toTableFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_format::TextTableFormat>
2) fn to_table_format(&self, cpp_utils::AsStruct) -> ::text_table_format::TextTableFormat
C++ method: QTextTableFormat QTextBlockFormat::toTableFormat() const

C++ documentation:

Returns this format as a table format.

fn top_margin(&self) -> c_double

C++ method: double QTextBlockFormat::topMargin() const

C++ documentation:

Returns the paragraph's top margin.

See also setTopMargin() and bottomMargin().

fn type_(&self) -> c_int

C++ method: int QTextBlockFormat::type() const

C++ documentation:

Returns the type of this format.

See also FormatType.