Module qt_gui::text_block_group::overloading

Traits

TextBlockGroupConnectArgs
This trait represents a set of arguments accepted by TextBlockGroup::connect method.

TextBlockGroupConnectStaticArgs
This trait represents a set of arguments accepted by TextBlockGroup::connect_static method.

TextBlockGroupDestroyedArgs
This trait represents a set of arguments accepted by TextBlockGroup::destroyed method.

TextBlockGroupDisconnectArgs
This trait represents a set of arguments accepted by TextBlockGroup::disconnect method.

TextBlockGroupDisconnectStaticArgs
This trait represents a set of arguments accepted by TextBlockGroup::disconnect_static method.

TextBlockGroupDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TextBlockGroup::dynamic_property_names method.

TextBlockGroupFormatArgs
This trait represents a set of arguments accepted by TextBlockGroup::format method.

TextBlockGroupObjectNameArgs
This trait represents a set of arguments accepted by TextBlockGroup::object_name method.

TextBlockGroupPropertyArgs
This trait represents a set of arguments accepted by TextBlockGroup::property method.

TextBlockGroupStartTimerArgs
This trait represents a set of arguments accepted by TextBlockGroup::start_timer method.

TextBlockGroupTrArgs
This trait represents a set of arguments accepted by TextBlockGroup::tr method.

TextBlockGroupTrUtf8Args
This trait represents a set of arguments accepted by TextBlockGroup::tr_utf8 method.