Struct qt_gui::text_block_user_data::TextBlockUserData

pub struct TextBlockUserData {
    // some fields omitted
}

C++ type: QTextBlockUserData

C++ documentation:

The QTextBlockUserData class is used to associate custom data with blocks of text.

QTextBlockUserData provides an abstract interface for container classes that are used to associate application-specific user data with text blocks in a QTextDocument.

Generally, subclasses of this class provide functions to allow data to be stored and retrieved, and instances are attached to blocks of text using QTextBlock::setUserData(). This makes it possible to store additional data per text block in a way that can be retrieved safely by the application.

Each subclass should provide a reimplementation of the destructor to ensure that any private data is automatically cleaned up when user data objects are deleted.