Struct qt_gui::text_char_format::TextCharFormat

pub struct TextCharFormat {
    // some fields omitted
}

C++ type: QTextCharFormat

C++ documentation:

The QTextCharFormat class provides formatting information for characters in a QTextDocument.

The character format of text in a document specifies the visual properties of the text, as well as information about its role in a hypertext document.

The font used can be set by supplying a font to the setFont() function, and each aspect of its appearance can be adjusted to give the desired effect. setFontFamily() and setFontPointSize() define the font's family (e.g. Times) and printed size; setFontWeight() and setFontItalic() provide control over the style of the font. setFontUnderline(), setFontOverline(), setFontStrikeOut(), and setFontFixedPitch() provide additional effects for text.

The color is set with setForeground(). If the text is intended to be used as an anchor (for hyperlinks), this can be enabled with setAnchor(). The setAnchorHref() and setAnchorNames() functions are used to specify the information about the hyperlink's destination and the anchor's name.

Methods

impl TextCharFormat

fn anchor_href<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatAnchorHrefArgs<'largs>

C++ method: QTextCharFormat::anchorHref

This is an overloaded function. Available variants:

Rust arguments:
1) fn anchor_href(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn anchor_href(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextCharFormat::anchorHref() const

C++ documentation:

Returns the text format's hypertext link, or an empty string if none has been set.

See also setAnchorHref().

fn anchor_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatAnchorNameArgs<'largs>

C++ method: QTextCharFormat::anchorName

This is an overloaded function. Available variants:

Rust arguments:
1) fn anchor_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn anchor_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextCharFormat::anchorName() const

C++ documentation:

This function is deprecated. Use anchorNames() instead.

Returns the anchor name associated with this text format, or an empty string if none has been set. If the anchor name is set, text with this format can be the destination of a hypertext link.

See also setAnchorName().

fn anchor_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatAnchorNamesArgs<'largs>

C++ method: QTextCharFormat::anchorNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn anchor_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string_list::StringList>
2) fn anchor_names(&self, cpp_utils::AsStruct) -> qt_core::string_list::StringList
C++ method: QStringList QTextCharFormat::anchorNames() const

C++ documentation:

Returns the anchor names associated with this text format, or an empty string list if none has been set. If the anchor names are set, text with this format can be the destination of a hypertext link.

This function was introduced in Qt 4.3.

See also setAnchorNames().

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatAsQVariantArgs<'largs>

C++ method: QTextCharFormat::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTextCharFormat::operator QVariant() const

C++ documentation:

Returns the text format as a QVariant

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatBackgroundArgs<'largs>

C++ method: QTextCharFormat::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextCharFormat::background() const

C++ documentation:

Returns the brush used to paint the document's background.

See also setBackground(), clearBackground(), and foreground().

fn bool_property(&self, property_id: c_int) -> bool

C++ method: bool QTextCharFormat::boolProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QTextFormat::Bool type, false is returned instead.

See also setProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn brush_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatBrushPropertyArgs<'largs>

C++ method: QTextCharFormat::brushProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::brush::Brush>
2) fn brush_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::brush::Brush
C++ method: QBrush QTextCharFormat::brushProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Brush type, Qt::NoBrush is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn clear_background(&mut self)

C++ method: void QTextCharFormat::clearBackground()

C++ documentation:

Clears the brush used to paint the document's background. The default brush will be used.

See also background(), setBackground(), and clearForeground().

fn clear_foreground(&mut self)

C++ method: void QTextCharFormat::clearForeground()

C++ documentation:

Clears the brush used to paint the document's foreground. The default brush will be used.

See also foreground(), setForeground(), and clearBackground().

fn clear_property(&mut self, property_id: c_int)

C++ method: void QTextCharFormat::clearProperty(int propertyId)

C++ documentation:

Clears the value of the property given by propertyId

See also Property.

fn color_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatColorPropertyArgs<'largs>

C++ method: QTextCharFormat::colorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn color_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::color::Color>
2) fn color_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::color::Color
C++ method: QColor QTextCharFormat::colorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Color type, an invalid color is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn double_property(&self, property_id: c_int) -> c_double

C++ method: double QTextCharFormat::doubleProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QVariant::Double or QMetaType::Float type, 0 is returned instead.

See also setProperty(), boolProperty(), intProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatFontArgs<'largs>

C++ method: QTextCharFormat::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> ::font::Font
C++ method: QFont QTextCharFormat::font() const

C++ documentation:

Returns the font for this character format.

See also setFont().

fn font_capitalization(&self) -> Capitalization

C++ method: QFont::Capitalization QTextCharFormat::fontCapitalization() const

C++ documentation:

Returns the current capitalization type of the font.

This function was introduced in Qt 4.4.

See also setFontCapitalization().

fn font_family<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatFontFamilyArgs<'largs>

C++ method: QTextCharFormat::fontFamily

This is an overloaded function. Available variants:

Rust arguments:
1) fn font_family(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn font_family(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextCharFormat::fontFamily() const

C++ documentation:

Returns the text format's font family.

See also setFontFamily() and font().

fn font_fixed_pitch(&self) -> bool

C++ method: bool QTextCharFormat::fontFixedPitch() const

C++ documentation:

Returns true if the text format's font is fixed pitch; otherwise returns false.

See also setFontFixedPitch() and font().

fn font_hinting_preference(&self) -> HintingPreference

C++ method: QFont::HintingPreference QTextCharFormat::fontHintingPreference() const

C++ documentation:

Returns the hinting preference set for this text format.

This function was introduced in Qt 4.8.

See also setFontHintingPreference(), font(), and QFont::hintingPreference().

fn font_italic(&self) -> bool

C++ method: bool QTextCharFormat::fontItalic() const

C++ documentation:

Returns true if the text format's font is italic; otherwise returns false.

See also setFontItalic() and font().

fn font_kerning(&self) -> bool

C++ method: bool QTextCharFormat::fontKerning() const

C++ documentation:

Returns true if the font kerning is enabled.

This function was introduced in Qt 4.5.

See also setFontKerning() and font().

fn font_letter_spacing(&self) -> c_double

C++ method: double QTextCharFormat::fontLetterSpacing() const

C++ documentation:

Returns the current letter spacing.

This function was introduced in Qt 4.4.

See also setFontLetterSpacing(), setFontLetterSpacingType(), and fontLetterSpacingType().

fn font_letter_spacing_type(&self) -> SpacingType

C++ method: QFont::SpacingType QTextCharFormat::fontLetterSpacingType() const

C++ documentation:

Returns the letter spacing type of this format..

This function was introduced in Qt 5.0.

See also setFontLetterSpacingType(), setFontLetterSpacing(), and fontLetterSpacing().

fn font_overline(&self) -> bool

C++ method: bool QTextCharFormat::fontOverline() const

C++ documentation:

Returns true if the text format's font is overlined; otherwise returns false.

See also setFontOverline() and font().

fn font_point_size(&self) -> c_double

C++ method: double QTextCharFormat::fontPointSize() const

C++ documentation:

Returns the font size used to display text in this format.

See also setFontPointSize() and font().

fn font_stretch(&self) -> c_int

C++ method: int QTextCharFormat::fontStretch() const

C++ documentation:

Returns the current font stretching.

This function was introduced in Qt 5.0.

See also setFontStretch().

fn font_strike_out(&self) -> bool

C++ method: bool QTextCharFormat::fontStrikeOut() const

C++ documentation:

Returns true if the text format's font is struck out (has a horizontal line drawn through it); otherwise returns false.

See also setFontStrikeOut() and font().

fn font_style_hint(&self) -> StyleHint

C++ method: QFont::StyleHint QTextCharFormat::fontStyleHint() const

C++ documentation:

Returns the font style hint.

This function was introduced in Qt 4.5.

See also setFontStyleHint() and font().

fn font_style_strategy(&self) -> StyleStrategy

C++ method: QFont::StyleStrategy QTextCharFormat::fontStyleStrategy() const

C++ documentation:

Returns the current font style strategy.

This function was introduced in Qt 4.5.

See also setFontStyleStrategy() and font().

fn font_underline(&self) -> bool

C++ method: bool QTextCharFormat::fontUnderline() const

C++ documentation:

Returns true if the text format's font is underlined; otherwise returns false.

See also setFontUnderline() and font().

fn font_weight(&self) -> c_int

C++ method: int QTextCharFormat::fontWeight() const

C++ documentation:

Returns the text format's font weight.

See also setFontWeight(), font(), and QFont::Weight.

fn font_word_spacing(&self) -> c_double

C++ method: double QTextCharFormat::fontWordSpacing() const

C++ documentation:

Returns the current word spacing value.

This function was introduced in Qt 4.4.

See also setFontWordSpacing().

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatForegroundArgs<'largs>

C++ method: QTextCharFormat::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextCharFormat::foreground() const

C++ documentation:

Returns the brush used to render foreground details, such as text, frame outlines, and table borders.

See also setForeground(), clearForeground(), and background().

fn has_property(&self, property_id: c_int) -> bool

C++ method: bool QTextCharFormat::hasProperty(int propertyId) const

C++ documentation:

Returns true if the text format has a property with the given propertyId; otherwise returns false.

See also properties() and Property.

fn int_property(&self, property_id: c_int) -> c_int

C++ method: int QTextCharFormat::intProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property is not of QTextFormat::Integer type, 0 is returned instead.

See also setProperty(), boolProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn is_anchor(&self) -> bool

C++ method: bool QTextCharFormat::isAnchor() const

C++ documentation:

Returns true if the text is formatted as an anchor; otherwise returns false.

See also setAnchor(), setAnchorHref(), and setAnchorNames().

fn is_block_format(&self) -> bool

C++ method: bool QTextCharFormat::isBlockFormat() const

C++ documentation:

Returns true if this text format is a BlockFormat; otherwise returns false.

fn is_char_format(&self) -> bool

C++ method: bool QTextCharFormat::isCharFormat() const

C++ documentation:

Returns true if this text format is a CharFormat; otherwise returns false.

fn is_empty(&self) -> bool

C++ method: bool QTextCharFormat::isEmpty() const

C++ documentation:

Returns true if the format does not store any properties; false otherwise.

This function was introduced in Qt 5.3.

See also propertyCount() and properties().

fn is_frame_format(&self) -> bool

C++ method: bool QTextCharFormat::isFrameFormat() const

C++ documentation:

Returns true if this text format is a FrameFormat; otherwise returns false.

fn is_image_format(&self) -> bool

C++ method: bool QTextCharFormat::isImageFormat() const

C++ documentation:

Returns true if this text format is an image format; otherwise returns false.

fn is_list_format(&self) -> bool

C++ method: bool QTextCharFormat::isListFormat() const

C++ documentation:

Returns true if this text format is a ListFormat; otherwise returns false.

fn is_table_cell_format(&self) -> bool

C++ method: bool QTextCharFormat::isTableCellFormat() const

C++ documentation:

Returns true if this text format is a TableCellFormat; otherwise returns false.

This function was introduced in Qt 4.4.

fn is_table_format(&self) -> bool

C++ method: bool QTextCharFormat::isTableFormat() const

C++ documentation:

Returns true if this text format is a TableFormat; otherwise returns false.

fn is_valid(&self) -> bool

C++ method: bool QTextCharFormat::isValid() const

C++ documentation:

Returns true if this character format is valid; otherwise returns false.

fn layout_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextCharFormat::layoutDirection() const

C++ documentation:

Returns the document's layout direction.

See also setLayoutDirection().

fn length_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatLengthPropertyArgs<'largs>

C++ method: QTextCharFormat::lengthProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_length::TextLength>
2) fn length_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_length::TextLength
C++ method: QTextLength QTextCharFormat::lengthProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthVectorProperty(), and Property.

fn length_vector_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatLengthVectorPropertyArgs<'largs>

C++ method: QTextCharFormat::lengthVectorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorTextLengthRef>
2) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorTextLengthRef
C++ method: QVector<QTextLength> QTextCharFormat::lengthVectorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId. If the property isn't of QTextFormat::LengthVector type, an empty length vector is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), and Property.

fn merge(&mut self, other: &TextFormat)

C++ method: void QTextCharFormat::merge(const QTextFormat& other)

C++ documentation:

Merges the other format with this format; where there are conflicts the other format takes precedence.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextCharFormatNewArgs

C++ method: QTextCharFormat::QTextCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
C++ method: [constructor] void QTextCharFormat::QTextCharFormat()

fn object_index(&self) -> c_int

C++ method: int QTextCharFormat::objectIndex() const

C++ documentation:

Returns the index of the format object, or -1 if the format object is invalid.

See also setObjectIndex().

fn object_type(&self) -> c_int

C++ method: int QTextCharFormat::objectType() const

C++ documentation:

Returns the text format's object type.

See also ObjectTypes and setObjectType().

fn op_eq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextCharFormat::operator==(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is the same as the other text format.

fn op_neq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextCharFormat::operator!=(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is different from the other text format.

fn pen_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatPenPropertyArgs<'largs>

C++ method: QTextCharFormat::penProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn pen_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pen::Pen>
2) fn pen_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::pen::Pen
C++ method: QPen QTextCharFormat::penProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Pen type, Qt::NoPen is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn properties<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatPropertiesArgs<'largs>

C++ method: QTextCharFormat::properties

This is an overloaded function. Available variants:

Rust arguments:
1) fn properties(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::map::MapCIntVariantRef>
2) fn properties(&self, cpp_utils::AsStruct) -> qt_core::map::MapCIntVariantRef
C++ method: QMap<int, QVariant> QTextCharFormat::properties() const

C++ documentation:

Returns a map with all properties of this text format.

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatPropertyArgs<'largs>

C++ method: QTextCharFormat::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QTextCharFormat::property(int propertyId) const

C++ documentation:

Returns the property specified by the given propertyId.

See also setProperty() and Property.

fn property_count(&self) -> c_int

C++ method: int QTextCharFormat::propertyCount() const

C++ documentation:

Returns the number of properties stored in the format.

This function was introduced in Qt 4.3.

fn set_anchor(&mut self, anchor: bool)

C++ method: void QTextCharFormat::setAnchor(bool anchor)

C++ documentation:

If anchor is true, text with this format represents an anchor, and is formatted in the appropriate way; otherwise the text is formatted normally. (Anchors are hyperlinks which are often shown underlined and in a different color from plain text.)

The way the text is rendered is independent of whether or not the format has a valid anchor defined. Use setAnchorHref(), and optionally setAnchorNames() to create a hypertext link.

See also isAnchor().

fn set_anchor_href(&mut self, value: &String)

C++ method: void QTextCharFormat::setAnchorHref(const QString& value)

C++ documentation:

Sets the hypertext link for the text format to the given value. This is typically a URL like "http://example.com/index.html".

The anchor will be displayed with the value as its display text; if you want to display different text call setAnchorNames().

To format the text as a hypertext link use setAnchor().

See also anchorHref().

fn set_anchor_name(&mut self, name: &String)

C++ method: void QTextCharFormat::setAnchorName(const QString& name)

C++ documentation:

This function is deprecated. Use setAnchorNames() instead.

Sets the text format's anchor name. For the anchor to work as a hyperlink, the destination must be set with setAnchorHref() and the anchor must be enabled with setAnchor().

See also anchorName().

fn set_anchor_names(&mut self, names: &StringList)

C++ method: void QTextCharFormat::setAnchorNames(const QStringList& names)

C++ documentation:

Sets the text format's anchor names. For the anchor to work as a hyperlink, the destination must be set with setAnchorHref() and the anchor must be enabled with setAnchor().

This function was introduced in Qt 4.3.

See also anchorNames().

fn set_background(&mut self, brush: &Brush)

C++ method: void QTextCharFormat::setBackground(const QBrush& brush)

C++ documentation:

Sets the brush use to paint the document's background to the brush specified.

See also background(), clearBackground(), and setForeground().

fn set_font<'largs, Args>(&'largs mut self, args: Args) where Args: TextCharFormatSetFontArgs<'largs>

C++ method: QTextCharFormat::setFont

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_font(&mut self, &::font::Font) -> ()
C++ method: void QTextCharFormat::setFont(const QFont& font)

C++ documentation:

This is an overloaded function.

Sets the text format's font.

See also font().

Variant 2

Rust arguments: fn set_font(&mut self, (&::font::Font, ::text_char_format::FontPropertiesInheritanceBehavior)) -> ()
C++ method: void QTextCharFormat::setFont(const QFont& font, QTextCharFormat::FontPropertiesInheritanceBehavior behavior)

C++ documentation:

Sets the text format's font.

If behavior is QTextCharFormat::FontPropertiesAll, the font property that has not been explicitly set is treated like as it were set with default value; If behavior is QTextCharFormat::FontPropertiesSpecifiedOnly, the font property that has not been explicitly set is ignored and the respective property value remains unchanged.

This function was introduced in Qt 5.3.

See also font().

fn set_font_capitalization(&mut self, capitalization: Capitalization)

C++ method: void QTextCharFormat::setFontCapitalization(QFont::Capitalization capitalization)

C++ documentation:

Sets the capitalization of the text that apppears in this font to capitalization.

A font's capitalization makes the text appear in the selected capitalization mode.

This function was introduced in Qt 4.4.

See also fontCapitalization().

fn set_font_family(&mut self, family: &String)

C++ method: void QTextCharFormat::setFontFamily(const QString& family)

C++ documentation:

Sets the text format's font family.

See also fontFamily() and setFont().

fn set_font_fixed_pitch(&mut self, fixed_pitch: bool)

C++ method: void QTextCharFormat::setFontFixedPitch(bool fixedPitch)

C++ documentation:

If fixedPitch is true, sets the text format's font to be fixed pitch; otherwise a non-fixed pitch font is used.

See also fontFixedPitch() and setFont().

fn set_font_hinting_preference(&mut self, hinting_preference: HintingPreference)

C++ method: void QTextCharFormat::setFontHintingPreference(QFont::HintingPreference hintingPreference)

C++ documentation:

Sets the hinting preference of the text format's font to be hintingPreference.

This function was introduced in Qt 4.8.

See also fontHintingPreference(), setFont(), and QFont::setHintingPreference().

fn set_font_italic(&mut self, italic: bool)

C++ method: void QTextCharFormat::setFontItalic(bool italic)

C++ documentation:

If italic is true, sets the text format's font to be italic; otherwise the font will be non-italic.

See also fontItalic() and setFont().

fn set_font_kerning(&mut self, enable: bool)

C++ method: void QTextCharFormat::setFontKerning(bool enable)

C++ documentation:

Enables kerning for this font if enable is true; otherwise disables it.

When kerning is enabled, glyph metrics do not add up anymore, even for Latin text. In other words, the assumption that width('a') + width('b') is equal to width("ab") is not neccesairly true.

This function was introduced in Qt 4.5.

See also fontKerning() and setFont().

fn set_font_letter_spacing(&mut self, spacing: c_double)

C++ method: void QTextCharFormat::setFontLetterSpacing(double spacing)

C++ documentation:

Sets the letter spacing of this format to the given spacing. The meaning of the value depends on the font letter spacing type.

For percentage spacing a value of 100 indicates default spacing; a value of 200 doubles the amount of space a letter takes.

This function was introduced in Qt 4.4.

See also fontLetterSpacing(), setFontLetterSpacingType(), and fontLetterSpacingType().

fn set_font_letter_spacing_type(&mut self, letter_spacing_type: SpacingType)

C++ method: void QTextCharFormat::setFontLetterSpacingType(QFont::SpacingType letterSpacingType)

C++ documentation:

Sets the letter spacing type of this format to letterSpacingType.

This function was introduced in Qt 5.0.

See also fontLetterSpacingType(), setFontLetterSpacing(), and fontLetterSpacing().

fn set_font_overline(&mut self, overline: bool)

C++ method: void QTextCharFormat::setFontOverline(bool overline)

C++ documentation:

If overline is true, sets the text format's font to be overlined; otherwise the font is displayed non-overlined.

See also fontOverline() and setFont().

fn set_font_point_size(&mut self, size: c_double)

C++ method: void QTextCharFormat::setFontPointSize(double size)

C++ documentation:

Sets the text format's font size.

See also fontPointSize() and setFont().

fn set_font_stretch(&mut self, factor: c_int)

C++ method: void QTextCharFormat::setFontStretch(int factor)

C++ documentation:

Sets the stretch factor for the font to factor.

The stretch factor changes the width of all characters in the font by factor percent. For example, setting factor to 150 results in all characters in the font being 1.5 times (ie. 150%) wider. The default stretch factor is 100. The minimum stretch factor is 1, and the maximum stretch factor is 4000.

The stretch factor is only applied to outline fonts. The stretch factor is ignored for bitmap fonts.

This function was introduced in Qt 5.0.

See also fontStretch().

fn set_font_strike_out(&mut self, strike_out: bool)

C++ method: void QTextCharFormat::setFontStrikeOut(bool strikeOut)

C++ documentation:

If strikeOut is true, sets the text format's font with strike-out enabled (with a horizontal line through it); otherwise it is displayed without strikeout.

See also fontStrikeOut() and setFont().

fn set_font_style_hint<'largs, Args>(&'largs mut self, args: Args) where Args: TextCharFormatSetFontStyleHintArgs<'largs>

C++ method: QTextCharFormat::setFontStyleHint

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_font_style_hint(&mut self, ::font::StyleHint) -> ()
2) fn set_font_style_hint(&mut self, (::font::StyleHint, ::font::StyleStrategy)) -> ()
C++ method: void QTextCharFormat::setFontStyleHint(QFont::StyleHint hint, QFont::StyleStrategy strategy = ?)

C++ documentation:

Sets the font style hint and strategy.

Qt does not support style hints on X11 since this information is not provided by the window system.

This function was introduced in Qt 4.5.

See also fontStyleHint(), setFont(), and QFont::setStyleHint().

fn set_font_style_strategy(&mut self, strategy: StyleStrategy)

C++ method: void QTextCharFormat::setFontStyleStrategy(QFont::StyleStrategy strategy)

C++ documentation:

Sets the font style strategy.

This function was introduced in Qt 4.5.

See also fontStyleStrategy(), setFont(), and QFont::setStyleStrategy().

fn set_font_underline(&mut self, underline: bool)

C++ method: void QTextCharFormat::setFontUnderline(bool underline)

C++ documentation:

If underline is true, sets the text format's font to be underlined; otherwise it is displayed non-underlined.

See also fontUnderline() and setFont().

fn set_font_weight(&mut self, weight: c_int)

C++ method: void QTextCharFormat::setFontWeight(int weight)

C++ documentation:

Sets the text format's font weight to weight.

See also fontWeight(), setFont(), and QFont::Weight.

fn set_font_word_spacing(&mut self, spacing: c_double)

C++ method: void QTextCharFormat::setFontWordSpacing(double spacing)

C++ documentation:

Sets the word spacing of this format to the given spacing, in pixels.

This function was introduced in Qt 4.4.

See also fontWordSpacing().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QTextCharFormat::setForeground(const QBrush& brush)

C++ documentation:

Sets the foreground brush to the specified brush. The foreground brush is mostly used to render text.

See also foreground(), clearForeground(), and setBackground().

fn set_layout_direction(&mut self, direction: LayoutDirection)

C++ method: void QTextCharFormat::setLayoutDirection(Qt::LayoutDirection direction)

C++ documentation:

Sets the document's layout direction to the specified direction.

See also layoutDirection().

fn set_object_index(&mut self, object: c_int)

C++ method: void QTextCharFormat::setObjectIndex(int object)

C++ documentation:

Sets the format object's object index.

See also objectIndex().

fn set_object_type(&mut self, type_: c_int)

C++ method: void QTextCharFormat::setObjectType(int type)

C++ documentation:

Sets the text format's object type to type.

See also ObjectTypes and objectType().

fn set_property<'largs, Args>(&'largs mut self, args: Args) where Args: TextCharFormatSetPropertyArgs<'largs>

C++ method: QTextCharFormat::setProperty

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_property(&mut self, (libc::c_int, &qt_core::variant::Variant)) -> ()
C++ method: void QTextCharFormat::setProperty(int propertyId, const QVariant& value)

C++ documentation:

Sets the property specified by the propertyId to the given value.

See also property() and Property.

Variant 2

Rust arguments: fn set_property(&mut self, (libc::c_int, &::vector::VectorTextLengthRef)) -> ()
C++ method: void QTextCharFormat::setProperty(int propertyId, const QVector<QTextLength>& lengths)

C++ documentation:

Sets the value of the property given by propertyId to value.

See also lengthVectorProperty() and Property.

fn set_table_cell_column_span(&mut self, table_cell_column_span: c_int)

C++ method: void QTextCharFormat::setTableCellColumnSpan(int tableCellColumnSpan)

fn set_table_cell_row_span(&mut self, table_cell_row_span: c_int)

C++ method: void QTextCharFormat::setTableCellRowSpan(int tableCellRowSpan)

fn set_text_outline(&mut self, pen: &Pen)

C++ method: void QTextCharFormat::setTextOutline(const QPen& pen)

C++ documentation:

Sets the pen used to draw the outlines of characters to the given pen.

See also textOutline().

fn set_tool_tip(&mut self, tip: &String)

C++ method: void QTextCharFormat::setToolTip(const QString& tip)

C++ documentation:

Sets the tool tip for a fragment of text to the given text.

This function was introduced in Qt 4.3.

See also toolTip().

fn set_underline_color(&mut self, color: &Color)

C++ method: void QTextCharFormat::setUnderlineColor(const QColor& color)

C++ documentation:

Sets the underline color used for the characters with this format to the color specified.

See also underlineColor().

fn set_underline_style(&mut self, style: UnderlineStyle)

C++ method: void QTextCharFormat::setUnderlineStyle(QTextCharFormat::UnderlineStyle style)

C++ documentation:

Sets the style of underlining the text to style.

This function was introduced in Qt 4.2.

See also underlineStyle().

fn set_vertical_alignment(&mut self, alignment: VerticalAlignment)

C++ method: void QTextCharFormat::setVerticalAlignment(QTextCharFormat::VerticalAlignment alignment)

C++ documentation:

Sets the vertical alignment used for the characters with this format to the alignment specified.

See also verticalAlignment().

fn string_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatStringPropertyArgs<'largs>

C++ method: QTextCharFormat::stringProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn string_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn string_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTextCharFormat::stringProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::String type, an empty string is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn swap(&mut self, other: &mut TextFormat)

C++ method: void QTextCharFormat::swap(QTextFormat& other)

C++ documentation:

Swaps this text format with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn table_cell_column_span(&self) -> c_int

C++ method: int QTextCharFormat::tableCellColumnSpan() const

fn table_cell_row_span(&self) -> c_int

C++ method: int QTextCharFormat::tableCellRowSpan() const

fn text_outline<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatTextOutlineArgs<'largs>

C++ method: QTextCharFormat::textOutline

This is an overloaded function. Available variants:

Rust arguments:
1) fn text_outline(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::pen::Pen>
2) fn text_outline(&self, cpp_utils::AsStruct) -> ::pen::Pen
C++ method: QPen QTextCharFormat::textOutline() const

C++ documentation:

Returns the pen used to draw the outlines of characters in this format.

See also setTextOutline().

fn to_block_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToBlockFormatArgs<'largs>

C++ method: QTextCharFormat::toBlockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_block_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
2) fn to_block_format(&self, cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
C++ method: QTextBlockFormat QTextCharFormat::toBlockFormat() const

C++ documentation:

Returns this format as a block format.

fn to_char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToCharFormatArgs<'largs>

C++ method: QTextCharFormat::toCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn to_char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextCharFormat::toCharFormat() const

C++ documentation:

Returns this format as a character format.

fn to_frame_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToFrameFormatArgs<'largs>

C++ method: QTextCharFormat::toFrameFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_frame_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame_format::TextFrameFormat>
2) fn to_frame_format(&self, cpp_utils::AsStruct) -> ::text_frame_format::TextFrameFormat
C++ method: QTextFrameFormat QTextCharFormat::toFrameFormat() const

C++ documentation:

Returns this format as a frame format.

fn to_image_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToImageFormatArgs<'largs>

C++ method: QTextCharFormat::toImageFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_image_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_image_format::TextImageFormat>
2) fn to_image_format(&self, cpp_utils::AsStruct) -> ::text_image_format::TextImageFormat
C++ method: QTextImageFormat QTextCharFormat::toImageFormat() const

C++ documentation:

Returns this format as an image format.

fn to_list_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToListFormatArgs<'largs>

C++ method: QTextCharFormat::toListFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_list_format::TextListFormat>
2) fn to_list_format(&self, cpp_utils::AsStruct) -> ::text_list_format::TextListFormat
C++ method: QTextListFormat QTextCharFormat::toListFormat() const

C++ documentation:

Returns this format as a list format.

fn to_table_cell_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToTableCellFormatArgs<'largs>

C++ method: QTextCharFormat::toTableCellFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_cell_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_cell_format::TextTableCellFormat>
2) fn to_table_cell_format(&self, cpp_utils::AsStruct) -> ::text_table_cell_format::TextTableCellFormat
C++ method: QTextTableCellFormat QTextCharFormat::toTableCellFormat() const

C++ documentation:

Returns this format as a table cell format.

This function was introduced in Qt 4.4.

fn to_table_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToTableFormatArgs<'largs>

C++ method: QTextCharFormat::toTableFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_format::TextTableFormat>
2) fn to_table_format(&self, cpp_utils::AsStruct) -> ::text_table_format::TextTableFormat
C++ method: QTextTableFormat QTextCharFormat::toTableFormat() const

C++ documentation:

Returns this format as a table format.

fn tool_tip<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatToolTipArgs<'largs>

C++ method: QTextCharFormat::toolTip

This is an overloaded function. Available variants:

Rust arguments:
1) fn tool_tip(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn tool_tip(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextCharFormat::toolTip() const

C++ documentation:

Returns the tool tip that is displayed for a fragment of text.

This function was introduced in Qt 4.3.

See also setToolTip().

fn type_(&self) -> c_int

C++ method: int QTextCharFormat::type() const

C++ documentation:

Returns the type of this format.

See also FormatType.

fn underline_color<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCharFormatUnderlineColorArgs<'largs>

C++ method: QTextCharFormat::underlineColor

This is an overloaded function. Available variants:

Rust arguments:
1) fn underline_color(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::color::Color>
2) fn underline_color(&self, cpp_utils::AsStruct) -> ::color::Color
C++ method: QColor QTextCharFormat::underlineColor() const

C++ documentation:

Returns the color used to underline the characters with this format.

See also setUnderlineColor().

fn underline_style(&self) -> UnderlineStyle

C++ method: QTextCharFormat::UnderlineStyle QTextCharFormat::underlineStyle() const

C++ documentation:

Returns the style of underlining the text.

This function was introduced in Qt 4.2.

See also setUnderlineStyle().

fn vertical_alignment(&self) -> VerticalAlignment

C++ method: QTextCharFormat::VerticalAlignment QTextCharFormat::verticalAlignment() const

C++ documentation:

Returns the vertical alignment used for characters with this format.

See also setVerticalAlignment().