Enum qt_gui::text_cursor::MoveOperation

pub enum MoveOperation {
    NoMove,
    Start,
    Up,
    StartOfLine,
    StartOfBlock,
    StartOfWord,
    PreviousBlock,
    PreviousCharacter,
    PreviousWord,
    Left,
    WordLeft,
    End,
    Down,
    EndOfLine,
    EndOfWord,
    EndOfBlock,
    NextBlock,
    NextCharacter,
    NextWord,
    Right,
    WordRight,
    NextCell,
    PreviousCell,
    NextRow,
    PreviousRow,
}

C++ type: QTextCursor::MoveOperation

C++ documentation:

See also movePosition().

Variants
NoMove

Keep the cursor where it is (C++ enum variant: NoMove = 0)

Start

Move to the start of the document. (C++ enum variant: Start = 1)

Up

Move up one line. (C++ enum variant: Up = 2)

StartOfLine

Move to the start of the current line. (C++ enum variant: StartOfLine = 3)

StartOfBlock

Move to the start of the current block. (C++ enum variant: StartOfBlock = 4)

StartOfWord

Move to the start of the current word. (C++ enum variant: StartOfWord = 5)

PreviousBlock

Move to the start of the previous block. (C++ enum variant: PreviousBlock = 6)

PreviousCharacter

Move to the previous character. (C++ enum variant: PreviousCharacter = 7)

PreviousWord

Move to the beginning of the previous word. (C++ enum variant: PreviousWord = 8)

Left

Move left one character. (C++ enum variant: Left = 9)

WordLeft

Move left one word. (C++ enum variant: WordLeft = 10)

End

Move to the end of the document. (C++ enum variant: End = 11)

Down

Move down one line. (C++ enum variant: Down = 12)

EndOfLine

Move to the end of the current line. (C++ enum variant: EndOfLine = 13)

EndOfWord

Move to the end of the current word. (C++ enum variant: EndOfWord = 14)

EndOfBlock

Move to the end of the current block. (C++ enum variant: EndOfBlock = 15)

NextBlock

Move to the beginning of the next block. (C++ enum variant: NextBlock = 16)

NextCharacter

Move to the next character. (C++ enum variant: NextCharacter = 17)

NextWord

Move to the next word. (C++ enum variant: NextWord = 18)

Right

Move right one character. (C++ enum variant: Right = 19)

WordRight

Move right one word. (C++ enum variant: WordRight = 20)

NextCell

Move to the beginning of the next table cell inside the current table. If the current cell is the last cell in the row, the cursor will move to the first cell in the next row. (C++ enum variant: NextCell = 21)

PreviousCell

Move to the beginning of the previous table cell inside the current table. If the current cell is the first cell in the row, the cursor will move to the last cell in the previous row. (C++ enum variant: PreviousCell = 22)

NextRow

Move to the first new cell of the next row in the current table. (C++ enum variant: NextRow = 23)

PreviousRow

Move to the last cell of the previous row in the current table. (C++ enum variant: PreviousRow = 24)