Module qt_gui::text_cursor::overloading

Traits

TextCursorBlockArgs
This trait represents a set of arguments accepted by TextCursor::block method.

TextCursorBlockCharFormatArgs
This trait represents a set of arguments accepted by TextCursor::block_char_format method.

TextCursorBlockFormatArgs
This trait represents a set of arguments accepted by TextCursor::block_format method.

TextCursorCharFormatArgs
This trait represents a set of arguments accepted by TextCursor::char_format method.

TextCursorCreateListArgs
This trait represents a set of arguments accepted by TextCursor::create_list method.

TextCursorInsertBlockArgs
This trait represents a set of arguments accepted by TextCursor::insert_block method.

TextCursorInsertImageArgs
This trait represents a set of arguments accepted by TextCursor::insert_image method.

TextCursorInsertListArgs
This trait represents a set of arguments accepted by TextCursor::insert_list method.

TextCursorInsertTableArgs
This trait represents a set of arguments accepted by TextCursor::insert_table method.

TextCursorInsertTextArgs
This trait represents a set of arguments accepted by TextCursor::insert_text method.

TextCursorMovePositionArgs
This trait represents a set of arguments accepted by TextCursor::move_position method.

TextCursorNewArgs
This trait represents a set of arguments accepted by TextCursor::new method.

TextCursorSelectedTextArgs
This trait represents a set of arguments accepted by TextCursor::selected_text method.

TextCursorSelectionArgs
This trait represents a set of arguments accepted by TextCursor::selection method.

TextCursorSetPositionArgs
This trait represents a set of arguments accepted by TextCursor::set_position method.