Struct qt_gui::text_cursor::TextCursor

pub struct TextCursor {
    // some fields omitted
}

C++ type: QTextCursor

C++ documentation:

The QTextCursor class offers an API to access and modify QTextDocuments.

Text cursors are objects that are used to access and modify the contents and underlying structure of text documents via a programming interface that mimics the behavior of a cursor in a text editor. QTextCursor contains information about both the cursor's position within a QTextDocument and any selection that it has made.

QTextCursor is modeled on the way a text cursor behaves in a text editor, providing a programmatic means of performing standard actions through the user interface. A document can be thought of as a single string of characters. The cursor's current position() then is always either between two consecutive characters in the string, or else before the very first character or after the very last character in the string. Documents can also contain tables, lists, images, and other objects in addition to text but, from the developer's point of view, the document can be treated as one long string. Some portions of that string can be considered to lie within particular blocks (e.g. paragraphs), or within a table's cell, or a list's item, or other structural elements. When we refer to "current character" we mean the character immediately before the cursor position() in the document. Similarly, the "current block" is the block that contains the cursor position().

A QTextCursor also has an anchor() position. The text that is between the anchor() and the position() is the selection. If anchor() == position() there is no selection.

The cursor position can be changed programmatically using setPosition() and movePosition(); the latter can also be used to select text. For selections see selectionStart(), selectionEnd(), hasSelection(), clearSelection(), and removeSelectedText().

If the position() is at the start of a block, atBlockStart() returns true; and if it is at the end of a block, atBlockEnd() returns true. The format of the current character is returned by charFormat(), and the format of the current block is returned by blockFormat().

Formatting can be applied to the current text document using the setCharFormat(), mergeCharFormat(), setBlockFormat() and mergeBlockFormat() functions. The 'set' functions will replace the cursor's current character or block format, while the 'merge' functions add the given format properties to the cursor's current format. If the cursor has a selection, the given format is applied to the current selection. Note that when only a part of a block is selected, the block format is applied to the entire block. The text at the current character position can be turned into a list using createList().

Deletions can be achieved using deleteChar(), deletePreviousChar(), and removeSelectedText().

Text strings can be inserted into the document with the insertText() function, blocks (representing new paragraphs) can be inserted with insertBlock().

Existing fragments of text can be inserted with insertFragment() but, if you want to insert pieces of text in various formats, it is usually still easier to use insertText() and supply a character format.

Various types of higher-level structure can also be inserted into the document with the cursor:

Lists are ordered sequences of block elements that are decorated with bullet points or symbols. These are inserted in a specified format with insertList().

Tables are inserted with the insertTable() function, and can be given an optional format. These contain an array of cells that can be traversed using the cursor.

Inline images are inserted with insertImage(). The image to be used can be specified in an image format, or by name.

Frames are inserted by calling insertFrame() with a specified format.

Actions can be grouped (i.e. treated as a single action for undo/redo) using beginEditBlock() and endEditBlock().

Cursor movements are limited to valid cursor positions. In Latin writing this is between any two consecutive characters in the text, before the first character, or after the last character. In some other writing systems cursor movements are limited to "clusters" (e.g. a syllable in Devanagari, or a base letter plus diacritics). Functions such as movePosition() and deleteChar() limit cursor movement to these valid positions.

Methods

impl TextCursor

fn anchor(&self) -> c_int

C++ method: int QTextCursor::anchor() const

C++ documentation:

Returns the anchor position; this is the same as position() unless there is a selection in which case position() marks one end of the selection and anchor() marks the other end. Just like the cursor position, the anchor position is between characters.

See also position(), setPosition(), movePosition(), selectionStart(), and selectionEnd().

fn at_block_end(&self) -> bool

C++ method: bool QTextCursor::atBlockEnd() const

C++ documentation:

Returns true if the cursor is at the end of a block; otherwise returns false.

See also atBlockStart() and atEnd().

fn at_block_start(&self) -> bool

C++ method: bool QTextCursor::atBlockStart() const

C++ documentation:

Returns true if the cursor is at the start of a block; otherwise returns false.

See also atBlockEnd() and atStart().

fn at_end(&self) -> bool

C++ method: bool QTextCursor::atEnd() const

C++ documentation:

Returns true if the cursor is at the end of the document; otherwise returns false.

This function was introduced in Qt 4.6.

See also atStart() and atBlockEnd().

fn at_start(&self) -> bool

C++ method: bool QTextCursor::atStart() const

C++ documentation:

Returns true if the cursor is at the start of the document; otherwise returns false.

See also atBlockStart() and atEnd().

fn begin_edit_block(&mut self)

C++ method: void QTextCursor::beginEditBlock()

C++ documentation:

Indicates the start of a block of editing operations on the document that should appear as a single operation from an undo/redo point of view.

For example:

QTextCursor cursor(textDocument);
cursor.beginEditBlock();
cursor.insertText("Hello");
cursor.insertText("World");
cursor.endEditBlock();

textDocument->undo();
The call to undo() will cause both insertions to be undone, causing both "World" and "Hello" to be removed.

It is possible to nest calls to beginEditBlock and endEditBlock. The top-most pair will determine the scope of the undo/redo operation.

See also endEditBlock().

fn block<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorBlockArgs<'largs>

C++ method: QTextCursor::block

This is an overloaded function. Available variants:

Rust arguments:
1) fn block(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block::TextBlock>
2) fn block(&self, cpp_utils::AsStruct) -> ::text_block::TextBlock
C++ method: QTextBlock QTextCursor::block() const

C++ documentation:

Returns the block that contains the cursor.

fn block_char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorBlockCharFormatArgs<'largs>

C++ method: QTextCursor::blockCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn block_char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn block_char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextCursor::blockCharFormat() const

C++ documentation:

Returns the block character format of the block the cursor is in.

The block char format is the format used when inserting text at the beginning of an empty block.

See also setBlockCharFormat().

fn block_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorBlockFormatArgs<'largs>

C++ method: QTextCursor::blockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn block_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
2) fn block_format(&self, cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
C++ method: QTextBlockFormat QTextCursor::blockFormat() const

C++ documentation:

Returns the block format of the block the cursor is in.

See also setBlockFormat() and charFormat().

fn block_number(&self) -> c_int

C++ method: int QTextCursor::blockNumber() const

C++ documentation:

Returns the number of the block the cursor is in, or 0 if the cursor is invalid.

Note that this function only makes sense in documents without complex objects such as tables or frames.

This function was introduced in Qt 4.2.

fn char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorCharFormatArgs<'largs>

C++ method: QTextCursor::charFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextCursor::charFormat() const

C++ documentation:

Returns the format of the character immediately before the cursor position(). If the cursor is positioned at the beginning of a text block that is not empty then the format of the character immediately after the cursor is returned.

See also setCharFormat(), insertText(), and blockFormat().

fn clear_selection(&mut self)

C++ method: void QTextCursor::clearSelection()

C++ documentation:

Clears the current selection by setting the anchor to the cursor position.

Note that it does not delete the text of the selection.

See also removeSelectedText() and hasSelection().

fn column_number(&self) -> c_int

C++ method: int QTextCursor::columnNumber() const

C++ documentation:

Returns the position of the cursor within its containing line.

Note that this is the column number relative to a wrapped line, not relative to the block (i.e. the paragraph).

You probably want to call positionInBlock() instead.

This function was introduced in Qt 4.2.

See also positionInBlock().

fn create_list<'largs, Args>(&'largs mut self, args: Args) -> *mut TextList where Args: TextCursorCreateListArgs<'largs>

C++ method: QTextCursor::createList

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn create_list(&mut self, ::text_list_format::Style) -> *mut ::text_list::TextList
C++ method: QTextList* QTextCursor::createList(QTextListFormat::Style style)

C++ documentation:

This is an overloaded function.

Creates and returns a new list with the given style, making the cursor's current paragraph the first list item.

The style to be used is defined by the QTextListFormat::Style enum.

See also insertList() and currentList().

Variant 2

Rust arguments: fn create_list(&mut self, &::text_list_format::TextListFormat) -> *mut ::text_list::TextList
C++ method: QTextList* QTextCursor::createList(const QTextListFormat& format)

C++ documentation:

Creates and returns a new list with the given format, and makes the current paragraph the cursor is in the first list item.

See also insertList() and currentList().

fn current_frame(&self) -> *mut TextFrame

C++ method: QTextFrame* QTextCursor::currentFrame() const

C++ documentation:

Returns a pointer to the current frame. Returns 0 if the cursor is invalid.

See also insertFrame().

fn current_list(&self) -> *mut TextList

C++ method: QTextList* QTextCursor::currentList() const

C++ documentation:

Returns the current list if the cursor position() is inside a block that is part of a list; otherwise returns 0.

See also insertList() and createList().

fn current_table(&self) -> *mut TextTable

C++ method: QTextTable* QTextCursor::currentTable() const

C++ documentation:

Returns a pointer to the current table if the cursor position() is inside a block that is part of a table; otherwise returns 0.

See also insertTable().

fn delete_char(&mut self)

C++ method: void QTextCursor::deleteChar()

C++ documentation:

If there is no selected text, deletes the character at the current cursor position; otherwise deletes the selected text.

See also deletePreviousChar(), hasSelection(), and clearSelection().

fn delete_previous_char(&mut self)

C++ method: void QTextCursor::deletePreviousChar()

C++ documentation:

If there is no selected text, deletes the character before the current cursor position; otherwise deletes the selected text.

See also deleteChar(), hasSelection(), and clearSelection().

fn document(&self) -> *mut TextDocument

C++ method: QTextDocument* QTextCursor::document() const

C++ documentation:

Returns the document this cursor is associated with.

This function was introduced in Qt 4.5.

fn end_edit_block(&mut self)

C++ method: void QTextCursor::endEditBlock()

C++ documentation:

Indicates the end of a block of editing operations on the document that should appear as a single operation from an undo/redo point of view.

See also beginEditBlock().

fn has_complex_selection(&self) -> bool

C++ method: bool QTextCursor::hasComplexSelection() const

C++ documentation:

Returns true if the cursor contains a selection that is not simply a range from selectionStart() to selectionEnd(); otherwise returns false.

Complex selections are ones that span at least two cells in a table; their extent is specified by selectedTableCells().

fn has_selection(&self) -> bool

C++ method: bool QTextCursor::hasSelection() const

C++ documentation:

Returns true if the cursor contains a selection; otherwise returns false.

fn insert_block<'largs, Args>(&'largs mut self, args: Args) where Args: TextCursorInsertBlockArgs<'largs>

C++ method: QTextCursor::insertBlock

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_block(&mut self, ()) -> ()
C++ method: void QTextCursor::insertBlock()

C++ documentation:

Inserts a new empty block at the cursor position() with the current blockFormat() and charFormat().

See also setBlockFormat().

Variant 2

Rust arguments: fn insert_block(&mut self, &::text_block_format::TextBlockFormat) -> ()
C++ method: void QTextCursor::insertBlock(const QTextBlockFormat& format)

C++ documentation:

This is an overloaded function.

Inserts a new empty block at the cursor position() with block format format and the current charFormat() as block char format.

See also setBlockFormat().

Variant 3

Rust arguments: fn insert_block(&mut self, (&::text_block_format::TextBlockFormat, &::text_char_format::TextCharFormat)) -> ()
C++ method: void QTextCursor::insertBlock(const QTextBlockFormat& format, const QTextCharFormat& charFormat)

C++ documentation:

This is an overloaded function.

Inserts a new empty block at the cursor position() with block format format and charFormat as block char format.

See also setBlockFormat().

fn insert_fragment(&mut self, fragment: &TextDocumentFragment)

C++ method: void QTextCursor::insertFragment(const QTextDocumentFragment& fragment)

C++ documentation:

Inserts the text fragment at the current position().

fn insert_frame(&mut self, format: &TextFrameFormat) -> *mut TextFrame

C++ method: QTextFrame* QTextCursor::insertFrame(const QTextFrameFormat& format)

C++ documentation:

Inserts a frame with the given format at the current cursor position(), moves the cursor position() inside the frame, and returns the frame.

If the cursor holds a selection, the whole selection is moved inside the frame.

See also hasSelection().

fn insert_html(&mut self, html: &String)

C++ method: void QTextCursor::insertHtml(const QString& html)

C++ documentation:

Inserts the text html at the current position(). The text is interpreted as HTML.

Note: When using this function with a style sheet, the style sheet will only apply to the current block in the document. In order to apply a style sheet throughout a document, use QTextDocument::setDefaultStyleSheet() instead.

This function was introduced in Qt 4.2.

fn insert_image<'largs, Args>(&'largs mut self, args: Args) where Args: TextCursorInsertImageArgs<'largs>

C++ method: QTextCursor::insertImage

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn insert_image(&mut self, &::image::Image) -> ()
2) fn insert_image(&mut self, (&::image::Image, &qt_core::string::String)) -> ()
C++ method: void QTextCursor::insertImage(const QImage& image, const QString& name = ?)

C++ documentation:

This is an overloaded function.

Convenience function for inserting the given image with an optional name at the current position().

This function was introduced in Qt 4.5.

Variant 2

Rust arguments: fn insert_image(&mut self, &qt_core::string::String) -> ()
C++ method: void QTextCursor::insertImage(const QString& name)

C++ documentation:

This is an overloaded function.

Convenience method for inserting the image with the given name at the current position().

QImage img = ...
textDocument->addResource(QTextDocument::ImageResource, QUrl("myimage"), img);
cursor.insertImage("myimage");

Variant 3

Rust arguments: fn insert_image(&mut self, &::text_image_format::TextImageFormat) -> ()
C++ method: void QTextCursor::insertImage(const QTextImageFormat& format)

C++ documentation:

Inserts the image defined by format at the current position().

Variant 4

Rust arguments: fn insert_image(&mut self, (&::text_image_format::TextImageFormat, ::text_frame_format::Position)) -> ()
C++ method: void QTextCursor::insertImage(const QTextImageFormat& format, QTextFrameFormat::Position alignment)

C++ documentation:

This is an overloaded function.

Inserts the image defined by the given format at the cursor's current position with the specified alignment.

This function was introduced in Qt 4.2.

See also position().

fn insert_list<'largs, Args>(&'largs mut self, args: Args) -> *mut TextList where Args: TextCursorInsertListArgs<'largs>

C++ method: QTextCursor::insertList

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_list(&mut self, ::text_list_format::Style) -> *mut ::text_list::TextList
C++ method: QTextList* QTextCursor::insertList(QTextListFormat::Style style)

C++ documentation:

This is an overloaded function.

Inserts a new block at the current position and makes it the first list item of a newly created list with the given style. Returns the created list.

See also currentList(), createList(), and insertBlock().

Variant 2

Rust arguments: fn insert_list(&mut self, &::text_list_format::TextListFormat) -> *mut ::text_list::TextList
C++ method: QTextList* QTextCursor::insertList(const QTextListFormat& format)

C++ documentation:

Inserts a new block at the current position and makes it the first list item of a newly created list with the given format. Returns the created list.

See also currentList(), createList(), and insertBlock().

fn insert_table<'largs, Args>(&'largs mut self, args: Args) -> *mut TextTable where Args: TextCursorInsertTableArgs<'largs>

C++ method: QTextCursor::insertTable

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_table(&mut self, (libc::c_int, libc::c_int)) -> *mut ::text_table::TextTable
C++ method: QTextTable* QTextCursor::insertTable(int rows, int cols)

C++ documentation:

This is an overloaded function.

Creates a new table with the given number of rows and columns, inserts it at the current cursor position() in the document, and returns the table object. The cursor is moved to the beginning of the first cell.

There must be at least one row and one column in the table.

See also currentTable().

Variant 2

Rust arguments: fn insert_table(&mut self, (libc::c_int, libc::c_int, &::text_table_format::TextTableFormat)) -> *mut ::text_table::TextTable
C++ method: QTextTable* QTextCursor::insertTable(int rows, int cols, const QTextTableFormat& format)

C++ documentation:

Creates a new table with the given number of rows and columns in the specified format, inserts it at the current cursor position() in the document, and returns the table object. The cursor is moved to the beginning of the first cell.

There must be at least one row and one column in the table.

See also currentTable().

fn insert_text<'largs, Args>(&'largs mut self, args: Args) where Args: TextCursorInsertTextArgs<'largs>

C++ method: QTextCursor::insertText

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert_text(&mut self, &qt_core::string::String) -> ()
C++ method: void QTextCursor::insertText(const QString& text)

C++ documentation:

Inserts text at the current position, using the current character format.

If there is a selection, the selection is deleted and replaced by text, for example:

cursor.clearSelection();
cursor.movePosition(QTextCursor::NextWord, QTextCursor::KeepAnchor);
cursor.insertText("Hello World");

This clears any existing selection, selects the word at the cursor (i.e. from position() forward), and replaces the selection with the phrase "Hello World".

Any ASCII linefeed characters (\n) in the inserted text are transformed into unicode block separators, corresponding to insertBlock() calls.

See also charFormat() and hasSelection().

Variant 2

Rust arguments: fn insert_text(&mut self, (&qt_core::string::String, &::text_char_format::TextCharFormat)) -> ()
C++ method: void QTextCursor::insertText(const QString& text, const QTextCharFormat& format)

C++ documentation:

This is an overloaded function.

Inserts text at the current position with the given format.

fn is_copy_of(&self, other: &TextCursor) -> bool

C++ method: bool QTextCursor::isCopyOf(const QTextCursor& other) const

C++ documentation:

Returns true if this cursor and other are copies of each other, i.e. one of them was created as a copy of the other and neither has moved since. This is much stricter than equality.

See also operator=() and operator==().

fn is_null(&self) -> bool

C++ method: bool QTextCursor::isNull() const

C++ documentation:

Returns true if the cursor is null; otherwise returns false. A null cursor is created by the default constructor.

fn join_previous_edit_block(&mut self)

C++ method: void QTextCursor::joinPreviousEditBlock()

C++ documentation:

Like beginEditBlock() indicates the start of a block of editing operations that should appear as a single operation for undo/redo. However unlike beginEditBlock() it does not start a new block but reverses the previous call to endEditBlock() and therefore makes following operations part of the previous edit block created.

For example:

QTextCursor cursor(textDocument);
cursor.beginEditBlock();
cursor.insertText("Hello");
cursor.insertText("World");
cursor.endEditBlock();

...

cursor.joinPreviousEditBlock(); cursor.insertText("Hey"); cursor.endEditBlock();

textDocument->undo();
The call to undo() will cause all three insertions to be undone.

See also beginEditBlock() and endEditBlock().

fn keep_position_on_insert(&self) -> bool

C++ method: bool QTextCursor::keepPositionOnInsert() const

C++ documentation:

Returns whether the cursor should keep its current position when text gets inserted at the position of the cursor.

The default is false;

This function was introduced in Qt 4.7.

See also setKeepPositionOnInsert().

fn merge_block_char_format(&mut self, modifier: &TextCharFormat)

C++ method: void QTextCursor::mergeBlockCharFormat(const QTextCharFormat& modifier)

C++ documentation:

Modifies the block char format of the current block (or all blocks that are contained in the selection) with the block format specified by modifier.

See also setBlockCharFormat().

fn merge_block_format(&mut self, modifier: &TextBlockFormat)

C++ method: void QTextCursor::mergeBlockFormat(const QTextBlockFormat& modifier)

C++ documentation:

Modifies the block format of the current block (or all blocks that are contained in the selection) with the block format specified by modifier.

See also setBlockFormat() and blockFormat().

fn merge_char_format(&mut self, modifier: &TextCharFormat)

C++ method: void QTextCursor::mergeCharFormat(const QTextCharFormat& modifier)

C++ documentation:

Merges the cursor's current character format with the properties described by format modifier. If the cursor has a selection, this function applies all the properties set in modifier to all the character formats that are part of the selection.

See also hasSelection() and setCharFormat().

fn move_position<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: TextCursorMovePositionArgs<'largs>

C++ method: QTextCursor::movePosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn move_position(&mut self, ::text_cursor::MoveOperation) -> bool
2) fn move_position(&mut self, (::text_cursor::MoveOperation, ::text_cursor::MoveMode)) -> bool
3) fn move_position(&mut self, (::text_cursor::MoveOperation, ::text_cursor::MoveMode, libc::c_int)) -> bool
C++ method: bool QTextCursor::movePosition(QTextCursor::MoveOperation op, QTextCursor::MoveMode arg2 = ?, int n = ?)

C++ documentation:

Moves the cursor by performing the given operation n times, using the specified mode, and returns true if all operations were completed successfully; otherwise returns false.

For example, if this function is repeatedly used to seek to the end of the next word, it will eventually fail when the end of the document is reached.

By default, the move operation is performed once (n = 1).

If mode is KeepAnchor, the cursor selects the text it moves over. This is the same effect that the user achieves when they hold down the Shift key and move the cursor with the cursor keys.

See also setVisualNavigation().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextCursorNewArgs

C++ method: QTextCursor::QTextCursor

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_cursor::TextCursor
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_cursor::TextCursor>
C++ method: [constructor] void QTextCursor::QTextCursor()

Variant 2

Rust arguments:
1) fn new((*mut ::text_document::TextDocument, cpp_utils::AsStruct)) -> ::text_cursor::TextCursor
2) fn new((*mut ::text_document::TextDocument, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_cursor::TextCursor>
C++ method: [constructor] void QTextCursor::QTextCursor(QTextDocument* document)

Variant 3

Rust arguments:
1) fn new((*mut ::text_frame::TextFrame, cpp_utils::AsStruct)) -> ::text_cursor::TextCursor
2) fn new((*mut ::text_frame::TextFrame, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_cursor::TextCursor>
C++ method: [constructor] void QTextCursor::QTextCursor(QTextFrame* frame)

Variant 4

Rust arguments:
1) fn new((&::text_block::TextBlock, cpp_utils::AsStruct)) -> ::text_cursor::TextCursor
2) fn new((&::text_block::TextBlock, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_cursor::TextCursor>
C++ method: [constructor] void QTextCursor::QTextCursor(const QTextBlock& block)

Variant 5

Rust arguments:
1) fn new((&::text_cursor::TextCursor, cpp_utils::AsStruct)) -> ::text_cursor::TextCursor
2) fn new((&::text_cursor::TextCursor, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_cursor::TextCursor>
C++ method: [constructor] void QTextCursor::QTextCursor(const QTextCursor& cursor)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TextCursor) -> &'l0 mut TextCursor

C++ method: QTextCursor& QTextCursor::operator=(const QTextCursor& other)

C++ documentation:

Makes a copy of cursor and assigns it to this QTextCursor. Note that QTextCursor is an implicitly shared class.

fn op_eq(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator==(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is at the same position in the document as this cursor; otherwise returns false.

fn op_ge(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator>=(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is positioned earlier or at the same position in the document as this cursor; otherwise returns false.

fn op_gt(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator>(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is positioned earlier in the document than this cursor; otherwise returns false.

fn op_le(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator<=(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is positioned later or at the same position in the document as this cursor; otherwise returns false.

fn op_lt(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator<(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is positioned later in the document than this cursor; otherwise returns false.

fn op_neq(&self, rhs: &TextCursor) -> bool

C++ method: bool QTextCursor::operator!=(const QTextCursor& rhs) const

C++ documentation:

Returns true if the other cursor is at a different position in the document as this cursor; otherwise returns false.

fn position(&self) -> c_int

C++ method: int QTextCursor::position() const

C++ documentation:

Returns the absolute position of the cursor within the document. The cursor is positioned between characters.

See also setPosition(), movePosition(), anchor(), and positionInBlock().

fn position_in_block(&self) -> c_int

C++ method: int QTextCursor::positionInBlock() const

C++ documentation:

Returns the relative position of the cursor within the block. The cursor is positioned between characters.

This is equivalent to position() - block().position().

This function was introduced in Qt 4.7.

See also position().

fn remove_selected_text(&mut self)

C++ method: void QTextCursor::removeSelectedText()

C++ documentation:

If there is a selection, its content is deleted; otherwise does nothing.

See also hasSelection().

fn select(&mut self, selection: SelectionType)

C++ method: void QTextCursor::select(QTextCursor::SelectionType selection)

C++ documentation:

Selects text in the document according to the given selection.

fn selected_table_cells(&self, first_row: *mut c_int, num_rows: *mut c_int, first_column: *mut c_int, num_columns: *mut c_int)

C++ method: void QTextCursor::selectedTableCells(int* firstRow, int* numRows, int* firstColumn, int* numColumns) const

C++ documentation:

If the selection spans over table cells, firstRow is populated with the number of the first row in the selection, firstColumn with the number of the first column in the selection, and numRows and numColumns with the number of rows and columns in the selection. If the selection does not span any table cells the results are harmless but undefined.

fn selected_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorSelectedTextArgs<'largs>

C++ method: QTextCursor::selectedText

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn selected_text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextCursor::selectedText() const

C++ documentation:

Returns the current selection's text (which may be empty). This only returns the text, with no rich text formatting information. If you want a document fragment (i.e. formatted rich text) use selection() instead.

Note: If the selection obtained from an editor spans a line break, the text will contain a Unicode U+2029 paragraph separator character instead of a newline \n character. Use QString::replace() to replace these characters with newlines.

fn selection<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCursorSelectionArgs<'largs>

C++ method: QTextCursor::selection

This is an overloaded function. Available variants:

Rust arguments:
1) fn selection(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
2) fn selection(&self, cpp_utils::AsStruct) -> ::text_document_fragment::TextDocumentFragment
C++ method: QTextDocumentFragment QTextCursor::selection() const

C++ documentation:

Returns the current selection (which may be empty) with all its formatting information. If you just want the selected text (i.e. plain text) use selectedText() instead.

Note: Unlike QTextDocumentFragment::toPlainText(), selectedText() may include special unicode characters such as QChar::ParagraphSeparator.

See also QTextDocumentFragment::toPlainText().

fn selection_end(&self) -> c_int

C++ method: int QTextCursor::selectionEnd() const

C++ documentation:

Returns the end of the selection or position() if the cursor doesn't have a selection.

See also selectionStart(), position(), and anchor().

fn selection_start(&self) -> c_int

C++ method: int QTextCursor::selectionStart() const

C++ documentation:

Returns the start of the selection or position() if the cursor doesn't have a selection.

See also selectionEnd(), position(), and anchor().

fn set_block_char_format(&mut self, format: &TextCharFormat)

C++ method: void QTextCursor::setBlockCharFormat(const QTextCharFormat& format)

C++ documentation:

Sets the block char format of the current block (or all blocks that are contained in the selection) to format.

See also blockCharFormat().

fn set_block_format(&mut self, format: &TextBlockFormat)

C++ method: void QTextCursor::setBlockFormat(const QTextBlockFormat& format)

C++ documentation:

Sets the block format of the current block (or all blocks that are contained in the selection) to format.

See also blockFormat() and mergeBlockFormat().

fn set_char_format(&mut self, format: &TextCharFormat)

C++ method: void QTextCursor::setCharFormat(const QTextCharFormat& format)

C++ documentation:

Sets the cursor's current character format to the given format. If the cursor has a selection, the given format is applied to the current selection.

See also charFormat(), hasSelection(), and mergeCharFormat().

fn set_keep_position_on_insert(&mut self, b: bool)

C++ method: void QTextCursor::setKeepPositionOnInsert(bool b)

C++ documentation:

Defines whether the cursor should keep its current position when text gets inserted at the current position of the cursor.

If b is true, the cursor keeps its current position when text gets inserted at the positing of the cursor. If b is false, the cursor moves along with the inserted text.

The default is false.

Note that a cursor always moves when text is inserted before the current position of the cursor, and it always keeps its position when text is inserted after the current position of the cursor.

This function was introduced in Qt 4.7.

See also keepPositionOnInsert().

fn set_position<'largs, Args>(&'largs mut self, args: Args) where Args: TextCursorSetPositionArgs<'largs>

C++ method: QTextCursor::setPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_position(&mut self, libc::c_int) -> ()
2) fn set_position(&mut self, (libc::c_int, ::text_cursor::MoveMode)) -> ()
C++ method: void QTextCursor::setPosition(int pos, QTextCursor::MoveMode mode = ?)

C++ documentation:

Moves the cursor to the absolute position in the document specified by pos using a MoveMode specified by m. The cursor is positioned between characters.

See also position(), movePosition(), and anchor().

fn set_vertical_movement_x(&mut self, x: c_int)

C++ method: void QTextCursor::setVerticalMovementX(int x)

C++ documentation:

Sets the visual x position for vertical cursor movements to x.

The vertical movement x position is cleared automatically when the cursor moves horizontally, and kept unchanged when the cursor moves vertically. The mechanism allows the cursor to move up and down on a visually straight line with proportional fonts, and to gently "jump" over short lines.

A value of -1 indicates no predefined x position. It will then be set automatically the next time the cursor moves up or down.

This function was introduced in Qt 4.7.

See also verticalMovementX().

fn set_visual_navigation(&mut self, b: bool)

C++ method: void QTextCursor::setVisualNavigation(bool b)

C++ documentation:

Sets visual navigation to b.

Visual navigation means skipping over hidden text paragraphs. The default is false.

This function was introduced in Qt 4.4.

See also visualNavigation() and movePosition().

fn swap(&mut self, other: &mut TextCursor)

C++ method: void QTextCursor::swap(QTextCursor& other)

C++ documentation:

Swaps this text cursor instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn vertical_movement_x(&self) -> c_int

C++ method: int QTextCursor::verticalMovementX() const

C++ documentation:

Returns the visual x position for vertical cursor movements.

A value of -1 indicates no predefined x position. It will then be set automatically the next time the cursor moves up or down.

This function was introduced in Qt 4.7.

See also setVerticalMovementX().

fn visual_navigation(&self) -> bool

C++ method: bool QTextCursor::visualNavigation() const

C++ documentation:

Returns true if the cursor does visual navigation; otherwise returns false.

Visual navigation means skipping over hidden text paragraphs. The default is false.

This function was introduced in Qt 4.4.

See also setVisualNavigation() and movePosition().