Struct qt_gui::text_document_fragment::TextDocumentFragment

pub struct TextDocumentFragment {
    // some fields omitted
}

C++ type: QTextDocumentFragment

C++ documentation:

The QTextDocumentFragment class represents a piece of formatted text from a QTextDocument.

A QTextDocumentFragment is a fragment of rich text, that can be inserted into a QTextDocument. A document fragment can be created from a QTextDocument, from a QTextCursor's selection, or from another document fragment. Document fragments can also be created by the static functions, fromPlainText() and fromHtml().

The contents of a document fragment can be obtained as plain text by using the toPlainText() function, or it can be obtained as HTML with toHtml().

Methods

impl TextDocumentFragment

fn from_html<Args>(args: Args) -> Args::ReturnType where Args: TextDocumentFragmentFromHtmlArgs

C++ method: QTextDocumentFragment::fromHtml

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_html((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
2) fn from_html((&qt_core::string::String, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
C++ method: static QTextDocumentFragment QTextDocumentFragment::fromHtml(const QString& html)

C++ documentation:

Returns a QTextDocumentFragment based on the arbitrary piece of HTML in the given text. The formatting is preserved as much as possible; for example, "<b>bold</b>" will become a document fragment with the text "bold" with a bold character format.

Variant 2

Rust arguments:
1) fn from_html((&qt_core::string::String, *const ::text_document::TextDocument, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
2) fn from_html((&qt_core::string::String, *const ::text_document::TextDocument, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
C++ method: static QTextDocumentFragment QTextDocumentFragment::fromHtml(const QString& html, const QTextDocument* resourceProvider)

C++ documentation:

Returns a QTextDocumentFragment based on the arbitrary piece of HTML in the given text. The formatting is preserved as much as possible; for example, "<b>bold</b>" will become a document fragment with the text "bold" with a bold character format.

If the provided HTML contains references to external resources such as imported style sheets, then they will be loaded through the resourceProvider.

This function was introduced in Qt 4.2.

fn from_plain_text<Args>(args: Args) -> Args::ReturnType where Args: TextDocumentFragmentFromPlainTextArgs

C++ method: QTextDocumentFragment::fromPlainText

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_plain_text((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
2) fn from_plain_text((&qt_core::string::String, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
C++ method: static QTextDocumentFragment QTextDocumentFragment::fromPlainText(const QString& plainText)

C++ documentation:

Returns a document fragment that contains the given plainText.

When inserting such a fragment into a QTextDocument the current char format of the QTextCursor used for insertion is used as format for the text.

fn is_empty(&self) -> bool

C++ method: bool QTextDocumentFragment::isEmpty() const

C++ documentation:

Returns true if the fragment is empty; otherwise returns false.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextDocumentFragmentNewArgs

C++ method: QTextDocumentFragment::QTextDocumentFragment

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_document_fragment::TextDocumentFragment
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
C++ method: [constructor] void QTextDocumentFragment::QTextDocumentFragment()

Variant 2

Rust arguments:
1) fn new((&::text_cursor::TextCursor, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
2) fn new((&::text_cursor::TextCursor, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
C++ method: [constructor] void QTextDocumentFragment::QTextDocumentFragment(const QTextCursor& range)

Variant 3

Rust arguments:
1) fn new((*const ::text_document::TextDocument, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
2) fn new((*const ::text_document::TextDocument, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
C++ method: [constructor] void QTextDocumentFragment::QTextDocumentFragment(const QTextDocument* document)

Variant 4

Rust arguments:
1) fn new((&::text_document_fragment::TextDocumentFragment, cpp_utils::AsStruct)) -> ::text_document_fragment::TextDocumentFragment
2) fn new((&::text_document_fragment::TextDocumentFragment, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_fragment::TextDocumentFragment>
C++ method: [constructor] void QTextDocumentFragment::QTextDocumentFragment(const QTextDocumentFragment& rhs)

fn op_assign<'l0, 'l1>(&'l0 mut self, rhs: &'l1 TextDocumentFragment) -> &'l0 mut TextDocumentFragment

C++ method: QTextDocumentFragment& QTextDocumentFragment::operator=(const QTextDocumentFragment& rhs)

C++ documentation:

Assigns the other fragment to this fragment.

fn to_html<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextDocumentFragmentToHtmlArgs<'largs>

C++ method: QTextDocumentFragment::toHtml

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_html(&self, (&qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn to_html(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
3) fn to_html(&self, (&qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> qt_core::string::String
4) fn to_html(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextDocumentFragment::toHtml(const QByteArray& encoding = ?) const

C++ documentation:

Returns the contents of the document fragment as HTML, using the specified encoding (e.g., "UTF-8", "ISO 8859-1").

This function was introduced in Qt 4.2.

See also toPlainText(), QTextDocument::toHtml(), and QTextCodec.

fn to_plain_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextDocumentFragmentToPlainTextArgs<'largs>

C++ method: QTextDocumentFragment::toPlainText

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_plain_text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn to_plain_text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextDocumentFragment::toPlainText() const

C++ documentation:

Returns the document fragment's text as plain text (i.e. with no formatting information).

See also toHtml().