Struct qt_gui::text_document_writer::TextDocumentWriter

pub struct TextDocumentWriter {
    // some fields omitted
}

C++ type: QTextDocumentWriter

C++ documentation:

The QTextDocumentWriter class provides a format-independent interface for writing a QTextDocument to files or other devices.

To write a document, construct a QTextDocumentWriter object with either a file name or a device object, and specify the document format to be written. You can construct a writer and set the format using setFormat() later.

Call write() to write the document to the device. If the document is successfully written, this function returns true. However, if an error occurs when writing the document, it will return false.

Call supportedDocumentFormats() for a list of formats that QTextDocumentWriter can write.

Since the capabilities of the supported output formats vary considerably, the writer simply outputs the appropriate subset of objects for each format. This typically includes the formatted text and images contained in a document.

Methods

impl TextDocumentWriter

fn codec(&self) -> *mut TextCodec

C++ method: QTextCodec* QTextDocumentWriter::codec() const

C++ documentation:

Returns the codec that is currently assigned to the writer.

See also setCodec().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QTextDocumentWriter::device() const

C++ documentation:

Returns the device currently assigned, or 0 if no device has been assigned.

See also setDevice().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextDocumentWriterFileNameArgs<'largs>

C++ method: QTextDocumentWriter::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextDocumentWriter::fileName() const

C++ documentation:

If the currently assigned device is a QFile, or if setFileName() has been called, this function returns the name of the file to be written to. In all other cases, it returns an empty string.

See also setFileName() and setDevice().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextDocumentWriterFormatArgs<'largs>

C++ method: QTextDocumentWriter::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::byte_array::ByteArray>
2) fn format(&self, cpp_utils::AsStruct) -> qt_core::byte_array::ByteArray
C++ method: QByteArray QTextDocumentWriter::format() const

C++ documentation:

Returns the format used for writing documents.

See also setFormat().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextDocumentWriterNewArgs

C++ method: QTextDocumentWriter::QTextDocumentWriter

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_document_writer::TextDocumentWriter
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_document_writer::TextDocumentWriter>
C++ method: [constructor] void QTextDocumentWriter::QTextDocumentWriter()

Variant 2

Rust arguments:
1) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::text_document_writer::TextDocumentWriter
2) fn new((*mut qt_core::qiod_evice::QiodEvice, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_writer::TextDocumentWriter>
C++ method: [constructor] void QTextDocumentWriter::QTextDocumentWriter(QIODevice* device, const QByteArray& format)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::text_document_writer::TextDocumentWriter
2) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::text_document_writer::TextDocumentWriter
3) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_writer::TextDocumentWriter>
4) fn new((&qt_core::string::String, &qt_core::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_document_writer::TextDocumentWriter>
C++ method: [constructor] void QTextDocumentWriter::QTextDocumentWriter(const QString& fileName, const QByteArray& format = ?)

fn set_codec(&mut self, codec: *mut TextCodec)

C++ method: void QTextDocumentWriter::setCodec(QTextCodec* codec)

C++ documentation:

Sets the codec for this stream to codec. The codec is used for encoding any data that is written. By default, QTextDocumentWriter uses UTF-8.

See also codec().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QTextDocumentWriter::setDevice(QIODevice* device)

C++ documentation:

Sets the writer's device to the device specified. If a device has already been set, the old device is removed but otherwise left unchanged.

If the device is not already open, QTextDocumentWriter will attempt to open the device in QIODevice::WriteOnly mode by calling open().

Note: This will not work for certain devices, such as QProcess, QTcpSocket and QUdpSocket, where some configuration is required before the device can be opened.

See also device() and setFileName().

fn set_file_name(&mut self, file_name: &String)

C++ method: void QTextDocumentWriter::setFileName(const QString& fileName)

C++ documentation:

Sets the name of the file to be written to fileName. Internally, QTextDocumentWriter will create a QFile and open it in QIODevice::WriteOnly mode, and use this file when writing the document.

See also fileName() and setDevice().

fn set_format(&mut self, format: &ByteArray)

C++ method: void QTextDocumentWriter::setFormat(const QByteArray& format)

C++ documentation:

Sets the format used to write documents to the format specified. format is a case insensitive text string. For example:

        QTextDocumentWriter writer;
        writer.setFormat("odf"); // same as writer.setFormat("ODF");

You can call supportedDocumentFormats() for the full list of formats QTextDocumentWriter supports.

See also format().

fn supported_document_formats<Args>(args: Args) -> Args::ReturnType where Args: TextDocumentWriterSupportedDocumentFormatsArgs

C++ method: QTextDocumentWriter::supportedDocumentFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn supported_document_formats(cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::list::ListByteArrayRef>
2) fn supported_document_formats(cpp_utils::AsStruct) -> qt_core::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTextDocumentWriter::supportedDocumentFormats()

C++ documentation:

Returns the list of document formats supported by QTextDocumentWriter.

By default, Qt can write the following formats:

Format	Description

plaintext	Plain text

HTML	HyperText Markup Language

ODF	OpenDocument Format

See also setFormat().

fn write<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: TextDocumentWriterWriteArgs<'largs>

C++ method: QTextDocumentWriter::write

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write(&mut self, *const ::text_document::TextDocument) -> bool
C++ method: bool QTextDocumentWriter::write(const QTextDocument* document)

C++ documentation:

Writes the given document to the assigned device or file and returns true if successful; otherwise returns false.

Variant 2

Rust arguments: fn write(&mut self, &::text_document_fragment::TextDocumentFragment) -> bool
C++ method: bool QTextDocumentWriter::write(const QTextDocumentFragment& fragment)

C++ documentation:

Writes the document fragment specified by fragment to the assigned device or file and returns true if successful; otherwise returns false.