Enum qt_gui::text_format::Property

pub enum Property {
    ObjectIndex,
    CssFloat,
    LayoutDirection,
    OutlinePen,
    BackgroundBrush,
    ForegroundBrush,
    BackgroundImageUrl,
    BlockAlignment,
    BlockTopMargin,
    BlockBottomMargin,
    BlockLeftMargin,
    BlockRightMargin,
    TextIndent,
    TabPositions,
    BlockIndent,
    LineHeight,
    LineHeightType,
    BlockNonBreakableLines,
    BlockTrailingHorizontalRulerWidth,
    FirstFontProperty,
    FontLetterSpacing,
    FontWordSpacing,
    FontStyleHint,
    FontStyleStrategy,
    FontKerning,
    FontHintingPreference,
    FontFamily,
    FontPointSize,
    FontSizeAdjustment,
    FontWeight,
    FontItalic,
    FontUnderline,
    FontOverline,
    FontStrikeOut,
    FontFixedPitch,
    FontPixelSize,
    TextUnderlineColor,
    TextVerticalAlignment,
    TextOutline,
    TextUnderlineStyle,
    TextToolTip,
    IsAnchor,
    AnchorHref,
    AnchorName,
    FontLetterSpacingType,
    FontStretch,
    ObjectType,
    ListStyle,
    ListIndent,
    ListNumberPrefix,
    ListNumberSuffix,
    FrameBorder,
    FrameMargin,
    FramePadding,
    FrameWidth,
    FrameHeight,
    FrameTopMargin,
    FrameBottomMargin,
    FrameLeftMargin,
    FrameRightMargin,
    FrameBorderBrush,
    FrameBorderStyle,
    TableColumns,
    TableColumnWidthConstraints,
    TableCellSpacing,
    TableCellPadding,
    TableHeaderRowCount,
    TableCellRowSpan,
    TableCellColumnSpan,
    TableCellTopPadding,
    TableCellBottomPadding,
    TableCellLeftPadding,
    TableCellRightPadding,
    ImageName,
    ImageWidth,
    ImageHeight,
    FullWidthSelection,
    PageBreakPolicy,
    UserProperty,
}

C++ type: QTextFormat::Property

C++ documentation:

This enum describes the different properties a format can have.

Paragraph and character properties

Paragraph properties

Character properties

List properties

Table and frame properties

Table cell properties

Image properties

Selection properties

Page break properties

See also property() and setProperty().

Variants
ObjectIndex

The index of the formatted object. See objectIndex(). (C++ enum variant: ObjectIndex = 0)

CssFloat

How a frame is located relative to the surrounding text (C++ enum variant: CssFloat = 2048)

LayoutDirection

The layout direction of the text in the document (Qt::LayoutDirection). (C++ enum variant: LayoutDirection = 2049)

OutlinePen

(C++ enum variant: OutlinePen = 2064)

BackgroundBrush

(C++ enum variant: BackgroundBrush = 2080)

ForegroundBrush

(C++ enum variant: ForegroundBrush = 2081)

BackgroundImageUrl

(C++ enum variant: BackgroundImageUrl = 2083)

BlockAlignment

(C++ enum variant: BlockAlignment = 4112)

BlockTopMargin

(C++ enum variant: BlockTopMargin = 4144)

BlockBottomMargin

(C++ enum variant: BlockBottomMargin = 4145)

BlockLeftMargin

(C++ enum variant: BlockLeftMargin = 4146)

BlockRightMargin

(C++ enum variant: BlockRightMargin = 4147)

TextIndent

(C++ enum variant: TextIndent = 4148)

TabPositions

Specifies the tab positions. The tab positions are structs of QTextOption::Tab which are stored in a QList (internally, in a QList<QVariant>). (C++ enum variant: TabPositions = 4149)

BlockIndent

(C++ enum variant: BlockIndent = 4160)

LineHeight

(C++ enum variant: LineHeight = 4168)

LineHeightType

(C++ enum variant: LineHeightType = 4169)

BlockNonBreakableLines

(C++ enum variant: BlockNonBreakableLines = 4176)

BlockTrailingHorizontalRulerWidth

The width of a horizontal ruler element. (C++ enum variant: BlockTrailingHorizontalRulerWidth = 4192)

FirstFontProperty

This variant corresponds to multiple C++ enum variants with the same value:

FirstFontProperty = 8160

FontCapitalization = 8160: Specifies the capitalization type that is to be applied to the text.

FontLetterSpacing

Changes the default spacing between individual letters in the font. The value is specified as a percentage or absolute value, depending on FontLetterSpacingType. The default value is 100%. (C++ enum variant: FontLetterSpacing = 8161)

FontWordSpacing

Changes the default spacing between individual words. A positive value increases the word spacing by the corresponding pixels; a negative value decreases the spacing. (C++ enum variant: FontWordSpacing = 8162)

FontStyleHint

Corresponds to the QFont::StyleHint property (C++ enum variant: FontStyleHint = 8163)

FontStyleStrategy

Corresponds to the QFont::StyleStrategy property (C++ enum variant: FontStyleStrategy = 8164)

FontKerning

Specifies whether the font has kerning turned on. (C++ enum variant: FontKerning = 8165)

FontHintingPreference

Controls the use of hinting according to values of the QFont::HintingPreference enum. (C++ enum variant: FontHintingPreference = 8166)

FontFamily

(C++ enum variant: FontFamily = 8192)

FontPointSize

(C++ enum variant: FontPointSize = 8193)

FontSizeAdjustment

This variant corresponds to multiple C++ enum variants with the same value:

FontSizeAdjustment = 8194: Specifies the change in size given to the fontsize already set using FontPointSize or FontPixelSize.

FontSizeIncrement = 8194

FontWeight

(C++ enum variant: FontWeight = 8195)

FontItalic

(C++ enum variant: FontItalic = 8196)

FontUnderline

This property has been deprecated. Use QTextFormat::TextUnderlineStyle instead. (C++ enum variant: FontUnderline = 8197)

FontOverline

(C++ enum variant: FontOverline = 8198)

FontStrikeOut

(C++ enum variant: FontStrikeOut = 8199)

FontFixedPitch

(C++ enum variant: FontFixedPitch = 8200)

FontPixelSize

This variant corresponds to multiple C++ enum variants with the same value:

FontPixelSize = 8201:

LastFontProperty = 8201

TextUnderlineColor

(C++ enum variant: TextUnderlineColor = 8208)

TextVerticalAlignment

(C++ enum variant: TextVerticalAlignment = 8225)

TextOutline

(C++ enum variant: TextOutline = 8226)

TextUnderlineStyle

(C++ enum variant: TextUnderlineStyle = 8227)

TextToolTip

Specifies the (optional) tool tip to be displayed for a fragment of text. (C++ enum variant: TextToolTip = 8228)

IsAnchor

(C++ enum variant: IsAnchor = 8240)

AnchorHref

(C++ enum variant: AnchorHref = 8241)

AnchorName

(C++ enum variant: AnchorName = 8242)

FontLetterSpacingType

Specifies the meaning of the FontLetterSpacing property. The default is QFont::PercentageSpacing. (C++ enum variant: FontLetterSpacingType = 8243)

FontStretch

Corresponds to the QFont::Stretch property (C++ enum variant: FontStretch = 8244)

ObjectType

(C++ enum variant: ObjectType = 12032)

ListStyle

Specifies the style used for the items in a list, described by values of the QTextListFormat::Style enum. (C++ enum variant: ListStyle = 12288)

ListIndent

Specifies the amount of indentation used for a list. (C++ enum variant: ListIndent = 12289)

ListNumberPrefix

Defines the text which is prepended to item numbers in numeric lists. (C++ enum variant: ListNumberPrefix = 12290)

ListNumberSuffix

Defines the text which is appended to item numbers in numeric lists. (C++ enum variant: ListNumberSuffix = 12291)

FrameBorder

(C++ enum variant: FrameBorder = 16384)

FrameMargin

(C++ enum variant: FrameMargin = 16385)

FramePadding

(C++ enum variant: FramePadding = 16386)

FrameWidth

(C++ enum variant: FrameWidth = 16387)

FrameHeight

(C++ enum variant: FrameHeight = 16388)

FrameTopMargin

(C++ enum variant: FrameTopMargin = 16389)

FrameBottomMargin

(C++ enum variant: FrameBottomMargin = 16390)

FrameLeftMargin

(C++ enum variant: FrameLeftMargin = 16391)

FrameRightMargin

(C++ enum variant: FrameRightMargin = 16392)

FrameBorderBrush

(C++ enum variant: FrameBorderBrush = 16393)

FrameBorderStyle

See the BorderStyle enum. (C++ enum variant: FrameBorderStyle = 16400)

TableColumns

(C++ enum variant: TableColumns = 16640)

TableColumnWidthConstraints

(C++ enum variant: TableColumnWidthConstraints = 16641)

TableCellSpacing

(C++ enum variant: TableCellSpacing = 16642)

TableCellPadding

(C++ enum variant: TableCellPadding = 16643)

TableHeaderRowCount

(C++ enum variant: TableHeaderRowCount = 16644)

TableCellRowSpan

C++ enum variant: TableCellRowSpan = 18448

TableCellColumnSpan

C++ enum variant: TableCellColumnSpan = 18449

TableCellTopPadding

C++ enum variant: TableCellTopPadding = 18450

TableCellBottomPadding

C++ enum variant: TableCellBottomPadding = 18451

TableCellLeftPadding

C++ enum variant: TableCellLeftPadding = 18452

TableCellRightPadding

C++ enum variant: TableCellRightPadding = 18453

ImageName

C++ enum variant: ImageName = 20480

ImageWidth

C++ enum variant: ImageWidth = 20496

ImageHeight

C++ enum variant: ImageHeight = 20497

FullWidthSelection

When set on the characterFormat of a selection, the whole width of the text will be shown selected. (C++ enum variant: FullWidthSelection = 24576)

PageBreakPolicy

Specifies how pages are broken. See the PageBreakFlag enum. (C++ enum variant: PageBreakPolicy = 28672)

UserProperty

(C++ enum variant: UserProperty = 1048576)