Module qt_gui::text_format::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

SwapArgs
This trait represents a set of arguments accepted by swap method.

TextFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextFormat::as_q_variant method.

TextFormatBackgroundArgs
This trait represents a set of arguments accepted by TextFormat::background method.

TextFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextFormat::brush_property method.

TextFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextFormat::color_property method.

TextFormatForegroundArgs
This trait represents a set of arguments accepted by TextFormat::foreground method.

TextFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextFormat::length_property method.

TextFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextFormat::length_vector_property method.

TextFormatNewArgs
This trait represents a set of arguments accepted by TextFormat::new method.

TextFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextFormat::pen_property method.

TextFormatPropertiesArgs
This trait represents a set of arguments accepted by TextFormat::properties method.

TextFormatPropertyArgs
This trait represents a set of arguments accepted by TextFormat::property method.

TextFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextFormat::set_property method.

TextFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextFormat::string_property method.

TextFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_block_format method.

TextFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_char_format method.

TextFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_frame_format method.

TextFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_image_format method.

TextFormatToListFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_list_format method.

TextFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_table_cell_format method.

TextFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextFormat::to_table_format method.