Struct qt_gui::text_format::TextFormat

pub struct TextFormat {
    // some fields omitted
}

C++ type: QTextFormat

C++ documentation:

The QTextFormat class provides formatting information for a QTextDocument.

A QTextFormat is a generic class used for describing the format of parts of a QTextDocument. The derived classes QTextCharFormat, QTextBlockFormat, QTextListFormat, and QTextTableFormat are usually more useful, and describe the formatting that is applied to specific parts of the document.

A format has a FormatType which specifies the kinds of text item it can format; e.g. a block of text, a list, a table, etc. A format also has various properties (some specific to particular format types), as described by the Property enum. Every property has a corresponding Property.

The format type is given by type(), and the format can be tested with isCharFormat(), isBlockFormat(), isListFormat(), isTableFormat(), isFrameFormat(), and isImageFormat(). If the type is determined, it can be retrieved with toCharFormat(), toBlockFormat(), toListFormat(), toTableFormat(), toFrameFormat(), and toImageFormat().

A format's properties can be set with the setProperty() functions, and retrieved with boolProperty(), intProperty(), doubleProperty(), and stringProperty() as appropriate. All the property IDs used in the format can be retrieved with allPropertyIds(). One format can be merged into another using merge().

A format's object index can be set with setObjectIndex(), and retrieved with objectIndex(). These methods can be used to associate the format with a QTextObject. It is used to represent lists, frames, and tables inside the document.

Methods

impl TextFormat

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatAsQVariantArgs<'largs>

C++ method: QTextFormat::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTextFormat::operator QVariant() const

C++ documentation:

Returns the text format as a QVariant

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatBackgroundArgs<'largs>

C++ method: QTextFormat::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextFormat::background() const

C++ documentation:

Returns the brush used to paint the document's background.

See also setBackground(), clearBackground(), and foreground().

fn bool_property(&self, property_id: c_int) -> bool

C++ method: bool QTextFormat::boolProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QTextFormat::Bool type, false is returned instead.

See also setProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn brush_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatBrushPropertyArgs<'largs>

C++ method: QTextFormat::brushProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::brush::Brush>
2) fn brush_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::brush::Brush
C++ method: QBrush QTextFormat::brushProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Brush type, Qt::NoBrush is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn clear_background(&mut self)

C++ method: void QTextFormat::clearBackground()

C++ documentation:

Clears the brush used to paint the document's background. The default brush will be used.

See also background(), setBackground(), and clearForeground().

fn clear_foreground(&mut self)

C++ method: void QTextFormat::clearForeground()

C++ documentation:

Clears the brush used to paint the document's foreground. The default brush will be used.

See also foreground(), setForeground(), and clearBackground().

fn clear_property(&mut self, property_id: c_int)

C++ method: void QTextFormat::clearProperty(int propertyId)

C++ documentation:

Clears the value of the property given by propertyId

See also Property.

fn color_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatColorPropertyArgs<'largs>

C++ method: QTextFormat::colorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn color_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::color::Color>
2) fn color_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::color::Color
C++ method: QColor QTextFormat::colorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Color type, an invalid color is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn double_property(&self, property_id: c_int) -> c_double

C++ method: double QTextFormat::doubleProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QVariant::Double or QMetaType::Float type, 0 is returned instead.

See also setProperty(), boolProperty(), intProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatForegroundArgs<'largs>

C++ method: QTextFormat::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextFormat::foreground() const

C++ documentation:

Returns the brush used to render foreground details, such as text, frame outlines, and table borders.

See also setForeground(), clearForeground(), and background().

fn has_property(&self, property_id: c_int) -> bool

C++ method: bool QTextFormat::hasProperty(int propertyId) const

C++ documentation:

Returns true if the text format has a property with the given propertyId; otherwise returns false.

See also properties() and Property.

fn int_property(&self, property_id: c_int) -> c_int

C++ method: int QTextFormat::intProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property is not of QTextFormat::Integer type, 0 is returned instead.

See also setProperty(), boolProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn is_block_format(&self) -> bool

C++ method: bool QTextFormat::isBlockFormat() const

C++ documentation:

Returns true if this text format is a BlockFormat; otherwise returns false.

fn is_char_format(&self) -> bool

C++ method: bool QTextFormat::isCharFormat() const

C++ documentation:

Returns true if this text format is a CharFormat; otherwise returns false.

fn is_empty(&self) -> bool

C++ method: bool QTextFormat::isEmpty() const

C++ documentation:

Returns true if the format does not store any properties; false otherwise.

This function was introduced in Qt 5.3.

See also propertyCount() and properties().

fn is_frame_format(&self) -> bool

C++ method: bool QTextFormat::isFrameFormat() const

C++ documentation:

Returns true if this text format is a FrameFormat; otherwise returns false.

fn is_image_format(&self) -> bool

C++ method: bool QTextFormat::isImageFormat() const

C++ documentation:

Returns true if this text format is an image format; otherwise returns false.

fn is_list_format(&self) -> bool

C++ method: bool QTextFormat::isListFormat() const

C++ documentation:

Returns true if this text format is a ListFormat; otherwise returns false.

fn is_table_cell_format(&self) -> bool

C++ method: bool QTextFormat::isTableCellFormat() const

C++ documentation:

Returns true if this text format is a TableCellFormat; otherwise returns false.

This function was introduced in Qt 4.4.

fn is_table_format(&self) -> bool

C++ method: bool QTextFormat::isTableFormat() const

C++ documentation:

Returns true if this text format is a TableFormat; otherwise returns false.

fn is_valid(&self) -> bool

C++ method: bool QTextFormat::isValid() const

C++ documentation:

Returns true if the format is valid (i.e. is not InvalidFormat); otherwise returns false.

fn layout_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextFormat::layoutDirection() const

C++ documentation:

Returns the document's layout direction.

See also setLayoutDirection().

fn length_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatLengthPropertyArgs<'largs>

C++ method: QTextFormat::lengthProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_length::TextLength>
2) fn length_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_length::TextLength
C++ method: QTextLength QTextFormat::lengthProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthVectorProperty(), and Property.

fn length_vector_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatLengthVectorPropertyArgs<'largs>

C++ method: QTextFormat::lengthVectorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorTextLengthRef>
2) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorTextLengthRef
C++ method: QVector<QTextLength> QTextFormat::lengthVectorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId. If the property isn't of QTextFormat::LengthVector type, an empty length vector is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), and Property.

fn merge(&mut self, other: &TextFormat)

C++ method: void QTextFormat::merge(const QTextFormat& other)

C++ documentation:

Merges the other format with this format; where there are conflicts the other format takes precedence.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextFormatNewArgs

C++ method: QTextFormat::QTextFormat

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_format::TextFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_format::TextFormat>
C++ method: [constructor] void QTextFormat::QTextFormat()

Variant 2

Rust arguments:
1) fn new((&::text_format::TextFormat, cpp_utils::AsStruct)) -> ::text_format::TextFormat
2) fn new((&::text_format::TextFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_format::TextFormat>
C++ method: [constructor] void QTextFormat::QTextFormat(const QTextFormat& rhs)

Variant 3

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::text_format::TextFormat
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_format::TextFormat>
C++ method: [constructor] void QTextFormat::QTextFormat(int type)

fn object_index(&self) -> c_int

C++ method: int QTextFormat::objectIndex() const

C++ documentation:

Returns the index of the format object, or -1 if the format object is invalid.

See also setObjectIndex().

fn object_type(&self) -> c_int

C++ method: int QTextFormat::objectType() const

C++ documentation:

Returns the text format's object type.

See also ObjectTypes and setObjectType().

fn op_assign<'l0, 'l1>(&'l0 mut self, rhs: &'l1 TextFormat) -> &'l0 mut TextFormat

C++ method: QTextFormat& QTextFormat::operator=(const QTextFormat& rhs)

C++ documentation:

Assigns the other text format to this text format, and returns a reference to this text format.

fn op_eq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextFormat::operator==(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is the same as the other text format.

fn op_neq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextFormat::operator!=(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is different from the other text format.

fn pen_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatPenPropertyArgs<'largs>

C++ method: QTextFormat::penProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn pen_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pen::Pen>
2) fn pen_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::pen::Pen
C++ method: QPen QTextFormat::penProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Pen type, Qt::NoPen is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn properties<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatPropertiesArgs<'largs>

C++ method: QTextFormat::properties

This is an overloaded function. Available variants:

Rust arguments:
1) fn properties(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::map::MapCIntVariantRef>
2) fn properties(&self, cpp_utils::AsStruct) -> qt_core::map::MapCIntVariantRef
C++ method: QMap<int, QVariant> QTextFormat::properties() const

C++ documentation:

Returns a map with all properties of this text format.

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatPropertyArgs<'largs>

C++ method: QTextFormat::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QTextFormat::property(int propertyId) const

C++ documentation:

Returns the property specified by the given propertyId.

See also setProperty() and Property.

fn property_count(&self) -> c_int

C++ method: int QTextFormat::propertyCount() const

C++ documentation:

Returns the number of properties stored in the format.

This function was introduced in Qt 4.3.

fn set_background(&mut self, brush: &Brush)

C++ method: void QTextFormat::setBackground(const QBrush& brush)

C++ documentation:

Sets the brush use to paint the document's background to the brush specified.

See also background(), clearBackground(), and setForeground().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QTextFormat::setForeground(const QBrush& brush)

C++ documentation:

Sets the foreground brush to the specified brush. The foreground brush is mostly used to render text.

See also foreground(), clearForeground(), and setBackground().

fn set_layout_direction(&mut self, direction: LayoutDirection)

C++ method: void QTextFormat::setLayoutDirection(Qt::LayoutDirection direction)

C++ documentation:

Sets the document's layout direction to the specified direction.

See also layoutDirection().

fn set_object_index(&mut self, object: c_int)

C++ method: void QTextFormat::setObjectIndex(int object)

C++ documentation:

Sets the format object's object index.

See also objectIndex().

fn set_object_type(&mut self, type_: c_int)

C++ method: void QTextFormat::setObjectType(int type)

C++ documentation:

Sets the text format's object type to type.

See also ObjectTypes and objectType().

fn set_property<'largs, Args>(&'largs mut self, args: Args) where Args: TextFormatSetPropertyArgs<'largs>

C++ method: QTextFormat::setProperty

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_property(&mut self, (libc::c_int, &qt_core::variant::Variant)) -> ()
C++ method: void QTextFormat::setProperty(int propertyId, const QVariant& value)

C++ documentation:

Sets the property specified by the propertyId to the given value.

See also property() and Property.

Variant 2

Rust arguments: fn set_property(&mut self, (libc::c_int, &::vector::VectorTextLengthRef)) -> ()
C++ method: void QTextFormat::setProperty(int propertyId, const QVector<QTextLength>& lengths)

C++ documentation:

Sets the value of the property given by propertyId to value.

See also lengthVectorProperty() and Property.

fn string_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatStringPropertyArgs<'largs>

C++ method: QTextFormat::stringProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn string_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn string_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTextFormat::stringProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::String type, an empty string is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn swap(&mut self, other: &mut TextFormat)

C++ method: void QTextFormat::swap(QTextFormat& other)

C++ documentation:

Swaps this text format with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn to_block_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToBlockFormatArgs<'largs>

C++ method: QTextFormat::toBlockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_block_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
2) fn to_block_format(&self, cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
C++ method: QTextBlockFormat QTextFormat::toBlockFormat() const

C++ documentation:

Returns this format as a block format.

fn to_char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToCharFormatArgs<'largs>

C++ method: QTextFormat::toCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn to_char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextFormat::toCharFormat() const

C++ documentation:

Returns this format as a character format.

fn to_frame_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToFrameFormatArgs<'largs>

C++ method: QTextFormat::toFrameFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_frame_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame_format::TextFrameFormat>
2) fn to_frame_format(&self, cpp_utils::AsStruct) -> ::text_frame_format::TextFrameFormat
C++ method: QTextFrameFormat QTextFormat::toFrameFormat() const

C++ documentation:

Returns this format as a frame format.

fn to_image_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToImageFormatArgs<'largs>

C++ method: QTextFormat::toImageFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_image_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_image_format::TextImageFormat>
2) fn to_image_format(&self, cpp_utils::AsStruct) -> ::text_image_format::TextImageFormat
C++ method: QTextImageFormat QTextFormat::toImageFormat() const

C++ documentation:

Returns this format as an image format.

fn to_list_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToListFormatArgs<'largs>

C++ method: QTextFormat::toListFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_list_format::TextListFormat>
2) fn to_list_format(&self, cpp_utils::AsStruct) -> ::text_list_format::TextListFormat
C++ method: QTextListFormat QTextFormat::toListFormat() const

C++ documentation:

Returns this format as a list format.

fn to_table_cell_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToTableCellFormatArgs<'largs>

C++ method: QTextFormat::toTableCellFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_cell_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_cell_format::TextTableCellFormat>
2) fn to_table_cell_format(&self, cpp_utils::AsStruct) -> ::text_table_cell_format::TextTableCellFormat
C++ method: QTextTableCellFormat QTextFormat::toTableCellFormat() const

C++ documentation:

Returns this format as a table cell format.

This function was introduced in Qt 4.4.

fn to_table_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextFormatToTableFormatArgs<'largs>

C++ method: QTextFormat::toTableFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_format::TextTableFormat>
2) fn to_table_format(&self, cpp_utils::AsStruct) -> ::text_table_format::TextTableFormat
C++ method: QTextTableFormat QTextFormat::toTableFormat() const

C++ documentation:

Returns this format as a table format.

fn type_(&self) -> c_int

C++ method: int QTextFormat::type() const

C++ documentation:

Returns the type of this format.

See also FormatType.