Struct qt_gui::text_frame::Iterator

pub struct Iterator {
    // some fields omitted
}

C++ type: QTextFrame::iterator

C++ documentation:

The iterator class provides an iterator for reading the contents of a QTextFrame.

A frame consists of an arbitrary sequence of QTextBlocks and child QTextFrames. This class provides a way to iterate over the child objects of a frame, and read their contents. It does not provide a way to modify the contents of the frame.

Methods

impl Iterator

fn at_end(&self) -> bool

C++ method: bool QTextFrame::iterator::atEnd() const

C++ documentation:

Returns true if the current item is the last item in the text block.

fn current_block<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorCurrentBlockArgs<'largs>

C++ method: QTextFrame::iterator::currentBlock

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_block(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block::TextBlock>
2) fn current_block(&self, cpp_utils::AsStruct) -> ::text_block::TextBlock
C++ method: QTextBlock QTextFrame::iterator::currentBlock() const

C++ documentation:

Returns the current block the iterator points to. If the iterator points to a child frame, the returned block is invalid.

See also currentFrame().

fn current_frame(&self) -> *mut TextFrame

C++ method: QTextFrame* QTextFrame::iterator::currentFrame() const

C++ documentation:

Returns the current frame pointed to by the iterator, or 0 if the iterator currently points to a block.

See also currentBlock().

fn new<Args>(args: Args) -> Args::ReturnType where Args: IteratorNewArgs

C++ method: QTextFrame::iterator::iterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_frame::Iterator
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame::Iterator>
C++ method: [constructor] void QTextFrame::iterator::iterator()

Variant 2

Rust arguments:
1) fn new((&::text_frame::Iterator, cpp_utils::AsStruct)) -> ::text_frame::Iterator
2) fn new((&::text_frame::Iterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_frame::Iterator>
C++ method: [constructor] void QTextFrame::iterator::iterator(const QTextFrame::iterator& o)

fn op_assign<'l0, 'l1>(&'l0 mut self, o: &'l1 Iterator) -> &'l0 mut Iterator

C++ method: QTextFrame::iterator& QTextFrame::iterator::operator=(const QTextFrame::iterator& o)

C++ documentation:

Assigns other to this iterator and returns a reference to this iterator.

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QTextFrame::iterator& QTextFrame::iterator::operator--()

C++ documentation:

The prefix -- operator (--i) makes the preceding item current and returns an iterator pointing to the new current item.

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpDecPostfixArgs<'largs>

C++ method: QTextFrame::iterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_frame::Iterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_frame::Iterator
C++ method: QTextFrame::iterator QTextFrame::iterator::operator--(int arg1)

C++ documentation:

The postfix -- operator (i--) makes the preceding item current and returns an iterator to the old current item.

fn op_eq(&self, o: &Iterator) -> bool

C++ method: bool QTextFrame::iterator::operator==(const QTextFrame::iterator& o) const

C++ documentation:

Retuns true if this iterator is the same as the other iterator; otherwise returns false.

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QTextFrame::iterator& QTextFrame::iterator::operator++()

C++ documentation:

The prefix ++ operator (++i) advances the iterator to the next item in the hash and returns an iterator to the new current item.

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpIncPostfixArgs<'largs>

C++ method: QTextFrame::iterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_frame::Iterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_frame::Iterator
C++ method: QTextFrame::iterator QTextFrame::iterator::operator++(int arg1)

C++ documentation:

The postfix ++ operator (i++) advances the iterator to the next item in the text block and returns an iterator to the old current item.

fn op_neq(&self, o: &Iterator) -> bool

C++ method: bool QTextFrame::iterator::operator!=(const QTextFrame::iterator& o) const

C++ documentation:

Retuns true if this iterator is different from the other iterator; otherwise returns false.

fn parent_frame(&self) -> *mut TextFrame

C++ method: QTextFrame* QTextFrame::iterator::parentFrame() const

C++ documentation:

Returns the parent frame of the current frame.

See also currentFrame() and QTextFrame::parentFrame().