Module qt_gui::text_frame_format::overloading

Traits

TextFrameFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextFrameFormat::as_q_variant method.

TextFrameFormatBackgroundArgs
This trait represents a set of arguments accepted by TextFrameFormat::background method.

TextFrameFormatBorderBrushArgs
This trait represents a set of arguments accepted by TextFrameFormat::border_brush method.

TextFrameFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::brush_property method.

TextFrameFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::color_property method.

TextFrameFormatForegroundArgs
This trait represents a set of arguments accepted by TextFrameFormat::foreground method.

TextFrameFormatHeightArgs
This trait represents a set of arguments accepted by TextFrameFormat::height method.

TextFrameFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::length_property method.

TextFrameFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::length_vector_property method.

TextFrameFormatNewArgs
This trait represents a set of arguments accepted by TextFrameFormat::new method.

TextFrameFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::pen_property method.

TextFrameFormatPropertiesArgs
This trait represents a set of arguments accepted by TextFrameFormat::properties method.

TextFrameFormatPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::property method.

TextFrameFormatSetHeightArgs
This trait represents a set of arguments accepted by TextFrameFormat::set_height method.

TextFrameFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::set_property method.

TextFrameFormatSetWidthArgs
This trait represents a set of arguments accepted by TextFrameFormat::set_width method.

TextFrameFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextFrameFormat::string_property method.

TextFrameFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_block_format method.

TextFrameFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_char_format method.

TextFrameFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_frame_format method.

TextFrameFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_image_format method.

TextFrameFormatToListFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_list_format method.

TextFrameFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_table_cell_format method.

TextFrameFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextFrameFormat::to_table_format method.

TextFrameFormatWidthArgs
This trait represents a set of arguments accepted by TextFrameFormat::width method.