Struct qt_gui::text_inline_object::TextInlineObject

pub struct TextInlineObject {
    // some fields omitted
}

C++ type: QTextInlineObject

C++ documentation:

The QTextInlineObject class represents an inline object in a QAbstractTextDocumentLayout and its implementations.

Normally, you do not need to create a QTextInlineObject. It is used by QAbstractTextDocumentLayout to handle inline objects when implementing a custom layout.

The inline object has various attributes that can be set, for example using, setWidth(), setAscent(), and setDescent(). The rectangle it occupies is given by rect(), and its direction by textDirection(). Its position in the text layout is given by textPosition(), and its format is given by format().

Methods

impl TextInlineObject

fn ascent(&self) -> c_double

C++ method: double QTextInlineObject::ascent() const

C++ documentation:

Returns the inline object's ascent.

See also setAscent(), descent(), width(), and rect().

fn descent(&self) -> c_double

C++ method: double QTextInlineObject::descent() const

C++ documentation:

Returns the inline object's descent.

See also setDescent(), ascent(), width(), and rect().

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextInlineObjectFormatArgs<'largs>

C++ method: QTextInlineObject::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_format::TextFormat>
2) fn format(&self, cpp_utils::AsStruct) -> ::text_format::TextFormat
C++ method: QTextFormat QTextInlineObject::format() const

C++ documentation:

Returns format of the inline object within the text layout.

fn format_index(&self) -> c_int

C++ method: int QTextInlineObject::formatIndex() const

C++ documentation:

Returns an integer describing the format of the inline object within the text layout.

fn height(&self) -> c_double

C++ method: double QTextInlineObject::height() const

C++ documentation:

Returns the inline object's total height. This is equal to ascent() + descent() + 1.

See also ascent(), descent(), width(), and rect().

fn is_valid(&self) -> bool

C++ method: bool QTextInlineObject::isValid() const

C++ documentation:

Returns true if this inline object is valid; otherwise returns false.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextInlineObjectNewArgs

C++ method: QTextInlineObject::QTextInlineObject

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_inline_object::TextInlineObject
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_inline_object::TextInlineObject>
C++ method: [constructor] void QTextInlineObject::QTextInlineObject()

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextInlineObjectRectArgs<'largs>

C++ method: QTextInlineObject::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTextInlineObject::rect() const

C++ documentation:

Returns the inline object's rectangle.

See also ascent(), descent(), and width().

fn set_ascent(&mut self, a: c_double)

C++ method: void QTextInlineObject::setAscent(double a)

C++ documentation:

Sets the inline object's ascent to a.

See also ascent(), setDescent(), width(), and rect().

fn set_descent(&mut self, d: c_double)

C++ method: void QTextInlineObject::setDescent(double d)

C++ documentation:

Sets the inline object's descent to d.

See also descent(), setAscent(), width(), and rect().

fn set_width(&mut self, w: c_double)

C++ method: void QTextInlineObject::setWidth(double w)

C++ documentation:

Sets the inline object's width to w.

See also width(), ascent(), descent(), and rect().

fn text_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextInlineObject::textDirection() const

C++ documentation:

Returns if the object should be laid out right-to-left or left-to-right.

fn text_position(&self) -> c_int

C++ method: int QTextInlineObject::textPosition() const

C++ documentation:

The position of the inline object within the text layout.

fn width(&self) -> c_double

C++ method: double QTextInlineObject::width() const

C++ documentation:

Returns the inline object's width.

See also setWidth(), ascent(), descent(), and rect().