Struct qt_gui::text_item::TextItem

pub struct TextItem {
    // some fields omitted
}

C++ type: QTextItem

C++ documentation:

The QTextItem class provides all the information required to draw text in a custom paint engine.

When you reimplement your own paint engine, you must reimplement QPaintEngine::drawTextItem(), a function that takes a QTextItem as one of its arguments.

Methods

impl TextItem

fn ascent(&self) -> c_double

C++ method: double QTextItem::ascent() const

C++ documentation:

Corresponds to the ascent of the piece of text that is drawn.

fn descent(&self) -> c_double

C++ method: double QTextItem::descent() const

C++ documentation:

Corresponds to the descent of the piece of text that is drawn.

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextItemFontArgs<'largs>

C++ method: QTextItem::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> ::font::Font
C++ method: QFont QTextItem::font() const

C++ documentation:

Returns the font that should be used to draw the text.

fn render_flags(&self) -> Flags<RenderFlag>

C++ method: QFlags<QTextItem::RenderFlag> QTextItem::renderFlags() const

C++ documentation:

Returns the render flags used.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextItemTextArgs<'largs>

C++ method: QTextItem::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextItem::text() const

C++ documentation:

Returns the text that should be drawn.

fn width(&self) -> c_double

C++ method: double QTextItem::width() const

C++ documentation:

Specifies the total width of the text to be drawn.