Module qt_gui::text_layout::overloading

Traits

TextLayoutAdditionalFormatsArgs
This trait represents a set of arguments accepted by TextLayout::additional_formats method.

TextLayoutBoundingRectArgs
This trait represents a set of arguments accepted by TextLayout::bounding_rect method.

TextLayoutCreateLineArgs
This trait represents a set of arguments accepted by TextLayout::create_line method.

TextLayoutDrawArgs
This trait represents a set of arguments accepted by TextLayout::draw method.

TextLayoutDrawCursorArgs
This trait represents a set of arguments accepted by TextLayout::draw_cursor method.

TextLayoutFontArgs
This trait represents a set of arguments accepted by TextLayout::font method.

TextLayoutGlyphRunsArgs
This trait represents a set of arguments accepted by TextLayout::glyph_runs method.

TextLayoutLineAtArgs
This trait represents a set of arguments accepted by TextLayout::line_at method.

TextLayoutLineForTextPositionArgs
This trait represents a set of arguments accepted by TextLayout::line_for_text_position method.

TextLayoutNewArgs
This trait represents a set of arguments accepted by TextLayout::new method.

TextLayoutNextCursorPositionArgs
This trait represents a set of arguments accepted by TextLayout::next_cursor_position method.

TextLayoutPositionArgs
This trait represents a set of arguments accepted by TextLayout::position method.

TextLayoutPreeditAreaTextArgs
This trait represents a set of arguments accepted by TextLayout::preedit_area_text method.

TextLayoutPreviousCursorPositionArgs
This trait represents a set of arguments accepted by TextLayout::previous_cursor_position method.

TextLayoutTextArgs
This trait represents a set of arguments accepted by TextLayout::text method.