Struct qt_gui::text_layout::FormatRange

pub struct FormatRange {
    // some fields omitted
}

C++ type: QTextLayout::FormatRange

C++ documentation:

The QTextLayout::FormatRange structure is used to apply extra formatting information for a specified area in the text layout's content.

Methods

impl FormatRange

fn format<'l0>(&'l0 self) -> &'l0 TextCharFormat

C++ method: const QTextCharFormat& QTextLayout::FormatRange::format() const

fn format_mut<'l0>(&'l0 mut self) -> &'l0 mut TextCharFormat

C++ method: QTextCharFormat& QTextLayout::FormatRange::format_mut()

fn length(&self) -> c_int

C++ method: int QTextLayout::FormatRange::length() const

fn set_format(&mut self, value: &TextCharFormat)

C++ method: void QTextLayout::FormatRange::set_format(QTextCharFormat value)

fn set_length(&mut self, value: c_int)

C++ method: void QTextLayout::FormatRange::set_length(int value)

fn set_start(&mut self, value: c_int)

C++ method: void QTextLayout::FormatRange::set_start(int value)

fn start(&self) -> c_int

C++ method: int QTextLayout::FormatRange::start() const