Struct qt_gui::text_layout::TextLayout

pub struct TextLayout {
    // some fields omitted
}

C++ type: QTextLayout

C++ documentation:

The QTextLayout class is used to lay out and render text.

It offers many features expected from a modern text layout engine, including Unicode compliant rendering, line breaking and handling of cursor positioning. It can also produce and render device independent layout, something that is important for WYSIWYG applications.

The class has a rather low level API and unless you intend to implement your own text rendering for some specialized widget, you probably won't need to use it directly.

QTextLayout can be used with both plain and rich text.

QTextLayout can be used to create a sequence of QTextLine instances with given widths and can position them independently on the screen. Once the layout is done, these lines can be drawn on a paint device.

The text to be laid out can be provided in the constructor or set with setText().

The layout can be seen as a sequence of QTextLine objects; use createLine() to create a QTextLine instance, and lineAt() or lineForTextPosition() to retrieve created lines.

Here is a code snippet that demonstrates the layout phase:

int leading = fontMetrics.leading();
qreal height = 0;
textLayout.setCacheEnabled(true);
textLayout.beginLayout();
while (1) {
    QTextLine line = textLayout.createLine();
    if (!line.isValid())
        break;

line.setLineWidth(lineWidth); height += leading; line.setPosition(QPointF(0, height)); height += line.height(); } textLayout.endLayout();
The text can then be rendered by calling the layout's draw() function:

QPainter painter(this);
textLayout.draw(&painter, QPoint(0, 0));

For a given position in the text you can find a valid cursor position with isValidCursorPosition(), nextCursorPosition(), and previousCursorPosition().

The QTextLayout itself can be positioned with setPosition(); it has a boundingRect(), and a minimumWidth() and a maximumWidth().

Methods

impl TextLayout

fn additional_formats<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutAdditionalFormatsArgs<'largs>

C++ method: QTextLayout::additionalFormats

This is an overloaded function. Available variants:

Rust arguments:
1) fn additional_formats(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListFormatRangeRef>
2) fn additional_formats(&self, cpp_utils::AsStruct) -> ::list::ListFormatRangeRef
C++ method: QList<QTextLayout::FormatRange> QTextLayout::additionalFormats() const

C++ documentation:

Returns the list of additional formats supported by the text layout.

See also setAdditionalFormats() and clearAdditionalFormats().

fn begin_layout(&mut self)

C++ method: void QTextLayout::beginLayout()

C++ documentation:

Begins the layout process.

See also endLayout().

fn bounding_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutBoundingRectArgs<'largs>

C++ method: QTextLayout::boundingRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounding_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn bounding_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTextLayout::boundingRect() const

C++ documentation:

The smallest rectangle that contains all the lines in the layout.

fn cache_enabled(&self) -> bool

C++ method: bool QTextLayout::cacheEnabled() const

C++ documentation:

Returns true if the complete layout information is cached; otherwise returns false.

See also setCacheEnabled().

fn clear_additional_formats(&mut self)

C++ method: void QTextLayout::clearAdditionalFormats()

C++ documentation:

Clears the list of additional formats supported by the text layout.

See also additionalFormats() and setAdditionalFormats().

fn clear_layout(&mut self)

C++ method: void QTextLayout::clearLayout()

C++ documentation:

Clears the line information in the layout. After having called this function, lineCount() returns 0.

This function was introduced in Qt 4.4.

fn create_line<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextLayoutCreateLineArgs<'largs>

C++ method: QTextLayout::createLine

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_line(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_line::TextLine>
2) fn create_line(&mut self, cpp_utils::AsStruct) -> ::text_line::TextLine
C++ method: QTextLine QTextLayout::createLine()

C++ documentation:

Returns a new text line to be laid out if there is text to be inserted into the layout; otherwise returns an invalid text line.

The text layout creates a new line object that starts after the last line in the layout, or at the beginning if the layout is empty. The layout maintains an internal cursor, and each line is filled with text from the cursor position onwards when the QTextLine::setLineWidth() function is called.

Once QTextLine::setLineWidth() is called, a new line can be created and filled with text. Repeating this process will lay out the whole block of text contained in the QTextLayout. If there is no text left to be inserted into the layout, the QTextLine returned will not be valid (isValid() will return false).

fn cursor_move_style(&self) -> CursorMoveStyle

C++ method: Qt::CursorMoveStyle QTextLayout::cursorMoveStyle() const

C++ documentation:

The cursor movement style of this QTextLayout. The default is Qt::LogicalMoveStyle.

See also setCursorMoveStyle().

fn draw<'largs, Args>(&'largs self, args: Args) where Args: TextLayoutDrawArgs<'largs>

C++ method: QTextLayout::draw

This is an overloaded function. Available variants:

Rust arguments:
1) fn draw(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF)) -> ()
2) fn draw(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF, &::vector::VectorFormatRangeRef)) -> ()
3) fn draw(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF, &::vector::VectorFormatRangeRef, &qt_core::rect_f::RectF)) -> ()
C++ method: void QTextLayout::draw(QPainter* p, const QPointF& pos, const QVector<QTextLayout::FormatRange>& selections = ?, const QRectF& clip = ?) const

C++ documentation:

Draws the whole layout on the painter p at the position specified by pos. The rendered layout includes the given selections and is clipped within the rectangle specified by clip.

fn draw_cursor<'largs, Args>(&'largs self, args: Args) where Args: TextLayoutDrawCursorArgs<'largs>

C++ method: QTextLayout::drawCursor

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn draw_cursor(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF, libc::c_int)) -> ()
C++ method: void QTextLayout::drawCursor(QPainter* p, const QPointF& pos, int cursorPosition) const

C++ documentation:

This is an overloaded function.

Draws a text cursor with the current pen at the given position using the painter specified. The corresponding position within the text is specified by cursorPosition.

Variant 2

Rust arguments: fn draw_cursor(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF, libc::c_int, libc::c_int)) -> ()
C++ method: void QTextLayout::drawCursor(QPainter* p, const QPointF& pos, int cursorPosition, int width) const

C++ documentation:

Draws a text cursor with the current pen and the specified width at the given position using the painter specified. The corresponding position within the text is specified by cursorPosition.

fn end_layout(&mut self)

C++ method: void QTextLayout::endLayout()

C++ documentation:

Ends the layout process.

See also beginLayout().

fn font<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutFontArgs<'largs>

C++ method: QTextLayout::font

This is an overloaded function. Available variants:

Rust arguments:
1) fn font(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::font::Font>
2) fn font(&self, cpp_utils::AsStruct) -> ::font::Font
C++ method: QFont QTextLayout::font() const

C++ documentation:

Returns the current font that is used for the layout, or a default font if none is set.

See also setFont().

fn glyph_runs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutGlyphRunsArgs<'largs>

C++ method: QTextLayout::glyphRuns

This is an overloaded function. Available variants:

Rust arguments:
1) fn glyph_runs(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
2) fn glyph_runs(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
3) fn glyph_runs(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
4) fn glyph_runs(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListGlyphRunRef
5) fn glyph_runs(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListGlyphRunRef
6) fn glyph_runs(&self, cpp_utils::AsStruct) -> ::list::ListGlyphRunRef
C++ method: QList<QGlyphRun> QTextLayout::glyphRuns(int from = ?, int length = ?) const

C++ documentation:

Returns the glyph indexes and positions for all glyphs corresponding to the length characters starting at the position from in this QTextLayout. This is an expensive function, and should not be called in a time sensitive context.

If from is less than zero, then the glyph run will begin at the first character in the layout. If length is less than zero, it will span the entire string from the start position.

This function was introduced in Qt 4.8.

See also draw() and QPainter::drawGlyphRun().

fn is_valid_cursor_position(&self, pos: c_int) -> bool

C++ method: bool QTextLayout::isValidCursorPosition(int pos) const

C++ documentation:

/ Returns true if position pos is a valid cursor position.

In a Unicode context some positions in the text are not valid cursor positions, because the position is inside a Unicode surrogate or a grapheme cluster.

A grapheme cluster is a sequence of two or more Unicode characters that form one indivisible entity on the screen. For example the latin character Ä' can be represented in Unicode by two characters,A' (0x41), and the combining diaresis (0x308). A text cursor can only validly be positioned before or after these two characters, never between them since that wouldn't make sense. In indic languages every syllable forms a grapheme cluster.

fn left_cursor_position(&self, old_pos: c_int) -> c_int

C++ method: int QTextLayout::leftCursorPosition(int oldPos) const

C++ documentation:

Returns the cursor position to the left of oldPos, next to it. It's dependent on the visual position of characters, after bi-directional reordering.

See also rightCursorPosition() and previousCursorPosition().

fn line_at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutLineAtArgs<'largs>

C++ method: QTextLayout::lineAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn line_at(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_line::TextLine>
2) fn line_at(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_line::TextLine
C++ method: QTextLine QTextLayout::lineAt(int i) const

C++ documentation:

Returns the i-th line of text in this text layout.

See also lineCount() and lineForTextPosition().

fn line_count(&self) -> c_int

C++ method: int QTextLayout::lineCount() const

C++ documentation:

Returns the number of lines in this text layout.

See also lineAt().

fn line_for_text_position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutLineForTextPositionArgs<'largs>

C++ method: QTextLayout::lineForTextPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn line_for_text_position(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_line::TextLine>
2) fn line_for_text_position(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_line::TextLine
C++ method: QTextLine QTextLayout::lineForTextPosition(int pos) const

C++ documentation:

Returns the line that contains the cursor position specified by pos.

See also isValidCursorPosition() and lineAt().

fn maximum_width(&self) -> c_double

C++ method: double QTextLayout::maximumWidth() const

C++ documentation:

The maximum width the layout could expand to; this is essentially the width of the entire text.

Warning: This function only returns a valid value after the layout has been done.

See also minimumWidth().

fn minimum_width(&self) -> c_double

C++ method: double QTextLayout::minimumWidth() const

C++ documentation:

The minimum width the layout needs. This is the width of the layout's smallest non-breakable substring.

Warning: This function only returns a valid value after the layout has been done.

See also maximumWidth().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextLayoutNewArgs

C++ method: QTextLayout::QTextLayout

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_layout::TextLayout
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_layout::TextLayout>
C++ method: [constructor] void QTextLayout::QTextLayout()

Variant 2

Rust arguments:
1) fn new((&qt_core::string::String, cpp_utils::AsStruct)) -> ::text_layout::TextLayout
2) fn new((&qt_core::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::TextLayout>
C++ method: [constructor] void QTextLayout::QTextLayout(const QString& text)

Variant 3

Rust arguments:
1) fn new((&qt_core::string::String, &::font::Font, cpp_utils::AsStruct)) -> ::text_layout::TextLayout
2) fn new((&qt_core::string::String, &::font::Font, *mut ::paint_device::PaintDevice, cpp_utils::AsStruct)) -> ::text_layout::TextLayout
3) fn new((&qt_core::string::String, &::font::Font, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::TextLayout>
4) fn new((&qt_core::string::String, &::font::Font, *mut ::paint_device::PaintDevice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::TextLayout>
C++ method: [constructor] void QTextLayout::QTextLayout(const QString& text, const QFont& font, QPaintDevice* paintdevice = ?)

Variant 4

Rust arguments:
1) fn new((&::text_block::TextBlock, cpp_utils::AsStruct)) -> ::text_layout::TextLayout
2) fn new((&::text_block::TextBlock, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::TextLayout>
C++ method: [constructor] void QTextLayout::QTextLayout(const QTextBlock& b)

fn next_cursor_position<'largs, Args>(&'largs self, args: Args) -> c_int where Args: TextLayoutNextCursorPositionArgs<'largs>

C++ method: QTextLayout::nextCursorPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_cursor_position(&self, libc::c_int) -> libc::c_int
2) fn next_cursor_position(&self, (libc::c_int, ::text_layout::CursorMode)) -> libc::c_int
C++ method: int QTextLayout::nextCursorPosition(int oldPos, QTextLayout::CursorMode mode = ?) const

C++ documentation:

Returns the next valid cursor position after oldPos that respects the given cursor mode. Returns value of oldPos, if oldPos is not a valid cursor position.

See also isValidCursorPosition() and previousCursorPosition().

fn position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutPositionArgs<'largs>

C++ method: QTextLayout::position

This is an overloaded function. Available variants:

Rust arguments:
1) fn position(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn position(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTextLayout::position() const

C++ documentation:

The global position of the layout. This is independent of the bounding rectangle and of the layout process.

This function was introduced in Qt 4.2.

See also setPosition().

fn preedit_area_position(&self) -> c_int

C++ method: int QTextLayout::preeditAreaPosition() const

C++ documentation:

Returns the position of the area in the text layout that will be processed before editing occurs.

See also preeditAreaText().

fn preedit_area_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutPreeditAreaTextArgs<'largs>

C++ method: QTextLayout::preeditAreaText

This is an overloaded function. Available variants:

Rust arguments:
1) fn preedit_area_text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn preedit_area_text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextLayout::preeditAreaText() const

C++ documentation:

Returns the text that is inserted in the layout before editing occurs.

See also preeditAreaPosition().

fn previous_cursor_position<'largs, Args>(&'largs self, args: Args) -> c_int where Args: TextLayoutPreviousCursorPositionArgs<'largs>

C++ method: QTextLayout::previousCursorPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_cursor_position(&self, libc::c_int) -> libc::c_int
2) fn previous_cursor_position(&self, (libc::c_int, ::text_layout::CursorMode)) -> libc::c_int
C++ method: int QTextLayout::previousCursorPosition(int oldPos, QTextLayout::CursorMode mode = ?) const

C++ documentation:

Returns the first valid cursor position before oldPos that respects the given cursor mode. Returns value of oldPos, if oldPos is not a valid cursor position.

See also isValidCursorPosition() and nextCursorPosition().

fn right_cursor_position(&self, old_pos: c_int) -> c_int

C++ method: int QTextLayout::rightCursorPosition(int oldPos) const

C++ documentation:

Returns the cursor position to the right of oldPos, next to it. It's dependent on the visual position of characters, after bi-directional reordering.

See also leftCursorPosition() and nextCursorPosition().

fn set_additional_formats(&mut self, overrides: &ListFormatRangeRef)

C++ method: void QTextLayout::setAdditionalFormats(const QList<QTextLayout::FormatRange>& overrides)

C++ documentation:

Sets the additional formats supported by the text layout to formatList. The formats are applied with preedit area text in place.

See also additionalFormats() and clearAdditionalFormats().

fn set_cache_enabled(&mut self, enable: bool)

C++ method: void QTextLayout::setCacheEnabled(bool enable)

C++ documentation:

Enables caching of the complete layout information if enable is true; otherwise disables layout caching. Usually QTextLayout throws most of the layouting information away after a call to endLayout() to reduce memory consumption. If you however want to draw the laid out text directly afterwards enabling caching might speed up drawing significantly.

See also cacheEnabled().

fn set_cursor_move_style(&mut self, style: CursorMoveStyle)

C++ method: void QTextLayout::setCursorMoveStyle(Qt::CursorMoveStyle style)

C++ documentation:

Sets the visual cursor movement style to the given style. If the QTextLayout is backed by a document, you can ignore this and use the option in QTextDocument, this option is for widgets like QLineEdit or custom widgets without a QTextDocument. Default value is Qt::LogicalMoveStyle.

See also cursorMoveStyle().

fn set_flags(&mut self, flags: c_int)

C++ method: void QTextLayout::setFlags(int flags)

fn set_font(&mut self, f: &Font)

C++ method: void QTextLayout::setFont(const QFont& f)

C++ documentation:

Sets the layout's font to the given font. The layout is invalidated and must be laid out again.

See also font().

fn set_position(&mut self, p: &PointF)

C++ method: void QTextLayout::setPosition(const QPointF& p)

C++ documentation:

Moves the text layout to point p.

See also position().

fn set_preedit_area(&mut self, position: c_int, text: &String)

C++ method: void QTextLayout::setPreeditArea(int position, const QString& text)

C++ documentation:

Sets the position and text of the area in the layout that is processed before editing occurs. The layout is invalidated and must be laid out again.

See also preeditAreaPosition() and preeditAreaText().

fn set_raw_font(&mut self, raw_font: &RawFont)

C++ method: void QTextLayout::setRawFont(const QRawFont& rawFont)

fn set_text(&mut self, string: &String)

C++ method: void QTextLayout::setText(const QString& string)

C++ documentation:

Sets the layout's text to the given string. The layout is invalidated and must be laid out again.

Notice that when using this QTextLayout as part of a QTextDocument this method will have no effect.

See also text().

fn set_text_option(&mut self, option: &TextOption)

C++ method: void QTextLayout::setTextOption(const QTextOption& option)

C++ documentation:

Sets the text option structure that controls the layout process to the given option.

See also textOption().

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLayoutTextArgs<'largs>

C++ method: QTextLayout::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn text(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextLayout::text() const

C++ documentation:

Returns the layout's text.

See also setText().

fn text_option<'l0>(&'l0 self) -> &'l0 TextOption

C++ method: const QTextOption& QTextLayout::textOption() const

C++ documentation:

Returns the current text option used to control the layout process.

See also setTextOption().