Struct qt_gui::text_length::TextLength

pub struct TextLength {
    // some fields omitted
}

C++ type: QTextLength

C++ documentation:

The QTextLength class encapsulates the different types of length used in a QTextDocument.

When we specify a value for the length of an element in a text document, we often need to provide some other information so that the length is used in the way we expect. For example, when we specify a table width, the value can represent a fixed number of pixels, or it can be a percentage value. This information changes both the meaning of the value and the way it is used.

Generally, this class is used to specify table widths. These can be specified either as a fixed amount of pixels, as a percentage of the containing frame's width, or by a variable width that allows it to take up just the space it requires.

Methods

impl TextLength

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLengthAsQVariantArgs<'largs>

C++ method: QTextLength::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTextLength::operator QVariant() const

C++ documentation:

Returns the text length as a QVariant

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextLengthNewArgs

C++ method: QTextLength::QTextLength

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_length::TextLength
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_length::TextLength>
C++ method: [constructor] void QTextLength::QTextLength()

Variant 2

Rust arguments:
1) fn new((::text_length::Type, libc::c_double, cpp_utils::AsStruct)) -> ::text_length::TextLength
2) fn new((::text_length::Type, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_length::TextLength>
C++ method: [constructor] void QTextLength::QTextLength(QTextLength::Type type, double value)

fn op_eq(&self, other: &TextLength) -> bool

C++ method: bool QTextLength::operator==(const QTextLength& other) const

C++ documentation:

Returns true if this text length is the same as the other text length.

fn op_neq(&self, other: &TextLength) -> bool

C++ method: bool QTextLength::operator!=(const QTextLength& other) const

C++ documentation:

Returns true if this text length is different from the other text length.

fn raw_value(&self) -> c_double

C++ method: double QTextLength::rawValue() const

C++ documentation:

Returns the constraint value that is specific for the type of the length. If the length is QTextLength::PercentageLength then the raw value is in percent, in the range of 0 to 100. If the length is QTextLength::FixedLength then that fixed amount is returned. For variable lengths, zero is returned.

fn type_(&self) -> Type

C++ method: QTextLength::Type QTextLength::type() const

C++ documentation:

Returns the type of this length object.

See also QTextLength::Type.

fn value(&self, maximum_length: c_double) -> c_double

C++ method: double QTextLength::value(double maximumLength) const

C++ documentation:

Returns the effective length, constrained by the type of the length object and the specified maximumLength.

See also type().