Struct qt_gui::text_line::TextLine

pub struct TextLine {
    // some fields omitted
}

C++ type: QTextLine

C++ documentation:

The QTextLine class represents a line of text inside a QTextLayout.

A text line is usually created by QTextLayout::createLine().

After being created, the line can be filled using the setLineWidth() or setNumColumns() functions. A line has a number of attributes including the rectangle it occupies, rect(), its coordinates, x() and y(), its textLength(), width() and naturalTextWidth(), and its ascent() and descent() relative to the text. The position of the cursor in terms of the line is available from cursorToX() and its inverse from xToCursor(). A line can be moved with setPosition().

Methods

impl TextLine

fn ascent(&self) -> c_double

C++ method: double QTextLine::ascent() const

C++ documentation:

Returns the line's ascent.

See also descent() and height().

fn cursor_to_x<'largs, Args>(&'largs self, args: Args) -> c_double where Args: TextLineCursorToXArgs<'largs>

C++ method: QTextLine::cursorToX

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn cursor_to_x(&self, libc::c_int) -> libc::c_double
2) fn cursor_to_x(&self, (libc::c_int, ::text_line::Edge)) -> libc::c_double
C++ method: double QTextLine::cursorToX(int cursorPos, QTextLine::Edge edge = ?) const

C++ documentation:

This is an overloaded function.

Variant 2

Rust arguments:
1) fn cursor_to_x(&self, *mut libc::c_int) -> libc::c_double
2) fn cursor_to_x(&self, (*mut libc::c_int, ::text_line::Edge)) -> libc::c_double
C++ method: double QTextLine::cursorToX(int* cursorPos, QTextLine::Edge edge = ?) const

C++ documentation:

Converts the cursor position cursorPos to the corresponding x position inside the line, taking account of the edge.

If cursorPos is not a valid cursor position, the nearest valid cursor position will be used instead, and cpos will be modified to point to this valid cursor position.

See also xToCursor().

fn descent(&self) -> c_double

C++ method: double QTextLine::descent() const

C++ documentation:

Returns the line's descent.

See also ascent() and height().

fn draw<'largs, Args>(&'largs self, args: Args) where Args: TextLineDrawArgs<'largs>

C++ method: QTextLine::draw

This is an overloaded function. Available variants:

Rust arguments:
1) fn draw(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF)) -> ()
2) fn draw(&self, (*mut ::painter::Painter, &qt_core::point_f::PointF, *const ::text_layout::FormatRange)) -> ()
C++ method: void QTextLine::draw(QPainter* p, const QPointF& point, const QTextLayout::FormatRange* selection = ?) const

C++ documentation:

Draws a line on the given painter at the specified position. The selection is reserved for internal use.

fn glyph_runs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLineGlyphRunsArgs<'largs>

C++ method: QTextLine::glyphRuns

This is an overloaded function. Available variants:

Rust arguments:
1) fn glyph_runs(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
2) fn glyph_runs(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
3) fn glyph_runs(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListGlyphRunRef>
4) fn glyph_runs(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListGlyphRunRef
5) fn glyph_runs(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListGlyphRunRef
6) fn glyph_runs(&self, cpp_utils::AsStruct) -> ::list::ListGlyphRunRef
C++ method: QList<QGlyphRun> QTextLine::glyphRuns(int from = ?, int length = ?) const

C++ documentation:

Returns the glyph indexes and positions for all glyphs in this QTextLine for characters in the range defined by from and length. The from index is relative to the beginning of the text in the containing QTextLayout, and the range must be within the range of QTextLine as given by functions textStart() and textLength().

If from is negative, it will default to textStart(), and if length is negative it will default to the return value of textLength().

This function was introduced in Qt 5.0.

See also QTextLayout::glyphRuns().

fn height(&self) -> c_double

C++ method: double QTextLine::height() const

C++ documentation:

Returns the line's height. This is equal to ascent() + descent() if leading is not included. If leading is included, this equals to ascent() + descent() + leading().

See also ascent(), descent(), leading(), and setLeadingIncluded().

fn horizontal_advance(&self) -> c_double

C++ method: double QTextLine::horizontalAdvance() const

C++ documentation:

Returns the horizontal advance of the text. The advance of the text is the distance from its position to the next position at which text would naturally be drawn.

By adding the advance to the position of the text line and using this as the position of a second text line, you will be able to position the two lines side-by-side without gaps in-between.

This function was introduced in Qt 4.7.

fn is_valid(&self) -> bool

C++ method: bool QTextLine::isValid() const

C++ documentation:

Returns true if this text line is valid; otherwise returns false.

fn leading(&self) -> c_double

C++ method: double QTextLine::leading() const

C++ documentation:

Returns the line's leading.

This function was introduced in Qt 4.6.

See also ascent(), descent(), and height().

fn leading_included(&self) -> bool

C++ method: bool QTextLine::leadingIncluded() const

C++ documentation:

Returns true if positive leading is included into the line's height; otherwise returns false.

By default, leading is not included.

This function was introduced in Qt 4.6.

See also setLeadingIncluded().

fn line_number(&self) -> c_int

C++ method: int QTextLine::lineNumber() const

C++ documentation:

Returns the position of the line in the text engine.

fn natural_text_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLineNaturalTextRectArgs<'largs>

C++ method: QTextLine::naturalTextRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn natural_text_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn natural_text_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTextLine::naturalTextRect() const

C++ documentation:

Returns the rectangle covered by the line.

fn natural_text_width(&self) -> c_double

C++ method: double QTextLine::naturalTextWidth() const

C++ documentation:

Returns the width of the line that is occupied by text. This is always <= to width(), and is the minimum width that could be used by layout() without changing the line break position.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextLineNewArgs

C++ method: QTextLine::QTextLine

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_line::TextLine
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_line::TextLine>
C++ method: [constructor] void QTextLine::QTextLine()

fn position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLinePositionArgs<'largs>

C++ method: QTextLine::position

This is an overloaded function. Available variants:

Rust arguments:
1) fn position(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn position(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTextLine::position() const

C++ documentation:

Returns the line's position relative to the text layout's position.

See also setPosition().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextLineRectArgs<'largs>

C++ method: QTextLine::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTextLine::rect() const

C++ documentation:

Returns the line's bounding rectangle.

See also x(), y(), textLength(), and width().

fn set_leading_included(&mut self, included: bool)

C++ method: void QTextLine::setLeadingIncluded(bool included)

C++ documentation:

Includes positive leading into the line's height if included is true; otherwise does not include leading.

By default, leading is not included.

Note that negative leading is ignored, it must be handled in the code using the text lines by letting the lines overlap.

This function was introduced in Qt 4.6.

See also leadingIncluded().

fn set_line_width(&mut self, width: c_double)

C++ method: void QTextLine::setLineWidth(double width)

C++ documentation:

Lays out the line with the given width. The line is filled from its starting position with as many characters as will fit into the line. In case the text cannot be split at the end of the line, it will be filled with additional characters to the next whitespace or end of the text.

fn set_num_columns<'largs, Args>(&'largs mut self, args: Args) where Args: TextLineSetNumColumnsArgs<'largs>

C++ method: QTextLine::setNumColumns

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_num_columns(&mut self, libc::c_int) -> ()
C++ method: void QTextLine::setNumColumns(int columns)

C++ documentation:

Lays out the line. The line is filled from its starting position with as many characters as are specified by numColumns. In case the text cannot be split until numColumns characters, the line will be filled with as many characters to the next whitespace or end of the text.

Variant 2

Rust arguments: fn set_num_columns(&mut self, (libc::c_int, libc::c_double)) -> ()
C++ method: void QTextLine::setNumColumns(int columns, double alignmentWidth)

C++ documentation:

Lays out the line. The line is filled from its starting position with as many characters as are specified by numColumns. In case the text cannot be split until numColumns characters, the line will be filled with as many characters to the next whitespace or end of the text. The provided alignmentWidth is used as reference width for alignment.

fn set_position(&mut self, pos: &PointF)

C++ method: void QTextLine::setPosition(const QPointF& pos)

C++ documentation:

Moves the line to position pos.

See also position().

fn text_length(&self) -> c_int

C++ method: int QTextLine::textLength() const

C++ documentation:

Returns the length of the text in the line.

See also naturalTextWidth().

fn text_start(&self) -> c_int

C++ method: int QTextLine::textStart() const

C++ documentation:

Returns the start of the line from the beginning of the string passed to the QTextLayout.

fn width(&self) -> c_double

C++ method: double QTextLine::width() const

C++ documentation:

Returns the line's width as specified by the layout() function.

See also naturalTextWidth(), x(), y(), textLength(), and rect().

fn x(&self) -> c_double

C++ method: double QTextLine::x() const

C++ documentation:

Returns the line's x position.

See also rect(), y(), textLength(), and width().

fn x_to_cursor<'largs, Args>(&'largs self, args: Args) -> c_int where Args: TextLineXToCursorArgs<'largs>

C++ method: QTextLine::xToCursor

This is an overloaded function. Available variants:

Rust arguments:
1) fn x_to_cursor(&self, libc::c_double) -> libc::c_int
2) fn x_to_cursor(&self, (libc::c_double, ::text_line::CursorPosition)) -> libc::c_int
C++ method: int QTextLine::xToCursor(double x, QTextLine::CursorPosition arg2 = ?) const

C++ documentation:

Converts the x-coordinate x, to the nearest matching cursor position, depending on the cursor position type, cpos. Note that result cursor position includes possible preedit area text.

See also cursorToX().

fn y(&self) -> c_double

C++ method: double QTextLine::y() const

C++ documentation:

Returns the line's y position.

See also x(), rect(), textLength(), and width().