Module qt_gui::text_list::overloading

Traits

TextListConnectArgs
This trait represents a set of arguments accepted by TextList::connect method.

TextListConnectStaticArgs
This trait represents a set of arguments accepted by TextList::connect_static method.

TextListDestroyedArgs
This trait represents a set of arguments accepted by TextList::destroyed method.

TextListDisconnectArgs
This trait represents a set of arguments accepted by TextList::disconnect method.

TextListDisconnectStaticArgs
This trait represents a set of arguments accepted by TextList::disconnect_static method.

TextListDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TextList::dynamic_property_names method.

TextListFormatArgs
This trait represents a set of arguments accepted by TextList::format method.

TextListItemArgs
This trait represents a set of arguments accepted by TextList::item method.

TextListItemTextArgs
This trait represents a set of arguments accepted by TextList::item_text method.

TextListNewArgs
This trait represents a set of arguments accepted by TextList::new method.

TextListObjectNameArgs
This trait represents a set of arguments accepted by TextList::object_name method.

TextListPropertyArgs
This trait represents a set of arguments accepted by TextList::property method.

TextListStartTimerArgs
This trait represents a set of arguments accepted by TextList::start_timer method.

TextListTrArgs
This trait represents a set of arguments accepted by TextList::tr method.

TextListTrUtf8Args
This trait represents a set of arguments accepted by TextList::tr_utf8 method.