Module qt_gui::text_list_format::overloading

Traits

TextListFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextListFormat::as_q_variant method.

TextListFormatBackgroundArgs
This trait represents a set of arguments accepted by TextListFormat::background method.

TextListFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::brush_property method.

TextListFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::color_property method.

TextListFormatForegroundArgs
This trait represents a set of arguments accepted by TextListFormat::foreground method.

TextListFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::length_property method.

TextListFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::length_vector_property method.

TextListFormatNewArgs
This trait represents a set of arguments accepted by TextListFormat::new method.

TextListFormatNumberPrefixArgs
This trait represents a set of arguments accepted by TextListFormat::number_prefix method.

TextListFormatNumberSuffixArgs
This trait represents a set of arguments accepted by TextListFormat::number_suffix method.

TextListFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::pen_property method.

TextListFormatPropertiesArgs
This trait represents a set of arguments accepted by TextListFormat::properties method.

TextListFormatPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::property method.

TextListFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::set_property method.

TextListFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextListFormat::string_property method.

TextListFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_block_format method.

TextListFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_char_format method.

TextListFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_frame_format method.

TextListFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_image_format method.

TextListFormatToListFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_list_format method.

TextListFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_table_cell_format method.

TextListFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextListFormat::to_table_format method.