Struct qt_gui::text_list_format::TextListFormat

pub struct TextListFormat {
    // some fields omitted
}

C++ type: QTextListFormat

C++ documentation:

The QTextListFormat class provides formatting information for lists in a QTextDocument.

A list is composed of one or more items, represented as text blocks. The list's format specifies the appearance of items in the list. In particular, it determines the indentation and the style of each item.

The indentation of the items is an integer value that causes each item to be offset from the left margin by a certain amount. This value is read with indent() and set with setIndent().

The style used to decorate each item is set with setStyle() and can be read with the style() function. The style controls the type of bullet points and numbering scheme used for items in the list. Note that lists that use the decimal numbering scheme begin counting at 1 rather than 0.

Style properties can be set to further configure the appearance of list items; for example, the ListNumberPrefix and ListNumberSuffix properties can be used to customize the numbers used in an ordered list so that they appear as (1), (2), (3), etc.:

QTextListFormat listFormat;

listFormat.setStyle(QTextListFormat::ListDecimal); listFormat.setNumberPrefix("("); listFormat.setNumberSuffix(")");

cursor.insertList(listFormat);

Methods

impl TextListFormat

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatAsQVariantArgs<'largs>

C++ method: QTextListFormat::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTextListFormat::operator QVariant() const

C++ documentation:

Returns the text format as a QVariant

fn background<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatBackgroundArgs<'largs>

C++ method: QTextListFormat::background

This is an overloaded function. Available variants:

Rust arguments:
1) fn background(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn background(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextListFormat::background() const

C++ documentation:

Returns the brush used to paint the document's background.

See also setBackground(), clearBackground(), and foreground().

fn bool_property(&self, property_id: c_int) -> bool

C++ method: bool QTextListFormat::boolProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QTextFormat::Bool type, false is returned instead.

See also setProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn brush_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatBrushPropertyArgs<'largs>

C++ method: QTextListFormat::brushProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn brush_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::brush::Brush>
2) fn brush_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::brush::Brush
C++ method: QBrush QTextListFormat::brushProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Brush type, Qt::NoBrush is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn clear_background(&mut self)

C++ method: void QTextListFormat::clearBackground()

C++ documentation:

Clears the brush used to paint the document's background. The default brush will be used.

See also background(), setBackground(), and clearForeground().

fn clear_foreground(&mut self)

C++ method: void QTextListFormat::clearForeground()

C++ documentation:

Clears the brush used to paint the document's foreground. The default brush will be used.

See also foreground(), setForeground(), and clearBackground().

fn clear_property(&mut self, property_id: c_int)

C++ method: void QTextListFormat::clearProperty(int propertyId)

C++ documentation:

Clears the value of the property given by propertyId

See also Property.

fn color_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatColorPropertyArgs<'largs>

C++ method: QTextListFormat::colorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn color_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::color::Color>
2) fn color_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::color::Color
C++ method: QColor QTextListFormat::colorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Color type, an invalid color is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn double_property(&self, property_id: c_int) -> c_double

C++ method: double QTextListFormat::doubleProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property isn't of QVariant::Double or QMetaType::Float type, 0 is returned instead.

See also setProperty(), boolProperty(), intProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn foreground<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatForegroundArgs<'largs>

C++ method: QTextListFormat::foreground

This is an overloaded function. Available variants:

Rust arguments:
1) fn foreground(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::brush::Brush>
2) fn foreground(&self, cpp_utils::AsStruct) -> ::brush::Brush
C++ method: QBrush QTextListFormat::foreground() const

C++ documentation:

Returns the brush used to render foreground details, such as text, frame outlines, and table borders.

See also setForeground(), clearForeground(), and background().

fn has_property(&self, property_id: c_int) -> bool

C++ method: bool QTextListFormat::hasProperty(int propertyId) const

C++ documentation:

Returns true if the text format has a property with the given propertyId; otherwise returns false.

See also properties() and Property.

fn indent(&self) -> c_int

C++ method: int QTextListFormat::indent() const

C++ documentation:

Returns the list format's indentation. The indentation is multiplied by the QTextDocument::indentWidth property to get the effective indent in pixels.

See also setIndent().

fn int_property(&self, property_id: c_int) -> c_int

C++ method: int QTextListFormat::intProperty(int propertyId) const

C++ documentation:

Returns the value of the property specified by propertyId. If the property is not of QTextFormat::Integer type, 0 is returned instead.

See also setProperty(), boolProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn is_block_format(&self) -> bool

C++ method: bool QTextListFormat::isBlockFormat() const

C++ documentation:

Returns true if this text format is a BlockFormat; otherwise returns false.

fn is_char_format(&self) -> bool

C++ method: bool QTextListFormat::isCharFormat() const

C++ documentation:

Returns true if this text format is a CharFormat; otherwise returns false.

fn is_empty(&self) -> bool

C++ method: bool QTextListFormat::isEmpty() const

C++ documentation:

Returns true if the format does not store any properties; false otherwise.

This function was introduced in Qt 5.3.

See also propertyCount() and properties().

fn is_frame_format(&self) -> bool

C++ method: bool QTextListFormat::isFrameFormat() const

C++ documentation:

Returns true if this text format is a FrameFormat; otherwise returns false.

fn is_image_format(&self) -> bool

C++ method: bool QTextListFormat::isImageFormat() const

C++ documentation:

Returns true if this text format is an image format; otherwise returns false.

fn is_list_format(&self) -> bool

C++ method: bool QTextListFormat::isListFormat() const

C++ documentation:

Returns true if this text format is a ListFormat; otherwise returns false.

fn is_table_cell_format(&self) -> bool

C++ method: bool QTextListFormat::isTableCellFormat() const

C++ documentation:

Returns true if this text format is a TableCellFormat; otherwise returns false.

This function was introduced in Qt 4.4.

fn is_table_format(&self) -> bool

C++ method: bool QTextListFormat::isTableFormat() const

C++ documentation:

Returns true if this text format is a TableFormat; otherwise returns false.

fn is_valid(&self) -> bool

C++ method: bool QTextListFormat::isValid() const

C++ documentation:

Returns true if this list format is valid; otherwise returns false.

fn layout_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextListFormat::layoutDirection() const

C++ documentation:

Returns the document's layout direction.

See also setLayoutDirection().

fn length_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatLengthPropertyArgs<'largs>

C++ method: QTextListFormat::lengthProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_length::TextLength>
2) fn length_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_length::TextLength
C++ method: QTextLength QTextListFormat::lengthProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthVectorProperty(), and Property.

fn length_vector_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatLengthVectorPropertyArgs<'largs>

C++ method: QTextListFormat::lengthVectorProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorTextLengthRef>
2) fn length_vector_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorTextLengthRef
C++ method: QVector<QTextLength> QTextListFormat::lengthVectorProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId. If the property isn't of QTextFormat::LengthVector type, an empty length vector is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), colorProperty(), lengthProperty(), and Property.

fn merge(&mut self, other: &TextFormat)

C++ method: void QTextListFormat::merge(const QTextFormat& other)

C++ documentation:

Merges the other format with this format; where there are conflicts the other format takes precedence.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextListFormatNewArgs

C++ method: QTextListFormat::QTextListFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_list_format::TextListFormat
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_list_format::TextListFormat>
C++ method: [constructor] void QTextListFormat::QTextListFormat()

fn number_prefix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatNumberPrefixArgs<'largs>

C++ method: QTextListFormat::numberPrefix

This is an overloaded function. Available variants:

Rust arguments:
1) fn number_prefix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn number_prefix(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextListFormat::numberPrefix() const

C++ documentation:

Returns the list format's number prefix.

This function was introduced in Qt 4.8.

See also setNumberPrefix().

fn number_suffix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatNumberSuffixArgs<'largs>

C++ method: QTextListFormat::numberSuffix

This is an overloaded function. Available variants:

Rust arguments:
1) fn number_suffix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn number_suffix(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTextListFormat::numberSuffix() const

C++ documentation:

Returns the list format's number suffix.

This function was introduced in Qt 4.8.

See also setNumberSuffix().

fn object_index(&self) -> c_int

C++ method: int QTextListFormat::objectIndex() const

C++ documentation:

Returns the index of the format object, or -1 if the format object is invalid.

See also setObjectIndex().

fn object_type(&self) -> c_int

C++ method: int QTextListFormat::objectType() const

C++ documentation:

Returns the text format's object type.

See also ObjectTypes and setObjectType().

fn op_eq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextListFormat::operator==(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is the same as the other text format.

fn op_neq(&self, rhs: &TextFormat) -> bool

C++ method: bool QTextListFormat::operator!=(const QTextFormat& rhs) const

C++ documentation:

Returns true if this text format is different from the other text format.

fn pen_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatPenPropertyArgs<'largs>

C++ method: QTextListFormat::penProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn pen_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::pen::Pen>
2) fn pen_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::pen::Pen
C++ method: QPen QTextListFormat::penProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::Pen type, Qt::NoPen is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), stringProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn properties<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatPropertiesArgs<'largs>

C++ method: QTextListFormat::properties

This is an overloaded function. Available variants:

Rust arguments:
1) fn properties(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::map::MapCIntVariantRef>
2) fn properties(&self, cpp_utils::AsStruct) -> qt_core::map::MapCIntVariantRef
C++ method: QMap<int, QVariant> QTextListFormat::properties() const

C++ documentation:

Returns a map with all properties of this text format.

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatPropertyArgs<'largs>

C++ method: QTextListFormat::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::variant::Variant
C++ method: QVariant QTextListFormat::property(int propertyId) const

C++ documentation:

Returns the property specified by the given propertyId.

See also setProperty() and Property.

fn property_count(&self) -> c_int

C++ method: int QTextListFormat::propertyCount() const

C++ documentation:

Returns the number of properties stored in the format.

This function was introduced in Qt 4.3.

fn set_background(&mut self, brush: &Brush)

C++ method: void QTextListFormat::setBackground(const QBrush& brush)

C++ documentation:

Sets the brush use to paint the document's background to the brush specified.

See also background(), clearBackground(), and setForeground().

fn set_foreground(&mut self, brush: &Brush)

C++ method: void QTextListFormat::setForeground(const QBrush& brush)

C++ documentation:

Sets the foreground brush to the specified brush. The foreground brush is mostly used to render text.

See also foreground(), clearForeground(), and setBackground().

fn set_indent(&mut self, indent: c_int)

C++ method: void QTextListFormat::setIndent(int indent)

C++ documentation:

Sets the list format's indentation. The indentation is multiplied by the QTextDocument::indentWidth property to get the effective indent in pixels.

See also indent().

fn set_layout_direction(&mut self, direction: LayoutDirection)

C++ method: void QTextListFormat::setLayoutDirection(Qt::LayoutDirection direction)

C++ documentation:

Sets the document's layout direction to the specified direction.

See also layoutDirection().

fn set_number_prefix(&mut self, number_prefix: &String)

C++ method: void QTextListFormat::setNumberPrefix(const QString& numberPrefix)

C++ documentation:

Sets the list format's number prefix to the string specified by numberPrefix. This can be used with all sorted list types. It does not have any effect on unsorted list types.

The default prefix is an empty string.

This function was introduced in Qt 4.8.

See also numberPrefix().

fn set_number_suffix(&mut self, number_suffix: &String)

C++ method: void QTextListFormat::setNumberSuffix(const QString& numberSuffix)

C++ documentation:

Sets the list format's number suffix to the string specified by numberSuffix. This can be used with all sorted list types. It does not have any effect on unsorted list types.

The default suffix is ".".

This function was introduced in Qt 4.8.

See also numberSuffix().

fn set_object_index(&mut self, object: c_int)

C++ method: void QTextListFormat::setObjectIndex(int object)

C++ documentation:

Sets the format object's object index.

See also objectIndex().

fn set_object_type(&mut self, type_: c_int)

C++ method: void QTextListFormat::setObjectType(int type)

C++ documentation:

Sets the text format's object type to type.

See also ObjectTypes and objectType().

fn set_property<'largs, Args>(&'largs mut self, args: Args) where Args: TextListFormatSetPropertyArgs<'largs>

C++ method: QTextListFormat::setProperty

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_property(&mut self, (libc::c_int, &qt_core::variant::Variant)) -> ()
C++ method: void QTextListFormat::setProperty(int propertyId, const QVariant& value)

C++ documentation:

Sets the property specified by the propertyId to the given value.

See also property() and Property.

Variant 2

Rust arguments: fn set_property(&mut self, (libc::c_int, &::vector::VectorTextLengthRef)) -> ()
C++ method: void QTextListFormat::setProperty(int propertyId, const QVector<QTextLength>& lengths)

C++ documentation:

Sets the value of the property given by propertyId to value.

See also lengthVectorProperty() and Property.

fn set_style(&mut self, style: Style)

C++ method: void QTextListFormat::setStyle(QTextListFormat::Style style)

C++ documentation:

Sets the list format's style.

See also style() and Style.

fn string_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatStringPropertyArgs<'largs>

C++ method: QTextListFormat::stringProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn string_property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::string::String>
2) fn string_property(&self, (libc::c_int, cpp_utils::AsStruct)) -> qt_core::string::String
C++ method: QString QTextListFormat::stringProperty(int propertyId) const

C++ documentation:

Returns the value of the property given by propertyId; if the property isn't of QVariant::String type, an empty string is returned instead.

See also setProperty(), boolProperty(), intProperty(), doubleProperty(), colorProperty(), lengthProperty(), lengthVectorProperty(), and Property.

fn style(&self) -> Style

C++ method: QTextListFormat::Style QTextListFormat::style() const

C++ documentation:

Returns the list format's style.

See also setStyle() and Style.

fn swap(&mut self, other: &mut TextFormat)

C++ method: void QTextListFormat::swap(QTextFormat& other)

C++ documentation:

Swaps this text format with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn to_block_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToBlockFormatArgs<'largs>

C++ method: QTextListFormat::toBlockFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_block_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_block_format::TextBlockFormat>
2) fn to_block_format(&self, cpp_utils::AsStruct) -> ::text_block_format::TextBlockFormat
C++ method: QTextBlockFormat QTextListFormat::toBlockFormat() const

C++ documentation:

Returns this format as a block format.

fn to_char_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToCharFormatArgs<'largs>

C++ method: QTextListFormat::toCharFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_char_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn to_char_format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextListFormat::toCharFormat() const

C++ documentation:

Returns this format as a character format.

fn to_frame_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToFrameFormatArgs<'largs>

C++ method: QTextListFormat::toFrameFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_frame_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame_format::TextFrameFormat>
2) fn to_frame_format(&self, cpp_utils::AsStruct) -> ::text_frame_format::TextFrameFormat
C++ method: QTextFrameFormat QTextListFormat::toFrameFormat() const

C++ documentation:

Returns this format as a frame format.

fn to_image_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToImageFormatArgs<'largs>

C++ method: QTextListFormat::toImageFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_image_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_image_format::TextImageFormat>
2) fn to_image_format(&self, cpp_utils::AsStruct) -> ::text_image_format::TextImageFormat
C++ method: QTextImageFormat QTextListFormat::toImageFormat() const

C++ documentation:

Returns this format as an image format.

fn to_list_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToListFormatArgs<'largs>

C++ method: QTextListFormat::toListFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_list_format::TextListFormat>
2) fn to_list_format(&self, cpp_utils::AsStruct) -> ::text_list_format::TextListFormat
C++ method: QTextListFormat QTextListFormat::toListFormat() const

C++ documentation:

Returns this format as a list format.

fn to_table_cell_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToTableCellFormatArgs<'largs>

C++ method: QTextListFormat::toTableCellFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_cell_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_cell_format::TextTableCellFormat>
2) fn to_table_cell_format(&self, cpp_utils::AsStruct) -> ::text_table_cell_format::TextTableCellFormat
C++ method: QTextTableCellFormat QTextListFormat::toTableCellFormat() const

C++ documentation:

Returns this format as a table cell format.

This function was introduced in Qt 4.4.

fn to_table_format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextListFormatToTableFormatArgs<'largs>

C++ method: QTextListFormat::toTableFormat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_table_format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_format::TextTableFormat>
2) fn to_table_format(&self, cpp_utils::AsStruct) -> ::text_table_format::TextTableFormat
C++ method: QTextTableFormat QTextListFormat::toTableFormat() const

C++ documentation:

Returns this format as a table format.

fn type_(&self) -> c_int

C++ method: int QTextListFormat::type() const

C++ documentation:

Returns the type of this format.

See also FormatType.