Module qt_gui::text_object::overloading

Traits

TextObjectConnectArgs
This trait represents a set of arguments accepted by TextObject::connect method.

TextObjectConnectStaticArgs
This trait represents a set of arguments accepted by TextObject::connect_static method.

TextObjectDestroyedArgs
This trait represents a set of arguments accepted by TextObject::destroyed method.

TextObjectDisconnectArgs
This trait represents a set of arguments accepted by TextObject::disconnect method.

TextObjectDisconnectStaticArgs
This trait represents a set of arguments accepted by TextObject::disconnect_static method.

TextObjectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TextObject::dynamic_property_names method.

TextObjectFormatArgs
This trait represents a set of arguments accepted by TextObject::format method.

TextObjectObjectNameArgs
This trait represents a set of arguments accepted by TextObject::object_name method.

TextObjectPropertyArgs
This trait represents a set of arguments accepted by TextObject::property method.

TextObjectStartTimerArgs
This trait represents a set of arguments accepted by TextObject::start_timer method.

TextObjectTrArgs
This trait represents a set of arguments accepted by TextObject::tr method.

TextObjectTrUtf8Args
This trait represents a set of arguments accepted by TextObject::tr_utf8 method.