Struct qt_gui::text_object_interface::TextObjectInterface

pub struct TextObjectInterface {
    // some fields omitted
}

C++ type: QTextObjectInterface

C++ documentation:

The QTextObjectInterface class allows drawing of custom text objects in QTextDocuments.

A text object describes the structure of one or more elements in a text document; for instance, images imported from HTML are implemented using text objects. A text object knows how to lay out and draw its elements when a document is being rendered.

Qt allows custom text objects to be inserted into a document by registering a custom object type with QTextCharFormat. A QTextObjectInterface must also be implemented for this type and be registered with the QAbstractTextDocumentLayout of the document. When the object type is encountered while rendering a QTextDocument, the intrinsicSize() and drawObject() functions of the interface are called.

The following list explains the required steps of inserting a custom text object into a document:

Choose an objectType. The objectType is an integer with a value greater or equal to QTextFormat::UserObject.

Create a QTextCharFormat object and set the object type to the chosen type using the setObjectType() function.

Implement the QTextObjectInterface class.

Call QAbstractTextDocumentLayout::registerHandler() with an instance of your QTextObjectInterface subclass to register your object type.

Insert QChar::ObjectReplacementCharacter with the aforementioned QTextCharFormat of the chosen object type into the document. As mentioned, the functions of QTextObjectInterface intrinsicSize() and drawObject() will then be called with the QTextFormat as parameter whenever the replacement character is encountered.

A class implementing a text object needs to inherit both QObject and QTextObjectInterface. QObject must be the first class inherited. For instance:

class SvgTextObject : public QObject, public QTextObjectInterface
{
    Q_OBJECT
    Q_INTERFACES(QTextObjectInterface)

The data of a text object is usually stored in the QTextCharFormat using QTextCharFormat::setProperty(), and then retrieved with QTextCharFormat::property().

Warning: Copy and Paste operations ignore custom text objects.

Methods

impl TextObjectInterface

fn draw_object(&mut self, painter: *mut Painter, rect: &RectF, doc: *mut TextDocument, pos_in_document: c_int, format: &TextFormat)

C++ method: pure virtual void QTextObjectInterface::drawObject(QPainter* painter, const QRectF& rect, QTextDocument* doc, int posInDocument, const QTextFormat& format)

C++ documentation:

Draws this text object using the specified painter.

The size of the rectangle, rect, to draw in is the size previously calculated by intrinsicSize(). The rectangles position is relative to the painter.

You also get the document (doc) and the position (posInDocument) of the format in that document.

See also intrinsicSize().

fn intrinsic_size<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextObjectInterfaceIntrinsicSizeArgs<'largs>

C++ method: QTextObjectInterface::intrinsicSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn intrinsic_size(&mut self, (*mut ::text_document::TextDocument, libc::c_int, &::text_format::TextFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::size_f::SizeF>
2) fn intrinsic_size(&mut self, (*mut ::text_document::TextDocument, libc::c_int, &::text_format::TextFormat, cpp_utils::AsStruct)) -> qt_core::size_f::SizeF
C++ method: pure virtual QSizeF QTextObjectInterface::intrinsicSize(QTextDocument* doc, int posInDocument, const QTextFormat& format)

C++ documentation:

The intrinsicSize() function returns the size of the text object represented by format in the given document (doc) at the given position (posInDocument).

The size calculated will be used for subsequent calls to drawObject() for this format.

See also drawObject().