Struct qt_gui::text_option::Tab

pub struct Tab {
    // some fields omitted
}

C++ type: QTextOption::Tab

C++ documentation:

Each tab definition is represented by this struct.

Methods

impl Tab

fn delimiter<'l0>(&'l0 self) -> &'l0 Char

C++ method: const QChar& QTextOption::Tab::delimiter() const

fn delimiter_mut<'l0>(&'l0 mut self) -> &'l0 mut Char

C++ method: QChar& QTextOption::Tab::delimiter_mut()

fn new<Args>(args: Args) -> Args::ReturnType where Args: TabNewArgs

C++ method: QTextOption::Tab::Tab

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_option::Tab
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_option::Tab>
C++ method: [constructor] void QTextOption::Tab::Tab()

Variant 2

Rust arguments:
1) fn new((libc::c_double, ::text_option::TabType, cpp_utils::AsStruct)) -> ::text_option::Tab
2) fn new((libc::c_double, ::text_option::TabType, &qt_core::char::Char, cpp_utils::AsStruct)) -> ::text_option::Tab
3) fn new((libc::c_double, ::text_option::TabType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_option::Tab>
4) fn new((libc::c_double, ::text_option::TabType, &qt_core::char::Char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_option::Tab>
C++ method: [constructor] void QTextOption::Tab::Tab(double pos, QTextOption::TabType tabType, QChar delim = ?)

fn op_eq(&self, other: &Tab) -> bool

C++ method: bool QTextOption::Tab::operator==(const QTextOption::Tab& other) const

C++ documentation:

Returns true if tab other is equal to this tab; otherwise returns false.

fn op_neq(&self, other: &Tab) -> bool

C++ method: bool QTextOption::Tab::operator!=(const QTextOption::Tab& other) const

C++ documentation:

Returns true if tab other is not equal to this tab; otherwise returns false.

fn position(&self) -> c_double

C++ method: double QTextOption::Tab::position() const

fn set_delimiter(&mut self, value: &Char)

C++ method: void QTextOption::Tab::set_delimiter(QChar value)

fn set_position(&mut self, value: c_double)

C++ method: void QTextOption::Tab::set_position(double value)

fn set_type(&mut self, value: TabType)

C++ method: void QTextOption::Tab::set_type(QTextOption::TabType value)

fn type_(&self) -> TabType

C++ method: QTextOption::TabType QTextOption::Tab::type() const