Struct qt_gui::text_option::TextOption

pub struct TextOption {
    // some fields omitted
}

C++ type: QTextOption

C++ documentation:

The QTextOption class provides a description of general rich text properties.

QTextOption is used to encapsulate common rich text properties in a single object. It contains information about text alignment, layout direction, word wrapping, and other standard properties associated with text rendering and layout.

Methods

impl TextOption

fn alignment(&self) -> Flags<AlignmentFlag>

C++ method: QFlags<Qt::AlignmentFlag> QTextOption::alignment() const

C++ documentation:

Returns the text alignment defined by the option.

See also setAlignment().

fn flags(&self) -> Flags<Flag>

C++ method: QFlags<QTextOption::Flag> QTextOption::flags() const

C++ documentation:

Returns the flags associated with the option.

See also setFlags().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextOptionNewArgs

C++ method: QTextOption::QTextOption

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_option::TextOption
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_option::TextOption>
C++ method: [constructor] void QTextOption::QTextOption()

Variant 2

Rust arguments:
1) fn new((qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, cpp_utils::AsStruct)) -> ::text_option::TextOption
2) fn new((qt_core::flags::Flags<qt_core::qt::AlignmentFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_option::TextOption>
C++ method: [constructor] void QTextOption::QTextOption(QFlags<Qt::AlignmentFlag> alignment)

Variant 3

Rust arguments:
1) fn new((&::text_option::TextOption, cpp_utils::AsStruct)) -> ::text_option::TextOption
2) fn new((&::text_option::TextOption, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_option::TextOption>
C++ method: [constructor] void QTextOption::QTextOption(const QTextOption& o)

fn op_assign<'l0, 'l1>(&'l0 mut self, o: &'l1 TextOption) -> &'l0 mut TextOption

C++ method: QTextOption& QTextOption::operator=(const QTextOption& o)

C++ documentation:

Returns true if the text option is the same as the other text option; otherwise returns false.

fn set_alignment(&mut self, alignment: Flags<AlignmentFlag>)

C++ method: void QTextOption::setAlignment(QFlags<Qt::AlignmentFlag> alignment)

C++ documentation:

Sets the option's text alignment to the specified alignment.

See also alignment().

fn set_flags(&mut self, flags: Flags<Flag>)

C++ method: void QTextOption::setFlags(QFlags<QTextOption::Flag> flags)

C++ documentation:

Sets the flags associated with the option to the given flags.

See also flags().

fn set_tab_array(&mut self, tab_stops: &ListCDouble)

C++ method: void QTextOption::setTabArray(const QList<double>& tabStops)

C++ documentation:

Sets the tab positions for the text layout to those specified by tabStops.

See also tabArray(), setTabStop(), and setTabs().

fn set_tab_stop(&mut self, tab_stop: c_double)

C++ method: void QTextOption::setTabStop(double tabStop)

C++ documentation:

Sets the default distance in device units between tab stops to the value specified by tabStop.

See also tabStop(), setTabArray(), setTabs(), and tabs().

fn set_tabs(&mut self, tab_stops: &ListTabRef)

C++ method: void QTextOption::setTabs(const QList<QTextOption::Tab>& tabStops)

C++ documentation:

Set the Tab properties to tabStops.

See also tabStop() and tabs().

fn set_text_direction(&mut self, a_direction: LayoutDirection)

C++ method: void QTextOption::setTextDirection(Qt::LayoutDirection aDirection)

C++ documentation:

Sets the direction of the text layout defined by the option to the given direction.

See also textDirection().

fn set_use_design_metrics(&mut self, b: bool)

C++ method: void QTextOption::setUseDesignMetrics(bool b)

C++ documentation:

If enable is true then the layout will use design metrics; otherwise it will use the metrics of the paint device (which is the default behavior).

See also useDesignMetrics().

fn set_wrap_mode(&mut self, wrap: WrapMode)

C++ method: void QTextOption::setWrapMode(QTextOption::WrapMode wrap)

C++ documentation:

Sets the option's text wrap mode to the given mode.

See also wrapMode().

fn tab_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextOptionTabArrayArgs<'largs>

C++ method: QTextOption::tabArray

This is an overloaded function. Available variants:

Rust arguments:
1) fn tab_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListCDouble>
2) fn tab_array(&self, cpp_utils::AsStruct) -> ::list::ListCDouble
C++ method: QList<double> QTextOption::tabArray() const

C++ documentation:

Returns a list of tab positions defined for the text layout.

See also setTabArray() and tabStop().

fn tab_stop(&self) -> c_double

C++ method: double QTextOption::tabStop() const

C++ documentation:

Returns the distance in device units between tab stops. Convenient function for the above method

See also setTabStop(), tabArray(), setTabs(), and tabs().

fn tabs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextOptionTabsArgs<'largs>

C++ method: QTextOption::tabs

This is an overloaded function. Available variants:

Rust arguments:
1) fn tabs(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListTabRef>
2) fn tabs(&self, cpp_utils::AsStruct) -> ::list::ListTabRef
C++ method: QList<QTextOption::Tab> QTextOption::tabs() const

C++ documentation:

Returns a list of tab positions defined for the text layout.

This function was introduced in Qt 4.4.

See also tabStop(), setTabs(), and setTabStop().

fn text_direction(&self) -> LayoutDirection

C++ method: Qt::LayoutDirection QTextOption::textDirection() const

C++ documentation:

Returns the direction of the text layout defined by the option.

See also setTextDirection().

fn use_design_metrics(&self) -> bool

C++ method: bool QTextOption::useDesignMetrics() const

C++ documentation:

Returns true if the layout uses design rather than device metrics; otherwise returns false.

See also setUseDesignMetrics().

fn wrap_mode(&self) -> WrapMode

C++ method: QTextOption::WrapMode QTextOption::wrapMode() const

C++ documentation:

Returns the text wrap mode defined by the option.

See also setWrapMode().