Module qt_gui::text_table::overloading

Traits

TextTableBeginArgs
This trait represents a set of arguments accepted by TextTable::begin method.

TextTableCellAtArgs
This trait represents a set of arguments accepted by TextTable::cell_at method.

TextTableChildFramesArgs
This trait represents a set of arguments accepted by TextTable::child_frames method.

TextTableConnectArgs
This trait represents a set of arguments accepted by TextTable::connect method.

TextTableConnectStaticArgs
This trait represents a set of arguments accepted by TextTable::connect_static method.

TextTableDestroyedArgs
This trait represents a set of arguments accepted by TextTable::destroyed method.

TextTableDisconnectArgs
This trait represents a set of arguments accepted by TextTable::disconnect method.

TextTableDisconnectStaticArgs
This trait represents a set of arguments accepted by TextTable::disconnect_static method.

TextTableDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by TextTable::dynamic_property_names method.

TextTableEndArgs
This trait represents a set of arguments accepted by TextTable::end method.

TextTableFirstCursorPositionArgs
This trait represents a set of arguments accepted by TextTable::first_cursor_position method.

TextTableFormatArgs
This trait represents a set of arguments accepted by TextTable::format method.

TextTableFrameFormatArgs
This trait represents a set of arguments accepted by TextTable::frame_format method.

TextTableLastCursorPositionArgs
This trait represents a set of arguments accepted by TextTable::last_cursor_position method.

TextTableMergeCellsArgs
This trait represents a set of arguments accepted by TextTable::merge_cells method.

TextTableNewArgs
This trait represents a set of arguments accepted by TextTable::new method.

TextTableObjectNameArgs
This trait represents a set of arguments accepted by TextTable::object_name method.

TextTablePropertyArgs
This trait represents a set of arguments accepted by TextTable::property method.

TextTableRowEndArgs
This trait represents a set of arguments accepted by TextTable::row_end method.

TextTableRowStartArgs
This trait represents a set of arguments accepted by TextTable::row_start method.

TextTableStartTimerArgs
This trait represents a set of arguments accepted by TextTable::start_timer method.

TextTableTrArgs
This trait represents a set of arguments accepted by TextTable::tr method.

TextTableTrUtf8Args
This trait represents a set of arguments accepted by TextTable::tr_utf8 method.