Struct qt_gui::text_table_cell::TextTableCell

pub struct TextTableCell {
    // some fields omitted
}

C++ type: QTextTableCell

C++ documentation:

The QTextTableCell class represents the properties of a cell in a QTextTable.

Table cells are pieces of document structure that belong to a table. The table orders cells into particular rows and columns; cells can also span multiple columns and rows.

Cells are usually created when a table is inserted into a document with QTextCursor::insertTable(), but they are also created and destroyed when a table is resized.

Cells contain information about their location in a table; you can obtain the row() and column() numbers of a cell, and its rowSpan() and columnSpan().

The format() of a cell describes the default character format of its contents. The firstCursorPosition() and lastCursorPosition() functions are used to obtain the extent of the cell in the document.

Methods

impl TextTableCell

fn begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextTableCellBeginArgs<'largs>

C++ method: QTextTableCell::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame::Iterator>
2) fn begin(&self, cpp_utils::AsStruct) -> ::text_frame::Iterator
C++ method: QTextFrame::iterator QTextTableCell::begin() const

C++ documentation:

Returns a frame iterator pointing to the beginning of the table's cell.

See also end().

fn column(&self) -> c_int

C++ method: int QTextTableCell::column() const

C++ documentation:

Returns the number of the column in the table that contains this cell.

See also row().

fn column_span(&self) -> c_int

C++ method: int QTextTableCell::columnSpan() const

C++ documentation:

Returns the number of columns this cell spans. The default is 1.

See also rowSpan().

fn end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextTableCellEndArgs<'largs>

C++ method: QTextTableCell::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_frame::Iterator>
2) fn end(&self, cpp_utils::AsStruct) -> ::text_frame::Iterator
C++ method: QTextFrame::iterator QTextTableCell::end() const

C++ documentation:

Returns a frame iterator pointing to the end of the table's cell.

See also begin().

fn first_cursor_position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextTableCellFirstCursorPositionArgs<'largs>

C++ method: QTextTableCell::firstCursorPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn first_cursor_position(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_cursor::TextCursor>
2) fn first_cursor_position(&self, cpp_utils::AsStruct) -> ::text_cursor::TextCursor
C++ method: QTextCursor QTextTableCell::firstCursorPosition() const

C++ documentation:

Returns the first valid cursor position in this cell.

See also lastCursorPosition().

fn first_position(&self) -> c_int

C++ method: int QTextTableCell::firstPosition() const

fn format<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextTableCellFormatArgs<'largs>

C++ method: QTextTableCell::format

This is an overloaded function. Available variants:

Rust arguments:
1) fn format(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_char_format::TextCharFormat>
2) fn format(&self, cpp_utils::AsStruct) -> ::text_char_format::TextCharFormat
C++ method: QTextCharFormat QTextTableCell::format() const

C++ documentation:

Returns the cell's character format.

See also setFormat().

fn is_valid(&self) -> bool

C++ method: bool QTextTableCell::isValid() const

C++ documentation:

Returns true if this is a valid table cell; otherwise returns false.

fn last_cursor_position<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextTableCellLastCursorPositionArgs<'largs>

C++ method: QTextTableCell::lastCursorPosition

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_cursor_position(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_cursor::TextCursor>
2) fn last_cursor_position(&self, cpp_utils::AsStruct) -> ::text_cursor::TextCursor
C++ method: QTextCursor QTextTableCell::lastCursorPosition() const

C++ documentation:

Returns the last valid cursor position in this cell.

See also firstCursorPosition().

fn last_position(&self) -> c_int

C++ method: int QTextTableCell::lastPosition() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextTableCellNewArgs

C++ method: QTextTableCell::QTextTableCell

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_table_cell::TextTableCell
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_table_cell::TextTableCell>
C++ method: [constructor] void QTextTableCell::QTextTableCell()

Variant 2

Rust arguments:
1) fn new((&::text_table_cell::TextTableCell, cpp_utils::AsStruct)) -> ::text_table_cell::TextTableCell
2) fn new((&::text_table_cell::TextTableCell, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_table_cell::TextTableCell>
C++ method: [constructor] void QTextTableCell::QTextTableCell(const QTextTableCell& o)

fn op_assign<'l0, 'l1>(&'l0 mut self, o: &'l1 TextTableCell) -> &'l0 mut TextTableCell

C++ method: QTextTableCell& QTextTableCell::operator=(const QTextTableCell& o)

C++ documentation:

Assigns the other table cell to this table cell.

fn op_eq(&self, other: &TextTableCell) -> bool

C++ method: bool QTextTableCell::operator==(const QTextTableCell& other) const

C++ documentation:

Returns true if this cell object and the other cell object describe the same cell; otherwise returns false.

fn op_neq(&self, other: &TextTableCell) -> bool

C++ method: bool QTextTableCell::operator!=(const QTextTableCell& other) const

C++ documentation:

Returns true if this cell object and the other cell object describe different cells; otherwise returns false.

fn row(&self) -> c_int

C++ method: int QTextTableCell::row() const

C++ documentation:

Returns the number of the row in the table that contains this cell.

See also column().

fn row_span(&self) -> c_int

C++ method: int QTextTableCell::rowSpan() const

C++ documentation:

Returns the number of rows this cell spans. The default is 1.

See also columnSpan().

fn set_format(&mut self, format: &TextCharFormat)

C++ method: void QTextTableCell::setFormat(const QTextCharFormat& format)

C++ documentation:

Sets the cell's character format to format. This can for example be used to change the background color of the entire cell:

QTextTableCell cell = table->cellAt(2, 3); QTextCharFormat format = cell.format(); format.setBackground(Qt::blue); cell.setFormat(format);

Note that the cell's row or column span cannot be changed through this function. You have to use QTextTable::mergeCells and QTextTable::splitCell instead.

This function was introduced in Qt 4.2.

See also format().

fn table_cell_format_index(&self) -> c_int

C++ method: int QTextTableCell::tableCellFormatIndex() const

C++ documentation:

Returns the index of the tableCell's format in the document's internal list of formats.

This function was introduced in Qt 4.5.

See also QTextDocument::allFormats().