Module qt_gui::text_table_cell_format::overloading

Traits

TextTableCellFormatAnchorHrefArgs
This trait represents a set of arguments accepted by TextTableCellFormat::anchor_href method.

TextTableCellFormatAnchorNameArgs
This trait represents a set of arguments accepted by TextTableCellFormat::anchor_name method.

TextTableCellFormatAnchorNamesArgs
This trait represents a set of arguments accepted by TextTableCellFormat::anchor_names method.

TextTableCellFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextTableCellFormat::as_q_variant method.

TextTableCellFormatBackgroundArgs
This trait represents a set of arguments accepted by TextTableCellFormat::background method.

TextTableCellFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::brush_property method.

TextTableCellFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::color_property method.

TextTableCellFormatFontArgs
This trait represents a set of arguments accepted by TextTableCellFormat::font method.

TextTableCellFormatFontFamilyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::font_family method.

TextTableCellFormatForegroundArgs
This trait represents a set of arguments accepted by TextTableCellFormat::foreground method.

TextTableCellFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::length_property method.

TextTableCellFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::length_vector_property method.

TextTableCellFormatNewArgs
This trait represents a set of arguments accepted by TextTableCellFormat::new method.

TextTableCellFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::pen_property method.

TextTableCellFormatPropertiesArgs
This trait represents a set of arguments accepted by TextTableCellFormat::properties method.

TextTableCellFormatPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::property method.

TextTableCellFormatSetFontArgs
This trait represents a set of arguments accepted by TextTableCellFormat::set_font method.

TextTableCellFormatSetFontStyleHintArgs
This trait represents a set of arguments accepted by TextTableCellFormat::set_font_style_hint method.

TextTableCellFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::set_property method.

TextTableCellFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextTableCellFormat::string_property method.

TextTableCellFormatTextOutlineArgs
This trait represents a set of arguments accepted by TextTableCellFormat::text_outline method.

TextTableCellFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_block_format method.

TextTableCellFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_char_format method.

TextTableCellFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_frame_format method.

TextTableCellFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_image_format method.

TextTableCellFormatToListFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_list_format method.

TextTableCellFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_table_cell_format method.

TextTableCellFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextTableCellFormat::to_table_format method.

TextTableCellFormatToolTipArgs
This trait represents a set of arguments accepted by TextTableCellFormat::tool_tip method.

TextTableCellFormatUnderlineColorArgs
This trait represents a set of arguments accepted by TextTableCellFormat::underline_color method.