Module qt_gui::text_table_format::overloading

Traits

TextTableFormatAsQVariantArgs
This trait represents a set of arguments accepted by TextTableFormat::as_q_variant method.

TextTableFormatBackgroundArgs
This trait represents a set of arguments accepted by TextTableFormat::background method.

TextTableFormatBorderBrushArgs
This trait represents a set of arguments accepted by TextTableFormat::border_brush method.

TextTableFormatBrushPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::brush_property method.

TextTableFormatColorPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::color_property method.

TextTableFormatColumnWidthConstraintsArgs
This trait represents a set of arguments accepted by TextTableFormat::column_width_constraints method.

TextTableFormatForegroundArgs
This trait represents a set of arguments accepted by TextTableFormat::foreground method.

TextTableFormatHeightArgs
This trait represents a set of arguments accepted by TextTableFormat::height method.

TextTableFormatLengthPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::length_property method.

TextTableFormatLengthVectorPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::length_vector_property method.

TextTableFormatNewArgs
This trait represents a set of arguments accepted by TextTableFormat::new method.

TextTableFormatPenPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::pen_property method.

TextTableFormatPropertiesArgs
This trait represents a set of arguments accepted by TextTableFormat::properties method.

TextTableFormatPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::property method.

TextTableFormatSetHeightArgs
This trait represents a set of arguments accepted by TextTableFormat::set_height method.

TextTableFormatSetPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::set_property method.

TextTableFormatSetWidthArgs
This trait represents a set of arguments accepted by TextTableFormat::set_width method.

TextTableFormatStringPropertyArgs
This trait represents a set of arguments accepted by TextTableFormat::string_property method.

TextTableFormatToBlockFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_block_format method.

TextTableFormatToCharFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_char_format method.

TextTableFormatToFrameFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_frame_format method.

TextTableFormatToImageFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_image_format method.

TextTableFormatToListFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_list_format method.

TextTableFormatToTableCellFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_table_cell_format method.

TextTableFormatToTableFormatArgs
This trait represents a set of arguments accepted by TextTableFormat::to_table_format method.

TextTableFormatWidthArgs
This trait represents a set of arguments accepted by TextTableFormat::width method.