Struct qt_gui::touch_device::TouchDevice

pub struct TouchDevice {
    // some fields omitted
}

C++ type: QTouchDevice

C++ documentation:

The QTouchDevice class describes the device from which touch events originate.

Each QTouchEvent contains a QTouchDevice pointer to allow accessing device-specific properties like type and capabilities. It is the responsibility of the platform or generic plug-ins to register the available touch devices via QWindowSystemInterface before generating any touch events. Applications do not need to instantiate this class, they should just access the global instances pointed to by QTouchEvent::device().

Methods

impl TouchDevice

fn capabilities(&self) -> Flags<CapabilityFlag>

C++ method: QFlags<QTouchDevice::CapabilityFlag> QTouchDevice::capabilities() const

C++ documentation:

Returns the touch device capabilities.

See also setCapabilities().

fn devices<Args>(args: Args) -> Args::ReturnType where Args: TouchDeviceDevicesArgs

C++ method: QTouchDevice::devices

This is an overloaded function. Available variants:

Rust arguments:
1) fn devices(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListTouchDevicePtr>
2) fn devices(cpp_utils::AsStruct) -> ::list::ListTouchDevicePtr
C++ method: static QList<const QTouchDevice*> QTouchDevice::devices()

C++ documentation:

Returns a list of all registered devices.

Note: The returned list cannot be used to add new devices. Use QWindowSystemInterface::registerTouchDevice() instead.

fn maximum_touch_points(&self) -> c_int

C++ method: int QTouchDevice::maximumTouchPoints() const

C++ documentation:

Returns the maximum number of simultaneous touch points (fingers) that can be detected.

This function was introduced in Qt 5.2.

See also setMaximumTouchPoints().

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchDeviceNameArgs<'largs>

C++ method: QTouchDevice::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::string::String>
2) fn name(&self, cpp_utils::AsStruct) -> qt_core::string::String
C++ method: QString QTouchDevice::name() const

C++ documentation:

Returns the touch device name.

This string may often be empty. It is however useful for systems that have more than one touch input device because there it can be used to differentiate between the devices (i.e. to tell from which device a QTouchEvent originates from).

See also setName().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TouchDeviceNewArgs

C++ method: QTouchDevice::QTouchDevice

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::touch_device::TouchDevice
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::touch_device::TouchDevice>
C++ method: [constructor] void QTouchDevice::QTouchDevice()

fn set_capabilities(&mut self, caps: Flags<CapabilityFlag>)

C++ method: void QTouchDevice::setCapabilities(QFlags<QTouchDevice::CapabilityFlag> caps)

C++ documentation:

Sets the capabilities caps supported by the device and its driver.

See also capabilities().

fn set_maximum_touch_points(&mut self, max: c_int)

C++ method: void QTouchDevice::setMaximumTouchPoints(int max)

C++ documentation:

Sets the maximum number of simultaneous touchpoints max supported by the device and its driver.

See also maximumTouchPoints().

fn set_name(&mut self, name: &String)

C++ method: void QTouchDevice::setName(const QString& name)

C++ documentation:

Sets the name (a unique identifier) for the device. In most systems it is enough to leave this unset and keep the default empty name. This identifier becomes important when having multiple touch devices and a need to differentiate between them.

See also name().

fn set_type(&mut self, dev_type: DeviceType)

C++ method: void QTouchDevice::setType(QTouchDevice::DeviceType devType)

C++ documentation:

Sets the device type devType.

See also type().

fn type_(&self) -> DeviceType

C++ method: QTouchDevice::DeviceType QTouchDevice::type() const

C++ documentation:

Returns the touch device type.

See also setType().