Module qt_gui::touch_event::overloading

Traits

TouchEventNewArgs
This trait represents a set of arguments accepted by TouchEvent::new method.

TouchEventRegisterEventTypeArgs
This trait represents a set of arguments accepted by TouchEvent::register_event_type method.

TouchPointLastNormalizedPosArgs
This trait represents a set of arguments accepted by TouchPoint::last_normalized_pos method.

TouchPointLastPosArgs
This trait represents a set of arguments accepted by TouchPoint::last_pos method.

TouchPointLastScenePosArgs
This trait represents a set of arguments accepted by TouchPoint::last_scene_pos method.

TouchPointLastScreenPosArgs
This trait represents a set of arguments accepted by TouchPoint::last_screen_pos method.

TouchPointNewArgs
This trait represents a set of arguments accepted by TouchPoint::new method.

TouchPointNormalizedPosArgs
This trait represents a set of arguments accepted by TouchPoint::normalized_pos method.

TouchPointPosArgs
This trait represents a set of arguments accepted by TouchPoint::pos method.

TouchPointRawScreenPositionsArgs
This trait represents a set of arguments accepted by TouchPoint::raw_screen_positions method.

TouchPointRectArgs
This trait represents a set of arguments accepted by TouchPoint::rect method.

TouchPointScenePosArgs
This trait represents a set of arguments accepted by TouchPoint::scene_pos method.

TouchPointSceneRectArgs
This trait represents a set of arguments accepted by TouchPoint::scene_rect method.

TouchPointScreenPosArgs
This trait represents a set of arguments accepted by TouchPoint::screen_pos method.

TouchPointScreenRectArgs
This trait represents a set of arguments accepted by TouchPoint::screen_rect method.

TouchPointStartNormalizedPosArgs
This trait represents a set of arguments accepted by TouchPoint::start_normalized_pos method.

TouchPointStartPosArgs
This trait represents a set of arguments accepted by TouchPoint::start_pos method.

TouchPointStartScenePosArgs
This trait represents a set of arguments accepted by TouchPoint::start_scene_pos method.

TouchPointStartScreenPosArgs
This trait represents a set of arguments accepted by TouchPoint::start_screen_pos method.

TouchPointVelocityArgs
This trait represents a set of arguments accepted by TouchPoint::velocity method.