Struct qt_gui::touch_event::TouchEvent

pub struct TouchEvent {
    // some fields omitted
}

C++ type: QTouchEvent

C++ documentation:

The QTouchEvent class contains parameters that describe a touch event.

Methods

impl TouchEvent

fn accept(&mut self)

C++ method: void QTouchEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn device(&self) -> *mut TouchDevice

C++ method: QTouchDevice* QTouchEvent::device() const

C++ documentation:

Returns the touch device from which this touch event originates.

fn ignore(&mut self)

C++ method: void QTouchEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QTouchEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QTouchEvent::modifiers() const

C++ documentation:

Returns the keyboard modifier flags that existed immediately before the event occurred.

See also QGuiApplication::keyboardModifiers().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TouchEventNewArgs

C++ method: QTouchEvent::QTouchEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((qt_core::event::Type, cpp_utils::AsStruct)) -> ::touch_event::TouchEvent
2) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, cpp_utils::AsStruct)) -> ::touch_event::TouchEvent
3) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::touch_event::TouchEvent
4) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::flags::Flags<qt_core::qt::TouchPointState>, cpp_utils::AsStruct)) -> ::touch_event::TouchEvent
5) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::flags::Flags<qt_core::qt::TouchPointState>, &::list::ListTouchPointRef, cpp_utils::AsStruct)) -> ::touch_event::TouchEvent
6) fn new((qt_core::event::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchEvent>
7) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchEvent>
8) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchEvent>
9) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::flags::Flags<qt_core::qt::TouchPointState>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchEvent>
10) fn new((qt_core::event::Type, *mut ::touch_device::TouchDevice, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::flags::Flags<qt_core::qt::TouchPointState>, &::list::ListTouchPointRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchEvent>
C++ method: [constructor] void QTouchEvent::QTouchEvent(QEvent::Type eventType, QTouchDevice* device = ?, QFlags<Qt::KeyboardModifier> modifiers = ?, QFlags<Qt::TouchPointState> touchPointStates = ?, const QList<QTouchEvent::TouchPoint>& touchPoints = ?)

fn register_event_type<Args>(args: Args) -> c_int where Args: TouchEventRegisterEventTypeArgs

C++ method: QTouchEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QTouchEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QTouchEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_device(&mut self, adevice: *mut TouchDevice)

C++ method: void QTouchEvent::setDevice(QTouchDevice* adevice)

fn set_modifiers(&mut self, amodifiers: Flags<KeyboardModifier>)

C++ method: void QTouchEvent::setModifiers(QFlags<Qt::KeyboardModifier> amodifiers)

fn set_target(&mut self, atarget: *mut Object)

C++ method: void QTouchEvent::setTarget(QObject* atarget)

fn set_timestamp(&mut self, atimestamp: c_ulong)

C++ method: void QTouchEvent::setTimestamp(unsigned long atimestamp)

fn set_touch_point_states(&mut self, a_touch_point_states: Flags<TouchPointState>)

C++ method: void QTouchEvent::setTouchPointStates(QFlags<Qt::TouchPointState> aTouchPointStates)

fn set_touch_points(&mut self, atouch_points: &ListTouchPointRef)

C++ method: void QTouchEvent::setTouchPoints(const QList<QTouchEvent::TouchPoint>& atouchPoints)

fn set_window(&mut self, awindow: *mut Window)

C++ method: void QTouchEvent::setWindow(QWindow* awindow)

fn spontaneous(&self) -> bool

C++ method: bool QTouchEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn target(&self) -> *mut Object

C++ method: QObject* QTouchEvent::target() const

C++ documentation:

Returns the target object within the window on which the event occurred. This is typically a QWidget or a QQuickItem. May be 0 when no specific target is available.

fn timestamp(&self) -> c_ulong

C++ method: unsigned long QTouchEvent::timestamp() const

C++ documentation:

Returns the window system's timestamp for this event.

fn touch_point_states(&self) -> Flags<TouchPointState>

C++ method: QFlags<Qt::TouchPointState> QTouchEvent::touchPointStates() const

C++ documentation:

Returns a bitwise OR of all the touch point states for this event.

fn touch_points<'l0>(&'l0 self) -> &'l0 ListTouchPointRef

C++ method: const QList<QTouchEvent::TouchPoint>& QTouchEvent::touchPoints() const

C++ documentation:

Returns the list of touch points contained in the touch event.

fn type_(&self) -> Type

C++ method: QEvent::Type QTouchEvent::type() const

C++ documentation:

Returns the event type.

fn window(&self) -> *mut Window

C++ method: QWindow* QTouchEvent::window() const

C++ documentation:

Returns the window on which the event occurred. Useful for doing global-local mapping on data like rawScreenPositions() which, for performance reasons, only stores the global positions in the touch event.