Struct qt_gui::touch_event::TouchPoint

pub struct TouchPoint {
    // some fields omitted
}

C++ type: QTouchEvent::TouchPoint

C++ documentation:

The TouchPoint class provides information about a touch point in a QTouchEvent.

Methods

impl TouchPoint

fn flags(&self) -> Flags<InfoFlag>

C++ method: QFlags<QTouchEvent::TouchPoint::InfoFlag> QTouchEvent::TouchPoint::flags() const

C++ documentation:

Returns additional information about the touch point.

See also QTouchEvent::TouchPoint::InfoFlags.

fn id(&self) -> c_int

C++ method: int QTouchEvent::TouchPoint::id() const

C++ documentation:

Returns the id number of this touch point.

Do not assume that id numbers start at zero or that they are sequential. Such an assumption is often false due to the way the underlying drivers work.

fn last_normalized_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointLastNormalizedPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::lastNormalizedPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_normalized_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_normalized_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::lastNormalizedPos() const

C++ documentation:

Returns the normalized position of this touch point from the previous touch event.

The coordinates are normalized to the size of the touch device, i.e. (0,0) is the top-left corner and (1,1) is the bottom-right corner.

See also normalizedPos() and startNormalizedPos().

fn last_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointLastPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::lastPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::lastPos() const

C++ documentation:

Returns the position of this touch point from the previous touch event, relative to the widget or QGraphicsItem that received the event.

See also pos() and startPos().

fn last_scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointLastScenePosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::lastScenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::lastScenePos() const

C++ documentation:

Returns the scene position of this touch point from the previous touch event.

The scene position is the position in QGraphicsScene coordinates if the QTouchEvent is handled by a QGraphicsItem::touchEvent() reimplementation, and identical to the screen position for widgets.

See also scenePos() and startScenePos().

fn last_screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointLastScreenPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::lastScreenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn last_screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::lastScreenPos() const

C++ documentation:

Returns the screen position of this touch point from the previous touch event.

See also screenPos() and startScreenPos().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TouchPointNewArgs

C++ method: QTouchEvent::TouchPoint::TouchPoint

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::touch_event::TouchPoint, cpp_utils::AsStruct)) -> ::touch_event::TouchPoint
2) fn new((&::touch_event::TouchPoint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchPoint>
C++ method: [constructor] void QTouchEvent::TouchPoint::TouchPoint(const QTouchEvent::TouchPoint& other)

Variant 2

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::touch_event::TouchPoint
2) fn new(cpp_utils::AsStruct) -> ::touch_event::TouchPoint
3) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::touch_event::TouchPoint>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::touch_event::TouchPoint>
C++ method: [constructor] void QTouchEvent::TouchPoint::TouchPoint(int id = ?)

fn normalized_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointNormalizedPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::normalizedPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn normalized_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::normalizedPos() const

C++ documentation:

Returns the normalized position of this touch point.

The coordinates are normalized to the size of the touch device, i.e. (0,0) is the top-left corner and (1,1) is the bottom-right corner.

See also startNormalizedPos(), lastNormalizedPos(), and pos().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TouchPoint) -> &'l0 mut TouchPoint

C++ method: QTouchEvent::TouchPoint& QTouchEvent::TouchPoint::operator=(const QTouchEvent::TouchPoint& other)

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::pos() const

C++ documentation:

Returns the position of this touch point, relative to the widget or QGraphicsItem that received the event.

See also startPos(), lastPos(), screenPos(), scenePos(), and normalizedPos().

fn pressure(&self) -> c_double

C++ method: double QTouchEvent::TouchPoint::pressure() const

C++ documentation:

Returns the pressure of this touch point. The return value is in the range 0.0 to 1.0.

fn raw_screen_positions<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointRawScreenPositionsArgs<'largs>

C++ method: QTouchEvent::TouchPoint::rawScreenPositions

This is an overloaded function. Available variants:

Rust arguments:
1) fn raw_screen_positions(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::vector::VectorPointFRef>
2) fn raw_screen_positions(&self, cpp_utils::AsStruct) -> qt_core::vector::VectorPointFRef
C++ method: QVector<QPointF> QTouchEvent::TouchPoint::rawScreenPositions() const

C++ documentation:

Returns the raw, unfiltered positions for the touch point. The positions are in native screen coordinates. To get local coordinates you can use mapFromGlobal() of the QWindow returned by QTouchEvent::window().

Note: Returns an empty vector if the touch device's capabilities do not include QTouchDevice::RawPositions.

Note: Native screen coordinates refer to the native orientation of the screen which, in case of mobile devices, is typically portrait. This means that on systems capable of screen orientation changes the positions in this list will not reflect the current orientation (unlike pos(), screenPos(), etc.) and will always be reported in the native orientation.

This function was introduced in Qt 5.0.

See also QTouchDevice::capabilities(), device(), and window().

fn rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointRectArgs<'largs>

C++ method: QTouchEvent::TouchPoint::rect

This is an overloaded function. Available variants:

Rust arguments:
1) fn rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTouchEvent::TouchPoint::rect() const

C++ documentation:

Returns the rect for this touch point, relative to the widget or QGraphicsItem that received the event. The rect is centered around the point returned by pos().

Note: This function returns an empty rect if the device does not report touch point sizes.

fn scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointScenePosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::scenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::scenePos() const

C++ documentation:

Returns the scene position of this touch point.

The scene position is the position in QGraphicsScene coordinates if the QTouchEvent is handled by a QGraphicsItem::touchEvent() reimplementation, and identical to the screen position for widgets.

See also startScenePos(), lastScenePos(), and pos().

fn scene_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointSceneRectArgs<'largs>

C++ method: QTouchEvent::TouchPoint::sceneRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn scene_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn scene_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTouchEvent::TouchPoint::sceneRect() const

C++ documentation:

Returns the rect for this touch point in scene coordinates.

Note: This function returns an empty rect if the device does not report touch point sizes.

See also scenePos() and rect().

fn screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointScreenPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::screenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::screenPos() const

C++ documentation:

Returns the screen position of this touch point.

See also startScreenPos(), lastScreenPos(), and pos().

fn screen_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointScreenRectArgs<'largs>

C++ method: QTouchEvent::TouchPoint::screenRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn screen_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn screen_rect(&self, cpp_utils::AsStruct) -> qt_core::rect_f::RectF
C++ method: QRectF QTouchEvent::TouchPoint::screenRect() const

C++ documentation:

Returns the rect for this touch point in screen coordinates.

Note: This function returns an empty rect if the device does not report touch point sizes.

See also screenPos() and rect().

fn set_flags(&mut self, flags: Flags<InfoFlag>)

C++ method: void QTouchEvent::TouchPoint::setFlags(QFlags<QTouchEvent::TouchPoint::InfoFlag> flags)

fn set_id(&mut self, id: c_int)

C++ method: void QTouchEvent::TouchPoint::setId(int id)

fn set_last_normalized_pos(&mut self, last_normalized_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setLastNormalizedPos(const QPointF& lastNormalizedPos)

fn set_last_pos(&mut self, last_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setLastPos(const QPointF& lastPos)

fn set_last_scene_pos(&mut self, last_scene_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setLastScenePos(const QPointF& lastScenePos)

fn set_last_screen_pos(&mut self, last_screen_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setLastScreenPos(const QPointF& lastScreenPos)

fn set_normalized_pos(&mut self, normalized_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setNormalizedPos(const QPointF& normalizedPos)

fn set_pos(&mut self, pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setPos(const QPointF& pos)

fn set_pressure(&mut self, pressure: c_double)

C++ method: void QTouchEvent::TouchPoint::setPressure(double pressure)

fn set_raw_screen_positions(&mut self, positions: &VectorPointFRef)

C++ method: void QTouchEvent::TouchPoint::setRawScreenPositions(const QVector<QPointF>& positions)

fn set_rect(&mut self, rect: &RectF)

C++ method: void QTouchEvent::TouchPoint::setRect(const QRectF& rect)

fn set_scene_pos(&mut self, scene_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setScenePos(const QPointF& scenePos)

fn set_scene_rect(&mut self, scene_rect: &RectF)

C++ method: void QTouchEvent::TouchPoint::setSceneRect(const QRectF& sceneRect)

fn set_screen_pos(&mut self, screen_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setScreenPos(const QPointF& screenPos)

fn set_screen_rect(&mut self, screen_rect: &RectF)

C++ method: void QTouchEvent::TouchPoint::setScreenRect(const QRectF& screenRect)

fn set_start_normalized_pos(&mut self, start_normalized_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setStartNormalizedPos(const QPointF& startNormalizedPos)

fn set_start_pos(&mut self, start_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setStartPos(const QPointF& startPos)

fn set_start_scene_pos(&mut self, start_scene_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setStartScenePos(const QPointF& startScenePos)

fn set_start_screen_pos(&mut self, start_screen_pos: &PointF)

C++ method: void QTouchEvent::TouchPoint::setStartScreenPos(const QPointF& startScreenPos)

fn set_state(&mut self, state: Flags<TouchPointState>)

C++ method: void QTouchEvent::TouchPoint::setState(QFlags<Qt::TouchPointState> state)

fn set_velocity(&mut self, v: &Vector2D)

C++ method: void QTouchEvent::TouchPoint::setVelocity(const QVector2D& v)

fn start_normalized_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointStartNormalizedPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::startNormalizedPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_normalized_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn start_normalized_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::startNormalizedPos() const

C++ documentation:

Returns the normalized starting position of this touch point.

The coordinates are normalized to the size of the touch device, i.e. (0,0) is the top-left corner and (1,1) is the bottom-right corner.

See also normalizedPos() and lastNormalizedPos().

fn start_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointStartPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::startPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn start_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::startPos() const

C++ documentation:

Returns the starting position of this touch point, relative to the widget or QGraphicsItem that received the event.

See also pos() and lastPos().

fn start_scene_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointStartScenePosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::startScenePos

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_scene_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn start_scene_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::startScenePos() const

C++ documentation:

Returns the starting scene position of this touch point.

The scene position is the position in QGraphicsScene coordinates if the QTouchEvent is handled by a QGraphicsItem::touchEvent() reimplementation, and identical to the screen position for widgets.

See also scenePos() and lastScenePos().

fn start_screen_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointStartScreenPosArgs<'largs>

C++ method: QTouchEvent::TouchPoint::startScreenPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_screen_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn start_screen_pos(&self, cpp_utils::AsStruct) -> qt_core::point_f::PointF
C++ method: QPointF QTouchEvent::TouchPoint::startScreenPos() const

C++ documentation:

Returns the starting screen position of this touch point.

See also screenPos() and lastScreenPos().

fn state(&self) -> TouchPointState

C++ method: Qt::TouchPointState QTouchEvent::TouchPoint::state() const

fn swap(&mut self, other: &mut TouchPoint)

C++ method: void QTouchEvent::TouchPoint::swap(QTouchEvent::TouchPoint& other)

fn velocity<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TouchPointVelocityArgs<'largs>

C++ method: QTouchEvent::TouchPoint::velocity

This is an overloaded function. Available variants:

Rust arguments:
1) fn velocity(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::qvector2d::Vector2D>
2) fn velocity(&self, cpp_utils::AsStruct) -> ::qvector2d::Vector2D
C++ method: QVector2D QTouchEvent::TouchPoint::velocity() const

C++ documentation:

Returns a velocity vector for this touch point. The vector is in the screen's coordinate system, using pixels per seconds for the magnitude.

Note: The returned vector is only valid if the touch device's capabilities include QTouchDevice::Velocity.

See also QTouchDevice::capabilities() and device().