Module qt_gui::transform::overloading

Traits

OpAddArgs
This trait represents a set of arguments accepted by op_add method.

OpDivArgs
This trait represents a set of arguments accepted by op_div method.

OpMulArgs
This trait represents a set of arguments accepted by op_mul method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpSubArgs
This trait represents a set of arguments accepted by op_sub method.

TransformAdjointArgs
This trait represents a set of arguments accepted by Transform::adjoint method.

TransformAsQVariantArgs
This trait represents a set of arguments accepted by Transform::as_q_variant method.

TransformFromScaleArgs
This trait represents a set of arguments accepted by Transform::from_scale method.

TransformFromTranslateArgs
This trait represents a set of arguments accepted by Transform::from_translate method.

TransformInvertedArgs
This trait represents a set of arguments accepted by Transform::inverted method.

TransformMapArgs
This trait represents a set of arguments accepted by Transform::map method.

TransformMapRectArgs
This trait represents a set of arguments accepted by Transform::map_rect method.

TransformMapToPolygonArgs
This trait represents a set of arguments accepted by Transform::map_to_polygon method.

TransformNewArgs
This trait represents a set of arguments accepted by Transform::new method.

TransformOpMulArgs
This trait represents a set of arguments accepted by Transform::op_mul method.

TransformOpMulAssignArgs
This trait represents a set of arguments accepted by Transform::op_mul_assign method.

TransformRotateArgs
This trait represents a set of arguments accepted by Transform::rotate method.

TransformRotateRadiansArgs
This trait represents a set of arguments accepted by Transform::rotate_radians method.

TransformTransposedArgs
This trait represents a set of arguments accepted by Transform::transposed method.