Struct qt_gui::transform::Transform

pub struct Transform {
    // some fields omitted
}

C++ type: QTransform

C++ documentation:

The QTransform class specifies 2D transformations of a coordinate system.

A transformation specifies how to translate, scale, shear, rotate or project the coordinate system, and is typically used when rendering graphics.

QTransform differs from QMatrix in that it is a true 3x3 matrix, allowing perspective transformations. QTransform's toAffine() method allows casting QTransform to QMatrix. If a perspective transformation has been specified on the matrix, then the conversion will cause loss of data.

QTransform is the recommended transformation class in Qt.

A QTransform object can be built using the setMatrix(), scale(), rotate(), translate() and shear() functions. Alternatively, it can be built by applying basic matrix operations. The matrix can also be defined when constructed, and it can be reset to the identity matrix (the default) using the reset() function.

The QTransform class supports mapping of graphic primitives: A given point, line, polygon, region, or painter path can be mapped to the coordinate system defined by this matrix using the map() function. In case of a rectangle, its coordinates can be transformed using the mapRect() function. A rectangle can also be transformed into a polygon (mapped to the coordinate system defined by this matrix), using the mapToPolygon() function.

QTransform provides the isIdentity() function which returns true if the matrix is the identity matrix, and the isInvertible() function which returns true if the matrix is non-singular (i.e. AB = BA = I). The inverted() function returns an inverted copy of this matrix if it is invertible (otherwise it returns the identity matrix), and adjoint() returns the matrix's classical adjoint. In addition, QTransform provides the determinant() function which returns the matrix's determinant.

Finally, the QTransform class supports matrix multiplication, addition and subtraction, and objects of the class can be streamed as well as compared.

Methods

impl Transform

fn adjoint<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformAdjointArgs<'largs>

C++ method: QTransform::adjoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn adjoint(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
2) fn adjoint(&self, cpp_utils::AsStruct) -> ::transform::Transform
C++ method: QTransform QTransform::adjoint() const

C++ documentation:

Returns the adjoint of this matrix.

fn as_q_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformAsQVariantArgs<'largs>

C++ method: QTransform::operator QVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn as_q_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::variant::Variant>
2) fn as_q_variant(&self, cpp_utils::AsStruct) -> qt_core::variant::Variant
C++ method: QVariant QTransform::operator QVariant() const

C++ documentation:

Returns the transform as a QVariant.

fn det(&self) -> c_double

C++ method: double QTransform::det() const

C++ documentation:

Returns the matrix's determinant. Use determinant() instead.

fn determinant(&self) -> c_double

C++ method: double QTransform::determinant() const

C++ documentation:

Returns the matrix's determinant.

fn dx(&self) -> c_double

C++ method: double QTransform::dx() const

C++ documentation:

Returns the horizontal translation factor.

See also m31(), translate(), and Basic Matrix Operations.

fn dy(&self) -> c_double

C++ method: double QTransform::dy() const

C++ documentation:

Returns the vertical translation factor.

See also translate() and Basic Matrix Operations.

fn from_scale<Args>(args: Args) -> Args::ReturnType where Args: TransformFromScaleArgs

C++ method: QTransform::fromScale

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_scale((libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
2) fn from_scale((libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::transform::Transform
C++ method: static QTransform QTransform::fromScale(double dx, double dy)

C++ documentation:

Creates a matrix which corresponds to a scaling of sx horizontally and sy vertically. This is the same as QTransform().scale(sx, sy) but slightly faster.

This function was introduced in Qt 4.5.

fn from_translate<Args>(args: Args) -> Args::ReturnType where Args: TransformFromTranslateArgs

C++ method: QTransform::fromTranslate

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_translate((libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
2) fn from_translate((libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::transform::Transform
C++ method: static QTransform QTransform::fromTranslate(double dx, double dy)

C++ documentation:

Creates a matrix which corresponds to a translation of dx along the x axis and dy along the y axis. This is the same as QTransform().translate(dx, dy) but slightly faster.

This function was introduced in Qt 4.5.

fn inverted<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformInvertedArgs<'largs>

C++ method: QTransform::inverted

This is an overloaded function. Available variants:

Rust arguments:
1) fn inverted(&self, (*mut bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
2) fn inverted(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
3) fn inverted(&self, (*mut bool, cpp_utils::AsStruct)) -> ::transform::Transform
4) fn inverted(&self, cpp_utils::AsStruct) -> ::transform::Transform
C++ method: QTransform QTransform::inverted(bool* invertible = ?) const

C++ documentation:

Returns an inverted copy of this matrix.

If the matrix is singular (not invertible), the returned matrix is the identity matrix. If invertible is valid (i.e. not 0), its value is set to true if the matrix is invertible, otherwise it is set to false.

See also isInvertible().

fn is_affine(&self) -> bool

C++ method: bool QTransform::isAffine() const

C++ documentation:

Returns true if the matrix represent an affine transformation, otherwise returns false.

fn is_identity(&self) -> bool

C++ method: bool QTransform::isIdentity() const

C++ documentation:

Returns true if the matrix is the identity matrix, otherwise returns false.

See also reset().

fn is_invertible(&self) -> bool

C++ method: bool QTransform::isInvertible() const

C++ documentation:

Returns true if the matrix is invertible, otherwise returns false.

See also inverted().

fn is_rotating(&self) -> bool

C++ method: bool QTransform::isRotating() const

C++ documentation:

Returns true if the matrix represents some kind of a rotating transformation, otherwise returns false.

Note: A rotation transformation of 180 degrees and/or 360 degrees is treated as a scaling transformation.

See also reset().

fn is_scaling(&self) -> bool

C++ method: bool QTransform::isScaling() const

C++ documentation:

Returns true if the matrix represents a scaling transformation, otherwise returns false.

See also reset().

fn is_translating(&self) -> bool

C++ method: bool QTransform::isTranslating() const

C++ documentation:

Returns true if the matrix represents a translating transformation, otherwise returns false.

See also reset().

fn m11(&self) -> c_double

C++ method: double QTransform::m11() const

C++ documentation:

Returns the horizontal scaling factor.

See also scale() and Basic Matrix Operations.

fn m12(&self) -> c_double

C++ method: double QTransform::m12() const

C++ documentation:

Returns the vertical shearing factor.

See also shear() and Basic Matrix Operations.

fn m13(&self) -> c_double

C++ method: double QTransform::m13() const

C++ documentation:

Returns the horizontal projection factor.

See also translate() and Basic Matrix Operations.

fn m21(&self) -> c_double

C++ method: double QTransform::m21() const

C++ documentation:

Returns the horizontal shearing factor.

See also shear() and Basic Matrix Operations.

fn m22(&self) -> c_double

C++ method: double QTransform::m22() const

C++ documentation:

Returns the vertical scaling factor.

See also scale() and Basic Matrix Operations.

fn m23(&self) -> c_double

C++ method: double QTransform::m23() const

C++ documentation:

Returns the vertical projection factor.

See also translate() and Basic Matrix Operations.

fn m31(&self) -> c_double

C++ method: double QTransform::m31() const

C++ documentation:

Returns the horizontal translation factor.

See also dx(), translate(), and Basic Matrix Operations.

fn m32(&self) -> c_double

C++ method: double QTransform::m32() const

C++ documentation:

Returns the vertical translation factor.

See also dy(), translate(), and Basic Matrix Operations.

fn m33(&self) -> c_double

C++ method: double QTransform::m33() const

C++ documentation:

Returns the division factor.

See also translate() and Basic Matrix Operations.

fn map<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformMapArgs<'largs>

C++ method: QTransform::map

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map(&self, (&qt_core::line::Line, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::line::Line>
2) fn map(&self, (&qt_core::line::Line, cpp_utils::AsStruct)) -> qt_core::line::Line
C++ method: QLine QTransform::map(const QLine& l) const

C++ documentation:

This is an overloaded function.

Creates and returns a QLineF object that is a copy of the given line, l, mapped into the coordinate system defined by this matrix.

Variant 2

Rust arguments:
1) fn map(&self, (&qt_core::line_f::LineF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::line_f::LineF>
2) fn map(&self, (&qt_core::line_f::LineF, cpp_utils::AsStruct)) -> qt_core::line_f::LineF
C++ method: QLineF QTransform::map(const QLineF& l) const

C++ documentation:

This is an overloaded function.

Creates and returns a QLine object that is a copy of the given line, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 3

Rust arguments:
1) fn map(&self, (&::painter_path::PainterPath, cpp_utils::AsBox)) -> cpp_utils::CppBox<::painter_path::PainterPath>
2) fn map(&self, (&::painter_path::PainterPath, cpp_utils::AsStruct)) -> ::painter_path::PainterPath
C++ method: QPainterPath QTransform::map(const QPainterPath& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPainterPath object that is a copy of the given path, mapped into the coordinate system defined by this matrix.

Variant 4

Rust arguments:
1) fn map(&self, (&qt_core::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn map(&self, (&qt_core::point::Point, cpp_utils::AsStruct)) -> qt_core::point::Point
C++ method: QPoint QTransform::map(const QPoint& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPoint object that is a copy of the given point, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 5

Rust arguments:
1) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::point_f::PointF>
2) fn map(&self, (&qt_core::point_f::PointF, cpp_utils::AsStruct)) -> qt_core::point_f::PointF
C++ method: QPointF QTransform::map(const QPointF& p) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPointF object that is a copy of the given point, p, mapped into the coordinate system defined by this matrix.

Variant 6

Rust arguments:
1) fn map(&self, (&::polygon::Polygon, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn map(&self, (&::polygon::Polygon, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QTransform::map(const QPolygon& a) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPolygon object that is a copy of the given polygon, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 7

Rust arguments:
1) fn map(&self, (&::polygon_f::PolygonF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon_f::PolygonF>
2) fn map(&self, (&::polygon_f::PolygonF, cpp_utils::AsStruct)) -> ::polygon_f::PolygonF
C++ method: QPolygonF QTransform::map(const QPolygonF& a) const

C++ documentation:

This is an overloaded function.

Creates and returns a QPolygonF object that is a copy of the given polygon, mapped into the coordinate system defined by this matrix.

Variant 8

Rust arguments:
1) fn map(&self, (&::region::Region, cpp_utils::AsBox)) -> cpp_utils::CppBox<::region::Region>
2) fn map(&self, (&::region::Region, cpp_utils::AsStruct)) -> ::region::Region
C++ method: QRegion QTransform::map(const QRegion& r) const

C++ documentation:

This is an overloaded function.

Creates and returns a QRegion object that is a copy of the given region, mapped into the coordinate system defined by this matrix.

Calling this method can be rather expensive if rotations or shearing are used.

Variant 9

Rust arguments: fn map(&self, (libc::c_double, libc::c_double, *mut libc::c_double, *mut libc::c_double)) -> ()
C++ method: void QTransform::map(double x, double y, double* tx, double* ty) const

C++ documentation:

Maps the given coordinates x and y into the coordinate system defined by this matrix. The resulting values are put in *tx and ty, respectively.

The coordinates are transformed using the following formulas:

x' = m11x + m21*y + dx
y' = m22*y + m12*x + dy
if (is not affine) {
    w' = m13x + m23y + m33
    x' /= w'
    y' /= w'
}

The point (x, y) is the original point, and (x', y') is the transformed point.

See also Basic Matrix Operations.

Variant 10

Rust arguments: fn map(&self, (libc::c_int, libc::c_int, *mut libc::c_int, *mut libc::c_int)) -> ()
C++ method: void QTransform::map(int x, int y, int* tx, int* ty) const

C++ documentation:

This is an overloaded function.

Maps the given coordinates x and y into the coordinate system defined by this matrix. The resulting values are put in *tx and *ty, respectively. Note that the transformed coordinates are rounded to the nearest integer.

fn map_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformMapRectArgs<'largs>

C++ method: QTransform::mapRect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect::Rect>
2) fn map_rect(&self, (&qt_core::rect::Rect, cpp_utils::AsStruct)) -> qt_core::rect::Rect
C++ method: QRect QTransform::mapRect(const QRect& arg1) const

C++ documentation:

This is an overloaded function.

Creates and returns a QRect object that is a copy of the given rectangle, mapped into the coordinate system defined by this matrix. Note that the transformed coordinates are rounded to the nearest integer.

Variant 2

Rust arguments:
1) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<qt_core::rect_f::RectF>
2) fn map_rect(&self, (&qt_core::rect_f::RectF, cpp_utils::AsStruct)) -> qt_core::rect_f::RectF
C++ method: QRectF QTransform::mapRect(const QRectF& arg1) const

C++ documentation:

Creates and returns a QRectF object that is a copy of the given rectangle, mapped into the coordinate system defined by this matrix.

The rectangle's coordinates are transformed using the following formulas:

x' = m11*x + m21*y + dx
y' = m22y + m12x + dy
if (is not affine) {
    w' = m13*x + m23*y + m33
    x' /= w'
    y' /= w'
}

If rotation or shearing has been specified, this function returns the bounding rectangle. To retrieve the exact region the given rectangle maps to, use the mapToPolygon() function instead.

See also mapToPolygon() and Basic Matrix Operations.

fn map_to_polygon<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformMapToPolygonArgs<'largs>

C++ method: QTransform::mapToPolygon

This is an overloaded function. Available variants:

Rust arguments:
1) fn map_to_polygon(&self, (&qt_core::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::polygon::Polygon>
2) fn map_to_polygon(&self, (&qt_core::rect::Rect, cpp_utils::AsStruct)) -> ::polygon::Polygon
C++ method: QPolygon QTransform::mapToPolygon(const QRect& r) const

C++ documentation:

Creates and returns a QPolygon representation of the given rectangle, mapped into the coordinate system defined by this matrix.

The rectangle's coordinates are transformed using the following formulas:

x' = m11x + m21y + dx
y' = m22*y + m12*x + dy
if (is not affine) {
    w' = m13x + m23y + m33
    x' /= w'
    y' /= w'
}

Polygons and rectangles behave slightly differently when transformed (due to integer rounding), so matrix.map(QPolygon(rectangle)) is not always the same as matrix.mapToPolygon(rectangle).

See also mapRect() and Basic Matrix Operations.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TransformNewArgs

C++ method: QTransform::QTransform

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::transform::Transform
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
C++ method: [constructor] void QTransform::QTransform()

Variant 2

Rust arguments:
1) fn new((qt_core::qt::Initialization, cpp_utils::AsStruct)) -> ::transform::Transform
2) fn new((qt_core::qt::Initialization, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
C++ method: [constructor] void QTransform::QTransform(Qt::Initialization arg1)

Variant 3

Rust arguments:
1) fn new((&::matrix::Matrix, cpp_utils::AsStruct)) -> ::transform::Transform
2) fn new((&::matrix::Matrix, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
C++ method: [constructor] void QTransform::QTransform(const QMatrix& mtx)

Variant 4

Rust arguments:
1) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::transform::Transform
2) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::transform::Transform
3) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
4) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
C++ method: [constructor] void QTransform::QTransform(double h11, double h12, double h13, double h21, double h22, double h23, double h31, double h32, double h33 = ?)

Variant 5

Rust arguments:
1) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::transform::Transform
2) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
C++ method: [constructor] void QTransform::QTransform(double h11, double h12, double h21, double h22, double dx, double dy)

fn op_add_assign<'l0>(&'l0 mut self, div: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::operator+=(double div)

C++ documentation:

This is an overloaded function.

Returns the matrix obtained by adding the given scalar to each element of this matrix.

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Transform) -> &'l0 mut Transform

C++ method: QTransform& QTransform::operator=(const QTransform& arg1)

C++ documentation:

Assigns the given matrix's values to this matrix.

fn op_div_assign<'l0>(&'l0 mut self, div: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::operator/=(double div)

C++ documentation:

This is an overloaded function.

Returns the result of performing an element-wise division of this matrix by the given scalar.

fn op_eq(&self, arg1: &Transform) -> bool

C++ method: bool QTransform::operator==(const QTransform& arg1) const

C++ documentation:

Returns true if this matrix is equal to the given matrix, otherwise returns false.

fn op_mul<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformOpMulArgs<'largs>

C++ method: QTransform::operator*

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_mul(&self, (&::transform::Transform, cpp_utils::AsBox)) -> cpp_utils::CppBox<::transform::Transform>
2) fn op_mul(&self, (&::transform::Transform, cpp_utils::AsStruct)) -> ::transform::Transform
C++ method: QTransform QTransform::operator*(const QTransform& o) const

C++ documentation:

Returns the result of multiplying this matrix by the given matrix.

Note that matrix multiplication is not commutative, i.e. a*b != b*a.

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Transform where Args: TransformOpMulAssignArgs<'largs>

C++ method: QTransform::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, &'l1 ::transform::Transform) -> &'l0 mut ::transform::Transform
C++ method: QTransform& QTransform::operator*=(const QTransform& arg1)

C++ documentation:

This is an overloaded function.

Returns the result of multiplying this matrix by the given matrix.

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_double) -> &'l0 mut ::transform::Transform
C++ method: QTransform& QTransform::operator*=(double div)

C++ documentation:

This is an overloaded function.

Returns the result of performing an element-wise multiplication of this matrix with the given scalar.

fn op_neq(&self, arg1: &Transform) -> bool

C++ method: bool QTransform::operator!=(const QTransform& arg1) const

C++ documentation:

Returns true if this matrix is not equal to the given matrix, otherwise returns false.

fn op_sub_assign<'l0>(&'l0 mut self, div: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::operator-=(double div)

C++ documentation:

This is an overloaded function.

Returns the matrix obtained by subtracting the given scalar from each element of this matrix.

fn quad_to_quad(one: &PolygonF, two: &PolygonF, result: &mut Transform) -> bool

C++ method: static bool QTransform::quadToQuad(const QPolygonF& one, const QPolygonF& two, QTransform& result)

C++ documentation:

Creates a transformation matrix, trans, that maps a four-sided polygon, one, to another four-sided polygon, two. Returns true if the transformation is possible; otherwise returns false.

This is a convenience method combining quadToSquare() and squareToQuad() methods. It allows the input quad to be transformed into any other quad.

See also squareToQuad() and quadToSquare().

fn quad_to_square(quad: &PolygonF, result: &mut Transform) -> bool

C++ method: static bool QTransform::quadToSquare(const QPolygonF& quad, QTransform& result)

C++ documentation:

Creates a transformation matrix, trans, that maps a four-sided polygon, quad, to a unit square. Returns true if the transformation is constructed or false if such a transformation does not exist.

See also squareToQuad() and quadToQuad().

fn reset(&mut self)

C++ method: void QTransform::reset()

C++ documentation:

Resets the matrix to an identity matrix, i.e. all elements are set to zero, except m11 and m22 (specifying the scale) and m33 which are set to 1.

See also QTransform(), isIdentity(), and Basic Matrix Operations.

fn rotate<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Transform where Args: TransformRotateArgs<'largs>

C++ method: QTransform::rotate

This is an overloaded function. Available variants:

Rust arguments:
1) fn rotate(&mut self, libc::c_double) -> &'l0 mut ::transform::Transform
2) fn rotate(&mut self, (libc::c_double, qt_core::qt::Axis)) -> &'l0 mut ::transform::Transform
C++ method: QTransform& QTransform::rotate(double a, Qt::Axis axis = ?)

C++ documentation:

Rotates the coordinate system counterclockwise by the given angle about the specified axis and returns a reference to the matrix.

Note that if you apply a QTransform to a point defined in widget coordinates, the direction of the rotation will be clockwise because the y-axis points downwards.

The angle is specified in degrees.

See also setMatrix().

fn rotate_radians<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Transform where Args: TransformRotateRadiansArgs<'largs>

C++ method: QTransform::rotateRadians

This is an overloaded function. Available variants:

Rust arguments:
1) fn rotate_radians(&mut self, libc::c_double) -> &'l0 mut ::transform::Transform
2) fn rotate_radians(&mut self, (libc::c_double, qt_core::qt::Axis)) -> &'l0 mut ::transform::Transform
C++ method: QTransform& QTransform::rotateRadians(double a, Qt::Axis axis = ?)

C++ documentation:

Rotates the coordinate system counterclockwise by the given angle about the specified axis and returns a reference to the matrix.

Note that if you apply a QTransform to a point defined in widget coordinates, the direction of the rotation will be clockwise because the y-axis points downwards.

The angle is specified in radians.

See also setMatrix().

fn scale<'l0>(&'l0 mut self, sx: c_double, sy: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::scale(double sx, double sy)

C++ documentation:

Scales the coordinate system by sx horizontally and sy vertically, and returns a reference to the matrix.

See also setMatrix().

fn set_matrix(&mut self, m11: c_double, m12: c_double, m13: c_double, m21: c_double, m22: c_double, m23: c_double, m31: c_double, m32: c_double, m33: c_double)

C++ method: void QTransform::setMatrix(double m11, double m12, double m13, double m21, double m22, double m23, double m31, double m32, double m33)

C++ documentation:

Sets the matrix elements to the specified values, m11, m12, m13 m21, m22, m23 m31, m32 and m33. Note that this function replaces the previous values. QTransform provides the translate(), rotate(), scale() and shear() convenience functions to manipulate the various matrix elements based on the currently defined coordinate system.

See also QTransform().

fn shear<'l0>(&'l0 mut self, sh: c_double, sv: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::shear(double sh, double sv)

C++ documentation:

Shears the coordinate system by sh horizontally and sv vertically, and returns a reference to the matrix.

See also setMatrix().

fn square_to_quad(square: &PolygonF, result: &mut Transform) -> bool

C++ method: static bool QTransform::squareToQuad(const QPolygonF& square, QTransform& result)

C++ documentation:

Creates a transformation matrix, trans, that maps a unit square to a four-sided polygon, quad. Returns true if the transformation is constructed or false if such a transformation does not exist.

See also quadToSquare() and quadToQuad().

fn to_affine<'l0>(&'l0 self) -> &'l0 Matrix

C++ method: const QMatrix& QTransform::toAffine() const

C++ documentation:

Returns the QTransform as an affine matrix.

Warning: If a perspective transformation has been specified, then the conversion will cause loss of data.

fn translate<'l0>(&'l0 mut self, dx: c_double, dy: c_double) -> &'l0 mut Transform

C++ method: QTransform& QTransform::translate(double dx, double dy)

C++ documentation:

Moves the coordinate system dx along the x axis and dy along the y axis, and returns a reference to the matrix.

See also setMatrix().

fn transposed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TransformTransposedArgs<'largs>

C++ method: QTransform::transposed

This is an overloaded function. Available variants:

Rust arguments:
1) fn transposed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::transform::Transform>
2) fn transposed(&self, cpp_utils::AsStruct) -> ::transform::Transform
C++ method: QTransform QTransform::transposed() const

C++ documentation:

Returns the transpose of this matrix.

fn type_(&self) -> TransformationType

C++ method: QTransform::TransformationType QTransform::type() const

C++ documentation:

Returns the transformation type of this matrix.

The transformation type is the highest enumeration value capturing all of the matrix's transformations. For example, if the matrix both scales and shears, the type would be TxShear, because TxShear has a higher enumeration value than TxScale.

Knowing the transformation type of a matrix is useful for optimization: you can often handle specific types more optimally than handling the generic case.