Module qt_gui::validator::overloading

Traits

ValidatorConnectArgs
This trait represents a set of arguments accepted by Validator::connect method.

ValidatorConnectStaticArgs
This trait represents a set of arguments accepted by Validator::connect_static method.

ValidatorDestroyedArgs
This trait represents a set of arguments accepted by Validator::destroyed method.

ValidatorDisconnectArgs
This trait represents a set of arguments accepted by Validator::disconnect method.

ValidatorDisconnectStaticArgs
This trait represents a set of arguments accepted by Validator::disconnect_static method.

ValidatorDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Validator::dynamic_property_names method.

ValidatorLocaleArgs
This trait represents a set of arguments accepted by Validator::locale method.

ValidatorObjectNameArgs
This trait represents a set of arguments accepted by Validator::object_name method.

ValidatorPropertyArgs
This trait represents a set of arguments accepted by Validator::property method.

ValidatorStartTimerArgs
This trait represents a set of arguments accepted by Validator::start_timer method.

ValidatorTrArgs
This trait represents a set of arguments accepted by Validator::tr method.

ValidatorTrUtf8Args
This trait represents a set of arguments accepted by Validator::tr_utf8 method.