Struct qt_gui::vector::VectorFormatRangeRef

pub struct VectorFormatRangeRef {
    // some fields omitted
}

C++ type: QVector<QTextLayout::FormatRange>

C++ documentation:

The QVector class is a template class that provides a dynamic array.

QVector<T> is one of Qt's generic container classes. It stores its items in adjacent memory locations and provides fast index-based access.

QList<T>, QLinkedList<T>, and QVarLengthArray<T> provide similar functionality. Here's an overview:

For most purposes, QList is the right class to use. Operations like prepend() and insert() are usually faster than with QVector because of the way QList stores its items in memory (see Algorithmic Complexity for details), and its index-based API is more convenient than QLinkedList's iterator-based API. It also expands to less code in your executable.

If you need a real linked list, with guarantees of constant time insertions in the middle of the list and iterators to items rather than indexes, use QLinkedList.

If you want the items to occupy adjacent memory positions, or if your items are larger than a pointer and you want to avoid the overhead of allocating them on the heap individually at insertion time, then use QVector.

If you want a low-level variable-size array, QVarLengthArray may be sufficient.

Here's an example of a QVector that stores integers and a QVector that stores QString values:

QVector<int> integerVector;
QVector<QString> stringVector;

QVector stores a vector (or array) of items. Typically, vectors are created with an initial size. For example, the following code constructs a QVector with 200 elements:

QVector<QString> vector(200);

The elements are automatically initialized with a default-constructed value. If you want to initialize the vector with a different value, pass that value as the second argument to the constructor:

QVector<QString> vector(200, "Pass");

You can also call fill() at any time to fill the vector with a value.

QVector uses 0-based indexes, just like C++ arrays. To access the item at a particular index position, you can use operator. On non-const vectors, operator returns a reference to the item that can be used on the left side of an assignment:

if (vector[0] == "Liz")
    vector[0] = "Elizabeth";

For read-only access, an alternative syntax is to use at():

for (int i = 0; i < vector.size(); ++i) {
    if (vector.at(i) == "Alfonso")
        cout << "Found Alfonso at position " << i << endl;
}

at() can be faster than operator, because it never causes a deep copy to occur.

Another way to access the data stored in a QVector is to call data(). The function returns a pointer to the first item in the vector. You can use the pointer to directly access and modify the elements stored in the vector. The pointer is also useful if you need to pass a QVector to a function that accepts a plain C++ array.

If you want to find all occurrences of a particular value in a vector, use indexOf() or lastIndexOf(). The former searches forward starting from a given index position, the latter searches backward. Both return the index of the matching item if they found one; otherwise, they return -1. For example:

int i = vector.indexOf("Harumi");
if (i != -1)
    cout << "First occurrence of Harumi is at position " << i << endl;

If you simply want to check whether a vector contains a particular value, use contains(). If you want to find out how many times a particular value occurs in the vector, use count().

QVector provides these basic functions to add, move, and remove items: insert(), replace(), remove(), prepend(), append(). With the exception of append() and replace(), these functions can be slow (linear time) for large vectors, because they require moving many items in the vector by one position in memory. If you want a container class that provides fast insertion/removal in the middle, use QList or QLinkedList instead.

Unlike plain C++ arrays, QVectors can be resized at any time by calling resize(). If the new size is larger than the old size, QVector might need to reallocate the whole vector. QVector tries to reduce the number of reallocations by preallocating up to twice as much memory as the actual data needs.

If you know in advance approximately how many items the QVector will contain, you can call reserve(), asking QVector to preallocate a certain amount of memory. You can also call capacity() to find out how much memory QVector actually allocated.

Note that using non-const operators and functions can cause QVector to do a deep copy of the data. This is due to implicit sharing.

QVector's value type must be an assignable data type. This covers most data types that are commonly used, but the compiler won't let you, for example, store a QWidget as a value; instead, store a QWidget *. A few functions have additional requirements; for example, indexOf() and lastIndexOf() expect the value type to support operator==(). These requirements are documented on a per-function basis.

Like the other container classes, QVector provides Java-style iterators (QVectorIterator and QMutableVectorIterator) and STL-style iterators (QVector::const_iterator and QVector::iterator). In practice, these are rarely used, because you can use indexes into the QVector.

In addition to QVector, Qt also provides QVarLengthArray, a very low-level class with little functionality that is optimized for speed.

QVector does not support inserting, prepending, appending or replacing with references to its own values. Doing so will cause your application to abort with an error message.

Methods

impl VectorFormatRangeRef

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: VectorFormatRangeRefAppendArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::text_layout::FormatRange) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::append(const QTextLayout::FormatRange& t)

C++ documentation:

Inserts value at the end of the vector.

Example:

QVector<QString> vector(0);
vector.append("one");
vector.append("two");
vector.append("three");
// vector: ["one", "two", "three"]

This is the same as calling resize(size() + 1) and assigning value to the new last element in the vector.

This operation is relatively fast, because QVector typically allocates more memory than necessary, so it can grow without reallocating the entire vector each time.

See also operator<<(), prepend(), and insert().

Variant 2

Rust arguments: fn append(&mut self, &::vector::VectorFormatRangeRef) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::append(const QVector<QTextLayout::FormatRange>& l)

C++ documentation:

This is an overloaded function.

Appends the items of the value vector to this vector.

This function was introduced in Qt 5.5.

See also operator<<() and operator+=().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::at(int i) const

C++ documentation:

Returns the item at index position i in the vector.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut FormatRange

C++ method: QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last().

fn capacity(&self) -> c_int

C++ method: int QVector<QTextLayout::FormatRange>::capacity() const

C++ documentation:

Returns the maximum number of items that can be stored in the vector without forcing a reallocation.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the vector, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::clear()

C++ documentation:

Removes all the elements from the vector and releases the memory used by the vector.

fn const_data(&self) -> *const FormatRange

C++ method: const QTextLayout::FormatRange* QVector<QTextLayout::FormatRange>::constData() const

C++ documentation:

Returns a const pointer to the data stored in the vector. The pointer can be used to access the items in the vector. The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also data() and operator[]().

fn count(&self) -> c_int

C++ method: int QVector<QTextLayout::FormatRange>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

fn data(&self) -> *const FormatRange

C++ method: const QTextLayout::FormatRange* QVector<QTextLayout::FormatRange>::data() const

C++ documentation:

This is an overloaded function.

fn data_mut(&mut self) -> *mut FormatRange

C++ method: QTextLayout::FormatRange* QVector<QTextLayout::FormatRange>::data()

C++ documentation:

Returns a pointer to the data stored in the vector. The pointer can be used to access and modify the items in the vector.

Example:

QVector<int> vector(10);
int data = vector.data();
for (int i = 0; i < 10; ++i)
    data[i] = 2  i;

The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also constData() and operator[]().

fn detach(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::detach()

fn empty(&self) -> bool

C++ method: bool QVector<QTextLayout::FormatRange>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the vector is empty; otherwise returns false.

fn fill<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorFormatRangeRef where Args: VectorFormatRangeRefFillArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::fill

This is an overloaded function. Available variants:

Rust arguments:
1) fn fill(&mut self, &'l1 ::text_layout::FormatRange) -> &'l0 mut ::vector::VectorFormatRangeRef
2) fn fill(&mut self, (&'l1 ::text_layout::FormatRange, libc::c_int)) -> &'l0 mut ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::fill(const QTextLayout::FormatRange& t, int size = ?)

C++ documentation:

Assigns value to all items in the vector. If size is different from -1 (the default), the vector is resized to size size beforehand.

Example:

QVector<QString> vector(3);
vector.fill("Yes");
// vector: ["Yes", "Yes", "Yes"]

vector.fill("oh", 5); // vector: ["oh", "oh", "oh", "oh", "oh"]
See also resize().

fn first<'l0>(&'l0 self) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut FormatRange

C++ method: QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::first()

C++ documentation:

Returns a reference to the first item in the vector. This function assumes that the vector isn't empty.

See also last() and isEmpty().

fn from_list<Args>(args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefFromListArgs

C++ method: QVector<QTextLayout::FormatRange>::fromList

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_list((&::list::ListFormatRangeRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
2) fn from_list((&::list::ListFormatRangeRef, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
C++ method: static QVector<QTextLayout::FormatRange> QVector<QTextLayout::FormatRange>::fromList(const QList<QTextLayout::FormatRange>& list)

C++ documentation:

Returns a QVector object with the data contained in list.

Example:

QStringList list;
list << "Sven" << "Kim" << "Ola";

QVector<QString> vect = QVector<QString>::fromList(list); // vect: ["Sven", "Kim", "Ola"]
See also toList() and QList::toVector().

fn front<'l0>(&'l0 self) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut FormatRange

C++ method: QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: VectorFormatRangeRefInsertArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, (libc::c_int, &::text_layout::FormatRange)) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::insert(int i, const QTextLayout::FormatRange& t)

C++ documentation:

Inserts value at index position i in the vector. If i is 0, the value is prepended to the vector. If i is size(), the value is appended to the vector.

Example:

QVector<QString> vector;
vector << "alpha" << "beta" << "delta";
vector.insert(2, "gamma");
// vector: ["alpha", "beta", "gamma", "delta"]

For large vectors, this operation can be slow (linear time), because it requires moving all the items at indexes i and above by one position further in memory. If you want a container class that provides a fast insert() function, use QLinkedList instead.

See also append(), prepend(), and remove().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, libc::c_int, &::text_layout::FormatRange)) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::insert(int i, int n, const QTextLayout::FormatRange& t)

C++ documentation:

This is an overloaded function.

Inserts count copies of value at index position i in the vector.

Example:

QVector<double> vector;
vector << 2.718 << 1.442 << 0.4342;
vector.insert(1, 3, 9.9);
// vector: [2.718, 9.9, 9.9, 9.9, 1.442, 0.4342]

fn is_detached(&self) -> bool

C++ method: bool QVector<QTextLayout::FormatRange>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QVector<QTextLayout::FormatRange>::isEmpty() const

C++ documentation:

Returns true if the vector has size 0; otherwise returns false.

See also size() and resize().

fn is_shared_with(&self, other: &VectorFormatRangeRef) -> bool

C++ method: bool QVector<QTextLayout::FormatRange>::isSharedWith(const QVector<QTextLayout::FormatRange>& other) const

fn last<'l0>(&'l0 self) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::last() const

C++ documentation:

This is an overloaded function.

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut FormatRange

C++ method: QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::last()

C++ documentation:

Returns a reference to the last item in the vector. This function assumes that the vector isn't empty.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QVector<QTextLayout::FormatRange>::length() const

C++ documentation:

Same as size() and count().

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also size(), count(), and QList::length().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefMidArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange> QVector<QTextLayout::FormatRange>::mid(int pos, int len = ?) const

C++ documentation:

Returns a sub-vector which contains elements from this vector, starting at position pos. If length is -1 (the default), all elements after pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn new<Args>(args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefNewArgs

C++ method: QVector<QTextLayout::FormatRange>::QVector

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::vector::VectorFormatRangeRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
C++ method: [constructor] void QVector<QTextLayout::FormatRange>::QVector()

Variant 2

Rust arguments:
1) fn new((&::vector::VectorFormatRangeRef, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
2) fn new((&::vector::VectorFormatRangeRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
C++ method: [constructor] void QVector<QTextLayout::FormatRange>::QVector(const QVector<QTextLayout::FormatRange>& v)

Variant 3

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
C++ method: [constructor] void QVector<QTextLayout::FormatRange>::QVector(int size)

Variant 4

Rust arguments:
1) fn new((libc::c_int, &::text_layout::FormatRange, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
2) fn new((libc::c_int, &::text_layout::FormatRange, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
C++ method: [constructor] void QVector<QTextLayout::FormatRange>::QVector(int size, const QTextLayout::FormatRange& t)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefOpAddArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::vector::VectorFormatRangeRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorFormatRangeRef>
2) fn op_add(&self, (&::vector::VectorFormatRangeRef, cpp_utils::AsStruct)) -> ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange> QVector<QTextLayout::FormatRange>::operator+(const QVector<QTextLayout::FormatRange>& l) const

C++ documentation:

Returns a vector that contains all the items in this vector followed by all the items in the other vector.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorFormatRangeRef where Args: VectorFormatRangeRefOpAddAssignArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::text_layout::FormatRange) -> &'l0 mut ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::operator+=(const QTextLayout::FormatRange& t)

C++ documentation:

This is an overloaded function.

Appends value to the vector.

See also append() and operator<<().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::vector::VectorFormatRangeRef) -> &'l0 mut ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::operator+=(const QVector<QTextLayout::FormatRange>& l)

C++ documentation:

Appends the items of the other vector to this vector and returns a reference to this vector.

See also operator+() and append().

fn op_assign<'l0, 'l1>(&'l0 mut self, v: &'l1 VectorFormatRangeRef) -> &'l0 mut VectorFormatRangeRef

C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::operator=(const QVector<QTextLayout::FormatRange>& v)

C++ documentation:

Assigns other to this vector and returns a reference to this vector.

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 FormatRange

C++ method: const QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(i).

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut FormatRange

C++ method: QTextLayout::FormatRange& QVector<QTextLayout::FormatRange>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

Note that using non-const operators can cause QVector to do a deep copy.

See also at() and value().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorFormatRangeRef where Args: VectorFormatRangeRefOpShlArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::text_layout::FormatRange) -> &'l0 mut ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::operator<<(const QTextLayout::FormatRange& t)

C++ documentation:

Appends value to the vector and returns a reference to this vector.

See also append() and operator+=().

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::vector::VectorFormatRangeRef) -> &'l0 mut ::vector::VectorFormatRangeRef
C++ method: QVector<QTextLayout::FormatRange>& QVector<QTextLayout::FormatRange>::operator<<(const QVector<QTextLayout::FormatRange>& l)

C++ documentation:

Appends other to the vector and returns a reference to the vector.

fn pop_back(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast().

fn pop_front(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst().

fn prepend(&mut self, t: &FormatRange)

C++ method: void QVector<QTextLayout::FormatRange>::prepend(const QTextLayout::FormatRange& t)

C++ documentation:

Inserts value at the beginning of the vector.

Example:

QVector<QString> vector;
vector.prepend("one");
vector.prepend("two");
vector.prepend("three");
// vector: ["three", "two", "one"]

This is the same as vector.insert(0, value).

For large vectors, this operation can be slow (linear time), because it requires moving all the items in the vector by one position further in memory. If you want a container class that provides a fast prepend() function, use QList or QLinkedList instead.

See also append() and insert().

fn push_back(&mut self, t: &FormatRange)

C++ method: void QVector<QTextLayout::FormatRange>::push_back(const QTextLayout::FormatRange& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &FormatRange)

C++ method: void QVector<QTextLayout::FormatRange>::push_front(const QTextLayout::FormatRange& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove<'largs, Args>(&'largs mut self, args: Args) where Args: VectorFormatRangeRefRemoveArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&mut self, libc::c_int) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::remove(int i)

C++ documentation:

This is an overloaded function.

Removes the element at index position i.

See also insert(), replace(), and fill().

Variant 2

Rust arguments: fn remove(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QVector<QTextLayout::FormatRange>::remove(int i, int n)

C++ documentation:

This is an overloaded function.

Removes count elements from the middle of the vector, starting at index position i.

See also insert(), replace(), and fill().

fn remove_at(&mut self, i: c_int)

C++ method: void QVector<QTextLayout::FormatRange>::removeAt(int i)

C++ documentation:

Removes the element at index position i. Equivalent to

remove(i);

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also remove() and QList::removeAt().

fn remove_first(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::removeFirst()

C++ documentation:

Removes the first item in the vector. Calling this function is equivalent to calling remove(0). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeFirst(), and isEmpty().

fn remove_last(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::removeLast()

C++ documentation:

Removes the last item in the vector. Calling this function is equivalent to calling remove(size() - 1). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeLast(), removeFirst(), and isEmpty().

fn replace(&mut self, i: c_int, t: &FormatRange)

C++ method: void QVector<QTextLayout::FormatRange>::replace(int i, const QTextLayout::FormatRange& t)

C++ documentation:

Replaces the item at index position i with value.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also operator[]() and remove().

fn reserve(&mut self, size: c_int)

C++ method: void QVector<QTextLayout::FormatRange>::reserve(int size)

C++ documentation:

Attempts to allocate memory for at least size elements. If you know in advance how large the vector will be, you can call this function, and if you call resize() often you are likely to get better performance. If size is an underestimate, the worst that will happen is that the QVector will be a bit slower.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to change the size of the vector, call resize().

See also squeeze() and capacity().

fn resize(&mut self, size: c_int)

C++ method: void QVector<QTextLayout::FormatRange>::resize(int size)

C++ documentation:

Sets the size of the vector to size. If size is greater than the current size, elements are added to the end; the new elements are initialized with a default-constructed value. If size is less than the current size, elements are removed from the end.

See also size().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QVector<QTextLayout::FormatRange>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QVector<QTextLayout::FormatRange>::size() const

C++ documentation:

Returns the number of items in the vector.

See also isEmpty() and resize().

fn squeeze(&mut self)

C++ method: void QVector<QTextLayout::FormatRange>::squeeze()

C++ documentation:

Releases any memory not required to store the items.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn swap(&mut self, other: &mut VectorFormatRangeRef)

C++ method: void QVector<QTextLayout::FormatRange>::swap(QVector<QTextLayout::FormatRange>& other)

C++ documentation:

Swaps vector other with this vector. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefTakeAtArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::FormatRange>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_layout::FormatRange
C++ method: QTextLayout::FormatRange QVector<QTextLayout::FormatRange>::takeAt(int i)

C++ documentation:

Removes the element at index position i and returns it.

Equivalent to

T t = at(i);
remove(i);
return t;

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also takeFirst(), takeLast(), and QList::takeAt().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefTakeFirstArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_layout::FormatRange>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> ::text_layout::FormatRange
C++ method: QTextLayout::FormatRange QVector<QTextLayout::FormatRange>::takeFirst()

C++ documentation:

Removes the first item in the vector and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also takeLast() and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefTakeLastArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::text_layout::FormatRange>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> ::text_layout::FormatRange
C++ method: QTextLayout::FormatRange QVector<QTextLayout::FormatRange>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

If you don't use the return value, removeLast() is more efficient.

This function was introduced in Qt 5.1.

See also takeFirst() and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorFormatRangeRefValueArgs<'largs>

C++ method: QVector<QTextLayout::FormatRange>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::FormatRange>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::text_layout::FormatRange
C++ method: QTextLayout::FormatRange QVector<QTextLayout::FormatRange>::value(int i) const

C++ documentation:

Returns the value at index position i in the vector.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that i is within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &::text_layout::FormatRange, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_layout::FormatRange>
2) fn value(&self, (libc::c_int, &::text_layout::FormatRange, cpp_utils::AsStruct)) -> ::text_layout::FormatRange
C++ method: QTextLayout::FormatRange QVector<QTextLayout::FormatRange>::value(int i, const QTextLayout::FormatRange& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.