Struct qt_gui::wheel_event::WheelEvent

pub struct WheelEvent {
    // some fields omitted
}

C++ type: QWheelEvent

C++ documentation:

The QWheelEvent class contains parameters that describe a wheel event.

Wheel events are sent to the widget under the mouse cursor, but if that widget does not handle the event they are sent to the focus widget. Wheel events are generated for both mouse wheels and trackpad scroll gestures. There are two ways to read the wheel event delta: angleDelta() returns the delta in wheel degrees. This value is always provided. pixelDelta() returns the delta in screen pixels and is available on platforms that have high-resolution trackpads, such as Mac OS X. If that is the case, source() will return Qt::MouseEventSynthesizedBySystem.

The functions pos() and globalPos() return the mouse cursor's location at the time of the event.

A wheel event contains a special accept flag that indicates whether the receiver wants the event. You should call ignore() if you do not handle the wheel event; this ensures that it will be sent to the parent widget.

The QWidget::setEnabled() function can be used to enable or disable mouse and keyboard events for a widget.

The event handler QWidget::wheelEvent() receives wheel events.

Methods

impl WheelEvent

fn accept(&mut self)

C++ method: void QWheelEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn angle_delta<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WheelEventAngleDeltaArgs<'largs>

C++ method: QWheelEvent::angleDelta

This is an overloaded function. Available variants:

Rust arguments:
1) fn angle_delta(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn angle_delta(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QWheelEvent::angleDelta() const

C++ documentation:

Returns the distance that the wheel is rotated, in eighths of a degree. A positive value indicates that the wheel was rotated forwards away from the user; a negative value indicates that the wheel was rotated backwards toward the user.

Most mouse types work in steps of 15 degrees, in which case the delta value is a multiple of 120; i.e., 120 units * 1/8 = 15 degrees.

However, some mice have finer-resolution wheels and send delta values that are less than 120 units (less than 15 degrees). To support this possibility, you can either cumulatively add the delta values from events until the value of 120 is reached, then scroll the widget, or you can partially scroll the widget in response to each wheel event.

Example:

void MyWidget::wheelEvent(QWheelEvent *event)
{
    QPoint numPixels = event->pixelDelta();
    QPoint numDegrees = event->angleDelta() / 8;

if (!numPixels.isNull()) { scrollWithPixels(numPixels); } else if (!numDegrees.isNull()) { QPoint numSteps = numDegrees / 15; scrollWithDegrees(numSteps); }

event->accept(); }
Note: On platforms that support scrolling phases, the delta may be null when:

scrolling is about to begin, but the distance did not yet change (Qt::ScrollBegin),

or scrolling has ended and the distance did not change anymore (Qt::ScrollEnd).

fn buttons(&self) -> Flags<MouseButton>

C++ method: QFlags<Qt::MouseButton> QWheelEvent::buttons() const

C++ documentation:

Returns the mouse state when the event occurred.

fn delta(&self) -> c_int

C++ method: int QWheelEvent::delta() const

C++ documentation:

This function has been deprecated, use pixelDelta() or angleDelta() instead.

fn global_pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WheelEventGlobalPosArgs<'largs>

C++ method: QWheelEvent::globalPos

This is an overloaded function. Available variants:

Rust arguments:
1) fn global_pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn global_pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QWheelEvent::globalPos() const

C++ documentation:

Returns the global position of the mouse pointer at the time of the event. This is important on asynchronous window systems such as X11; whenever you move your widgets around in response to mouse events, globalPos() can differ a lot from the current cursor position returned by QCursor::pos().

See also globalX() and globalY().

fn global_pos_f<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QWheelEvent::globalPosF() const

C++ documentation:

Returns the global position of the mouse pointer at the time of the event. This is important on asynchronous window systems such as X11; whenever you move your widgets around in response to mouse events, globalPosF() can differ a lot from the current cursor position returned by QCursor::pos().

See also posF().

fn global_x(&self) -> c_int

C++ method: int QWheelEvent::globalX() const

C++ documentation:

Returns the global x position of the mouse cursor at the time of the event.

See also globalY() and globalPos().

fn global_y(&self) -> c_int

C++ method: int QWheelEvent::globalY() const

C++ documentation:

Returns the global y position of the mouse cursor at the time of the event.

See also globalX() and globalPos().

fn ignore(&mut self)

C++ method: void QWheelEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QWheelEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn modifiers(&self) -> Flags<KeyboardModifier>

C++ method: QFlags<Qt::KeyboardModifier> QWheelEvent::modifiers() const

C++ documentation:

Returns the keyboard modifier flags that existed immediately before the event occurred.

See also QGuiApplication::keyboardModifiers().

fn new<Args>(args: Args) -> Args::ReturnType where Args: WheelEventNewArgs

C++ method: QWheelEvent::QWheelEvent

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
2) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
C++ method: [constructor] void QWheelEvent::QWheelEvent(const QPointF& pos, const QPointF& globalPos, QPoint pixelDelta, QPoint angleDelta, int qt4Delta, Qt::Orientation qt4Orientation, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers)

Variant 2

Rust arguments:
1) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::ScrollPhase, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
2) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::ScrollPhase, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
C++ method: [constructor] void QWheelEvent::QWheelEvent(const QPointF& pos, const QPointF& globalPos, QPoint pixelDelta, QPoint angleDelta, int qt4Delta, Qt::Orientation qt4Orientation, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers, Qt::ScrollPhase phase)

Variant 3

Rust arguments:
1) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::ScrollPhase, qt_core::qt::MouseEventSource, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
2) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, &qt_core::point::Point, &qt_core::point::Point, libc::c_int, qt_core::qt::Orientation, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::ScrollPhase, qt_core::qt::MouseEventSource, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
C++ method: [constructor] void QWheelEvent::QWheelEvent(const QPointF& pos, const QPointF& globalPos, QPoint pixelDelta, QPoint angleDelta, int qt4Delta, Qt::Orientation qt4Orientation, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers, Qt::ScrollPhase phase, Qt::MouseEventSource source)

Variant 4

Rust arguments:
1) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
2) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::Orientation, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
3) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
4) fn new((&qt_core::point_f::PointF, &qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::Orientation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
C++ method: [constructor] void QWheelEvent::QWheelEvent(const QPointF& pos, const QPointF& globalPos, int delta, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers, Qt::Orientation orient = ?)

Variant 5

Rust arguments:
1) fn new((&qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
2) fn new((&qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::Orientation, cpp_utils::AsStruct)) -> ::wheel_event::WheelEvent
3) fn new((&qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
4) fn new((&qt_core::point_f::PointF, libc::c_int, qt_core::flags::Flags<qt_core::qt::MouseButton>, qt_core::flags::Flags<qt_core::qt::KeyboardModifier>, qt_core::qt::Orientation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::wheel_event::WheelEvent>
C++ method: [constructor] void QWheelEvent::QWheelEvent(const QPointF& pos, int delta, QFlags<Qt::MouseButton> buttons, QFlags<Qt::KeyboardModifier> modifiers, Qt::Orientation orient = ?)

fn orientation(&self) -> Orientation

C++ method: Qt::Orientation QWheelEvent::orientation() const

C++ documentation:

Returns the wheel's orientation.

Use angleDelta() instead.

fn phase(&self) -> ScrollPhase

C++ method: Qt::ScrollPhase QWheelEvent::phase() const

C++ documentation:

Returns the scrolling phase of this wheel event.

Note: The Qt::ScrollBegin and Qt::ScrollEnd phases are currently supported only on Mac OS X.

This function was introduced in Qt 5.2.

fn pixel_delta<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WheelEventPixelDeltaArgs<'largs>

C++ method: QWheelEvent::pixelDelta

This is an overloaded function. Available variants:

Rust arguments:
1) fn pixel_delta(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pixel_delta(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QWheelEvent::pixelDelta() const

C++ documentation:

Returns the scrolling distance in pixels on screen. This value is provided on platforms that support high-resolution pixel-based delta values, such as Mac OS X. The value should be used directly to scroll content on screen.

Example:

void MyWidget::wheelEvent(QWheelEvent *event)
{
    QPoint numPixels = event->pixelDelta();
    QPoint numDegrees = event->angleDelta() / 8;

if (!numPixels.isNull()) { scrollWithPixels(numPixels); } else if (!numDegrees.isNull()) { QPoint numSteps = numDegrees / 15; scrollWithDegrees(numSteps); }

event->accept(); }
Note: On platforms that support scrolling phases, the delta may be null when:

scrolling is about to begin, but the distance did not yet change (Qt::ScrollBegin),

or scrolling has ended and the distance did not change anymore (Qt::ScrollEnd).

fn pos<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WheelEventPosArgs<'largs>

C++ method: QWheelEvent::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<qt_core::point::Point>
2) fn pos(&self, cpp_utils::AsStruct) -> qt_core::point::Point
C++ method: QPoint QWheelEvent::pos() const

C++ documentation:

Returns the position of the mouse cursor relative to the widget that received the event.

If you move your widgets around in response to mouse events, use globalPos() instead of this function.

See also x(), y(), and globalPos().

fn pos_f<'l0>(&'l0 self) -> &'l0 PointF

C++ method: const QPointF& QWheelEvent::posF() const

C++ documentation:

Returns the position of the mouse cursor relative to the widget that received the event.

If you move your widgets around in response to mouse events, use globalPosF() instead of this function.

See also globalPosF().

fn register_event_type<Args>(args: Args) -> c_int where Args: WheelEventRegisterEventTypeArgs

C++ method: QWheelEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QWheelEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QWheelEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn set_modifiers(&mut self, amodifiers: Flags<KeyboardModifier>)

C++ method: void QWheelEvent::setModifiers(QFlags<Qt::KeyboardModifier> amodifiers)

fn set_timestamp(&mut self, atimestamp: c_ulong)

C++ method: void QWheelEvent::setTimestamp(unsigned long atimestamp)

fn source(&self) -> MouseEventSource

C++ method: Qt::MouseEventSource QWheelEvent::source() const

C++ documentation:

Returns information about the wheel event source.

The source can be used to distinguish between events that come from a mouse with a physical wheel and events that are generated by some other means, such as a flick gesture on a touchpad.

Note: Many platforms provide no such information. On such platforms Qt::MouseEventNotSynthesized is returned always.

This function was introduced in Qt 5.5.

See also Qt::MouseEventSource.

fn spontaneous(&self) -> bool

C++ method: bool QWheelEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn timestamp(&self) -> c_ulong

C++ method: unsigned long QWheelEvent::timestamp() const

C++ documentation:

Returns the window system's timestamp for this event.

fn type_(&self) -> Type

C++ method: QEvent::Type QWheelEvent::type() const

C++ documentation:

Returns the event type.

fn x(&self) -> c_int

C++ method: int QWheelEvent::x() const

C++ documentation:

Returns the x position of the mouse cursor, relative to the widget that received the event.

See also y() and pos().

fn y(&self) -> c_int

C++ method: int QWheelEvent::y() const

C++ documentation:

Returns the y position of the mouse cursor, relative to the widget that received the event.

See also x() and pos().